Network.AWS.EC2

Description

Amazon Elastic Compute Cloud (Amazon EC2) is a web service that provides resizable compute capacity in the cloud. It is designed to make web-scale computing easier for developers. Amazon EC2’s simple web service interface allows you to obtain and configure capacity with minimal friction. It provides you with complete control of your computing resources and lets you run on Amazon’s proven computing environment. Amazon EC2 reduces the time required to obtain and boot new server instances to minutes, allowing you to quickly scale capacity, both up and down, as your computing requirements change. Amazon EC2 changes the economics of computing by allowing you to pay only for capacity that you actually use. Amazon EC2 provides developers the tools to build failure resilient applications and isolate themselves from common failure scenarios.

Documentation

module Network.AWS.EC2.AcceptVpcPeeringConnection

module Network.AWS.EC2.AllocateAddress

module Network.AWS.EC2.AssignPrivateIpAddresses

module Network.AWS.EC2.AssociateAddress

module Network.AWS.EC2.AssociateDhcpOptions

module Network.AWS.EC2.AssociateRouteTable

module Network.AWS.EC2.AttachClassicLinkVpc

module Network.AWS.EC2.AttachInternetGateway

module Network.AWS.EC2.AttachNetworkInterface

module Network.AWS.EC2.AttachVolume

module Network.AWS.EC2.AttachVpnGateway

module Network.AWS.EC2.AuthorizeSecurityGroupEgress

module Network.AWS.EC2.AuthorizeSecurityGroupIngress

module Network.AWS.EC2.BundleInstance

module Network.AWS.EC2.CancelBundleTask

module Network.AWS.EC2.CancelConversionTask

module Network.AWS.EC2.CancelExportTask

module Network.AWS.EC2.CancelReservedInstancesListing

module Network.AWS.EC2.CancelSpotInstanceRequests

module Network.AWS.EC2.ConfirmProductInstance

module Network.AWS.EC2.CopyImage

module Network.AWS.EC2.CopySnapshot

module Network.AWS.EC2.CreateCustomerGateway

module Network.AWS.EC2.CreateDhcpOptions

module Network.AWS.EC2.CreateImage

module Network.AWS.EC2.CreateInstanceExportTask

module Network.AWS.EC2.CreateInternetGateway

module Network.AWS.EC2.CreateKeyPair

module Network.AWS.EC2.CreateNetworkAcl

module Network.AWS.EC2.CreateNetworkAclEntry

module Network.AWS.EC2.CreateNetworkInterface

module Network.AWS.EC2.CreatePlacementGroup

module Network.AWS.EC2.CreateReservedInstancesListing

module Network.AWS.EC2.CreateRoute

module Network.AWS.EC2.CreateRouteTable

module Network.AWS.EC2.CreateSecurityGroup

module Network.AWS.EC2.CreateSnapshot

module Network.AWS.EC2.CreateSpotDatafeedSubscription

module Network.AWS.EC2.CreateSubnet

module Network.AWS.EC2.CreateTags

module Network.AWS.EC2.CreateVolume

module Network.AWS.EC2.CreateVpc

module Network.AWS.EC2.CreateVpcPeeringConnection

module Network.AWS.EC2.CreateVpnConnection

module Network.AWS.EC2.CreateVpnConnectionRoute

module Network.AWS.EC2.CreateVpnGateway

module Network.AWS.EC2.DeleteCustomerGateway

module Network.AWS.EC2.DeleteDhcpOptions

module Network.AWS.EC2.DeleteInternetGateway

module Network.AWS.EC2.DeleteKeyPair

module Network.AWS.EC2.DeleteNetworkAcl

module Network.AWS.EC2.DeleteNetworkAclEntry

module Network.AWS.EC2.DeleteNetworkInterface

module Network.AWS.EC2.DeletePlacementGroup

module Network.AWS.EC2.DeleteRoute

module Network.AWS.EC2.DeleteRouteTable

module Network.AWS.EC2.DeleteSecurityGroup

module Network.AWS.EC2.DeleteSnapshot

module Network.AWS.EC2.DeleteSpotDatafeedSubscription

module Network.AWS.EC2.DeleteSubnet

module Network.AWS.EC2.DeleteTags

module Network.AWS.EC2.DeleteVolume

module Network.AWS.EC2.DeleteVpc

module Network.AWS.EC2.DeleteVpcPeeringConnection

module Network.AWS.EC2.DeleteVpnConnection

module Network.AWS.EC2.DeleteVpnConnectionRoute

module Network.AWS.EC2.DeleteVpnGateway

module Network.AWS.EC2.DeregisterImage

module Network.AWS.EC2.DescribeAccountAttributes

module Network.AWS.EC2.DescribeAddresses

module Network.AWS.EC2.DescribeAvailabilityZones

module Network.AWS.EC2.DescribeBundleTasks

module Network.AWS.EC2.DescribeClassicLinkInstances

module Network.AWS.EC2.DescribeConversionTasks

module Network.AWS.EC2.DescribeCustomerGateways

module Network.AWS.EC2.DescribeDhcpOptions

module Network.AWS.EC2.DescribeExportTasks

module Network.AWS.EC2.DescribeImageAttribute

module Network.AWS.EC2.DescribeImages

module Network.AWS.EC2.DescribeInstanceAttribute

module Network.AWS.EC2.DescribeInstanceStatus

module Network.AWS.EC2.DescribeInstances

module Network.AWS.EC2.DescribeInternetGateways

module Network.AWS.EC2.DescribeKeyPairs

module Network.AWS.EC2.DescribeNetworkAcls

module Network.AWS.EC2.DescribeNetworkInterfaceAttribute

module Network.AWS.EC2.DescribeNetworkInterfaces

module Network.AWS.EC2.DescribePlacementGroups

module Network.AWS.EC2.DescribeRegions

module Network.AWS.EC2.DescribeReservedInstances

module Network.AWS.EC2.DescribeReservedInstancesListings

module Network.AWS.EC2.DescribeReservedInstancesModifications

module Network.AWS.EC2.DescribeReservedInstancesOfferings

module Network.AWS.EC2.DescribeRouteTables

module Network.AWS.EC2.DescribeSecurityGroups

module Network.AWS.EC2.DescribeSnapshotAttribute

module Network.AWS.EC2.DescribeSnapshots

module Network.AWS.EC2.DescribeSpotDatafeedSubscription

module Network.AWS.EC2.DescribeSpotInstanceRequests

module Network.AWS.EC2.DescribeSpotPriceHistory

module Network.AWS.EC2.DescribeSubnets

module Network.AWS.EC2.DescribeTags

module Network.AWS.EC2.DescribeVolumeAttribute

module Network.AWS.EC2.DescribeVolumeStatus

module Network.AWS.EC2.DescribeVolumes

module Network.AWS.EC2.DescribeVpcAttribute

module Network.AWS.EC2.DescribeVpcClassicLink

module Network.AWS.EC2.DescribeVpcPeeringConnections

module Network.AWS.EC2.DescribeVpcs

module Network.AWS.EC2.DescribeVpnConnections

module Network.AWS.EC2.DescribeVpnGateways

module Network.AWS.EC2.DetachClassicLinkVpc

module Network.AWS.EC2.DetachInternetGateway

module Network.AWS.EC2.DetachNetworkInterface

module Network.AWS.EC2.DetachVolume

module Network.AWS.EC2.DetachVpnGateway

module Network.AWS.EC2.DisableVgwRoutePropagation

module Network.AWS.EC2.DisableVpcClassicLink

module Network.AWS.EC2.DisassociateAddress

module Network.AWS.EC2.DisassociateRouteTable

module Network.AWS.EC2.EnableVgwRoutePropagation

module Network.AWS.EC2.EnableVolumeIO

module Network.AWS.EC2.EnableVpcClassicLink

module Network.AWS.EC2.GetConsoleOutput

module Network.AWS.EC2.GetPasswordData

module Network.AWS.EC2.ImportInstance

module Network.AWS.EC2.ImportKeyPair

module Network.AWS.EC2.ImportVolume

module Network.AWS.EC2.ModifyImageAttribute

module Network.AWS.EC2.ModifyInstanceAttribute

module Network.AWS.EC2.ModifyNetworkInterfaceAttribute

module Network.AWS.EC2.ModifyReservedInstances

module Network.AWS.EC2.ModifySnapshotAttribute

module Network.AWS.EC2.ModifySubnetAttribute

module Network.AWS.EC2.ModifyVolumeAttribute

module Network.AWS.EC2.ModifyVpcAttribute

module Network.AWS.EC2.MonitorInstances

module Network.AWS.EC2.PurchaseReservedInstancesOffering

module Network.AWS.EC2.RebootInstances

module Network.AWS.EC2.RegisterImage

module Network.AWS.EC2.RejectVpcPeeringConnection

module Network.AWS.EC2.ReleaseAddress

module Network.AWS.EC2.ReplaceNetworkAclAssociation

module Network.AWS.EC2.ReplaceNetworkAclEntry

module Network.AWS.EC2.ReplaceRoute

module Network.AWS.EC2.ReplaceRouteTableAssociation

module Network.AWS.EC2.ReportInstanceStatus

module Network.AWS.EC2.RequestSpotInstances

module Network.AWS.EC2.ResetImageAttribute

module Network.AWS.EC2.ResetInstanceAttribute

module Network.AWS.EC2.ResetNetworkInterfaceAttribute

module Network.AWS.EC2.ResetSnapshotAttribute

module Network.AWS.EC2.RevokeSecurityGroupEgress

module Network.AWS.EC2.RevokeSecurityGroupIngress

module Network.AWS.EC2.RunInstances

module Network.AWS.EC2.StartInstances

module Network.AWS.EC2.StopInstances

module Network.AWS.EC2.TerminateInstances

module Network.AWS.EC2.Types

module Network.AWS.EC2.UnassignPrivateIpAddresses

module Network.AWS.EC2.UnmonitorInstances

module Network.AWS.EC2.Waiters