Network.AWS.EC2.AcceptReservedInstancesExchangeQuote

Description

Accepts the Convertible Reserved Instance exchange quote described in the GetReservedInstancesExchangeQuote call.

Creating a Request

acceptReservedInstancesExchangeQuote :: AcceptReservedInstancesExchangeQuote

Creates a value of AcceptReservedInstancesExchangeQuote with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

arieqTargetConfigurations - The configurations of the Convertible Reserved Instance offerings that you are purchasing in this exchange.

arieqDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

arieqReservedInstanceIds - The IDs of the Convertible Reserved Instances to exchange for other Convertible Reserved Instances of the same or higher value.

data AcceptReservedInstancesExchangeQuote

Contains the parameters for accepting the quote.

See: acceptReservedInstancesExchangeQuote smart constructor.

Request Lenses

arieqTargetConfigurations :: Lens' AcceptReservedInstancesExchangeQuote [TargetConfigurationRequest]

The configurations of the Convertible Reserved Instance offerings that you are purchasing in this exchange.

arieqDryRun :: Lens' AcceptReservedInstancesExchangeQuote (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

arieqReservedInstanceIds :: Lens' AcceptReservedInstancesExchangeQuote [Text]

The IDs of the Convertible Reserved Instances to exchange for other Convertible Reserved Instances of the same or higher value.

Destructuring the Response

acceptReservedInstancesExchangeQuoteResponse

Arguments

:: Int
arieqrsResponseStatus

-> AcceptReservedInstancesExchangeQuoteResponse

Creates a value of AcceptReservedInstancesExchangeQuoteResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

arieqrsExchangeId - The ID of the successful exchange.

arieqrsResponseStatus - -- | The response status code.

data AcceptReservedInstancesExchangeQuoteResponse

The result of the exchange and whether it was successful .

See: acceptReservedInstancesExchangeQuoteResponse smart constructor.

Response Lenses

arieqrsExchangeId :: Lens' AcceptReservedInstancesExchangeQuoteResponse (Maybe Text)

The ID of the successful exchange.

arieqrsResponseStatus :: Lens' AcceptReservedInstancesExchangeQuoteResponse Int

- | The response status code.