Network.AWS.EC2.AcceptVpcPeeringConnection

Description

Accept a VPC peering connection request. To accept a request, the VPC peering connection must be in the pending-acceptance state, and you must be the owner of the peer VPC. Use the DescribeVpcPeeringConnections request to view your outstanding VPC peering connection requests. Example This example accepts the specified VPC peering connection request. https:/ec2.amazonaws.com?Action=AcceptVpcPeeringConnection &amp;vpcPeeringConnectionId=pcx-1a2b3c4d &amp;AUTHPARAMS &lt;AcceptVpcPeeringConnectionResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;vpcPeeringConnection&gt; &lt;vpcPeeringConnectionId&gt;pcx-1a2b3c4d&lt;/vpcPeeringConnectionId&gt; &lt;requesterVpcInfo&gt; &lt;ownerId&gt;123456789012&lt;/ownerId&gt; &lt;vpcId&gt;vpc-1a2b3c4d&lt;/vpcId&gt; &lt;cidrBlock&gt;10.0.0.028&lt;cidrBlock&gt; &lt;/requesterVpcInfo&gt; &lt;accepterVpcInfo&gt; &lt;ownerId&gt;777788889999&lt;/ownerId&gt; &lt;vpcId&gt;vpc-111aaa22&lt;/vpcId&gt; &lt;cidrBlock&gt;10.0.1.028&lt;cidrBlock&gt; &lt;/accepterVpcInfo&gt; &lt;status&gt; &lt;code&gt;active&lt;/code&gt; &lt;message&gt;Active&lt;message&gt; &lt;status&gt; &lt;tagSet/&gt; &lt;vpcPeeringConnection&gt; &lt;AcceptVpcPeeringConnectionResponse&gt;".

Request

data AcceptVpcPeeringConnection

Request constructor

acceptVpcPeeringConnection :: AcceptVpcPeeringConnection

Smart constructor for the minimum required parameters to construct a valid AcceptVpcPeeringConnection request.

The fields accessible through corresponding lenses are:

VpcPeeringConnectionId :: Maybe Text

Request lenses

avpcVpcPeeringConnectionId :: Lens' AcceptVpcPeeringConnection (Maybe Text)

The ID of the VPC peering connection.

Response

data AcceptVpcPeeringConnectionResponse

Response constructor

acceptVpcPeeringConnectionResponse :: AcceptVpcPeeringConnectionResponse

Smart constructor for the minimum required parameters to construct a valid AcceptVpcPeeringConnectionResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

VpcPeeringConnection :: Maybe VpcPeeringConnection

Response lenses

avpcrVpcPeeringConnection :: Lens' AcceptVpcPeeringConnectionResponse (Maybe VpcPeeringConnection)

Information about the VPC peering connection.