Network.AWS.EC2.AllocateAddress

Description

Acquires an Elastic IP address. An Elastic IP address is for use either in the EC2-Classic platform or in a VPC. For more information, see Elastic IP Addresses in the Amazon Elastic Compute Cloud User Guide. Example for EC2-Classic This example request allocates an Elastic IP address for use with instances in EC2-Classic. https:/ec2.amazonaws.com?Action=AllocateAddress &amp;AUTHPARAMS &lt;AllocateAddressResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;publicIp&gt;192.0.2.1&lt;/publicIp&gt; &lt;domain&gt;standard&lt;domain&gt; &lt;AllocateAddressResponse&gt; Example for EC2-VPC This example request allocates an Elastic IP address for use with instances in a VPC. https:/ec2.amazonaws.com?Action=AllocateAddress Domain=vpc &amp;AUTHPARAMS &lt;AllocateAddressResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;publicIp&gt;198.51.100.1&lt;/publicIp&gt; &lt;domain&gt;vpc&lt;/domain&gt; &lt;allocationId&gt;eipalloc-5723d13e&lt;/allocationId&gt; &lt;/AllocateAddressResponse&gt;.

Request

data AllocateAddress

Request constructor

mkAllocateAddress :: AllocateAddress

Smart constructor for the minimum required parameters to construct a valid AllocateAddress request.

The fields accessible through corresponding lenses are:

Domain :: Maybe DomainType

Request lenses

aaDomain :: Lens' AllocateAddress (Maybe DomainType)

Set to vpc to allocate the address for use with instances in a VPC. Default: The address is for use with instances in EC2-Classic.

Response

data AllocateAddressResponse

Response constructor

mkAllocateAddressResponse :: AllocateAddressResponse

Smart constructor for the minimum required parameters to construct a valid AllocateAddressResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

PublicIp :: Maybe Text

Domain :: Maybe DomainType

AllocationId :: Maybe Text

Response lenses

aarPublicIp :: Lens' AllocateAddressResponse (Maybe Text)

The Elastic IP address.

aarDomain :: Lens' AllocateAddressResponse (Maybe DomainType)

Indicates whether this Elastic IP address is for use with instances in EC2-Classic (standard) or instances in a VPC (vpc).

aarAllocationId :: Lens' AllocateAddressResponse (Maybe Text)

EC2-VPC

The ID that AWS assigns to represent the allocation of the Elastic IP address for use with instances in a VPC.