Network.AWS.EC2.AllocateHosts

Description

Allocates a Dedicated Host to your account. At minimum you need to specify the instance size type, Availability Zone, and quantity of hosts you want to allocate.

Creating a Request

allocateHosts

Arguments

:: Text
ahAvailabilityZone

-> Text
ahInstanceType

-> Int
ahQuantity

-> AllocateHosts

Creates a value of AllocateHosts with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ahClientToken - Unique, case-sensitive identifier you provide to ensure idempotency of the request. For more information, see How to Ensure Idempotency in the Amazon Elastic Compute Cloud User Guide .

ahAutoPlacement - This is enabled by default. This property allows instances to be automatically placed onto available Dedicated Hosts, when you are launching instances without specifying a host ID. Default: Enabled

ahAvailabilityZone - The Availability Zone for the Dedicated Hosts.

ahInstanceType - Specify the instance type that you want your Dedicated Hosts to be configured for. When you specify the instance type, that is the only instance type that you can launch onto that host.

ahQuantity - The number of Dedicated Hosts you want to allocate to your account with these parameters.

data AllocateHosts

Contains the parameters for AllocateHosts.

See: allocateHosts smart constructor.

Request Lenses

ahClientToken :: Lens' AllocateHosts (Maybe Text)

Unique, case-sensitive identifier you provide to ensure idempotency of the request. For more information, see How to Ensure Idempotency in the Amazon Elastic Compute Cloud User Guide .

ahAutoPlacement :: Lens' AllocateHosts (Maybe AutoPlacement)

This is enabled by default. This property allows instances to be automatically placed onto available Dedicated Hosts, when you are launching instances without specifying a host ID. Default: Enabled

ahAvailabilityZone :: Lens' AllocateHosts Text

The Availability Zone for the Dedicated Hosts.

ahInstanceType :: Lens' AllocateHosts Text

Specify the instance type that you want your Dedicated Hosts to be configured for. When you specify the instance type, that is the only instance type that you can launch onto that host.

ahQuantity :: Lens' AllocateHosts Int

The number of Dedicated Hosts you want to allocate to your account with these parameters.

Destructuring the Response

allocateHostsResponse

Arguments

:: Int
ahrsResponseStatus

-> AllocateHostsResponse

Creates a value of AllocateHostsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ahrsHostIds - The ID of the allocated Dedicated Host. This is used when you want to launch an instance onto a specific host.

ahrsResponseStatus - -- | The response status code.

data AllocateHostsResponse

Contains the output of AllocateHosts.

See: allocateHostsResponse smart constructor.

Response Lenses

ahrsHostIds :: Lens' AllocateHostsResponse [Text]

The ID of the allocated Dedicated Host. This is used when you want to launch an instance onto a specific host.

ahrsResponseStatus :: Lens' AllocateHostsResponse Int

- | The response status code.