Network.AWS.EC2.AssignIPv6Addresses

Description

Assigns one or more IPv6 addresses to the specified network interface. You can specify one or more specific IPv6 addresses, or you can specify the number of IPv6 addresses to be automatically assigned from within the subnet's IPv6 CIDR block range. You can assign as many IPv6 addresses to a network interface as you can assign private IPv4 addresses, and the limit varies per instance type. For information, see IP Addresses Per Network Interface Per Instance Type in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

assignIPv6Addresses

Arguments

:: Text
aiaNetworkInterfaceId

-> AssignIPv6Addresses

Creates a value of AssignIPv6Addresses with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aiaIPv6AddressCount - The number of IPv6 addresses to assign to the network interface. Amazon EC2 automatically selects the IPv6 addresses from the subnet range. You can't use this option if specifying specific IPv6 addresses.

aiaIPv6Addresses - One or more specific IPv6 addresses to be assigned to the network interface. You can't use this option if you're specifying a number of IPv6 addresses.

aiaNetworkInterfaceId - The ID of the network interface.

data AssignIPv6Addresses

See: assignIPv6Addresses smart constructor.

Request Lenses

aiaIPv6AddressCount :: Lens' AssignIPv6Addresses (Maybe Int)

The number of IPv6 addresses to assign to the network interface. Amazon EC2 automatically selects the IPv6 addresses from the subnet range. You can't use this option if specifying specific IPv6 addresses.

aiaIPv6Addresses :: Lens' AssignIPv6Addresses [Text]

One or more specific IPv6 addresses to be assigned to the network interface. You can't use this option if you're specifying a number of IPv6 addresses.

aiaNetworkInterfaceId :: Lens' AssignIPv6Addresses Text

The ID of the network interface.

Destructuring the Response

assignIPv6AddressesResponse

Arguments

:: Int
aiarsResponseStatus

-> AssignIPv6AddressesResponse

Creates a value of AssignIPv6AddressesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aiarsNetworkInterfaceId - The ID of the network interface.

aiarsAssignedIPv6Addresses - The IPv6 addresses assigned to the network interface.

aiarsResponseStatus - -- | The response status code.

data AssignIPv6AddressesResponse

See: assignIPv6AddressesResponse smart constructor.

Response Lenses

aiarsNetworkInterfaceId :: Lens' AssignIPv6AddressesResponse (Maybe Text)

The ID of the network interface.

aiarsAssignedIPv6Addresses :: Lens' AssignIPv6AddressesResponse [Text]

The IPv6 addresses assigned to the network interface.

aiarsResponseStatus :: Lens' AssignIPv6AddressesResponse Int

- | The response status code.