Network.AWS.EC2.AssignPrivateIPAddresses

Description

Assigns one or more secondary private IP addresses to the specified network interface. You can specify one or more specific secondary IP addresses, or you can specify the number of secondary IP addresses to be automatically assigned within the subnet's CIDR block range. The number of secondary IP addresses that you can assign to an instance varies by instance type. For information about instance types, see Instance Types in the Amazon Elastic Compute Cloud User Guide . For more information about Elastic IP addresses, see Elastic IP Addresses in the Amazon Elastic Compute Cloud User Guide .

AssignPrivateIpAddresses is available only in EC2-VPC.

Creating a Request

assignPrivateIPAddresses

Arguments

:: Text
apiaNetworkInterfaceId

-> AssignPrivateIPAddresses

Creates a value of AssignPrivateIPAddresses with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

apiaPrivateIPAddresses - One or more IP addresses to be assigned as a secondary private IP address to the network interface. You can't specify this parameter when also specifying a number of secondary IP addresses. If you don't specify an IP address, Amazon EC2 automatically selects an IP address within the subnet range.

apiaAllowReassignment - Indicates whether to allow an IP address that is already assigned to another network interface or instance to be reassigned to the specified network interface.

apiaSecondaryPrivateIPAddressCount - The number of secondary IP addresses to assign to the network interface. You can't specify this parameter when also specifying private IP addresses.

apiaNetworkInterfaceId - The ID of the network interface.

data AssignPrivateIPAddresses

Contains the parameters for AssignPrivateIpAddresses.

See: assignPrivateIPAddresses smart constructor.

Request Lenses

apiaPrivateIPAddresses :: Lens' AssignPrivateIPAddresses [Text]

One or more IP addresses to be assigned as a secondary private IP address to the network interface. You can't specify this parameter when also specifying a number of secondary IP addresses. If you don't specify an IP address, Amazon EC2 automatically selects an IP address within the subnet range.

apiaAllowReassignment :: Lens' AssignPrivateIPAddresses (Maybe Bool)

Indicates whether to allow an IP address that is already assigned to another network interface or instance to be reassigned to the specified network interface.

apiaSecondaryPrivateIPAddressCount :: Lens' AssignPrivateIPAddresses (Maybe Int)

The number of secondary IP addresses to assign to the network interface. You can't specify this parameter when also specifying private IP addresses.

apiaNetworkInterfaceId :: Lens' AssignPrivateIPAddresses Text

The ID of the network interface.

Destructuring the Response

assignPrivateIPAddressesResponse :: AssignPrivateIPAddressesResponse

Creates a value of AssignPrivateIPAddressesResponse with the minimum fields required to make a request.

data AssignPrivateIPAddressesResponse

See: assignPrivateIPAddressesResponse smart constructor.