Network.AWS.EC2.AssociateIAMInstanceProfile

Description

Associates an IAM instance profile with a running or stopped instance. You cannot associate more than one IAM instance profile with an instance.

Creating a Request

associateIAMInstanceProfile

Arguments

:: IAMInstanceProfileSpecification
aiapIAMInstanceProfile

-> Text
aiapInstanceId

-> AssociateIAMInstanceProfile

Creates a value of AssociateIAMInstanceProfile with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aiapIAMInstanceProfile - The IAM instance profile.

aiapInstanceId - The ID of the instance.

data AssociateIAMInstanceProfile

See: associateIAMInstanceProfile smart constructor.

Request Lenses

aiapIAMInstanceProfile :: Lens' AssociateIAMInstanceProfile IAMInstanceProfileSpecification

The IAM instance profile.

aiapInstanceId :: Lens' AssociateIAMInstanceProfile Text

The ID of the instance.

Destructuring the Response

associateIAMInstanceProfileResponse

Arguments

:: Int
aiaprsResponseStatus

-> AssociateIAMInstanceProfileResponse

Creates a value of AssociateIAMInstanceProfileResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aiaprsIAMInstanceProfileAssociation - Information about the IAM instance profile association.

aiaprsResponseStatus - -- | The response status code.

data AssociateIAMInstanceProfileResponse

See: associateIAMInstanceProfileResponse smart constructor.

Response Lenses

aiaprsIAMInstanceProfileAssociation :: Lens' AssociateIAMInstanceProfileResponse (Maybe IAMInstanceProfileAssociation)

Information about the IAM instance profile association.

aiaprsResponseStatus :: Lens' AssociateIAMInstanceProfileResponse Int

- | The response status code.