Network.AWS.EC2.AssociateVPCCidrBlock

Description

Associates a CIDR block with your VPC. You can only associate a single Amazon-provided IPv6 CIDR block with your VPC. The IPv6 CIDR block size is fixed at /56.

Creating a Request

associateVPCCidrBlock

Arguments

:: Text
avcbVPCId

-> AssociateVPCCidrBlock

Creates a value of AssociateVPCCidrBlock with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

avcbAmazonProvidedIPv6CidrBlock - Requests an Amazon-provided IPv6 CIDR block with a /56 prefix length for the VPC. You cannot specify the range of IPv6 addresses, or the size of the CIDR block.

avcbVPCId - The ID of the VPC.

data AssociateVPCCidrBlock

See: associateVPCCidrBlock smart constructor.

Request Lenses

avcbAmazonProvidedIPv6CidrBlock :: Lens' AssociateVPCCidrBlock (Maybe Bool)

Requests an Amazon-provided IPv6 CIDR block with a /56 prefix length for the VPC. You cannot specify the range of IPv6 addresses, or the size of the CIDR block.

avcbVPCId :: Lens' AssociateVPCCidrBlock Text

The ID of the VPC.

Destructuring the Response

associateVPCCidrBlockResponse

Arguments

:: Int
avcbrsResponseStatus

-> AssociateVPCCidrBlockResponse

Creates a value of AssociateVPCCidrBlockResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

avcbrsVPCId - The ID of the VPC.

avcbrsIPv6CidrBlockAssociation - Information about the IPv6 CIDR block association.

avcbrsResponseStatus - -- | The response status code.

data AssociateVPCCidrBlockResponse

See: associateVPCCidrBlockResponse smart constructor.

Response Lenses

avcbrsVPCId :: Lens' AssociateVPCCidrBlockResponse (Maybe Text)

The ID of the VPC.

avcbrsIPv6CidrBlockAssociation :: Lens' AssociateVPCCidrBlockResponse (Maybe VPCIPv6CidrBlockAssociation)

Information about the IPv6 CIDR block association.

avcbrsResponseStatus :: Lens' AssociateVPCCidrBlockResponse Int

- | The response status code.