Network.AWS.EC2.AttachClassicLinkVpc

Description

Links an EC2-Classic instance to a ClassicLink-enabled VPC through one or more of the VPC's security groups. You cannot link an EC2-Classic instance to more than one VPC at a time. You can only link an instance that's in the running state. An instance is automatically unlinked from a VPC when it's stopped - you can link it to the VPC again when you restart it.

After you've linked an instance, you cannot change the VPC security groups that are associated with it. To change the security groups, you must first unlink the instance, and then link it again.

Linking your instance to a VPC is sometimes referred to as attaching your instance.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-AttachClassicLinkVpc.html

Request

data AttachClassicLinkVpc

Request constructor

attachClassicLinkVpc

Arguments

:: Text
aclvInstanceId

-> Text
aclvVpcId

-> AttachClassicLinkVpc

AttachClassicLinkVpc constructor.

The fields accessible through corresponding lenses are:

aclvDryRun :: Maybe Bool

aclvGroups :: [Text]

aclvInstanceId :: Text

aclvVpcId :: Text

Request lenses

aclvDryRun :: Lens' AttachClassicLinkVpc (Maybe Bool)

aclvGroups :: Lens' AttachClassicLinkVpc [Text]

The ID of one or more of the VPC's security groups. You cannot specify security groups from a different VPC.

aclvInstanceId :: Lens' AttachClassicLinkVpc Text

The ID of an EC2-Classic instance to link to the ClassicLink-enabled VPC.

aclvVpcId :: Lens' AttachClassicLinkVpc Text

The ID of a ClassicLink-enabled VPC.

Response

data AttachClassicLinkVpcResponse

Response constructor

attachClassicLinkVpcResponse :: AttachClassicLinkVpcResponse

AttachClassicLinkVpcResponse constructor.

The fields accessible through corresponding lenses are:

aclvrReturn :: Maybe Bool

Response lenses

aclvrReturn :: Lens' AttachClassicLinkVpcResponse (Maybe Bool)

Returns true if the request succeeds; otherwise, it returns an error.