Network.AWS.EC2.AttachInternetGateway

Description

Attaches an Internet gateway to a VPC, enabling connectivity between the Internet and the VPC. For more information about your VPC and Internet gateway, see the Amazon Virtual Private Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-AttachInternetGateway.html

Request

data AttachInternetGateway

Request constructor

attachInternetGateway

Arguments

:: Text
aigInternetGatewayId

-> Text
aigVpcId

-> AttachInternetGateway

AttachInternetGateway constructor.

The fields accessible through corresponding lenses are:

aigDryRun :: Maybe Bool

aigInternetGatewayId :: Text

aigVpcId :: Text

Request lenses

aigDryRun :: Lens' AttachInternetGateway (Maybe Bool)

aigInternetGatewayId :: Lens' AttachInternetGateway Text

The ID of the Internet gateway.

aigVpcId :: Lens' AttachInternetGateway Text

The ID of the VPC.

Response

data AttachInternetGatewayResponse

Response constructor

attachInternetGatewayResponse :: AttachInternetGatewayResponse

AttachInternetGatewayResponse constructor.