Network.AWS.EC2.AttachNetworkInterface

Description

Attaches a network interface to an instance.

Creating a Request

attachNetworkInterface

Arguments

:: Int
aniDeviceIndex

-> Text
aniInstanceId

-> Text
aniNetworkInterfaceId

-> AttachNetworkInterface

Creates a value of AttachNetworkInterface with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aniDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

aniDeviceIndex - The index of the device for the network interface attachment.

aniInstanceId - The ID of the instance.

aniNetworkInterfaceId - The ID of the network interface.

data AttachNetworkInterface

Contains the parameters for AttachNetworkInterface.

See: attachNetworkInterface smart constructor.

Request Lenses

aniDryRun :: Lens' AttachNetworkInterface (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

aniDeviceIndex :: Lens' AttachNetworkInterface Int

The index of the device for the network interface attachment.

aniInstanceId :: Lens' AttachNetworkInterface Text

The ID of the instance.

aniNetworkInterfaceId :: Lens' AttachNetworkInterface Text

The ID of the network interface.

Destructuring the Response

attachNetworkInterfaceResponse

Arguments

:: Int
anirsResponseStatus

-> AttachNetworkInterfaceResponse

Creates a value of AttachNetworkInterfaceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

anirsAttachmentId - The ID of the network interface attachment.

anirsResponseStatus - -- | The response status code.

data AttachNetworkInterfaceResponse

Contains the output of AttachNetworkInterface.

See: attachNetworkInterfaceResponse smart constructor.

Response Lenses

anirsAttachmentId :: Lens' AttachNetworkInterfaceResponse (Maybe Text)

The ID of the network interface attachment.

anirsResponseStatus :: Lens' AttachNetworkInterfaceResponse Int

- | The response status code.