Network.AWS.EC2.AttachVPNGateway

Description

Attaches a virtual private gateway to a VPC. You can attach one virtual private gateway to one VPC at a time.

For more information, see Adding a Hardware Virtual Private Gateway to Your VPC in the Amazon Virtual Private Cloud User Guide .

Creating a Request

attachVPNGateway

Arguments

:: Text
avgVPCId

-> Text
avgVPNGatewayId

-> AttachVPNGateway

Creates a value of AttachVPNGateway with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

avgDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

avgVPCId - The ID of the VPC.

avgVPNGatewayId - The ID of the virtual private gateway.

data AttachVPNGateway

Contains the parameters for AttachVpnGateway.

See: attachVPNGateway smart constructor.

Request Lenses

avgDryRun :: Lens' AttachVPNGateway (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

avgVPCId :: Lens' AttachVPNGateway Text

The ID of the VPC.

avgVPNGatewayId :: Lens' AttachVPNGateway Text

The ID of the virtual private gateway.

Destructuring the Response

attachVPNGatewayResponse

Arguments

:: Int
avgrsResponseStatus

-> AttachVPNGatewayResponse

Creates a value of AttachVPNGatewayResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

avgrsVPCAttachment - Information about the attachment.

avgrsResponseStatus - -- | The response status code.

data AttachVPNGatewayResponse

Contains the output of AttachVpnGateway.

See: attachVPNGatewayResponse smart constructor.

Response Lenses

avgrsVPCAttachment :: Lens' AttachVPNGatewayResponse (Maybe VPCAttachment)

Information about the attachment.

avgrsResponseStatus :: Lens' AttachVPNGatewayResponse Int

- | The response status code.