Network.AWS.EC2.AttachVolume

Description

Attaches an Amazon EBS volume to a running or stopped instance and exposes it to the instance with the specified device name. Encrypted Amazon EBS volumes may only be attached to instances that support Amazon EBS encryption. For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide. For a list of supported device names, see Attaching an Amazon EBS Volume to an Instance. Any device names that aren't reserved for instance store volumes can be used for Amazon EBS volumes. For more information, see Amazon EC2 Instance Store in the Amazon Elastic Compute Cloud User Guide. If a volume has an AWS Marketplace product code: The volume can only be attached as the root device of a stopped instance. You must be subscribed to the AWS Marketplace code that is on the volume. The configuration (instance type, operating system) of the instance must support that specific AWS Marketplace code. For example, you cannot take a volume from a Windows instance and attach it to a Linux instance. AWS Marketplace product codes are copied from the volume to the instance. For an overview of the AWS Marketplace, see https:/aws.amazon.commarketplacehelp200900000. For more information about how to use the AWS Marketplace, see AWS Marketplace. For more information about Amazon EBS volumes, see Attaching Amazon EBS Volumes in the Amazon Elastic Compute Cloud User Guide. Example 1 This example request attaches the volume with the ID vol-1a2b3c4d to the instance with the ID i-1a2b3c4d and exposes it as devsdh. https:/ec2.amazonaws.com?Action=AttachVolume &amp;VolumeId=vol-1a2b3c4d &amp;InstanceId=i-1a2b3c4d &amp;Device=devsdh &amp;AUTHPARAMS &lt;AttachVolumeResponse xmlns="http://ec2.amazonaws.com/doc/2014-05-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;volumeId&gt;vol-1a2b3c4d&lt;/volumeId&gt; &lt;instanceId&gt;i-1a2b3c4d&lt;/instanceId&gt; &lt;device&gt;devsdh&lt;/device&gt; &lt;status&gt;attaching&lt;/status&gt; &lt;attachTime&gt;YYYY-MM-DDTHH:MM:SS.000Z&lt;/attachTime&gt; &lt;/AttachVolumeResponse&gt;.

Request

data AttachVolume

Request constructor

attachVolume

Arguments

:: Text
avVolumeId

-> Text
avInstanceId

-> Text
avDevice

-> AttachVolume

Smart constructor for the minimum required parameters to construct a valid AttachVolume request.

The fields accessible through corresponding lenses are:

VolumeId :: Text

InstanceId :: Text

Device :: Text

Request lenses

avVolumeId :: Lens' AttachVolume Text

The ID of the Amazon EBS volume. The volume and instance must be within the same Availability Zone.

avInstanceId :: Lens' AttachVolume Text

The ID of the instance.

avDevice :: Lens' AttachVolume Text

The device name to expose to the instance (for example, devsdh or xvdh).

Response

data AttachVolumeResponse

Response constructor

attachVolumeResponse :: AttachVolumeResponse

Smart constructor for the minimum required parameters to construct a valid AttachVolumeResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

VolumeId :: Maybe Text

InstanceId :: Maybe Text

Device :: Maybe Text

State :: Maybe VolumeAttachmentState

AttachTime :: Maybe ISO8601

DeleteOnTermination :: Maybe Bool

Response lenses

avrVolumeId :: Lens' AttachVolumeResponse (Maybe Text)

The ID of the volume.

avrInstanceId :: Lens' AttachVolumeResponse (Maybe Text)

The ID of the instance.

avrDevice :: Lens' AttachVolumeResponse (Maybe Text)

The device name.

avrState :: Lens' AttachVolumeResponse (Maybe VolumeAttachmentState)

The attachment state of the volume.

avrAttachTime :: Lens' AttachVolumeResponse (Maybe ISO8601)

The time stamp when the attachment initiated.

avrDeleteOnTermination :: Lens' AttachVolumeResponse (Maybe Bool)

Indicates whether the Amazon EBS volume is deleted on instance termination.