Network.AWS.EC2.AuthorizeSecurityGroupEgress

Description

Adds one or more egress rules to a security group for use with a VPC. Specifically, this action permits instances to send traffic to one or more destination CIDR IP address ranges, or to one or more destination security groups for the same VPC.

You can have up to 50 rules per security group (covering both ingress and egress rules).

A security group is for use with instances either in the EC2-Classic platform or in a specific VPC. This action doesn't apply to security groups for use in EC2-Classic. For more information, see Security Groups for Your VPC in the Amazon Virtual Private Cloud User Guide.

Each rule consists of the protocol (for example, TCP), plus either a CIDR range or a source group. For the TCP and UDP protocols, you must also specify the destination port or port range. For the ICMP protocol, you must also specify the ICMP type and code. You can use -1 for the type or code to mean all types or all codes.

Rule changes are propagated to affected instances as quickly as possible. However, a small delay might occur.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-AuthorizeSecurityGroupEgress.html

Request

data AuthorizeSecurityGroupEgress

Request constructor

authorizeSecurityGroupEgress

Arguments

:: Text
asgeGroupId

-> AuthorizeSecurityGroupEgress

AuthorizeSecurityGroupEgress constructor.

The fields accessible through corresponding lenses are:

asgeCidrIp :: Maybe Text

asgeDryRun :: Maybe Bool

asgeFromPort :: Maybe Int

asgeGroupId :: Text

asgeIpPermissions :: [IpPermission]

asgeIpProtocol :: Maybe Text

asgeSourceSecurityGroupName :: Maybe Text

asgeSourceSecurityGroupOwnerId :: Maybe Text

asgeToPort :: Maybe Int

Request lenses

asgeCidrIp :: Lens' AuthorizeSecurityGroupEgress (Maybe Text)

The CIDR IP address range. You can't specify this parameter when specifying a source security group.

asgeDryRun :: Lens' AuthorizeSecurityGroupEgress (Maybe Bool)

asgeFromPort :: Lens' AuthorizeSecurityGroupEgress (Maybe Int)

The start of port range for the TCP and UDP protocols, or an ICMP type number. For the ICMP type number, use '-1' to specify all ICMP types.

asgeGroupId :: Lens' AuthorizeSecurityGroupEgress Text

The ID of the security group.

asgeIpPermissions :: Lens' AuthorizeSecurityGroupEgress [IpPermission]

A set of IP permissions. You can't specify a destination security group and a CIDR IP address range.

asgeIpProtocol :: Lens' AuthorizeSecurityGroupEgress (Maybe Text)

The IP protocol name (tcp, udp, icmp) or number (see Protocol Numbers). Use '-1' to specify all.

asgeSourceSecurityGroupName :: Lens' AuthorizeSecurityGroupEgress (Maybe Text)

EC2-Classic, default VPC

The name of the destination security group. You can't specify a destination security group and a CIDR IP address range.

asgeSourceSecurityGroupOwnerId :: Lens' AuthorizeSecurityGroupEgress (Maybe Text)

The ID of the destination security group. You can't specify a destination security group and a CIDR IP address range.

asgeToPort :: Lens' AuthorizeSecurityGroupEgress (Maybe Int)

The end of port range for the TCP and UDP protocols, or an ICMP code number. For the ICMP code number, use '-1' to specify all ICMP codes for the ICMP type.

Response

data AuthorizeSecurityGroupEgressResponse

Response constructor

authorizeSecurityGroupEgressResponse :: AuthorizeSecurityGroupEgressResponse

AuthorizeSecurityGroupEgressResponse constructor.