Network.AWS.EC2.AuthorizeSecurityGroupIngress

Description

Adds one or more ingress rules to a security group.

EC2-Classic: You can have up to 100 rules per group.

EC2-VPC: You can have up to 50 rules per group (covering both ingress and egress rules).

Rule changes are propagated to instances within the security group as quickly as possible. However, a small delay might occur.

EC2-Classic

This action gives one or more CIDR IP address ranges permission to access a security group in your account, or gives one or more security groups (called the source groups) permission to access a security group for your account. A source group can be for your own AWS account, or another.

EC2-VPC

This action gives one or more CIDR IP address ranges permission to access a security group in your VPC, or gives one or more other security groups (called the source groups) permission to access a security group for your VPC. The security groups must all be for the same VPC.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-AuthorizeSecurityGroupIngress.html

Request

data AuthorizeSecurityGroupIngress

Request constructor

authorizeSecurityGroupIngress :: AuthorizeSecurityGroupIngress

AuthorizeSecurityGroupIngress constructor.

The fields accessible through corresponding lenses are:

asgiCidrIp :: Maybe Text

asgiDryRun :: Maybe Bool

asgiFromPort :: Maybe Int

asgiGroupId :: Maybe Text

asgiGroupName :: Maybe Text

asgiIpPermissions :: [IpPermission]

asgiIpProtocol :: Maybe Text

asgiSourceSecurityGroupName :: Maybe Text

asgiSourceSecurityGroupOwnerId :: Maybe Text

asgiToPort :: Maybe Int

Request lenses

asgiCidrIp :: Lens' AuthorizeSecurityGroupIngress (Maybe Text)

The CIDR IP address range. You can't specify this parameter when specifying a source security group.

asgiDryRun :: Lens' AuthorizeSecurityGroupIngress (Maybe Bool)

asgiFromPort :: Lens' AuthorizeSecurityGroupIngress (Maybe Int)

The start of port range for the TCP and UDP protocols, or an ICMP type number. For the ICMP type number, use '-1' to specify all ICMP types.

asgiGroupId :: Lens' AuthorizeSecurityGroupIngress (Maybe Text)

The ID of the security group. Required for a nondefault VPC.

asgiGroupName :: Lens' AuthorizeSecurityGroupIngress (Maybe Text)

EC2-Classic, default VPC

The name of the security group.

asgiIpPermissions :: Lens' AuthorizeSecurityGroupIngress [IpPermission]

A set of IP permissions. Can be used to specify multiple rules in a single command.

asgiIpProtocol :: Lens' AuthorizeSecurityGroupIngress (Maybe Text)

The IP protocol name (tcp, udp, icmp) or number (see Protocol Numbers). (VPC only) Use '-1' to specify all.

asgiSourceSecurityGroupName :: Lens' AuthorizeSecurityGroupIngress (Maybe Text)

EC2-Classic, default VPC

The name of the source security group. You can't specify a source security group and a CIDR IP address range.

asgiSourceSecurityGroupOwnerId :: Lens' AuthorizeSecurityGroupIngress (Maybe Text)

The ID of the source security group. You can't specify a source security group and a CIDR IP address range.

asgiToPort :: Lens' AuthorizeSecurityGroupIngress (Maybe Int)

The end of port range for the TCP and UDP protocols, or an ICMP code number. For the ICMP code number, use '-1' to specify all ICMP codes for the ICMP type.

Response

data AuthorizeSecurityGroupIngressResponse

Response constructor

authorizeSecurityGroupIngressResponse :: AuthorizeSecurityGroupIngressResponse

AuthorizeSecurityGroupIngressResponse constructor.