Network.AWS.EC2.BundleInstance

Description

Bundles an Amazon instance store-backed Windows instance.

During bundling, only the root device volume (C:) is bundled. Data on other instance store volumes is not preserved.

This action is not applicable for Linux/Unix instances or Windows instances that are backed by Amazon EBS.

For more information, see Creating an Instance Store-Backed Windows AMI.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-BundleInstance.html

Request

data BundleInstance

Request constructor

bundleInstance

Arguments

:: Text
biInstanceId

-> Storage
biStorage

-> BundleInstance

BundleInstance constructor.

The fields accessible through corresponding lenses are:

biDryRun :: Maybe Bool

biInstanceId :: Text

biStorage :: Storage

Request lenses

biDryRun :: Lens' BundleInstance (Maybe Bool)

biInstanceId :: Lens' BundleInstance Text

The ID of the instance to bundle.

Type: String

Default: None

Required: Yes

biStorage :: Lens' BundleInstance Storage

The bucket in which to store the AMI. You can specify a bucket that you already own or a new bucket that Amazon EC2 creates on your behalf. If you specify a bucket that belongs to someone else, Amazon EC2 returns an error.

Response

data BundleInstanceResponse

Response constructor

bundleInstanceResponse :: BundleInstanceResponse

BundleInstanceResponse constructor.

The fields accessible through corresponding lenses are:

birBundleTask :: Maybe BundleTask

Response lenses

birBundleTask :: Lens' BundleInstanceResponse (Maybe BundleTask)

Information about the bundle task.