Network.AWS.EC2.CancelBundleTask

Description

Cancels a bundling operation for an instance store-backed Windows instance.

Creating a Request

cancelBundleTask

Arguments

:: Text
cbtBundleId

-> CancelBundleTask

Creates a value of CancelBundleTask with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cbtDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cbtBundleId - The ID of the bundle task.

data CancelBundleTask

Contains the parameters for CancelBundleTask.

See: cancelBundleTask smart constructor.

Request Lenses

cbtDryRun :: Lens' CancelBundleTask (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cbtBundleId :: Lens' CancelBundleTask Text

The ID of the bundle task.

Destructuring the Response

cancelBundleTaskResponse

Arguments

:: Int
cbtrsResponseStatus

-> CancelBundleTaskResponse

Creates a value of CancelBundleTaskResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cbtrsBundleTask - Information about the bundle task.

cbtrsResponseStatus - -- | The response status code.

data CancelBundleTaskResponse

Contains the output of CancelBundleTask.

See: cancelBundleTaskResponse smart constructor.

Response Lenses

cbtrsBundleTask :: Lens' CancelBundleTaskResponse (Maybe BundleTask)

Information about the bundle task.

cbtrsResponseStatus :: Lens' CancelBundleTaskResponse Int

- | The response status code.