Network.AWS.EC2.CancelConversionTask

Description

Cancels an active conversion task. The task can be the import of an instance or volume. The action removes all artifacts of the conversion, including a partially uploaded volume or instance. If the conversion is complete or is in the process of transferring the final disk image, the command fails and returns an exception.

For more information, see Using the Command Line Tools to Import YourVirtual Machine to Amazon EC2 in the Amazon Elastic Compute Cloud User Guidefor Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CancelConversionTask.html

Request

data CancelConversionTask

Request constructor

cancelConversionTask

Arguments

:: Text
cctConversionTaskId

-> CancelConversionTask

CancelConversionTask constructor.

The fields accessible through corresponding lenses are:

cctConversionTaskId :: Text

cctDryRun :: Maybe Bool

cctReasonMessage :: Maybe Text

Request lenses

cctConversionTaskId :: Lens' CancelConversionTask Text

The ID of the conversion task.

cctDryRun :: Lens' CancelConversionTask (Maybe Bool)

cctReasonMessage :: Lens' CancelConversionTask (Maybe Text)

Response

data CancelConversionTaskResponse

Response constructor

cancelConversionTaskResponse :: CancelConversionTaskResponse

CancelConversionTaskResponse constructor.