Network.AWS.EC2.CancelExportTask

Description

Cancels an active export task. The request removes all artifacts of the export, including any partially-created Amazon S3 objects. If the export task is complete or is in the process of transferring the final disk image, the command fails and returns an error.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CancelExportTask.html

Request

data CancelExportTask

Request constructor

cancelExportTask

Arguments

:: Text
cetExportTaskId

-> CancelExportTask

CancelExportTask constructor.

The fields accessible through corresponding lenses are:

cetExportTaskId :: Text

Request lenses

cetExportTaskId :: Lens' CancelExportTask Text

The ID of the export task. This is the ID returned by CreateInstanceExportTask .

Response

data CancelExportTaskResponse

Response constructor

cancelExportTaskResponse :: CancelExportTaskResponse

CancelExportTaskResponse constructor.