Network.AWS.EC2.CancelImportTask

Description

Cancels an in-process import virtual machine or import snapshot task.

Creating a Request

cancelImportTask :: CancelImportTask

Creates a value of CancelImportTask with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

citCancelReason - The reason for canceling the task.

citImportTaskId - The ID of the import image or import snapshot task to be canceled.

citDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data CancelImportTask

Contains the parameters for CancelImportTask.

See: cancelImportTask smart constructor.

Request Lenses

citCancelReason :: Lens' CancelImportTask (Maybe Text)

The reason for canceling the task.

citImportTaskId :: Lens' CancelImportTask (Maybe Text)

The ID of the import image or import snapshot task to be canceled.

citDryRun :: Lens' CancelImportTask (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

cancelImportTaskResponse

Arguments

:: Int
citrsResponseStatus

-> CancelImportTaskResponse

Creates a value of CancelImportTaskResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

citrsState - The current state of the task being canceled.

citrsImportTaskId - The ID of the task being canceled.

citrsPreviousState - The current state of the task being canceled.

citrsResponseStatus - -- | The response status code.

data CancelImportTaskResponse

Contains the output for CancelImportTask.

See: cancelImportTaskResponse smart constructor.

Response Lenses

citrsState :: Lens' CancelImportTaskResponse (Maybe Text)

The current state of the task being canceled.

citrsImportTaskId :: Lens' CancelImportTaskResponse (Maybe Text)

The ID of the task being canceled.

citrsPreviousState :: Lens' CancelImportTaskResponse (Maybe Text)

The current state of the task being canceled.

citrsResponseStatus :: Lens' CancelImportTaskResponse Int

- | The response status code.