Network.AWS.EC2.CancelSpotFleetRequests

Description

Cancels the specified Spot fleet requests.

After you cancel a Spot fleet request, the Spot fleet launches no new Spot instances. You must specify whether the Spot fleet should also terminate its Spot instances. If you terminate the instances, the Spot fleet request enters the cancelled_terminating state. Otherwise, the Spot fleet request enters the cancelled_running state and the instances continue to run until they are interrupted or you terminate them manually.

Creating a Request

cancelSpotFleetRequests

Arguments

:: Bool
csfrTerminateInstances

-> CancelSpotFleetRequests

Creates a value of CancelSpotFleetRequests with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csfrDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

csfrSpotFleetRequestIds - The IDs of the Spot fleet requests.

csfrTerminateInstances - Indicates whether to terminate instances for a Spot fleet request if it is canceled successfully.

data CancelSpotFleetRequests

Contains the parameters for CancelSpotFleetRequests.

See: cancelSpotFleetRequests smart constructor.

Request Lenses

csfrDryRun :: Lens' CancelSpotFleetRequests (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

csfrSpotFleetRequestIds :: Lens' CancelSpotFleetRequests [Text]

The IDs of the Spot fleet requests.

csfrTerminateInstances :: Lens' CancelSpotFleetRequests Bool

Indicates whether to terminate instances for a Spot fleet request if it is canceled successfully.

Destructuring the Response

cancelSpotFleetRequestsResponse

Arguments

:: Int
csfrrsResponseStatus

-> CancelSpotFleetRequestsResponse

Creates a value of CancelSpotFleetRequestsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csfrrsSuccessfulFleetRequests - Information about the Spot fleet requests that are successfully canceled.

csfrrsUnsuccessfulFleetRequests - Information about the Spot fleet requests that are not successfully canceled.

csfrrsResponseStatus - -- | The response status code.

data CancelSpotFleetRequestsResponse

Contains the output of CancelSpotFleetRequests.

See: cancelSpotFleetRequestsResponse smart constructor.

Response Lenses

csfrrsSuccessfulFleetRequests :: Lens' CancelSpotFleetRequestsResponse [CancelSpotFleetRequestsSuccessItem]

Information about the Spot fleet requests that are successfully canceled.

csfrrsUnsuccessfulFleetRequests :: Lens' CancelSpotFleetRequestsResponse [CancelSpotFleetRequestsErrorItem]

Information about the Spot fleet requests that are not successfully canceled.

csfrrsResponseStatus :: Lens' CancelSpotFleetRequestsResponse Int

- | The response status code.