Network.AWS.EC2.CancelSpotInstanceRequests

Description

Cancels one or more Spot Instance requests. Spot Instances are instances that Amazon EC2 starts on your behalf when the bid price that you specify exceeds the current Spot Price. Amazon EC2 periodically sets the Spot Price based on available Spot Instance capacity and current Spot Instance requests. For more information, see Spot Instance Requests in the Amazon Elastic Compute CloudUser Guide for Linux.

Canceling a Spot Instance request does not terminate running Spot Instances associated with the request.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CancelSpotInstanceRequests.html

Request

data CancelSpotInstanceRequests

Request constructor

cancelSpotInstanceRequests :: CancelSpotInstanceRequests

CancelSpotInstanceRequests constructor.

The fields accessible through corresponding lenses are:

csirDryRun :: Maybe Bool

csirSpotInstanceRequestIds :: [Text]

Request lenses

csirDryRun :: Lens' CancelSpotInstanceRequests (Maybe Bool)

csirSpotInstanceRequestIds :: Lens' CancelSpotInstanceRequests [Text]

One or more Spot Instance request IDs.

Response

data CancelSpotInstanceRequestsResponse

Response constructor

cancelSpotInstanceRequestsResponse :: CancelSpotInstanceRequestsResponse

CancelSpotInstanceRequestsResponse constructor.

The fields accessible through corresponding lenses are:

csirrCancelledSpotInstanceRequests :: [CancelledSpotInstanceRequest]

Response lenses

csirrCancelledSpotInstanceRequests :: Lens' CancelSpotInstanceRequestsResponse [CancelledSpotInstanceRequest]

One or more Spot Instance requests.