Network.AWS.EC2.ConfirmProductInstance

Description

Determines whether a product code is associated with an instance. This action can only be used by the owner of the product code. It is useful when a product code owner needs to verify whether another user's instance is eligible for support.

Creating a Request

confirmProductInstance

Arguments

:: Text
cpiInstanceId

-> Text
cpiProductCode

-> ConfirmProductInstance

Creates a value of ConfirmProductInstance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpiDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cpiInstanceId - The ID of the instance.

cpiProductCode - The product code. This must be a product code that you own.

data ConfirmProductInstance

Contains the parameters for ConfirmProductInstance.

See: confirmProductInstance smart constructor.

Request Lenses

cpiDryRun :: Lens' ConfirmProductInstance (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cpiInstanceId :: Lens' ConfirmProductInstance Text

The ID of the instance.

cpiProductCode :: Lens' ConfirmProductInstance Text

The product code. This must be a product code that you own.

Destructuring the Response

confirmProductInstanceResponse

Arguments

:: Int
cpirsResponseStatus

-> ConfirmProductInstanceResponse

Creates a value of ConfirmProductInstanceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpirsReturn - The return value of the request. Returns true if the specified product code is owned by the requester and associated with the specified instance.

cpirsOwnerId - The AWS account ID of the instance owner. This is only present if the product code is attached to the instance.

cpirsResponseStatus - -- | The response status code.

data ConfirmProductInstanceResponse

Contains the output of ConfirmProductInstance.

See: confirmProductInstanceResponse smart constructor.

Response Lenses

cpirsReturn :: Lens' ConfirmProductInstanceResponse (Maybe Bool)

The return value of the request. Returns true if the specified product code is owned by the requester and associated with the specified instance.

cpirsOwnerId :: Lens' ConfirmProductInstanceResponse (Maybe Text)

The AWS account ID of the instance owner. This is only present if the product code is attached to the instance.

cpirsResponseStatus :: Lens' ConfirmProductInstanceResponse Int

- | The response status code.