Network.AWS.EC2.CopyImage

Description

Initiates the copy of an AMI from the specified source region to the current region. You specify the destination region by using its endpoint when making the request. AMIs that use encrypted Amazon EBS snapshots cannot be copied with this method.

For more information, see Copying AMIs in the Amazon Elastic Compute CloudUser Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CopyImage.html

Request

data CopyImage

Request constructor

copyImage

Arguments

:: Text
ciSourceRegion

-> Text
ciSourceImageId

-> Text
ciName

-> CopyImage

CopyImage constructor.

The fields accessible through corresponding lenses are:

ciClientToken :: Maybe Text

ciDescription :: Maybe Text

ciDryRun :: Maybe Bool

ciName :: Text

ciSourceImageId :: Text

ciSourceRegion :: Text

Request lenses

ciClientToken :: Lens' CopyImage (Maybe Text)

Unique, case-sensitive identifier you provide to ensure idempotency of the request. For more information, see How to Ensure Idempotency in the AmazonElastic Compute Cloud User Guide for Linux.

ciDescription :: Lens' CopyImage (Maybe Text)

A description for the new AMI in the destination region.

ciDryRun :: Lens' CopyImage (Maybe Bool)

ciName :: Lens' CopyImage Text

The name of the new AMI in the destination region.

ciSourceImageId :: Lens' CopyImage Text

The ID of the AMI to copy.

ciSourceRegion :: Lens' CopyImage Text

The name of the region that contains the AMI to copy.

Response

data CopyImageResponse

Response constructor

copyImageResponse :: CopyImageResponse

CopyImageResponse constructor.

The fields accessible through corresponding lenses are:

cir1ImageId :: Maybe Text

Response lenses

cir1ImageId :: Lens' CopyImageResponse (Maybe Text)

The ID of the new AMI.