Network.AWS.EC2.CopySnapshot

Description

Copies a point-in-time snapshot of an Amazon EBS volume and stores it in Amazon S3. You can copy the snapshot within the same region or from one region to another. You can use the snapshot to create Amazon EBS volumes or Amazon Machine Images (AMIs). The snapshot is copied to the regional endpoint that you send the HTTP request to. Copies of encrypted Amazon EBS snapshots remain encrypted. Copies of unencrypted snapshots remain unencrypted. For more information, see Copying an Amazon EBS Snapshot in the Amazon Elastic Compute Cloud User Guide. Example This example request copies the snapshot in the us-west-1 region with the ID snap-1a2b3c4d. https:/ec2.amazonaws.com?Action=CopySnapshot &amp;SourceRegion=us-west-1 &amp;SourceSnapshotId=snap-1a2b3c4d &amp;Description=My_snapshot &amp;AUTHPARAMS &lt;CopySnapshotResponse xmlns="http://ec2.amazonaws.com/doc/2014-05-01/"&gt; &lt;requestId&gt;60bc441d-fa2c-494d-b155-5d6a3EXAMPLE&lt;/requestId&gt; &lt;snapshotId&gt;snap-2a2b3c4d&lt;/snapshotId&gt; &lt;/CopySnapshotResponse&gt;.

Request

data CopySnapshot

Request constructor

mkCopySnapshot

Arguments

:: Text
csSourceRegion

-> Text
csSourceSnapshotId

-> CopySnapshot

Smart constructor for the minimum required parameters to construct a valid CopySnapshot request.

The fields accessible through corresponding lenses are:

SourceRegion :: Text

SourceSnapshotId :: Text

Description :: Maybe Text

DestinationRegion :: Maybe Text

PresignedUrl :: Maybe Text

Request lenses

csSourceRegion :: Lens' CopySnapshot Text

The ID of the region that contains the snapshot to be copied.

csSourceSnapshotId :: Lens' CopySnapshot Text

The ID of the Amazon EBS snapshot to copy.

csDescription :: Lens' CopySnapshot (Maybe Text)

A description for the new Amazon EBS snapshot.

csDestinationRegion :: Lens' CopySnapshot (Maybe Text)

The destination region of the snapshot copy operation. This parameter is required in the PresignedUrl.

csPresignedUrl :: Lens' CopySnapshot (Maybe Text)

The pre-signed URL that facilitates copying an encrypted snapshot. This parameter is only required when copying an encrypted snapshot with the Amazon EC2 Query API; it is available as an optional parameter in all other cases. The PresignedUrl should use the snapshot source endpoint, the CopySnapshot action, and include the SourceRegion, SourceSnapshotId, and DestinationRegion parameters. The PresignedUrl must be signed using AWS Signature Version 4. Because Amazon EBS snapshots are stored in Amazon S3, the signing algorithm for this parameter uses the same logic that is described in Authenticating Requests by Using Query Parameters (AWS Signature Version 4) in the Amazon Simple Storage Service API Reference. An invalid or improperly signed PresignedUrl will cause the copy operation to fail asynchronously, and the snapshot will move to an error state.

Response

data CopySnapshotResponse

Response constructor

mkCopySnapshotResponse :: CopySnapshotResponse

Smart constructor for the minimum required parameters to construct a valid CopySnapshotResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

SnapshotId :: Maybe Text

Response lenses

csrSnapshotId :: Lens' CopySnapshotResponse (Maybe Text)

The ID of the new snapshot.