Network.AWS.EC2.CreateCustomerGateway

Description

Provides information to AWS about your VPN customer gateway device. The customer gateway is the appliance at your end of the VPN connection. (The device on the AWS side of the VPN connection is the virtual private gateway.) You must provide the Internet-routable IP address of the customer gateway's external interface. The IP address must be static and can't be behind a device performing network address translation (NAT). For devices that use Border Gateway Protocol (BGP), you can also provide the device's BGP Autonomous System Number (ASN). You can use an existing ASN assigned to your network. If you don't have an ASN already, you can use a private ASN (in the 64512 - 65534 range). For more information about VPN customer gateways, see Adding a Hardware Virtual Private Gateway to Your VPC in the Amazon Virtual Private Cloud User Guide.

Request

data CreateCustomerGateway

Request constructor

createCustomerGateway

Arguments

:: Text
ccgType

-> Text
ccgPublicIp

-> Int
ccgBgpAsn

-> CreateCustomerGateway

CreateCustomerGateway constructor.

The fields accessible through corresponding lenses are:

ccgBgpAsn :: Int

ccgDryRun :: Maybe Bool

ccgPublicIp :: Text

ccgType :: Text

Request lenses

ccgBgpAsn :: Lens' CreateCustomerGateway Int

For devices that support BGP, the customer gateway's BGP ASN. Default: 65000.

ccgDryRun :: Lens' CreateCustomerGateway (Maybe Bool)

ccgPublicIp :: Lens' CreateCustomerGateway Text

The Internet-routable IP address for the customer gateway's outside interface. The address must be static.

ccgType :: Lens' CreateCustomerGateway Text

The type of VPN connection that this customer gateway supports (ipsec.1).

Response

data CreateCustomerGatewayResult

Response constructor

createCustomerGatewayResult :: CreateCustomerGatewayResult

CreateCustomerGatewayResult constructor.

The fields accessible through corresponding lenses are:

ccgrCustomerGateway :: Maybe CustomerGateway

Response lenses

ccgrCustomerGateway :: Lens' CreateCustomerGatewayResult (Maybe CustomerGateway)

Information about the customer gateway.