Network.AWS.EC2.CreateDefaultVPC

Description

Creates a default VPC with a size /16 IPv4 CIDR block and a default subnet in each Availability Zone. For more information about the components of a default VPC, see Default VPC and Default Subnets in the Amazon Virtual Private Cloud User Guide . You cannot specify the components of the default VPC yourself.

You can create a default VPC if you deleted your previous default VPC. You cannot have more than one default VPC per region.

If your account supports EC2-Classic, you cannot use this action to create a default VPC in a region that supports EC2-Classic. If you want a default VPC in a region that supports EC2-Classic, see "I really want a default VPC for my existing EC2 account. Is that possible?" in the Default VPCs FAQ .

Creating a Request

createDefaultVPC :: CreateDefaultVPC

Creates a value of CreateDefaultVPC with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdvDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data CreateDefaultVPC

Contains the parameters for CreateDefaultVpc.

See: createDefaultVPC smart constructor.

Request Lenses

cdvDryRun :: Lens' CreateDefaultVPC (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

createDefaultVPCResponse

Arguments

:: Int
cdvrsResponseStatus

-> CreateDefaultVPCResponse

Creates a value of CreateDefaultVPCResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdvrsVPC - Information about the VPC.

cdvrsResponseStatus - -- | The response status code.

data CreateDefaultVPCResponse

Contains the output of CreateDefaultVpc.

See: createDefaultVPCResponse smart constructor.

Response Lenses

cdvrsVPC :: Lens' CreateDefaultVPCResponse (Maybe VPC)

Information about the VPC.

cdvrsResponseStatus :: Lens' CreateDefaultVPCResponse Int

- | The response status code.