Network.AWS.EC2.CreateDhcpOptions

Description

Creates a set of DHCP options for your VPC. After creating the set, you must associate it with the VPC, causing all existing and new instances that you launch in the VPC to use this set of DHCP options. The following are the individual DHCP options you can specify. For more information about the options, see RFC 2132.

'domain-name-servers' - The IP addresses of up to four domain name servers, or AmazonProvidedDNS. The default DHCP option set specifies AmazonProvidedDNS. If specifying more than one domain name server, specify the IP addresses in a single parameter, separated by commas. 'domain-name' - If you're using AmazonProvidedDNS in 'us-east-1', specify 'ec2.internal'. If you're using AmazonProvidedDNS in another region, specify 'region.compute.internal' (for example, 'ap-northeast-1.compute.internal'). Otherwise, specify a domain name (for example, com). Important: Some Linux operating systems accept multiple domain names separated by spaces. However, Windows and other Linux operating systems treat the value as a single domain, which results in unexpected behavior. If your DHCP options set is associated with a VPC that has instances with multiple operating systems, specify only one domain name. 'ntp-servers' - The IP addresses of up to four Network Time Protocol (NTP) servers. 'netbios-name-servers' - The IP addresses of up to four NetBIOS name servers. 'netbios-node-type' - The NetBIOS node type (1, 2, 4, or 8). We recommend that you specify 2 (broadcast and multicast are not currently supported). For more information about these node types, see RFC 2132. Your VPC automatically starts out with a set of DHCP options that includes only a DNS server that we provide (AmazonProvidedDNS). If you create a set of options, and if your VPC has an Internet gateway, make sure to set the 'domain-name-servers' option either to AmazonProvidedDNS or to a domain name server of your choice. For more information about DHCP options, see DHCP Options Sets in the Amazon Virtual Private Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateDhcpOptions.html

Request

data CreateDhcpOptions

Request constructor

createDhcpOptions :: CreateDhcpOptions

CreateDhcpOptions constructor.

The fields accessible through corresponding lenses are:

cdoDhcpConfigurations :: [NewDhcpConfiguration]

cdoDryRun :: Maybe Bool

Request lenses

cdoDhcpConfigurations :: Lens' CreateDhcpOptions [NewDhcpConfiguration]

A DHCP configuration option.

cdoDryRun :: Lens' CreateDhcpOptions (Maybe Bool)

Response

data CreateDhcpOptionsResponse

Response constructor

createDhcpOptionsResponse :: CreateDhcpOptionsResponse

CreateDhcpOptionsResponse constructor.

The fields accessible through corresponding lenses are:

cdorDhcpOptions :: Maybe DhcpOptions

Response lenses

cdorDhcpOptions :: Lens' CreateDhcpOptionsResponse (Maybe DhcpOptions)

A set of DHCP options.