Network.AWS.EC2.CreateImage

Description

Creates an Amazon EBS-backed AMI from an Amazon EBS-backed instance that is either running or stopped.

If you customized your instance with instance store volumes or EBS volumes in addition to the root device volume, the new AMI contains block device mapping information for those volumes. When you launch an instance from this new AMI, the instance automatically launches with those additional volumes.

For more information, see Creating Amazon EBS-Backed Linux AMIs in the Amazon Elastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateImage.html

Request

data CreateImage

Request constructor

createImage

Arguments

:: Text
ci1InstanceId

-> Text
ci1Name

-> CreateImage

CreateImage constructor.

The fields accessible through corresponding lenses are:

ci1BlockDeviceMappings :: [BlockDeviceMapping]

ci1Description :: Maybe Text

ci1DryRun :: Maybe Bool

ci1InstanceId :: Text

ci1Name :: Text

ci1NoReboot :: Maybe Bool

Request lenses

ci1BlockDeviceMappings :: Lens' CreateImage [BlockDeviceMapping]

Information about one or more block device mappings.

ci1Description :: Lens' CreateImage (Maybe Text)

A description for the new image.

ci1DryRun :: Lens' CreateImage (Maybe Bool)

ci1InstanceId :: Lens' CreateImage Text

The ID of the instance.

ci1Name :: Lens' CreateImage Text

A name for the new image.

Constraints: 3-128 alphanumeric characters, parentheses (()), square brackets ([]), spaces ( ), periods (.), slashes (/), dashes (-), single quotes ('), at-signs (@), or underscores(_)

ci1NoReboot :: Lens' CreateImage (Maybe Bool)

By default, this parameter is set to false, which means Amazon EC2 attempts to shut down the instance cleanly before image creation and then reboots the instance. When the parameter is set to true, Amazon EC2 doesn't shut down the instance before creating the image. When this option is used, file system integrity on the created image can't be guaranteed.

Response

data CreateImageResponse

Response constructor

createImageResponse :: CreateImageResponse

CreateImageResponse constructor.

The fields accessible through corresponding lenses are:

cirImageId :: Maybe Text

Response lenses

cirImageId :: Lens' CreateImageResponse (Maybe Text)

The ID of the new AMI.