Network.AWS.EC2.CreateInstanceExportTask

Description

Exports a running or stopped instance to an Amazon S3 bucket.

For information about the supported operating systems, image formats, and known limitations for the types of instances you can export, see ExportingEC2 Instances in the Amazon Elastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateInstanceExportTask.html

Request

data CreateInstanceExportTask

Request constructor

createInstanceExportTask

Arguments

:: Text
cietInstanceId

-> CreateInstanceExportTask

CreateInstanceExportTask constructor.

The fields accessible through corresponding lenses are:

cietDescription :: Maybe Text

cietExportToS3Task :: Maybe ExportToS3TaskSpecification

cietInstanceId :: Text

cietTargetEnvironment :: Maybe ExportEnvironment

Request lenses

cietDescription :: Lens' CreateInstanceExportTask (Maybe Text)

A description for the conversion task or the resource being exported. The maximum length is 255 bytes.

cietExportToS3Task :: Lens' CreateInstanceExportTask (Maybe ExportToS3TaskSpecification)

cietInstanceId :: Lens' CreateInstanceExportTask Text

The ID of the instance.

cietTargetEnvironment :: Lens' CreateInstanceExportTask (Maybe ExportEnvironment)

The target virtualization environment.

Response

data CreateInstanceExportTaskResponse

Response constructor

createInstanceExportTaskResponse :: CreateInstanceExportTaskResponse

CreateInstanceExportTaskResponse constructor.

The fields accessible through corresponding lenses are:

cietrExportTask :: Maybe ExportTask

Response lenses

cietrExportTask :: Lens' CreateInstanceExportTaskResponse (Maybe ExportTask)