Network.AWS.EC2.CreateInternetGateway

Description

Creates an Internet gateway for use with a VPC. After creating the Internet gateway, you attach it to a VPC using AttachInternetGateway.

For more information about your VPC and Internet gateway, see the AmazonVirtual Private Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateInternetGateway.html

Request

data CreateInternetGateway

Request constructor

createInternetGateway :: CreateInternetGateway

CreateInternetGateway constructor.

The fields accessible through corresponding lenses are:

cigDryRun :: Maybe Bool

Request lenses

cigDryRun :: Lens' CreateInternetGateway (Maybe Bool)

Response

data CreateInternetGatewayResponse

Response constructor

createInternetGatewayResponse :: CreateInternetGatewayResponse

CreateInternetGatewayResponse constructor.

The fields accessible through corresponding lenses are:

cigrInternetGateway :: Maybe InternetGateway

Response lenses

cigrInternetGateway :: Lens' CreateInternetGatewayResponse (Maybe InternetGateway)

Information about the Internet gateway.