Network.AWS.EC2.CreateKeyPair

Description

Creates a 2048-bit RSA key pair with the specified name. Amazon EC2 stores the public key and displays the private key for you to save to a file. The private key is returned as an unencrypted PEM encoded PKCS#8 private key. If a key with the specified name already exists, Amazon EC2 returns an error. You can have up to five thousand key pairs per region. The key pair returned to you is available only in the region in which you create it. To create a key pair that is available in all regions, use ImportKeyPair. For more information about key pairs, see Key Pairs in the Amazon Elastic Compute Cloud User Guide.

Request

data CreateKeyPair

Request constructor

createKeyPair

Arguments

:: Text
ckpKeyName

-> CreateKeyPair

CreateKeyPair constructor.

The fields accessible through corresponding lenses are:

ckpDryRun :: Maybe Bool

ckpKeyName :: Text

Request lenses

ckpDryRun :: Lens' CreateKeyPair (Maybe Bool)

ckpKeyName :: Lens' CreateKeyPair Text

A unique name for the key pair. Constraints: Up to 255 ASCII characters.

Response

data KeyPair

Response constructor

keyPair :: KeyPair

KeyPair constructor.

The fields accessible through corresponding lenses are:

kpKeyFingerprint :: Maybe Text

kpKeyMaterial :: Maybe Text

kpKeyName :: Maybe Text

Response lenses

kpKeyFingerprint :: Lens' KeyPair (Maybe Text)

The SHA-1 digest of the DER encoded private key.

kpKeyMaterial :: Lens' KeyPair (Maybe Text)

An unencrypted PEM encoded RSA private key.

kpKeyName :: Lens' KeyPair (Maybe Text)

The name of the key pair.