Network.AWS.EC2.CreateNetworkACLEntry

Description

Creates an entry (a rule) in a network ACL with the specified rule number. Each network ACL has a set of numbered ingress rules and a separate set of numbered egress rules. When determining whether a packet should be allowed in or out of a subnet associated with the ACL, we process the entries in the ACL according to the rule numbers, in ascending order. Each network ACL has a set of ingress rules and a separate set of egress rules.

We recommend that you leave room between the rule numbers (for example, 100, 110, 120, ...), and not number them one right after the other (for example, 101, 102, 103, ...). This makes it easier to add a rule between existing ones without having to renumber the rules.

After you add an entry, you can't modify it; you must either replace it, or create an entry and delete the old one.

For more information about network ACLs, see Network ACLs in the Amazon Virtual Private Cloud User Guide .

Creating a Request

createNetworkACLEntry

Arguments

:: Bool
cnaeEgress

-> Text
cnaeNetworkACLId

-> Text
cnaeProtocol

-> RuleAction
cnaeRuleAction

-> Int
cnaeRuleNumber

-> CreateNetworkACLEntry

Creates a value of CreateNetworkACLEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cnaeIPv6CidrBlock - The IPv6 network range to allow or deny, in CIDR notation (for example 2001:db8:1234:1a00::/64 ).

cnaeICMPTypeCode - ICMP protocol: The ICMP or ICMPv6 type and code. Required if specifying the ICMP protocol, or protocol 58 (ICMPv6) with an IPv6 CIDR block.

cnaePortRange - TCP or UDP protocols: The range of ports the rule applies to.

cnaeCidrBlock - The IPv4 network range to allow or deny, in CIDR notation (for example 172.16.0.0/24 ).

cnaeDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cnaeEgress - Indicates whether this is an egress rule (rule is applied to traffic leaving the subnet).

cnaeNetworkACLId - The ID of the network ACL.

cnaeProtocol - The protocol. A value of -1 or all means all protocols. If you specify all , -1 , or a protocol number other than tcp , udp , or icmp , traffic on all ports is allowed, regardless of any ports or ICMP types or codes you specify. If you specify protocol 58 (ICMPv6) and specify an IPv4 CIDR block, traffic for all ICMP types and codes allowed, regardless of any that you specify. If you specify protocol 58 (ICMPv6) and specify an IPv6 CIDR block, you must specify an ICMP type and code.

cnaeRuleAction - Indicates whether to allow or deny the traffic that matches the rule.

cnaeRuleNumber - The rule number for the entry (for example, 100). ACL entries are processed in ascending order by rule number. Constraints: Positive integer from 1 to 32766. The range 32767 to 65535 is reserved for internal use.

data CreateNetworkACLEntry

Contains the parameters for CreateNetworkAclEntry.

See: createNetworkACLEntry smart constructor.

Request Lenses

cnaeIPv6CidrBlock :: Lens' CreateNetworkACLEntry (Maybe Text)

The IPv6 network range to allow or deny, in CIDR notation (for example 2001:db8:1234:1a00::/64 ).

cnaeICMPTypeCode :: Lens' CreateNetworkACLEntry (Maybe ICMPTypeCode)

ICMP protocol: The ICMP or ICMPv6 type and code. Required if specifying the ICMP protocol, or protocol 58 (ICMPv6) with an IPv6 CIDR block.

cnaePortRange :: Lens' CreateNetworkACLEntry (Maybe PortRange)

TCP or UDP protocols: The range of ports the rule applies to.

cnaeCidrBlock :: Lens' CreateNetworkACLEntry (Maybe Text)

The IPv4 network range to allow or deny, in CIDR notation (for example 172.16.0.0/24 ).

cnaeDryRun :: Lens' CreateNetworkACLEntry (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cnaeEgress :: Lens' CreateNetworkACLEntry Bool

Indicates whether this is an egress rule (rule is applied to traffic leaving the subnet).

cnaeNetworkACLId :: Lens' CreateNetworkACLEntry Text

The ID of the network ACL.

cnaeProtocol :: Lens' CreateNetworkACLEntry Text

The protocol. A value of -1 or all means all protocols. If you specify all , -1 , or a protocol number other than tcp , udp , or icmp , traffic on all ports is allowed, regardless of any ports or ICMP types or codes you specify. If you specify protocol 58 (ICMPv6) and specify an IPv4 CIDR block, traffic for all ICMP types and codes allowed, regardless of any that you specify. If you specify protocol 58 (ICMPv6) and specify an IPv6 CIDR block, you must specify an ICMP type and code.

cnaeRuleAction :: Lens' CreateNetworkACLEntry RuleAction

Indicates whether to allow or deny the traffic that matches the rule.

cnaeRuleNumber :: Lens' CreateNetworkACLEntry Int

The rule number for the entry (for example, 100). ACL entries are processed in ascending order by rule number. Constraints: Positive integer from 1 to 32766. The range 32767 to 65535 is reserved for internal use.

Destructuring the Response

createNetworkACLEntryResponse :: CreateNetworkACLEntryResponse

Creates a value of CreateNetworkACLEntryResponse with the minimum fields required to make a request.

data CreateNetworkACLEntryResponse

See: createNetworkACLEntryResponse smart constructor.