Network.AWS.EC2.CreateNetworkAclEntry

Description

Creates an entry (a rule) in a network ACL with the specified rule number. Each network ACL has a set of numbered ingress rules and a separate set of numbered egress rules. When determining whether a packet should be allowed in or out of a subnet associated with the ACL, we process the entries in the ACL according to the rule numbers, in ascending order. Each network ACL has a set of ingress rules and a separate set of egress rules.

We recommend that you leave room between the rule numbers (for example, 100, 110, 120, ...), and not number them one right after the other (for example, 101, 102, 103, ...). This makes it easier to add a rule between existing ones without having to renumber the rules.

After you add an entry, you can't modify it; you must either replace it, or create an entry and delete the old one.

For more information about network ACLs, see Network ACLs in the AmazonVirtual Private Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateNetworkAclEntry.html

Request

data CreateNetworkAclEntry

Request constructor

createNetworkAclEntry

Arguments

:: Text
cnaeNetworkAclId

-> Int
cnaeRuleNumber

-> Text
cnaeProtocol

-> RuleAction
cnaeRuleAction

-> Bool
cnaeEgress

-> Text
cnaeCidrBlock

-> CreateNetworkAclEntry

CreateNetworkAclEntry constructor.

The fields accessible through corresponding lenses are:

cnaeCidrBlock :: Text

cnaeDryRun :: Maybe Bool

cnaeEgress :: Bool

cnaeIcmpTypeCode :: Maybe IcmpTypeCode

cnaeNetworkAclId :: Text

cnaePortRange :: Maybe PortRange

cnaeProtocol :: Text

cnaeRuleAction :: RuleAction

cnaeRuleNumber :: Int

Request lenses

cnaeCidrBlock :: Lens' CreateNetworkAclEntry Text

The network range to allow or deny, in CIDR notation (for example '172.16.0.0/24').

cnaeDryRun :: Lens' CreateNetworkAclEntry (Maybe Bool)

cnaeEgress :: Lens' CreateNetworkAclEntry Bool

Indicates whether this is an egress rule (rule is applied to traffic leaving the subnet).

cnaeIcmpTypeCode :: Lens' CreateNetworkAclEntry (Maybe IcmpTypeCode)

ICMP protocol: The ICMP type and code. Required if specifying ICMP for the protocol.

cnaeNetworkAclId :: Lens' CreateNetworkAclEntry Text

The ID of the network ACL.

cnaePortRange :: Lens' CreateNetworkAclEntry (Maybe PortRange)

TCP or UDP protocols: The range of ports the rule applies to.

cnaeProtocol :: Lens' CreateNetworkAclEntry Text

The protocol. A value of -1 means all protocols.

cnaeRuleAction :: Lens' CreateNetworkAclEntry RuleAction

Indicates whether to allow or deny the traffic that matches the rule.

cnaeRuleNumber :: Lens' CreateNetworkAclEntry Int

The rule number for the entry (for example, 100). ACL entries are processed in ascending order by rule number.

Constraints: Positive integer from 1 to 32766

Response

data CreateNetworkAclEntryResponse

Response constructor

createNetworkAclEntryResponse :: CreateNetworkAclEntryResponse

CreateNetworkAclEntryResponse constructor.