Network.AWS.EC2.CreateNetworkInterface

Description

Creates a network interface in the specified subnet.

For more information about network interfaces, see Elastic Network Interfaces in the Amazon Elastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateNetworkInterface.html

Request

data CreateNetworkInterface

Request constructor

createNetworkInterface

Arguments

:: Text
cniSubnetId

-> CreateNetworkInterface

CreateNetworkInterface constructor.

The fields accessible through corresponding lenses are:

cniDescription :: Maybe Text

cniDryRun :: Maybe Bool

cniGroups :: [Text]

cniPrivateIpAddress :: Maybe Text

cniPrivateIpAddresses :: [PrivateIpAddressSpecification]

cniSecondaryPrivateIpAddressCount :: Maybe Int

cniSubnetId :: Text

Request lenses

cniDescription :: Lens' CreateNetworkInterface (Maybe Text)

A description for the network interface.

cniDryRun :: Lens' CreateNetworkInterface (Maybe Bool)

cniGroups :: Lens' CreateNetworkInterface [Text]

The IDs of one or more security groups.

cniPrivateIpAddress :: Lens' CreateNetworkInterface (Maybe Text)

The primary private IP address of the network interface. If you don't specify an IP address, Amazon EC2 selects one for you from the subnet range. If you specify an IP address, you cannot indicate any IP addresses specified in privateIpAddresses as primary (only one IP address can be designated as primary).

cniPrivateIpAddresses :: Lens' CreateNetworkInterface [PrivateIpAddressSpecification]

One or more private IP addresses.

cniSecondaryPrivateIpAddressCount :: Lens' CreateNetworkInterface (Maybe Int)

The number of secondary private IP addresses to assign to a network interface. When you specify a number of secondary IP addresses, Amazon EC2 selects these IP addresses within the subnet range. You can't specify this option and specify more than one private IP address using privateIpAddresses.

The number of IP addresses you can assign to a network interface varies by instance type. For more information, see Private IP Addresses Per ENI PerInstance Type in the Amazon Elastic Compute Cloud User Guide for Linux.

cniSubnetId :: Lens' CreateNetworkInterface Text

The ID of the subnet to associate with the network interface.

Response

data CreateNetworkInterfaceResponse

Response constructor

createNetworkInterfaceResponse :: CreateNetworkInterfaceResponse

CreateNetworkInterfaceResponse constructor.

The fields accessible through corresponding lenses are:

cnirNetworkInterface :: Maybe NetworkInterface

Response lenses

cnirNetworkInterface :: Lens' CreateNetworkInterfaceResponse (Maybe NetworkInterface)

Information about the network interface.