Network.AWS.EC2.CreateRouteTable

Description

Creates a route table for the specified VPC. After you create a route table, you can add routes and associate the table with a subnet.

For more information about route tables, see Route Tables in the AmazonVirtual Private Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateRouteTable.html

Request

data CreateRouteTable

Request constructor

createRouteTable

Arguments

:: Text
crtVpcId

-> CreateRouteTable

CreateRouteTable constructor.

The fields accessible through corresponding lenses are:

crtDryRun :: Maybe Bool

crtVpcId :: Text

Request lenses

crtDryRun :: Lens' CreateRouteTable (Maybe Bool)

crtVpcId :: Lens' CreateRouteTable Text

The ID of the VPC.

Response

data CreateRouteTableResponse

Response constructor

createRouteTableResponse :: CreateRouteTableResponse

CreateRouteTableResponse constructor.

The fields accessible through corresponding lenses are:

crtrRouteTable :: Maybe RouteTable

Response lenses

crtrRouteTable :: Lens' CreateRouteTableResponse (Maybe RouteTable)

Information about the route table.