Network.AWS.EC2.CreateSecurityGroup

Description

Creates a security group. A security group is for use with instances either in the EC2-Classic platform or in a specific VPC. For more information, see Amazon EC2 Security Groups in the Amazon Elastic Compute Cloud User Guide and Security Groups for Your VPC in the Amazon Virtual Private Cloud User Guide. EC2-Classic: You can have up to 500 security groups. EC2-VPC: You can create up to 100 security groups per VPC. When you create a security group, you specify a friendly name of your choice. You can have a security group for use in EC2-Classic with the same name as a security group for use in a VPC. However, you can't have two security groups for use in EC2-Classic with the same name or two security groups for use in a VPC with the same name. You have a default security group for use in EC2-Classic and a default security group for use in your VPC. If you don't specify a security group when you launch an instance, the instance is launched into the appropriate default security group. A default security group includes a default rule that grants instances unrestricted network access to each other. You can add or remove rules from your security groups using AuthorizeSecurityGroupIngress, AuthorizeSecurityGroupEgress, RevokeSecurityGroupIngress, and RevokeSecurityGroupEgress.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateSecurityGroup.html

Request

data CreateSecurityGroup

Request constructor

createSecurityGroup

Arguments

:: Text
csgGroupName

-> Text
csgDescription

-> CreateSecurityGroup

CreateSecurityGroup constructor.

The fields accessible through corresponding lenses are:

csgDescription :: Text

csgDryRun :: Maybe Bool

csgGroupName :: Text

csgVpcId :: Maybe Text

Request lenses

csgDescription :: Lens' CreateSecurityGroup Text

A description for the security group. This is informational only. Constraints: Up to 255 characters in length Constraints for EC2-Classic: ASCII characters Constraints for EC2-VPC: a-z, A-Z, 0-9, spaces, and ._-:/()#,@[]+=&amp;;{}!$*.

csgDryRun :: Lens' CreateSecurityGroup (Maybe Bool)

csgGroupName :: Lens' CreateSecurityGroup Text

The name of the security group. Constraints: Up to 255 characters in length Constraints for EC2-Classic: ASCII characters Constraints for EC2-VPC: a-z, A-Z, 0-9, spaces, and ._-:/()#,@[]+=&amp;;{}!$*.

csgVpcId :: Lens' CreateSecurityGroup (Maybe Text)

EC2-VPC

The ID of the VPC. Required for EC2-VPC.

Response

data CreateSecurityGroupResponse

Response constructor

createSecurityGroupResponse

Arguments

:: Text
csgrGroupId

-> CreateSecurityGroupResponse

CreateSecurityGroupResponse constructor.

The fields accessible through corresponding lenses are:

csgrGroupId :: Text

Response lenses

csgrGroupId :: Lens' CreateSecurityGroupResponse Text

The ID of the security group.