Network.AWS.EC2.CreateSnapshot

Description

Creates a snapshot of an Amazon EBS volume and stores it in Amazon S3. You can use snapshots for backups, to make copies of Amazon EBS volumes, and to save data before shutting down an instance. When a snapshot is created, any AWS Marketplace product codes that are associated with the source volume are propagated to the snapshot. You can take a snapshot of an attached volume that is in use. However, snapshots only capture data that has been written to your Amazon EBS volume at the time the snapshot command is issued; this may exclude any data that has been cached by any applications or the operating system. If you can pause any file systems on the volume long enough to take a snapshot, your snapshot should be complete. However, if you cannot pause all file writes to the volume, you should unmount the volume from within the instance, issue the snapshot command, and then remount the volume to ensure a consistent and complete snapshot. You may remount and use your volume while the snapshot status is pending. To create a snapshot for Amazon EBS volumes that serve as root devices, you should stop the instance before taking the snapshot. Snapshots that are taken from encrypted volumes are automatically encrypted. Volumes that are created from encrypted snapshots are also automatically encrypted. Your encrypted volumes and any associated snapshots always remain protected. For more information, see Amazon Elastic Block Store and Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide.

Request

data CreateSnapshot

Request constructor

createSnapshot

Arguments

:: Text
cs2VolumeId

-> CreateSnapshot

CreateSnapshot constructor.

The fields accessible through corresponding lenses are:

cs2Description :: Maybe Text

cs2DryRun :: Maybe Bool

cs2VolumeId :: Text

Request lenses

cs2Description :: Lens' CreateSnapshot (Maybe Text)

A description for the snapshot.

cs2DryRun :: Lens' CreateSnapshot (Maybe Bool)

cs2VolumeId :: Lens' CreateSnapshot Text

The ID of the Amazon EBS volume.

Response

data Snapshot

Response constructor

snapshot :: Snapshot

Snapshot constructor.

The fields accessible through corresponding lenses are:

sDescription :: Maybe Text

sEncrypted :: Maybe Bool

sOwnerAlias :: Maybe Text

sOwnerId :: Maybe Text

sProgress :: Maybe Text

sSnapshotId :: Maybe Text

sStartTime :: Maybe UTCTime

sState :: Maybe Text

sTags :: [Tag]

sVolumeId :: Maybe Text

sVolumeSize :: Maybe Int

Response lenses

sDescription :: Lens' Snapshot (Maybe Text)

The description for the snapshot.

sEncrypted :: Lens' Snapshot (Maybe Bool)

Indicates whether the snapshot is encrypted.

sOwnerAlias :: Lens' Snapshot (Maybe Text)

The AWS account alias (for example, amazon, self) or AWS account ID that owns the snapshot.

sOwnerId :: Lens' Snapshot (Maybe Text)

The AWS account ID of the Amazon EBS snapshot owner.

sProgress :: Lens' Snapshot (Maybe Text)

The progress of the snapshot, as a percentage.

sSnapshotId :: Lens' Snapshot (Maybe Text)

The ID of the snapshot.

sStartTime :: Lens' Snapshot (Maybe UTCTime)

The time stamp when the snapshot was initiated.

sState :: Lens' Snapshot (Maybe Text)

The snapshot state.

sTags :: Lens' Snapshot [Tag]

Any tags assigned to the snapshot.

sVolumeId :: Lens' Snapshot (Maybe Text)

The ID of the volume.

sVolumeSize :: Lens' Snapshot (Maybe Int)

The size of the volume, in GiB.