Network.AWS.EC2.CreateSpotDatafeedSubscription

Description

Creates a datafeed for Spot Instances, enabling you to view Spot Instance usage logs. You can create one data feed per AWS account. For more information, see <http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/using-spot-instances.html Spot Instances> in the Amazon Elastic Compute Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateSpotDatafeedSubscription.html

Request

data CreateSpotDatafeedSubscription

Request constructor

createSpotDatafeedSubscription

Arguments

:: Text
csdsBucket

-> CreateSpotDatafeedSubscription

CreateSpotDatafeedSubscription constructor.

The fields accessible through corresponding lenses are:

csdsBucket :: Text

csdsDryRun :: Maybe Bool

csdsPrefix :: Maybe Text

Request lenses

csdsBucket :: Lens' CreateSpotDatafeedSubscription Text

The Amazon S3 bucket in which to store the Spot Instance datafeed. Constraints: Must be a valid bucket associated with your AWS account.

csdsDryRun :: Lens' CreateSpotDatafeedSubscription (Maybe Bool)

csdsPrefix :: Lens' CreateSpotDatafeedSubscription (Maybe Text)

A prefix for the datafeed file names.

Response

data CreateSpotDatafeedSubscriptionResponse

Response constructor

createSpotDatafeedSubscriptionResponse :: CreateSpotDatafeedSubscriptionResponse

CreateSpotDatafeedSubscriptionResponse constructor.

The fields accessible through corresponding lenses are:

csdsrSpotDatafeedSubscription :: Maybe SpotDatafeedSubscription

Response lenses

csdsrSpotDatafeedSubscription :: Lens' CreateSpotDatafeedSubscriptionResponse (Maybe SpotDatafeedSubscription)

The Spot Instance datafeed subscription.