Network.AWS.EC2.CreateSubnet

Description

Creates a subnet in an existing VPC.

When you create each subnet, you provide the VPC ID and the CIDR block you want for the subnet. After you create a subnet, you can't change its CIDR block. The subnet's CIDR block can be the same as the VPC's CIDR block (assuming you want only a single subnet in the VPC), or a subset of the VPC's CIDR block. If you create more than one subnet in a VPC, the subnets' CIDR blocks must not overlap. The smallest subnet (and VPC) you can create uses a 28 netmask (16 IP addresses), and the largest uses a 16 netmask (65,536 IP addresses).

AWS reserves both the first four and the last IP address in each subnet's CIDR block. They're not available for use.

If you add more than one subnet to a VPC, they're set up in a star topology with a logical router in the middle.

If you launch an instance in a VPC using an Amazon EBS-backed AMI, the IP address doesn't change if you stop and restart the instance (unlike a similar instance launched outside a VPC, which gets a new IP address when restarted). It's therefore possible to have a subnet with no running instances (they're all stopped), but no remaining IP addresses available.

For more information about subnets, see Your VPC and Subnets in the AmazonVirtual Private Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateSubnet.html

Request

data CreateSubnet

Request constructor

createSubnet

Arguments

:: Text
cs1VpcId

-> Text
cs1CidrBlock

-> CreateSubnet

CreateSubnet constructor.

The fields accessible through corresponding lenses are:

cs1AvailabilityZone :: Maybe Text

cs1CidrBlock :: Text

cs1DryRun :: Maybe Bool

cs1VpcId :: Text

Request lenses

cs1AvailabilityZone :: Lens' CreateSubnet (Maybe Text)

The Availability Zone for the subnet.

Default: Amazon EC2 selects one for you (recommended).

cs1CidrBlock :: Lens' CreateSubnet Text

The network range for the subnet, in CIDR notation. For example, '10.0.0.0/24'.

cs1DryRun :: Lens' CreateSubnet (Maybe Bool)

cs1VpcId :: Lens' CreateSubnet Text

The ID of the VPC.

Response

data CreateSubnetResponse

Response constructor

createSubnetResponse :: CreateSubnetResponse

CreateSubnetResponse constructor.

The fields accessible through corresponding lenses are:

csrSubnet :: Maybe Subnet

Response lenses

csrSubnet :: Lens' CreateSubnetResponse (Maybe Subnet)

Information about the subnet.