Network.AWS.EC2.CreateTags

Description

Adds or overwrites one or more tags for the specified Amazon EC2 resource or resources. Each resource can have a maximum of 10 tags. Each tag consists of a key and optional value. Tag keys must be unique per resource.

For more information about tags, see Tagging Your Resources in the AmazonElastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateTags.html

Request

data CreateTags

Request constructor

createTags :: CreateTags

CreateTags constructor.

The fields accessible through corresponding lenses are:

ct1DryRun :: Maybe Bool

ct1Resources :: [Text]

ct1Tags :: [Tag]

Request lenses

ct1DryRun :: Lens' CreateTags (Maybe Bool)

ct1Resources :: Lens' CreateTags [Text]

The IDs of one or more resources to tag. For example, ami-1a2b3c4d.

ct1Tags :: Lens' CreateTags [Tag]

One or more tags. The value parameter is required, but if you don't want the tag to have a value, specify the parameter with no value, and we set the value to an empty string.

Response

data CreateTagsResponse

Response constructor

createTagsResponse :: CreateTagsResponse

CreateTagsResponse constructor.