Network.AWS.EC2.CreateVPCPeeringConnection

Description

Requests a VPC peering connection between two VPCs: a requester VPC that you own and a peer VPC with which to create the connection. The peer VPC can belong to another AWS account. The requester VPC and peer VPC cannot have overlapping CIDR blocks.

The owner of the peer VPC must accept the peering request to activate the peering connection. The VPC peering connection request expires after 7 days, after which it cannot be accepted or rejected.

If you try to create a VPC peering connection between VPCs that have overlapping CIDR blocks, the VPC peering connection status goes to failed .

Creating a Request

createVPCPeeringConnection :: CreateVPCPeeringConnection

Creates a value of CreateVPCPeeringConnection with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cvpcPeerVPCId - The ID of the VPC with which you are creating the VPC peering connection.

cvpcVPCId - The ID of the requester VPC.

cvpcPeerOwnerId - The AWS account ID of the owner of the peer VPC. Default: Your AWS account ID

cvpcDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data CreateVPCPeeringConnection

Contains the parameters for CreateVpcPeeringConnection.

See: createVPCPeeringConnection smart constructor.

Request Lenses

cvpcPeerVPCId :: Lens' CreateVPCPeeringConnection (Maybe Text)

The ID of the VPC with which you are creating the VPC peering connection.

cvpcVPCId :: Lens' CreateVPCPeeringConnection (Maybe Text)

The ID of the requester VPC.

cvpcPeerOwnerId :: Lens' CreateVPCPeeringConnection (Maybe Text)

The AWS account ID of the owner of the peer VPC. Default: Your AWS account ID

cvpcDryRun :: Lens' CreateVPCPeeringConnection (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

createVPCPeeringConnectionResponse

Arguments

:: Int
cvpcrsResponseStatus

-> CreateVPCPeeringConnectionResponse

Creates a value of CreateVPCPeeringConnectionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cvpcrsVPCPeeringConnection - Information about the VPC peering connection.

cvpcrsResponseStatus - -- | The response status code.

data CreateVPCPeeringConnectionResponse

Contains the output of CreateVpcPeeringConnection.

See: createVPCPeeringConnectionResponse smart constructor.

Response Lenses

cvpcrsVPCPeeringConnection :: Lens' CreateVPCPeeringConnectionResponse (Maybe VPCPeeringConnection)

Information about the VPC peering connection.

cvpcrsResponseStatus :: Lens' CreateVPCPeeringConnectionResponse Int

- | The response status code.