Network.AWS.EC2.CreateVPNConnection

Description

Creates a VPN connection between an existing virtual private gateway and a VPN customer gateway. The only supported connection type is ipsec.1 .

The response includes information that you need to give to your network administrator to configure your customer gateway.

Important: We strongly recommend that you use HTTPS when calling this operation because the response contains sensitive cryptographic information for configuring your customer gateway.

If you decide to shut down your VPN connection for any reason and later create a new VPN connection, you must reconfigure your customer gateway with the new information returned from this call.

This is an idempotent operation. If you perform the operation more than once, Amazon EC2 doesn't return an error.

For more information about VPN connections, see Adding a Hardware Virtual Private Gateway to Your VPC in the Amazon Virtual Private Cloud User Guide .

Creating a Request

createVPNConnection

Arguments

:: Text
cvcCustomerGatewayId

-> Text
cvcType

-> Text
cvcVPNGatewayId

-> CreateVPNConnection

Creates a value of CreateVPNConnection with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cvcOptions - Indicates whether the VPN connection requires static routes. If you are creating a VPN connection for a device that does not support BGP, you must specify true . Default: false

cvcDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cvcCustomerGatewayId - The ID of the customer gateway.

cvcType - The type of VPN connection (ipsec.1 ).

cvcVPNGatewayId - The ID of the virtual private gateway.

data CreateVPNConnection

Contains the parameters for CreateVpnConnection.

See: createVPNConnection smart constructor.

Request Lenses

cvcOptions :: Lens' CreateVPNConnection (Maybe VPNConnectionOptionsSpecification)

Indicates whether the VPN connection requires static routes. If you are creating a VPN connection for a device that does not support BGP, you must specify true . Default: false

cvcDryRun :: Lens' CreateVPNConnection (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cvcCustomerGatewayId :: Lens' CreateVPNConnection Text

The ID of the customer gateway.

cvcType :: Lens' CreateVPNConnection Text

The type of VPN connection (ipsec.1 ).

cvcVPNGatewayId :: Lens' CreateVPNConnection Text

The ID of the virtual private gateway.

Destructuring the Response

createVPNConnectionResponse

Arguments

:: Int
cvcrsResponseStatus

-> CreateVPNConnectionResponse

Creates a value of CreateVPNConnectionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cvcrsVPNConnection - Information about the VPN connection.

cvcrsResponseStatus - -- | The response status code.

data CreateVPNConnectionResponse

Contains the output of CreateVpnConnection.

See: createVPNConnectionResponse smart constructor.

Response Lenses

cvcrsVPNConnection :: Lens' CreateVPNConnectionResponse (Maybe VPNConnection)

Information about the VPN connection.

cvcrsResponseStatus :: Lens' CreateVPNConnectionResponse Int

- | The response status code.