Network.AWS.EC2.CreateVPNConnectionRoute

Description

Creates a static route associated with a VPN connection between an existing virtual private gateway and a VPN customer gateway. The static route allows traffic to be routed from the virtual private gateway to the VPN customer gateway.

For more information about VPN connections, see Adding a Hardware Virtual Private Gateway to Your VPC in the Amazon Virtual Private Cloud User Guide .

Creating a Request

createVPNConnectionRoute

Arguments

:: Text
cvcrDestinationCidrBlock

-> Text
cvcrVPNConnectionId

-> CreateVPNConnectionRoute

Creates a value of CreateVPNConnectionRoute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cvcrDestinationCidrBlock - The CIDR block associated with the local subnet of the customer network.

cvcrVPNConnectionId - The ID of the VPN connection.

data CreateVPNConnectionRoute

Contains the parameters for CreateVpnConnectionRoute.

See: createVPNConnectionRoute smart constructor.

Request Lenses

cvcrDestinationCidrBlock :: Lens' CreateVPNConnectionRoute Text

The CIDR block associated with the local subnet of the customer network.

cvcrVPNConnectionId :: Lens' CreateVPNConnectionRoute Text

The ID of the VPN connection.

Destructuring the Response

createVPNConnectionRouteResponse :: CreateVPNConnectionRouteResponse

Creates a value of CreateVPNConnectionRouteResponse with the minimum fields required to make a request.

data CreateVPNConnectionRouteResponse

See: createVPNConnectionRouteResponse smart constructor.