Network.AWS.EC2.CreateVPNGateway

Description

Creates a virtual private gateway. A virtual private gateway is the endpoint on the VPC side of your VPN connection. You can create a virtual private gateway before creating the VPC itself.

For more information about virtual private gateways, see Adding a Hardware Virtual Private Gateway to Your VPC in the Amazon Virtual Private Cloud User Guide .

Creating a Request

createVPNGateway

Arguments

:: GatewayType
cvgType

-> CreateVPNGateway

Creates a value of CreateVPNGateway with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cvgAvailabilityZone - The Availability Zone for the virtual private gateway.

cvgDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cvgType - The type of VPN connection this virtual private gateway supports.

data CreateVPNGateway

Contains the parameters for CreateVpnGateway.

See: createVPNGateway smart constructor.

Request Lenses

cvgAvailabilityZone :: Lens' CreateVPNGateway (Maybe Text)

The Availability Zone for the virtual private gateway.

cvgDryRun :: Lens' CreateVPNGateway (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cvgType :: Lens' CreateVPNGateway GatewayType

The type of VPN connection this virtual private gateway supports.

Destructuring the Response

createVPNGatewayResponse

Arguments

:: Int
cvgrsResponseStatus

-> CreateVPNGatewayResponse

Creates a value of CreateVPNGatewayResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cvgrsVPNGateway - Information about the virtual private gateway.

cvgrsResponseStatus - -- | The response status code.

data CreateVPNGatewayResponse

Contains the output of CreateVpnGateway.

See: createVPNGatewayResponse smart constructor.

Response Lenses

cvgrsVPNGateway :: Lens' CreateVPNGatewayResponse (Maybe VPNGateway)

Information about the virtual private gateway.

cvgrsResponseStatus :: Lens' CreateVPNGatewayResponse Int

- | The response status code.