Network.AWS.EC2.CreateVpnConnectionRoute

Description

Creates a static route associated with a VPN connection between an existing virtual private gateway and a VPN customer gateway. The static route allows traffic to be routed from the virtual private gateway to the VPN customer gateway. For more information about VPN connections, see Adding a Hardware Virtual Private Gateway to Your VPC in the Amazon Virtual Private Cloud User Guide. Example This example creates a static route to the VPN connection for the VPN connection with the ID vpn-83ad48ea to the destination CIDR block 11.12.0.0/16. Note that when using the Query API the "/" is denoted as "%2F". https:/ec2.amazonaws.com?Action=CreateVpnConnectionRoute &amp;DestinationCidrBlock=11.12.0.0%2F16 &amp;VpnConnectionId=vpn-83ad48ea &amp;AUTHPARAMS &lt;CreateVpnConnectionRouteResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;4f35a1b2-c2c3-4093-b51f-abb9d7311990&lt;/requestId&gt; &lt;return&gt;true&lt;/return&gt; &lt;/CreateVpnConnectionRouteResponse&gt;.

Request

data CreateVpnConnectionRoute

Request constructor

createVpnConnectionRoute

Arguments

:: Text
cvcr1VpnConnectionId

-> Text
cvcr1DestinationCidrBlock

-> CreateVpnConnectionRoute

Smart constructor for the minimum required parameters to construct a valid CreateVpnConnectionRoute request.

The fields accessible through corresponding lenses are:

VpnConnectionId :: Text

DestinationCidrBlock :: Text

Request lenses

cvcr1VpnConnectionId :: Lens' CreateVpnConnectionRoute Text

The ID of the VPN connection.

cvcr1DestinationCidrBlock :: Lens' CreateVpnConnectionRoute Text

The CIDR block associated with the local subnet of the customer network.

Response

data CreateVpnConnectionRouteResponse

Response constructor

createVpnConnectionRouteResponse :: CreateVpnConnectionRouteResponse

Smart constructor for the minimum required parameters to construct a valid CreateVpnConnectionRouteResponse response.

This constructor is provided for convenience and testing purposes.