Network.AWS.EC2.CreateVpnGateway

Description

Creates a virtual private gateway. A virtual private gateway is the endpoint on the VPC side of your VPN connection. You can create a virtual private gateway before creating the VPC itself. For more information about virtual private gateways, see Adding a Hardware Virtual Private Gateway to Your VPC in the Amazon Virtual Private Cloud User Guide. Example This example creates a virtual private gateway. https:/ec2.amazonaws.com?Action=CreateVpnGateway &amp;Type=ipsec.1 &amp;AUTHPARAMS &lt;CreateVpnGatewayResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;vpnGateway&gt; &lt;vpnGatewayId&gt;vgw-8db04f81&lt;/vpnGatewayId&gt; &lt;state&gt;pending&lt;state&gt; &lt;type&gt;ipsec.1&lt;type&gt; &lt;availabilityZone&gt;us-east-1a&lt;/availabilityZone&gt; &lt;attachments&gt; &lt;tagSet&gt; &lt;/vpnGateway&gt; &lt;/CreateVpnGatewayResponse&gt;.

Request

data CreateVpnGateway

Request constructor

createVpnGateway

Arguments

:: GatewayType
cvgType

-> CreateVpnGateway

Smart constructor for the minimum required parameters to construct a valid CreateVpnGateway request.

The fields accessible through corresponding lenses are:

Type :: GatewayType

AvailabilityZone :: Maybe Text

Request lenses

cvgType :: Lens' CreateVpnGateway GatewayType

The type of VPN connection this virtual private gateway supports.

cvgAvailabilityZone :: Lens' CreateVpnGateway (Maybe Text)

The Availability Zone for the virtual private gateway.

Response

data CreateVpnGatewayResponse

Response constructor

createVpnGatewayResponse :: CreateVpnGatewayResponse

Smart constructor for the minimum required parameters to construct a valid CreateVpnGatewayResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

VpnGateway :: Maybe VpnGateway

Response lenses

cvgrVpnGateway :: Lens' CreateVpnGatewayResponse (Maybe VpnGateway)

Information about the virtual private gateway.