Network.AWS.EC2.DeleteCustomerGateway

Description

Deletes the specified customer gateway. You must delete the VPN connection before you can delete the customer gateway. Example This example deletes the specified customer gateway. https:/ec2.amazonaws.com?Action=DeleteCustomerGateway &amp;CustomerGatewayId=cgw-b4dc3961 &amp;AUTHPARAMS &lt;DeleteCustomerGatewayResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;DeleteCustomerGatewayResponse&gt;.

Request

data DeleteCustomerGateway

Request constructor

mkDeleteCustomerGateway

Arguments

:: Text
dcgCustomerGatewayId

-> DeleteCustomerGateway

Smart constructor for the minimum required parameters to construct a valid DeleteCustomerGateway request.

The fields accessible through corresponding lenses are:

CustomerGatewayId :: Text

Request lenses

dcgCustomerGatewayId :: Lens' DeleteCustomerGateway Text

The ID of the customer gateway.

Response

data DeleteCustomerGatewayResponse

Response constructor

mkDeleteCustomerGatewayResponse :: DeleteCustomerGatewayResponse

Smart constructor for the minimum required parameters to construct a valid DeleteCustomerGatewayResponse response.

This constructor is provided for convenience and testing purposes.