Network.AWS.EC2.DeleteDHCPOptions

Description

Deletes the specified set of DHCP options. You must disassociate the set of DHCP options before you can delete it. You can disassociate the set of DHCP options by associating either a new set of options or the default set of options with the VPC.

Creating a Request

deleteDHCPOptions

Arguments

:: Text
ddhcpoDHCPOptionsId

-> DeleteDHCPOptions

Creates a value of DeleteDHCPOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddhcpoDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

ddhcpoDHCPOptionsId - The ID of the DHCP options set.

data DeleteDHCPOptions

Contains the parameters for DeleteDhcpOptions.

See: deleteDHCPOptions smart constructor.

Request Lenses

ddhcpoDryRun :: Lens' DeleteDHCPOptions (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

ddhcpoDHCPOptionsId :: Lens' DeleteDHCPOptions Text

The ID of the DHCP options set.

Destructuring the Response

deleteDHCPOptionsResponse :: DeleteDHCPOptionsResponse

Creates a value of DeleteDHCPOptionsResponse with the minimum fields required to make a request.

data DeleteDHCPOptionsResponse

See: deleteDHCPOptionsResponse smart constructor.