Network.AWS.EC2.DeleteDhcpOptions

Description

Deletes the specified set of DHCP options. You must disassociate the set of DHCP options before you can delete it. You can disassociate the set of DHCP options by associating either a new set of options or the default set of options with the VPC. Example This example deletes the specified set of DHCP options. https:/ec2.amazonaws.com?Action=DeleteDhcpOptions &amp;DhcpOptionsId=dopt-7a8b9c2d &amp;AUTHPARAMS &lt;DeleteDhcpOptionsResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;DeleteDhcpOptionsResponse&gt;.

Request

data DeleteDhcpOptions

Request constructor

deleteDhcpOptions

Arguments

:: Text
ddoDhcpOptionsId

-> DeleteDhcpOptions

Smart constructor for the minimum required parameters to construct a valid DeleteDhcpOptions request.

The fields accessible through corresponding lenses are:

DhcpOptionsId :: Text

Request lenses

ddoDhcpOptionsId :: Lens' DeleteDhcpOptions Text

The ID of the DHCP options set.

Response

data DeleteDhcpOptionsResponse

Response constructor

deleteDhcpOptionsResponse :: DeleteDhcpOptionsResponse

Smart constructor for the minimum required parameters to construct a valid DeleteDhcpOptionsResponse response.

This constructor is provided for convenience and testing purposes.