Network.AWS.EC2.DeleteEgressOnlyInternetGateway

Description

Deletes an egress-only Internet gateway.

Creating a Request

deleteEgressOnlyInternetGateway

Arguments

:: Text
deoigeEgressOnlyInternetGatewayId

-> DeleteEgressOnlyInternetGateway

Creates a value of DeleteEgressOnlyInternetGateway with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

deoigeDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

deoigeEgressOnlyInternetGatewayId - The ID of the egress-only Internet gateway.

data DeleteEgressOnlyInternetGateway

See: deleteEgressOnlyInternetGateway smart constructor.

Request Lenses

deoigeDryRun :: Lens' DeleteEgressOnlyInternetGateway (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

deoigeEgressOnlyInternetGatewayId :: Lens' DeleteEgressOnlyInternetGateway Text

The ID of the egress-only Internet gateway.

Destructuring the Response

deleteEgressOnlyInternetGatewayResponse

Arguments

:: Int
delrsResponseStatus

-> DeleteEgressOnlyInternetGatewayResponse

Creates a value of DeleteEgressOnlyInternetGatewayResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

delrsReturnCode - Returns true if the request succeeds; otherwise, it returns an error.

delrsResponseStatus - -- | The response status code.

data DeleteEgressOnlyInternetGatewayResponse

See: deleteEgressOnlyInternetGatewayResponse smart constructor.

Response Lenses

delrsReturnCode :: Lens' DeleteEgressOnlyInternetGatewayResponse (Maybe Bool)

Returns true if the request succeeds; otherwise, it returns an error.

delrsResponseStatus :: Lens' DeleteEgressOnlyInternetGatewayResponse Int

- | The response status code.