Network.AWS.EC2.DeleteInternetGateway

Description

Deletes the specified Internet gateway. You must detach the Internet gateway from the VPC before you can delete it.

Creating a Request

deleteInternetGateway

Arguments

:: Text
digiInternetGatewayId

-> DeleteInternetGateway

Creates a value of DeleteInternetGateway with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

digiDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

digiInternetGatewayId - The ID of the Internet gateway.

data DeleteInternetGateway

Contains the parameters for DeleteInternetGateway.

See: deleteInternetGateway smart constructor.

Request Lenses

digiDryRun :: Lens' DeleteInternetGateway (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

digiInternetGatewayId :: Lens' DeleteInternetGateway Text

The ID of the Internet gateway.

Destructuring the Response

deleteInternetGatewayResponse :: DeleteInternetGatewayResponse

Creates a value of DeleteInternetGatewayResponse with the minimum fields required to make a request.

data DeleteInternetGatewayResponse

See: deleteInternetGatewayResponse smart constructor.