Network.AWS.EC2.DeleteNetworkACLEntry

Description

Deletes the specified ingress or egress entry (rule) from the specified network ACL.

Creating a Request

deleteNetworkACLEntry

Arguments

:: Bool
dnaeEgress

-> Text
dnaeNetworkACLId

-> Int
dnaeRuleNumber

-> DeleteNetworkACLEntry

Creates a value of DeleteNetworkACLEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dnaeDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dnaeEgress - Indicates whether the rule is an egress rule.

dnaeNetworkACLId - The ID of the network ACL.

dnaeRuleNumber - The rule number of the entry to delete.

data DeleteNetworkACLEntry

Contains the parameters for DeleteNetworkAclEntry.

See: deleteNetworkACLEntry smart constructor.

Request Lenses

dnaeDryRun :: Lens' DeleteNetworkACLEntry (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dnaeEgress :: Lens' DeleteNetworkACLEntry Bool

Indicates whether the rule is an egress rule.

dnaeNetworkACLId :: Lens' DeleteNetworkACLEntry Text

The ID of the network ACL.

dnaeRuleNumber :: Lens' DeleteNetworkACLEntry Int

The rule number of the entry to delete.

Destructuring the Response

deleteNetworkACLEntryResponse :: DeleteNetworkACLEntryResponse

Creates a value of DeleteNetworkACLEntryResponse with the minimum fields required to make a request.

data DeleteNetworkACLEntryResponse

See: deleteNetworkACLEntryResponse smart constructor.