Network.AWS.EC2.DeleteNetworkInterface

Description

Deletes the specified network interface. You must detach the network interface before you can delete it.

Creating a Request

deleteNetworkInterface

Arguments

:: Text
dninNetworkInterfaceId

-> DeleteNetworkInterface

Creates a value of DeleteNetworkInterface with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dninDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dninNetworkInterfaceId - The ID of the network interface.

data DeleteNetworkInterface

Contains the parameters for DeleteNetworkInterface.

See: deleteNetworkInterface smart constructor.

Request Lenses

dninDryRun :: Lens' DeleteNetworkInterface (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dninNetworkInterfaceId :: Lens' DeleteNetworkInterface Text

The ID of the network interface.

Destructuring the Response

deleteNetworkInterfaceResponse :: DeleteNetworkInterfaceResponse

Creates a value of DeleteNetworkInterfaceResponse with the minimum fields required to make a request.

data DeleteNetworkInterfaceResponse

See: deleteNetworkInterfaceResponse smart constructor.