Network.AWS.EC2.DeleteNetworkInterfacePermission

Description

Deletes a permission for a network interface. By default, you cannot delete the permission if the account for which you're removing the permission has attached the network interface to an instance. However, you can force delete the permission, regardless of any attachment.

Creating a Request

deleteNetworkInterfacePermission

Arguments

:: Text
dnipNetworkInterfacePermissionId

-> DeleteNetworkInterfacePermission

Creates a value of DeleteNetworkInterfacePermission with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dnipForce - Specify true to remove the permission even if the network interface is attached to an instance.

dnipDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dnipNetworkInterfacePermissionId - The ID of the network interface permission.

data DeleteNetworkInterfacePermission

Contains the parameters for DeleteNetworkInterfacePermission.

See: deleteNetworkInterfacePermission smart constructor.

Request Lenses

dnipForce :: Lens' DeleteNetworkInterfacePermission (Maybe Bool)

Specify true to remove the permission even if the network interface is attached to an instance.

dnipDryRun :: Lens' DeleteNetworkInterfacePermission (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dnipNetworkInterfacePermissionId :: Lens' DeleteNetworkInterfacePermission Text

The ID of the network interface permission.

Destructuring the Response

deleteNetworkInterfacePermissionResponse

Arguments

:: Int
dniprsResponseStatus

-> DeleteNetworkInterfacePermissionResponse

Creates a value of DeleteNetworkInterfacePermissionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dniprsReturn - Returns true if the request succeeds, otherwise returns an error.

dniprsResponseStatus - -- | The response status code.

data DeleteNetworkInterfacePermissionResponse

Contains the output for DeleteNetworkInterfacePermission.

See: deleteNetworkInterfacePermissionResponse smart constructor.

Response Lenses

dniprsReturn :: Lens' DeleteNetworkInterfacePermissionResponse (Maybe Bool)

Returns true if the request succeeds, otherwise returns an error.

dniprsResponseStatus :: Lens' DeleteNetworkInterfacePermissionResponse Int

- | The response status code.