Network.AWS.EC2.DeleteRouteTable

Description

Deletes the specified route table. You must disassociate the route table from any subnets before you can delete it. You can't delete the main route table.

Creating a Request

deleteRouteTable

Arguments

:: Text
drtrRouteTableId

-> DeleteRouteTable

Creates a value of DeleteRouteTable with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drtrDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

drtrRouteTableId - The ID of the route table.

data DeleteRouteTable

Contains the parameters for DeleteRouteTable.

See: deleteRouteTable smart constructor.

Request Lenses

drtrDryRun :: Lens' DeleteRouteTable (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

drtrRouteTableId :: Lens' DeleteRouteTable Text

The ID of the route table.

Destructuring the Response

deleteRouteTableResponse :: DeleteRouteTableResponse

Creates a value of DeleteRouteTableResponse with the minimum fields required to make a request.

data DeleteRouteTableResponse

See: deleteRouteTableResponse smart constructor.