Network.AWS.EC2.DeleteSecurityGroup

Description

Deletes a security group.

If you attempt to delete a security group that is associated with an instance, or is referenced by another security group, the operation fails with InUse in EC2-Classic or DependencyViolation in EC2-VPC.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DeleteSecurityGroup.html

Request

data DeleteSecurityGroup

Request constructor

deleteSecurityGroup :: DeleteSecurityGroup

DeleteSecurityGroup constructor.

The fields accessible through corresponding lenses are:

dsgDryRun :: Maybe Bool

dsgGroupId :: Maybe Text

dsgGroupName :: Maybe Text

Request lenses

dsgDryRun :: Lens' DeleteSecurityGroup (Maybe Bool)

dsgGroupId :: Lens' DeleteSecurityGroup (Maybe Text)

The ID of the security group. Required for a nondefault VPC.

dsgGroupName :: Lens' DeleteSecurityGroup (Maybe Text)

EC2-Classic, default VPC

The name of the security group. You can specify either the security group name or the security group ID.

Response

data DeleteSecurityGroupResponse

Response constructor

deleteSecurityGroupResponse :: DeleteSecurityGroupResponse

DeleteSecurityGroupResponse constructor.