Network.AWS.EC2.DeleteSnapshot

Description

Deletes the specified snapshot.

When you make periodic snapshots of a volume, the snapshots are incremental, and only the blocks on the device that have changed since your last snapshot are saved in the new snapshot. When you delete a snapshot, only the data not needed for any other snapshot is removed. So regardless of which prior snapshots have been deleted, all active snapshots will have access to all the information needed to restore the volume.

You cannot delete a snapshot of the root device of an Amazon EBS volume used by a registered AMI. You must first de-register the AMI before you can delete the snapshot.

For more information, see Deleting an Amazon EBS Snapshot in the AmazonElastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DeleteSnapshot.html

Request

data DeleteSnapshot

Request constructor

deleteSnapshot

Arguments

:: Text
ds3SnapshotId

-> DeleteSnapshot

DeleteSnapshot constructor.

The fields accessible through corresponding lenses are:

ds3DryRun :: Maybe Bool

ds3SnapshotId :: Text

Request lenses

ds3DryRun :: Lens' DeleteSnapshot (Maybe Bool)

ds3SnapshotId :: Lens' DeleteSnapshot Text

The ID of the Amazon EBS snapshot.

Response

data DeleteSnapshotResponse

Response constructor

deleteSnapshotResponse :: DeleteSnapshotResponse

DeleteSnapshotResponse constructor.