Network.AWS.EC2.DeleteSubnet

Description

Deletes the specified subnet. You must terminate all running instances in the subnet before you can delete the subnet.

Creating a Request

deleteSubnet

Arguments

:: Text
ddSubnetId

-> DeleteSubnet

Creates a value of DeleteSubnet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

ddSubnetId - The ID of the subnet.

data DeleteSubnet

Contains the parameters for DeleteSubnet.

See: deleteSubnet smart constructor.

Request Lenses

ddDryRun :: Lens' DeleteSubnet (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

ddSubnetId :: Lens' DeleteSubnet Text

The ID of the subnet.

Destructuring the Response

deleteSubnetResponse :: DeleteSubnetResponse

Creates a value of DeleteSubnetResponse with the minimum fields required to make a request.

data DeleteSubnetResponse

See: deleteSubnetResponse smart constructor.