Network.AWS.EC2.DeleteTags

Description

Deletes the specified set of tags from the specified set of resources. This call is designed to follow a DescribeTags request.

For more information about tags, see Tagging Your Resources in the AmazonElastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DeleteTags.html

Request

data DeleteTags

Request constructor

deleteTags :: DeleteTags

DeleteTags constructor.

The fields accessible through corresponding lenses are:

dt1DryRun :: Maybe Bool

dt1Resources :: [Text]

dt1Tags :: [Tag]

Request lenses

dt1DryRun :: Lens' DeleteTags (Maybe Bool)

dt1Resources :: Lens' DeleteTags [Text]

The ID of the resource. For example, ami-1a2b3c4d. You can specify more than one resource ID.

dt1Tags :: Lens' DeleteTags [Tag]

One or more tags to delete. If you omit the value parameter, we delete the tag regardless of its value. If you specify this parameter with an empty string as the value, we delete the key only if its value is an empty string.

Response

data DeleteTagsResponse

Response constructor

deleteTagsResponse :: DeleteTagsResponse

DeleteTagsResponse constructor.