Network.AWS.EC2.DeleteVPC

Description

Deletes the specified VPC. You must detach or delete all gateways and resources that are associated with the VPC before you can delete it. For example, you must terminate all instances running in the VPC, delete all security groups associated with the VPC (except the default one), delete all route tables associated with the VPC (except the default one), and so on.

Creating a Request

deleteVPC

Arguments

:: Text
delVPCId

-> DeleteVPC

Creates a value of DeleteVPC with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

delDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

delVPCId - The ID of the VPC.

data DeleteVPC

Contains the parameters for DeleteVpc.

See: deleteVPC smart constructor.

Request Lenses

delDryRun :: Lens' DeleteVPC (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

delVPCId :: Lens' DeleteVPC Text

The ID of the VPC.

Destructuring the Response

deleteVPCResponse :: DeleteVPCResponse

Creates a value of DeleteVPCResponse with the minimum fields required to make a request.

data DeleteVPCResponse

See: deleteVPCResponse smart constructor.