Network.AWS.EC2.DeleteVPCEndpoints

Description

Deletes one or more specified VPC endpoints. Deleting the endpoint also deletes the endpoint routes in the route tables that were associated with the endpoint.

Creating a Request

deleteVPCEndpoints :: DeleteVPCEndpoints

Creates a value of DeleteVPCEndpoints with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dveDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dveVPCEndpointIds - One or more endpoint IDs.

data DeleteVPCEndpoints

Contains the parameters for DeleteVpcEndpoints.

See: deleteVPCEndpoints smart constructor.

Request Lenses

dveDryRun :: Lens' DeleteVPCEndpoints (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dveVPCEndpointIds :: Lens' DeleteVPCEndpoints [Text]

One or more endpoint IDs.

Destructuring the Response

deleteVPCEndpointsResponse

Arguments

:: Int
dversResponseStatus

-> DeleteVPCEndpointsResponse

Creates a value of DeleteVPCEndpointsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dversUnsuccessful - Information about the endpoints that were not successfully deleted.

dversResponseStatus - -- | The response status code.

data DeleteVPCEndpointsResponse

Contains the output of DeleteVpcEndpoints.

See: deleteVPCEndpointsResponse smart constructor.

Response Lenses

dversUnsuccessful :: Lens' DeleteVPCEndpointsResponse [UnsuccessfulItem]

Information about the endpoints that were not successfully deleted.

dversResponseStatus :: Lens' DeleteVPCEndpointsResponse Int

- | The response status code.