Network.AWS.EC2.DeleteVPNConnectionRoute

Description

Deletes the specified static route associated with a VPN connection between an existing virtual private gateway and a VPN customer gateway. The static route allows traffic to be routed from the virtual private gateway to the VPN customer gateway.

Creating a Request

deleteVPNConnectionRoute

Arguments

:: Text
dvcrDestinationCidrBlock

-> Text
dvcrVPNConnectionId

-> DeleteVPNConnectionRoute

Creates a value of DeleteVPNConnectionRoute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvcrDestinationCidrBlock - The CIDR block associated with the local subnet of the customer network.

dvcrVPNConnectionId - The ID of the VPN connection.

data DeleteVPNConnectionRoute

Contains the parameters for DeleteVpnConnectionRoute.

See: deleteVPNConnectionRoute smart constructor.

Request Lenses

dvcrDestinationCidrBlock :: Lens' DeleteVPNConnectionRoute Text

The CIDR block associated with the local subnet of the customer network.

dvcrVPNConnectionId :: Lens' DeleteVPNConnectionRoute Text

The ID of the VPN connection.

Destructuring the Response

deleteVPNConnectionRouteResponse :: DeleteVPNConnectionRouteResponse

Creates a value of DeleteVPNConnectionRouteResponse with the minimum fields required to make a request.

data DeleteVPNConnectionRouteResponse

See: deleteVPNConnectionRouteResponse smart constructor.