Network.AWS.EC2.DeleteVPNGateway

Description

Deletes the specified virtual private gateway. We recommend that before you delete a virtual private gateway, you detach it from the VPC and delete the VPN connection. Note that you don't need to delete the virtual private gateway if you plan to delete and recreate the VPN connection between your VPC and your network.

Creating a Request

deleteVPNGateway

Arguments

:: Text
dvgVPNGatewayId

-> DeleteVPNGateway

Creates a value of DeleteVPNGateway with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvgDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dvgVPNGatewayId - The ID of the virtual private gateway.

data DeleteVPNGateway

Contains the parameters for DeleteVpnGateway.

See: deleteVPNGateway smart constructor.

Request Lenses

dvgDryRun :: Lens' DeleteVPNGateway (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dvgVPNGatewayId :: Lens' DeleteVPNGateway Text

The ID of the virtual private gateway.

Destructuring the Response

deleteVPNGatewayResponse :: DeleteVPNGatewayResponse

Creates a value of DeleteVPNGatewayResponse with the minimum fields required to make a request.

data DeleteVPNGatewayResponse

See: deleteVPNGatewayResponse smart constructor.