Network.AWS.EC2.DeleteVolume

Description

Deletes the specified Amazon EBS volume. The volume must be in the available state (not attached to an instance).

The volume may remain in the deleting state for several minutes.

For more information, see Deleting an Amazon EBS Volume in the AmazonElastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DeleteVolume.html

Request

data DeleteVolume

Request constructor

deleteVolume

Arguments

:: Text
dv4VolumeId

-> DeleteVolume

DeleteVolume constructor.

The fields accessible through corresponding lenses are:

dv4DryRun :: Maybe Bool

dv4VolumeId :: Text

Request lenses

dv4DryRun :: Lens' DeleteVolume (Maybe Bool)

dv4VolumeId :: Lens' DeleteVolume Text

The ID of the volume.

Response

data DeleteVolumeResponse

Response constructor

deleteVolumeResponse :: DeleteVolumeResponse

DeleteVolumeResponse constructor.