Network.AWS.EC2.DeleteVpnConnection

Description

Deletes the specified VPN connection.

If you're deleting the VPC and its associated components, we recommend that you detach the virtual private gateway from the VPC and delete the VPC before deleting the VPN connection. If you believe that the tunnel credentials for your VPN connection have been compromised, you can delete the VPN connection and create a new one that has new keys, without needing to delete the VPC or virtual private gateway. If you create a new VPN connection, you must reconfigure the customer gateway using the new configuration information returned with the new VPN connection ID.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DeleteVpnConnection.html

Request

data DeleteVpnConnection

Request constructor

deleteVpnConnection

Arguments

:: Text
dvcVpnConnectionId

-> DeleteVpnConnection

DeleteVpnConnection constructor.

The fields accessible through corresponding lenses are:

dvcDryRun :: Maybe Bool

dvcVpnConnectionId :: Text

Request lenses

dvcDryRun :: Lens' DeleteVpnConnection (Maybe Bool)

dvcVpnConnectionId :: Lens' DeleteVpnConnection Text

The ID of the VPN connection.

Response

data DeleteVpnConnectionResponse

Response constructor

deleteVpnConnectionResponse :: DeleteVpnConnectionResponse

DeleteVpnConnectionResponse constructor.