Network.AWS.EC2.DescribeAccountAttributes

Description

Describes the specified attribute of your AWS account. Example This example describes the platforms that are supported by your AWS account. The first response is for an account that supports only EC2-VPC. The second response if for an account that supports both EC2-Classic and EC2-VPC. https:/ec2.amazonaws.com?Action=DescribeAccountAttributes &amp;AttributeName.1=supported-platforms &amp;AUTHPARAMS &lt;DescribeAccountAttributesResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;accountAttributeSet&gt; &lt;item&gt; &lt;attributeName&gt;supported-platforms&lt;/attributeName&gt; &lt;attributeValueSet&gt; &lt;item&gt; &lt;attributeValue&gt;VPC&lt;attributeValue&gt; &lt;item&gt; &lt;attributeValueSet&gt; &lt;item&gt; &lt;/accountAttributeSet&gt; &lt;/DescribeAccountAttributesResponse&gt; &lt;DescribeAccountAttributesResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;accountAttributeSet&gt; &lt;item&gt; &lt;attributeName&gt;supported-platforms&lt;/attributeName&gt; &lt;attributeValueSet&gt; &lt;item&gt; &lt;attributeValue&gt;EC2&lt;attributeValue&gt; &lt;item&gt; &lt;item&gt; &lt;attributeValue&gt;VPC&lt;attributeValue&gt; &lt;item&gt; &lt;attributeValueSet&gt; &lt;item&gt; &lt;/accountAttributeSet&gt; &lt;/DescribeAccountAttributesResponse&gt; Example 2 This example describes the ID of your default VPC. The first response is for an account that supports only EC2-VPC. The second response if for an account that supports both EC2-Classic and EC2-VPC. https:/ec2.amazonaws.com?Action=DescribeAccountAttributes &amp;AttributeName.1=default-vpc &amp;AUTHPARAMS &lt;DescribeAccountAttributesResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;accountAttributeSet&gt; &lt;item&gt; &lt;attributeName&gt;default-vpc&lt;/attributeName&gt; &lt;attributeValueSet&gt; &lt;item&gt; &lt;attributeValue&gt;vpc-xxxxxxxx&lt;attributeValue&gt; &lt;item&gt; &lt;attributeValueSet&gt; &lt;item&gt; &lt;/accountAttributeSet&gt; &lt;/DescribeAccountAttributesResponse&gt; &lt;DescribeAccountAttributesResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;accountAttributeSet&gt; &lt;item&gt; &lt;attributeName&gt;default-vpc&lt;/attributeName&gt; &lt;attributeValueSet&gt; &lt;item&gt; &lt;attributeValue&gt;none&lt;attributeValue&gt; &lt;item&gt; &lt;attributeValueSet&gt; &lt;item&gt; &lt;/accountAttributeSet&gt; &lt;/DescribeAccountAttributesResponse&gt;.

Request

data DescribeAccountAttributes

Request constructor

describeAccountAttributes :: DescribeAccountAttributes

Smart constructor for the minimum required parameters to construct a valid DescribeAccountAttributes request.

The fields accessible through corresponding lenses are:

AttributeNames :: [AccountAttributeName]

Request lenses

daaAttributeNames :: Lens' DescribeAccountAttributes [AccountAttributeName]

One or more account attribute names.

Response

data DescribeAccountAttributesResponse

Response constructor

describeAccountAttributesResponse :: DescribeAccountAttributesResponse

Smart constructor for the minimum required parameters to construct a valid DescribeAccountAttributesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

AccountAttributes :: [AccountAttribute]

Response lenses

daarAccountAttributes :: Lens' DescribeAccountAttributesResponse [AccountAttribute]

Information about one or more account attributes.