Network.AWS.EC2.DescribeAddresses

Description

Describes one or more of your Elastic IP addresses.

An Elastic IP address is for use in either the EC2-Classic platform or in a VPC. For more information, see Elastic IP Addresses in the Amazon ElasticCompute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeAddresses.html

Request

data DescribeAddresses

Request constructor

describeAddresses :: DescribeAddresses

DescribeAddresses constructor.

The fields accessible through corresponding lenses are:

daAllocationIds :: [Text]

daDryRun :: Maybe Bool

daFilters :: [Filter]

daPublicIps :: [Text]

Request lenses

daAllocationIds :: Lens' DescribeAddresses [Text]

EC2-VPC

One or more allocation IDs.

Default: Describes all your Elastic IP addresses.

daDryRun :: Lens' DescribeAddresses (Maybe Bool)

daFilters :: Lens' DescribeAddresses [Filter]

One or more filters. Filter names and values are case-sensitive.

'allocation-id' - [EC2-VPC] The allocation ID for the address.

'association-id' - [EC2-VPC] The association ID for the address.

domain - Indicates whether the address is for use in EC2-Classic (standard) or in a VPC ( vpc).

'instance-id' - The ID of the instance the address is associated with, if any.

'network-interface-id' - [EC2-VPC] The ID of the network interface that the address is associated with, if any.

'network-interface-owner-id' - The AWS account ID of the owner.

'private-ip-address' - [EC2-VPC] The private IP address associated with the Elastic IP address.

'public-ip' - The Elastic IP address.

daPublicIps :: Lens' DescribeAddresses [Text]

EC2-Classic

One or more Elastic IP addresses.

Default: Describes all your Elastic IP addresses.

Response

data DescribeAddressesResponse

Response constructor

describeAddressesResponse :: DescribeAddressesResponse

DescribeAddressesResponse constructor.

The fields accessible through corresponding lenses are:

darAddresses :: [Address]

Response lenses

darAddresses :: Lens' DescribeAddressesResponse [Address]

Information about one or more Elastic IP addresses.