Network.AWS.EC2.DescribeAvailabilityZones

Description

Describes one or more of the Availability Zones that are available to you. The results include zones only for the region you're currently using. If there is an event impacting an Availability Zone, you can use this request to view the state and any provided message for that Availability Zone. For more information, see Regions and Availability Zones in the Amazon Elastic Compute Cloud User Guide. Example This example request describes the Availability Zones that are available to you. The response includes Availability Zones only for the current region. https:/ec2.amazonaws.com?Action=DescribeAvailabilityZones &amp;AUTHPARAMS &lt;DescribeAvailabilityZonesResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;availabilityZoneInfo&gt; &lt;item&gt; &lt;zoneName&gt;us-east-1a&lt;/zoneName&gt; &lt;zoneState&gt;available&lt;/zoneState&gt; &lt;regionName&gt;us-east-1&lt;regionName&gt; &lt;messageSet&gt; &lt;item&gt; &lt;item&gt; &lt;zoneName&gt;us-east-1b&lt;zoneName&gt; &lt;zoneState&gt;available&lt;/zoneState&gt; &lt;regionName&gt;us-east-1&lt;regionName&gt; &lt;messageSet&gt; &lt;item&gt; &lt;item&gt; &lt;zoneName&gt;us-east-1c&lt;zoneName&gt; &lt;zoneState&gt;available&lt;/zoneState&gt; &lt;regionName&gt;us-east-1&lt;regionName&gt; &lt;messageSet&gt; &lt;item&gt; &lt;item&gt; &lt;zoneName&gt;us-east-1d&lt;zoneName&gt; &lt;zoneState&gt;available&lt;/zoneState&gt; &lt;regionName&gt;us-east-1&lt;regionName&gt; &lt;messageSet&gt; &lt;item&gt; &lt;availabilityZoneInfo&gt; &lt;/DescribeAvailabilityZonesResponse&gt;.

Request

data DescribeAvailabilityZones

Request constructor

describeAvailabilityZones :: DescribeAvailabilityZones

Smart constructor for the minimum required parameters to construct a valid DescribeAvailabilityZones request.

The fields accessible through corresponding lenses are:

ZoneNames :: [Text]

Filters :: [Filter]

Request lenses

dazZoneNames :: Lens' DescribeAvailabilityZones [Text]

The names of one or more Availability Zones.

dazFilters :: Lens' DescribeAvailabilityZones [Filter]

One or more filters. message - Information about the Availability Zone. region-name - The name of the region for the Availability Zone (for example, us-east-1). state - The state of the Availability Zone (available | impaired | unavailable). zone-name - The name of the Availability Zone (for example, us-east-1a).

Response

data DescribeAvailabilityZonesResponse

Response constructor

describeAvailabilityZonesResponse :: DescribeAvailabilityZonesResponse

Smart constructor for the minimum required parameters to construct a valid DescribeAvailabilityZonesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

AvailabilityZones :: [AvailabilityZone]

Response lenses

dazrAvailabilityZones :: Lens' DescribeAvailabilityZonesResponse [AvailabilityZone]

Information about one or more Availability Zones.