Network.AWS.EC2.DescribeBundleTasks

Description

Describes one or more of your bundling tasks.

Completed bundle tasks are listed for only a limited time. If your bundle task is no longer in the list, you can still register an AMI from it. Just use RegisterImage with the Amazon S3 bucket name and image manifest name you provided to the bundle task.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeBundleTasks.html

Request

data DescribeBundleTasks

Request constructor

describeBundleTasks :: DescribeBundleTasks

DescribeBundleTasks constructor.

The fields accessible through corresponding lenses are:

dbtBundleIds :: [Text]

dbtDryRun :: Maybe Bool

dbtFilters :: [Filter]

Request lenses

dbtBundleIds :: Lens' DescribeBundleTasks [Text]

One or more bundle task IDs.

Default: Describes all your bundle tasks.

dbtDryRun :: Lens' DescribeBundleTasks (Maybe Bool)

dbtFilters :: Lens' DescribeBundleTasks [Filter]

One or more filters.

'bundle-id' - The ID of the bundle task.

'error-code' - If the task failed, the error code returned.

'error-message' - If the task failed, the error message returned.

'instance-id' - The ID of the instance.

progress - The level of task completion, as a percentage (for example, 20%).

's3-bucket' - The Amazon S3 bucket to store the AMI.

's3-prefix' - The beginning of the AMI name.

'start-time' - The time the task started (for example, 2013-09-15T17:15:20.000Z).

state - The state of the task (pending | 'waiting-for-shutdown' | bundling | storing | cancelling | complete | failed).

'update-time' - The time of the most recent update for the task.

Response

data DescribeBundleTasksResponse

Response constructor

describeBundleTasksResponse :: DescribeBundleTasksResponse

DescribeBundleTasksResponse constructor.

The fields accessible through corresponding lenses are:

dbtrBundleTasks :: [BundleTask]

Response lenses

dbtrBundleTasks :: Lens' DescribeBundleTasksResponse [BundleTask]

Information about one or more bundle tasks.