Network.AWS.EC2.DescribeClassicLinkInstances

Description

Describes one or more of your linked EC2-Classic instances. This request only returns information about EC2-Classic instances linked to a VPC through ClassicLink; you cannot use this request to return information about other instances.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeClassicLinkInstances.html

Request

data DescribeClassicLinkInstances

Request constructor

describeClassicLinkInstances :: DescribeClassicLinkInstances

DescribeClassicLinkInstances constructor.

The fields accessible through corresponding lenses are:

dcliDryRun :: Maybe Bool

dcliFilters :: [Filter]

dcliInstanceIds :: [Text]

dcliMaxResults :: Maybe Int

dcliNextToken :: Maybe Text

Request lenses

dcliDryRun :: Lens' DescribeClassicLinkInstances (Maybe Bool)

dcliFilters :: Lens' DescribeClassicLinkInstances [Filter]

One or more filters.

'group-id' - The ID of a VPC security group that's associated with the instance.

'instance-id' - The ID of the instance.

tag:key=value - The key/value combination of a tag assigned to the resource.

'tag-key' - The key of a tag assigned to the resource. This filter is independent of the 'tag-value' filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag:key=value filter.

'tag-value' - The value of a tag assigned to the resource. This filter is independent of the 'tag-key' filter.

'vpc-id' - The ID of the VPC that the instance is linked to.

dcliInstanceIds :: Lens' DescribeClassicLinkInstances [Text]

One or more instance IDs. Must be instances linked to a VPC through ClassicLink.

dcliMaxResults :: Lens' DescribeClassicLinkInstances (Maybe Int)

The maximum number of results to return for the request in a single page. The remaining results of the initial request can be seen by sending another request with the returned NextToken value. This value can be between 5 and 1000; if MaxResults is given a value larger than 1000, only 1000 results are returned. You cannot specify this parameter and the instance IDs parameter in the same request.

Constraint: If the value is greater than 1000, we return only 1000 items.

dcliNextToken :: Lens' DescribeClassicLinkInstances (Maybe Text)

The token to retrieve the next page of results.

Response

data DescribeClassicLinkInstancesResponse

Response constructor

describeClassicLinkInstancesResponse :: DescribeClassicLinkInstancesResponse

DescribeClassicLinkInstancesResponse constructor.

The fields accessible through corresponding lenses are:

dclirInstances :: [ClassicLinkInstance]

dclirNextToken :: Maybe Text

Response lenses

dclirInstances :: Lens' DescribeClassicLinkInstancesResponse [ClassicLinkInstance]

Information about one or more linked EC2-Classic instances.

dclirNextToken :: Lens' DescribeClassicLinkInstancesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.