Network.AWS.EC2.DescribeConversionTasks

Description

Describes one or more of your conversion tasks. For more information, see Using the Command Line Tools to Import Your Virtual Machine to Amazon EC2 in the Amazon Elastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeConversionTasks.html

Request

data DescribeConversionTasks

Request constructor

describeConversionTasks :: DescribeConversionTasks

DescribeConversionTasks constructor.

The fields accessible through corresponding lenses are:

dctConversionTaskIds :: [Text]

dctDryRun :: Maybe Bool

dctFilters :: [Filter]

Request lenses

dctConversionTaskIds :: Lens' DescribeConversionTasks [Text]

One or more conversion task IDs.

dctDryRun :: Lens' DescribeConversionTasks (Maybe Bool)

dctFilters :: Lens' DescribeConversionTasks [Filter]

Response

data DescribeConversionTasksResponse

Response constructor

describeConversionTasksResponse :: DescribeConversionTasksResponse

DescribeConversionTasksResponse constructor.

The fields accessible through corresponding lenses are:

dctrConversionTasks :: [ConversionTask]

Response lenses

dctrConversionTasks :: Lens' DescribeConversionTasksResponse [ConversionTask]