Network.AWS.EC2.DescribeCustomerGateways

Description

Describes one or more of your VPN customer gateways.

For more information about VPN customer gateways, see Adding a HardwareVirtual Private Gateway to Your VPC in the Amazon Virtual Private Cloud UserGuide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeCustomerGateways.html

Request

data DescribeCustomerGateways

Request constructor

describeCustomerGateways :: DescribeCustomerGateways

DescribeCustomerGateways constructor.

The fields accessible through corresponding lenses are:

dcgCustomerGatewayIds :: [Text]

dcgDryRun :: Maybe Bool

dcgFilters :: [Filter]

Request lenses

dcgCustomerGatewayIds :: Lens' DescribeCustomerGateways [Text]

One or more customer gateway IDs.

Default: Describes all your customer gateways.

dcgDryRun :: Lens' DescribeCustomerGateways (Maybe Bool)

dcgFilters :: Lens' DescribeCustomerGateways [Filter]

One or more filters.

'bgp-asn' - The customer gateway's Border Gateway Protocol (BGP) Autonomous System Number (ASN).

'customer-gateway-id' - The ID of the customer gateway.

'ip-address' - The IP address of the customer gateway's Internet-routable external interface.

state - The state of the customer gateway (pending | available | deleting | deleted).

'type' - The type of customer gateway. Currently, the only supported type is 'ipsec.1'.

tag:key=value - The key/value combination of a tag assigned to the resource.

'tag-key' - The key of a tag assigned to the resource. This filter is independent of the 'tag-value' filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag:key=value filter.

'tag-value' - The value of a tag assigned to the resource. This filter is independent of the 'tag-key' filter.

Response

data DescribeCustomerGatewaysResponse

Response constructor

describeCustomerGatewaysResponse :: DescribeCustomerGatewaysResponse

DescribeCustomerGatewaysResponse constructor.

The fields accessible through corresponding lenses are:

dcgrCustomerGateways :: [CustomerGateway]

Response lenses

dcgrCustomerGateways :: Lens' DescribeCustomerGatewaysResponse [CustomerGateway]

Information about one or more customer gateways.