Network.AWS.EC2.DescribeDHCPOptions

Description

Describes one or more of your DHCP options sets.

For more information about DHCP options sets, see DHCP Options Sets in the Amazon Virtual Private Cloud User Guide .

Creating a Request

describeDHCPOptions :: DescribeDHCPOptions

Creates a value of DescribeDHCPOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddoFilters - One or more filters. * dhcp-options-id - The ID of a set of DHCP options. * key - The key for one of the options (for example, domain-name ). * value - The value for one of the options. * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter.

ddoDHCPOptionsIds - The IDs of one or more DHCP options sets. Default: Describes all your DHCP options sets.

ddoDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data DescribeDHCPOptions

Contains the parameters for DescribeDhcpOptions.

See: describeDHCPOptions smart constructor.

Request Lenses

ddoFilters :: Lens' DescribeDHCPOptions [Filter]

One or more filters. * dhcp-options-id - The ID of a set of DHCP options. * key - The key for one of the options (for example, domain-name ). * value - The value for one of the options. * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter.

ddoDHCPOptionsIds :: Lens' DescribeDHCPOptions [Text]

The IDs of one or more DHCP options sets. Default: Describes all your DHCP options sets.

ddoDryRun :: Lens' DescribeDHCPOptions (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

describeDHCPOptionsResponse

Arguments

:: Int
ddorsResponseStatus

-> DescribeDHCPOptionsResponse

Creates a value of DescribeDHCPOptionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ddorsDHCPOptions - Information about one or more DHCP options sets.

ddorsResponseStatus - -- | The response status code.

data DescribeDHCPOptionsResponse

Contains the output of DescribeDhcpOptions.

See: describeDHCPOptionsResponse smart constructor.

Response Lenses

ddorsDHCPOptions :: Lens' DescribeDHCPOptionsResponse [DHCPOptions]

Information about one or more DHCP options sets.

ddorsResponseStatus :: Lens' DescribeDHCPOptionsResponse Int

- | The response status code.