Network.AWS.EC2.DescribeDhcpOptions

Description

Describes one or more of your DHCP options sets.

For more information about DHCP options sets, see DHCP Options Sets in the Amazon Virtual Private Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeDhcpOptions.html

Request

data DescribeDhcpOptions

Request constructor

describeDhcpOptions :: DescribeDhcpOptions

DescribeDhcpOptions constructor.

The fields accessible through corresponding lenses are:

ddoDhcpOptionsIds :: [Text]

ddoDryRun :: Maybe Bool

ddoFilters :: [Filter]

Request lenses

ddoDhcpOptionsIds :: Lens' DescribeDhcpOptions [Text]

The IDs of one or more DHCP options sets.

Default: Describes all your DHCP options sets.

ddoDryRun :: Lens' DescribeDhcpOptions (Maybe Bool)

ddoFilters :: Lens' DescribeDhcpOptions [Filter]

One or more filters.

'dhcp-options-id' - The ID of a set of DHCP options.

key - The key for one of the options (for example, 'domain-name').

value - The value for one of the options.

tag:key=value - The key/value combination of a tag assigned to the resource.

'tag-key' - The key of a tag assigned to the resource. This filter is independent of the 'tag-value' filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag:key=value filter.

'tag-value' - The value of a tag assigned to the resource. This filter is independent of the 'tag-key' filter.

Response

data DescribeDhcpOptionsResponse

Response constructor

describeDhcpOptionsResponse :: DescribeDhcpOptionsResponse

DescribeDhcpOptionsResponse constructor.

The fields accessible through corresponding lenses are:

ddorDhcpOptions :: [DhcpOptions]

Response lenses

ddorDhcpOptions :: Lens' DescribeDhcpOptionsResponse [DhcpOptions]

Information about one or more DHCP options sets.