Network.AWS.EC2.DescribeEgressOnlyInternetGateways

Description

Describes one or more of your egress-only Internet gateways.

Creating a Request

describeEgressOnlyInternetGateways :: DescribeEgressOnlyInternetGateways

Creates a value of DescribeEgressOnlyInternetGateways with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

deoigEgressOnlyInternetGatewayIds - One or more egress-only Internet gateway IDs.

deoigNextToken - The token to retrieve the next page of results.

deoigDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

deoigMaxResults - The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned NextToken value. This value can be between 5 and 1000; if MaxResults is given a value larger than 1000, only 1000 results are returned.

data DescribeEgressOnlyInternetGateways

See: describeEgressOnlyInternetGateways smart constructor.

Request Lenses

deoigEgressOnlyInternetGatewayIds :: Lens' DescribeEgressOnlyInternetGateways [Text]

One or more egress-only Internet gateway IDs.

deoigNextToken :: Lens' DescribeEgressOnlyInternetGateways (Maybe Text)

The token to retrieve the next page of results.

deoigDryRun :: Lens' DescribeEgressOnlyInternetGateways (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

deoigMaxResults :: Lens' DescribeEgressOnlyInternetGateways (Maybe Int)

The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned NextToken value. This value can be between 5 and 1000; if MaxResults is given a value larger than 1000, only 1000 results are returned.

Destructuring the Response

describeEgressOnlyInternetGatewaysResponse

Arguments

:: Int
deoigrsResponseStatus

-> DescribeEgressOnlyInternetGatewaysResponse

Creates a value of DescribeEgressOnlyInternetGatewaysResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

deoigrsEgressOnlyInternetGateways - Information about the egress-only Internet gateways.

deoigrsNextToken - The token to use to retrieve the next page of results.

deoigrsResponseStatus - -- | The response status code.

data DescribeEgressOnlyInternetGatewaysResponse

See: describeEgressOnlyInternetGatewaysResponse smart constructor.

Response Lenses

deoigrsEgressOnlyInternetGateways :: Lens' DescribeEgressOnlyInternetGatewaysResponse [EgressOnlyInternetGateway]

Information about the egress-only Internet gateways.

deoigrsNextToken :: Lens' DescribeEgressOnlyInternetGatewaysResponse (Maybe Text)

The token to use to retrieve the next page of results.

deoigrsResponseStatus :: Lens' DescribeEgressOnlyInternetGatewaysResponse Int

- | The response status code.