Network.AWS.EC2.DescribeElasticGpus

Description

Describes the Elastic GPUs associated with your instances. For more information about Elastic GPUs, see Amazon EC2 Elastic GPUs .

Creating a Request

describeElasticGpus :: DescribeElasticGpus

Creates a value of DescribeElasticGpus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

degFilters - One or more filters. * availability-zone - The Availability Zone in which the Elastic GPU resides. * elastic-gpu-health - The status of the Elastic GPU (OK | IMPAIRED ). * elastic-gpu-state - The state of the Elastic GPU (ATTACHED ). * elastic-gpu-type - The type of Elastic GPU; for example, eg1.medium . * instance-id - The ID of the instance to which the Elastic GPU is associated.

degNextToken - The token to request the next page of results.

degDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

degMaxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. This value can be between 5 and 1000.

degElasticGpuIds - One or more Elastic GPU IDs.

data DescribeElasticGpus

See: describeElasticGpus smart constructor.

Request Lenses

degFilters :: Lens' DescribeElasticGpus [Filter]

One or more filters. * availability-zone - The Availability Zone in which the Elastic GPU resides. * elastic-gpu-health - The status of the Elastic GPU (OK | IMPAIRED ). * elastic-gpu-state - The state of the Elastic GPU (ATTACHED ). * elastic-gpu-type - The type of Elastic GPU; for example, eg1.medium . * instance-id - The ID of the instance to which the Elastic GPU is associated.

degNextToken :: Lens' DescribeElasticGpus (Maybe Text)

The token to request the next page of results.

degDryRun :: Lens' DescribeElasticGpus (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

degMaxResults :: Lens' DescribeElasticGpus (Maybe Int)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. This value can be between 5 and 1000.

degElasticGpuIds :: Lens' DescribeElasticGpus [Text]

One or more Elastic GPU IDs.

Destructuring the Response

describeElasticGpusResponse

Arguments

:: Int
degrsResponseStatus

-> DescribeElasticGpusResponse

Creates a value of DescribeElasticGpusResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

degrsElasticGpuSet - Information about the Elastic GPUs.

degrsNextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

degrsMaxResults - The total number of items to return. If the total number of items available is more than the value specified in max-items then a Next-Token will be provided in the output that you can use to resume pagination.

degrsResponseStatus - -- | The response status code.

data DescribeElasticGpusResponse

See: describeElasticGpusResponse smart constructor.

Response Lenses

degrsElasticGpuSet :: Lens' DescribeElasticGpusResponse [ElasticGpus]

Information about the Elastic GPUs.

degrsNextToken :: Lens' DescribeElasticGpusResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

degrsMaxResults :: Lens' DescribeElasticGpusResponse (Maybe Int)

The total number of items to return. If the total number of items available is more than the value specified in max-items then a Next-Token will be provided in the output that you can use to resume pagination.

degrsResponseStatus :: Lens' DescribeElasticGpusResponse Int

- | The response status code.