Network.AWS.EC2.DescribeExportTasks

Description

Describes one or more of your export tasks. Example This example describes a single export task. https:/ec2.amazonaws.com?Action=DescribeExportTasks &amp;exportTaskId.1=export-i-1234wxyz &amp;AUTHPARAMS &lt;DescribeExportTasksResponse xmlns="http://ec2.amazonaws.com/doc/2013-06-15/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;exportTaskSet&gt; &lt;item&gt; &lt;exportTaskId&gt;export-i-1234wxyz&lt;/exportTaskId&gt; &lt;description&gt;Example for docs&lt;/description&gt; &lt;state&gt;active&lt;/state&gt; &lt;statusMessage&gt;Running&lt;/statusMessage&gt; &lt;instanceExport&gt; &lt;instanceId&gt;i-12345678&lt;/instanceId&gt; &lt;targetEnvironment&gt;VMWare&lt;/targetEnvironment&gt; &lt;/instanceExport&gt; &lt;exportToS3&gt; &lt;diskImageFormat&gt;VMDK&lt;/diskImageFormat&gt; &lt;containerFormat&gt;OVA&lt;/containerFormat&gt; &lt;s3Bucket&gt;my-bucket-for-exported-vm&lt;/s3Bucket&gt; &lt;s3Key&gt;my-exports export-i-1234wxyz .ova&lt;s3Key&gt; &lt;exportToS3&gt; &lt;item&gt; &lt;exportTaskSet&gt; &lt; DescribeExportTasksResponse&gt;.

Request

data DescribeExportTasks

Request constructor

describeExportTasks :: DescribeExportTasks

Smart constructor for the minimum required parameters to construct a valid DescribeExportTasks request.

The fields accessible through corresponding lenses are:

ExportTaskIds :: [Text]

Request lenses

detExportTaskIds :: Lens' DescribeExportTasks [Text]

One or more export task IDs.

Response

data DescribeExportTasksResponse

Response constructor

describeExportTasksResponse :: DescribeExportTasksResponse

Smart constructor for the minimum required parameters to construct a valid DescribeExportTasksResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ExportTasks :: [ExportTask]

Response lenses

detrExportTasks :: Lens' DescribeExportTasksResponse [ExportTask]