Network.AWS.EC2.DescribeHostReservationOfferings

Description

Describes the Dedicated Host Reservations that are available to purchase.

The results describe all the Dedicated Host Reservation offerings, including offerings that may not match the instance family and region of your Dedicated Hosts. When purchasing an offering, ensure that the the instance family and region of the offering matches that of the Dedicated Hosts it will be associated with. For an overview of supported instance types, see <http:docs.aws.amazon.comAWSEC2latestUserGuidededicated-hosts-overview.html Dedicated Hosts Overview> in the Amazon Elastic Compute Cloud User Guide/ .

Creating a Request

describeHostReservationOfferings :: DescribeHostReservationOfferings

Creates a value of DescribeHostReservationOfferings with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dhroMaxDuration - This is the maximum duration of the reservation you'd like to purchase, specified in seconds. Reservations are available in one-year and three-year terms. The number of seconds specified must be the number of seconds in a year (365x24x60x60) times one of the supported durations (1 or 3). For example, specify 94608000 for three years.

dhroNextToken - The token to use to retrieve the next page of results.

dhroMinDuration - This is the minimum duration of the reservation you'd like to purchase, specified in seconds. Reservations are available in one-year and three-year terms. The number of seconds specified must be the number of seconds in a year (365x24x60x60) times one of the supported durations (1 or 3). For example, specify 31536000 for one year.

dhroOfferingId - The ID of the reservation offering.

dhroFilter - One or more filters. * instance-family - The instance family of the offering (e.g., m4 ). * payment-option - The payment option (NoUpfront | PartialUpfront | AllUpfront ).

dhroMaxResults - The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned nextToken value. This value can be between 5 and 500; if maxResults is given a larger value than 500, you will receive an error.

data DescribeHostReservationOfferings

See: describeHostReservationOfferings smart constructor.

Request Lenses

dhroMaxDuration :: Lens' DescribeHostReservationOfferings (Maybe Int)

This is the maximum duration of the reservation you'd like to purchase, specified in seconds. Reservations are available in one-year and three-year terms. The number of seconds specified must be the number of seconds in a year (365x24x60x60) times one of the supported durations (1 or 3). For example, specify 94608000 for three years.

dhroNextToken :: Lens' DescribeHostReservationOfferings (Maybe Text)

The token to use to retrieve the next page of results.

dhroMinDuration :: Lens' DescribeHostReservationOfferings (Maybe Int)

This is the minimum duration of the reservation you'd like to purchase, specified in seconds. Reservations are available in one-year and three-year terms. The number of seconds specified must be the number of seconds in a year (365x24x60x60) times one of the supported durations (1 or 3). For example, specify 31536000 for one year.

dhroOfferingId :: Lens' DescribeHostReservationOfferings (Maybe Text)

The ID of the reservation offering.

dhroFilter :: Lens' DescribeHostReservationOfferings [Filter]

One or more filters. * instance-family - The instance family of the offering (e.g., m4 ). * payment-option - The payment option (NoUpfront | PartialUpfront | AllUpfront ).

dhroMaxResults :: Lens' DescribeHostReservationOfferings (Maybe Int)

The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned nextToken value. This value can be between 5 and 500; if maxResults is given a larger value than 500, you will receive an error.

Destructuring the Response

describeHostReservationOfferingsResponse

Arguments

:: Int
dhrorsResponseStatus

-> DescribeHostReservationOfferingsResponse

Creates a value of DescribeHostReservationOfferingsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dhrorsOfferingSet - Information about the offerings.

dhrorsNextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

dhrorsResponseStatus - -- | The response status code.

data DescribeHostReservationOfferingsResponse

See: describeHostReservationOfferingsResponse smart constructor.

Response Lenses

dhrorsOfferingSet :: Lens' DescribeHostReservationOfferingsResponse [HostOffering]

Information about the offerings.

dhrorsNextToken :: Lens' DescribeHostReservationOfferingsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

dhrorsResponseStatus :: Lens' DescribeHostReservationOfferingsResponse Int

- | The response status code.