Network.AWS.EC2.DescribeHostReservations

Description

Describes Dedicated Host Reservations which are associated with Dedicated Hosts in your account.

Creating a Request

describeHostReservations :: DescribeHostReservations

Creates a value of DescribeHostReservations with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dhrNextToken - The token to use to retrieve the next page of results.

dhrHostReservationIdSet - One or more host reservation IDs.

dhrFilter - One or more filters. * instance-family - The instance family (e.g., m4 ). * payment-option - The payment option (NoUpfront | PartialUpfront | AllUpfront ). * state - The state of the reservation (payment-pending | payment-failed | active | retired ).

dhrMaxResults - The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned nextToken value. This value can be between 5 and 500; if maxResults is given a larger value than 500, you will receive an error.

data DescribeHostReservations

See: describeHostReservations smart constructor.

Request Lenses

dhrNextToken :: Lens' DescribeHostReservations (Maybe Text)

The token to use to retrieve the next page of results.

dhrHostReservationIdSet :: Lens' DescribeHostReservations [Text]

One or more host reservation IDs.

dhrFilter :: Lens' DescribeHostReservations [Filter]

One or more filters. * instance-family - The instance family (e.g., m4 ). * payment-option - The payment option (NoUpfront | PartialUpfront | AllUpfront ). * state - The state of the reservation (payment-pending | payment-failed | active | retired ).

dhrMaxResults :: Lens' DescribeHostReservations (Maybe Int)

The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned nextToken value. This value can be between 5 and 500; if maxResults is given a larger value than 500, you will receive an error.

Destructuring the Response

describeHostReservationsResponse

Arguments

:: Int
dhrrsResponseStatus

-> DescribeHostReservationsResponse

Creates a value of DescribeHostReservationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dhrrsNextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

dhrrsHostReservationSet - Details about the reservation's configuration.

dhrrsResponseStatus - -- | The response status code.

data DescribeHostReservationsResponse

See: describeHostReservationsResponse smart constructor.

Response Lenses

dhrrsNextToken :: Lens' DescribeHostReservationsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

dhrrsHostReservationSet :: Lens' DescribeHostReservationsResponse [HostReservation]

Details about the reservation's configuration.

dhrrsResponseStatus :: Lens' DescribeHostReservationsResponse Int

- | The response status code.