Network.AWS.EC2.DescribeHosts

Description

Describes one or more of your Dedicated Hosts.

The results describe only the Dedicated Hosts in the region you're currently using. All listed instances consume capacity on your Dedicated Host. Dedicated Hosts that have recently been released will be listed with the state released .

Creating a Request

describeHosts :: DescribeHosts

Creates a value of DescribeHosts with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dhNextToken - The token to retrieve the next page of results.

dhFilter - One or more filters. * instance-type - The instance type size that the Dedicated Host is configured to support. * auto-placement - Whether auto-placement is enabled or disabled (on | off ). * host-reservation-id - The ID of the reservation assigned to this host. * client-token - The idempotency token you provided when you launched the instance * state - The allocation state of the Dedicated Host (available | under-assessment | permanent-failure | released | released-permanent-failure ). * availability-zone - The Availability Zone of the host.

dhHostIds - The IDs of the Dedicated Hosts. The IDs are used for targeted instance launches.

dhMaxResults - The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned nextToken value. This value can be between 5 and 500; if maxResults is given a larger value than 500, you will receive an error. You cannot specify this parameter and the host IDs parameter in the same request.

data DescribeHosts

Contains the parameters for DescribeHosts.

See: describeHosts smart constructor.

Request Lenses

dhNextToken :: Lens' DescribeHosts (Maybe Text)

The token to retrieve the next page of results.

dhFilter :: Lens' DescribeHosts [Filter]

One or more filters. * instance-type - The instance type size that the Dedicated Host is configured to support. * auto-placement - Whether auto-placement is enabled or disabled (on | off ). * host-reservation-id - The ID of the reservation assigned to this host. * client-token - The idempotency token you provided when you launched the instance * state - The allocation state of the Dedicated Host (available | under-assessment | permanent-failure | released | released-permanent-failure ). * availability-zone - The Availability Zone of the host.

dhHostIds :: Lens' DescribeHosts [Text]

The IDs of the Dedicated Hosts. The IDs are used for targeted instance launches.

dhMaxResults :: Lens' DescribeHosts (Maybe Int)

The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned nextToken value. This value can be between 5 and 500; if maxResults is given a larger value than 500, you will receive an error. You cannot specify this parameter and the host IDs parameter in the same request.

Destructuring the Response

describeHostsResponse

Arguments

:: Int
dhrsResponseStatus

-> DescribeHostsResponse

Creates a value of DescribeHostsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dhrsHosts - Information about the Dedicated Hosts.

dhrsNextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

dhrsResponseStatus - -- | The response status code.

data DescribeHostsResponse

Contains the output of DescribeHosts.

See: describeHostsResponse smart constructor.

Response Lenses

dhrsHosts :: Lens' DescribeHostsResponse [Host]

Information about the Dedicated Hosts.

dhrsNextToken :: Lens' DescribeHostsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

dhrsResponseStatus :: Lens' DescribeHostsResponse Int

- | The response status code.