Network.AWS.EC2.DescribeIAMInstanceProfileAssociations

Description

Describes your IAM instance profile associations.

Creating a Request

describeIAMInstanceProfileAssociations :: DescribeIAMInstanceProfileAssociations

Creates a value of DescribeIAMInstanceProfileAssociations with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diapaFilters - One or more filters. * instance-id - The ID of the instance. * state - The state of the association (associating | associated | disassociating | disassociated ).

diapaNextToken - The token to request the next page of results.

diapaAssociationIds - One or more IAM instance profile associations.

diapaMaxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value.

data DescribeIAMInstanceProfileAssociations

See: describeIAMInstanceProfileAssociations smart constructor.

Request Lenses

diapaFilters :: Lens' DescribeIAMInstanceProfileAssociations [Filter]

One or more filters. * instance-id - The ID of the instance. * state - The state of the association (associating | associated | disassociating | disassociated ).

diapaNextToken :: Lens' DescribeIAMInstanceProfileAssociations (Maybe Text)

The token to request the next page of results.

diapaAssociationIds :: Lens' DescribeIAMInstanceProfileAssociations [Text]

One or more IAM instance profile associations.

diapaMaxResults :: Lens' DescribeIAMInstanceProfileAssociations (Maybe Natural)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value.

Destructuring the Response

describeIAMInstanceProfileAssociationsResponse

Arguments

:: Int
diaparsResponseStatus

-> DescribeIAMInstanceProfileAssociationsResponse

Creates a value of DescribeIAMInstanceProfileAssociationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diaparsIAMInstanceProfileAssociations - Information about one or more IAM instance profile associations.

diaparsNextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

diaparsResponseStatus - -- | The response status code.

data DescribeIAMInstanceProfileAssociationsResponse

See: describeIAMInstanceProfileAssociationsResponse smart constructor.

Response Lenses

diaparsIAMInstanceProfileAssociations :: Lens' DescribeIAMInstanceProfileAssociationsResponse [IAMInstanceProfileAssociation]

Information about one or more IAM instance profile associations.

diaparsNextToken :: Lens' DescribeIAMInstanceProfileAssociationsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

diaparsResponseStatus :: Lens' DescribeIAMInstanceProfileAssociationsResponse Int

- | The response status code.