Network.AWS.EC2.DescribeIdFormat

Description

Describes the ID format settings for your resources on a per-region basis, for example, to view which resource types are enabled for longer IDs. This request only returns information about resource types whose ID formats can be modified; it does not return information about other resource types.

The following resource types support longer IDs: instance | reservation | snapshot | volume .

These settings apply to the IAM user who makes the request; they do not apply to the entire AWS account. By default, an IAM user defaults to the same settings as the root user, unless they explicitly override the settings by running the ModifyIdFormat command. Resources created with longer IDs are visible to all IAM users, regardless of these settings and provided that they have permission to use the relevant Describe command for the resource type.

Creating a Request

describeIdFormat :: DescribeIdFormat

Creates a value of DescribeIdFormat with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

difResource - The type of resource: instance | reservation | snapshot | volume

data DescribeIdFormat

Contains the parameters for DescribeIdFormat.

See: describeIdFormat smart constructor.

Request Lenses

difResource :: Lens' DescribeIdFormat (Maybe Text)

The type of resource: instance | reservation | snapshot | volume

Destructuring the Response

describeIdFormatResponse

Arguments

:: Int
difrsResponseStatus

-> DescribeIdFormatResponse

Creates a value of DescribeIdFormatResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

difrsStatuses - Information about the ID format for the resource.

difrsResponseStatus - -- | The response status code.

data DescribeIdFormatResponse

Contains the output of DescribeIdFormat.

See: describeIdFormatResponse smart constructor.

Response Lenses

difrsStatuses :: Lens' DescribeIdFormatResponse [IdFormat]

Information about the ID format for the resource.

difrsResponseStatus :: Lens' DescribeIdFormatResponse Int

- | The response status code.