Network.AWS.EC2.DescribeIdentityIdFormat

Description

Describes the ID format settings for resources for the specified IAM user, IAM role, or root user. For example, you can view the resource types that are enabled for longer IDs. This request only returns information about resource types whose ID formats can be modified; it does not return information about other resource types. For more information, see Resource IDs in the Amazon Elastic Compute Cloud User Guide .

The following resource types support longer IDs: instance | reservation | snapshot | volume .

These settings apply to the principal specified in the request. They do not apply to the principal that makes the request.

Creating a Request

describeIdentityIdFormat

Arguments

:: Text
diifPrincipalARN

-> DescribeIdentityIdFormat

Creates a value of DescribeIdentityIdFormat with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diifResource - The type of resource: instance | reservation | snapshot | volume

diifPrincipalARN - The ARN of the principal, which can be an IAM role, IAM user, or the root user.

data DescribeIdentityIdFormat

Contains the parameters for DescribeIdentityIdFormat.

See: describeIdentityIdFormat smart constructor.

Request Lenses

diifResource :: Lens' DescribeIdentityIdFormat (Maybe Text)

The type of resource: instance | reservation | snapshot | volume

diifPrincipalARN :: Lens' DescribeIdentityIdFormat Text

The ARN of the principal, which can be an IAM role, IAM user, or the root user.

Destructuring the Response

describeIdentityIdFormatResponse

Arguments

:: Int
diifrsResponseStatus

-> DescribeIdentityIdFormatResponse

Creates a value of DescribeIdentityIdFormatResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diifrsStatuses - Information about the ID format for the resources.

diifrsResponseStatus - -- | The response status code.

data DescribeIdentityIdFormatResponse

Contains the output of DescribeIdentityIdFormat.

See: describeIdentityIdFormatResponse smart constructor.

Response Lenses

diifrsStatuses :: Lens' DescribeIdentityIdFormatResponse [IdFormat]

Information about the ID format for the resources.

diifrsResponseStatus :: Lens' DescribeIdentityIdFormatResponse Int

- | The response status code.