Network.AWS.EC2.DescribeImageAttribute

Description

Describes the specified attribute of the specified AMI. You can specify only one attribute at a time.

Creating a Request

describeImageAttribute

Arguments

:: ImageAttributeName
diaiAttribute

-> Text
diaiImageId

-> DescribeImageAttribute

Creates a value of DescribeImageAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diaiDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

diaiAttribute - The AMI attribute. Note : Depending on your account privileges, the blockDeviceMapping attribute may return a Client.AuthFailure error. If this happens, use DescribeImages to get information about the block device mapping for the AMI.

diaiImageId - The ID of the AMI.

data DescribeImageAttribute

Contains the parameters for DescribeImageAttribute.

See: describeImageAttribute smart constructor.

Request Lenses

diaiDryRun :: Lens' DescribeImageAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

diaiAttribute :: Lens' DescribeImageAttribute ImageAttributeName

The AMI attribute. Note : Depending on your account privileges, the blockDeviceMapping attribute may return a Client.AuthFailure error. If this happens, use DescribeImages to get information about the block device mapping for the AMI.

diaiImageId :: Lens' DescribeImageAttribute Text

The ID of the AMI.

Destructuring the Response

describeImageAttributeResponse

Arguments

:: Int
diarsResponseStatus

-> DescribeImageAttributeResponse

Creates a value of DescribeImageAttributeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diarsLaunchPermissions - One or more launch permissions.

diarsRAMDiskId - The RAM disk ID.

diarsKernelId - The kernel ID.

diarsSRIOVNetSupport - Indicates whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

diarsImageId - The ID of the AMI.

diarsProductCodes - One or more product codes.

diarsDescription - A description for the AMI.

diarsBlockDeviceMappings - One or more block device mapping entries.

diarsResponseStatus - -- | The response status code.

data DescribeImageAttributeResponse

Describes an image attribute.

See: describeImageAttributeResponse smart constructor.

Response Lenses

diarsLaunchPermissions :: Lens' DescribeImageAttributeResponse [LaunchPermission]

One or more launch permissions.

diarsRAMDiskId :: Lens' DescribeImageAttributeResponse (Maybe AttributeValue)

The RAM disk ID.

diarsKernelId :: Lens' DescribeImageAttributeResponse (Maybe AttributeValue)

The kernel ID.

diarsSRIOVNetSupport :: Lens' DescribeImageAttributeResponse (Maybe AttributeValue)

Indicates whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

diarsImageId :: Lens' DescribeImageAttributeResponse (Maybe Text)

The ID of the AMI.

diarsProductCodes :: Lens' DescribeImageAttributeResponse [ProductCode]

One or more product codes.

diarsDescription :: Lens' DescribeImageAttributeResponse (Maybe AttributeValue)

A description for the AMI.

diarsBlockDeviceMappings :: Lens' DescribeImageAttributeResponse [BlockDeviceMapping]

One or more block device mapping entries.

diarsResponseStatus :: Lens' DescribeImageAttributeResponse Int

- | The response status code.