Network.AWS.EC2.DescribeImages

Description

Describes one or more of the images (AMIs, AKIs, and ARIs) available to you. Images available to you include public images, private images that you own, and private images owned by other AWS accounts but for which you have explicit launch permissions.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeImages.html

Request

data DescribeImages

Request constructor

describeImages :: DescribeImages

DescribeImages constructor.

The fields accessible through corresponding lenses are:

di2DryRun :: Maybe Bool

di2ExecutableUsers :: [Text]

di2Filters :: [Filter]

di2ImageIds :: [Text]

di2Owners :: [Text]

Request lenses

di2DryRun :: Lens' DescribeImages (Maybe Bool)

di2ExecutableUsers :: Lens' DescribeImages [Text]

Scopes the images by users with explicit launch permissions. Specify an AWS account ID, self (the sender of the request), or all (public AMIs).

di2Filters :: Lens' DescribeImages [Filter]

One or more filters. architecture - The image architecture (i386 | x86_64). block-device-mapping.delete-on-termination - A Boolean value that indicates whether the Amazon EBS volume is deleted on instance termination. block-device-mapping.device-name - The device name for the Amazon EBS volume (for example, devsdh). block-device-mapping.snapshot-id - The ID of the snapshot used for the Amazon EBS volume. block-device-mapping.volume-size - The volume size of the Amazon EBS volume, in GiB. block-device-mapping.volume-type - The volume type of the Amazon EBS volume (gp2 | standard | io1). description - The description of the image (provided during image creation). hypervisor - The hypervisor type (ovm | xen). image-id - The ID of the image. image-type - The image type (machine | kernel | ramdisk). is-public - A Boolean that indicates whether the image is public. kernel-id - The kernel ID. manifest-location - The location of the image manifest. name - The name of the AMI (provided during image creation). owner-alias - The AWS account alias (for example, amazon). owner-id - The AWS account ID of the image owner. platform - The platform. To only list Windows-based AMIs, use windows. product-code - The product code. product-code.type - The type of the product code (devpay | marketplace). ramdisk-id - The RAM disk ID. root-device-name - The name of the root device volume (for example, devsda1). root-device-type - The type of the root device volume (ebs | instance-store). state - The state of the image (available | pending | failed). state-reason-code - The reason code for the state change. state-reason-message - The message for the state change. tag:key=value - The key/value combination of a tag assigned to the resource. tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag:key=value filter. tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter. virtualization-type - The virtualization type (paravirtual | hvm).

di2ImageIds :: Lens' DescribeImages [Text]

One or more image IDs. Default: Describes all images available to you.

di2Owners :: Lens' DescribeImages [Text]

Filters the images by the owner. Specify an AWS account ID, amazon (owner is Amazon), aws-marketplace (owner is AWS Marketplace), self (owner is the sender of the request), or all (all owners).

Response

data DescribeImagesResponse

Response constructor

describeImagesResponse :: DescribeImagesResponse

DescribeImagesResponse constructor.

The fields accessible through corresponding lenses are:

dirImages :: [Image]

Response lenses

dirImages :: Lens' DescribeImagesResponse [Image]

Information about one or more images.