Network.AWS.EC2.DescribeImportSnapshotTasks

Description

Describes your import snapshot tasks.

Creating a Request

describeImportSnapshotTasks :: DescribeImportSnapshotTasks

Creates a value of DescribeImportSnapshotTasks with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

distFilters - One or more filters.

distImportTaskIds - A list of import snapshot task IDs.

distNextToken - A token that indicates the next page of results.

distDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

distMaxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value.

data DescribeImportSnapshotTasks

Contains the parameters for DescribeImportSnapshotTasks.

See: describeImportSnapshotTasks smart constructor.

Request Lenses

distFilters :: Lens' DescribeImportSnapshotTasks [Filter]

One or more filters.

distImportTaskIds :: Lens' DescribeImportSnapshotTasks [Text]

A list of import snapshot task IDs.

distNextToken :: Lens' DescribeImportSnapshotTasks (Maybe Text)

A token that indicates the next page of results.

distDryRun :: Lens' DescribeImportSnapshotTasks (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

distMaxResults :: Lens' DescribeImportSnapshotTasks (Maybe Int)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value.

Destructuring the Response

describeImportSnapshotTasksResponse

Arguments

:: Int
distrsResponseStatus

-> DescribeImportSnapshotTasksResponse

Creates a value of DescribeImportSnapshotTasksResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

distrsNextToken - The token to use to get the next page of results. This value is null when there are no more results to return.

distrsImportSnapshotTasks - A list of zero or more import snapshot tasks that are currently active or were completed or canceled in the previous 7 days.

distrsResponseStatus - -- | The response status code.

data DescribeImportSnapshotTasksResponse

Contains the output for DescribeImportSnapshotTasks.

See: describeImportSnapshotTasksResponse smart constructor.

Response Lenses

distrsNextToken :: Lens' DescribeImportSnapshotTasksResponse (Maybe Text)

The token to use to get the next page of results. This value is null when there are no more results to return.

distrsImportSnapshotTasks :: Lens' DescribeImportSnapshotTasksResponse [ImportSnapshotTask]

A list of zero or more import snapshot tasks that are currently active or were completed or canceled in the previous 7 days.

distrsResponseStatus :: Lens' DescribeImportSnapshotTasksResponse Int

- | The response status code.