Network.AWS.EC2.DescribeInstanceAttribute

Description

Describes the specified attribute of the specified instance. You can specify only one attribute at a time. Valid attribute values are: instanceType | kernel | ramdisk | userData | disableApiTermination | instanceInitiatedShutdownBehavior | rootDeviceName | blockDeviceMapping | productCodes | sourceDestCheck | groupSet | ebsOptimized | sriovNetSupport.

Request

data DescribeInstanceAttribute

Request constructor

describeInstanceAttribute

Arguments

:: Text
dia1InstanceId

-> Text
dia1Attribute

-> DescribeInstanceAttribute

DescribeInstanceAttribute constructor.

The fields accessible through corresponding lenses are:

dia1Attribute :: Text

dia1DryRun :: Maybe Bool

dia1InstanceId :: Text

Request lenses

dia1Attribute :: Lens' DescribeInstanceAttribute Text

The instance attribute.

dia1DryRun :: Lens' DescribeInstanceAttribute (Maybe Bool)

dia1InstanceId :: Lens' DescribeInstanceAttribute Text

The ID of the instance.

Response

data InstanceAttribute

Response constructor

instanceAttribute :: InstanceAttribute

InstanceAttribute constructor.

The fields accessible through corresponding lenses are:

iaBlockDeviceMappings :: [InstanceBlockDeviceMapping]

iaDisableApiTermination :: Maybe AttributeBooleanValue

iaEbsOptimized :: Maybe AttributeBooleanValue

iaGroups :: [GroupIdentifier]

iaInstanceId :: Maybe Text

iaInstanceInitiatedShutdownBehavior :: Maybe AttributeValue

iaInstanceType :: Maybe AttributeValue

iaKernelId :: Maybe AttributeValue

iaProductCodes :: [ProductCode]

iaRamdiskId :: Maybe AttributeValue

iaRootDeviceName :: Maybe AttributeValue

iaSourceDestCheck :: Maybe AttributeBooleanValue

iaSriovNetSupport :: Maybe AttributeValue

iaUserData :: Maybe AttributeValue

Response lenses

iaBlockDeviceMappings :: Lens' InstanceAttribute [InstanceBlockDeviceMapping]

The block device mapping of the instance.

iaDisableApiTermination :: Lens' InstanceAttribute (Maybe AttributeBooleanValue)

If the value is true, you can't terminate the instance through the Amazon EC2 console, CLI, or API; otherwise, you can.

iaEbsOptimized :: Lens' InstanceAttribute (Maybe AttributeBooleanValue)

Indicates whether the instance is optimized for EBS I/O.

iaGroups :: Lens' InstanceAttribute [GroupIdentifier]

The security groups associated with the instance.

iaInstanceId :: Lens' InstanceAttribute (Maybe Text)

The ID of the instance.

iaInstanceInitiatedShutdownBehavior :: Lens' InstanceAttribute (Maybe AttributeValue)

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

iaInstanceType :: Lens' InstanceAttribute (Maybe AttributeValue)

The instance type.

iaKernelId :: Lens' InstanceAttribute (Maybe AttributeValue)

The kernel ID.

iaProductCodes :: Lens' InstanceAttribute [ProductCode]

A list of product codes.

iaRamdiskId :: Lens' InstanceAttribute (Maybe AttributeValue)

The RAM disk ID.

iaRootDeviceName :: Lens' InstanceAttribute (Maybe AttributeValue)

The name of the root device (for example, devsda1).

iaSourceDestCheck :: Lens' InstanceAttribute (Maybe AttributeBooleanValue)

Indicates whether source/destination checking is enabled. A value of true means checking is enabled, and false means checking is disabled. This value must be false for a NAT instance to perform NAT.

iaSriovNetSupport :: Lens' InstanceAttribute (Maybe AttributeValue)

iaUserData :: Lens' InstanceAttribute (Maybe AttributeValue)

The Base64-encoded MIME user data.