Network.AWS.EC2.DescribeInternetGateways

Description

Describes one or more of your Internet gateways. Example This example describes all your Internet gateways. https:/ec2.amazonaws.com?Action=DescribeInternetGateways &amp;AUTHPARAMS 59dbff89-35bd-4eac-99ed-be587EXAMPLE igw-eaad4883EXAMPLE vpc-11ad4878 available.

Request

data DescribeInternetGateways

Request constructor

describeInternetGateways :: DescribeInternetGateways

Smart constructor for the minimum required parameters to construct a valid DescribeInternetGateways request.

The fields accessible through corresponding lenses are:

InternetGatewayIds :: [Text]

Filters :: [Filter]

Request lenses

dig1InternetGatewayIds :: Lens' DescribeInternetGateways [Text]

One or more Internet gateway IDs. Default: Describes all your Internet gateways.

dig1Filters :: Lens' DescribeInternetGateways [Filter]

One or more filters. attachment.state - The current state of the attachment between the gateway and the VPC. Present only if a VPC is attached. attachment.vpc-id - The ID of an attached VPC. internet-gateway-id - The ID of the Internet gateway. tag:key=value - The key/value combination of a tag assigned to the resource. tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag:key=value filter. tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter.

Response

data DescribeInternetGatewaysResponse

Response constructor

describeInternetGatewaysResponse :: DescribeInternetGatewaysResponse

Smart constructor for the minimum required parameters to construct a valid DescribeInternetGatewaysResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

InternetGateways :: [InternetGateway]

Response lenses

digrInternetGateways :: Lens' DescribeInternetGatewaysResponse [InternetGateway]

Information about one or more Internet gateways.