Network.AWS.EC2.DescribeMovingAddresses

Description

Describes your Elastic IP addresses that are being moved to the EC2-VPC platform, or that are being restored to the EC2-Classic platform. This request does not return information about any other Elastic IP addresses in your account.

Creating a Request

describeMovingAddresses :: DescribeMovingAddresses

Creates a value of DescribeMovingAddresses with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmaFilters - One or more filters. * moving-status - The status of the Elastic IP address (MovingToVpc | RestoringToClassic ).

dmaPublicIPs - One or more Elastic IP addresses.

dmaNextToken - The token to use to retrieve the next page of results.

dmaDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dmaMaxResults - The maximum number of results to return for the request in a single page. The remaining results of the initial request can be seen by sending another request with the returned NextToken value. This value can be between 5 and 1000; if MaxResults is given a value outside of this range, an error is returned. Default: If no value is provided, the default is 1000.

data DescribeMovingAddresses

Contains the parameters for DescribeMovingAddresses.

See: describeMovingAddresses smart constructor.

Request Lenses

dmaFilters :: Lens' DescribeMovingAddresses [Filter]

One or more filters. * moving-status - The status of the Elastic IP address (MovingToVpc | RestoringToClassic ).

dmaPublicIPs :: Lens' DescribeMovingAddresses [Text]

One or more Elastic IP addresses.

dmaNextToken :: Lens' DescribeMovingAddresses (Maybe Text)

The token to use to retrieve the next page of results.

dmaDryRun :: Lens' DescribeMovingAddresses (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dmaMaxResults :: Lens' DescribeMovingAddresses (Maybe Int)

The maximum number of results to return for the request in a single page. The remaining results of the initial request can be seen by sending another request with the returned NextToken value. This value can be between 5 and 1000; if MaxResults is given a value outside of this range, an error is returned. Default: If no value is provided, the default is 1000.

Destructuring the Response

describeMovingAddressesResponse

Arguments

:: Int
dmarsResponseStatus

-> DescribeMovingAddressesResponse

Creates a value of DescribeMovingAddressesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmarsMovingAddressStatuses - The status for each Elastic IP address.

dmarsNextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

dmarsResponseStatus - -- | The response status code.

data DescribeMovingAddressesResponse

Contains the output of DescribeMovingAddresses.

See: describeMovingAddressesResponse smart constructor.

Response Lenses

dmarsMovingAddressStatuses :: Lens' DescribeMovingAddressesResponse [MovingAddressStatus]

The status for each Elastic IP address.

dmarsNextToken :: Lens' DescribeMovingAddressesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

dmarsResponseStatus :: Lens' DescribeMovingAddressesResponse Int

- | The response status code.