Network.AWS.EC2.DescribeNatGateways

Description

Describes one or more of the your NAT gateways.

This operation returns paginated results.

Creating a Request

describeNatGateways :: DescribeNatGateways

Creates a value of DescribeNatGateways with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dngNatGatewayIds - One or more NAT gateway IDs.

dngNextToken - The token to retrieve the next page of results.

dngFilter - One or more filters. * nat-gateway-id - The ID of the NAT gateway. * state - The state of the NAT gateway (pending | failed | available | deleting | deleted ). * subnet-id - The ID of the subnet in which the NAT gateway resides. * vpc-id - The ID of the VPC in which the NAT gateway resides.

dngMaxResults - The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results. Constraint: If the value specified is greater than 1000, we return only 1000 items.

data DescribeNatGateways

Contains the parameters for DescribeNatGateways.

See: describeNatGateways smart constructor.

Request Lenses

dngNatGatewayIds :: Lens' DescribeNatGateways [Text]

One or more NAT gateway IDs.

dngNextToken :: Lens' DescribeNatGateways (Maybe Text)

The token to retrieve the next page of results.

dngFilter :: Lens' DescribeNatGateways [Filter]

One or more filters. * nat-gateway-id - The ID of the NAT gateway. * state - The state of the NAT gateway (pending | failed | available | deleting | deleted ). * subnet-id - The ID of the subnet in which the NAT gateway resides. * vpc-id - The ID of the VPC in which the NAT gateway resides.

dngMaxResults :: Lens' DescribeNatGateways (Maybe Int)

The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results. Constraint: If the value specified is greater than 1000, we return only 1000 items.

Destructuring the Response

describeNatGatewaysResponse

Arguments

:: Int
dngrsResponseStatus

-> DescribeNatGatewaysResponse

Creates a value of DescribeNatGatewaysResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dngrsNatGateways - Information about the NAT gateways.

dngrsNextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

dngrsResponseStatus - -- | The response status code.

data DescribeNatGatewaysResponse

Contains the output of DescribeNatGateways.

See: describeNatGatewaysResponse smart constructor.

Response Lenses

dngrsNatGateways :: Lens' DescribeNatGatewaysResponse [NatGateway]

Information about the NAT gateways.

dngrsNextToken :: Lens' DescribeNatGatewaysResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

dngrsResponseStatus :: Lens' DescribeNatGatewaysResponse Int

- | The response status code.