Network.AWS.EC2.DescribeNetworkAcls

Description

Describes one or more of your network ACLs.

For more information about network ACLs, see Network ACLs in the AmazonVirtual Private Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeNetworkAcls.html

Request

data DescribeNetworkAcls

Request constructor

describeNetworkAcls :: DescribeNetworkAcls

DescribeNetworkAcls constructor.

The fields accessible through corresponding lenses are:

dna1DryRun :: Maybe Bool

dna1Filters :: [Filter]

dna1NetworkAclIds :: [Text]

Request lenses

dna1DryRun :: Lens' DescribeNetworkAcls (Maybe Bool)

dna1Filters :: Lens' DescribeNetworkAcls [Filter]

One or more filters.

'association.association-id' - The ID of an association ID for the ACL.

'association.network-acl-id' - The ID of the network ACL involved in the association.

'association.subnet-id' - The ID of the subnet involved in the association.

'default' - Indicates whether the ACL is the default network ACL for the VPC.

'entry.cidr' - The CIDR range specified in the entry.

'entry.egress' - Indicates whether the entry applies to egress traffic.

'entry.icmp.code' - The ICMP code specified in the entry, if any.

'entry.icmp.type' - The ICMP type specified in the entry, if any.

'entry.port-range.from' - The start of the port range specified in the entry.

'entry.port-range.to' - The end of the port range specified in the entry.

'entry.protocol' - The protocol specified in the entry (tcp | udp | icmp or a protocol number).

'entry.rule-action' - Allows or denies the matching traffic (allow | deny).

'entry.rule-number' - The number of an entry (in other words, rule) in the ACL's set of entries.

'network-acl-id' - The ID of the network ACL.

tag:key=value - The key/value combination of a tag assigned to the resource.

'tag-key' - The key of a tag assigned to the resource. This filter is independent of the 'tag-value' filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag:key=value filter.

'tag-value' - The value of a tag assigned to the resource. This filter is independent of the 'tag-key' filter.

'vpc-id' - The ID of the VPC for the network ACL.

dna1NetworkAclIds :: Lens' DescribeNetworkAcls [Text]

One or more network ACL IDs.

Default: Describes all your network ACLs.

Response

data DescribeNetworkAclsResponse

Response constructor

describeNetworkAclsResponse :: DescribeNetworkAclsResponse

DescribeNetworkAclsResponse constructor.

The fields accessible through corresponding lenses are:

dnarNetworkAcls :: [NetworkAcl]

Response lenses

dnarNetworkAcls :: Lens' DescribeNetworkAclsResponse [NetworkAcl]

Information about one or more network ACLs.