Network.AWS.EC2.DescribeNetworkInterfaceAttribute

Description

Describes a network interface attribute. You can specify only one attribute at a time. Example This example describes the sourceDestCheck attribute of the specified network interface. https:/ec2.amazonaws.com?Action=DescribeNetworkInterfaceAttribute &amp;NetworkInterfaceId=eni-686ea200 &amp;Attribute=sourceDestCheck &amp;AUTHPARAMS &lt;DescribeNetworkInterfaceAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;7a20c6b2-d71c-45fb-bba7-37306850544b&lt;/requestId&gt; &lt;networkInterfaceId&gt;eni-686ea200&lt;/networkInterfaceId&gt; &lt;sourceDestCheck&gt; &lt;value&gt;true&lt;/value&gt; &lt;/sourceDestCheck&gt; &lt;/DescribeNetworkInterfaceAttributeResponse&gt;.

Request

data DescribeNetworkInterfaceAttribute

Request constructor

mkDescribeNetworkInterfaceAttribute

Arguments

:: Text
dniaNetworkInterfaceId

-> DescribeNetworkInterfaceAttribute

Smart constructor for the minimum required parameters to construct a valid DescribeNetworkInterfaceAttribute request.

The fields accessible through corresponding lenses are:

NetworkInterfaceId :: Text

Attribute :: Maybe NetworkInterfaceAttribute

Request lenses

dniaNetworkInterfaceId :: Lens' DescribeNetworkInterfaceAttribute Text

The ID of the network interface.

dniaAttribute :: Lens' DescribeNetworkInterfaceAttribute (Maybe NetworkInterfaceAttribute)

The attribute of the network interface.

Response

data DescribeNetworkInterfaceAttributeResponse

Response constructor

mkDescribeNetworkInterfaceAttributeResponse :: DescribeNetworkInterfaceAttributeResponse

Smart constructor for the minimum required parameters to construct a valid DescribeNetworkInterfaceAttributeResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

NetworkInterfaceId :: Maybe Text

Description :: Maybe AttributeValue

SourceDestCheck :: Maybe AttributeBooleanValue

Groups :: [GroupIdentifier]

Attachment :: Maybe NetworkInterfaceAttachment

Response lenses

dniarNetworkInterfaceId :: Lens' DescribeNetworkInterfaceAttributeResponse (Maybe Text)

The ID of the network interface.

dniarDescription :: Lens' DescribeNetworkInterfaceAttributeResponse (Maybe AttributeValue)

The description of the network interface.

dniarSourceDestCheck :: Lens' DescribeNetworkInterfaceAttributeResponse (Maybe AttributeBooleanValue)

Indicates whether source/destination checking is enabled.

dniarGroups :: Lens' DescribeNetworkInterfaceAttributeResponse [GroupIdentifier]

The security groups associated with the network interface.

dniarAttachment :: Lens' DescribeNetworkInterfaceAttributeResponse (Maybe NetworkInterfaceAttachment)

The attachment (if any) of the network interface.