Network.AWS.EC2.DescribeNetworkInterfacePermissions

Description

Describes the permissions for your network interfaces.

Creating a Request

describeNetworkInterfacePermissions :: DescribeNetworkInterfacePermissions

Creates a value of DescribeNetworkInterfacePermissions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dnipFilters - One or more filters. * network-interface-permission.network-interface-permission-id - The ID of the permission. * network-interface-permission.network-interface-id - The ID of the network interface. * network-interface-permission.aws-account-id - The AWS account ID. * network-interface-permission.aws-service - The AWS service. * network-interface-permission.permission - The type of permission (INSTANCE-ATTACH | EIP-ASSOCIATE ).

dnipNextToken - The token to request the next page of results.

dnipNetworkInterfacePermissionIds - One or more network interface permission IDs.

dnipMaxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. If this parameter is not specified, up to 50 results are returned by default.

data DescribeNetworkInterfacePermissions

Contains the parameters for DescribeNetworkInterfacePermissions.

See: describeNetworkInterfacePermissions smart constructor.

Request Lenses

dnipFilters :: Lens' DescribeNetworkInterfacePermissions [Filter]

One or more filters. * network-interface-permission.network-interface-permission-id - The ID of the permission. * network-interface-permission.network-interface-id - The ID of the network interface. * network-interface-permission.aws-account-id - The AWS account ID. * network-interface-permission.aws-service - The AWS service. * network-interface-permission.permission - The type of permission (INSTANCE-ATTACH | EIP-ASSOCIATE ).

dnipNextToken :: Lens' DescribeNetworkInterfacePermissions (Maybe Text)

The token to request the next page of results.

dnipNetworkInterfacePermissionIds :: Lens' DescribeNetworkInterfacePermissions [Text]

One or more network interface permission IDs.

dnipMaxResults :: Lens' DescribeNetworkInterfacePermissions (Maybe Int)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. If this parameter is not specified, up to 50 results are returned by default.

Destructuring the Response

describeNetworkInterfacePermissionsResponse

Arguments

:: Int
drsResponseStatus

-> DescribeNetworkInterfacePermissionsResponse

Creates a value of DescribeNetworkInterfacePermissionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsNetworkInterfacePermissions - The network interface permissions.

drsNextToken - The token to use to retrieve the next page of results.

drsResponseStatus - -- | The response status code.

data DescribeNetworkInterfacePermissionsResponse

Contains the output for DescribeNetworkInterfacePermissions.

See: describeNetworkInterfacePermissionsResponse smart constructor.

Response Lenses

drsNetworkInterfacePermissions :: Lens' DescribeNetworkInterfacePermissionsResponse [NetworkInterfacePermission]

The network interface permissions.

drsNextToken :: Lens' DescribeNetworkInterfacePermissionsResponse (Maybe Text)

The token to use to retrieve the next page of results.

drsResponseStatus :: Lens' DescribeNetworkInterfacePermissionsResponse Int

- | The response status code.