Network.AWS.EC2.DescribePlacementGroups

Description

Describes one or more of your placement groups. For more information about placement groups and cluster instances, see Cluster Instances in the Amazon Elastic Compute Cloud User Guide. Example This example describes the placement group named XYZ-cluster. https:/ec2.amazonaws.com?Action=DescribePlacementGroups &amp;GroupName.1=XYZ-cluster &amp;AUTHPARAMS d4904fd9-82c2-4ea5-adfe-a9cc3EXAMPLE XYZ-cluster cluster available Example This example filters the response to include only placement groups that include the string Project in the name. https:/ec2.amazonaws.com?Action=DescribePlacementGroups &amp;Filter.1.Name=group-name &amp;Filter.1.Value=*Project* &amp;AUTHPARAMS d4904fd9-82c2-4ea5-adfe-a9cc3EXAMPLE Project-cluster cluster available.

Request

data DescribePlacementGroups

Request constructor

mkDescribePlacementGroups :: DescribePlacementGroups

Smart constructor for the minimum required parameters to construct a valid DescribePlacementGroups request.

The fields accessible through corresponding lenses are:

GroupNames :: [Text]

Filters :: [Filter]

Request lenses

dpg1GroupNames :: Lens' DescribePlacementGroups [Text]

One or more placement group names. Default: Describes all your placement groups, or only those otherwise specified.

dpg1Filters :: Lens' DescribePlacementGroups [Filter]

One or more filters. group-name - The name of the placement group. state - The state of the placement group (pending | available | deleting | deleted). strategy - The strategy of the placement group (cluster).

Response

data DescribePlacementGroupsResponse

Response constructor

mkDescribePlacementGroupsResponse :: DescribePlacementGroupsResponse

Smart constructor for the minimum required parameters to construct a valid DescribePlacementGroupsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

PlacementGroups :: [PlacementGroup]

Response lenses

dpgrPlacementGroups :: Lens' DescribePlacementGroupsResponse [PlacementGroup]

One or more placement groups.