Network.AWS.EC2.DescribePrefixLists

Description

Describes available AWS services in a prefix list format, which includes the prefix list name and prefix list ID of the service and the IP address range for the service. A prefix list ID is required for creating an outbound security group rule that allows traffic from a VPC to access an AWS service through a VPC endpoint.

Creating a Request

describePrefixLists :: DescribePrefixLists

Creates a value of DescribePrefixLists with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dplFilters - One or more filters. * prefix-list-id : The ID of a prefix list. * prefix-list-name : The name of a prefix list.

dplPrefixListIds - One or more prefix list IDs.

dplNextToken - The token for the next set of items to return. (You received this token from a prior call.)

dplDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dplMaxResults - The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results. Constraint: If the value specified is greater than 1000, we return only 1000 items.

data DescribePrefixLists

Contains the parameters for DescribePrefixLists.

See: describePrefixLists smart constructor.

Request Lenses

dplFilters :: Lens' DescribePrefixLists [Filter]

One or more filters. * prefix-list-id : The ID of a prefix list. * prefix-list-name : The name of a prefix list.

dplPrefixListIds :: Lens' DescribePrefixLists [Text]

One or more prefix list IDs.

dplNextToken :: Lens' DescribePrefixLists (Maybe Text)

The token for the next set of items to return. (You received this token from a prior call.)

dplDryRun :: Lens' DescribePrefixLists (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dplMaxResults :: Lens' DescribePrefixLists (Maybe Int)

The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results. Constraint: If the value specified is greater than 1000, we return only 1000 items.

Destructuring the Response

describePrefixListsResponse

Arguments

:: Int
dplrsResponseStatus

-> DescribePrefixListsResponse

Creates a value of DescribePrefixListsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dplrsNextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dplrsPrefixLists - All available prefix lists.

dplrsResponseStatus - -- | The response status code.

data DescribePrefixListsResponse

Contains the output of DescribePrefixLists.

See: describePrefixListsResponse smart constructor.

Response Lenses

dplrsNextToken :: Lens' DescribePrefixListsResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dplrsPrefixLists :: Lens' DescribePrefixListsResponse [PrefixList]

All available prefix lists.

dplrsResponseStatus :: Lens' DescribePrefixListsResponse Int

- | The response status code.