Network.AWS.EC2.DescribeRegions

Description

Describes one or more regions that are currently available to you.

For a list of the regions supported by Amazon EC2, see Regions and Endpoints.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeRegions.html

Request

data DescribeRegions

Request constructor

describeRegions :: DescribeRegions

DescribeRegions constructor.

The fields accessible through corresponding lenses are:

dr1DryRun :: Maybe Bool

dr1Filters :: [Filter]

dr1RegionNames :: [Text]

Request lenses

dr1DryRun :: Lens' DescribeRegions (Maybe Bool)

dr1Filters :: Lens' DescribeRegions [Filter]

One or more filters.

endpoint - The endpoint of the region (for example, 'ec2.us-east-1.amazonaws.com').

'region-name' - The name of the region (for example, 'us-east-1').

dr1RegionNames :: Lens' DescribeRegions [Text]

The names of one or more regions.

Response

data DescribeRegionsResponse

Response constructor

describeRegionsResponse :: DescribeRegionsResponse

DescribeRegionsResponse constructor.

The fields accessible through corresponding lenses are:

drrRegions :: [Region]

Response lenses

drrRegions :: Lens' DescribeRegionsResponse [Region]

Information about one or more regions.