Network.AWS.EC2.DescribeReservedInstances

Description

Describes one or more of the Reserved Instances that you purchased.

For more information about Reserved Instances, see Reserved Instances in the Amazon Elastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeReservedInstances.html

Request

data DescribeReservedInstances

Request constructor

describeReservedInstances :: DescribeReservedInstances

DescribeReservedInstances constructor.

The fields accessible through corresponding lenses are:

driDryRun :: Maybe Bool

driFilters :: [Filter]

driOfferingType :: Maybe OfferingTypeValues

driReservedInstancesIds :: [Text]

Request lenses

driDryRun :: Lens' DescribeReservedInstances (Maybe Bool)

driFilters :: Lens' DescribeReservedInstances [Filter]

One or more filters.

'availability-zone' - The Availability Zone where the Reserved Instance can be used.

duration - The duration of the Reserved Instance (one year or three years), in seconds ('31536000' | '94608000').

end - The time when the Reserved Instance expires (for example, 2014-08-07T11:54:42.000Z).

'fixed-price' - The purchase price of the Reserved Instance (for example, 9800.0).

'instance-type' - The instance type on which the Reserved Instance can be used.

'product-description' - The product description of the Reserved Instance ('Linux/UNIX' | 'Linux/UNIX (Amazon VPC)' | Windows | 'Windows (Amazon VPC)').

'reserved-instances-id' - The ID of the Reserved Instance.

start - The time at which the Reserved Instance purchase request was placed (for example, 2014-08-07T11:54:42.000Z).

state - The state of the Reserved Instance ('pending-payment' | active | 'payment-failed' | retired).

tag:key=value - The key/value combination of a tag assigned to the resource.

'tag-key' - The key of a tag assigned to the resource. This filter is independent of the 'tag-value' filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag:key=value filter.

'tag-value' - The value of a tag assigned to the resource. This filter is independent of the 'tag-key' filter.

'usage-price' - The usage price of the Reserved Instance, per hour (for example, 0.84).

driOfferingType :: Lens' DescribeReservedInstances (Maybe OfferingTypeValues)

The Reserved Instance offering type. If you are using tools that predate the 2011-11-01 API version, you only have access to the 'Medium Utilization' Reserved Instance offering type.

driReservedInstancesIds :: Lens' DescribeReservedInstances [Text]

One or more Reserved Instance IDs.

Default: Describes all your Reserved Instances, or only those otherwise specified.

Response

data DescribeReservedInstancesResponse

Response constructor

describeReservedInstancesResponse :: DescribeReservedInstancesResponse

DescribeReservedInstancesResponse constructor.

The fields accessible through corresponding lenses are:

drirReservedInstances :: [ReservedInstances]

Response lenses

drirReservedInstances :: Lens' DescribeReservedInstancesResponse [ReservedInstances]

A list of Reserved Instances.