Network.AWS.EC2.DescribeReservedInstancesModifications

Description

Describes the modifications made to your Reserved Instances. If no parameter is specified, information about all your Reserved Instances modification requests is returned. If a modification ID is specified, only information about the specific modification is returned. Example 1 https:/ec2.amazonaws.com?Action=DescribeReservedInstancesModifications &amp;AUTHPARAMS Example 2 This example filters the response to include only Reserved Instances modification requests with status processing. https:/ec2.amazonaws.com?Action=DescribeReservedInstancesModifications &amp;Filter.1.Name=status &amp;Filter.1.Value.1=processing &amp;AUTHPARAMS.

Request

data DescribeReservedInstancesModifications

Request constructor

describeReservedInstancesModifications :: DescribeReservedInstancesModifications

Smart constructor for the minimum required parameters to construct a valid DescribeReservedInstancesModifications request.

The fields accessible through corresponding lenses are:

ReservedInstancesModificationIds :: [Text]

NextToken :: Maybe Text

Filters :: [Filter]

Request lenses

drimReservedInstancesModificationIds :: Lens' DescribeReservedInstancesModifications [Text]

IDs for the submitted modification request.

drimNextToken :: Lens' DescribeReservedInstancesModifications (Maybe Text)

The token for the next page of data.

drimFilters :: Lens' DescribeReservedInstancesModifications [Filter]

One or more filters. client-token - The idempotency token for the modification request. create-date - The time when the modification request was created. effective-date - The time when the modification becomes effective. modification-result.reserved-instances-id - The ID for the Reserved Instances created as part of the modification request. This ID is only available when the status of the modification is fulfilled. modification-result.target-configuration.availability-zone - The Availability Zone for the new Reserved Instances. modification-result.target-configuration.instance-count - The number of new Reserved Instances. modification-result.target-configuration.instance-type - The instance type of the new Reserved Instances. modification-result.target-configuration.platform - The network platform of the new Reserved Instances (EC2-Classic | EC2-VPC). reserved-instances-id - The ID of the Reserved Instances modified. reserved-instances-modification-id - The ID of the modification request. status - The status of the Reserved Instances modification request (processing | fulfilled | failed). status-message - The reason for the status. update-date - The time when the modification request was last updated.

Response

data DescribeReservedInstancesModificationsResponse

Response constructor

describeReservedInstancesModificationsResponse :: DescribeReservedInstancesModificationsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeReservedInstancesModificationsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ReservedInstancesModifications :: [ReservedInstancesModification]

NextToken :: Maybe Text

Response lenses

drimrReservedInstancesModifications :: Lens' DescribeReservedInstancesModificationsResponse [ReservedInstancesModification]

The Reserved Instance modification information.

drimrNextToken :: Lens' DescribeReservedInstancesModificationsResponse (Maybe Text)

The token for the next page of data.