Network.AWS.EC2.DescribeReservedInstancesOfferings

Description

Describes Reserved Instance offerings that are available for purchase. With Reserved Instances, you purchase the right to launch instances for a period of time. During that time period, you do not receive insufficient capacity errors, and you pay a lower usage rate than the rate charged for On-Demand instances for the actual time used.

For more information, see Reserved Instance Marketplace in the AmazonElastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeReservedInstancesOfferings.html

Request

data DescribeReservedInstancesOfferings

Request constructor

describeReservedInstancesOfferings :: DescribeReservedInstancesOfferings

DescribeReservedInstancesOfferings constructor.

The fields accessible through corresponding lenses are:

drioAvailabilityZone :: Maybe Text

drioDryRun :: Maybe Bool

drioFilters :: [Filter]

drioIncludeMarketplace :: Maybe Bool

drioInstanceTenancy :: Maybe Tenancy

drioInstanceType :: Maybe InstanceType

drioMaxDuration :: Maybe Integer

drioMaxInstanceCount :: Maybe Int

drioMaxResults :: Maybe Int

drioMinDuration :: Maybe Integer

drioNextToken :: Maybe Text

drioOfferingType :: Maybe OfferingTypeValues

drioProductDescription :: Maybe RIProductDescription

drioReservedInstancesOfferingIds :: [Text]

Request lenses

drioAvailabilityZone :: Lens' DescribeReservedInstancesOfferings (Maybe Text)

The Availability Zone in which the Reserved Instance can be used.

drioDryRun :: Lens' DescribeReservedInstancesOfferings (Maybe Bool)

drioFilters :: Lens' DescribeReservedInstancesOfferings [Filter]

One or more filters.

'availability-zone' - The Availability Zone where the Reserved Instance can be used.

duration - The duration of the Reserved Instance (for example, one year or three years), in seconds ('31536000' | '94608000').

'fixed-price' - The purchase price of the Reserved Instance (for example, 9800.0).

'instance-type' - The instance type on which the Reserved Instance can be used.

marketplace - Set to true to show only Reserved Instance Marketplace offerings. When this filter is not used, which is the default behavior, all offerings from AWS and Reserved Instance Marketplace are listed.

'product-description' - The description of the Reserved Instance ('Linux/UNIX' | 'Linux/UNIX (Amazon VPC)' | Windows | 'Windows (Amazon VPC)').

'reserved-instances-offering-id' - The Reserved Instances offering ID.

'usage-price' - The usage price of the Reserved Instance, per hour (for example, 0.84).

drioIncludeMarketplace :: Lens' DescribeReservedInstancesOfferings (Maybe Bool)

Include Marketplace offerings in the response.

drioInstanceTenancy :: Lens' DescribeReservedInstancesOfferings (Maybe Tenancy)

The tenancy of the Reserved Instance offering. A Reserved Instance with dedicated tenancy runs on single-tenant hardware and can only be launched within a VPC.

Default: 'default'

drioInstanceType :: Lens' DescribeReservedInstancesOfferings (Maybe InstanceType)

The instance type on which the Reserved Instance can be used. For more information, see Instance Types in the Amazon Elastic Compute Cloud UserGuide for Linux.

drioMaxDuration :: Lens' DescribeReservedInstancesOfferings (Maybe Integer)

The maximum duration (in seconds) to filter when searching for offerings.

Default: 94608000 (3 years)

drioMaxInstanceCount :: Lens' DescribeReservedInstancesOfferings (Maybe Int)

The maximum number of instances to filter when searching for offerings.

Default: 20

drioMaxResults :: Lens' DescribeReservedInstancesOfferings (Maybe Int)

The maximum number of results to return for the request in a single page. The remaining results of the initial request can be seen by sending another request with the returned NextToken value. The maximum is 100.

Default: 100

drioMinDuration :: Lens' DescribeReservedInstancesOfferings (Maybe Integer)

The minimum duration (in seconds) to filter when searching for offerings.

Default: 2592000 (1 month)

drioNextToken :: Lens' DescribeReservedInstancesOfferings (Maybe Text)

The token to retrieve the next page of results.

drioOfferingType :: Lens' DescribeReservedInstancesOfferings (Maybe OfferingTypeValues)

The Reserved Instance offering type. If you are using tools that predate the 2011-11-01 API version, you only have access to the 'Medium Utilization' Reserved Instance offering type.

drioProductDescription :: Lens' DescribeReservedInstancesOfferings (Maybe RIProductDescription)

The Reserved Instance description. Instances that include '(Amazon VPC)' in the description are for use with Amazon VPC.

drioReservedInstancesOfferingIds :: Lens' DescribeReservedInstancesOfferings [Text]

One or more Reserved Instances offering IDs.

Response

data DescribeReservedInstancesOfferingsResponse

Response constructor

describeReservedInstancesOfferingsResponse :: DescribeReservedInstancesOfferingsResponse

DescribeReservedInstancesOfferingsResponse constructor.

The fields accessible through corresponding lenses are:

driorNextToken :: Maybe Text

driorReservedInstancesOfferings :: [ReservedInstancesOffering]

Response lenses

driorNextToken :: Lens' DescribeReservedInstancesOfferingsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

driorReservedInstancesOfferings :: Lens' DescribeReservedInstancesOfferingsResponse [ReservedInstancesOffering]

A list of Reserved Instances offerings.