Network.AWS.EC2.DescribeScheduledInstanceAvailability

Description

Finds available schedules that meet the specified criteria.

You can search for an available schedule no more than 3 months in advance. You must meet the minimum required duration of 1,200 hours per year. For example, the minimum daily schedule is 4 hours, the minimum weekly schedule is 24 hours, and the minimum monthly schedule is 100 hours.

After you find a schedule that meets your needs, call PurchaseScheduledInstances to purchase Scheduled Instances with that schedule.

Creating a Request

describeScheduledInstanceAvailability

Arguments

:: SlotDateTimeRangeRequest
dsiaFirstSlotStartTimeRange

-> ScheduledInstanceRecurrenceRequest
dsiaRecurrence

-> DescribeScheduledInstanceAvailability

Creates a value of DescribeScheduledInstanceAvailability with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsiaMinSlotDurationInHours - The minimum available duration, in hours. The minimum required duration is 1,200 hours per year. For example, the minimum daily schedule is 4 hours, the minimum weekly schedule is 24 hours, and the minimum monthly schedule is 100 hours.

dsiaFilters - One or more filters. * availability-zone - The Availability Zone (for example, us-west-2a ). * instance-type - The instance type (for example, c4.large ). * network-platform - The network platform (EC2-Classic or EC2-VPC ). * platform - The platform (Linux/UNIX or Windows ).

dsiaNextToken - The token for the next set of results.

dsiaMaxSlotDurationInHours - The maximum available duration, in hours. This value must be greater than MinSlotDurationInHours and less than 1,720.

dsiaDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dsiaMaxResults - The maximum number of results to return in a single call. This value can be between 5 and 300. The default value is 300. To retrieve the remaining results, make another call with the returned NextToken value.

dsiaFirstSlotStartTimeRange - The time period for the first schedule to start.

dsiaRecurrence - The schedule recurrence.

data DescribeScheduledInstanceAvailability

Contains the parameters for DescribeScheduledInstanceAvailability.

See: describeScheduledInstanceAvailability smart constructor.

Request Lenses

dsiaMinSlotDurationInHours :: Lens' DescribeScheduledInstanceAvailability (Maybe Int)

The minimum available duration, in hours. The minimum required duration is 1,200 hours per year. For example, the minimum daily schedule is 4 hours, the minimum weekly schedule is 24 hours, and the minimum monthly schedule is 100 hours.

dsiaFilters :: Lens' DescribeScheduledInstanceAvailability [Filter]

One or more filters. * availability-zone - The Availability Zone (for example, us-west-2a ). * instance-type - The instance type (for example, c4.large ). * network-platform - The network platform (EC2-Classic or EC2-VPC ). * platform - The platform (Linux/UNIX or Windows ).

dsiaNextToken :: Lens' DescribeScheduledInstanceAvailability (Maybe Text)

The token for the next set of results.

dsiaMaxSlotDurationInHours :: Lens' DescribeScheduledInstanceAvailability (Maybe Int)

The maximum available duration, in hours. This value must be greater than MinSlotDurationInHours and less than 1,720.

dsiaDryRun :: Lens' DescribeScheduledInstanceAvailability (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dsiaMaxResults :: Lens' DescribeScheduledInstanceAvailability (Maybe Int)

The maximum number of results to return in a single call. This value can be between 5 and 300. The default value is 300. To retrieve the remaining results, make another call with the returned NextToken value.

dsiaFirstSlotStartTimeRange :: Lens' DescribeScheduledInstanceAvailability SlotDateTimeRangeRequest

The time period for the first schedule to start.

dsiaRecurrence :: Lens' DescribeScheduledInstanceAvailability ScheduledInstanceRecurrenceRequest

The schedule recurrence.

Destructuring the Response

describeScheduledInstanceAvailabilityResponse

Arguments

:: Int
dsiarsResponseStatus

-> DescribeScheduledInstanceAvailabilityResponse

Creates a value of DescribeScheduledInstanceAvailabilityResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsiarsScheduledInstanceAvailabilitySet - Information about the available Scheduled Instances.

dsiarsNextToken - The token required to retrieve the next set of results. This value is null when there are no more results to return.

dsiarsResponseStatus - -- | The response status code.

data DescribeScheduledInstanceAvailabilityResponse

Contains the output of DescribeScheduledInstanceAvailability.

See: describeScheduledInstanceAvailabilityResponse smart constructor.

Response Lenses

dsiarsScheduledInstanceAvailabilitySet :: Lens' DescribeScheduledInstanceAvailabilityResponse [ScheduledInstanceAvailability]

Information about the available Scheduled Instances.

dsiarsNextToken :: Lens' DescribeScheduledInstanceAvailabilityResponse (Maybe Text)

The token required to retrieve the next set of results. This value is null when there are no more results to return.

dsiarsResponseStatus :: Lens' DescribeScheduledInstanceAvailabilityResponse Int

- | The response status code.