Network.AWS.EC2.DescribeScheduledInstances

Description

Describes one or more of your Scheduled Instances.

Creating a Request

describeScheduledInstances :: DescribeScheduledInstances

Creates a value of DescribeScheduledInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsiFilters - One or more filters. * availability-zone - The Availability Zone (for example, us-west-2a ). * instance-type - The instance type (for example, c4.large ). * network-platform - The network platform (EC2-Classic or EC2-VPC ). * platform - The platform (Linux/UNIX or Windows ).

dsiSlotStartTimeRange - The time period for the first schedule to start.

dsiNextToken - The token for the next set of results.

dsiScheduledInstanceIds - One or more Scheduled Instance IDs.

dsiDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dsiMaxResults - The maximum number of results to return in a single call. This value can be between 5 and 300. The default value is 100. To retrieve the remaining results, make another call with the returned NextToken value.

data DescribeScheduledInstances

Contains the parameters for DescribeScheduledInstances.

See: describeScheduledInstances smart constructor.

Request Lenses

dsiFilters :: Lens' DescribeScheduledInstances [Filter]

One or more filters. * availability-zone - The Availability Zone (for example, us-west-2a ). * instance-type - The instance type (for example, c4.large ). * network-platform - The network platform (EC2-Classic or EC2-VPC ). * platform - The platform (Linux/UNIX or Windows ).

dsiSlotStartTimeRange :: Lens' DescribeScheduledInstances (Maybe SlotStartTimeRangeRequest)

The time period for the first schedule to start.

dsiNextToken :: Lens' DescribeScheduledInstances (Maybe Text)

The token for the next set of results.

dsiScheduledInstanceIds :: Lens' DescribeScheduledInstances [Text]

One or more Scheduled Instance IDs.

dsiDryRun :: Lens' DescribeScheduledInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dsiMaxResults :: Lens' DescribeScheduledInstances (Maybe Int)

The maximum number of results to return in a single call. This value can be between 5 and 300. The default value is 100. To retrieve the remaining results, make another call with the returned NextToken value.

Destructuring the Response

describeScheduledInstancesResponse

Arguments

:: Int
dsirsResponseStatus

-> DescribeScheduledInstancesResponse

Creates a value of DescribeScheduledInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsirsNextToken - The token required to retrieve the next set of results. This value is null when there are no more results to return.

dsirsScheduledInstanceSet - Information about the Scheduled Instances.

dsirsResponseStatus - -- | The response status code.

data DescribeScheduledInstancesResponse

Contains the output of DescribeScheduledInstances.

See: describeScheduledInstancesResponse smart constructor.

Response Lenses

dsirsNextToken :: Lens' DescribeScheduledInstancesResponse (Maybe Text)

The token required to retrieve the next set of results. This value is null when there are no more results to return.

dsirsScheduledInstanceSet :: Lens' DescribeScheduledInstancesResponse [ScheduledInstance]

Information about the Scheduled Instances.

dsirsResponseStatus :: Lens' DescribeScheduledInstancesResponse Int

- | The response status code.