Network.AWS.EC2.DescribeSecurityGroupReferences

Description

EC2-VPC only

Describes the VPCs on the other side of a VPC peering connection that are referencing the security groups you've specified in this request.

Creating a Request

describeSecurityGroupReferences :: DescribeSecurityGroupReferences

Creates a value of DescribeSecurityGroupReferences with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsgrDryRun - Checks whether you have the required permissions for the operation, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

dsgrGroupId - One or more security group IDs in your account.

data DescribeSecurityGroupReferences

See: describeSecurityGroupReferences smart constructor.

Request Lenses

dsgrDryRun :: Lens' DescribeSecurityGroupReferences (Maybe Bool)

Checks whether you have the required permissions for the operation, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

dsgrGroupId :: Lens' DescribeSecurityGroupReferences [Text]

One or more security group IDs in your account.

Destructuring the Response

describeSecurityGroupReferencesResponse

Arguments

:: Int
dsgrrsResponseStatus

-> DescribeSecurityGroupReferencesResponse

Creates a value of DescribeSecurityGroupReferencesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsgrrsSecurityGroupReferenceSet - Information about the VPCs with the referencing security groups.

dsgrrsResponseStatus - -- | The response status code.

data DescribeSecurityGroupReferencesResponse

See: describeSecurityGroupReferencesResponse smart constructor.

Response Lenses

dsgrrsSecurityGroupReferenceSet :: Lens' DescribeSecurityGroupReferencesResponse [SecurityGroupReference]

Information about the VPCs with the referencing security groups.

dsgrrsResponseStatus :: Lens' DescribeSecurityGroupReferencesResponse Int

- | The response status code.