Network.AWS.EC2.DescribeSnapshotAttribute

Description

Describes the specified attribute of the specified snapshot. You can specify only one attribute at a time.

For more information about Amazon EBS snapshots, see Amazon EBS Snapshots in the Amazon Elastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeSnapshotAttribute.html

Request

data DescribeSnapshotAttribute

Request constructor

describeSnapshotAttribute

Arguments

:: Text
dsaSnapshotId

-> SnapshotAttributeName
dsaAttribute

-> DescribeSnapshotAttribute

DescribeSnapshotAttribute constructor.

The fields accessible through corresponding lenses are:

dsaAttribute :: SnapshotAttributeName

dsaDryRun :: Maybe Bool

dsaSnapshotId :: Text

Request lenses

dsaAttribute :: Lens' DescribeSnapshotAttribute SnapshotAttributeName

The snapshot attribute you would like to view.

dsaDryRun :: Lens' DescribeSnapshotAttribute (Maybe Bool)

dsaSnapshotId :: Lens' DescribeSnapshotAttribute Text

The ID of the Amazon EBS snapshot.

Response

data DescribeSnapshotAttributeResponse

Response constructor

describeSnapshotAttributeResponse :: DescribeSnapshotAttributeResponse

DescribeSnapshotAttributeResponse constructor.

The fields accessible through corresponding lenses are:

dsarCreateVolumePermissions :: [CreateVolumePermission]

dsarProductCodes :: [ProductCode]

dsarSnapshotId :: Maybe Text

Response lenses

dsarCreateVolumePermissions :: Lens' DescribeSnapshotAttributeResponse [CreateVolumePermission]

A list of permissions for creating volumes from the snapshot.

dsarProductCodes :: Lens' DescribeSnapshotAttributeResponse [ProductCode]

A list of product codes.

dsarSnapshotId :: Lens' DescribeSnapshotAttributeResponse (Maybe Text)

The ID of the Amazon EBS snapshot.