Network.AWS.EC2.DescribeSpotFleetInstances

Description

Describes the running instances for the specified Spot fleet.

Creating a Request

describeSpotFleetInstances

Arguments

:: Text
dsfiSpotFleetRequestId

-> DescribeSpotFleetInstances

Creates a value of DescribeSpotFleetInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsfiNextToken - The token for the next set of results.

dsfiDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dsfiMaxResults - The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

dsfiSpotFleetRequestId - The ID of the Spot fleet request.

data DescribeSpotFleetInstances

Contains the parameters for DescribeSpotFleetInstances.

See: describeSpotFleetInstances smart constructor.

Request Lenses

dsfiNextToken :: Lens' DescribeSpotFleetInstances (Maybe Text)

The token for the next set of results.

dsfiDryRun :: Lens' DescribeSpotFleetInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dsfiMaxResults :: Lens' DescribeSpotFleetInstances (Maybe Int)

The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

dsfiSpotFleetRequestId :: Lens' DescribeSpotFleetInstances Text

The ID of the Spot fleet request.

Destructuring the Response

describeSpotFleetInstancesResponse

Arguments

:: Int
dsfirsResponseStatus

-> Text
dsfirsSpotFleetRequestId

-> DescribeSpotFleetInstancesResponse

Creates a value of DescribeSpotFleetInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsfirsNextToken - The token required to retrieve the next set of results. This value is null when there are no more results to return.

dsfirsResponseStatus - -- | The response status code.

dsfirsActiveInstances - The running instances. Note that this list is refreshed periodically and might be out of date.

dsfirsSpotFleetRequestId - The ID of the Spot fleet request.

data DescribeSpotFleetInstancesResponse

Contains the output of DescribeSpotFleetInstances.

See: describeSpotFleetInstancesResponse smart constructor.

Response Lenses

dsfirsNextToken :: Lens' DescribeSpotFleetInstancesResponse (Maybe Text)

The token required to retrieve the next set of results. This value is null when there are no more results to return.

dsfirsResponseStatus :: Lens' DescribeSpotFleetInstancesResponse Int

- | The response status code.

dsfirsActiveInstances :: Lens' DescribeSpotFleetInstancesResponse [ActiveInstance]

The running instances. Note that this list is refreshed periodically and might be out of date.

dsfirsSpotFleetRequestId :: Lens' DescribeSpotFleetInstancesResponse Text

The ID of the Spot fleet request.