Network.AWS.EC2.DescribeSpotFleetRequestHistory

Description

Describes the events for the specified Spot fleet request during the specified time.

Spot fleet events are delayed by up to 30 seconds before they can be described. This ensures that you can query by the last evaluated time and not miss a recorded event.

Creating a Request

describeSpotFleetRequestHistory

Arguments

:: Text
dsfrhSpotFleetRequestId

-> UTCTime
dsfrhStartTime

-> DescribeSpotFleetRequestHistory

Creates a value of DescribeSpotFleetRequestHistory with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsfrhNextToken - The token for the next set of results.

dsfrhEventType - The type of events to describe. By default, all events are described.

dsfrhDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dsfrhMaxResults - The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

dsfrhSpotFleetRequestId - The ID of the Spot fleet request.

dsfrhStartTime - The starting date and time for the events, in UTC format (for example, YYYY -MM -DD THH :MM :SS Z).

data DescribeSpotFleetRequestHistory

Contains the parameters for DescribeSpotFleetRequestHistory.

See: describeSpotFleetRequestHistory smart constructor.

Request Lenses

dsfrhNextToken :: Lens' DescribeSpotFleetRequestHistory (Maybe Text)

The token for the next set of results.

dsfrhEventType :: Lens' DescribeSpotFleetRequestHistory (Maybe EventType)

The type of events to describe. By default, all events are described.

dsfrhDryRun :: Lens' DescribeSpotFleetRequestHistory (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dsfrhMaxResults :: Lens' DescribeSpotFleetRequestHistory (Maybe Int)

The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

dsfrhSpotFleetRequestId :: Lens' DescribeSpotFleetRequestHistory Text

The ID of the Spot fleet request.

dsfrhStartTime :: Lens' DescribeSpotFleetRequestHistory UTCTime

The starting date and time for the events, in UTC format (for example, YYYY -MM -DD THH :MM :SS Z).

Destructuring the Response

describeSpotFleetRequestHistoryResponse

Arguments

:: Int
dsfrhrsResponseStatus

-> UTCTime
dsfrhrsLastEvaluatedTime

-> Text
dsfrhrsSpotFleetRequestId

-> UTCTime
dsfrhrsStartTime

-> DescribeSpotFleetRequestHistoryResponse

Creates a value of DescribeSpotFleetRequestHistoryResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsfrhrsNextToken - The token required to retrieve the next set of results. This value is null when there are no more results to return.

dsfrhrsResponseStatus - -- | The response status code.

dsfrhrsHistoryRecords - Information about the events in the history of the Spot fleet request.

dsfrhrsLastEvaluatedTime - The last date and time for the events, in UTC format (for example, YYYY -MM -DD THH :MM :SS Z). All records up to this time were retrieved. If nextToken indicates that there are more results, this value is not present.

dsfrhrsSpotFleetRequestId - The ID of the Spot fleet request.

dsfrhrsStartTime - The starting date and time for the events, in UTC format (for example, YYYY -MM -DD THH :MM :SS Z).

data DescribeSpotFleetRequestHistoryResponse

Contains the output of DescribeSpotFleetRequestHistory.

See: describeSpotFleetRequestHistoryResponse smart constructor.

Response Lenses

dsfrhrsNextToken :: Lens' DescribeSpotFleetRequestHistoryResponse (Maybe Text)

The token required to retrieve the next set of results. This value is null when there are no more results to return.

dsfrhrsResponseStatus :: Lens' DescribeSpotFleetRequestHistoryResponse Int

- | The response status code.

dsfrhrsHistoryRecords :: Lens' DescribeSpotFleetRequestHistoryResponse [HistoryRecord]

Information about the events in the history of the Spot fleet request.

dsfrhrsLastEvaluatedTime :: Lens' DescribeSpotFleetRequestHistoryResponse UTCTime

The last date and time for the events, in UTC format (for example, YYYY -MM -DD THH :MM :SS Z). All records up to this time were retrieved. If nextToken indicates that there are more results, this value is not present.

dsfrhrsSpotFleetRequestId :: Lens' DescribeSpotFleetRequestHistoryResponse Text

The ID of the Spot fleet request.

dsfrhrsStartTime :: Lens' DescribeSpotFleetRequestHistoryResponse UTCTime

The starting date and time for the events, in UTC format (for example, YYYY -MM -DD THH :MM :SS Z).