Network.AWS.EC2.DescribeSpotFleetRequests

Description

Describes your Spot fleet requests.

Spot fleet requests are deleted 48 hours after they are canceled and their instances are terminated.

This operation returns paginated results.

Creating a Request

describeSpotFleetRequests :: DescribeSpotFleetRequests

Creates a value of DescribeSpotFleetRequests with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsfrSpotFleetRequestIds - The IDs of the Spot fleet requests.

dsfrNextToken - The token for the next set of results.

dsfrDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dsfrMaxResults - The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

data DescribeSpotFleetRequests

Contains the parameters for DescribeSpotFleetRequests.

See: describeSpotFleetRequests smart constructor.

Request Lenses

dsfrSpotFleetRequestIds :: Lens' DescribeSpotFleetRequests [Text]

The IDs of the Spot fleet requests.

dsfrNextToken :: Lens' DescribeSpotFleetRequests (Maybe Text)

The token for the next set of results.

dsfrDryRun :: Lens' DescribeSpotFleetRequests (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dsfrMaxResults :: Lens' DescribeSpotFleetRequests (Maybe Int)

The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

Destructuring the Response

describeSpotFleetRequestsResponse

Arguments

:: Int
dsfrrsResponseStatus

-> DescribeSpotFleetRequestsResponse

Creates a value of DescribeSpotFleetRequestsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsfrrsNextToken - The token required to retrieve the next set of results. This value is null when there are no more results to return.

dsfrrsResponseStatus - -- | The response status code.

dsfrrsSpotFleetRequestConfigs - Information about the configuration of your Spot fleet.

data DescribeSpotFleetRequestsResponse

Contains the output of DescribeSpotFleetRequests.

See: describeSpotFleetRequestsResponse smart constructor.

Response Lenses

dsfrrsNextToken :: Lens' DescribeSpotFleetRequestsResponse (Maybe Text)

The token required to retrieve the next set of results. This value is null when there are no more results to return.

dsfrrsResponseStatus :: Lens' DescribeSpotFleetRequestsResponse Int

- | The response status code.

dsfrrsSpotFleetRequestConfigs :: Lens' DescribeSpotFleetRequestsResponse [SpotFleetRequestConfig]

Information about the configuration of your Spot fleet.