Network.AWS.EC2.DescribeSpotPriceHistory

Description

Describes the Spot Price history. Spot Instances are instances that Amazon EC2 starts on your behalf when the maximum price that you specify exceeds the current Spot Price. Amazon EC2 periodically sets the Spot Price based on available Spot Instance capacity and current Spot Instance requests. For more information about Spot Instances, see Spot Instances in the Amazon Elastic Compute Cloud User Guide. When you specify an Availability Zone, this operation describes the price history for the specified Availability Zone with the most recent set of prices listed first. If you don't specify an Availability Zone, you get the prices across all Availability Zones, starting with the most recent set. However, if you're using an API version earlier than 2011-05-15, you get the lowest price across the region for the specified time period. The prices returned are listed in chronological order, from the oldest to the most recent. When you specify the start and end time options, this operation returns two pieces of data: the prices of the instance types within the time range that you specified and the time when the price changed. The price is valid within the time period that you specified; the response merely indicates the last time that the price changed.

Request

data DescribeSpotPriceHistory

Request constructor

describeSpotPriceHistory :: DescribeSpotPriceHistory

DescribeSpotPriceHistory constructor.

The fields accessible through corresponding lenses are:

dsphAvailabilityZone :: Maybe Text

dsphDryRun :: Maybe Bool

dsphEndTime :: Maybe UTCTime

dsphFilters :: [Filter]

dsphInstanceTypes :: [Text]

dsphMaxResults :: Maybe Int

dsphNextToken :: Maybe Text

dsphProductDescriptions :: [Text]

dsphStartTime :: Maybe UTCTime

Request lenses

dsphAvailabilityZone :: Lens' DescribeSpotPriceHistory (Maybe Text)

The Availability Zone.

dsphDryRun :: Lens' DescribeSpotPriceHistory (Maybe Bool)

dsphEndTime :: Lens' DescribeSpotPriceHistory (Maybe UTCTime)

The end date and time of the Spot Price history data.

dsphFilters :: Lens' DescribeSpotPriceHistory [Filter]

One or more filters. availability-zone - The Availability Zone for which prices should be returned. instance-type - The type of instance (for example, m1.small). product-description - The product description for the Spot Price (LinuxUNIX | SUSE Linux | Windows | LinuxUNIX (Amazon VPC) | SUSE Linux (Amazon VPC) | Windows (Amazon VPC)). spot-price - The Spot Price. The value must match exactly (or use wildcards; greater than or less than comparison is not supported). timestamp - The timestamp of the Spot Price history (for example, 2010-08-16T05:06:11.000Z). You can use wildcards (* and ?). Greater than or less than comparison is not supported.

dsphInstanceTypes :: Lens' DescribeSpotPriceHistory [Text]

One or more instance types.

dsphMaxResults :: Lens' DescribeSpotPriceHistory (Maybe Int)

The number of rows to return.

dsphNextToken :: Lens' DescribeSpotPriceHistory (Maybe Text)

The next set of rows to return.

dsphProductDescriptions :: Lens' DescribeSpotPriceHistory [Text]

One or more basic product descriptions.

dsphStartTime :: Lens' DescribeSpotPriceHistory (Maybe UTCTime)

The start date and time of the Spot Price history data.

Response

data DescribeSpotPriceHistoryResult

Response constructor

describeSpotPriceHistoryResult :: DescribeSpotPriceHistoryResult

DescribeSpotPriceHistoryResult constructor.

The fields accessible through corresponding lenses are:

dsphrNextToken :: Maybe Text

dsphrSpotPriceHistory :: [SpotPrice]

Response lenses

dsphrNextToken :: Lens' DescribeSpotPriceHistoryResult (Maybe Text)

The string marking the next set of results. This is empty if there are no more results.

dsphrSpotPriceHistory :: Lens' DescribeSpotPriceHistoryResult [SpotPrice]

The historical Spot Prices.