Network.AWS.EC2.DescribeTags

Description

Describes one or more of the tags for your EC2 resources. For more information about tags, see Tagging Your Resources in the Amazon Elastic Compute Cloud User Guide.

Request

data DescribeTags

Request constructor

describeTags :: DescribeTags

DescribeTags constructor.

The fields accessible through corresponding lenses are:

dtDryRun :: Maybe Bool

dtFilters :: [Filter]

dtMaxResults :: Maybe Int

dtNextToken :: Maybe Text

Request lenses

dtDryRun :: Lens' DescribeTags (Maybe Bool)

dtFilters :: Lens' DescribeTags [Filter]

One or more filters. key - The tag key. resource-id - The resource ID. resource-type - The resource type (customer-gateway | dhcp-options | image | instance | internet-gateway | network-acl | network-interface | reserved-instances | route-table | security-group | snapshot | spot-instances-request | subnet | volume | vpc | vpn-connection | vpn-gateway). value - The tag value.

dtMaxResults :: Lens' DescribeTags (Maybe Int)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results. If the value is greater than 1000, we return only 1000 items.

dtNextToken :: Lens' DescribeTags (Maybe Text)

The token for the next set of items to return. (You received this token from a prior call.).

Response

data DescribeTagsResult

Response constructor

describeTagsResult :: DescribeTagsResult

DescribeTagsResult constructor.

The fields accessible through corresponding lenses are:

dtrNextToken :: Maybe Text

dtrTags :: [TagDescription]

Response lenses

dtrNextToken :: Lens' DescribeTagsResult (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dtrTags :: Lens' DescribeTagsResult [TagDescription]

A list of tags.