Network.AWS.EC2.DescribeVPCClassicLink

Description

Describes the ClassicLink status of one or more VPCs.

Creating a Request

describeVPCClassicLink :: DescribeVPCClassicLink

Creates a value of DescribeVPCClassicLink with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvclFilters - One or more filters. * is-classic-link-enabled - Whether the VPC is enabled for ClassicLink (true | false ). * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter.

dvclVPCIds - One or more VPCs for which you want to describe the ClassicLink status.

dvclDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data DescribeVPCClassicLink

Contains the parameters for DescribeVpcClassicLink.

See: describeVPCClassicLink smart constructor.

Request Lenses

dvclFilters :: Lens' DescribeVPCClassicLink [Filter]

One or more filters. * is-classic-link-enabled - Whether the VPC is enabled for ClassicLink (true | false ). * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter.

dvclVPCIds :: Lens' DescribeVPCClassicLink [Text]

One or more VPCs for which you want to describe the ClassicLink status.

dvclDryRun :: Lens' DescribeVPCClassicLink (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

describeVPCClassicLinkResponse

Arguments

:: Int
dvclrsResponseStatus

-> DescribeVPCClassicLinkResponse

Creates a value of DescribeVPCClassicLinkResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvclrsVPCs - The ClassicLink status of one or more VPCs.

dvclrsResponseStatus - -- | The response status code.

data DescribeVPCClassicLinkResponse

Contains the output of DescribeVpcClassicLink.

See: describeVPCClassicLinkResponse smart constructor.

Response Lenses

dvclrsVPCs :: Lens' DescribeVPCClassicLinkResponse [VPCClassicLink]

The ClassicLink status of one or more VPCs.

dvclrsResponseStatus :: Lens' DescribeVPCClassicLinkResponse Int

- | The response status code.