Network.AWS.EC2.DescribeVPCClassicLinkDNSSupport

Description

Describes the ClassicLink DNS support status of one or more VPCs. If enabled, the DNS hostname of a linked EC2-Classic instance resolves to its private IP address when addressed from an instance in the VPC to which it's linked. Similarly, the DNS hostname of an instance in a VPC resolves to its private IP address when addressed from a linked EC2-Classic instance. For more information, see ClassicLink in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

describeVPCClassicLinkDNSSupport :: DescribeVPCClassicLinkDNSSupport

Creates a value of DescribeVPCClassicLinkDNSSupport with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvcldsNextToken - The token for the next set of items to return. (You received this token from a prior call.)

dvcldsVPCIds - One or more VPC IDs.

dvcldsMaxResults - The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results.

data DescribeVPCClassicLinkDNSSupport

Contains the parameters for DescribeVpcClassicLinkDnsSupport.

See: describeVPCClassicLinkDNSSupport smart constructor.

Request Lenses

dvcldsNextToken :: Lens' DescribeVPCClassicLinkDNSSupport (Maybe Text)

The token for the next set of items to return. (You received this token from a prior call.)

dvcldsVPCIds :: Lens' DescribeVPCClassicLinkDNSSupport [Text]

One or more VPC IDs.

dvcldsMaxResults :: Lens' DescribeVPCClassicLinkDNSSupport (Maybe Natural)

The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results.

Destructuring the Response

describeVPCClassicLinkDNSSupportResponse

Arguments

:: Int
dvpccldnssrsResponseStatus

-> DescribeVPCClassicLinkDNSSupportResponse

Creates a value of DescribeVPCClassicLinkDNSSupportResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvpccldnssrsVPCs - Information about the ClassicLink DNS support status of the VPCs.

dvpccldnssrsNextToken - The token to use when requesting the next set of items.

dvpccldnssrsResponseStatus - -- | The response status code.

data DescribeVPCClassicLinkDNSSupportResponse

Contains the output of DescribeVpcClassicLinkDnsSupport.

See: describeVPCClassicLinkDNSSupportResponse smart constructor.

Response Lenses

dvpccldnssrsVPCs :: Lens' DescribeVPCClassicLinkDNSSupportResponse [ClassicLinkDNSSupport]

Information about the ClassicLink DNS support status of the VPCs.

dvpccldnssrsNextToken :: Lens' DescribeVPCClassicLinkDNSSupportResponse (Maybe Text)

The token to use when requesting the next set of items.

dvpccldnssrsResponseStatus :: Lens' DescribeVPCClassicLinkDNSSupportResponse Int

- | The response status code.