Network.AWS.EC2.DescribeVPCEndpoints

Description

Describes one or more of your VPC endpoints.

Creating a Request

describeVPCEndpoints :: DescribeVPCEndpoints

Creates a value of DescribeVPCEndpoints with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvpceFilters - One or more filters. * service-name : The name of the AWS service. * vpc-id : The ID of the VPC in which the endpoint resides. * vpc-endpoint-id : The ID of the endpoint. * vpc-endpoint-state : The state of the endpoint. (pending | available | deleting | deleted )

dvpceNextToken - The token for the next set of items to return. (You received this token from a prior call.)

dvpceVPCEndpointIds - One or more endpoint IDs.

dvpceDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dvpceMaxResults - The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results. Constraint: If the value is greater than 1000, we return only 1000 items.

data DescribeVPCEndpoints

Contains the parameters for DescribeVpcEndpoints.

See: describeVPCEndpoints smart constructor.

Request Lenses

dvpceFilters :: Lens' DescribeVPCEndpoints [Filter]

One or more filters. * service-name : The name of the AWS service. * vpc-id : The ID of the VPC in which the endpoint resides. * vpc-endpoint-id : The ID of the endpoint. * vpc-endpoint-state : The state of the endpoint. (pending | available | deleting | deleted )

dvpceNextToken :: Lens' DescribeVPCEndpoints (Maybe Text)

The token for the next set of items to return. (You received this token from a prior call.)

dvpceVPCEndpointIds :: Lens' DescribeVPCEndpoints [Text]

One or more endpoint IDs.

dvpceDryRun :: Lens' DescribeVPCEndpoints (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dvpceMaxResults :: Lens' DescribeVPCEndpoints (Maybe Int)

The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results. Constraint: If the value is greater than 1000, we return only 1000 items.

Destructuring the Response

describeVPCEndpointsResponse

Arguments

:: Int
dvpcersResponseStatus

-> DescribeVPCEndpointsResponse

Creates a value of DescribeVPCEndpointsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvpcersNextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dvpcersVPCEndpoints - Information about the endpoints.

dvpcersResponseStatus - -- | The response status code.

data DescribeVPCEndpointsResponse

Contains the output of DescribeVpcEndpoints.

See: describeVPCEndpointsResponse smart constructor.

Response Lenses

dvpcersNextToken :: Lens' DescribeVPCEndpointsResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dvpcersVPCEndpoints :: Lens' DescribeVPCEndpointsResponse [VPCEndpoint]

Information about the endpoints.

dvpcersResponseStatus :: Lens' DescribeVPCEndpointsResponse Int

- | The response status code.