Network.AWS.EC2.DescribeVPCPeeringConnections

Description

Describes one or more of your VPC peering connections.

Creating a Request

describeVPCPeeringConnections :: DescribeVPCPeeringConnections

Creates a value of DescribeVPCPeeringConnections with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvpcpcFilters - One or more filters. * accepter-vpc-info.cidr-block - The IPv4 CIDR block of the peer VPC. * accepter-vpc-info.owner-id - The AWS account ID of the owner of the peer VPC. * accepter-vpc-info.vpc-id - The ID of the peer VPC. * expiration-time - The expiration date and time for the VPC peering connection. * requester-vpc-info.cidr-block - The IPv4 CIDR block of the requester's VPC. * requester-vpc-info.owner-id - The AWS account ID of the owner of the requester VPC. * requester-vpc-info.vpc-id - The ID of the requester VPC. * status-code - The status of the VPC peering connection (pending-acceptance | failed | expired | provisioning | active | deleted | rejected ). * status-message - A message that provides more information about the status of the VPC peering connection, if applicable. * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter. * vpc-peering-connection-id - The ID of the VPC peering connection.

dvpcpcVPCPeeringConnectionIds - One or more VPC peering connection IDs. Default: Describes all your VPC peering connections.

dvpcpcDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data DescribeVPCPeeringConnections

Contains the parameters for DescribeVpcPeeringConnections.

See: describeVPCPeeringConnections smart constructor.

Request Lenses

dvpcpcFilters :: Lens' DescribeVPCPeeringConnections [Filter]

One or more filters. * accepter-vpc-info.cidr-block - The IPv4 CIDR block of the peer VPC. * accepter-vpc-info.owner-id - The AWS account ID of the owner of the peer VPC. * accepter-vpc-info.vpc-id - The ID of the peer VPC. * expiration-time - The expiration date and time for the VPC peering connection. * requester-vpc-info.cidr-block - The IPv4 CIDR block of the requester's VPC. * requester-vpc-info.owner-id - The AWS account ID of the owner of the requester VPC. * requester-vpc-info.vpc-id - The ID of the requester VPC. * status-code - The status of the VPC peering connection (pending-acceptance | failed | expired | provisioning | active | deleted | rejected ). * status-message - A message that provides more information about the status of the VPC peering connection, if applicable. * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter. * vpc-peering-connection-id - The ID of the VPC peering connection.

dvpcpcVPCPeeringConnectionIds :: Lens' DescribeVPCPeeringConnections [Text]

One or more VPC peering connection IDs. Default: Describes all your VPC peering connections.

dvpcpcDryRun :: Lens' DescribeVPCPeeringConnections (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

describeVPCPeeringConnectionsResponse

Arguments

:: Int
dvpcpcrsResponseStatus

-> DescribeVPCPeeringConnectionsResponse

Creates a value of DescribeVPCPeeringConnectionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvpcpcrsVPCPeeringConnections - Information about the VPC peering connections.

dvpcpcrsResponseStatus - -- | The response status code.

data DescribeVPCPeeringConnectionsResponse

Contains the output of DescribeVpcPeeringConnections.

See: describeVPCPeeringConnectionsResponse smart constructor.

Response Lenses

dvpcpcrsVPCPeeringConnections :: Lens' DescribeVPCPeeringConnectionsResponse [VPCPeeringConnection]

Information about the VPC peering connections.

dvpcpcrsResponseStatus :: Lens' DescribeVPCPeeringConnectionsResponse Int

- | The response status code.