Network.AWS.EC2.DescribeVPCs

Description

Describes one or more of your VPCs.

Creating a Request

describeVPCs :: DescribeVPCs

Creates a value of DescribeVPCs with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvsFilters - One or more filters. * cidr - The IPv4 CIDR block of the VPC. The CIDR block you specify must exactly match the VPC's CIDR block for information to be returned for the VPC. Must contain the slash followed by one or two digits (for example, /28 ). * dhcp-options-id - The ID of a set of DHCP options. * ipv6-cidr-block-association.ipv6-cidr-block - An IPv6 CIDR block associated with the VPC. * ipv6-cidr-block-association.association-id - The association ID for an IPv6 CIDR block associated with the VPC. * ipv6-cidr-block-association.state - The state of an IPv6 CIDR block associated with the VPC. * isDefault - Indicates whether the VPC is the default VPC. * state - The state of the VPC (pending | available ). * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter. * vpc-id - The ID of the VPC.

dvsVPCIds - One or more VPC IDs. Default: Describes all your VPCs.

dvsDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data DescribeVPCs

Contains the parameters for DescribeVpcs.

See: describeVPCs smart constructor.

Request Lenses

dvsFilters :: Lens' DescribeVPCs [Filter]

One or more filters. * cidr - The IPv4 CIDR block of the VPC. The CIDR block you specify must exactly match the VPC's CIDR block for information to be returned for the VPC. Must contain the slash followed by one or two digits (for example, /28 ). * dhcp-options-id - The ID of a set of DHCP options. * ipv6-cidr-block-association.ipv6-cidr-block - An IPv6 CIDR block associated with the VPC. * ipv6-cidr-block-association.association-id - The association ID for an IPv6 CIDR block associated with the VPC. * ipv6-cidr-block-association.state - The state of an IPv6 CIDR block associated with the VPC. * isDefault - Indicates whether the VPC is the default VPC. * state - The state of the VPC (pending | available ). * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter. * vpc-id - The ID of the VPC.

dvsVPCIds :: Lens' DescribeVPCs [Text]

One or more VPC IDs. Default: Describes all your VPCs.

dvsDryRun :: Lens' DescribeVPCs (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

describeVPCsResponse

Arguments

:: Int
dvrsResponseStatus

-> DescribeVPCsResponse

Creates a value of DescribeVPCsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvrsVPCs - Information about one or more VPCs.

dvrsResponseStatus - -- | The response status code.

data DescribeVPCsResponse

Contains the output of DescribeVpcs.

See: describeVPCsResponse smart constructor.

Response Lenses

dvrsVPCs :: Lens' DescribeVPCsResponse [VPC]

Information about one or more VPCs.

dvrsResponseStatus :: Lens' DescribeVPCsResponse Int

- | The response status code.