Network.AWS.EC2.DescribeVPNConnections

Description

Describes one or more of your VPN connections.

For more information about VPN connections, see Adding a Hardware Virtual Private Gateway to Your VPC in the Amazon Virtual Private Cloud User Guide .

Creating a Request

describeVPNConnections :: DescribeVPNConnections

Creates a value of DescribeVPNConnections with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvpncFilters - One or more filters. * customer-gateway-configuration - The configuration information for the customer gateway. * customer-gateway-id - The ID of a customer gateway associated with the VPN connection. * state - The state of the VPN connection (pending | available | deleting | deleted ). * option.static-routes-only - Indicates whether the connection has static routes only. Used for devices that do not support Border Gateway Protocol (BGP). * route.destination-cidr-block - The destination CIDR block. This corresponds to the subnet used in a customer data center. * bgp-asn - The BGP Autonomous System Number (ASN) associated with a BGP device. * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter. * type - The type of VPN connection. Currently the only supported type is ipsec.1 . * vpn-connection-id - The ID of the VPN connection. * vpn-gateway-id - The ID of a virtual private gateway associated with the VPN connection.

dvpncVPNConnectionIds - One or more VPN connection IDs. Default: Describes your VPN connections.

dvpncDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data DescribeVPNConnections

Contains the parameters for DescribeVpnConnections.

See: describeVPNConnections smart constructor.

Request Lenses

dvpncFilters :: Lens' DescribeVPNConnections [Filter]

One or more filters. * customer-gateway-configuration - The configuration information for the customer gateway. * customer-gateway-id - The ID of a customer gateway associated with the VPN connection. * state - The state of the VPN connection (pending | available | deleting | deleted ). * option.static-routes-only - Indicates whether the connection has static routes only. Used for devices that do not support Border Gateway Protocol (BGP). * route.destination-cidr-block - The destination CIDR block. This corresponds to the subnet used in a customer data center. * bgp-asn - The BGP Autonomous System Number (ASN) associated with a BGP device. * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter. * type - The type of VPN connection. Currently the only supported type is ipsec.1 . * vpn-connection-id - The ID of the VPN connection. * vpn-gateway-id - The ID of a virtual private gateway associated with the VPN connection.

dvpncVPNConnectionIds :: Lens' DescribeVPNConnections [Text]

One or more VPN connection IDs. Default: Describes your VPN connections.

dvpncDryRun :: Lens' DescribeVPNConnections (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

describeVPNConnectionsResponse

Arguments

:: Int
dvcrsResponseStatus

-> DescribeVPNConnectionsResponse

Creates a value of DescribeVPNConnectionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvcrsVPNConnections - Information about one or more VPN connections.

dvcrsResponseStatus - -- | The response status code.

data DescribeVPNConnectionsResponse

Contains the output of DescribeVpnConnections.

See: describeVPNConnectionsResponse smart constructor.

Response Lenses

dvcrsVPNConnections :: Lens' DescribeVPNConnectionsResponse [VPNConnection]

Information about one or more VPN connections.

dvcrsResponseStatus :: Lens' DescribeVPNConnectionsResponse Int

- | The response status code.