Network.AWS.EC2.DescribeVPNGateways

Description

Describes one or more of your virtual private gateways.

For more information about virtual private gateways, see Adding an IPsec Hardware VPN to Your VPC in the Amazon Virtual Private Cloud User Guide .

Creating a Request

describeVPNGateways :: DescribeVPNGateways

Creates a value of DescribeVPNGateways with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvgsFilters - One or more filters. * attachment.state - The current state of the attachment between the gateway and the VPC (attaching | attached | detaching | detached ). * attachment.vpc-id - The ID of an attached VPC. * availability-zone - The Availability Zone for the virtual private gateway (if applicable). * state - The state of the virtual private gateway (pending | available | deleting | deleted ). * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter. * type - The type of virtual private gateway. Currently the only supported type is ipsec.1 . * vpn-gateway-id - The ID of the virtual private gateway.

dvgsVPNGatewayIds - One or more virtual private gateway IDs. Default: Describes all your virtual private gateways.

dvgsDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data DescribeVPNGateways

Contains the parameters for DescribeVpnGateways.

See: describeVPNGateways smart constructor.

Request Lenses

dvgsFilters :: Lens' DescribeVPNGateways [Filter]

One or more filters. * attachment.state - The current state of the attachment between the gateway and the VPC (attaching | attached | detaching | detached ). * attachment.vpc-id - The ID of an attached VPC. * availability-zone - The Availability Zone for the virtual private gateway (if applicable). * state - The state of the virtual private gateway (pending | available | deleting | deleted ). * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter. * type - The type of virtual private gateway. Currently the only supported type is ipsec.1 . * vpn-gateway-id - The ID of the virtual private gateway.

dvgsVPNGatewayIds :: Lens' DescribeVPNGateways [Text]

One or more virtual private gateway IDs. Default: Describes all your virtual private gateways.

dvgsDryRun :: Lens' DescribeVPNGateways (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

describeVPNGatewaysResponse

Arguments

:: Int
dvgrsResponseStatus

-> DescribeVPNGatewaysResponse

Creates a value of DescribeVPNGatewaysResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvgrsVPNGateways - Information about one or more virtual private gateways.

dvgrsResponseStatus - -- | The response status code.

data DescribeVPNGatewaysResponse

Contains the output of DescribeVpnGateways.

See: describeVPNGatewaysResponse smart constructor.

Response Lenses

dvgrsVPNGateways :: Lens' DescribeVPNGatewaysResponse [VPNGateway]

Information about one or more virtual private gateways.

dvgrsResponseStatus :: Lens' DescribeVPNGatewaysResponse Int

- | The response status code.