Network.AWS.EC2.DescribeVolumeAttribute

Description

Describes the specified attribute of the specified volume. You can specify only one attribute at a time.

For more information about Amazon EBS volumes, see Amazon EBS Volumes in the Amazon Elastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeVolumeAttribute.html

Request

data DescribeVolumeAttribute

Request constructor

describeVolumeAttribute

Arguments

:: Text
dvaVolumeId

-> DescribeVolumeAttribute

DescribeVolumeAttribute constructor.

The fields accessible through corresponding lenses are:

dvaAttribute :: Maybe VolumeAttributeName

dvaDryRun :: Maybe Bool

dvaVolumeId :: Text

Request lenses

dvaAttribute :: Lens' DescribeVolumeAttribute (Maybe VolumeAttributeName)

The instance attribute.

dvaDryRun :: Lens' DescribeVolumeAttribute (Maybe Bool)

dvaVolumeId :: Lens' DescribeVolumeAttribute Text

The ID of the volume.

Response

data DescribeVolumeAttributeResponse

Response constructor

describeVolumeAttributeResponse :: DescribeVolumeAttributeResponse

DescribeVolumeAttributeResponse constructor.

The fields accessible through corresponding lenses are:

dvarAutoEnableIO :: Maybe AttributeBooleanValue

dvarProductCodes :: [ProductCode]

dvarVolumeId :: Maybe Text

Response lenses

dvarAutoEnableIO :: Lens' DescribeVolumeAttributeResponse (Maybe AttributeBooleanValue)

The state of autoEnableIO attribute.

dvarProductCodes :: Lens' DescribeVolumeAttributeResponse [ProductCode]

A list of product codes.

dvarVolumeId :: Lens' DescribeVolumeAttributeResponse (Maybe Text)

The ID of the volume.