Network.AWS.EC2.DescribeVpcAttribute

Description

Describes the specified attribute of the specified VPC. You can specify only one attribute at a time. Example 1 This example describes the enableDnsSupport attribute of the specified VPC. The sample response indicates that DNS resolution is supported. https:/ec2.amazonaws.com?Action=DescribeVpcAttribute &amp;VpcId=vpc-1a2b3c4d &amp;Attribute=enableDnsSupport &amp;AUTHPARAMS &lt;DescribeVpcAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;vpcId&gt;vpc-1a2b3c4d&lt;/vpcId&gt; &lt;enableDnsSupport&gt; &lt;value&gt;true&lt;value&gt; &lt;enableDnsSupport&gt; &lt;/DescribeVpcAttributeResponse&gt; Example 2 This request describes the enableDnsHostnames attribute of the specified VPC. The sample response indicates that DNS hostnames are supported. https:/ec2.amazonaws.com?Action=DescribeVpcAttribute &amp;VpcId=vpc-1a2b3c4d &amp;Attribute=enableDnsHostnames &amp;AUTHPARAMS &lt;DescribeVpcAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;vpcId&gt;vpc-1a2b3c4d&lt;/vpcId&gt; &lt;enableDnsHostnames&gt; &lt;value&gt;true&lt;value&gt; &lt;enableDnsHostnames&gt; &lt;/DescribeVpcAttributeResponse&gt;.

Request

data DescribeVpcAttribute

Request constructor

mkDescribeVpcAttribute

Arguments

:: Text
dva1VpcId

-> DescribeVpcAttribute

Smart constructor for the minimum required parameters to construct a valid DescribeVpcAttribute request.

The fields accessible through corresponding lenses are:

VpcId :: Text

Attribute :: Maybe VpcAttributeName

Request lenses

dva1VpcId :: Lens' DescribeVpcAttribute Text

The ID of the VPC.

dva1Attribute :: Lens' DescribeVpcAttribute (Maybe VpcAttributeName)

The VPC attribute.

Response

data DescribeVpcAttributeResponse

Response constructor

mkDescribeVpcAttributeResponse :: DescribeVpcAttributeResponse

Smart constructor for the minimum required parameters to construct a valid DescribeVpcAttributeResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

VpcId :: Maybe Text

EnableDnsSupport :: Maybe AttributeBooleanValue

EnableDnsHostnames :: Maybe AttributeBooleanValue

Response lenses

dvarrVpcId :: Lens' DescribeVpcAttributeResponse (Maybe Text)

The ID of the VPC.

dvarrEnableDnsSupport :: Lens' DescribeVpcAttributeResponse (Maybe AttributeBooleanValue)

Indicates whether DNS resolution is enabled for the VPC. If this attribute is true, the Amazon DNS server resolves DNS hostnames for your instances to their corresponding IP addresses; otherwise, it does not.

dvarrEnableDnsHostnames :: Lens' DescribeVpcAttributeResponse (Maybe AttributeBooleanValue)

Indicates whether the instances launched in the VPC get DNS hostnames. If this attribute is true, instances in the VPC get DNS hostnames; otherwise, they do not.