Network.AWS.EC2.DescribeVpcs

Description

Describes one or more of your VPCs.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeVpcs.html

Request

data DescribeVpcs

Request constructor

describeVpcs :: DescribeVpcs

DescribeVpcs constructor.

The fields accessible through corresponding lenses are:

dv1DryRun :: Maybe Bool

dv1Filters :: [Filter]

dv1VpcIds :: [Text]

Request lenses

dv1DryRun :: Lens' DescribeVpcs (Maybe Bool)

dv1Filters :: Lens' DescribeVpcs [Filter]

One or more filters.

cidr - The CIDR block of the VPC. The CIDR block you specify must exactly match the VPC's CIDR block for information to be returned for the VPC. Must contain the slash followed by one or two digits (for example, '/28').

'dhcp-options-id' - The ID of a set of DHCP options.

isDefault - Indicates whether the VPC is the default VPC.

state - The state of the VPC (pending | available).

tag:key=value - The key/value combination of a tag assigned to the resource.

'tag-key' - The key of a tag assigned to the resource. This filter is independent of the 'tag-value' filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag:key=value filter.

'tag-value' - The value of a tag assigned to the resource. This filter is independent of the 'tag-key' filter.

'vpc-id' - The ID of the VPC.

dv1VpcIds :: Lens' DescribeVpcs [Text]

One or more VPC IDs.

Default: Describes all your VPCs.

Response

data DescribeVpcsResponse

Response constructor

describeVpcsResponse :: DescribeVpcsResponse

DescribeVpcsResponse constructor.

The fields accessible through corresponding lenses are:

dvrVpcs :: [Vpc]

Response lenses

dvrVpcs :: Lens' DescribeVpcsResponse [Vpc]

Information about one or more VPCs.