Network.AWS.EC2.DescribeVpnConnections

Description

Describes one or more of your VPN connections.

For more information about VPN connections, see Adding a Hardware VirtualPrivate Gateway to Your VPC in the Amazon Virtual Private Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeVpnConnections.html

Request

data DescribeVpnConnections

Request constructor

describeVpnConnections :: DescribeVpnConnections

DescribeVpnConnections constructor.

The fields accessible through corresponding lenses are:

dvc1DryRun :: Maybe Bool

dvc1Filters :: [Filter]

dvc1VpnConnectionIds :: [Text]

Request lenses

dvc1DryRun :: Lens' DescribeVpnConnections (Maybe Bool)

dvc1Filters :: Lens' DescribeVpnConnections [Filter]

One or more filters.

'customer-gateway-configuration' - The configuration information for the customer gateway.

'customer-gateway-id' - The ID of a customer gateway associated with the VPN connection.

state - The state of the VPN connection (pending | available | deleting | deleted).

'option.static-routes-only' - Indicates whether the connection has static routes only. Used for devices that do not support Border Gateway Protocol (BGP).

'route.destination-cidr-block' - The destination CIDR block. This corresponds to the subnet used in a customer data center.

'bgp-asn' - The BGP Autonomous System Number (ASN) associated with a BGP device.

tag:key=value - The key/value combination of a tag assigned to the resource.

'tag-key' - The key of a tag assigned to the resource. This filter is independent of the 'tag-value' filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag:key=value filter.

'tag-value' - The value of a tag assigned to the resource. This filter is independent of the 'tag-key' filter.

'type' - The type of VPN connection. Currently the only supported type is 'ipsec.1'.

'vpn-connection-id' - The ID of the VPN connection.

'vpn-gateway-id' - The ID of a virtual private gateway associated with the VPN connection.

dvc1VpnConnectionIds :: Lens' DescribeVpnConnections [Text]

One or more VPN connection IDs.

Default: Describes your VPN connections.

Response

data DescribeVpnConnectionsResponse

Response constructor

describeVpnConnectionsResponse :: DescribeVpnConnectionsResponse

DescribeVpnConnectionsResponse constructor.

The fields accessible through corresponding lenses are:

dvcrVpnConnections :: [VpnConnection]

Response lenses

dvcrVpnConnections :: Lens' DescribeVpnConnectionsResponse [VpnConnection]

Information about one or more VPN connections.