Network.AWS.EC2.DetachClassicLinkVPC

Description

Unlinks (detaches) a linked EC2-Classic instance from a VPC. After the instance has been unlinked, the VPC security groups are no longer associated with it. An instance is automatically unlinked from a VPC when it's stopped.

Creating a Request

detachClassicLinkVPC

Arguments

:: Text
dclvInstanceId

-> Text
dclvVPCId

-> DetachClassicLinkVPC

Creates a value of DetachClassicLinkVPC with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dclvDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dclvInstanceId - The ID of the instance to unlink from the VPC.

dclvVPCId - The ID of the VPC to which the instance is linked.

data DetachClassicLinkVPC

Contains the parameters for DetachClassicLinkVpc.

See: detachClassicLinkVPC smart constructor.

Request Lenses

dclvDryRun :: Lens' DetachClassicLinkVPC (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dclvInstanceId :: Lens' DetachClassicLinkVPC Text

The ID of the instance to unlink from the VPC.

dclvVPCId :: Lens' DetachClassicLinkVPC Text

The ID of the VPC to which the instance is linked.

Destructuring the Response

detachClassicLinkVPCResponse

Arguments

:: Int
dclvrsResponseStatus

-> DetachClassicLinkVPCResponse

Creates a value of DetachClassicLinkVPCResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dclvrsReturn - Returns true if the request succeeds; otherwise, it returns an error.

dclvrsResponseStatus - -- | The response status code.

data DetachClassicLinkVPCResponse

Contains the output of DetachClassicLinkVpc.

See: detachClassicLinkVPCResponse smart constructor.

Response Lenses

dclvrsReturn :: Lens' DetachClassicLinkVPCResponse (Maybe Bool)

Returns true if the request succeeds; otherwise, it returns an error.

dclvrsResponseStatus :: Lens' DetachClassicLinkVPCResponse Int

- | The response status code.