Network.AWS.EC2.DetachInternetGateway

Description

Detaches an Internet gateway from a VPC, disabling connectivity between the Internet and the VPC. The VPC must not contain any running instances with Elastic IP addresses.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DetachInternetGateway.html

Request

data DetachInternetGateway

Request constructor

detachInternetGateway

Arguments

:: Text
digInternetGatewayId

-> Text
digVpcId

-> DetachInternetGateway

DetachInternetGateway constructor.

The fields accessible through corresponding lenses are:

digDryRun :: Maybe Bool

digInternetGatewayId :: Text

digVpcId :: Text

Request lenses

digDryRun :: Lens' DetachInternetGateway (Maybe Bool)

digInternetGatewayId :: Lens' DetachInternetGateway Text

The ID of the Internet gateway.

digVpcId :: Lens' DetachInternetGateway Text

The ID of the VPC.

Response

data DetachInternetGatewayResponse

Response constructor

detachInternetGatewayResponse :: DetachInternetGatewayResponse

DetachInternetGatewayResponse constructor.