Network.AWS.EC2.DetachNetworkInterface

Description

Detaches a network interface from an instance. Example This example detaches the specified elastic network interface (ENI). https:/ec2.amazonaws.com?Action=DetachNetworkInterface &amp;AttachmentId=eni-attach-d94b09b0 &amp;AUTHPARAMS &lt;DetachNetworkInterfaceResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;ce540707-0635-46bc-97da-33a8a362a0e8&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;DetachNetworkInterfaceResponse&gt;.

Request

data DetachNetworkInterface

Request constructor

detachNetworkInterface

Arguments

:: Text
dni2AttachmentId

-> DetachNetworkInterface

Smart constructor for the minimum required parameters to construct a valid DetachNetworkInterface request.

The fields accessible through corresponding lenses are:

AttachmentId :: Text

Force :: Maybe Bool

Request lenses

dni2AttachmentId :: Lens' DetachNetworkInterface Text

The ID of the attachment.

dni2Force :: Lens' DetachNetworkInterface (Maybe Bool)

Specifies whether to force a detachment.

Response

data DetachNetworkInterfaceResponse

Response constructor

detachNetworkInterfaceResponse :: DetachNetworkInterfaceResponse

Smart constructor for the minimum required parameters to construct a valid DetachNetworkInterfaceResponse response.

This constructor is provided for convenience and testing purposes.