Network.AWS.EC2.DetachVolume

Description

Detaches an Amazon EBS volume from an instance. Make sure to unmount any file systems on the device within your operating system before detaching the volume. Failure to do so results in the volume being stuck in a busy state while detaching.

If an Amazon EBS volume is the root device of an instance, it can't be detached while the instance is running. To detach the root volume, stop the instance first.

When a volume with an AWS Marketplace product code is detached from an instance, the product code is no longer associated with the instance.

For more information, see Detaching an Amazon EBS Volume in the AmazonElastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DetachVolume.html

Request

data DetachVolume

Request constructor

detachVolume

Arguments

:: Text
dvVolumeId

-> DetachVolume

DetachVolume constructor.

The fields accessible through corresponding lenses are:

dvDevice :: Maybe Text

dvDryRun :: Maybe Bool

dvForce :: Maybe Bool

dvInstanceId :: Maybe Text

dvVolumeId :: Text

Request lenses

dvDevice :: Lens' DetachVolume (Maybe Text)

The device name.

dvDryRun :: Lens' DetachVolume (Maybe Bool)

dvForce :: Lens' DetachVolume (Maybe Bool)

Forces detachment if the previous detachment attempt did not occur cleanly (for example, logging into an instance, unmounting the volume, and detaching normally). This option can lead to data loss or a corrupted file system. Use this option only as a last resort to detach a volume from a failed instance. The instance won't have an opportunity to flush file system caches or file system metadata. If you use this option, you must perform file system check and repair procedures.

dvInstanceId :: Lens' DetachVolume (Maybe Text)

The ID of the instance.

dvVolumeId :: Lens' DetachVolume Text

The ID of the volume.

Response

data DetachVolumeResponse

Response constructor

detachVolumeResponse :: DetachVolumeResponse

DetachVolumeResponse constructor.

The fields accessible through corresponding lenses are:

dvrAttachTime :: Maybe UTCTime

dvrDeleteOnTermination :: Maybe Bool

dvrDevice :: Maybe Text

dvrInstanceId :: Maybe Text

dvrState :: Maybe VolumeAttachmentState

dvrVolumeId :: Maybe Text

Response lenses

dvrAttachTime :: Lens' DetachVolumeResponse (Maybe UTCTime)

The time stamp when the attachment initiated.

dvrDeleteOnTermination :: Lens' DetachVolumeResponse (Maybe Bool)

Indicates whether the Amazon EBS volume is deleted on instance termination.

dvrDevice :: Lens' DetachVolumeResponse (Maybe Text)

The device name.

dvrInstanceId :: Lens' DetachVolumeResponse (Maybe Text)

The ID of the instance.

dvrState :: Lens' DetachVolumeResponse (Maybe VolumeAttachmentState)

The attachment state of the volume.

dvrVolumeId :: Lens' DetachVolumeResponse (Maybe Text)

The ID of the volume.