Network.AWS.EC2.DisableVPCClassicLink

Description

Disables ClassicLink for a VPC. You cannot disable ClassicLink for a VPC that has EC2-Classic instances linked to it.

Creating a Request

disableVPCClassicLink

Arguments

:: Text
dvpcclVPCId

-> DisableVPCClassicLink

Creates a value of DisableVPCClassicLink with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvpcclDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dvpcclVPCId - The ID of the VPC.

data DisableVPCClassicLink

Contains the parameters for DisableVpcClassicLink.

See: disableVPCClassicLink smart constructor.

Request Lenses

dvpcclDryRun :: Lens' DisableVPCClassicLink (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dvpcclVPCId :: Lens' DisableVPCClassicLink Text

The ID of the VPC.

Destructuring the Response

disableVPCClassicLinkResponse

Arguments

:: Int
dvpcclrsResponseStatus

-> DisableVPCClassicLinkResponse

Creates a value of DisableVPCClassicLinkResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvpcclrsReturn - Returns true if the request succeeds; otherwise, it returns an error.

dvpcclrsResponseStatus - -- | The response status code.

data DisableVPCClassicLinkResponse

Contains the output of DisableVpcClassicLink.

See: disableVPCClassicLinkResponse smart constructor.

Response Lenses

dvpcclrsReturn :: Lens' DisableVPCClassicLinkResponse (Maybe Bool)

Returns true if the request succeeds; otherwise, it returns an error.

dvpcclrsResponseStatus :: Lens' DisableVPCClassicLinkResponse Int

- | The response status code.