Network.AWS.EC2.DisableVPCClassicLinkDNSSupport

Description

Disables ClassicLink DNS support for a VPC. If disabled, DNS hostnames resolve to public IP addresses when addressed between a linked EC2-Classic instance and instances in the VPC to which it's linked. For more information about ClassicLink, see ClassicLink in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

disableVPCClassicLinkDNSSupport :: DisableVPCClassicLinkDNSSupport

Creates a value of DisableVPCClassicLinkDNSSupport with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvcldsVPCId - The ID of the VPC.

data DisableVPCClassicLinkDNSSupport

Contains the parameters for DisableVpcClassicLinkDnsSupport.

See: disableVPCClassicLinkDNSSupport smart constructor.

Request Lenses

dvcldsVPCId :: Lens' DisableVPCClassicLinkDNSSupport (Maybe Text)

The ID of the VPC.

Destructuring the Response

disableVPCClassicLinkDNSSupportResponse

Arguments

:: Int
dvcldsrsResponseStatus

-> DisableVPCClassicLinkDNSSupportResponse

Creates a value of DisableVPCClassicLinkDNSSupportResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvcldsrsReturn - Returns true if the request succeeds; otherwise, it returns an error.

dvcldsrsResponseStatus - -- | The response status code.

data DisableVPCClassicLinkDNSSupportResponse

Contains the output of DisableVpcClassicLinkDnsSupport.

See: disableVPCClassicLinkDNSSupportResponse smart constructor.

Response Lenses

dvcldsrsReturn :: Lens' DisableVPCClassicLinkDNSSupportResponse (Maybe Bool)

Returns true if the request succeeds; otherwise, it returns an error.

dvcldsrsResponseStatus :: Lens' DisableVPCClassicLinkDNSSupportResponse Int

- | The response status code.