Network.AWS.EC2.DisassociateAddress

Description

Disassociates an Elastic IP address from the instance or network interface it's associated with. This is an idempotent operation. If you perform the operation more than once, Amazon EC2 doesn't return an error. Example for EC2-Classic This example disassociates the specified Elastic IP address from the instance in EC2-Classic to which it is associated. https:/ec2.amazonaws.com?Action=DisassociateAddress &amp;PublicIp=192.0.2.1 &amp;AUTHPARAMS Example for EC2-VPC This example disassociates the specified Elastic IP address from the instance in a VPC to which it is associated. https:/ec2.amazonaws.com?Action=DisassociateAddress &amp;AssociationId=eipassoc-aa7486c3 &amp;AUTHPARAMS.

Request

data DisassociateAddress

Request constructor

disassociateAddress :: DisassociateAddress

Smart constructor for the minimum required parameters to construct a valid DisassociateAddress request.

The fields accessible through corresponding lenses are:

PublicIp :: Maybe Text

AssociationId :: Maybe Text

Request lenses

da1PublicIp :: Lens' DisassociateAddress (Maybe Text)

EC2-Classic

The Elastic IP address.

da1AssociationId :: Lens' DisassociateAddress (Maybe Text)

EC2-VPC

The association ID.

Response

data DisassociateAddressResponse

Response constructor

disassociateAddressResponse :: DisassociateAddressResponse

Smart constructor for the minimum required parameters to construct a valid DisassociateAddressResponse response.

This constructor is provided for convenience and testing purposes.