Network.AWS.EC2.DisassociateIAMInstanceProfile

Description

Disassociates an IAM instance profile from a running or stopped instance.

Use DescribeIamInstanceProfileAssociations to get the association ID.

Creating a Request

disassociateIAMInstanceProfile

Arguments

:: Text
diapAssociationId

-> DisassociateIAMInstanceProfile

Creates a value of DisassociateIAMInstanceProfile with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diapAssociationId - The ID of the IAM instance profile association.

data DisassociateIAMInstanceProfile

See: disassociateIAMInstanceProfile smart constructor.

Request Lenses

diapAssociationId :: Lens' DisassociateIAMInstanceProfile Text

The ID of the IAM instance profile association.

Destructuring the Response

disassociateIAMInstanceProfileResponse

Arguments

:: Int
diaprsResponseStatus

-> DisassociateIAMInstanceProfileResponse

Creates a value of DisassociateIAMInstanceProfileResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diaprsIAMInstanceProfileAssociation - Information about the IAM instance profile association.

diaprsResponseStatus - -- | The response status code.

data DisassociateIAMInstanceProfileResponse

See: disassociateIAMInstanceProfileResponse smart constructor.

Response Lenses

diaprsIAMInstanceProfileAssociation :: Lens' DisassociateIAMInstanceProfileResponse (Maybe IAMInstanceProfileAssociation)

Information about the IAM instance profile association.

diaprsResponseStatus :: Lens' DisassociateIAMInstanceProfileResponse Int

- | The response status code.