Network.AWS.EC2.DisassociateRouteTable

Description

Disassociates a subnet from a route table. After you perform this action, the subnet no longer uses the routes in the route table. Instead, it uses the routes in the VPC's main route table. For more information about route tables, see Route Tables in the Amazon Virtual Private Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DisassociateRouteTable.html

Request

data DisassociateRouteTable

Request constructor

disassociateRouteTable

Arguments

:: Text
drtAssociationId

-> DisassociateRouteTable

DisassociateRouteTable constructor.

The fields accessible through corresponding lenses are:

drtAssociationId :: Text

drtDryRun :: Maybe Bool

Request lenses

drtAssociationId :: Lens' DisassociateRouteTable Text

The association ID representing the current association between the route table and subnet.

drtDryRun :: Lens' DisassociateRouteTable (Maybe Bool)

Response

data DisassociateRouteTableResponse

Response constructor

disassociateRouteTableResponse :: DisassociateRouteTableResponse

DisassociateRouteTableResponse constructor.