Network.AWS.EC2.DisassociateSubnetCidrBlock

Description

Disassociates a CIDR block from a subnet. Currently, you can disassociate an IPv6 CIDR block only. You must detach or delete all gateways and resources that are associated with the CIDR block before you can disassociate it.

Creating a Request

disassociateSubnetCidrBlock

Arguments

:: Text
dscbAssociationId

-> DisassociateSubnetCidrBlock

Creates a value of DisassociateSubnetCidrBlock with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dscbAssociationId - The association ID for the CIDR block.

data DisassociateSubnetCidrBlock

See: disassociateSubnetCidrBlock smart constructor.

Request Lenses

dscbAssociationId :: Lens' DisassociateSubnetCidrBlock Text

The association ID for the CIDR block.

Destructuring the Response

disassociateSubnetCidrBlockResponse

Arguments

:: Int
dscbrsResponseStatus

-> DisassociateSubnetCidrBlockResponse

Creates a value of DisassociateSubnetCidrBlockResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dscbrsSubnetId - The ID of the subnet.

dscbrsIPv6CidrBlockAssociation - Information about the IPv6 CIDR block association.

dscbrsResponseStatus - -- | The response status code.

data DisassociateSubnetCidrBlockResponse

See: disassociateSubnetCidrBlockResponse smart constructor.

Response Lenses

dscbrsSubnetId :: Lens' DisassociateSubnetCidrBlockResponse (Maybe Text)

The ID of the subnet.

dscbrsIPv6CidrBlockAssociation :: Lens' DisassociateSubnetCidrBlockResponse (Maybe SubnetIPv6CidrBlockAssociation)

Information about the IPv6 CIDR block association.

dscbrsResponseStatus :: Lens' DisassociateSubnetCidrBlockResponse Int

- | The response status code.