Network.AWS.EC2.DisassociateVPCCidrBlock

Description

Disassociates a CIDR block from a VPC. Currently, you can disassociate an IPv6 CIDR block only. You must detach or delete all gateways and resources that are associated with the CIDR block before you can disassociate it.

Creating a Request

disassociateVPCCidrBlock

Arguments

:: Text
dvcbAssociationId

-> DisassociateVPCCidrBlock

Creates a value of DisassociateVPCCidrBlock with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvcbAssociationId - The association ID for the CIDR block.

data DisassociateVPCCidrBlock

See: disassociateVPCCidrBlock smart constructor.

Request Lenses

dvcbAssociationId :: Lens' DisassociateVPCCidrBlock Text

The association ID for the CIDR block.

Destructuring the Response

disassociateVPCCidrBlockResponse

Arguments

:: Int
dvcbrsResponseStatus

-> DisassociateVPCCidrBlockResponse

Creates a value of DisassociateVPCCidrBlockResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvcbrsVPCId - The ID of the VPC.

dvcbrsIPv6CidrBlockAssociation - Information about the IPv6 CIDR block association.

dvcbrsResponseStatus - -- | The response status code.

data DisassociateVPCCidrBlockResponse

See: disassociateVPCCidrBlockResponse smart constructor.

Response Lenses

dvcbrsVPCId :: Lens' DisassociateVPCCidrBlockResponse (Maybe Text)

The ID of the VPC.

dvcbrsIPv6CidrBlockAssociation :: Lens' DisassociateVPCCidrBlockResponse (Maybe VPCIPv6CidrBlockAssociation)

Information about the IPv6 CIDR block association.

dvcbrsResponseStatus :: Lens' DisassociateVPCCidrBlockResponse Int

- | The response status code.