Network.AWS.EC2.EnableVGWRoutePropagation

Description

Enables a virtual private gateway (VGW) to propagate routes to the specified route table of a VPC.

Creating a Request

enableVGWRoutePropagation

Arguments

:: Text
evrpGatewayId

-> Text
evrpRouteTableId

-> EnableVGWRoutePropagation

Creates a value of EnableVGWRoutePropagation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

evrpGatewayId - The ID of the virtual private gateway.

evrpRouteTableId - The ID of the route table.

data EnableVGWRoutePropagation

Contains the parameters for EnableVgwRoutePropagation.

See: enableVGWRoutePropagation smart constructor.

Request Lenses

evrpGatewayId :: Lens' EnableVGWRoutePropagation Text

The ID of the virtual private gateway.

evrpRouteTableId :: Lens' EnableVGWRoutePropagation Text

The ID of the route table.

Destructuring the Response

enableVGWRoutePropagationResponse :: EnableVGWRoutePropagationResponse

Creates a value of EnableVGWRoutePropagationResponse with the minimum fields required to make a request.

data EnableVGWRoutePropagationResponse

See: enableVGWRoutePropagationResponse smart constructor.