Network.AWS.EC2.EnableVPCClassicLink

Description

Enables a VPC for ClassicLink. You can then link EC2-Classic instances to your ClassicLink-enabled VPC to allow communication over private IP addresses. You cannot enable your VPC for ClassicLink if any of your VPC's route tables have existing routes for address ranges within the 10.0.0.0/8 IP address range, excluding local routes for VPCs in the 10.0.0.0/16 and 10.1.0.0/16 IP address ranges. For more information, see ClassicLink in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

enableVPCClassicLink

Arguments

:: Text
evclVPCId

-> EnableVPCClassicLink

Creates a value of EnableVPCClassicLink with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

evclDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

evclVPCId - The ID of the VPC.

data EnableVPCClassicLink

Contains the parameters for EnableVpcClassicLink.

See: enableVPCClassicLink smart constructor.

Request Lenses

evclDryRun :: Lens' EnableVPCClassicLink (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

evclVPCId :: Lens' EnableVPCClassicLink Text

The ID of the VPC.

Destructuring the Response

enableVPCClassicLinkResponse

Arguments

:: Int
evclrsResponseStatus

-> EnableVPCClassicLinkResponse

Creates a value of EnableVPCClassicLinkResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

evclrsReturn - Returns true if the request succeeds; otherwise, it returns an error.

evclrsResponseStatus - -- | The response status code.

data EnableVPCClassicLinkResponse

Contains the output of EnableVpcClassicLink.

See: enableVPCClassicLinkResponse smart constructor.

Response Lenses

evclrsReturn :: Lens' EnableVPCClassicLinkResponse (Maybe Bool)

Returns true if the request succeeds; otherwise, it returns an error.

evclrsResponseStatus :: Lens' EnableVPCClassicLinkResponse Int

- | The response status code.