Network.AWS.EC2.EnableVolumeIO

Description

Enables IO operations for a volume that had IO operations disabled because the data on the volume was potentially inconsistent.

Creating a Request

enableVolumeIO

Arguments

:: Text
evioVolumeId

-> EnableVolumeIO

Creates a value of EnableVolumeIO with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

evioDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

evioVolumeId - The ID of the volume.

data EnableVolumeIO

Contains the parameters for EnableVolumeIO.

See: enableVolumeIO smart constructor.

Request Lenses

evioDryRun :: Lens' EnableVolumeIO (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

evioVolumeId :: Lens' EnableVolumeIO Text

The ID of the volume.

Destructuring the Response

enableVolumeIOResponse :: EnableVolumeIOResponse

Creates a value of EnableVolumeIOResponse with the minimum fields required to make a request.

data EnableVolumeIOResponse

See: enableVolumeIOResponse smart constructor.