Network.AWS.EC2.GetConsoleScreenshot

Description

Retrieve a JPG-format screenshot of a running instance to help with troubleshooting.

The returned content is Base64-encoded.

Creating a Request

getConsoleScreenshot

Arguments

:: Text
gcsInstanceId

-> GetConsoleScreenshot

Creates a value of GetConsoleScreenshot with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcsWakeUp - When set to true , acts as keystroke input and wakes up an instance that's in standby or "sleep" mode.

gcsDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

gcsInstanceId - The ID of the instance.

data GetConsoleScreenshot

Contains the parameters for the request.

See: getConsoleScreenshot smart constructor.

Request Lenses

gcsWakeUp :: Lens' GetConsoleScreenshot (Maybe Bool)

When set to true , acts as keystroke input and wakes up an instance that's in standby or "sleep" mode.

gcsDryRun :: Lens' GetConsoleScreenshot (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

gcsInstanceId :: Lens' GetConsoleScreenshot Text

The ID of the instance.

Destructuring the Response

getConsoleScreenshotResponse

Arguments

:: Int
gcsrsResponseStatus

-> GetConsoleScreenshotResponse

Creates a value of GetConsoleScreenshotResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcsrsInstanceId - The ID of the instance.

gcsrsImageData - The data that comprises the image.

gcsrsResponseStatus - -- | The response status code.

data GetConsoleScreenshotResponse

Contains the output of the request.

See: getConsoleScreenshotResponse smart constructor.

Response Lenses

gcsrsInstanceId :: Lens' GetConsoleScreenshotResponse (Maybe Text)

The ID of the instance.

gcsrsImageData :: Lens' GetConsoleScreenshotResponse (Maybe Text)

The data that comprises the image.

gcsrsResponseStatus :: Lens' GetConsoleScreenshotResponse Int

- | The response status code.