Network.AWS.EC2.GetPasswordData

Description

Retrieves the encrypted administrator password for an instance running Windows. The Windows password is only generated the first time an AMI is launched. It is not generated for rebundled AMIs or after the password is changed on an instance. The password is encrypted using the key pair that you specified when you launched the instance. You must provide the corresponding key pair file. Password generation and encryption takes a few moments. We recommend that you wait up to 15 minutes after launching an instance before trying to retrieve the generated password. Example This example returns the encrypted version of the administrator password for the specified instance. https:/ec2.amazonaws.com?Action=GetPasswordData &amp;InstanceId=i-10a64379 &amp;AUTHPARAMS &lt;GetPasswordDataResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;instanceId&gt;i-2574e22a&lt;/instanceId&gt; &lt;timestamp&gt;2009-10-24 15:00:00&lt;/timestamp&gt; &lt;passwordData&gt;TGludXggdmVyc2lvbiAyLjYuMTYteGVuVSAoYnVpbGRlckBwYXRjaGJhdC5hbWF6b25zYSkgKGdj&lt;/passwordData&gt; &lt;/GetPasswordDataResponse&gt;.

Request

data GetPasswordData

Request constructor

getPasswordData

Arguments

:: Text
gpdInstanceId

-> GetPasswordData

Smart constructor for the minimum required parameters to construct a valid GetPasswordData request.

The fields accessible through corresponding lenses are:

InstanceId :: Text

Request lenses

gpdInstanceId :: Lens' GetPasswordData Text

The ID of the Windows instance.

Response

data GetPasswordDataResponse

Response constructor

getPasswordDataResponse :: GetPasswordDataResponse

Smart constructor for the minimum required parameters to construct a valid GetPasswordDataResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

InstanceId :: Maybe Text

Timestamp :: Maybe ISO8601

PasswordData :: Maybe Text

Response lenses

gpdrInstanceId :: Lens' GetPasswordDataResponse (Maybe Text)

The ID of the Windows instance.

gpdrTimestamp :: Lens' GetPasswordDataResponse (Maybe ISO8601)

The time the data was last updated.

gpdrPasswordData :: Lens' GetPasswordDataResponse (Maybe Text)

The password of the instance.