Network.AWS.EC2.GetReservedInstancesExchangeQuote

Description

Returns details about the values and term of your specified Convertible Reserved Instances. When a target configuration is specified, it returns information about whether the exchange is valid and can be performed.

Creating a Request

getReservedInstancesExchangeQuote :: GetReservedInstancesExchangeQuote

Creates a value of GetReservedInstancesExchangeQuote with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grieqTargetConfigurations - The configuration requirements of the Convertible Reserved Instances to exchange for your current Convertible Reserved Instances.

grieqDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

grieqReservedInstanceIds - The IDs of the Convertible Reserved Instances to exchange.

data GetReservedInstancesExchangeQuote

Contains the parameters for GetReservedInstanceExchangeQuote.

See: getReservedInstancesExchangeQuote smart constructor.

Request Lenses

grieqTargetConfigurations :: Lens' GetReservedInstancesExchangeQuote [TargetConfigurationRequest]

The configuration requirements of the Convertible Reserved Instances to exchange for your current Convertible Reserved Instances.

grieqDryRun :: Lens' GetReservedInstancesExchangeQuote (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

grieqReservedInstanceIds :: Lens' GetReservedInstancesExchangeQuote [Text]

The IDs of the Convertible Reserved Instances to exchange.

Destructuring the Response

getReservedInstancesExchangeQuoteResponse

Arguments

:: Int
grieqrsResponseStatus

-> GetReservedInstancesExchangeQuoteResponse

Creates a value of GetReservedInstancesExchangeQuoteResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grieqrsValidationFailureReason - Describes the reason why the exchange cannot be completed.

grieqrsTargetConfigurationValueRollup - The cost associated with the Reserved Instance.

grieqrsCurrencyCode - The currency of the transaction.

grieqrsTargetConfigurationValueSet - The values of the target Convertible Reserved Instances.

grieqrsReservedInstanceValueRollup - The cost associated with the Reserved Instance.

grieqrsOutputReservedInstancesWillExpireAt - The new end date of the reservation term.

grieqrsReservedInstanceValueSet - The configuration of your Convertible Reserved Instances.

grieqrsIsValidExchange - If true , the exchange is valid. If false , the exchange cannot be completed.

grieqrsPaymentDue - The total true upfront charge for the exchange.

grieqrsResponseStatus - -- | The response status code.

data GetReservedInstancesExchangeQuoteResponse

Contains the output of GetReservedInstancesExchangeQuote.

See: getReservedInstancesExchangeQuoteResponse smart constructor.

Response Lenses

grieqrsValidationFailureReason :: Lens' GetReservedInstancesExchangeQuoteResponse (Maybe Text)

Describes the reason why the exchange cannot be completed.

grieqrsTargetConfigurationValueRollup :: Lens' GetReservedInstancesExchangeQuoteResponse (Maybe ReservationValue)

The cost associated with the Reserved Instance.

grieqrsCurrencyCode :: Lens' GetReservedInstancesExchangeQuoteResponse (Maybe Text)

The currency of the transaction.

grieqrsTargetConfigurationValueSet :: Lens' GetReservedInstancesExchangeQuoteResponse [TargetReservationValue]

The values of the target Convertible Reserved Instances.

grieqrsReservedInstanceValueRollup :: Lens' GetReservedInstancesExchangeQuoteResponse (Maybe ReservationValue)

The cost associated with the Reserved Instance.

grieqrsOutputReservedInstancesWillExpireAt :: Lens' GetReservedInstancesExchangeQuoteResponse (Maybe UTCTime)

The new end date of the reservation term.

grieqrsReservedInstanceValueSet :: Lens' GetReservedInstancesExchangeQuoteResponse [ReservedInstanceReservationValue]

The configuration of your Convertible Reserved Instances.

grieqrsIsValidExchange :: Lens' GetReservedInstancesExchangeQuoteResponse (Maybe Bool)

If true , the exchange is valid. If false , the exchange cannot be completed.

grieqrsPaymentDue :: Lens' GetReservedInstancesExchangeQuoteResponse (Maybe Text)

The total true upfront charge for the exchange.

grieqrsResponseStatus :: Lens' GetReservedInstancesExchangeQuoteResponse Int

- | The response status code.