Network.AWS.EC2.ImportImage

Description

Import single or multi-volume disk images or EBS snapshots into an Amazon Machine Image (AMI). For more information, see Importing a VM as an Image Using VM Import/Export in the VM Import/Export User Guide .

Creating a Request

importImage :: ImportImage

Creates a value of ImportImage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

impHypervisor - The target hypervisor platform. Valid values: xen

impPlatform - The operating system of the virtual machine. Valid values: Windows | Linux

impClientToken - The token to enable idempotency for VM import requests.

impLicenseType - The license type to be used for the Amazon Machine Image (AMI) after importing. Note: You may only use BYOL if you have existing licenses with rights to use these licenses in a third party cloud like AWS. For more information, see Prerequisites in the VM Import/Export User Guide. Valid values: AWS | BYOL

impRoleName - The name of the role to use when not using the default role, vmimport.

impArchitecture - The architecture of the virtual machine. Valid values: i386 | x86_64

impDescription - A description string for the import image task.

impDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

impClientData - The client-specific data.

impDiskContainers - Information about the disk containers.

data ImportImage

Contains the parameters for ImportImage.

See: importImage smart constructor.

Request Lenses

impHypervisor :: Lens' ImportImage (Maybe Text)

The target hypervisor platform. Valid values: xen

impPlatform :: Lens' ImportImage (Maybe Text)

The operating system of the virtual machine. Valid values: Windows | Linux

impClientToken :: Lens' ImportImage (Maybe Text)

The token to enable idempotency for VM import requests.

impLicenseType :: Lens' ImportImage (Maybe Text)

The license type to be used for the Amazon Machine Image (AMI) after importing. Note: You may only use BYOL if you have existing licenses with rights to use these licenses in a third party cloud like AWS. For more information, see Prerequisites in the VM Import/Export User Guide. Valid values: AWS | BYOL

impRoleName :: Lens' ImportImage (Maybe Text)

The name of the role to use when not using the default role, vmimport.

impArchitecture :: Lens' ImportImage (Maybe Text)

The architecture of the virtual machine. Valid values: i386 | x86_64

impDescription :: Lens' ImportImage (Maybe Text)

A description string for the import image task.

impDryRun :: Lens' ImportImage (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

impClientData :: Lens' ImportImage (Maybe ClientData)

The client-specific data.

impDiskContainers :: Lens' ImportImage [ImageDiskContainer]

Information about the disk containers.

Destructuring the Response

importImageResponse

Arguments

:: Int
irsResponseStatus

-> ImportImageResponse

Creates a value of ImportImageResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

irsStatus - A brief status of the task.

irsHypervisor - The target hypervisor of the import task.

irsPlatform - The operating system of the virtual machine.

irsProgress - The progress of the task.

irsLicenseType - The license type of the virtual machine.

irsSnapshotDetails - Information about the snapshots.

irsStatusMessage - A detailed status message of the import task.

irsImageId - The ID of the Amazon Machine Image (AMI) created by the import task.

irsImportTaskId - The task ID of the import image task.

irsArchitecture - The architecture of the virtual machine.

irsDescription - A description of the import task.

irsResponseStatus - -- | The response status code.

data ImportImageResponse

Contains the output for ImportImage.

See: importImageResponse smart constructor.

Response Lenses

irsStatus :: Lens' ImportImageResponse (Maybe Text)

A brief status of the task.

irsHypervisor :: Lens' ImportImageResponse (Maybe Text)

The target hypervisor of the import task.

irsPlatform :: Lens' ImportImageResponse (Maybe Text)

The operating system of the virtual machine.

irsProgress :: Lens' ImportImageResponse (Maybe Text)

The progress of the task.

irsLicenseType :: Lens' ImportImageResponse (Maybe Text)

The license type of the virtual machine.

irsSnapshotDetails :: Lens' ImportImageResponse [SnapshotDetail]

Information about the snapshots.

irsStatusMessage :: Lens' ImportImageResponse (Maybe Text)

A detailed status message of the import task.

irsImageId :: Lens' ImportImageResponse (Maybe Text)

The ID of the Amazon Machine Image (AMI) created by the import task.

irsImportTaskId :: Lens' ImportImageResponse (Maybe Text)

The task ID of the import image task.

irsArchitecture :: Lens' ImportImageResponse (Maybe Text)

The architecture of the virtual machine.

irsDescription :: Lens' ImportImageResponse (Maybe Text)

A description of the import task.

irsResponseStatus :: Lens' ImportImageResponse Int

- | The response status code.