Network.AWS.EC2.ImportInstance

Description

Creates an import instance task using metadata from the specified disk image. After importing the image, you then upload it using the 'ec2-import-volume' command in the EC2 command line tools. For more information, see Using theCommand Line Tools to Import Your Virtual Machine to Amazon EC2 in the AmazonElastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-ImportInstance.html

Request

data ImportInstance

Request constructor

importInstance

Arguments

:: PlatformValues
iiPlatform

-> ImportInstance

ImportInstance constructor.

The fields accessible through corresponding lenses are:

iiDescription :: Maybe Text

iiDiskImages :: [DiskImage]

iiDryRun :: Maybe Bool

iiLaunchSpecification :: Maybe ImportInstanceLaunchSpecification

iiPlatform :: PlatformValues

Request lenses

iiDescription :: Lens' ImportInstance (Maybe Text)

A description for the instance being imported.

iiDiskImages :: Lens' ImportInstance [DiskImage]

iiDryRun :: Lens' ImportInstance (Maybe Bool)

iiLaunchSpecification :: Lens' ImportInstance (Maybe ImportInstanceLaunchSpecification)

iiPlatform :: Lens' ImportInstance PlatformValues

The instance operating system.

Response

data ImportInstanceResponse

Response constructor

importInstanceResponse :: ImportInstanceResponse

ImportInstanceResponse constructor.

The fields accessible through corresponding lenses are:

iirConversionTask :: Maybe ConversionTask

Response lenses

iirConversionTask :: Lens' ImportInstanceResponse (Maybe ConversionTask)