Network.AWS.EC2.ImportSnapshot

Description

Imports a disk into an EBS snapshot.

Creating a Request

importSnapshot :: ImportSnapshot

Creates a value of ImportSnapshot with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

isDiskContainer - Information about the disk container.

isClientToken - Token to enable idempotency for VM import requests.

isRoleName - The name of the role to use when not using the default role, vmimport.

isDescription - The description string for the import snapshot task.

isDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

isClientData - The client-specific data.

data ImportSnapshot

Contains the parameters for ImportSnapshot.

See: importSnapshot smart constructor.

Request Lenses

isDiskContainer :: Lens' ImportSnapshot (Maybe SnapshotDiskContainer)

Information about the disk container.

isClientToken :: Lens' ImportSnapshot (Maybe Text)

Token to enable idempotency for VM import requests.

isRoleName :: Lens' ImportSnapshot (Maybe Text)

The name of the role to use when not using the default role, vmimport.

isDescription :: Lens' ImportSnapshot (Maybe Text)

The description string for the import snapshot task.

isDryRun :: Lens' ImportSnapshot (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

isClientData :: Lens' ImportSnapshot (Maybe ClientData)

The client-specific data.

Destructuring the Response

importSnapshotResponse

Arguments

:: Int
isrsResponseStatus

-> ImportSnapshotResponse

Creates a value of ImportSnapshotResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

isrsSnapshotTaskDetail - Information about the import snapshot task.

isrsImportTaskId - The ID of the import snapshot task.

isrsDescription - A description of the import snapshot task.

isrsResponseStatus - -- | The response status code.

data ImportSnapshotResponse

Contains the output for ImportSnapshot.

See: importSnapshotResponse smart constructor.

Response Lenses

isrsSnapshotTaskDetail :: Lens' ImportSnapshotResponse (Maybe SnapshotTaskDetail)

Information about the import snapshot task.

isrsImportTaskId :: Lens' ImportSnapshotResponse (Maybe Text)

The ID of the import snapshot task.

isrsDescription :: Lens' ImportSnapshotResponse (Maybe Text)

A description of the import snapshot task.

isrsResponseStatus :: Lens' ImportSnapshotResponse Int

- | The response status code.