Network.AWS.EC2.ImportVolume

Description

Creates an import volume task using metadata from the specified disk image.For more information, see Importing Disks to Amazon EBS .

For information about the import manifest referenced by this API action, see VM Import Manifest .

Creating a Request

importVolume

Arguments

:: Text
ivAvailabilityZone

-> DiskImageDetail
ivImage

-> VolumeDetail
ivVolume

-> ImportVolume

Creates a value of ImportVolume with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ivDescription - A description of the volume.

ivDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

ivAvailabilityZone - The Availability Zone for the resulting EBS volume.

ivImage - The disk image.

ivVolume - The volume size.

data ImportVolume

Contains the parameters for ImportVolume.

See: importVolume smart constructor.

Request Lenses

ivDescription :: Lens' ImportVolume (Maybe Text)

A description of the volume.

ivDryRun :: Lens' ImportVolume (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

ivAvailabilityZone :: Lens' ImportVolume Text

The Availability Zone for the resulting EBS volume.

ivImage :: Lens' ImportVolume DiskImageDetail

The disk image.

ivVolume :: Lens' ImportVolume VolumeDetail

The volume size.

Destructuring the Response

importVolumeResponse

Arguments

:: Int
ivrsResponseStatus

-> ImportVolumeResponse

Creates a value of ImportVolumeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ivrsConversionTask - Information about the conversion task.

ivrsResponseStatus - -- | The response status code.

data ImportVolumeResponse

Contains the output for ImportVolume.

See: importVolumeResponse smart constructor.

Response Lenses

ivrsConversionTask :: Lens' ImportVolumeResponse (Maybe ConversionTask)

Information about the conversion task.

ivrsResponseStatus :: Lens' ImportVolumeResponse Int

- | The response status code.