Network.AWS.EC2.ModifyHosts

Description

Modify the auto-placement setting of a Dedicated Host. When auto-placement is enabled, AWS will place instances that you launch with a tenancy of host , but without targeting a specific host ID, onto any available Dedicated Host in your account which has auto-placement enabled. When auto-placement is disabled, you need to provide a host ID if you want the instance to launch onto a specific host. If no host ID is provided, the instance will be launched onto a suitable host which has auto-placement enabled.

Creating a Request

modifyHosts

Arguments

:: AutoPlacement
mhAutoPlacement

-> ModifyHosts

Creates a value of ModifyHosts with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mhAutoPlacement - Specify whether to enable or disable auto-placement.

mhHostIds - The host IDs of the Dedicated Hosts you want to modify.

data ModifyHosts

Contains the parameters for ModifyHosts.

See: modifyHosts smart constructor.

Request Lenses

mhAutoPlacement :: Lens' ModifyHosts AutoPlacement

Specify whether to enable or disable auto-placement.

mhHostIds :: Lens' ModifyHosts [Text]

The host IDs of the Dedicated Hosts you want to modify.

Destructuring the Response

modifyHostsResponse

Arguments

:: Int
mhrsResponseStatus

-> ModifyHostsResponse

Creates a value of ModifyHostsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mhrsUnsuccessful - The IDs of the Dedicated Hosts that could not be modified. Check whether the setting you requested can be used.

mhrsSuccessful - The IDs of the Dedicated Hosts that were successfully modified.

mhrsResponseStatus - -- | The response status code.

data ModifyHostsResponse

Contains the output of ModifyHosts.

See: modifyHostsResponse smart constructor.

Response Lenses

mhrsUnsuccessful :: Lens' ModifyHostsResponse [UnsuccessfulItem]

The IDs of the Dedicated Hosts that could not be modified. Check whether the setting you requested can be used.

mhrsSuccessful :: Lens' ModifyHostsResponse [Text]

The IDs of the Dedicated Hosts that were successfully modified.

mhrsResponseStatus :: Lens' ModifyHostsResponse Int

- | The response status code.