Network.AWS.EC2.ModifyIdFormat

Description

Modifies the ID format for the specified resource on a per-region basis. You can specify that resources should receive longer IDs (17-character IDs) when they are created. The following resource types support longer IDs: instance | reservation | snapshot | volume .

This setting applies to the IAM user who makes the request; it does not apply to the entire AWS account. By default, an IAM user defaults to the same settings as the root user. If you're using this action as the root user, then these settings apply to the entire account, unless an IAM user explicitly overrides these settings for themselves. For more information, see Resource IDs in the Amazon Elastic Compute Cloud User Guide .

Resources created with longer IDs are visible to all IAM roles and users, regardless of these settings and provided that they have permission to use the relevant Describe command for the resource type.

Creating a Request

modifyIdFormat

Arguments

:: Text
mifResource

-> Bool
mifUseLongIds

-> ModifyIdFormat

Creates a value of ModifyIdFormat with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mifResource - The type of resource: instance | reservation | snapshot | volume

mifUseLongIds - Indicate whether the resource should use longer IDs (17-character IDs).

data ModifyIdFormat

Contains the parameters of ModifyIdFormat.

See: modifyIdFormat smart constructor.

Request Lenses

mifResource :: Lens' ModifyIdFormat Text

The type of resource: instance | reservation | snapshot | volume

mifUseLongIds :: Lens' ModifyIdFormat Bool

Indicate whether the resource should use longer IDs (17-character IDs).

Destructuring the Response

modifyIdFormatResponse :: ModifyIdFormatResponse

Creates a value of ModifyIdFormatResponse with the minimum fields required to make a request.

data ModifyIdFormatResponse

See: modifyIdFormatResponse smart constructor.