Network.AWS.EC2.ModifyImageAttribute

Description

Modifies the specified attribute of the specified AMI. You can specify only one attribute at a time.

AWS Marketplace product codes cannot be modified. Images with an AWS Marketplace product code cannot be made public.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-ModifyImageAttribute.html

Request

data ModifyImageAttribute

Request constructor

modifyImageAttribute

Arguments

:: Text
miaImageId

-> ModifyImageAttribute

ModifyImageAttribute constructor.

The fields accessible through corresponding lenses are:

miaAttribute :: Maybe Text

miaDescription :: Maybe AttributeValue

miaDryRun :: Maybe Bool

miaImageId :: Text

miaLaunchPermission :: Maybe LaunchPermissionModifications

miaOperationType :: Maybe Text

miaProductCodes :: [Text]

miaUserGroups :: [Text]

miaUserIds :: [Text]

miaValue :: Maybe Text

Request lenses

miaAttribute :: Lens' ModifyImageAttribute (Maybe Text)

The name of the attribute to modify.

miaDescription :: Lens' ModifyImageAttribute (Maybe AttributeValue)

A description for the AMI.

miaDryRun :: Lens' ModifyImageAttribute (Maybe Bool)

miaImageId :: Lens' ModifyImageAttribute Text

The ID of the AMI.

miaLaunchPermission :: Lens' ModifyImageAttribute (Maybe LaunchPermissionModifications)

A launch permission modification.

miaOperationType :: Lens' ModifyImageAttribute (Maybe Text)

The operation type.

miaProductCodes :: Lens' ModifyImageAttribute [Text]

One or more product codes. After you add a product code to an AMI, it can't be removed. This is only valid when modifying the productCodes attribute.

miaUserGroups :: Lens' ModifyImageAttribute [Text]

One or more user groups. This is only valid when modifying the launchPermission attribute.

miaUserIds :: Lens' ModifyImageAttribute [Text]

One or more AWS account IDs. This is only valid when modifying the launchPermission attribute.

miaValue :: Lens' ModifyImageAttribute (Maybe Text)

The value of the attribute being modified. This is only valid when modifying the description attribute.

Response

data ModifyImageAttributeResponse

Response constructor

modifyImageAttributeResponse :: ModifyImageAttributeResponse

ModifyImageAttributeResponse constructor.