Network.AWS.EC2.ModifyInstanceAttribute

Description

Modifies the specified attribute of the specified instance. You can specify only one attribute at a time.

To modify some attributes, the instance must be stopped. For more information, see Modifying Attributes of a Stopped Instance in the AmazonElastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-ModifyInstanceAttribute.html

Request

data ModifyInstanceAttribute

Request constructor

modifyInstanceAttribute

Arguments

:: Text
mia1InstanceId

-> ModifyInstanceAttribute

ModifyInstanceAttribute constructor.

The fields accessible through corresponding lenses are:

mia1Attribute :: Maybe InstanceAttributeName

mia1BlockDeviceMappings :: [InstanceBlockDeviceMappingSpecification]

mia1DisableApiTermination :: Maybe AttributeBooleanValue

mia1DryRun :: Maybe Bool

mia1EbsOptimized :: Maybe AttributeBooleanValue

mia1Groups :: [Text]

mia1InstanceId :: Text

mia1InstanceInitiatedShutdownBehavior :: Maybe AttributeValue

mia1InstanceType :: Maybe AttributeValue

mia1Kernel :: Maybe AttributeValue

mia1Ramdisk :: Maybe AttributeValue

mia1SourceDestCheck :: Maybe AttributeBooleanValue

mia1SriovNetSupport :: Maybe AttributeValue

mia1UserData :: Maybe BlobAttributeValue

mia1Value :: Maybe Text

Request lenses

mia1Attribute :: Lens' ModifyInstanceAttribute (Maybe InstanceAttributeName)

The name of the attribute.

mia1BlockDeviceMappings :: Lens' ModifyInstanceAttribute [InstanceBlockDeviceMappingSpecification]

Modifies the DeleteOnTermination attribute for volumes that are currently attached. The volume must be owned by the caller. If no value is specified for DeleteOnTermination, the default is true and the volume is deleted when the instance is terminated.

To add instance store volumes to an Amazon EBS-backed instance, you must add them when you launch the instance. For more information, see Updating theBlock Device Mapping when Launching an Instance in the Amazon Elastic ComputeCloud User Guide for Linux.

mia1DisableApiTermination :: Lens' ModifyInstanceAttribute (Maybe AttributeBooleanValue)

If the value is true, you can't terminate the instance using the Amazon EC2 console, CLI, or API; otherwise, you can.

mia1DryRun :: Lens' ModifyInstanceAttribute (Maybe Bool)

mia1EbsOptimized :: Lens' ModifyInstanceAttribute (Maybe AttributeBooleanValue)

Specifies whether the instance is optimized for EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS Optimized instance.

mia1Groups :: Lens' ModifyInstanceAttribute [Text]

EC2-VPC

Changes the security groups of the instance. You must specify at least one security group, even if it's just the default security group for the VPC. You must specify the security group ID, not the security group name.

mia1InstanceId :: Lens' ModifyInstanceAttribute Text

The ID of the instance.

mia1InstanceInitiatedShutdownBehavior :: Lens' ModifyInstanceAttribute (Maybe AttributeValue)

Specifies whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

mia1InstanceType :: Lens' ModifyInstanceAttribute (Maybe AttributeValue)

Changes the instance type to the specified value. For more information, see Instance Types. If the instance type is not valid, the error returned is InvalidInstanceAttributeValue.

mia1Kernel :: Lens' ModifyInstanceAttribute (Maybe AttributeValue)

Changes the instance's kernel to the specified value. We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB.

mia1Ramdisk :: Lens' ModifyInstanceAttribute (Maybe AttributeValue)

Changes the instance's RAM disk to the specified value. We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB.

mia1SourceDestCheck :: Lens' ModifyInstanceAttribute (Maybe AttributeBooleanValue)

Specifies whether source/destination checking is enabled. A value of true means that checking is enabled, and false means checking is disabled. This value must be false for a NAT instance to perform NAT.

mia1SriovNetSupport :: Lens' ModifyInstanceAttribute (Maybe AttributeValue)

Set to simple to enable enhanced networking for the instance.

There is no way to disable enhanced networking at this time.

This option is supported only for HVM instances. Specifying this option with a PV instance can make it unreachable.

mia1UserData :: Lens' ModifyInstanceAttribute (Maybe BlobAttributeValue)

Changes the instance's user data to the specified value.

mia1Value :: Lens' ModifyInstanceAttribute (Maybe Text)

A new value for the attribute. Use only with the kernel, ramdisk, userData, disableApiTermination, or intanceInitiateShutdownBehavior attribute.

Response

data ModifyInstanceAttributeResponse

Response constructor

modifyInstanceAttributeResponse :: ModifyInstanceAttributeResponse

ModifyInstanceAttributeResponse constructor.