Network.AWS.EC2.ModifyInstancePlacement

Description

Set the instance affinity value for a specific stopped instance and modify the instance tenancy setting.

Instance affinity is disabled by default. When instance affinity is host and it is not associated with a specific Dedicated Host, the next time it is launched it will automatically be associated with the host it lands on. This relationship will persist if the instance is stopped/started, or rebooted.

You can modify the host ID associated with a stopped instance. If a stopped instance has a new host ID association, the instance will target that host when restarted.

You can modify the tenancy of a stopped instance with a tenancy of host or dedicated .

Affinity, hostID, and tenancy are not required parameters, but at least one of them must be specified in the request. Affinity and tenancy can be modified in the same request, but tenancy can only be modified on instances that are stopped.

Creating a Request

modifyInstancePlacement

Arguments

:: Text
mipInstanceId

-> ModifyInstancePlacement

Creates a value of ModifyInstancePlacement with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mipAffinity - The new affinity setting for the instance.

mipHostId - The ID of the Dedicated Host that the instance will have affinity with.

mipTenancy - The tenancy of the instance that you are modifying.

mipInstanceId - The ID of the instance that you are modifying.

data ModifyInstancePlacement

Contains the parameters for ModifyInstancePlacement.

See: modifyInstancePlacement smart constructor.

Request Lenses

mipAffinity :: Lens' ModifyInstancePlacement (Maybe Affinity)

The new affinity setting for the instance.

mipHostId :: Lens' ModifyInstancePlacement (Maybe Text)

The ID of the Dedicated Host that the instance will have affinity with.

mipTenancy :: Lens' ModifyInstancePlacement (Maybe HostTenancy)

The tenancy of the instance that you are modifying.

mipInstanceId :: Lens' ModifyInstancePlacement Text

The ID of the instance that you are modifying.

Destructuring the Response

modifyInstancePlacementResponse

Arguments

:: Int
miprsResponseStatus

-> ModifyInstancePlacementResponse

Creates a value of ModifyInstancePlacementResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

miprsReturn - Is true if the request succeeds, and an error otherwise.

miprsResponseStatus - -- | The response status code.

data ModifyInstancePlacementResponse

Contains the output of ModifyInstancePlacement.

See: modifyInstancePlacementResponse smart constructor.

Response Lenses

miprsReturn :: Lens' ModifyInstancePlacementResponse (Maybe Bool)

Is true if the request succeeds, and an error otherwise.

miprsResponseStatus :: Lens' ModifyInstancePlacementResponse Int

- | The response status code.