Network.AWS.EC2.ModifySnapshotAttribute

Description

Adds or removes permission settings for the specified snapshot. You may add or remove specified AWS account IDs from a snapshot's list of create volume permissions, but you cannot do both in a single API call. If you need to both add and remove account IDs for a snapshot, you must use multiple API calls.

For more information on modifying snapshot permissions, see Sharing Snapshots in the Amazon Elastic Compute Cloud User Guide for Linux.

Snapshots with AWS Marketplace product codes cannot be made public.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-ModifySnapshotAttribute.html

Request

data ModifySnapshotAttribute

Request constructor

modifySnapshotAttribute

Arguments

:: Text
msaSnapshotId

-> ModifySnapshotAttribute

ModifySnapshotAttribute constructor.

The fields accessible through corresponding lenses are:

msaAttribute :: Maybe SnapshotAttributeName

msaCreateVolumePermission :: Maybe CreateVolumePermissionModifications

msaDryRun :: Maybe Bool

msaGroupNames :: [Text]

msaOperationType :: Maybe Text

msaSnapshotId :: Text

msaUserIds :: [Text]

Request lenses

msaAttribute :: Lens' ModifySnapshotAttribute (Maybe SnapshotAttributeName)

The snapshot attribute to modify.

msaCreateVolumePermission :: Lens' ModifySnapshotAttribute (Maybe CreateVolumePermissionModifications)

A JSON representation of the snapshot attribute modification.

msaDryRun :: Lens' ModifySnapshotAttribute (Maybe Bool)

msaGroupNames :: Lens' ModifySnapshotAttribute [Text]

The group to modify for the snapshot.

msaOperationType :: Lens' ModifySnapshotAttribute (Maybe Text)

The type of operation to perform to the attribute.

msaSnapshotId :: Lens' ModifySnapshotAttribute Text

The ID of the snapshot.

msaUserIds :: Lens' ModifySnapshotAttribute [Text]

The account ID to modify for the snapshot.

Response

data ModifySnapshotAttributeResponse

Response constructor

modifySnapshotAttributeResponse :: ModifySnapshotAttributeResponse

ModifySnapshotAttributeResponse constructor.