Network.AWS.EC2.ModifySpotFleetRequest

Description

Modifies the specified Spot fleet request.

While the Spot fleet request is being modified, it is in the modifying state.

To scale up your Spot fleet, increase its target capacity. The Spot fleet launches the additional Spot instances according to the allocation strategy for the Spot fleet request. If the allocation strategy is lowestPrice , the Spot fleet launches instances using the Spot pool with the lowest price. If the allocation strategy is diversified , the Spot fleet distributes the instances across the Spot pools.

To scale down your Spot fleet, decrease its target capacity. First, the Spot fleet cancels any open bids that exceed the new target capacity. You can request that the Spot fleet terminate Spot instances until the size of the fleet no longer exceeds the new target capacity. If the allocation strategy is lowestPrice , the Spot fleet terminates the instances with the highest price per unit. If the allocation strategy is diversified , the Spot fleet terminates instances across the Spot pools. Alternatively, you can request that the Spot fleet keep the fleet at its current size, but not replace any Spot instances that are interrupted or that you terminate manually.

Creating a Request

modifySpotFleetRequest

Arguments

:: Text
msfrSpotFleetRequestId

-> ModifySpotFleetRequest

Creates a value of ModifySpotFleetRequest with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

msfrTargetCapacity - The size of the fleet.

msfrExcessCapacityTerminationPolicy - Indicates whether running Spot instances should be terminated if the target capacity of the Spot fleet request is decreased below the current size of the Spot fleet.

msfrSpotFleetRequestId - The ID of the Spot fleet request.

data ModifySpotFleetRequest

Contains the parameters for ModifySpotFleetRequest.

See: modifySpotFleetRequest smart constructor.

Request Lenses

msfrTargetCapacity :: Lens' ModifySpotFleetRequest (Maybe Int)

The size of the fleet.

msfrExcessCapacityTerminationPolicy :: Lens' ModifySpotFleetRequest (Maybe ExcessCapacityTerminationPolicy)

Indicates whether running Spot instances should be terminated if the target capacity of the Spot fleet request is decreased below the current size of the Spot fleet.

msfrSpotFleetRequestId :: Lens' ModifySpotFleetRequest Text

The ID of the Spot fleet request.

Destructuring the Response

modifySpotFleetRequestResponse

Arguments

:: Int
msfrrsResponseStatus

-> ModifySpotFleetRequestResponse

Creates a value of ModifySpotFleetRequestResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

msfrrsReturn - Is true if the request succeeds, and an error otherwise.

msfrrsResponseStatus - -- | The response status code.

data ModifySpotFleetRequestResponse

Contains the output of ModifySpotFleetRequest.

See: modifySpotFleetRequestResponse smart constructor.

Response Lenses

msfrrsReturn :: Lens' ModifySpotFleetRequestResponse (Maybe Bool)

Is true if the request succeeds, and an error otherwise.

msfrrsResponseStatus :: Lens' ModifySpotFleetRequestResponse Int

- | The response status code.