Network.AWS.EC2.ModifyVPCAttribute

Description

Modifies the specified attribute of the specified VPC.

Creating a Request

modifyVPCAttribute

Arguments

:: Text
mvaVPCId

-> ModifyVPCAttribute

Creates a value of ModifyVPCAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mvaEnableDNSHostnames - Indicates whether the instances launched in the VPC get DNS hostnames. If enabled, instances in the VPC get DNS hostnames; otherwise, they do not. You cannot modify the DNS resolution and DNS hostnames attributes in the same request. Use separate requests for each attribute. You can only enable DNS hostnames if you've enabled DNS support.

mvaEnableDNSSupport - Indicates whether the DNS resolution is supported for the VPC. If enabled, queries to the Amazon provided DNS server at the 169.254.169.253 IP address, or the reserved IP address at the base of the VPC network range "plus two" will succeed. If disabled, the Amazon provided DNS service in the VPC that resolves public DNS hostnames to IP addresses is not enabled. You cannot modify the DNS resolution and DNS hostnames attributes in the same request. Use separate requests for each attribute.

mvaVPCId - The ID of the VPC.

data ModifyVPCAttribute

Contains the parameters for ModifyVpcAttribute.

See: modifyVPCAttribute smart constructor.

Request Lenses

mvaEnableDNSHostnames :: Lens' ModifyVPCAttribute (Maybe AttributeBooleanValue)

Indicates whether the instances launched in the VPC get DNS hostnames. If enabled, instances in the VPC get DNS hostnames; otherwise, they do not. You cannot modify the DNS resolution and DNS hostnames attributes in the same request. Use separate requests for each attribute. You can only enable DNS hostnames if you've enabled DNS support.

mvaEnableDNSSupport :: Lens' ModifyVPCAttribute (Maybe AttributeBooleanValue)

Indicates whether the DNS resolution is supported for the VPC. If enabled, queries to the Amazon provided DNS server at the 169.254.169.253 IP address, or the reserved IP address at the base of the VPC network range "plus two" will succeed. If disabled, the Amazon provided DNS service in the VPC that resolves public DNS hostnames to IP addresses is not enabled. You cannot modify the DNS resolution and DNS hostnames attributes in the same request. Use separate requests for each attribute.

mvaVPCId :: Lens' ModifyVPCAttribute Text

The ID of the VPC.

Destructuring the Response

modifyVPCAttributeResponse :: ModifyVPCAttributeResponse

Creates a value of ModifyVPCAttributeResponse with the minimum fields required to make a request.

data ModifyVPCAttributeResponse

See: modifyVPCAttributeResponse smart constructor.