Network.AWS.EC2.ModifyVPCPeeringConnectionOptions

Description

Modifies the VPC peering connection options on one side of a VPC peering connection. You can do the following:

Enable/disable communication over the peering connection between an EC2-Classic instance that's linked to your VPC (using ClassicLink) and instances in the peer VPC.

Enable/disable communication over the peering connection between instances in your VPC and an EC2-Classic instance that's linked to the peer VPC.

Enable/disable a local VPC to resolve public DNS hostnames to private IP addresses when queried from instances in the peer VPC.

If the peered VPCs are in different accounts, each owner must initiate a separate request to modify the peering connection options, depending on whether their VPC was the requester or accepter for the VPC peering connection. If the peered VPCs are in the same account, you can modify the requester and accepter options in the same request. To confirm which VPC is the accepter and requester for a VPC peering connection, use the DescribeVpcPeeringConnections command.

Creating a Request

modifyVPCPeeringConnectionOptions

Arguments

:: Text
mvpcoVPCPeeringConnectionId

-> ModifyVPCPeeringConnectionOptions

Creates a value of ModifyVPCPeeringConnectionOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mvpcoRequesterPeeringConnectionOptions - The VPC peering connection options for the requester VPC.

mvpcoAccepterPeeringConnectionOptions - The VPC peering connection options for the accepter VPC.

mvpcoDryRun - Checks whether you have the required permissions for the operation, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

mvpcoVPCPeeringConnectionId - The ID of the VPC peering connection.

data ModifyVPCPeeringConnectionOptions

See: modifyVPCPeeringConnectionOptions smart constructor.

Request Lenses

mvpcoRequesterPeeringConnectionOptions :: Lens' ModifyVPCPeeringConnectionOptions (Maybe PeeringConnectionOptionsRequest)

The VPC peering connection options for the requester VPC.

mvpcoAccepterPeeringConnectionOptions :: Lens' ModifyVPCPeeringConnectionOptions (Maybe PeeringConnectionOptionsRequest)

The VPC peering connection options for the accepter VPC.

mvpcoDryRun :: Lens' ModifyVPCPeeringConnectionOptions (Maybe Bool)

Checks whether you have the required permissions for the operation, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

mvpcoVPCPeeringConnectionId :: Lens' ModifyVPCPeeringConnectionOptions Text

The ID of the VPC peering connection.

Destructuring the Response

modifyVPCPeeringConnectionOptionsResponse

Arguments

:: Int
mvpcorsResponseStatus

-> ModifyVPCPeeringConnectionOptionsResponse

Creates a value of ModifyVPCPeeringConnectionOptionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mvpcorsRequesterPeeringConnectionOptions - Information about the VPC peering connection options for the requester VPC.

mvpcorsAccepterPeeringConnectionOptions - Information about the VPC peering connection options for the accepter VPC.

mvpcorsResponseStatus - -- | The response status code.

data ModifyVPCPeeringConnectionOptionsResponse

See: modifyVPCPeeringConnectionOptionsResponse smart constructor.

Response Lenses

mvpcorsRequesterPeeringConnectionOptions :: Lens' ModifyVPCPeeringConnectionOptionsResponse (Maybe PeeringConnectionOptions)

Information about the VPC peering connection options for the requester VPC.

mvpcorsAccepterPeeringConnectionOptions :: Lens' ModifyVPCPeeringConnectionOptionsResponse (Maybe PeeringConnectionOptions)

Information about the VPC peering connection options for the accepter VPC.

mvpcorsResponseStatus :: Lens' ModifyVPCPeeringConnectionOptionsResponse Int

- | The response status code.