Network.AWS.EC2.ModifyVolumeAttribute

Description

Modifies a volume attribute.

By default, all I/O operations for the volume are suspended when the data on the volume is determined to be potentially inconsistent, to prevent undetectable, latent data corruption. The I/O access to the volume can be resumed by first enabling I/O access and then checking the data consistency on your volume.

You can change the default behavior to resume I/O operations. We recommend that you change this only for boot volumes or for volumes that are stateless or disposable.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-ModifyVolumeAttribute.html

Request

data ModifyVolumeAttribute

Request constructor

modifyVolumeAttribute

Arguments

:: Text
mvaVolumeId

-> ModifyVolumeAttribute

ModifyVolumeAttribute constructor.

The fields accessible through corresponding lenses are:

mvaAutoEnableIO :: Maybe AttributeBooleanValue

mvaDryRun :: Maybe Bool

mvaVolumeId :: Text

Request lenses

mvaAutoEnableIO :: Lens' ModifyVolumeAttribute (Maybe AttributeBooleanValue)

Indicates whether the volume should be auto-enabled for I/O operations.

mvaDryRun :: Lens' ModifyVolumeAttribute (Maybe Bool)

mvaVolumeId :: Lens' ModifyVolumeAttribute Text

The ID of the volume.

Response

data ModifyVolumeAttributeResponse

Response constructor

modifyVolumeAttributeResponse :: ModifyVolumeAttributeResponse

ModifyVolumeAttributeResponse constructor.