Network.AWS.EC2.MonitorInstances

Description

Enables monitoring for a running instance. For more information about monitoring instances, see Monitoring Your Instances and Volumes in the Amazon Elastic Compute Cloud User Guide. Example This example enables monitoring for two instances. https:/ec2.amazonaws.com?Action=MonitorInstances &amp;InstanceId.1=i-43a4412a &amp;InstanceId.2=i-23a3397d &amp;AUTHPARAMS &lt;MonitorInstancesResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;instancesSet&gt; &lt;item&gt; &lt;instanceId&gt;i-43a4412a&lt;/instanceId&gt; &lt;monitoring&gt; &lt;state&gt;pending&lt;state&gt; &lt;monitoring&gt; &lt;/item&gt; &lt;item&gt; &lt;instanceId&gt;i-23a3397d&lt;/instanceId&gt; &lt;monitoring&gt; &lt;state&gt;pending&lt;state&gt; &lt;monitoring&gt; &lt;item&gt; &lt;instancesSet&gt; &lt;/MonitorInstancesResponse&gt;.

Request

data MonitorInstances

Request constructor

monitorInstances

Arguments

:: [Text]
miInstanceIds

-> MonitorInstances

Smart constructor for the minimum required parameters to construct a valid MonitorInstances request.

The fields accessible through corresponding lenses are:

InstanceIds :: [Text]

Request lenses

miInstanceIds :: Lens' MonitorInstances [Text]

One or more instance IDs.

Response

data MonitorInstancesResponse

Response constructor

monitorInstancesResponse :: MonitorInstancesResponse

Smart constructor for the minimum required parameters to construct a valid MonitorInstancesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

InstanceMonitorings :: [InstanceMonitoring]

Response lenses

mirInstanceMonitorings :: Lens' MonitorInstancesResponse [InstanceMonitoring]

Monitoring information for one or more instances.