Network.AWS.EC2.PurchaseHostReservation

Description

Purchase a reservation with configurations that match those of your Dedicated Host. You must have active Dedicated Hosts in your account before you purchase a reservation. This action results in the specified reservation being purchased and charged to your account.

Creating a Request

purchaseHostReservation

Arguments

:: Text
phrOfferingId

-> PurchaseHostReservation

Creates a value of PurchaseHostReservation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

phrCurrencyCode - The currency in which the totalUpfrontPrice , LimitPrice , and totalHourlyPrice amounts are specified. At this time, the only supported currency is USD .

phrClientToken - Unique, case-sensitive identifier you provide to ensure idempotency of the request. For more information, see How to Ensure Idempotency in the Amazon Elastic Compute Cloud User Guide .

phrLimitPrice - The specified limit is checked against the total upfront cost of the reservation (calculated as the offering's upfront cost multiplied by the host count). If the total upfront cost is greater than the specified price limit, the request will fail. This is used to ensure that the purchase does not exceed the expected upfront cost of the purchase. At this time, the only supported currency is USD . For example, to indicate a limit price of USD 100, specify 100.00.

phrHostIdSet - The IDs of the Dedicated Hosts that the reservation will be associated with.

phrOfferingId - The ID of the offering.

data PurchaseHostReservation

See: purchaseHostReservation smart constructor.

Request Lenses

phrCurrencyCode :: Lens' PurchaseHostReservation (Maybe CurrencyCodeValues)

The currency in which the totalUpfrontPrice , LimitPrice , and totalHourlyPrice amounts are specified. At this time, the only supported currency is USD .

phrClientToken :: Lens' PurchaseHostReservation (Maybe Text)

Unique, case-sensitive identifier you provide to ensure idempotency of the request. For more information, see How to Ensure Idempotency in the Amazon Elastic Compute Cloud User Guide .

phrLimitPrice :: Lens' PurchaseHostReservation (Maybe Text)

The specified limit is checked against the total upfront cost of the reservation (calculated as the offering's upfront cost multiplied by the host count). If the total upfront cost is greater than the specified price limit, the request will fail. This is used to ensure that the purchase does not exceed the expected upfront cost of the purchase. At this time, the only supported currency is USD . For example, to indicate a limit price of USD 100, specify 100.00.

phrHostIdSet :: Lens' PurchaseHostReservation [Text]

The IDs of the Dedicated Hosts that the reservation will be associated with.

phrOfferingId :: Lens' PurchaseHostReservation Text

The ID of the offering.

Destructuring the Response

purchaseHostReservationResponse

Arguments

:: Int
phrrsResponseStatus

-> PurchaseHostReservationResponse

Creates a value of PurchaseHostReservationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

phrrsCurrencyCode - The currency in which the totalUpfrontPrice and totalHourlyPrice amounts are specified. At this time, the only supported currency is USD .

phrrsClientToken - Unique, case-sensitive identifier you provide to ensure idempotency of the request. For more information, see How to Ensure Idempotency in the Amazon Elastic Compute Cloud User Guide

phrrsTotalHourlyPrice - The total hourly price of the reservation calculated per hour.

phrrsTotalUpfrontPrice - The total amount that will be charged to your account when you purchase the reservation.

phrrsPurchase - Describes the details of the purchase.

phrrsResponseStatus - -- | The response status code.

data PurchaseHostReservationResponse

See: purchaseHostReservationResponse smart constructor.

Response Lenses

phrrsCurrencyCode :: Lens' PurchaseHostReservationResponse (Maybe CurrencyCodeValues)

The currency in which the totalUpfrontPrice and totalHourlyPrice amounts are specified. At this time, the only supported currency is USD .

phrrsClientToken :: Lens' PurchaseHostReservationResponse (Maybe Text)

Unique, case-sensitive identifier you provide to ensure idempotency of the request. For more information, see How to Ensure Idempotency in the Amazon Elastic Compute Cloud User Guide

phrrsTotalHourlyPrice :: Lens' PurchaseHostReservationResponse (Maybe Text)

The total hourly price of the reservation calculated per hour.

phrrsTotalUpfrontPrice :: Lens' PurchaseHostReservationResponse (Maybe Text)

The total amount that will be charged to your account when you purchase the reservation.

phrrsPurchase :: Lens' PurchaseHostReservationResponse [Purchase]

Describes the details of the purchase.

phrrsResponseStatus :: Lens' PurchaseHostReservationResponse Int

- | The response status code.