Network.AWS.EC2.PurchaseReservedInstancesOffering

Description

Purchases a Reserved Instance for use with your account. With Amazon EC2 Reserved Instances, you obtain a capacity reservation for a certain instance configuration over a specified period of time. You pay a lower usage rate than with On-Demand instances for the time that you actually use the capacity reservation.

Use DescribeReservedInstancesOfferings to get a list of Reserved Instance offerings that match your specifications. After you've purchased a Reserved Instance, you can check for your new Reserved Instance with DescribeReservedInstances.

For more information, see Reserved Instances and Reserved InstanceMarketplace in the Amazon Elastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-PurchaseReservedInstancesOffering.html

Request

data PurchaseReservedInstancesOffering

Request constructor

purchaseReservedInstancesOffering

Arguments

:: Text
prioReservedInstancesOfferingId

-> Int
prioInstanceCount

-> PurchaseReservedInstancesOffering

PurchaseReservedInstancesOffering constructor.

The fields accessible through corresponding lenses are:

prioDryRun :: Maybe Bool

prioInstanceCount :: Int

prioLimitPrice :: Maybe ReservedInstanceLimitPrice

prioReservedInstancesOfferingId :: Text

Request lenses

prioDryRun :: Lens' PurchaseReservedInstancesOffering (Maybe Bool)

prioInstanceCount :: Lens' PurchaseReservedInstancesOffering Int

The number of Reserved Instances to purchase.

prioLimitPrice :: Lens' PurchaseReservedInstancesOffering (Maybe ReservedInstanceLimitPrice)

Specified for Reserved Instance Marketplace offerings to limit the total order and ensure that the Reserved Instances are not purchased at unexpected prices.

prioReservedInstancesOfferingId :: Lens' PurchaseReservedInstancesOffering Text

The ID of the Reserved Instance offering to purchase.

Response

data PurchaseReservedInstancesOfferingResponse

Response constructor

purchaseReservedInstancesOfferingResponse :: PurchaseReservedInstancesOfferingResponse

PurchaseReservedInstancesOfferingResponse constructor.

The fields accessible through corresponding lenses are:

priorReservedInstancesId :: Maybe Text

Response lenses

priorReservedInstancesId :: Lens' PurchaseReservedInstancesOfferingResponse (Maybe Text)

The IDs of the purchased Reserved Instances.