Network.AWS.EC2.RebootInstances

Description

Requests a reboot of one or more instances. This operation is asynchronous; it only queues a request to reboot the specified instances. The operation succeeds if the instances are valid and belong to you. Requests to reboot terminated instances are ignored.

If a Linux/Unix instance does not cleanly shut down within four minutes, Amazon EC2 performs a hard reboot.

For more information about troubleshooting, see Getting Console Output andRebooting Instances in the Amazon Elastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-RebootInstances.html

Request

data RebootInstances

Request constructor

rebootInstances :: RebootInstances

RebootInstances constructor.

The fields accessible through corresponding lenses are:

ri2DryRun :: Maybe Bool

ri2InstanceIds :: [Text]

Request lenses

ri2DryRun :: Lens' RebootInstances (Maybe Bool)

ri2InstanceIds :: Lens' RebootInstances [Text]

One or more instance IDs.

Response

data RebootInstancesResponse

Response constructor

rebootInstancesResponse :: RebootInstancesResponse

RebootInstancesResponse constructor.