Network.AWS.EC2.RegisterImage

Description

Registers an AMI. When you're creating an AMI, this is the final step you must complete before you can launch an instance from the AMI. For more information about creating AMIs, see Creating Your Own AMIs in the Amazon Elastic Compute Cloud User Guide. You can also use RegisterImage to create an Amazon EBS-backed AMI from a snapshot of a root device volume. For more information, see Launching an Instance from a Snapshot in the Amazon Elastic Compute Cloud User Guide. If needed, you can deregister an AMI at any time. Any modifications you make to an AMI backed by an instance store volume invalidates its registration. If you make changes to an image, deregister the previous image and register the new image.

Request

data RegisterImage

Request constructor

registerImage

Arguments

:: Text
riName

-> RegisterImage

RegisterImage constructor.

The fields accessible through corresponding lenses are:

riArchitecture :: Maybe Text

riBlockDeviceMappings :: [BlockDeviceMapping]

riDescription :: Maybe Text

riDryRun :: Maybe Bool

riImageLocation :: Maybe Text

riKernelId :: Maybe Text

riName :: Text

riRamdiskId :: Maybe Text

riRootDeviceName :: Maybe Text

riSriovNetSupport :: Maybe Text

riVirtualizationType :: Maybe Text

Request lenses

riArchitecture :: Lens' RegisterImage (Maybe Text)

The architecture of the AMI. Default: For Amazon EBS-backed AMIs, i386. For instance store-backed AMIs, the architecture specified in the manifest file.

riBlockDeviceMappings :: Lens' RegisterImage [BlockDeviceMapping]

One or more block device mapping entries.

riDescription :: Lens' RegisterImage (Maybe Text)

A description for your AMI.

riDryRun :: Lens' RegisterImage (Maybe Bool)

riImageLocation :: Lens' RegisterImage (Maybe Text)

The full path to your AMI manifest in Amazon S3 storage.

riKernelId :: Lens' RegisterImage (Maybe Text)

The ID of the kernel.

riName :: Lens' RegisterImage Text

A name for your AMI. Constraints: 3-128 alphanumeric characters, parentheses (()), square brackets ([]), spaces ( ), periods (.), slashes (/), dashes (-), single quotes ('), at-signs (@), or underscores(_).

riRamdiskId :: Lens' RegisterImage (Maybe Text)

The ID of the RAM disk.

riRootDeviceName :: Lens' RegisterImage (Maybe Text)

The name of the root device (for example, devsda1, or xvda).

riSriovNetSupport :: Lens' RegisterImage (Maybe Text)

Set to simple to enable enhanced networking for the AMI and any instances that you launch from the AMI. There is no way to disable enhanced networking at this time. This option is supported only for HVM AMIs. Specifying this option with a PV AMI can make instances launched from the AMI unreachable.

riVirtualizationType :: Lens' RegisterImage (Maybe Text)

The type of virtualization. Default: paravirtual.

Response

data RegisterImageResult

Response constructor

registerImageResult :: RegisterImageResult

RegisterImageResult constructor.

The fields accessible through corresponding lenses are:

rirImageId :: Maybe Text

Response lenses

rirImageId :: Lens' RegisterImageResult (Maybe Text)

The ID of the newly registered AMI.