Network.AWS.EC2.RejectVPCPeeringConnection

Description

Rejects a VPC peering connection request. The VPC peering connection must be in the pending-acceptance state. Use the DescribeVpcPeeringConnections request to view your outstanding VPC peering connection requests. To delete an active VPC peering connection, or to delete a VPC peering connection request that you initiated, use DeleteVpcPeeringConnection .

Creating a Request

rejectVPCPeeringConnection

Arguments

:: Text
rvpcVPCPeeringConnectionId

-> RejectVPCPeeringConnection

Creates a value of RejectVPCPeeringConnection with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rvpcDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rvpcVPCPeeringConnectionId - The ID of the VPC peering connection.

data RejectVPCPeeringConnection

Contains the parameters for RejectVpcPeeringConnection.

See: rejectVPCPeeringConnection smart constructor.

Request Lenses

rvpcDryRun :: Lens' RejectVPCPeeringConnection (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rvpcVPCPeeringConnectionId :: Lens' RejectVPCPeeringConnection Text

The ID of the VPC peering connection.

Destructuring the Response

rejectVPCPeeringConnectionResponse

Arguments

:: Int
rvpcrsResponseStatus

-> RejectVPCPeeringConnectionResponse

Creates a value of RejectVPCPeeringConnectionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rvpcrsReturn - Returns true if the request succeeds; otherwise, it returns an error.

rvpcrsResponseStatus - -- | The response status code.

data RejectVPCPeeringConnectionResponse

Contains the output of RejectVpcPeeringConnection.

See: rejectVPCPeeringConnectionResponse smart constructor.

Response Lenses

rvpcrsReturn :: Lens' RejectVPCPeeringConnectionResponse (Maybe Bool)

Returns true if the request succeeds; otherwise, it returns an error.

rvpcrsResponseStatus :: Lens' RejectVPCPeeringConnectionResponse Int

- | The response status code.