Network.AWS.EC2.ReleaseHosts

Description

When you no longer want to use an On-Demand Dedicated Host it can be released. On-Demand billing is stopped and the host goes into released state. The host ID of Dedicated Hosts that have been released can no longer be specified in another request, e.g., ModifyHosts. You must stop or terminate all instances on a host before it can be released.

When Dedicated Hosts are released, it make take some time for them to stop counting toward your limit and you may receive capacity errors when trying to allocate new Dedicated hosts. Try waiting a few minutes, and then try again.

Released hosts will still appear in a DescribeHosts response.

Creating a Request

releaseHosts :: ReleaseHosts

Creates a value of ReleaseHosts with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rhHostIds - The IDs of the Dedicated Hosts you want to release.

data ReleaseHosts

Contains the parameters for ReleaseHosts.

See: releaseHosts smart constructor.

Request Lenses

rhHostIds :: Lens' ReleaseHosts [Text]

The IDs of the Dedicated Hosts you want to release.

Destructuring the Response

releaseHostsResponse

Arguments

:: Int
rhrsResponseStatus

-> ReleaseHostsResponse

Creates a value of ReleaseHostsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rhrsUnsuccessful - The IDs of the Dedicated Hosts that could not be released, including an error message.

rhrsSuccessful - The IDs of the Dedicated Hosts that were successfully released.

rhrsResponseStatus - -- | The response status code.

data ReleaseHostsResponse

Contains the output of ReleaseHosts.

See: releaseHostsResponse smart constructor.

Response Lenses

rhrsUnsuccessful :: Lens' ReleaseHostsResponse [UnsuccessfulItem]

The IDs of the Dedicated Hosts that could not be released, including an error message.

rhrsSuccessful :: Lens' ReleaseHostsResponse [Text]

The IDs of the Dedicated Hosts that were successfully released.

rhrsResponseStatus :: Lens' ReleaseHostsResponse Int

- | The response status code.