Network.AWS.EC2.ReplaceIAMInstanceProfileAssociation

Description

Replaces an IAM instance profile for the specified running instance. You can use this action to change the IAM instance profile that's associated with an instance without having to disassociate the existing IAM instance profile first.

Use DescribeIamInstanceProfileAssociations to get the association ID.

Creating a Request

replaceIAMInstanceProfileAssociation

Arguments

:: IAMInstanceProfileSpecification
riapaIAMInstanceProfile

-> Text
riapaAssociationId

-> ReplaceIAMInstanceProfileAssociation

Creates a value of ReplaceIAMInstanceProfileAssociation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

riapaIAMInstanceProfile - The IAM instance profile.

riapaAssociationId - The ID of the existing IAM instance profile association.

data ReplaceIAMInstanceProfileAssociation

See: replaceIAMInstanceProfileAssociation smart constructor.

Request Lenses

riapaIAMInstanceProfile :: Lens' ReplaceIAMInstanceProfileAssociation IAMInstanceProfileSpecification

The IAM instance profile.

riapaAssociationId :: Lens' ReplaceIAMInstanceProfileAssociation Text

The ID of the existing IAM instance profile association.

Destructuring the Response

replaceIAMInstanceProfileAssociationResponse

Arguments

:: Int
riaparsResponseStatus

-> ReplaceIAMInstanceProfileAssociationResponse

Creates a value of ReplaceIAMInstanceProfileAssociationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

riaparsIAMInstanceProfileAssociation - Information about the IAM instance profile association.

riaparsResponseStatus - -- | The response status code.

data ReplaceIAMInstanceProfileAssociationResponse

See: replaceIAMInstanceProfileAssociationResponse smart constructor.

Response Lenses

riaparsIAMInstanceProfileAssociation :: Lens' ReplaceIAMInstanceProfileAssociationResponse (Maybe IAMInstanceProfileAssociation)

Information about the IAM instance profile association.

riaparsResponseStatus :: Lens' ReplaceIAMInstanceProfileAssociationResponse Int

- | The response status code.