Network.AWS.EC2.ReplaceNetworkACLAssociation

Description

Changes which network ACL a subnet is associated with. By default when you create a subnet, it's automatically associated with the default network ACL. For more information about network ACLs, see Network ACLs in the Amazon Virtual Private Cloud User Guide .

Creating a Request

replaceNetworkACLAssociation

Arguments

:: Text
rnaaAssociationId

-> Text
rnaaNetworkACLId

-> ReplaceNetworkACLAssociation

Creates a value of ReplaceNetworkACLAssociation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rnaaDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rnaaAssociationId - The ID of the current association between the original network ACL and the subnet.

rnaaNetworkACLId - The ID of the new network ACL to associate with the subnet.

data ReplaceNetworkACLAssociation

Contains the parameters for ReplaceNetworkAclAssociation.

See: replaceNetworkACLAssociation smart constructor.

Request Lenses

rnaaDryRun :: Lens' ReplaceNetworkACLAssociation (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rnaaAssociationId :: Lens' ReplaceNetworkACLAssociation Text

The ID of the current association between the original network ACL and the subnet.

rnaaNetworkACLId :: Lens' ReplaceNetworkACLAssociation Text

The ID of the new network ACL to associate with the subnet.

Destructuring the Response

replaceNetworkACLAssociationResponse

Arguments

:: Int
rnaarsResponseStatus

-> ReplaceNetworkACLAssociationResponse

Creates a value of ReplaceNetworkACLAssociationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rnaarsNewAssociationId - The ID of the new association.

rnaarsResponseStatus - -- | The response status code.

data ReplaceNetworkACLAssociationResponse

Contains the output of ReplaceNetworkAclAssociation.

See: replaceNetworkACLAssociationResponse smart constructor.

Response Lenses

rnaarsNewAssociationId :: Lens' ReplaceNetworkACLAssociationResponse (Maybe Text)

The ID of the new association.

rnaarsResponseStatus :: Lens' ReplaceNetworkACLAssociationResponse Int

- | The response status code.