Network.AWS.EC2.ReplaceNetworkAclEntry

Description

Replaces an entry (rule) in a network ACL. For more information about network ACLs, see Network ACLs in the Amazon Virtual Private Cloud User Guide. Example This example replaces the egress entry numbered 110 in the network ACL with ID acl-2cb85d45. The new rule denies egress traffic destined for anywhere (0.0.0.0/0) on TCP port 139. https:/ec2.amazonaws.com?Action=ReplaceNetworkAclEntry &amp;NetworkAclId=acl-2cb85d45 &amp;RuleNumber=110 &amp;Protocol=tcp &amp;RuleAction=deny &amp;Egress=true &amp;CidrBlock=0.0.0.0/0 &amp;PortRange.From=139 &amp;PortRange.To=139 &amp;AUTHPARAMS &lt;ReplaceNetworkAclEntryResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;ReplaceNetworkAclEntryResponse&gt;.

Request

data ReplaceNetworkAclEntry

Request constructor

replaceNetworkAclEntry

Arguments

:: Text
rnaeNetworkAclId

-> Integer
rnaeRuleNumber

-> Text
rnaeProtocol

-> RuleAction
rnaeRuleAction

-> Bool
rnaeEgress

-> Text
rnaeCidrBlock

-> ReplaceNetworkAclEntry

Smart constructor for the minimum required parameters to construct a valid ReplaceNetworkAclEntry request.

The fields accessible through corresponding lenses are:

NetworkAclId :: Text

RuleNumber :: Integer

Protocol :: Text

RuleAction :: RuleAction

Egress :: Bool

CidrBlock :: Text

IcmpTypeCode :: Maybe IcmpTypeCode

PortRange :: Maybe PortRange

Request lenses

rnaeNetworkAclId :: Lens' ReplaceNetworkAclEntry Text

The ID of the ACL.

rnaeRuleNumber :: Lens' ReplaceNetworkAclEntry Integer

The rule number of the entry to replace.

rnaeProtocol :: Lens' ReplaceNetworkAclEntry Text

The IP protocol. You can specify all or -1 to mean all protocols.

rnaeRuleAction :: Lens' ReplaceNetworkAclEntry RuleAction

Indicates whether to allow or deny the traffic that matches the rule.

rnaeEgress :: Lens' ReplaceNetworkAclEntry Bool

Indicates whether to replace the egress rule. Default: If no value is specified, we replace the ingress rule.

rnaeCidrBlock :: Lens' ReplaceNetworkAclEntry Text

The network range to allow or deny, in CIDR notation.

rnaeIcmpTypeCode :: Lens' ReplaceNetworkAclEntry (Maybe IcmpTypeCode)

ICMP protocol: The ICMP type and code.

rnaePortRange :: Lens' ReplaceNetworkAclEntry (Maybe PortRange)

TCP or UDP protocols: The range of ports the rule applies to.

Response

data ReplaceNetworkAclEntryResponse

Response constructor

replaceNetworkAclEntryResponse :: ReplaceNetworkAclEntryResponse

Smart constructor for the minimum required parameters to construct a valid ReplaceNetworkAclEntryResponse response.

This constructor is provided for convenience and testing purposes.