Network.AWS.EC2.ReplaceRoute

Description

Replaces an existing route within a route table in a VPC. You must provide only one of the following: Internet gateway or virtual private gateway, NAT instance, VPC peering connection, or network interface.

For more information about route tables, see Route Tables in the AmazonVirtual Private Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-ReplaceRoute.html

Request

data ReplaceRoute

Request constructor

replaceRoute

Arguments

:: Text
rrRouteTableId

-> Text
rrDestinationCidrBlock

-> ReplaceRoute

ReplaceRoute constructor.

The fields accessible through corresponding lenses are:

rrDestinationCidrBlock :: Text

rrDryRun :: Maybe Bool

rrGatewayId :: Maybe Text

rrInstanceId :: Maybe Text

rrNetworkInterfaceId :: Maybe Text

rrRouteTableId :: Text

rrVpcPeeringConnectionId :: Maybe Text

Request lenses

rrDestinationCidrBlock :: Lens' ReplaceRoute Text

The CIDR address block used for the destination match. The value you provide must match the CIDR of an existing route in the table.

rrDryRun :: Lens' ReplaceRoute (Maybe Bool)

rrGatewayId :: Lens' ReplaceRoute (Maybe Text)

The ID of an Internet gateway or virtual private gateway.

rrInstanceId :: Lens' ReplaceRoute (Maybe Text)

The ID of a NAT instance in your VPC.

rrNetworkInterfaceId :: Lens' ReplaceRoute (Maybe Text)

The ID of a network interface.

rrRouteTableId :: Lens' ReplaceRoute Text

The ID of the route table.

rrVpcPeeringConnectionId :: Lens' ReplaceRoute (Maybe Text)

The ID of a VPC peering connection.

Response

data ReplaceRouteResponse

Response constructor

replaceRouteResponse :: ReplaceRouteResponse

ReplaceRouteResponse constructor.