Network.AWS.EC2.ReplaceRouteTableAssociation

Description

Changes the route table associated with a given subnet in a VPC. After the operation completes, the subnet uses the routes in the new route table it's associated with. For more information about route tables, see Route Tables in the Amazon Virtual Private Cloud User Guide.

You can also use ReplaceRouteTableAssociation to change which table is the main route table in the VPC. You just specify the main route table's association ID and the route table to be the new main route table.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-ReplaceRouteTableAssociation.html

Request

data ReplaceRouteTableAssociation

Request constructor

replaceRouteTableAssociation

Arguments

:: Text
rrtaAssociationId

-> Text
rrtaRouteTableId

-> ReplaceRouteTableAssociation

ReplaceRouteTableAssociation constructor.

The fields accessible through corresponding lenses are:

rrtaAssociationId :: Text

rrtaDryRun :: Maybe Bool

rrtaRouteTableId :: Text

Request lenses

rrtaAssociationId :: Lens' ReplaceRouteTableAssociation Text

The association ID.

rrtaDryRun :: Lens' ReplaceRouteTableAssociation (Maybe Bool)

rrtaRouteTableId :: Lens' ReplaceRouteTableAssociation Text

The ID of the new route table to associate with the subnet.

Response

data ReplaceRouteTableAssociationResponse

Response constructor

replaceRouteTableAssociationResponse :: ReplaceRouteTableAssociationResponse

ReplaceRouteTableAssociationResponse constructor.

The fields accessible through corresponding lenses are:

rrtarNewAssociationId :: Maybe Text

Response lenses

rrtarNewAssociationId :: Lens' ReplaceRouteTableAssociationResponse (Maybe Text)

The ID of the new association.