Network.AWS.EC2.RequestSpotFleet

Description

Creates a Spot fleet request.

You can submit a single request that includes multiple launch specifications that vary by instance type, AMI, Availability Zone, or subnet.

By default, the Spot fleet requests Spot instances in the Spot pool where the price per unit is the lowest. Each launch specification can include its own instance weighting that reflects the value of the instance type to your application workload.

Alternatively, you can specify that the Spot fleet distribute the target capacity across the Spot pools included in its launch specifications. By ensuring that the Spot instances in your Spot fleet are in different Spot pools, you can improve the availability of your fleet.

For more information, see Spot Fleet Requests in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

requestSpotFleet

Arguments

:: SpotFleetRequestConfigData
rsfSpotFleetRequestConfig

-> RequestSpotFleet

Creates a value of RequestSpotFleet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsfDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rsfSpotFleetRequestConfig - The configuration for the Spot fleet request.

data RequestSpotFleet

Contains the parameters for RequestSpotFleet.

See: requestSpotFleet smart constructor.

Request Lenses

rsfDryRun :: Lens' RequestSpotFleet (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rsfSpotFleetRequestConfig :: Lens' RequestSpotFleet SpotFleetRequestConfigData

The configuration for the Spot fleet request.

Destructuring the Response

requestSpotFleetResponse

Arguments

:: Int
rsfrsResponseStatus

-> Text
rsfrsSpotFleetRequestId

-> RequestSpotFleetResponse

Creates a value of RequestSpotFleetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsfrsResponseStatus - -- | The response status code.

rsfrsSpotFleetRequestId - The ID of the Spot fleet request.

data RequestSpotFleetResponse

Contains the output of RequestSpotFleet.

See: requestSpotFleetResponse smart constructor.

Response Lenses

rsfrsResponseStatus :: Lens' RequestSpotFleetResponse Int

- | The response status code.

rsfrsSpotFleetRequestId :: Lens' RequestSpotFleetResponse Text

The ID of the Spot fleet request.