Network.AWS.EC2.RequestSpotInstances

Description

Creates a Spot Instance request. Spot Instances are instances that Amazon EC2 launches when the bid price that you specify exceeds the current Spot Price. Amazon EC2 periodically sets the Spot Price based on available Spot Instance capacity and current Spot Instance requests. For more information, see SpotInstance Requests in the Amazon Elastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-RequestSpotInstances.html

Request

data RequestSpotInstances

Request constructor

requestSpotInstances

Arguments

:: Text
rsiSpotPrice

-> RequestSpotInstances

RequestSpotInstances constructor.

The fields accessible through corresponding lenses are:

rsiAvailabilityZoneGroup :: Maybe Text

rsiDryRun :: Maybe Bool

rsiInstanceCount :: Maybe Int

rsiLaunchGroup :: Maybe Text

rsiLaunchSpecification :: Maybe RequestSpotLaunchSpecification

rsiSpotPrice :: Text

rsiType :: Maybe SpotInstanceType

rsiValidFrom :: Maybe UTCTime

rsiValidUntil :: Maybe UTCTime

Request lenses

rsiAvailabilityZoneGroup :: Lens' RequestSpotInstances (Maybe Text)

The user-specified name for a logical grouping of bids.

When you specify an Availability Zone group in a Spot Instance request, all Spot Instances in the request are launched in the same Availability Zone. Instance proximity is maintained with this parameter, but the choice of Availability Zone is not. The group applies only to bids for Spot Instances of the same instance type. Any additional Spot Instance requests that are specified with the same Availability Zone group name are launched in that same Availability Zone, as long as at least one instance from the group is still active.

If there is no active instance running in the Availability Zone group that you specify for a new Spot Instance request (all instances are terminated, the bid is expired, or the bid falls below current market), then Amazon EC2 launches the instance in any Availability Zone where the constraint can be met. Consequently, the subsequent set of Spot Instances could be placed in a different zone from the original request, even if you specified the same Availability Zone group.

Default: Instances are launched in any available Availability Zone.

rsiDryRun :: Lens' RequestSpotInstances (Maybe Bool)

rsiInstanceCount :: Lens' RequestSpotInstances (Maybe Int)

The maximum number of Spot Instances to launch.

Default: 1

rsiLaunchGroup :: Lens' RequestSpotInstances (Maybe Text)

The instance launch group. Launch groups are Spot Instances that launch together and terminate together.

Default: Instances are launched and terminated individually

rsiLaunchSpecification :: Lens' RequestSpotInstances (Maybe RequestSpotLaunchSpecification)

rsiSpotPrice :: Lens' RequestSpotInstances Text

The maximum hourly price (bid) for any Spot Instance launched to fulfill the request.

rsiType :: Lens' RequestSpotInstances (Maybe SpotInstanceType)

The Spot Instance request type.

Default: 'one-time'

rsiValidFrom :: Lens' RequestSpotInstances (Maybe UTCTime)

The start date of the request. If this is a one-time request, the request becomes active at this date and time and remains active until all instances launch, the request expires, or the request is canceled. If the request is persistent, the request becomes active at this date and time and remains active until it expires or is canceled.

Default: The request is effective indefinitely.

rsiValidUntil :: Lens' RequestSpotInstances (Maybe UTCTime)

The end date of the request. If this is a one-time request, the request remains active until all instances launch, the request is canceled, or this date is reached. If the request is persistent, it remains active until it is canceled or this date and time is reached.

Default: The request is effective indefinitely.

Response

data RequestSpotInstancesResponse

Response constructor

requestSpotInstancesResponse :: RequestSpotInstancesResponse

RequestSpotInstancesResponse constructor.

The fields accessible through corresponding lenses are:

rsirSpotInstanceRequests :: [SpotInstanceRequest]

Response lenses

rsirSpotInstanceRequests :: Lens' RequestSpotInstancesResponse [SpotInstanceRequest]

One or more Spot Instance requests.