Network.AWS.EC2.ResetImageAttribute

Description

Resets an attribute of an AMI to its default value.

Creating a Request

resetImageAttribute

Arguments

:: ResetImageAttributeName
resAttribute

-> Text
resImageId

-> ResetImageAttribute

Creates a value of ResetImageAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

resDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

resAttribute - The attribute to reset (currently you can only reset the launch permission attribute).

resImageId - The ID of the AMI.

data ResetImageAttribute

Contains the parameters for ResetImageAttribute.

See: resetImageAttribute smart constructor.

Request Lenses

resDryRun :: Lens' ResetImageAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

resAttribute :: Lens' ResetImageAttribute ResetImageAttributeName

The attribute to reset (currently you can only reset the launch permission attribute).

resImageId :: Lens' ResetImageAttribute Text

The ID of the AMI.

Destructuring the Response

resetImageAttributeResponse :: ResetImageAttributeResponse

Creates a value of ResetImageAttributeResponse with the minimum fields required to make a request.

data ResetImageAttributeResponse

See: resetImageAttributeResponse smart constructor.