Network.AWS.EC2.ResetInstanceAttribute

Description

Resets an attribute of an instance to its default value. To reset the kernel or ramdisk, the instance must be in a stopped state. To reset the SourceDestCheck, the instance can be either running or stopped. The SourceDestCheck attribute controls whether source/destination checking is enabled. The default value is true, which means checking is enabled. This value must be false for a NAT instance to perform NAT. For more information, see NAT Instances in the Amazon Virtual Private Cloud User Guide.

Request

data ResetInstanceAttribute

Request constructor

resetInstanceAttribute

Arguments

:: Text
ria1InstanceId

-> Text
ria1Attribute

-> ResetInstanceAttribute

ResetInstanceAttribute constructor.

The fields accessible through corresponding lenses are:

ria1Attribute :: Text

ria1DryRun :: Maybe Bool

ria1InstanceId :: Text

Request lenses

ria1Attribute :: Lens' ResetInstanceAttribute Text

The attribute to reset.

ria1DryRun :: Lens' ResetInstanceAttribute (Maybe Bool)

ria1InstanceId :: Lens' ResetInstanceAttribute Text

The ID of the instance.

Response

data ResetInstanceAttributeResponse

Response constructor

resetInstanceAttributeResponse :: ResetInstanceAttributeResponse

ResetInstanceAttributeResponse constructor.