Network.AWS.EC2.ResetNetworkInterfaceAttribute

Description

Resets a network interface attribute. You can specify only one attribute at a time.

Creating a Request

resetNetworkInterfaceAttribute

Arguments

:: Text
rniaNetworkInterfaceId

-> ResetNetworkInterfaceAttribute

Creates a value of ResetNetworkInterfaceAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rniaSourceDestCheck - The source/destination checking attribute. Resets the value to true .

rniaDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rniaNetworkInterfaceId - The ID of the network interface.

data ResetNetworkInterfaceAttribute

Contains the parameters for ResetNetworkInterfaceAttribute.

See: resetNetworkInterfaceAttribute smart constructor.

Request Lenses

rniaSourceDestCheck :: Lens' ResetNetworkInterfaceAttribute (Maybe Text)

The source/destination checking attribute. Resets the value to true .

rniaDryRun :: Lens' ResetNetworkInterfaceAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rniaNetworkInterfaceId :: Lens' ResetNetworkInterfaceAttribute Text

The ID of the network interface.

Destructuring the Response

resetNetworkInterfaceAttributeResponse :: ResetNetworkInterfaceAttributeResponse

Creates a value of ResetNetworkInterfaceAttributeResponse with the minimum fields required to make a request.

data ResetNetworkInterfaceAttributeResponse

See: resetNetworkInterfaceAttributeResponse smart constructor.