Network.AWS.EC2.ResetSnapshotAttribute

Description

Resets permission settings for the specified snapshot.

For more information on modifying snapshot permissions, see Sharing Snapshots in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

resetSnapshotAttribute

Arguments

:: SnapshotAttributeName
rsaAttribute

-> Text
rsaSnapshotId

-> ResetSnapshotAttribute

Creates a value of ResetSnapshotAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsaDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rsaAttribute - The attribute to reset. Currently, only the attribute for permission to create volumes can be reset.

rsaSnapshotId - The ID of the snapshot.

data ResetSnapshotAttribute

Contains the parameters for ResetSnapshotAttribute.

See: resetSnapshotAttribute smart constructor.

Request Lenses

rsaDryRun :: Lens' ResetSnapshotAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rsaAttribute :: Lens' ResetSnapshotAttribute SnapshotAttributeName

The attribute to reset. Currently, only the attribute for permission to create volumes can be reset.

rsaSnapshotId :: Lens' ResetSnapshotAttribute Text

The ID of the snapshot.

Destructuring the Response

resetSnapshotAttributeResponse :: ResetSnapshotAttributeResponse

Creates a value of ResetSnapshotAttributeResponse with the minimum fields required to make a request.

data ResetSnapshotAttributeResponse

See: resetSnapshotAttributeResponse smart constructor.