Network.AWS.EC2.RestoreAddressToClassic

Description

Restores an Elastic IP address that was previously moved to the EC2-VPC platform back to the EC2-Classic platform. You cannot move an Elastic IP address that was originally allocated for use in EC2-VPC. The Elastic IP address must not be associated with an instance or network interface.

Creating a Request

restoreAddressToClassic

Arguments

:: Text
ratcPublicIP

-> RestoreAddressToClassic

Creates a value of RestoreAddressToClassic with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ratcDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

ratcPublicIP - The Elastic IP address.

data RestoreAddressToClassic

Contains the parameters for RestoreAddressToClassic.

See: restoreAddressToClassic smart constructor.

Request Lenses

ratcDryRun :: Lens' RestoreAddressToClassic (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

ratcPublicIP :: Lens' RestoreAddressToClassic Text

The Elastic IP address.

Destructuring the Response

restoreAddressToClassicResponse

Arguments

:: Int
ratcrsResponseStatus

-> RestoreAddressToClassicResponse

Creates a value of RestoreAddressToClassicResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ratcrsStatus - The move status for the IP address.

ratcrsPublicIP - The Elastic IP address.

ratcrsResponseStatus - -- | The response status code.

data RestoreAddressToClassicResponse

Contains the output of RestoreAddressToClassic.

See: restoreAddressToClassicResponse smart constructor.

Response Lenses

ratcrsStatus :: Lens' RestoreAddressToClassicResponse (Maybe AddressStatus)

The move status for the IP address.

ratcrsPublicIP :: Lens' RestoreAddressToClassicResponse (Maybe Text)

The Elastic IP address.

ratcrsResponseStatus :: Lens' RestoreAddressToClassicResponse Int

- | The response status code.