Network.AWS.EC2.RevokeSecurityGroupEgress

Description

Removes one or more egress rules from a security group for EC2-VPC. The values that you specify in the revoke request (for example, ports) must match the existing rule's values for the rule to be revoked.

Each rule consists of the protocol and the CIDR range or source security group. For the TCP and UDP protocols, you must also specify the destination port or range of ports. For the ICMP protocol, you must also specify the ICMP type and code.

Rule changes are propagated to instances within the security group as quickly as possible. However, a small delay might occur.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-RevokeSecurityGroupEgress.html

Request

data RevokeSecurityGroupEgress

Request constructor

revokeSecurityGroupEgress

Arguments

:: Text
rsgeGroupId

-> RevokeSecurityGroupEgress

RevokeSecurityGroupEgress constructor.

The fields accessible through corresponding lenses are:

rsgeCidrIp :: Maybe Text

rsgeDryRun :: Maybe Bool

rsgeFromPort :: Maybe Int

rsgeGroupId :: Text

rsgeIpPermissions :: [IpPermission]

rsgeIpProtocol :: Maybe Text

rsgeSourceSecurityGroupName :: Maybe Text

rsgeSourceSecurityGroupOwnerId :: Maybe Text

rsgeToPort :: Maybe Int

Request lenses

rsgeCidrIp :: Lens' RevokeSecurityGroupEgress (Maybe Text)

The CIDR IP address range. You can't specify this parameter when specifying a source security group.

rsgeDryRun :: Lens' RevokeSecurityGroupEgress (Maybe Bool)

rsgeFromPort :: Lens' RevokeSecurityGroupEgress (Maybe Int)

The start of port range for the TCP and UDP protocols, or an ICMP type number. For the ICMP type number, use '-1' to specify all ICMP types.

rsgeGroupId :: Lens' RevokeSecurityGroupEgress Text

The ID of the security group.

rsgeIpPermissions :: Lens' RevokeSecurityGroupEgress [IpPermission]

A set of IP permissions. You can't specify a destination security group and a CIDR IP address range.

rsgeIpProtocol :: Lens' RevokeSecurityGroupEgress (Maybe Text)

The IP protocol name (tcp, udp, icmp) or number (see Protocol Numbers). Use '-1' to specify all.

rsgeSourceSecurityGroupName :: Lens' RevokeSecurityGroupEgress (Maybe Text)

EC2-Classic, default VPC

The name of the destination security group. You can't specify a destination security group and a CIDR IP address range.

rsgeSourceSecurityGroupOwnerId :: Lens' RevokeSecurityGroupEgress (Maybe Text)

The ID of the destination security group. You can't specify a destination security group and a CIDR IP address range.

rsgeToPort :: Lens' RevokeSecurityGroupEgress (Maybe Int)

The end of port range for the TCP and UDP protocols, or an ICMP code number. For the ICMP code number, use '-1' to specify all ICMP codes for the ICMP type.

Response

data RevokeSecurityGroupEgressResponse

Response constructor

revokeSecurityGroupEgressResponse :: RevokeSecurityGroupEgressResponse

RevokeSecurityGroupEgressResponse constructor.