Network.AWS.EC2.RevokeSecurityGroupIngress

Description

Removes one or more ingress rules from a security group. The values that you specify in the revoke request (for example, ports) must match the existing rule's values for the rule to be removed.

Each rule consists of the protocol and the CIDR range or source security group. For the TCP and UDP protocols, you must also specify the destination port or range of ports. For the ICMP protocol, you must also specify the ICMP type and code.

Rule changes are propagated to instances within the security group as quickly as possible. However, a small delay might occur.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-RevokeSecurityGroupIngress.html

Request

data RevokeSecurityGroupIngress

Request constructor

revokeSecurityGroupIngress :: RevokeSecurityGroupIngress

RevokeSecurityGroupIngress constructor.

The fields accessible through corresponding lenses are:

rsgiCidrIp :: Maybe Text

rsgiDryRun :: Maybe Bool

rsgiFromPort :: Maybe Int

rsgiGroupId :: Maybe Text

rsgiGroupName :: Maybe Text

rsgiIpPermissions :: [IpPermission]

rsgiIpProtocol :: Maybe Text

rsgiSourceSecurityGroupName :: Maybe Text

rsgiSourceSecurityGroupOwnerId :: Maybe Text

rsgiToPort :: Maybe Int

Request lenses

rsgiCidrIp :: Lens' RevokeSecurityGroupIngress (Maybe Text)

The CIDR IP address range. You can't specify this parameter when specifying a source security group.

rsgiDryRun :: Lens' RevokeSecurityGroupIngress (Maybe Bool)

rsgiFromPort :: Lens' RevokeSecurityGroupIngress (Maybe Int)

The start of port range for the TCP and UDP protocols, or an ICMP type number. For the ICMP type number, use '-1' to specify all ICMP types.

rsgiGroupId :: Lens' RevokeSecurityGroupIngress (Maybe Text)

The ID of the security group.

rsgiGroupName :: Lens' RevokeSecurityGroupIngress (Maybe Text)

EC2-Classic, default VPC

The name of the security group.

rsgiIpPermissions :: Lens' RevokeSecurityGroupIngress [IpPermission]

A set of IP permissions. You can't specify a source security group and a CIDR IP address range.

rsgiIpProtocol :: Lens' RevokeSecurityGroupIngress (Maybe Text)

The IP protocol name (tcp, udp, icmp) or number (see Protocol Numbers). Use '-1' to specify all.

rsgiSourceSecurityGroupName :: Lens' RevokeSecurityGroupIngress (Maybe Text)

EC2-Classic, default VPC

The name of the source security group. You can't specify a source security group and a CIDR IP address range.

rsgiSourceSecurityGroupOwnerId :: Lens' RevokeSecurityGroupIngress (Maybe Text)

The ID of the source security group. You can't specify a source security group and a CIDR IP address range.

rsgiToPort :: Lens' RevokeSecurityGroupIngress (Maybe Int)

The end of port range for the TCP and UDP protocols, or an ICMP code number. For the ICMP code number, use '-1' to specify all ICMP codes for the ICMP type.

Response

data RevokeSecurityGroupIngressResponse

Response constructor

revokeSecurityGroupIngressResponse :: RevokeSecurityGroupIngressResponse

RevokeSecurityGroupIngressResponse constructor.