Network.AWS.EC2.RunInstances

Description

Launches the specified number of instances using an AMI for which you have permissions. When you launch an instance, it enters the pending state. After the instance is ready for you, it enters the running state. To check the state of your instance, call DescribeInstances. If you don't specify a security group when launching an instance, Amazon EC2 uses the default security group. For more information, see Security Groups in the Amazon Elastic Compute Cloud User Guide. Linux instances have access to the public key of the key pair at boot. You can use this key to provide secure access to the instance. Amazon EC2 public images use this feature to provide secure access without passwords. For more information, see Key Pairs in the Amazon Elastic Compute Cloud User Guide. You can provide optional user data when launching an instance. For more information, see Instance Metadata in the Amazon Elastic Compute Cloud User Guide. If any of the AMIs have a product code attached for which the user has not subscribed, RunInstances fails. T2 instance types can only be launched into a VPC. If you do not have a default VPC, or if you do not specify a subnet ID in the request, RunInstances fails. For more information about troubleshooting, see What To Do If An Instance Immediately Terminates, and Troubleshooting Connecting to Your Instance in the Amazon Elastic Compute Cloud User Guide.

Request

data RunInstances

Request constructor

runInstances

Arguments

:: Text
ri1ImageId

-> Int
ri1MinCount

-> Int
ri1MaxCount

-> RunInstances

RunInstances constructor.

The fields accessible through corresponding lenses are:

ri1AdditionalInfo :: Maybe Text

ri1BlockDeviceMappings :: [BlockDeviceMapping]

ri1ClientToken :: Maybe Text

ri1DisableApiTermination :: Maybe Bool

ri1DryRun :: Maybe Bool

ri1EbsOptimized :: Maybe Bool

ri1IamInstanceProfile :: Maybe IamInstanceProfileSpecification

ri1ImageId :: Text

ri1InstanceInitiatedShutdownBehavior :: Maybe Text

ri1InstanceType :: Maybe Text

ri1KernelId :: Maybe Text

ri1KeyName :: Maybe Text

ri1MaxCount :: Int

ri1MinCount :: Int

ri1Monitoring :: Maybe RunInstancesMonitoringEnabled

ri1NetworkInterfaces :: [InstanceNetworkInterfaceSpecification]

ri1Placement :: Maybe Placement

ri1PrivateIpAddress :: Maybe Text

ri1RamdiskId :: Maybe Text

ri1SecurityGroupIds :: [Text]

ri1SecurityGroups :: [Text]

ri1SubnetId :: Maybe Text

ri1UserData :: Maybe Text

Request lenses

ri1AdditionalInfo :: Lens' RunInstances (Maybe Text)

Reserved.

ri1BlockDeviceMappings :: Lens' RunInstances [BlockDeviceMapping]

The block device mapping.

ri1ClientToken :: Lens' RunInstances (Maybe Text)

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency in the Amazon Elastic Compute Cloud User Guide. Constraints: Maximum 64 ASCII characters.

ri1DisableApiTermination :: Lens' RunInstances (Maybe Bool)

If you set this parameter to true, you can't terminate the instance using the Amazon EC2 console, CLI, or API; otherwise, you can. If you set this parameter to true and then later want to be able to terminate the instance, you must first change the value of the disableApiTermination attribute to false using ModifyInstanceAttribute. Alternatively, if you set InstanceInitiatedShutdownBehavior to terminate, you can terminate the instance by running the shutdown command from the instance. Default: false.

ri1DryRun :: Lens' RunInstances (Maybe Bool)

ri1EbsOptimized :: Lens' RunInstances (Maybe Bool)

Indicates whether the instance is optimized for EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal Amazon EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS-optimized instance. Default: false.

ri1IamInstanceProfile :: Lens' RunInstances (Maybe IamInstanceProfileSpecification)

The IAM instance profile.

ri1ImageId :: Lens' RunInstances Text

The ID of the AMI, which you can get by calling DescribeImages.

ri1InstanceInitiatedShutdownBehavior :: Lens' RunInstances (Maybe Text)

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown). Default: stop.

ri1InstanceType :: Lens' RunInstances (Maybe Text)

The instance type. For more information, see Instance Types in the Amazon Elastic Compute Cloud User Guide. Default: m1.small.

ri1KernelId :: Lens' RunInstances (Maybe Text)

The ID of the kernel. We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB in the Amazon Elastic Compute Cloud User Guide.

ri1KeyName :: Lens' RunInstances (Maybe Text)

The name of the key pair. You can create a key pair using CreateKeyPair or ImportKeyPair. If you launch an instance without specifying a key pair, you can't connect to the instance.

ri1MaxCount :: Lens' RunInstances Int

The maximum number of instances to launch. If you specify more instances than Amazon EC2 can launch in the target Availability Zone, Amazon EC2 launches the largest possible number of instances above MinCount. Constraints: Between 1 and the maximum number you're allowed for the specified instance type. For more information about the default limits, and how to request an increase, see How many instances can I run in Amazon EC2 in the Amazon EC2 General FAQ.

ri1MinCount :: Lens' RunInstances Int

The minimum number of instances to launch. If you specify a minimum that is more instances than Amazon EC2 can launch in the target Availability Zone, Amazon EC2 launches no instances. Constraints: Between 1 and the maximum number you're allowed for the specified instance type. For more information about the default limits, and how to request an increase, see How many instances can I run in Amazon EC2 in the Amazon EC2 General FAQ.

ri1Monitoring :: Lens' RunInstances (Maybe RunInstancesMonitoringEnabled)

The monitoring for the instance.

ri1NetworkInterfaces :: Lens' RunInstances [InstanceNetworkInterfaceSpecification]

One or more network interfaces.

ri1Placement :: Lens' RunInstances (Maybe Placement)

The placement for the instance.

ri1PrivateIpAddress :: Lens' RunInstances (Maybe Text)

EC2-VPC

The primary IP address. You must specify a value from the IP address range of the subnet. Only one private IP address can be designated as primary. Therefore, you can't specify this parameter if PrivateIpAddresses.n.Primary is set to true and PrivateIpAddresses.n.PrivateIpAddress is set to an IP address. Default: We select an IP address from the IP address range of the subnet.

ri1RamdiskId :: Lens' RunInstances (Maybe Text)

The ID of the RAM disk. We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB in the Amazon Elastic Compute Cloud User Guide.

ri1SecurityGroupIds :: Lens' RunInstances [Text]

One or more security group IDs. You can create a security group using CreateSecurityGroup. Default: Amazon EC2 uses the default security group.

ri1SecurityGroups :: Lens' RunInstances [Text]

EC2-Classic, default VPC

One or more security group names. For a nondefault VPC, you must use security group IDs instead. Default: Amazon EC2 uses the default security group.

ri1SubnetId :: Lens' RunInstances (Maybe Text)

EC2-VPC

The ID of the subnet to launch the instance into.

ri1UserData :: Lens' RunInstances (Maybe Text)

The Base64-encoded MIME user data for the instances.

Response

data Reservation

Response constructor

reservation :: Reservation

Reservation constructor.

The fields accessible through corresponding lenses are:

rGroups :: [GroupIdentifier]

rInstances :: [Instance]

rOwnerId :: Maybe Text

rRequesterId :: Maybe Text

rReservationId :: Maybe Text

Response lenses

rGroups :: Lens' Reservation [GroupIdentifier]

One or more security groups.

rInstances :: Lens' Reservation [Instance]

One or more instances.

rOwnerId :: Lens' Reservation (Maybe Text)

The ID of the AWS account that owns the reservation.

rRequesterId :: Lens' Reservation (Maybe Text)

The ID of the requester that launched the instances on your behalf (for example, AWS Management Console or Auto Scaling).

rReservationId :: Lens' Reservation (Maybe Text)

The ID of the reservation.