Network.AWS.EC2.StartInstances

Description

Starts an Amazon EBS-backed AMI that you've previously stopped.

Instances that use Amazon EBS volumes as their root devices can be quickly stopped and started. When an instance is stopped, the compute resources are released and you are not billed for hourly instance usage. However, your root partition Amazon EBS volume remains, continues to persist your data, and you are charged for Amazon EBS volume usage. You can restart your instance at any time. Each time you transition an instance from stopped to started, Amazon EC2 charges a full instance hour, even if transitions happen multiple times within a single hour.

Before stopping an instance, make sure it is in a state from which it can be restarted. Stopping an instance does not preserve data stored in RAM.

Performing this operation on an instance that uses an instance store as its root device returns an error.

For more information, see Stopping Instances in the Amazon Elastic ComputeCloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-StartInstances.html

Request

data StartInstances

Request constructor

startInstances :: StartInstances

StartInstances constructor.

The fields accessible through corresponding lenses are:

si1AdditionalInfo :: Maybe Text

si1DryRun :: Maybe Bool

si1InstanceIds :: [Text]

Request lenses

si1AdditionalInfo :: Lens' StartInstances (Maybe Text)

Reserved.

si1DryRun :: Lens' StartInstances (Maybe Bool)

si1InstanceIds :: Lens' StartInstances [Text]

One or more instance IDs.

Response

data StartInstancesResponse

Response constructor

startInstancesResponse :: StartInstancesResponse

StartInstancesResponse constructor.

The fields accessible through corresponding lenses are:

sirStartingInstances :: [InstanceStateChange]

Response lenses

sirStartingInstances :: Lens' StartInstancesResponse [InstanceStateChange]

Information about one or more started instances.