Network.AWS.EC2.TerminateInstances

Description

Shuts down one or more instances. This operation is idempotent; if you terminate an instance more than once, each call succeeds.

Terminated instances remain visible after termination (for approximately one hour).

By default, Amazon EC2 deletes all Amazon EBS volumes that were attached when the instance launched. Volumes attached after instance launch continue running.

You can stop, start, and terminate EBS-backed instances. You can only terminate instance store-backed instances. What happens to an instance differs if you stop it or terminate it. For example, when you stop an instance, the root device and any other devices attached to the instance persist. When you terminate an instance, the root device and any other devices attached during the instance launch are automatically deleted. For more information about the differences between stopping and terminating instances, see Instance Lifecycle in the Amazon Elastic Compute Cloud UserGuide for Linux.

For more information about troubleshooting, see Troubleshooting TerminatingYour Instance in the Amazon Elastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-TerminateInstances.html

Request

data TerminateInstances

Request constructor

terminateInstances :: TerminateInstances

TerminateInstances constructor.

The fields accessible through corresponding lenses are:

tiDryRun :: Maybe Bool

tiInstanceIds :: [Text]

Request lenses

tiDryRun :: Lens' TerminateInstances (Maybe Bool)

tiInstanceIds :: Lens' TerminateInstances [Text]

One or more instance IDs.

Response

data TerminateInstancesResponse

Response constructor

terminateInstancesResponse :: TerminateInstancesResponse

TerminateInstancesResponse constructor.

The fields accessible through corresponding lenses are:

tirTerminatingInstances :: [InstanceStateChange]

Response lenses

tirTerminatingInstances :: Lens' TerminateInstancesResponse [InstanceStateChange]

Information about one or more terminated instances.