Network.AWS.EC2.UnassignIPv6Addresses

Description

Unassigns one or more IPv6 addresses from a network interface.

Creating a Request

unassignIPv6Addresses

Arguments

:: Text
uiaNetworkInterfaceId

-> UnassignIPv6Addresses

Creates a value of UnassignIPv6Addresses with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uiaIPv6Addresses - The IPv6 addresses to unassign from the network interface.

uiaNetworkInterfaceId - The ID of the network interface.

data UnassignIPv6Addresses

See: unassignIPv6Addresses smart constructor.

Request Lenses

uiaIPv6Addresses :: Lens' UnassignIPv6Addresses [Text]

The IPv6 addresses to unassign from the network interface.

uiaNetworkInterfaceId :: Lens' UnassignIPv6Addresses Text

The ID of the network interface.

Destructuring the Response

unassignIPv6AddressesResponse

Arguments

:: Int
uiarsResponseStatus

-> UnassignIPv6AddressesResponse

Creates a value of UnassignIPv6AddressesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uiarsNetworkInterfaceId - The ID of the network interface.

uiarsUnassignedIPv6Addresses - The IPv6 addresses that have been unassigned from the network interface.

uiarsResponseStatus - -- | The response status code.

data UnassignIPv6AddressesResponse

See: unassignIPv6AddressesResponse smart constructor.

Response Lenses

uiarsNetworkInterfaceId :: Lens' UnassignIPv6AddressesResponse (Maybe Text)

The ID of the network interface.

uiarsUnassignedIPv6Addresses :: Lens' UnassignIPv6AddressesResponse [Text]

The IPv6 addresses that have been unassigned from the network interface.

uiarsResponseStatus :: Lens' UnassignIPv6AddressesResponse Int

- | The response status code.