Network.AWS.EC2.UnassignPrivateIPAddresses

Description

Unassigns one or more secondary private IP addresses from a network interface.

Creating a Request

unassignPrivateIPAddresses

Arguments

:: Text
upiaNetworkInterfaceId

-> UnassignPrivateIPAddresses

Creates a value of UnassignPrivateIPAddresses with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

upiaNetworkInterfaceId - The ID of the network interface.

upiaPrivateIPAddresses - The secondary private IP addresses to unassign from the network interface. You can specify this option multiple times to unassign more than one IP address.

data UnassignPrivateIPAddresses

Contains the parameters for UnassignPrivateIpAddresses.

See: unassignPrivateIPAddresses smart constructor.

Request Lenses

upiaNetworkInterfaceId :: Lens' UnassignPrivateIPAddresses Text

The ID of the network interface.

upiaPrivateIPAddresses :: Lens' UnassignPrivateIPAddresses [Text]

The secondary private IP addresses to unassign from the network interface. You can specify this option multiple times to unassign more than one IP address.

Destructuring the Response

unassignPrivateIPAddressesResponse :: UnassignPrivateIPAddressesResponse

Creates a value of UnassignPrivateIPAddressesResponse with the minimum fields required to make a request.

data UnassignPrivateIPAddressesResponse

See: unassignPrivateIPAddressesResponse smart constructor.