Network.AWS.EC2.UnassignPrivateIpAddresses

Description

Unassigns one or more secondary private IP addresses from a network interface. Example The following example unassigns two secondary private IP addresses from the specified network interface. https:/ec2.amazonaws.com?Action=UnassignPrivateIpAddresses &amp;NetworkInterfaceId=eni-197d9972 &amp;PrivateIpAddress.0=10.0.2.60 &amp;PrivateIpAddress.1=10.0.2.65 &amp;AUTHPARAMS &lt;UnassignPrivateIpAddresses xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;UnassignPrivateIpAddresses&gt;.

Request

data UnassignPrivateIpAddresses

Request constructor

unassignPrivateIpAddresses

Arguments

:: Text
upiaNetworkInterfaceId

-> [Text]
upiaPrivateIpAddresses

-> UnassignPrivateIpAddresses

Smart constructor for the minimum required parameters to construct a valid UnassignPrivateIpAddresses request.

The fields accessible through corresponding lenses are:

NetworkInterfaceId :: Text

PrivateIpAddresses :: [Text]

Request lenses

upiaNetworkInterfaceId :: Lens' UnassignPrivateIpAddresses Text

The ID of the network interface.

upiaPrivateIpAddresses :: Lens' UnassignPrivateIpAddresses [Text]

The secondary private IP addresses to unassign from the network interface. You can specify this option multiple times to unassign more than one IP address.

Response

data UnassignPrivateIpAddressesResponse

Response constructor

unassignPrivateIpAddressesResponse :: UnassignPrivateIpAddressesResponse

Smart constructor for the minimum required parameters to construct a valid UnassignPrivateIpAddressesResponse response.

This constructor is provided for convenience and testing purposes.