Network.AWS.EC2.V2014_06_15.AssociateDhcpOptions

Description

Associates a set of DHCP options (that you've previously created) with the specified VPC, or associates no DHCP options with the VPC. After you associate the options with the VPC, any existing instances and all new instances that you launch in that VPC use the options. You don't need to restart or relaunch the instances. They automatically pick up the changes within a few hours, depending on how frequently the instance renews its DHCP lease. You can explicitly renew the lease using the operating system on the instance. For more information, see DHCP Options Sets in the Amazon Virtual Private Cloud User Guide. Example 1 This example associates the DHCP options with the ID dopt-7a8b9c2d with the VPC with the ID vpc-1a2b3c4d. https:/ec2.amazonaws.com?Action=AssociateDhcpOptions &amp;DhcpOptionsId=dopt-7a8b9c2d &amp;VpcId=vpc-1a2b3c4d &amp;AUTHPARAMS &lt;AssociateDhcpOptionsResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;AssociateDhcpOptionsResponse&gt; Example 2 This example changes the VPC with the ID vpc-1a2b3c4d to have no associated DHCP options set. https:/ec2.amazonaws.com?Action=AssociateDhcpOptions &amp;DhcpOptionsId=default &amp;VpcId=vpc-1a2b3c4d &amp;AUTHPARAMS &lt;AssociateDhcpOptionsResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;AssociateDhcpOptionsResponse&gt;.

Request

data AssociateDhcpOptions

Request constructor

mkAssociateDhcpOptions

Arguments

:: Text
adoDhcpOptionsId

-> Text
adoVpcId

-> AssociateDhcpOptions

Smart constructor for the minimum required parameters to construct a valid AssociateDhcpOptions request.

Request lenses

adoDhcpOptionsId :: Lens' AssociateDhcpOptions Text

The ID of the DHCP options set, or default to associate no DHCP options with the VPC.

adoVpcId :: Lens' AssociateDhcpOptions Text

The ID of the VPC.

Response

data AssociateDhcpOptionsResponse