Network.AWS.EC2.V2014_06_15.AssociateRouteTable

Description

Associates a subnet with a route table. The subnet and route table must be in the same VPC. This association causes traffic originating from the subnet to be routed according to the routes in the route table. The action returns an association ID, which you need in order to disassociate the route table from the subnet later. A route table can be associated with multiple subnets. For more information about route tables, see Route Tables in the Amazon Virtual Private Cloud User Guide. Example This example associates a route table with the ID rtb-e4ad488d with a subnet with the ID subnet-15ad487c. https:/ec2.amazonaws.com?Action=AssociateRouteTable &amp;RouteTableId=rtb-e4ad488d &amp;SubnetId=subnet-15ad487c &lt;AssociateRouteTableResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;associationId&gt;rtbassoc-f8ad4891&lt;/associationId&gt; &lt;/AssociateRouteTableResponse&gt;.

Request

data AssociateRouteTable

Request constructor

mkAssociateRouteTable

Arguments

:: Text
artSubnetId

-> Text
artRouteTableId

-> AssociateRouteTable

Smart constructor for the minimum required parameters to construct a valid AssociateRouteTable request.

Request lenses

artSubnetId :: Lens' AssociateRouteTable Text

The ID of the subnet.

artRouteTableId :: Lens' AssociateRouteTable Text

The ID of the route table.

Response

data AssociateRouteTableResponse

Response lenses

artrsAssociationId :: Lens' AssociateRouteTableResponse (Maybe Text)

The route table association ID (needed to disassociate the route table).