Network.AWS.EC2.V2014_06_15.AttachInternetGateway

Description

Attaches an Internet gateway to a VPC, enabling connectivity between the Internet and the VPC. For more information about your VPC and Internet gateway, see the Amazon Virtual Private Cloud User Guide. Example This example attaches the Internet gateway with the ID igw-eaad4883 to the VPC with the ID vpc-11ad4878. https:/ec2.amazonaws.com?Action=AttachInternetGateway &amp;InternetGatewayId=igw-eaad4883 &amp;VpcId=vpc-11ad4878 &amp;AUTHPARAMS &lt;AttachInternetGatewayResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;AttachInternetGatewayResponse&gt;.

Documentation

attachInternetGateway

Arguments

:: Text
_aigrVpcId

-> Text
_aigrInternetGatewayId

-> AttachInternetGateway

Minimum specification for a AttachInternetGateway request.

data AttachInternetGateway

Constructors

AttachInternetGateway

Fields

_aigrVpcId :: Text

The ID of the VPC.

_aigrInternetGatewayId :: Text

The ID of the Internet gateway.

_aigrDryRun :: Maybe Bool

aigrVpcId :: Lens' AttachInternetGateway Text

aigrInternetGatewayId :: Lens' AttachInternetGateway Text

aigrDryRun :: Lens' AttachInternetGateway (Maybe Bool)

data AttachInternetGatewayResponse

Constructors

AttachInternetGatewayResponse