Network.AWS.EC2.V2014_06_15.AttachNetworkInterface

Description

Attaches a network interface to an instance. Example This example attaches the specified network interface to the specified instance. https:/ec2.amazonaws.com?Action=AttachNetworkInterface &amp;DeviceIndex=1 &amp;InstanceId=i-9cc316fe &amp;NetworkInterfaceId=eni-ffda3197 &amp;AUTHPARAMS &lt;AttachNetworkInterfaceResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;ace8cd1e-e685-4e44-90fb-92014d907212&lt;/requestId&gt; &lt;attachmentId&gt;eni-attach-d94b09b0&lt;/attachmentId&gt; &lt;/AttachNetworkInterfaceResponse&gt;.

Documentation

attachNetworkInterface

Arguments

:: Integer
_anirDeviceIndex

-> Text
_anirInstanceId

-> Text
_anirNetworkInterfaceId

-> AttachNetworkInterface

Minimum specification for a AttachNetworkInterface request.

data AttachNetworkInterface

Constructors

AttachNetworkInterface

Fields

_anirDeviceIndex :: Integer

The index of the device for the network interface attachment.

_anirInstanceId :: Text

The ID of the instance.

_anirNetworkInterfaceId :: Text

The ID of the network interface.

_anirDryRun :: Maybe Bool

anirNetworkInterfaceId :: Lens' AttachNetworkInterface Text

anirInstanceId :: Lens' AttachNetworkInterface Text

anirDryRun :: Lens' AttachNetworkInterface (Maybe Bool)

anirDeviceIndex :: Lens' AttachNetworkInterface Integer

data AttachNetworkInterfaceResponse

Constructors

AttachNetworkInterfaceResponse

Fields

_anisAttachmentId :: Maybe Text

The ID of the network interface attachment.

anisAttachmentId :: Iso' AttachNetworkInterfaceResponse (Maybe Text)