Network.AWS.EC2.V2014_06_15.CancelBundleTask

Description

Cancels a bundling operation for an instance store-backed Windows instance. Example This example request cancels the specified bundle task. https:/ec2.amazonaws.com?Action=CancelBundleTask &amp;BundleId=bun-cla322b9 &amp;AUTHPARAMS &lt;CancelBundleTaskResponse xmlns="http://ec2.amazonaws.com/doc/2014-02-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;bundleInstanceTask&gt; &lt;instanceId&gt;i-12345678&lt;/instanceId&gt; &lt;bundleId&gt;bun-cla322b9&lt;/bundleId&gt; &lt;state&gt;canceling&lt;/state&gt; &lt;startTime&gt;2008-10-07T11:41:50.000Z&lt;/startTime&gt; &lt;updateTime&gt;2008-10-07T11:51:50.000Z&lt;/updateTime&gt; &lt;progress&gt;20%&lt;/progress&gt; &lt;storage&gt; &lt;S3&gt; &lt;bucket&gt;myawsbucket&lt;/bucket&gt; &lt;prefix&gt;my-new-image&lt;prefix&gt; &lt;S3&gt; &lt;/storage&gt; &lt;bundleInstanceTask&gt; &lt;CancelBundleTaskResponse&gt;.

Request

data CancelBundleTask

Request constructor

mkCancelBundleTask

Arguments

:: Text
cbtBundleId

-> CancelBundleTask

Smart constructor for the minimum required parameters to construct a valid CancelBundleTask request.

Request lenses

cbtBundleId :: Lens' CancelBundleTask Text

The ID of the bundle task.

Response

data CancelBundleTaskResponse

Response lenses

cbtrBundleTask :: Lens' CancelBundleTaskResponse (Maybe BundleTask)

The bundle task.