Network.AWS.EC2.V2014_06_15.CancelConversionTask

Description

Cancels an active conversion task. The task can be the import of an instance or volume. The action removes all artifacts of the conversion, including a partially uploaded volume or instance. If the conversion is complete or is in the process of transferring the final disk image, the command fails and returns an exception. For more information, see Using the Command Line Tools to Import Your Virtual Machine to Amazon EC2 in the Amazon Elastic Compute Cloud User Guide. Example This example request cancels the conversion task with the ID import-i-fh95npoc. https:/ec2.amazonaws.com?Action=CancelConversionTask &amp;ConversionTaskId=import-i-fh95npoc &amp;AUTHPARAMS 59dbff89-35bd-4eac-99ed-be587EXAMPLE true.

Documentation

cancelConversionTask

Arguments

:: Text
_ccrConversionTaskId

-> CancelConversionTask

Minimum specification for a CancelConversionTask request.

data CancelConversionTask

Constructors

CancelConversionTask

Fields

_ccrConversionTaskId :: Text

The ID of the conversion task.

_ccrDryRun :: Maybe Bool

_ccrReasonMessage :: Maybe Text

ccrReasonMessage :: Lens' CancelConversionTask (Maybe Text)

ccrDryRun :: Lens' CancelConversionTask (Maybe Bool)

ccrConversionTaskId :: Lens' CancelConversionTask Text

data CancelConversionTaskResponse

Constructors

CancelConversionTaskResponse