Network.AWS.EC2.V2014_06_15.CancelExportTask

Description

Cancels an active export task. The request removes all artifacts of the export, including any partially-created Amazon S3 objects. If the export task is complete or is in the process of transferring the final disk image, the command fails and returns an error. Example This example request cancels the export task with the ID export-i-1234wxyz. https:/ec2.amazonaws.com?Action=CancelExportTask &amp;exportTaskId=export-i-1234wxyz &amp;AUTHPARAMS 59dbff89-35bd-4eac-99ed-be587EXAMPLE true.

Documentation

data CancelExportTask

Constructors

CancelExportTask

Fields

_cetrExportTaskId :: Text

The ID of the export task. This is the ID returned by CreateInstanceExportTask.

cetrExportTaskId :: Iso' CancelExportTask Text

data CancelExportTaskResponse

Constructors

CancelExportTaskResponse