Network.AWS.EC2.V2014_06_15.CancelReservedInstancesListing

Description

Cancels the specified Reserved Instance listing in the Reserved Instance Marketplace. For more information, see Reserved Instance Marketplace in the Amazon Elastic Compute Cloud User Guide. Example This example request cancels a Reserved Instance listing in the Reserved Instance Marketplace. https:/ec2.amazonaws.com?Action=CancelReservedInstancesListing &amp;ReservedInstancesListingId=3ebe97b5-f273-43b6-a204-7a18cEXAMPLE &amp;AUTHPARAMS bec2cf62-98ef-434a-8a15-886fcexample 3ebe97b5-f273-43b6-a204-7a18cEXAMPLE e5a2ff3b-7d14-494f-90af-0b5d0EXAMPLE 2012-07-12T16:55:28.000Z 2012-07-12T16:55:28.000Z cancelled CANCELLED Available 0 Sold 0 Cancelled 1 Pending 0 5 166.64 USD false 4 133.32 USD false 3 99.99 USD false 2 66.66 USD false 1 33.33 USD false XqJIt1342112125076.

Documentation

data CancelReservedInstancesListing

Constructors

CancelReservedInstancesListing

Fields

_crilrReservedInstancesListingId :: Text

The ID of the Reserved Instance listing.

crilrReservedInstancesListingId :: Iso' CancelReservedInstancesListing Text

data CancelReservedInstancesListingResponse

Constructors

CancelReservedInstancesListingResponse

Fields

_crilsReservedInstancesListings :: [ReservedInstancesListing]

The Reserved Instance listing.

crilsReservedInstancesListings :: Iso' CancelReservedInstancesListingResponse [ReservedInstancesListing]