Network.AWS.EC2.V2014_06_15.CancelSpotInstanceRequests

Description

Cancels one or more Spot Instance requests. Spot Instances are instances that Amazon EC2 starts on your behalf when the maximum price that you specify exceeds the current Spot Price. Amazon EC2 periodically sets the Spot Price based on available Spot Instance capacity and current Spot Instance requests. For more information about Spot Instances, see Spot Instances in the Amazon Elastic Compute Cloud User Guide. Canceling a Spot Instance request does not terminate running Spot Instances associated with the request. Example This example cancels the specified Spot Instance request. https:/ec2.amazonaws.com?Action=CancelSpotInstanceRequests &amp;SpotInstanceRequestId.1=sir-1a2b3c4d &amp;AUTHPARAMS &lt;CancelSpotInstanceRequestsResponse xmlns="http://ec2.amazonaws.com/doc/2013-06-15/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;spotInstanceRequestSet&gt; &lt;item&gt; &lt;spotInstanceRequestId&gt;sir-1a2b3c4d&lt;/spotInstanceRequestId&gt; &lt;state&gt;cancelled&lt;state&gt; &lt;item&gt; &lt;/spotInstanceRequestSet&gt; &lt;/CancelSpotInstanceRequestsResponse&gt;.

Request

data CancelSpotInstanceRequests

Request constructor

mkCancelSpotInstanceRequests

Arguments

:: [Text]
csirSpotInstanceRequestIds

-> CancelSpotInstanceRequests

Smart constructor for the minimum required parameters to construct a valid CancelSpotInstanceRequests request.

Request lenses

csirSpotInstanceRequestIds :: Lens' CancelSpotInstanceRequests [Text]

One or more Spot Instance request IDs.

Response

data CancelSpotInstanceRequestsResponse

Response lenses

csirrsCancelledSpotInstanceRequests :: Lens' CancelSpotInstanceRequestsResponse [CancelledSpotInstanceRequest]

One or more Spot Instance requests.