Network.AWS.EC2.V2014_06_15.CopyImage

Description

Initiates the copy of an AMI from the specified source region to the region in which the request was made. You specify the destination region by using its endpoint when making the request. AMIs that use encrypted Amazon EBS snapshots cannot be copied with this method. For more information, see Copying AMIs in the Amazon Elastic Compute Cloud User Guide. Example This example request copies the AMI in us-west-2 with the ID ami-1a2b3c4d, naming the new AMI My-Standard-AMI. https:/ec2.amazonaws.com?Action=CopyImage &amp;SourceRegion=us-west-2 &amp;SourceImageId=ami-1a2b3c4d &amp;Name=My-Standard-AMI &amp;Description=This%20is%20the%20new%20version%20of%20My-Standard-AMI &amp;ClientToken=550e8400-e29b-41d4-a716-446655440000 &amp;AUTHPARAMS &lt;CopyImageResponse xmlns="http://ec2.amazonaws.com/doc/2014-02-01/"&gt; &lt;requestId&gt;60bc441d-fa2c-494d-b155-5d6a3EXAMPLE&lt;/requestId&gt; &lt;imageId&gt;ami-4d3c2b1a&lt;imageId&gt; &lt;CopyImageResponse&gt;.

Request

data CopyImage

Request constructor

mkCopyImage

Arguments

:: Text
ciSourceRegion

-> Text
ciSourceImageId

-> CopyImage

Smart constructor for the minimum required parameters to construct a valid CopyImage request.

The fields accessible through corresponding lenses are:

SourceRegion :: Text

SourceImageId :: Text

Name :: Maybe Text

Description :: Maybe Text

ClientToken :: Maybe Text

Request lenses

ciSourceRegion :: Lens' CopyImage Text

The name of the region that contains the AMI to copy.

ciSourceImageId :: Lens' CopyImage Text

The ID of the AMI to copy.

ciName :: Lens' CopyImage (Maybe Text)

The name of the new AMI in the destination region.

ciDescription :: Lens' CopyImage (Maybe Text)

A description for the new AMI in the destination region.

ciClientToken :: Lens' CopyImage (Maybe Text)

Unique, case-sensitive identifier you provide to ensure idempotency of the request. For more information, see How to Ensure Idempotency in the Amazon Elastic Compute Cloud User Guide.

Response

data CopyImageResponse

Response constructor

mkCopyImageResponse :: CopyImageResponse

Smart constructor for the minimum required parameters to construct a valid CopyImageResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ImageId :: Maybe Text

Response lenses

cirImageId :: Lens' CopyImageResponse (Maybe Text)

The ID of the new AMI.