Network.AWS.EC2.V2014_06_15.CreateInternetGateway

Description

Creates an Internet gateway for use with a VPC. After creating the Internet gateway, you attach it to a VPC using AttachInternetGateway. For more information about your VPC and Internet gateway, see the Amazon Virtual Private Cloud User Guide. Example This example creates an Internet gateway. https:/ec2.amazonaws.com?Action=CreateInternetGateway &amp;AUTHPARAMS 59dbff89-35bd-4eac-99ed-be587EXAMPLE igw-eaad4883.

Request

data CreateInternetGateway

Request constructor

mkCreateInternetGateway :: CreateInternetGateway

Smart constructor for the minimum required parameters to construct a valid CreateInternetGateway request.

Response

data CreateInternetGatewayResponse

Response lenses

cigrsInternetGateway :: Lens' CreateInternetGatewayResponse (Maybe InternetGateway)

Information about the Internet gateway.