Network.AWS.EC2.V2014_06_15.CreateNetworkAcl

Description

Creates a network ACL in a VPC. Network ACLs provide an optional layer of security (in addition to security groups) for the instances in your VPC. For more information about network ACLs, see Network ACLs in the Amazon Virtual Private Cloud User Guide. Example This example creates a network ACL in the specified VPC. The response includes a default entry for egress, and another for ingress, each with a very high rule number. These are the last entries we process to decide whether traffic is allowed in or out of an associated subnet. If the traffic doesn't match any rules with a lower rule number, then these default entries ultimately deny the traffic. https:/ec2.amazonaws.com?Action=CreateNetworkAcl &amp;VpcId=vpc-11ad4878 &amp;AUTHPARAMS 59dbff89-35bd-4eac-99ed-be587EXAMPLE acl-5fb85d36 vpc-11ad4878 false 32767 all deny true 0.0.0.0/0 32767 all deny false 0.0.0.0/0.

Documentation

createNetworkAcl

Arguments

:: Text
_cnarVpcId

-> CreateNetworkAcl

Minimum specification for a CreateNetworkAcl request.

data CreateNetworkAcl

Constructors

CreateNetworkAcl

Fields

_cnarVpcId :: Text

The ID of the VPC.

_cnarDryRun :: Maybe Bool

cnarVpcId :: Lens' CreateNetworkAcl Text

cnarDryRun :: Lens' CreateNetworkAcl (Maybe Bool)

data CreateNetworkAclResponse

Constructors

CreateNetworkAclResponse

Fields

_cnasNetworkAcl :: Maybe NetworkAcl

Information about the network ACL.

cnasNetworkAcl :: Iso' CreateNetworkAclResponse (Maybe NetworkAcl)