Network.AWS.EC2.V2014_06_15.CreatePlacementGroup

Description

Creates a placement group that you launch cluster instances into. You must give the group a name that's unique within the scope of your account. For more information about placement groups and cluster instances, see Cluster Instances in the Amazon Elastic Compute Cloud User Guide. Example This example creates a placement group named XYZ-cluster. https:/ec2.amazonaws.com?Action=CreatePlacementGroup &amp;GroupName=XYZ-cluster &amp;Strategy=cluster &amp;AUTHPARAMS &lt;CreatePlacementGroupResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;d4904fd9-82c2-4ea5-adfe-a9cc3EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;CreatePlacementGroupResponse&gt;.

Documentation

createPlacementGroup

Arguments

:: PlacementStrategy
_cpgrStrategy

-> Text
_cpgrGroupName

-> CreatePlacementGroup

Minimum specification for a CreatePlacementGroup request.

data CreatePlacementGroup

Constructors

CreatePlacementGroup

Fields

_cpgrStrategy :: PlacementStrategy

The placement strategy.

_cpgrGroupName :: Text

A name for the placement group. Constraints: Up to 255 ASCII characters.

_cpgrDryRun :: Maybe Bool

cpgrStrategy :: Lens' CreatePlacementGroup PlacementStrategy

cpgrGroupName :: Lens' CreatePlacementGroup Text

cpgrDryRun :: Lens' CreatePlacementGroup (Maybe Bool)

data CreatePlacementGroupResponse

Constructors

CreatePlacementGroupResponse