Network.AWS.EC2.V2014_06_15.CreateSpotDatafeedSubscription

Description

Creates a datafeed for Spot Instances, enabling you to view Spot Instance usage logs. You can create one data feed per AWS account. For more information, see Spot Instances in the Amazon Elastic Compute Cloud User Guide. Example This example creates a Spot Instance datafeed for the account. https:/ec2.amazonaws.com?Action=CreateSpotDatafeedSubscription &amp;Bucket=my-s3-bucket &amp;AUTHPARAMS &lt;CreateSpotDatafeedSubscriptionResponse xmlns="http://ec2.amazonaws.com/doc/2013-06-15/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;spotDatafeedSubscription&gt; &lt;ownerId&gt;123456789012&lt;/ownerId&gt; &lt;bucket&gt;my-s3-bucket&lt;/bucket&gt; &lt;prefix&gt;spotdata_&lt;prefix&gt; &lt;state&gt;Active&lt;state&gt; &lt;/spotDatafeedSubscription&gt; &lt;/CreateSpotDatafeedSubscriptionResponse&gt;.

Request

data CreateSpotDatafeedSubscription

Request constructor

mkCreateSpotDatafeedSubscription

Arguments

:: Text
csdsBucket

-> CreateSpotDatafeedSubscription

Smart constructor for the minimum required parameters to construct a valid CreateSpotDatafeedSubscription request.

Request lenses

csdsBucket :: Lens' CreateSpotDatafeedSubscription Text

The Amazon S3 bucket in which to store the Spot Instance datafeed. Constraints: Must be a valid bucket associated with your AWS account.

csdsPrefix :: Lens' CreateSpotDatafeedSubscription (Maybe Text)

A prefix for the datafeed file names.

Response

data CreateSpotDatafeedSubscriptionResponse

Response lenses

csdsrsSpotDatafeedSubscription :: Lens' CreateSpotDatafeedSubscriptionResponse (Maybe SpotDatafeedSubscription)

The Spot Instance datafeed subscription.