Network.AWS.EC2.V2014_06_15.CreateTags

Description

Adds or overwrites one or more tags for the specified EC2 resource or resources. Each resource can have a maximum of 10 tags. Each tag consists of a key and optional value. Tag keys must be unique per resource. For more information about tags, see Tagging Your Resources in the Amazon Elastic Compute Cloud User Guide. Example This example request adds (or overwrites) two tags for an AMI and an instance. One of the tags is just a key (webserver), with no value (we set the value to an empty string). The other tag consists of a key (stack) and value (Production). https:/ec2.amazonaws.com?Action=CreateTags &amp;ResourceId.1=ami-1a2b3c4d &amp;ResourceId.2=i-7f4d3a2b &amp;Tag.1.Key=webserver &amp;Tag.1.Value= &amp;Tag.2.Key=stack &amp;Tag.2.Value=Production &amp;AUTHPARAMS xmlns="http://ec2.amazonaws.com/doc/2014-05-01/"> 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE true.

Request

data CreateTags

Request constructor

mkCreateTags

Arguments

:: [Text]
ctResources

-> [Tag]
ctTags

-> CreateTags

Smart constructor for the minimum required parameters to construct a valid CreateTags request.

Request lenses

ctResources :: Lens' CreateTags [Text]

The IDs of one or more resources to tag. For example, ami-1a2b3c4d.

ctTags :: Lens' CreateTags [Tag]

One or more tags. The value parameter is required, but if you don't want the tag to have a value, specify the parameter with no value, and we set the value to an empty string.

Response

data CreateTagsResponse

Response constructor

mkCreateTagsResponse :: CreateTagsResponse

Smart constructor for the minimum required parameters to construct a valid CreateTagsResponse response.

This constructor is provided for convenience and testing purposes.