Network.AWS.EC2.V2014_06_15.CreateVolume

Description

Creates an Amazon EBS volume that can be attached to an instance in the same Availability Zone. The volume is created in the specified region. You can create a new empty volume or restore a volume from an Amazon EBS snapshot. Any AWS Marketplace product codes from the snapshot are propagated to the volume. You can create encrypted volumes with the Encrypted parameter. Encrypted volumes may only be attached to instances that support Amazon EBS encryption. Volumes that are created from encrypted snapshots are also automatically encrypted. For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide. For more information, see Creating or Restoring an Amazon EBS Volume in the Amazon Elastic Compute Cloud User Guide. Example This example request creates an 80 GiB encrypted volume in the Availability Zone us-east-1a. https:/ec2.amazonaws.com?Action=CreateVolume &amp;Size=80 &amp;AvailabilityZone=us-east-1a &amp;Encrypted=1 &amp;AUTHPARAMS &lt;CreateVolumeResponse xmlns="http://ec2.amazonaws.com/doc/2014-05-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;volumeId&gt;vol-1a2b3c4d&lt;volumeId&gt; &lt;size&gt;80&lt;size&gt; &lt;snapshotId/&gt; &lt;availabilityZone&gt;us-east-1a&lt;/availabilityZone&gt; &lt;status&gt;creating&lt;/status&gt; &lt;createTime&gt;YYYY-MM-DDTHH:MM:SS.000Z&lt;/createTime&gt; &lt;volumeType&gt;standard&lt;/volumeType&gt; &lt;encrypted&gt;true&lt;encrypted&gt; &lt;CreateVolumeResponse&gt;.

Documentation

createVolume

Arguments

:: Text
_cvtAvailabilityZone

-> CreateVolume

Minimum specification for a CreateVolume request.

data CreateVolume

Constructors

CreateVolume

Fields

_cvtAvailabilityZone :: Text

The Availability Zone in which to create the volume. Use DescribeAvailabilityZones to list the Availability Zones that are currently available to you.

_cvtEncrypted :: Maybe Bool

Specifies whether the volume should be encrypted.

_cvtDryRun :: Maybe Bool

_cvtSize :: Maybe Integer

The size of the volume, in GiBs. Constraints: If the volume type is io1, the minimum size of the volume is 10 GiB. Default: If you're creating the volume from a snapshot and don't specify a volume size, the default is the snapshot size.

_cvtIops :: Maybe Integer

Only valid for Provisioned IOPS (SSD) volumes. The number of I/O operations per second (IOPS) to provision for the volume.

_cvtSnapshotId :: Maybe Text

The snapshot from which to create the volume.

_cvtVolumeType :: Maybe VolumeType

The volume type. This can be gp2 for General Purpose (SSD) volumes, io1 for Provisioned IOPS (SSD) volumes, or standard for Magnetic volumes. Default: standard.

cvtVolumeType :: Lens' CreateVolume (Maybe VolumeType)

cvtSnapshotId :: Lens' CreateVolume (Maybe Text)

cvtSize :: Lens' CreateVolume (Maybe Integer)

cvtIops :: Lens' CreateVolume (Maybe Integer)

cvtEncrypted :: Lens' CreateVolume (Maybe Bool)

cvtDryRun :: Lens' CreateVolume (Maybe Bool)

cvtAvailabilityZone :: Lens' CreateVolume Text

data CreateVolumeResponse

Constructors

CreateVolumeResponse

Fields

_vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhEncrypted :: Maybe Bool

Indicates whether the volume is encrypted.

_vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhCreateTime :: Maybe ISO8601

The time stamp when volume creation was initiated.

_vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhSize :: Maybe Integer

The size of the volume, in GiBs.

_vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhIops :: Maybe Integer

The number of I/O operations per second (IOPS) that the volume supports. For Provisioned IOPS (SSD) volumes, this represents the number of IOPS that are provisioned for the volume. For General Purpose (SSD) volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates I/O credits for bursting. For more information on General Purpose (SSD) baseline performance, I/O credits, and bursting, see Amazon EBS Volume Types in the Amazon Elastic Compute Cloud User Guide. Constraint: Range is 100 to 4000 for Provisioned IOPS (SSD) volumes and 3 to 3072 for General Purpose (SSD) volumes. Condition: This parameter is required for requests to create io1 volumes; it is not used in requests to create standard or gp2 volumes.

_vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhTags :: [Tag]

Any tags assigned to the volume.

_vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhAvailabilityZone :: Maybe Text

The Availability Zone for the volume.

_vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhVolumeId :: Maybe Text

The ID of the volume.

_vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhSnapshotId :: Maybe Text

The snapshot from which the volume was created, if applicable.

_vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhAttachments :: [VolumeAttachment]

_vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhState :: Maybe VolumeState

The volume state.

_vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhVolumeType :: Maybe VolumeType

The volume type. This can be gp2 for General Purpose (SSD) volumes, io1 for Provisioned IOPS (SSD) volumes, or standard for Magnetic volumes.

vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhVolumeType :: Lens' CreateVolumeResponse (Maybe VolumeType)

vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhVolumeId :: Lens' CreateVolumeResponse (Maybe Text)

vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhTags :: Lens' CreateVolumeResponse [Tag]

vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhState :: Lens' CreateVolumeResponse (Maybe VolumeState)

vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhSnapshotId :: Lens' CreateVolumeResponse (Maybe Text)

vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhSize :: Lens' CreateVolumeResponse (Maybe Integer)

vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhIops :: Lens' CreateVolumeResponse (Maybe Integer)

vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhEncrypted :: Lens' CreateVolumeResponse (Maybe Bool)

vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhCreateTime :: Lens' CreateVolumeResponse (Maybe ISO8601)

vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhAvailabilityZone :: Lens' CreateVolumeResponse (Maybe Text)

vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvhAttachments :: Lens' CreateVolumeResponse [VolumeAttachment]