Network.AWS.EC2.V2014_06_15.CreateVpcPeeringConnection

Description

Requests a VPC peering connection between two VPCs: a requester VPC that you own and a peer VPC with which to create the connection. The peer VPC can belong to another AWS account. The requester VPC and peer VPC cannot have overlapping CIDR blocks. The owner of the peer VPC must accept the peering request to activate the peering connection. The VPC peering connection request expires after 7 days, after which it cannot be accepted or rejected. A CreateVpcPeeringConnection request between VPCs with overlapping CIDR blocks results in the VPC peering connection having a status of failed. Example 1 This example requests a peering connection between your VPC (vpc-1a2b3c4d), and a VPC (vpc-a1b2c3d4) that belongs to AWS account 123456789012. https:/ec2.amazonaws.com?Action=CreateVpcPeeringConnection &amp;VpcId=vpc-1a2b3c4d &amp;PeerVpcId=vpc-a1b2c3d4 &amp;PeerOwnerId=123456789012 &amp;AUTHPARAMS &lt;CreateVpcPeeringConnectionResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;vpcPeeringConnection&gt; &lt;vpcPeeringConnectionId&gt;pcx-73a5401a&lt;/vpcPeeringConnectionId&gt; &lt;requesterVpcInfo&gt; &lt;ownerId&gt;777788889999&lt;/ownerId&gt; &lt;vpcId&gt;vpc-vpc-1a2b3c4d&lt;/vpcId&gt; &lt;cidrBlock&gt;10.0.0.028&lt;cidrBlock&gt; &lt;/requesterVpcInfo&gt; &lt;accepterVpcInfo&gt; &lt;ownerId&gt;123456789012&lt;/ownerId&gt; &lt;vpcId&gt;vpc-a1b2c3d4&lt;vpcId&gt; &lt;accepterVpcInfo&gt; &lt;status&gt; &lt;code&gt;initiating-request&lt;/code&gt; &lt;message&gt;Initiating Request to 123456789012&lt;/message&gt; &lt;/status&gt; &lt;expirationTime&gt;2014-02-18T14:37:25.000Z&lt;/expirationTime&gt; &lt;tagSet&gt; &lt;vpcPeeringConnection&gt; &lt;/CreateVpcPeeringConnectionResponse&gt; Example 2 This example requests a peering connection between your VPCs vpc-1a2b3c4d and vpc-11122233. https:/ec2.amazonaws.com?Action=CreateVpcPeeringConnection &amp;VpcId=vpc-1a2b3c4d &amp;PeerVpcId=vpc-11122233 &amp;AUTHPARAMS.

Documentation

createVpcPeeringConnection :: CreateVpcPeeringConnection

Minimum specification for a CreateVpcPeeringConnection request.

data CreateVpcPeeringConnection

Constructors

CreateVpcPeeringConnection

Fields

_cvpcrDryRun :: Maybe Bool

_cvpcrPeerVpcId :: Maybe Text

The ID of the VPC with which you are creating the VPC peering connection.

_cvpcrVpcId :: Maybe Text

The ID of the requester VPC.

_cvpcrPeerOwnerId :: Maybe Text

The AWS account ID of the owner of the peer VPC. Default: Your AWS account ID.

cvpcrVpcId :: Lens' CreateVpcPeeringConnection (Maybe Text)

cvpcrPeerVpcId :: Lens' CreateVpcPeeringConnection (Maybe Text)

cvpcrPeerOwnerId :: Lens' CreateVpcPeeringConnection (Maybe Text)

cvpcrDryRun :: Lens' CreateVpcPeeringConnection (Maybe Bool)

data CreateVpcPeeringConnectionResponse

Constructors

CreateVpcPeeringConnectionResponse

Fields

_cvpcsVpcPeeringConnection :: Maybe VpcPeeringConnection

Information about the VPC peering connection.

cvpcsVpcPeeringConnection :: Iso' CreateVpcPeeringConnectionResponse (Maybe VpcPeeringConnection)