Network.AWS.EC2.V2014_06_15.CreateVpnConnectionRoute

Description

Creates a static route associated with a VPN connection between an existing virtual private gateway and a VPN customer gateway. The static route allows traffic to be routed from the virtual private gateway to the VPN customer gateway. For more information about VPN connections, see Adding a Hardware Virtual Private Gateway to Your VPC in the Amazon Virtual Private Cloud User Guide. Example This example creates a static route to the VPN connection for the VPN connection with the ID vpn-83ad48ea to the destination CIDR block 11.12.0.0/16. Note that when using the Query API the "/" is denoted as "%2F". https:/ec2.amazonaws.com?Action=CreateVpnConnectionRoute &amp;DestinationCidrBlock=11.12.0.0%2F16 &amp;VpnConnectionId=vpn-83ad48ea &amp;AUTHPARAMS &lt;CreateVpnConnectionRouteResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;4f35a1b2-c2c3-4093-b51f-abb9d7311990&lt;/requestId&gt; &lt;return&gt;true&lt;/return&gt; &lt;/CreateVpnConnectionRouteResponse&gt;.

Documentation

data CreateVpnConnectionRoute

Constructors

CreateVpnConnectionRoute

Fields

_cvcrrVpnConnectionId :: Text

The ID of the VPN connection.

_cvcrrDestinationCidrBlock :: Text

The CIDR block associated with the local subnet of the customer network.

cvcrrVpnConnectionId :: Lens' CreateVpnConnectionRoute Text

cvcrrDestinationCidrBlock :: Lens' CreateVpnConnectionRoute Text

data CreateVpnConnectionRouteResponse

Constructors

CreateVpnConnectionRouteResponse