Network.AWS.EC2.V2014_06_15.DeleteNetworkInterface

Description

Deletes the specified network interface. You must detach the network interface before you can delete it. Example This example deletes the specified network interface. https:/ec2.amazonaws.com?Action=DeleteNetworkInterface &amp;NetworkInterfaceId=eni-ffda3197 &amp;AUTHPARAMS &lt;DeleteNetworkInterfaceResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;e1c6d73b-edaa-4e62-9909-6611404e1739&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;DeleteNetworkInterfaceResponse&gt;.

Request

data DeleteNetworkInterface

Request constructor

mkDeleteNetworkInterface

Arguments

:: Text
dniNetworkInterfaceId

-> DeleteNetworkInterface

Smart constructor for the minimum required parameters to construct a valid DeleteNetworkInterface request.

Request lenses

dniNetworkInterfaceId :: Lens' DeleteNetworkInterface Text

The ID of the network interface.

Response

data DeleteNetworkInterfaceResponse