Network.AWS.EC2.V2014_06_15.DeletePlacementGroup

Description

Deletes the specified placement group. You must terminate all instances in the placement group before you can delete the placement group. For more information about placement groups and cluster instances, see Cluster Instances in the Amazon Elastic Compute Cloud User Guide. Example This example deletes the placement group named XYZ-cluster. https:/ec2.amazonaws.com?Action=DeletePlacementGroup &amp;GroupName=XYZ-cluster &amp;AUTHPARAMS &lt;DeletePlacementGroupResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;d4904fd9-82c2-4ea5-adfe-a9cc3EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;DeletePlacementGroupResponse&gt;.

Documentation

deletePlacementGroup

Arguments

:: Text
_dpgrGroupName

-> DeletePlacementGroup

Minimum specification for a DeletePlacementGroup request.

data DeletePlacementGroup

Constructors

DeletePlacementGroup

Fields

_dpgrGroupName :: Text

The name of the placement group.

_dpgrDryRun :: Maybe Bool

dpgrGroupName :: Lens' DeletePlacementGroup Text

dpgrDryRun :: Lens' DeletePlacementGroup (Maybe Bool)

data DeletePlacementGroupResponse

Constructors

DeletePlacementGroupResponse