Network.AWS.EC2.V2014_06_15.DeleteRoute

Description

Deletes the specified route from the specified route table. Example This example deletes the route with destination CIDR 172.16.1.0/24 from the specified route table. https:/ec2.amazonaws.com?Action=DeleteRoute &amp;RouteTableId=rtb-e4ad488d &amp;DestinationCidrBlock=172.16.1.0/24 &amp;AUTHPARMS &lt;DeleteRouteResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;DeleteRouteResponse&gt;.

Documentation

deleteRoute

Arguments

:: Text
_drrRouteTableId

-> Text
_drrDestinationCidrBlock

-> DeleteRoute

Minimum specification for a DeleteRoute request.

data DeleteRoute

Constructors

DeleteRoute

Fields

_drrRouteTableId :: Text

The ID of the route table.

_drrDestinationCidrBlock :: Text

The CIDR range for the route. The value you specify must match the CIDR for the route exactly.

_drrDryRun :: Maybe Bool

drrRouteTableId :: Lens' DeleteRoute Text

drrDryRun :: Lens' DeleteRoute (Maybe Bool)

drrDestinationCidrBlock :: Lens' DeleteRoute Text

data DeleteRouteResponse

Constructors

DeleteRouteResponse