Network.AWS.EC2.V2014_06_15.DeleteSecurityGroup

Description

Deletes a security group. If you attempt to delete a security group that is associated with an instance, or is referenced by another security group, the operation fails with InvalidGroup.InUse in EC2-Classic or DependencyViolation in EC2-VPC. Example for EC2-Classic This example deletes the specified security group for EC2-Classic. https:/ec2.amazonaws.com?Action=DeleteSecurityGroup &amp;GroupName=websrv &amp;AUTHPARAMS &lt;DeleteSecurityGroupResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;DeleteSecurityGroupResponse&gt; Example for EC2-VPC his example deletes the specified security group for EC2-VPC. https:/ec2.amazonaws.com?Action=DeleteSecurityGroup &amp;GroupId=sg-1a2b3c4d &amp;AUTHPARAMS &lt;DeleteSecurityGroupResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;DeleteSecurityGroupResponse&gt;.

Documentation

deleteSecurityGroup :: DeleteSecurityGroup

Minimum specification for a DeleteSecurityGroup request.

data DeleteSecurityGroup

Constructors

DeleteSecurityGroup

Fields

_dsgrDryRun :: Maybe Bool

_dsgrGroupId :: Maybe Text

The ID of the security group.

_dsgrGroupName :: Maybe Text

EC2-Classic, default VPC

The name of the security group.

dsgrGroupName :: Lens' DeleteSecurityGroup (Maybe Text)

dsgrGroupId :: Lens' DeleteSecurityGroup (Maybe Text)

dsgrDryRun :: Lens' DeleteSecurityGroup (Maybe Bool)

data DeleteSecurityGroupResponse

Constructors

DeleteSecurityGroupResponse