Network.AWS.EC2.V2014_06_15.DeleteSpotDatafeedSubscription

Description

Deletes the datafeed for Spot Instances. For more information, see Spot Instances in the Amazon Elastic Compute Cloud User Guide. Example This example request deletes the datafeed for the AWS account. https:/ec2.amazonaws.com?Action=DeleteSpotDatafeedSubscription &amp;AUTHPARAMS &lt;DeleteSpotDatafeedSubscriptionResponse xmlns="http://ec2.amazonaws.com/doc/2013-06-15/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;/return&gt; &lt;/DeleteSpotDatafeedSubscriptionResponse&gt;.

Request

data DeleteSpotDatafeedSubscription

Request constructor

mkDeleteSpotDatafeedSubscription :: DeleteSpotDatafeedSubscription

Smart constructor for the minimum required parameters to construct a valid DeleteSpotDatafeedSubscription request.

Response

data DeleteSpotDatafeedSubscriptionResponse

Response constructor

mkDeleteSpotDatafeedSubscriptionResponse :: DeleteSpotDatafeedSubscriptionResponse

Smart constructor for the minimum required parameters to construct a valid DeleteSpotDatafeedSubscriptionResponse response.

This constructor is provided for convenience and testing purposes.