Network.AWS.EC2.V2014_06_15.DeleteVpnConnection

Description

Deletes the specified VPN connection. If you're deleting the VPC and its associated components, we recommend that you detach the virtual private gateway from the VPC and delete the VPC before deleting the VPN connection. Example This example deletes the specified VPN connection. https:/ec2.amazonaws.com?Action=DeleteVpnConnection &amp;vpnConnectionId=vpn-44a8938f &amp;AUTHPARAMS &lt;DeleteVpnConnectionResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;DeleteVpnConnectionResponse&gt;.

Documentation

deleteVpnConnection

Arguments

:: Text
_dvcrVpnConnectionId

-> DeleteVpnConnection

Minimum specification for a DeleteVpnConnection request.

data DeleteVpnConnection

Constructors

DeleteVpnConnection

Fields

_dvcrVpnConnectionId :: Text

The ID of the VPN connection.

_dvcrDryRun :: Maybe Bool

dvcrVpnConnectionId :: Lens' DeleteVpnConnection Text

dvcrDryRun :: Lens' DeleteVpnConnection (Maybe Bool)

data DeleteVpnConnectionResponse

Constructors

DeleteVpnConnectionResponse