Network.AWS.EC2.V2014_06_15.DeleteVpnGateway

Description

Deletes the specified virtual private gateway. We recommend that before you delete a virtual private gateway, you detach it from the VPC and delete the VPN connection. Note that you don't need to delete the virtual private gateway if you plan to delete and recreate the VPN connection between your VPC and your network. Example This example deletes the specified virtual private gateway. https:/ec2.amazonaws.com?Action=DeleteVpnGateway &amp;vpnGatewayId=vgw-8db04f81 &amp;AUTHPARAMS &lt;DeleteVpnGatewayResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;DeleteVpnGatewayResponse&gt;.

Documentation

deleteVpnGateway

Arguments

:: Text
_dvgrVpnGatewayId

-> DeleteVpnGateway

Minimum specification for a DeleteVpnGateway request.

data DeleteVpnGateway

Constructors

DeleteVpnGateway

Fields

_dvgrVpnGatewayId :: Text

The ID of the virtual private gateway.

_dvgrDryRun :: Maybe Bool

dvgrVpnGatewayId :: Lens' DeleteVpnGateway Text

dvgrDryRun :: Lens' DeleteVpnGateway (Maybe Bool)

data DeleteVpnGatewayResponse

Constructors

DeleteVpnGatewayResponse