Network.AWS.EC2.V2014_06_15.DescribeBundleTasks

Description

Describes one or more of your bundling tasks. Completed bundle tasks are listed for only a limited time. If your bundle task is no longer in the list, you can still register an AMI from it. Just use RegisterImage with the Amazon S3 bucket name and image manifest name you provided to the bundle task. Example 1 This example describes the status of the specified bundle task. https:/ec2.amazonaws.com?Action=DescribeBundleTasks &amp;bundleId.1=bun-c1a540a8 &amp;AUTHPARAMS 59dbff89-35bd-4eac-99ed-be587EXAMPLE i-12345678 bun-c1a540a8 cancelling 2008-10-07T11:41:50.000Z 2008-10-07T11:51:50.000Z myawsbucket winami 20% Example 2 This example filters the response to include only bundle tasks whose state is either complete or failed, and in addition are targeted for the Amazon S3 bucket named myawsbucket. https:/ec2.amazonaws.com?Action=DescribeBundleTasks &amp;Filter.1.Name=s3-bucket &amp;Filter.1.Value.1=myawsbucket &amp;Filter.2.Name=state &amp;Filter.2.Name.1=complete &amp;Filter.2.Name.2=failed &amp;AUTHPARAMS.

Request

data DescribeBundleTasks

Request constructor

mkDescribeBundleTasks :: DescribeBundleTasks

Smart constructor for the minimum required parameters to construct a valid DescribeBundleTasks request.

The fields accessible through corresponding lenses are:

BundleIds :: [Text]

Filters :: [Filter]

Request lenses

dbtBundleIds :: Lens' DescribeBundleTasks [Text]

One or more bundle task IDs. Default: Describes all your bundle tasks.

dbtFilters :: Lens' DescribeBundleTasks [Filter]

One or more filters. bundle-id - The ID of the bundle task. error-code - If the task failed, the error code returned. error-message - If the task failed, the error message returned. instance-id - The ID of the instance. progress - The level of task completion, as a percentage (for example, 20%). s3-bucket - The Amazon S3 bucket to store the AMI. s3-prefix - The beginning of the AMI name. start-time - The time the task started (for example, 2013-09-15T17:15:20.000Z). state - The state of the task (pending | waiting-for-shutdown | bundling | storing | cancelling | complete | failed). update-time - The time of the most recent update for the task.

Response

data DescribeBundleTasksResponse

Response constructor

mkDescribeBundleTasksResponse :: DescribeBundleTasksResponse

Smart constructor for the minimum required parameters to construct a valid DescribeBundleTasksResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

BundleTasks :: [BundleTask]

Response lenses

dbtrBundleTasks :: Lens' DescribeBundleTasksResponse [BundleTask]

Information about one or more bundle tasks.