Network.AWS.EC2.V2014_06_15.DescribeConversionTasks

Description

Describes one or more of your conversion tasks. For more information, see Using the Command Line Tools to Import Your Virtual Machine to Amazon EC2 in the Amazon Elastic Compute Cloud User Guide. Example This example describes all your conversion tasks. https:/ec2.amazonaws.com?Action=DescribeConversionTasks &amp;AUTHPARAMS import-i-fh95npoc 2010-12-22T12:01Z 1000 us-east-1a VDMK 128696320 https:/s3.amazonaws.commyawsbucket​a3a5e1b6-590d-43cc-97c1-15c7325d3f41​Win_2008_Server_Data_Center_SP2_32-bit.​vmdkmanifest.xml?AWSAccessKeyId=​AKIAIOSFODNN7EXAMPLE&amp;​Expires=1294855591&amp;​Signature=5snej01TlTtL0uR7KExtEXAMPLE%3D 8 vol-34d8a2ff active.

Documentation

describeConversionTasks :: DescribeConversionTasks

Minimum specification for a DescribeConversionTasks request.

data DescribeConversionTasks

Constructors

DescribeConversionTasks

Fields

_dctrDryRun :: Maybe Bool

_dctrConversionTaskIds :: [Text]

One or more conversion task IDs.

_dctrFilters :: [Filter]

dctrFilters :: Lens' DescribeConversionTasks [Filter]

dctrDryRun :: Lens' DescribeConversionTasks (Maybe Bool)

dctrConversionTaskIds :: Lens' DescribeConversionTasks [Text]

data DescribeConversionTasksResponse

Constructors

DescribeConversionTasksResponse

Fields

_dctsConversionTasks :: [ConversionTask]

dctsConversionTasks :: Iso' DescribeConversionTasksResponse [ConversionTask]