Network.AWS.EC2.V2014_06_15.DescribeCustomerGateways

Description

Describes one or more of your VPN customer gateways. For more information about VPN customer gateways, see Adding a Hardware Virtual Private Gateway to Your VPC in the Amazon Virtual Private Cloud User Guide. Example 1 This example request describes the specified customer gateway. https:/ec2.amazonaws.com?Action=DescribeCustomerGateways &amp;CustomerGatewayId.1=cgw-b4dc3961 &amp;AUTHPARAMS &lt;DescribeCustomerGatewaysResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;customerGatewaySet&gt; &lt;item&gt; &lt;customerGatewayId&gt;cgw-b4dc3961&lt;/customerGatewayId&gt; &lt;state&gt;available&lt;state&gt; &lt;type&gt;ipsec.1&lt;type&gt; &lt;ipAddress&gt;12.1.2.3&lt;/ipAddress&gt; &lt;bgpAsn&gt;65534&lt;bgpasn&gt; &lt;tagSet&gt; &lt;/item&gt; &lt;customerGatewaySet&gt; &lt;DescribeCustomerGatewaysResponse&gt; Example 2 This example request uses filters to describe any customer gateway you own whose IP address is 12.1.2.3, and whose state is either pending or available. https:/ec2.amazonaws.com?Action=DescribeCustomerGateways &amp;Filter.1.Name=ip-address &amp;Filter.1.Value.1=12.1.2.3 &amp;Filter.2.Name=state &amp;Filter.2.Value.1=pending &amp;Filter.2.Value.2=available &amp;AUTHPARAMS.

Documentation

describeCustomerGateways :: DescribeCustomerGateways

Minimum specification for a DescribeCustomerGateways request.

data DescribeCustomerGateways

Constructors

DescribeCustomerGateways

Fields

_dcgrDryRun :: Maybe Bool

_dcgrCustomerGatewayIds :: [Text]

One or more customer gateway IDs. Default: Describes all your customer gateways.

_dcgrFilters :: [Filter]

One or more filters. bgp-asn - The customer gateway's Border Gateway Protocol (BGP) Autonomous System Number (ASN). customer-gateway-id - The ID of the customer gateway. ip-address - The IP address of the customer gateway's Internet-routable external interface. state - The state of the customer gateway (pending | available | deleting | deleted). type - The type of customer gateway. Currently, the only supported type is ipsec.1. tag:key=value - The key/value combination of a tag assigned to the resource. tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag:key=value filter. tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter.

dcgrFilters :: Lens' DescribeCustomerGateways [Filter]

dcgrDryRun :: Lens' DescribeCustomerGateways (Maybe Bool)

dcgrCustomerGatewayIds :: Lens' DescribeCustomerGateways [Text]

data DescribeCustomerGatewaysResponse

Constructors

DescribeCustomerGatewaysResponse

Fields

_dcgsCustomerGateways :: [CustomerGateway]

Information about one or more customer gateways.

dcgsCustomerGateways :: Iso' DescribeCustomerGatewaysResponse [CustomerGateway]