Network.AWS.EC2.V2014_06_15.DescribeInstanceAttribute

Description

Describes the specified attribute of the specified instance. You can specify only one attribute at a time. Example 1 This example lists the instance type of the specified instance. https:/ec2.amazonaws.com?Action=DescribeInstanceAttribute &amp;InstanceId=i-10a64379 &amp;Attribute=instanceType &amp;AUTHPARAMS &lt;DescribeInstanceAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;instanceId&gt;i-10a64379&lt;/instanceId&gt; &lt;instanceType&gt; &lt;value&gt;t1.micro&lt;value&gt; &lt;instanceType&gt; &lt;/DescribeInstanceAttributeResponse&gt; Example 2 This example lists the current value of the InstanceInitiatedShutdownBehavior attribute for the specified instance. https:/ec2.amazonaws.com?Action=DescribeInstanceAttribute &amp;InstanceId=i-10a64379 &amp;Attribute=instanceInitiatedShutdownBehavior &amp;AUTHPARAMS &lt;DescribeInstanceAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;instanceId&gt;i-10a64379&lt;/instanceId&gt; &lt;instanceInitiatedShutdownBehavior&gt; &lt;value&gt;stop&lt;/value&gt; &lt;/instanceInitiatedShutdownBehavior&gt; &lt;/DescribeInstanceAttributeResponse&gt; Example 3 This example lists the current value of the DisableApiTermination attribute for the specified instance. https:/ec2.amazonaws.com?Action=DescribeInstanceAttribute &amp;InstanceId=i-10a64379 &amp;Attribute=disableApiTermination &amp;AUTHPARAMS &lt;DescribeInstanceAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;instanceId&gt;i-10a64379&lt;/instanceId&gt; &lt;disableApiTermination&gt; &lt;value&gt;false&lt;/value&gt; &lt;disableApiTermination&gt; &lt;DescribeInstanceAttributeResponse&gt;.

Request

data DescribeInstanceAttribute

Request constructor

mkDescribeInstanceAttribute

Arguments

:: Text
dia1InstanceId

-> InstanceAttributeName
dia1Attribute

-> DescribeInstanceAttribute

Smart constructor for the minimum required parameters to construct a valid DescribeInstanceAttribute request.

Request lenses

dia1InstanceId :: Lens' DescribeInstanceAttribute Text

The ID of the instance.

dia1Attribute :: Lens' DescribeInstanceAttribute InstanceAttributeName

The instance attribute.

Response

data DescribeInstanceAttributeResponse

Response lenses

diarsrsInstanceId :: Lens' DescribeInstanceAttributeResponse (Maybe Text)

The ID of the instance.

diarsrsInstanceType :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

The instance type.

diarsrsKernelId :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

The kernel ID.

diarsrsRamdiskId :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

The RAM disk ID.

diarsrsUserData :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

The Base64-encoded MIME user data.

diarsrsDisableApiTermination :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeBooleanValue)

If the value is true, you can't terminate the instance through the Amazon EC2 console, CLI, or API; otherwise, you can.

diarsrsInstanceInitiatedShutdownBehavior :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

diarsrsRootDeviceName :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

The name of the root device (for example, devsda1).

diarsrsBlockDeviceMappings :: Lens' DescribeInstanceAttributeResponse [InstanceBlockDeviceMapping]

The block device mapping of the instance.

diarsrsProductCodes :: Lens' DescribeInstanceAttributeResponse [ProductCode]

A list of product codes.

diarsrsEbsOptimized :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeBooleanValue)

Indicates whether the instance is optimized for EBS I/O.

diarsrsSriovNetSupport :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

diarsrsSourceDestCheck :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeBooleanValue)

Indicates whether source/destination checking is enabled. A value of true means checking is enabled, and false means checking is disabled. This value must be false for a NAT instance to perform NAT.