Network.AWS.EC2.V2014_06_15.DescribeKeyPairs

Description

Describes one or more of your key pairs. For more information about key pairs, see Key Pairs in the Amazon Elastic Compute Cloud User Guide. Example This example describes the keypair with name my-key-pair. https:/ec2.amazonaws.com?Action=DescribeKeyPairs &amp;KeyName.1=my-key-pair &amp;AUTHPARAMS 59dbff89-35bd-4eac-99ed-be587EXAMPLE my-key-pair 1f:51:ae:28:bf:89:e9:d8:1f:25:5d:37:2d:7d:b8:ca:9f:f5:f1:6f Example This example filters the response to include only key pairs whose names include the string Dave. https:/ec2.amazonaws.com?Action=DescribeKeyPairs &amp;Filter.1.Name=key-name &amp;Filter.1.Value.1=*Dave* &amp;AUTHPARAMS.

Documentation

describeKeyPairs :: DescribeKeyPairs

Minimum specification for a DescribeKeyPairs request.

data DescribeKeyPairs

Constructors

DescribeKeyPairs

Fields

_dkpsDryRun :: Maybe Bool

_dkpsFilters :: [Filter]

One or more filters. fingerprint - The fingerprint of the key pair. key-name - The name of the key pair.

_dkpsKeyNames :: [Text]

One or more key pair names. Default: Describes all your key pairs.

dkpsKeyNames :: Lens' DescribeKeyPairs [Text]

dkpsFilters :: Lens' DescribeKeyPairs [Filter]

dkpsDryRun :: Lens' DescribeKeyPairs (Maybe Bool)

data DescribeKeyPairsResponse

Constructors

DescribeKeyPairsResponse

Fields

_dkptKeyPairs :: [KeyPairInfo]

Information about one or more key pairs.

dkptKeyPairs :: Iso' DescribeKeyPairsResponse [KeyPairInfo]