Network.AWS.EC2.V2014_06_15.DescribeReservedInstancesListings

Description

Describes your account's Reserved Instance listings in the Reserved Instance Marketplace. For more information, see Reserved Instance Marketplace in the Amazon Elastic Compute Cloud User Guide. Example This example shows all the listings associated with your account. https:/ec2.amazonaws.com?Action=DescribeReservedInstancesListings &amp;AUTHPARAMS cec5c904-8f3a-4de5-8f5a-ff7f9EXAMPLE 253dfbf9-c335-4808-b956-d942cEXAMPLE e5a2ff3b-7d14-494f-90af-0b5d0EXAMPLE 2012-07-06T19:35:29.000Z 2012-07-06T19:35:30.000Z active ACTIVE Available 20 Sold 0 Cancelled 0 Pending 0 8 480.0 USD false 7 420.0 USD false 6 360.0 USD active 5 300.0 USD false 4 240.0 USD false 3 180.0 USD false 2 120.0 USD false 1 60.0 USD false myclienttoken1.

Documentation

describeReservedInstancesListings :: DescribeReservedInstancesListings

Minimum specification for a DescribeReservedInstancesListings request.

data DescribeReservedInstancesListings

Constructors

DescribeReservedInstancesListings

Fields

_drilrFilters :: [Filter]

One or more filters. reserved-instances-id - The ID of the Reserved Instances. reserved-instances-listing-id - The ID of the Reserved Instances listing. status - The status of the Reserved Instance listing (pending | active | cancelled | closed). status-message - The reason for the status.

_drilrReservedInstancesId :: Maybe Text

One or more Reserved Instance IDs.

_drilrReservedInstancesListingId :: Maybe Text

One or more Reserved Instance Listing IDs.

drilrReservedInstancesListingId :: Lens' DescribeReservedInstancesListings (Maybe Text)

drilrReservedInstancesId :: Lens' DescribeReservedInstancesListings (Maybe Text)

drilrFilters :: Lens' DescribeReservedInstancesListings [Filter]

data DescribeReservedInstancesListingsResponse

Constructors

DescribeReservedInstancesListingsResponse

Fields

_drilsReservedInstancesListings :: [ReservedInstancesListing]

Information about the Reserved Instance listing.

drilsReservedInstancesListings :: Iso' DescribeReservedInstancesListingsResponse [ReservedInstancesListing]