Network.AWS.EC2.V2014_06_15.DescribeVolumeAttribute

Description

Describes the specified attribute of the specified volume. You can specify only one attribute at a time. For more information about Amazon EBS volumes, see Amazon EBS Volumes in the Amazon Elastic Compute Cloud User Guide. Example This example describes the autoEnableIO attribute of the volume vol-12345678. https:/ec2.amazonaws.com?Action=DescribeVolumeAttribute &amp;Attribute=autoEnableIO &amp;VolumeId=vol-12345678 &amp;AUTHPARAMS &lt;DescribeVolumeAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2014-05-01/"&gt; &lt;requestId&gt;5jkdf074-37ed-4004-8671-a78ee82bf1cbEXAMPLE&lt;/requestId&gt; &lt;volumeId&gt;vol-12345678&lt;/volumeId&gt; &lt;autoEnableIO&gt; &lt;value&gt;false&lt;value&gt; &lt;autoEnableIO&gt; &lt;/DescribeVolumeAttributeResponse&gt; Example This example describes the productCodes attribute of the volume vol-12345678. https:/ec2.amazonaws.com?Action=DescribeVolumeAttribute &amp;Attribute=productCodes &amp;VolumeId=vol-12345678 &amp;AUTHPARAMS &lt;DescribeVolumeAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2014-05-01/"&gt; &lt;requestId&gt;5jkdf074-37ed-4004-8671-a78ee82bf1cbEXAMPLE&lt;/requestId&gt; &lt;volumeId&gt;vol-12345678&lt;/volumeId&gt; &lt;productCodes&gt; &lt;item&gt; &lt;productCode&gt;a1b2c3d4e5f6g7h8i9j10k11&lt;/productCode&gt; &lt;type&gt;marketplace&lt;type&gt; &lt;item&gt; &lt;/productCodes&gt; &lt;/DescribeVolumeAttributeResponse&gt;.

Documentation

describeVolumeAttribute

Arguments

:: Text
_dvarVolumeId

-> DescribeVolumeAttribute

Minimum specification for a DescribeVolumeAttribute request.

data DescribeVolumeAttribute

Constructors

DescribeVolumeAttribute

Fields

_dvarVolumeId :: Text

The ID of the volume.

_dvarDryRun :: Maybe Bool

_dvarAttribute :: Maybe VolumeAttributeName

The instance attribute.

dvarVolumeId :: Lens' DescribeVolumeAttribute Text

dvarDryRun :: Lens' DescribeVolumeAttribute (Maybe Bool)

dvarAttribute :: Lens' DescribeVolumeAttribute (Maybe VolumeAttributeName)

data DescribeVolumeAttributeResponse

Constructors

DescribeVolumeAttributeResponse

Fields

_dvasAutoEnableIO :: Maybe AttributeBooleanValue

The state of autoEnableIO attribute.

_dvasProductCodes :: [ProductCode]

A list of product codes.

_dvasVolumeId :: Maybe Text

The ID of the volume.

dvasVolumeId :: Lens' DescribeVolumeAttributeResponse (Maybe Text)

dvasProductCodes :: Lens' DescribeVolumeAttributeResponse [ProductCode]

dvasAutoEnableIO :: Lens' DescribeVolumeAttributeResponse (Maybe AttributeBooleanValue)