Network.AWS.EC2.V2014_06_15.DescribeVpcAttribute

Description

Describes the specified attribute of the specified VPC. You can specify only one attribute at a time. Example 1 This example describes the enableDnsSupport attribute of the specified VPC. The sample response indicates that DNS resolution is supported. https:/ec2.amazonaws.com?Action=DescribeVpcAttribute &amp;VpcId=vpc-1a2b3c4d &amp;Attribute=enableDnsSupport &amp;AUTHPARAMS &lt;DescribeVpcAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;vpcId&gt;vpc-1a2b3c4d&lt;/vpcId&gt; &lt;enableDnsSupport&gt; &lt;value&gt;true&lt;value&gt; &lt;enableDnsSupport&gt; &lt;/DescribeVpcAttributeResponse&gt; Example 2 This request describes the enableDnsHostnames attribute of the specified VPC. The sample response indicates that DNS hostnames are supported. https:/ec2.amazonaws.com?Action=DescribeVpcAttribute &amp;VpcId=vpc-1a2b3c4d &amp;Attribute=enableDnsHostnames &amp;AUTHPARAMS &lt;DescribeVpcAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;vpcId&gt;vpc-1a2b3c4d&lt;/vpcId&gt; &lt;enableDnsHostnames&gt; &lt;value&gt;true&lt;value&gt; &lt;enableDnsHostnames&gt; &lt;/DescribeVpcAttributeResponse&gt;.

Documentation

describeVpcAttribute

Arguments

:: Text
_dvatVpcId

-> DescribeVpcAttribute

Minimum specification for a DescribeVpcAttribute request.

data DescribeVpcAttribute

Constructors

DescribeVpcAttribute

Fields

_dvatVpcId :: Text

The ID of the VPC.

_dvatDryRun :: Maybe Bool

_dvatAttribute :: Maybe VpcAttributeName

The VPC attribute.

dvatVpcId :: Lens' DescribeVpcAttribute Text

dvatDryRun :: Lens' DescribeVpcAttribute (Maybe Bool)

dvatAttribute :: Lens' DescribeVpcAttribute (Maybe VpcAttributeName)

data DescribeVpcAttributeResponse

Constructors

DescribeVpcAttributeResponse

Fields

_dvauEnableDnsHostnames :: Maybe AttributeBooleanValue

Indicates whether the instances launched in the VPC get DNS hostnames. If this attribute is true, instances in the VPC get DNS hostnames; otherwise, they do not.

_dvauEnableDnsSupport :: Maybe AttributeBooleanValue

Indicates whether DNS resolution is enabled for the VPC. If this attribute is true, the Amazon DNS server resolves DNS hostnames for your instances to their corresponding IP addresses; otherwise, it does not.

_dvauVpcId :: Maybe Text

The ID of the VPC.

dvauVpcId :: Lens' DescribeVpcAttributeResponse (Maybe Text)

dvauEnableDnsSupport :: Lens' DescribeVpcAttributeResponse (Maybe AttributeBooleanValue)

dvauEnableDnsHostnames :: Lens' DescribeVpcAttributeResponse (Maybe AttributeBooleanValue)