Network.AWS.EC2.V2014_06_15.DetachVolume

Description

Detaches an Amazon EBS volume from an instance. Make sure to unmount any file systems on the device within your operating system before detaching the volume. Failure to do so results in the volume being stuck in a busy state while detaching. If an Amazon EBS volume is the root device of an instance, it can't be detached while the instance is running. To detach the root volume, stop the instance first. If the root volume is detached from an instance with an AWS Marketplace product code, then the AWS Marketplace product codes from that volume are no longer associated with the instance. For more information, see Detaching an Amazon EBS Volume in the Amazon Elastic Compute Cloud User Guide. Example This example detaches volume vol-1a2b3c4d. https:/ec2.amazonaws.com?Action=DetachVolume &amp;VolumeId=vol-1a2b3c4d &amp;AUTHPARAMS &lt;DetachVolumeResponse xmlns="http://ec2.amazonaws.com/doc/2014-05-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;volumeId&gt;vol-1a2b3c4d&lt;/volumeId&gt; &lt;instanceId&gt;i-1a2b3c4d&lt;/instanceId&gt; &lt;device&gt;devsdh&lt;/device&gt; &lt;status&gt;detaching&lt;/status&gt; &lt;attachTime&gt;YYYY-MM-DDTHH:MM:SS.000Z&lt;/attachTime&gt; &lt;/DetachVolumeResponse&gt;.

Request

data DetachVolume

Request constructor

mkDetachVolume

Arguments

:: Text
dv4VolumeId

-> DetachVolume

Smart constructor for the minimum required parameters to construct a valid DetachVolume request.

Request lenses

dv4VolumeId :: Lens' DetachVolume Text

The ID of the volume.

dv4InstanceId :: Lens' DetachVolume (Maybe Text)

The ID of the instance.

dv4Device :: Lens' DetachVolume (Maybe Text)

The device name.

dv4Force :: Lens' DetachVolume (Maybe Bool)

Forces detachment if the previous detachment attempt did not occur cleanly (for example, logging into an instance, unmounting the volume, and detaching normally). This option can lead to data loss or a corrupted file system. Use this option only as a last resort to detach a volume from a failed instance. The instance won't have an opportunity to flush file system caches or file system metadata. If you use this option, you must perform file system check and repair procedures.

Response

data DetachVolumeResponse

Response lenses

dvr1VolumeId :: Lens' DetachVolumeResponse (Maybe Text)

The ID of the volume.

dvr1InstanceId :: Lens' DetachVolumeResponse (Maybe Text)

The ID of the instance.

dvr1Device :: Lens' DetachVolumeResponse (Maybe Text)

The device name.

dvr1State :: Lens' DetachVolumeResponse (Maybe VolumeAttachmentState)

The attachment state of the volume.

dvr1AttachTime :: Lens' DetachVolumeResponse (Maybe ISO8601)

The time stamp when the attachment initiated.

dvr1DeleteOnTermination :: Lens' DetachVolumeResponse (Maybe Bool)

Indicates whether the Amazon EBS volume is deleted on instance termination.