Network.AWS.EC2.V2014_06_15.DetachVpnGateway

Description

Detaches a virtual private gateway from a VPC. You do this if you're planning to turn off the VPC and not use it anymore. You can confirm a virtual private gateway has been completely detached from a VPC by describing the virtual private gateway (any attachments to the virtual private gateway are also described). You must wait for the attachment's state to switch to detached before you can delete the VPC or attach a different VPC to the virtual private gateway. Example This example detaches the specified virtual private gateway from the specified VPC. https:/ec2.amazonaws.com?Action=DetachVpnGateway &amp;VpnGatewayId=vgw-8db04f81 &amp;VpcId=vpc-1a2b3c4d &amp;AUTHPARAMS &lt;DetachVpnGatewayResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;DetachVpnGatewayResponse&gt;.

Request

data DetachVpnGateway

Request constructor

mkDetachVpnGateway

Arguments

:: Text
dvg2VpnGatewayId

-> Text
dvg2VpcId

-> DetachVpnGateway

Smart constructor for the minimum required parameters to construct a valid DetachVpnGateway request.

Request lenses

dvg2VpnGatewayId :: Lens' DetachVpnGateway Text

The ID of the virtual private gateway.

dvg2VpcId :: Lens' DetachVpnGateway Text

The ID of the VPC.

Response

data DetachVpnGatewayResponse

Response constructor

mkDetachVpnGatewayResponse :: DetachVpnGatewayResponse

Smart constructor for the minimum required parameters to construct a valid DetachVpnGatewayResponse response.

This constructor is provided for convenience and testing purposes.