Network.AWS.EC2.V2014_06_15.DisassociateRouteTable

Description

Disassociates a subnet from a route table. After you perform this action, the subnet no longer uses the routes in the route table. Instead, it uses the routes in the VPC's main route table. For more information about route tables, see Route Tables in the Amazon Virtual Private Cloud User Guide. Example This example disassociates the specified route table from the subnet it's associated to. https:/ec2.amazonaws.com?Action=DisassociateRouteTable &amp;AssociationId=rtbassoc-fdad4894 &amp;AUTHPARAMS &lt;DisassociateRouteTableResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;DisassociateRouteTableResponse&gt;.

Documentation

disassociateRouteTable

Arguments

:: Text
_drtrAssociationId

-> DisassociateRouteTable

Minimum specification for a DisassociateRouteTable request.

data DisassociateRouteTable

Constructors

DisassociateRouteTable

Fields

_drtrAssociationId :: Text

The association ID representing the current association between the route table and subnet.

_drtrDryRun :: Maybe Bool

drtrDryRun :: Lens' DisassociateRouteTable (Maybe Bool)

drtrAssociationId :: Lens' DisassociateRouteTable Text

data DisassociateRouteTableResponse

Constructors

DisassociateRouteTableResponse