Network.AWS.EC2.V2014_06_15.EnableVgwRoutePropagation

Description

Enables a virtual private gateway (VGW) to propagate routes to the routing tables of a VPC. Example This example enables the specified virtual private gateway to propagate routes automatically to the routing table with the ID rtb-c98a35a0. https:/ec2.amazonaws.com?Action=EnableVgwRoutePropagation &amp;RouteTableID=rtb-c98a35a0 &amp;GatewayId= vgw-d8e09e8a &amp;AUTHPARAMS &lt;EnableVgwRoutePropagation xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;4f35a1b2-c2c3-4093-b51f-abb9d7311990&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;EnableVgwRoutePropagation&gt;.

Documentation

data EnableVgwRoutePropagation

Constructors

EnableVgwRoutePropagation

Fields

_evrprRouteTableId :: Text

The ID of the routing table.

_evrprGatewayId :: Text

The ID of the virtual private gateway.

evrprRouteTableId :: Lens' EnableVgwRoutePropagation Text

evrprGatewayId :: Lens' EnableVgwRoutePropagation Text

data EnableVgwRoutePropagationResponse

Constructors

EnableVgwRoutePropagationResponse