Network.AWS.EC2.V2014_06_15.ImportInstance

Description

Creates an import instance task using metadata from the specified disk image. After importing the image, you then upload it using the ec2-import-volume command in the EC2 command line tools. For more information, see Using the Command Line Tools to Import Your Virtual Machine to Amazon EC2 in the Amazon Elastic Compute Cloud User Guide. Example This example creates an import instance task that migrates a Windows Server 2008 SP2 (32-bit) VM into the AWS us-east-1 region. https:/ec2.amazonaws.com?Action=ImportInstance &amp;LaunchSpecification.Architecture=x86_64 &amp;LaunchSpecification.InstanceType=m1.xlarge &amp;DiskImage.1.Image.Format=VMDK &amp;DiskImage.1.Image.Bytes=1179593728 &amp;DiskImage.1.Image.ImportManifestUrl=https:/s3.amazonaws.commyawsbucket​a3a5e1b6-590d-43cc-97c1-15c7325d3f41​Win_2008_Server_Data_Center_SP2_32-bit.​vmdkmanifest.xml?AWSAccessKeyId=​AKIAIOSFODNN7EXAMPLE&amp;​Expires=1294855591&amp;​Signature=5snej01TlTtL0uR7KExtEXAMPLE%3D &amp;DiskImage.1.Volume.Size=12 &amp;Platform=Windows &amp;AUTHPARAMS import-i-ffvko9js 2010-12-22T12:01Z 0 us-east-1a VMDK 1179593728 https:/s3.amazonaws.commyawsbucket​a3a5e1b6-590d-43cc-97c1-15c7325d3f41​Win_2008_Server_Data_Center_SP2_32-bit.​vmdkmanifest.xml?AWSAccessKeyId=​AKIAIOSFODNN7EXAMPLE&amp;​Expires=1294855591&amp;​Signature=5snej01TlTtL0uR7KExtEXAMPLE%3D 12 vol-1a2b3c4d active i-12655a7f.

Documentation

importInstance

Arguments

:: PlatformValues
_iirPlatform

-> ImportInstance

Minimum specification for a ImportInstance request.

data ImportInstance

Constructors

ImportInstance

Fields

_iirPlatform :: PlatformValues

The instance operating system.

_iirDryRun :: Maybe Bool

_iirDiskImages :: [DiskImage]

_iirLaunchSpecification :: Maybe ImportInstanceLaunchSpecification

_iirDescription :: Maybe Text

A description for the instance being imported.

iirPlatform :: Lens' ImportInstance PlatformValues

iirLaunchSpecification :: Lens' ImportInstance (Maybe ImportInstanceLaunchSpecification)

iirDryRun :: Lens' ImportInstance (Maybe Bool)

iirDiskImages :: Lens' ImportInstance [DiskImage]

iirDescription :: Lens' ImportInstance (Maybe Text)

data ImportInstanceResponse

Constructors

ImportInstanceResponse

Fields

_iisConversionTask :: Maybe ConversionTask

iisConversionTask :: Iso' ImportInstanceResponse (Maybe ConversionTask)