Network.AWS.EC2.V2014_06_15.ImportVolume

Description

Creates an import volume task using metadata from the specified disk image. After importing the image, you then upload it using the ec2-import-volume command in the Amazon EC2 command-line interface (CLI) tools. For more information, see Using the Command Line Tools to Import Your Virtual Machine to Amazon EC2 in the Amazon Elastic Compute Cloud User Guide. Example This example creates an import volume task that migrates a Windows Server 2008 SP2 (32-bit) volume into the AWS us-east-1 region. https:/ec2.amazonaws.com?Action=ImportVolume &amp;AvailabilityZone=us-east-1c &amp;Image.Format=VMDK &amp;Image.Bytes=128696320 &amp;Image.ImportManifestUrl=https:/s3.amazonaws.commyawsbucket​a3a5e1b6-590d-43cc-97c1-15c7325d3f41​Win_2008_Server_Data_Center_SP2_32-bit.​vmdkmanifest.xml?AWSAccessKeyId=​AKIAIOSFODNN7EXAMPLE&amp;​Expires=1294855591&amp;​Signature=5snej01TlTtL0uR7KExtEXAMPLE%3D &amp;VolumeSize=8 &amp;AUTHPARAMS&gt; import-i-fh95npoc 2010-12-22T12:01Z 0 us-east-1c VDMK 128696320 https:/s3.amazonaws.commyawsbucket​a3a5e1b6-590d-43cc-97c1-15c7325d3f41​Win_2008_Server_Data_Center_SP2_32-bit.​vmdkmanifest.xml?AWSAccessKeyId=​AKIAIOSFODNN7EXAMPLE&amp;​Expires=1294855591&amp;​Signature=5snej01TlTtL0uR7KExtEXAMPLE%3D ccb1b0536a4a70e86016b85229b5c6b10b14a4eb 8 vol-34d8a2ff active.

Documentation

importVolume

Arguments

:: DiskImageDetail
_ivrImage

-> Text
_ivrAvailabilityZone

-> VolumeDetail
_ivrVolume

-> ImportVolume

Minimum specification for a ImportVolume request.

data ImportVolume

Constructors

ImportVolume

Fields

_ivrImage :: DiskImageDetail

_ivrAvailabilityZone :: Text

The Availability Zone for the resulting Amazon EBS volume.

_ivrVolume :: VolumeDetail

_ivrDryRun :: Maybe Bool

_ivrDescription :: Maybe Text

An optional description for the volume being imported.

ivrVolume :: Lens' ImportVolume VolumeDetail

ivrImage :: Lens' ImportVolume DiskImageDetail

ivrDryRun :: Lens' ImportVolume (Maybe Bool)

ivrDescription :: Lens' ImportVolume (Maybe Text)

ivrAvailabilityZone :: Lens' ImportVolume Text

data ImportVolumeResponse

Constructors

ImportVolumeResponse

Fields

_ivsConversionTask :: Maybe ConversionTask

ivsConversionTask :: Iso' ImportVolumeResponse (Maybe ConversionTask)