Network.AWS.EC2.V2014_06_15.ModifyImageAttribute

Description

Modifies the specified attribute of the specified AMI. You can specify only one attribute at a time. AWS Marketplace product codes cannot be modified. Images with an AWS Marketplace product code cannot be made public. Example.

Request

data ModifyImageAttribute

Request constructor

mkModifyImageAttribute

Arguments

:: Text
miaImageId

-> ModifyImageAttribute

Smart constructor for the minimum required parameters to construct a valid ModifyImageAttribute request.

Request lenses

miaImageId :: Lens' ModifyImageAttribute Text

The ID of the AMI.

miaAttribute :: Lens' ModifyImageAttribute (Maybe Text)

The name of the attribute to modify.

miaOperationType :: Lens' ModifyImageAttribute (Maybe Text)

The operation type.

miaUserIds :: Lens' ModifyImageAttribute [Text]

One or more AWS account IDs. This is only valid when modifying the launchPermission attribute.

miaUserGroups :: Lens' ModifyImageAttribute [Text]

One or more user groups. This is only valid when modifying the launchPermission attribute.

miaProductCodes :: Lens' ModifyImageAttribute [Text]

One or more product codes. After you add a product code to an AMI, it can't be removed. This is only valid when modifying the productCodes attribute.

miaValue :: Lens' ModifyImageAttribute (Maybe Text)

The value of the attribute being modified. This is only valid when modifying the description attribute.

miaLaunchPermission :: Lens' ModifyImageAttribute (Maybe LaunchPermissionModifications)

miaDescription :: Lens' ModifyImageAttribute (Maybe AttributeValue)

A description for the AMI.

Response

data ModifyImageAttributeResponse