Network.AWS.EC2.V2014_06_15.ModifyNetworkInterfaceAttribute

Description

Modifies the specified network interface attribute. You can specify only one attribute at a time. Example This example sets source/destination checking to false for the specified network interface. https:/ec2.amazonaws.com?Action=ModifyNetworkInterfaceAttribute &amp;NetworkInterfaceId=eni-ffda3197 &amp;SourceDestCheck.Value=false &amp;AUTHPARAMS &lt;ModifyNetworkInterfaceAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;657a4623-5620-4232-b03b-427e852d71cf&lt;/requestId&gt; &lt;return&gt;true&lt;/return&gt; &lt;/ModifyNetworkInterfaceAttributeResponse&gt;.

Documentation

modifyNetworkInterfaceAttribute

Arguments

:: Text
_mniarNetworkInterfaceId

-> ModifyNetworkInterfaceAttribute

Minimum specification for a ModifyNetworkInterfaceAttribute request.

data ModifyNetworkInterfaceAttribute

Constructors

ModifyNetworkInterfaceAttribute

Fields

_mniarNetworkInterfaceId :: Text

The ID of the network interface.

_mniarSourceDestCheck :: Maybe AttributeBooleanValue

Indicates whether source/destination checking is enabled. A value of true means checking is enabled, and false means checking is disabled. This value must be false for a NAT instance to perform NAT. For more information, see NAT Instances in the Amazon Virtual Private Cloud User Guide.

_mniarDescription :: Maybe AttributeValue

A description for the network interface.

_mniarDryRun :: Maybe Bool

_mniarAttachment :: Maybe NetworkInterfaceAttachmentChanges

The ID of the interface attachment.

_mniarGroups :: [Text]

Changes the security groups for the network interface. The new set of groups you specify replaces the current set. You must specify at least one group, even if it's just the default security group in the VPC. You must specify the ID of the security group, not the name.

mniarSourceDestCheck :: Lens' ModifyNetworkInterfaceAttribute (Maybe AttributeBooleanValue)

mniarNetworkInterfaceId :: Lens' ModifyNetworkInterfaceAttribute Text

mniarGroups :: Lens' ModifyNetworkInterfaceAttribute [Text]

mniarDryRun :: Lens' ModifyNetworkInterfaceAttribute (Maybe Bool)

mniarDescription :: Lens' ModifyNetworkInterfaceAttribute (Maybe AttributeValue)

mniarAttachment :: Lens' ModifyNetworkInterfaceAttribute (Maybe NetworkInterfaceAttachmentChanges)

data ModifyNetworkInterfaceAttributeResponse

Constructors

ModifyNetworkInterfaceAttributeResponse