Network.AWS.EC2.V2014_06_15.ModifySnapshotAttribute

Description

Adds or removes permission settings for the specified snapshot. You may add or remove specified AWS account IDs from a snapshot's list of create volume permissions, but you cannot do both in a single API call. If you need to both add and remove account IDs for a snapshot, you must use multiple API calls. For more information on modifying snapshot permissions, see Sharing Snapshots in the Amazon Elastic Compute Cloud User Guide. Snapshots with AWS Marketplace product codes cannot be made public. Example This example makes the snap-1a2b3c4d snapshot public, and gives the account with ID 111122223333 permission to create volumes from the snapshot. https:/ec2.amazonaws.com?Action=ModifySnapshotAttribute &amp;snapshotId=snap-1a2b3c4d &amp;CreateVolumePermission.Add.1.UserId=111122223333 &amp;CreateVolumePermission.Add.1.Group=all &amp;AUTHPARAMS &lt;ModifySnapshotAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2014-05-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;ModifySnapshotAttributeResponse&gt; Example This example makes the snap-1a2b3c4d snapshot public, and removes the account with ID 111122223333 from the list of users with permission to create volumes from the snapshot. https:/ec2.amazonaws.com?Action=ModifySnapshotAttribute &amp;snapshotId=snap-1a2b3c4d &amp;CreateVolumePermission.Remove.1.UserId=111122223333 &amp;CreateVolumePermission.Add.1.Group=all &amp;AUTHPARAMS &lt;ModifySnapshotAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2014-05-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;ModifySnapshotAttributeResponse&gt;.

Documentation

modifySnapshotAttribute

Arguments

:: Text
_msasSnapshotId

-> ModifySnapshotAttribute

Minimum specification for a ModifySnapshotAttribute request.

data ModifySnapshotAttribute

Constructors

ModifySnapshotAttribute

Fields

_msasSnapshotId :: Text

The ID of the snapshot.

_msasDryRun :: Maybe Bool

_msasCreateVolumePermission :: Maybe CreateVolumePermissionModifications

A JSON representation of the snapshot attribute modification.

_msasGroupNames :: [Text]

The group to modify for the snapshot.

_msasAttribute :: Maybe SnapshotAttributeName

The snapshot attribute to modify.

_msasUserIds :: [Text]

The account ID to modify for the snapshot.

_msasOperationType :: Maybe Text

The type of operation to perform to the attribute.

msasUserIds :: Lens' ModifySnapshotAttribute [Text]

msasSnapshotId :: Lens' ModifySnapshotAttribute Text

msasOperationType :: Lens' ModifySnapshotAttribute (Maybe Text)

msasGroupNames :: Lens' ModifySnapshotAttribute [Text]

msasDryRun :: Lens' ModifySnapshotAttribute (Maybe Bool)

msasCreateVolumePermission :: Lens' ModifySnapshotAttribute (Maybe CreateVolumePermissionModifications)

msasAttribute :: Lens' ModifySnapshotAttribute (Maybe SnapshotAttributeName)

data ModifySnapshotAttributeResponse

Constructors

ModifySnapshotAttributeResponse