Network.AWS.EC2.V2014_06_15.ModifyVolumeAttribute

Description

Modifies a volume attribute. By default, all I/O operations for the volume are suspended when the data on the volume is determined to be potentially inconsistent, to prevent undetectable, latent data corruption. The I/O access to the volume can be resumed by first enabling I/O access and then checking the data consistency on your volume. You can change the default behavior to resume I/O operations. We recommend that you change this only for boot volumes or for volumes that are stateless or disposable. Example This example modifies the attribute of the volume vol-12345678. https:/ec2.amazonaws.com?Action=ModifyVolumeAttribute &amp;VolumeId=vol-12345678 &amp;AutoEnableIO.Value=true &amp;AUTHPARAMS &lt;ModifyVolumeAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2014-05-01/"&gt; &lt;requestId&gt;5jkdf074-37ed-4004-8671-a78ee82bf1cbEXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;ModifyVolumeAttributeResponse&gt;.

Documentation

modifyVolumeAttribute

Arguments

:: Text
_mvarVolumeId

-> ModifyVolumeAttribute

Minimum specification for a ModifyVolumeAttribute request.

data ModifyVolumeAttribute

Constructors

ModifyVolumeAttribute

Fields

_mvarVolumeId :: Text

The ID of the volume.

_mvarAutoEnableIO :: Maybe AttributeBooleanValue

Indicates whether the volume should be auto-enabled for I/O operations.

_mvarDryRun :: Maybe Bool

mvarVolumeId :: Lens' ModifyVolumeAttribute Text

mvarDryRun :: Lens' ModifyVolumeAttribute (Maybe Bool)

mvarAutoEnableIO :: Lens' ModifyVolumeAttribute (Maybe AttributeBooleanValue)

data ModifyVolumeAttributeResponse

Constructors

ModifyVolumeAttributeResponse