Network.AWS.EC2.V2014_06_15.PurchaseReservedInstancesOffering

Description

Purchases a Reserved Instance for use with your account. With Amazon EC2 Reserved Instances, you obtain a capacity reservation for a certain instance configuration over a specified period of time. You pay a lower usage rate than with On-Demand instances for the time that you actually use the capacity reservation. For more information, see Reserved Instance Marketplace in the Amazon Elastic Compute Cloud User Guide. Example 1 This example uses a limit price to limit the total purchase order of Reserved Instances from Reserved Instance Marketplace. https:/ec2.amazonaws.com?Action=PurchaseReservedInstancesOffering &amp;ReservedInstancesOfferingId=4b2293b4-5813-4cc8-9ce3-1957fEXAMPLE &amp;LimitPrice.Amount=200 &amp;InstanceCount=2 &amp;AUTHPARAMS 59dbff89-35bd-4eac-99ed-be587EXAMPLE e5a2ff3b-7d14-494f-90af-0b5d0EXAMPLE Example 2 This example illustrates a purchase of a Reserved Instances offering. https:/ec2.amazonaws.com?Action=PurchaseReservedInstancesOffering &amp;ReservedInstancesOfferingId=4b2293b4-5813-4cc8-9ce3-1957fEXAMPLE &amp;InstanceCount=2 &amp;AUTHPARAMS &lt;PurchaseReservedInstancesOfferingResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;reservedInstancesId&gt;e5a2ff3b-7d14-494f-90af-0b5d0EXAMPLE&lt;/reservedInstancesId&gt; &lt;/PurchaseReservedInstancesOfferingResponse&gt;.

Request

data PurchaseReservedInstancesOffering

Request constructor

mkPurchaseReservedInstancesOffering

Arguments

:: Text
prioReservedInstancesOfferingId

-> Integer
prioInstanceCount

-> PurchaseReservedInstancesOffering

Smart constructor for the minimum required parameters to construct a valid PurchaseReservedInstancesOffering request.

Request lenses

prioReservedInstancesOfferingId :: Lens' PurchaseReservedInstancesOffering Text

The ID of the Reserved Instance offering to purchase.

prioInstanceCount :: Lens' PurchaseReservedInstancesOffering Integer

The number of Reserved Instances to purchase.

prioLimitPrice :: Lens' PurchaseReservedInstancesOffering (Maybe ReservedInstanceLimitPrice)

Specified for Reserved Instance Marketplace offerings to limit the total order and ensure that the Reserved Instances are not purchased at unexpected prices.

Response

data PurchaseReservedInstancesOfferingResponse

Response lenses

priorReservedInstancesId :: Lens' PurchaseReservedInstancesOfferingResponse (Maybe Text)

The IDs of the purchased Reserved Instances.