Network.AWS.EC2.V2014_06_15.RebootInstances

Description

Requests a reboot of one or more instances. This operation is asynchronous; it only queues a request to reboot the specified instances. The operation succeeds if the instances are valid and belong to you. Requests to reboot terminated instances are ignored. If a Linux/Unix instance does not cleanly shut down within four minutes, Amazon EC2 performs a hard reboot. For more information about troubleshooting, see Getting Console Output and Rebooting Instances in the Amazon Elastic Compute Cloud User Guide. Example This example reboots two instances. https:/ec2.amazonaws.com?Action=RebootInstances &amp;InstanceId.1=i-1a2b3c4d &amp;InstanceId.2=i-4d3acf62 &amp;AUTHPARAMS &lt;RebootInstancesResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;RebootInstancesResponse&gt;.

Request

data RebootInstances

Request constructor

mkRebootInstances

Arguments

:: [Text]
ri1InstanceIds

-> RebootInstances

Smart constructor for the minimum required parameters to construct a valid RebootInstances request.

Request lenses

ri1InstanceIds :: Lens' RebootInstances [Text]

One or more instance IDs.

Response

data RebootInstancesResponse

Response constructor

mkRebootInstancesResponse :: RebootInstancesResponse

Smart constructor for the minimum required parameters to construct a valid RebootInstancesResponse response.

This constructor is provided for convenience and testing purposes.