Network.AWS.EC2.V2014_06_15.ReplaceNetworkAclEntry

Description

Replaces an entry (rule) in a network ACL. For more information about network ACLs, see Network ACLs in the Amazon Virtual Private Cloud User Guide. Example This example replaces the egress entry numbered 110 in the network ACL with ID acl-2cb85d45. The new rule denies egress traffic destined for anywhere (0.0.0.0/0) on TCP port 139. https:/ec2.amazonaws.com?Action=ReplaceNetworkAclEntry &amp;NetworkAclId=acl-2cb85d45 &amp;RuleNumber=110 &amp;Protocol=tcp &amp;RuleAction=deny &amp;Egress=true &amp;CidrBlock=0.0.0.0/0 &amp;PortRange.From=139 &amp;PortRange.To=139 &amp;AUTHPARAMS &lt;ReplaceNetworkAclEntryResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;ReplaceNetworkAclEntryResponse&gt;.

Documentation

replaceNetworkAclEntry

Arguments

:: Bool
_rnaerEgress

-> Integer
_rnaerRuleNumber

-> RuleAction
_rnaerRuleAction

-> Text
_rnaerNetworkAclId

-> Text
_rnaerProtocol

-> Text
_rnaerCidrBlock

-> ReplaceNetworkAclEntry

Minimum specification for a ReplaceNetworkAclEntry request.

data ReplaceNetworkAclEntry

Constructors

ReplaceNetworkAclEntry

Fields

_rnaerEgress :: Bool

Indicates whether to replace the egress rule. Default: If no value is specified, we replace the ingress rule.

_rnaerRuleNumber :: Integer

The rule number of the entry to replace.

_rnaerRuleAction :: RuleAction

Indicates whether to allow or deny the traffic that matches the rule.

_rnaerNetworkAclId :: Text

The ID of the ACL.

_rnaerProtocol :: Text

The IP protocol. You can specify all or -1 to mean all protocols.

_rnaerCidrBlock :: Text

The network range to allow or deny, in CIDR notation.

_rnaerDryRun :: Maybe Bool

_rnaerIcmpTypeCode :: Maybe IcmpTypeCode

ICMP protocol: The ICMP type and code.

_rnaerPortRange :: Maybe PortRange

TCP or UDP protocols: The range of ports the rule applies to.

rnaerRuleNumber :: Lens' ReplaceNetworkAclEntry Integer

rnaerRuleAction :: Lens' ReplaceNetworkAclEntry RuleAction

rnaerProtocol :: Lens' ReplaceNetworkAclEntry Text

rnaerPortRange :: Lens' ReplaceNetworkAclEntry (Maybe PortRange)

rnaerNetworkAclId :: Lens' ReplaceNetworkAclEntry Text

rnaerIcmpTypeCode :: Lens' ReplaceNetworkAclEntry (Maybe IcmpTypeCode)

rnaerEgress :: Lens' ReplaceNetworkAclEntry Bool

rnaerDryRun :: Lens' ReplaceNetworkAclEntry (Maybe Bool)

rnaerCidrBlock :: Lens' ReplaceNetworkAclEntry Text

data ReplaceNetworkAclEntryResponse

Constructors

ReplaceNetworkAclEntryResponse