Network.AWS.EC2.V2014_06_15.ReplaceRoute

Description

Replaces an existing route within a route table in a VPC. You must provide only one of the following: Internet gateway, NAT instance, VPC peering connection, or network interface. For more information about route tables, see Route Tables in the Amazon Virtual Private Cloud User Guide. Example This example replaces a route in the specified route table. The new route matches the CIDR 10.0.0.0/8 and sends the traffic to the virtual private gateway with the ID vgw-1d00376e. https:/ec2.amazonaws.com?Action=ReplaceRoute &amp;RouteTableId=rtb-e4ad488d &amp;DestinationCidrBlock=10.0.0.0/8 &amp;GatewayId=vgw-1d00376e &amp;AUTHPARAMS &lt;ReplaceRouteResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;ReplaceRouteResponse&gt;.

Request

data ReplaceRoute

Request constructor

mkReplaceRoute

Arguments

:: Text
rr1RouteTableId

-> Text
rr1DestinationCidrBlock

-> ReplaceRoute

Smart constructor for the minimum required parameters to construct a valid ReplaceRoute request.

The fields accessible through corresponding lenses are:

RouteTableId :: Text

DestinationCidrBlock :: Text

GatewayId :: Maybe Text

InstanceId :: Maybe Text

NetworkInterfaceId :: Maybe Text

VpcPeeringConnectionId :: Maybe Text

Request lenses

rr1RouteTableId :: Lens' ReplaceRoute Text

The ID of the route table.

rr1DestinationCidrBlock :: Lens' ReplaceRoute Text

The CIDR address block used for the destination match. The value you provide must match the CIDR of an existing route in the table.

rr1GatewayId :: Lens' ReplaceRoute (Maybe Text)

The ID of an Internet gateway attached to your VPC.

rr1InstanceId :: Lens' ReplaceRoute (Maybe Text)

The ID of a NAT instance in your VPC.

rr1NetworkInterfaceId :: Lens' ReplaceRoute (Maybe Text)

The ID of a network interface.

rr1VpcPeeringConnectionId :: Lens' ReplaceRoute (Maybe Text)

The ID of a VPC peering connection.

Response

data ReplaceRouteResponse

Response constructor

mkReplaceRouteResponse :: ReplaceRouteResponse

Smart constructor for the minimum required parameters to construct a valid ReplaceRouteResponse response.

This constructor is provided for convenience and testing purposes.