Network.AWS.EC2.V2014_06_15.ReplaceRouteTableAssociation

Description

Changes the route table associated with a given subnet in a VPC. After the operation completes, the subnet uses the routes in the new route table it's associated with. For more information about route tables, see Route Tables in the Amazon Virtual Private Cloud User Guide. You can also use ReplaceRouteTableAssociation to change which table is the main route table in the VPC. You just specify the main route table's association ID and the route table to be the new main route table. Example This example starts with a route table associated with a subnet, and a corresponding association ID rtbassoc-f8ad4891. You want to associate a different route table (table rtb-f9ad4890) to the subnet. The result is a new association ID representing the new association. https:/ec2.amazonaws.com?Action=ReplaceRouteTableAssociation &amp;AssociationId=rtbassoc-f8ad4891 &amp;RouteTableId=rtb-f9ad4890 &amp;AUTHPARAMS 59dbff89-35bd-4eac-99ed-be587EXAMPLE rtbassoc-faad4893.

Documentation

replaceRouteTableAssociation

Arguments

:: Text
_rrtarAssociationId

-> Text
_rrtarRouteTableId

-> ReplaceRouteTableAssociation

Minimum specification for a ReplaceRouteTableAssociation request.

data ReplaceRouteTableAssociation

Constructors

ReplaceRouteTableAssociation

Fields

_rrtarAssociationId :: Text

The association ID.

_rrtarRouteTableId :: Text

The ID of the new route table to associate with the subnet.

_rrtarDryRun :: Maybe Bool

rrtarRouteTableId :: Lens' ReplaceRouteTableAssociation Text

rrtarDryRun :: Lens' ReplaceRouteTableAssociation (Maybe Bool)

rrtarAssociationId :: Lens' ReplaceRouteTableAssociation Text

data ReplaceRouteTableAssociationResponse

Constructors

ReplaceRouteTableAssociationResponse

Fields

_rrtasNewAssociationId :: Maybe Text

The ID of the new association.

rrtasNewAssociationId :: Iso' ReplaceRouteTableAssociationResponse (Maybe Text)