Network.AWS.EC2.V2014_06_15.ResetImageAttribute

Description

Resets an attribute of an AMI to its default value. Example This example resets the launchPermission attribute for the specified AMI. https:/ec2.amazonaws.com?Action=ResetImageAttribute &amp;ImageId=ami-61a54008 &amp;Attribute=launchPermission &amp;AUTHPARAMS &lt;ResetImageAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2014-02-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;ResetImageAttributeResponse&gt;.

Request

data ResetImageAttribute

Request constructor

mkResetImageAttribute

Arguments

:: Text
riaImageId

-> ResetImageAttributeName
riaAttribute

-> ResetImageAttribute

Smart constructor for the minimum required parameters to construct a valid ResetImageAttribute request.

The fields accessible through corresponding lenses are:

ImageId :: Text

Attribute :: ResetImageAttributeName

Request lenses

riaImageId :: Lens' ResetImageAttribute Text

The ID of the AMI.

riaAttribute :: Lens' ResetImageAttribute ResetImageAttributeName

The attribute to reset (currently you can only reset the launch permission attribute).

Response

data ResetImageAttributeResponse

Response constructor

mkResetImageAttributeResponse :: ResetImageAttributeResponse

Smart constructor for the minimum required parameters to construct a valid ResetImageAttributeResponse response.

This constructor is provided for convenience and testing purposes.