Network.AWS.EC2.V2014_06_15.ResetInstanceAttribute

Description

Resets an attribute of an instance to its default value. To reset the kernel or RAM disk, the instance must be in a stopped state. To reset the SourceDestCheck, the instance can be either running or stopped. The SourceDestCheck attribute controls whether source/destination checking is enabled. The default value is true, which means checking is enabled. This value must be false for a NAT instance to perform NAT. For more information, see NAT Instances in the Amazon Virtual Private Cloud User Guide. Example This example resets the sourceDestCheck attribute. https:/ec2.amazonaws.com?Action=ResetInstanceAttribute &amp;InstanceId=i-1a2b3c4d &amp;Attribute=sourceDestCheck &amp;AUTHPARAMS &lt;ResetInstanceAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;ResetInstanceAttributeResponse&gt;.

Request

data ResetInstanceAttribute

Request constructor

mkResetInstanceAttribute

Arguments

:: Text
ria1InstanceId

-> InstanceAttributeName
ria1Attribute

-> ResetInstanceAttribute

Smart constructor for the minimum required parameters to construct a valid ResetInstanceAttribute request.

The fields accessible through corresponding lenses are:

InstanceId :: Text

Attribute :: InstanceAttributeName

Request lenses

ria1InstanceId :: Lens' ResetInstanceAttribute Text

The ID of the instance.

ria1Attribute :: Lens' ResetInstanceAttribute InstanceAttributeName

The attribute to reset.

Response

data ResetInstanceAttributeResponse

Response constructor

mkResetInstanceAttributeResponse :: ResetInstanceAttributeResponse

Smart constructor for the minimum required parameters to construct a valid ResetInstanceAttributeResponse response.

This constructor is provided for convenience and testing purposes.