Network.AWS.EC2.V2014_06_15.ResetNetworkInterfaceAttribute

Description

Resets a network interface attribute. You can specify only one attribute at a time. Example This example resets the sourceDestCheck attribute for the specified network interface. https:/ec2.amazonaws.com?Action=ResetNetworkInterfaceAttribute &amp;NetworkInterfaceId=eni-ffda3197 &amp;Attribute=sourceDestCheck &amp;AUTHPARAMS &lt;ResetNetworkInterfaceAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;5187642e-3f16-44a3-b05f-24c3848b5162&lt;/requestId&gt; &lt;return&gt;true&lt;/return&gt; &lt;/ResetNetworkInterfaceAttributeResponse&gt;.

Documentation

resetNetworkInterfaceAttribute

Arguments

:: Text
_rniarNetworkInterfaceId

-> ResetNetworkInterfaceAttribute

Minimum specification for a ResetNetworkInterfaceAttribute request.

data ResetNetworkInterfaceAttribute

Constructors

ResetNetworkInterfaceAttribute

Fields

_rniarNetworkInterfaceId :: Text

The ID of the network interface.

_rniarDryRun :: Maybe Bool

_rniarSourceDestCheck :: Maybe Text

Indicates whether source/destination checking is enabled. A value of true means checking is enabled, and false means checking is disabled. This value must be false for a NAT instance to perform NAT.

rniarSourceDestCheck :: Lens' ResetNetworkInterfaceAttribute (Maybe Text)

rniarNetworkInterfaceId :: Lens' ResetNetworkInterfaceAttribute Text

rniarDryRun :: Lens' ResetNetworkInterfaceAttribute (Maybe Bool)

data ResetNetworkInterfaceAttributeResponse

Constructors

ResetNetworkInterfaceAttributeResponse