Network.AWS.EC2.V2014_06_15.ResetSnapshotAttribute

Description

Resets permission settings for the specified snapshot. For more information on modifying snapshot permissions, see Sharing Snapshots in the Amazon Elastic Compute Cloud User Guide. Example This example resets the permissions for snap-1a2b3c4d, making it a private snapshot that can only be used by the account that created it. https:/ec2.amazonaws.com?Action=ResetSnapshotAttribute &amp;SnapshotId=snap-1a2b3c4d &amp;Attribute=createVolumePermission &amp;AUTHPARAMS &lt;ResetSnapshotAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2014-05-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;ResetSnapshotAttributeResponse&gt;.

Documentation

resetSnapshotAttribute

Arguments

:: SnapshotAttributeName
_rsarAttribute

-> Text
_rsarSnapshotId

-> ResetSnapshotAttribute

Minimum specification for a ResetSnapshotAttribute request.

data ResetSnapshotAttribute

Constructors

ResetSnapshotAttribute

Fields

_rsarAttribute :: SnapshotAttributeName

The attribute to reset (currently only the attribute for permission to create volumes can be reset).

_rsarSnapshotId :: Text

The ID of the snapshot.

_rsarDryRun :: Maybe Bool

rsarSnapshotId :: Lens' ResetSnapshotAttribute Text

rsarDryRun :: Lens' ResetSnapshotAttribute (Maybe Bool)

rsarAttribute :: Lens' ResetSnapshotAttribute SnapshotAttributeName

data ResetSnapshotAttributeResponse

Constructors

ResetSnapshotAttributeResponse