Network.AWS.EC2.V2014_06_15.RevokeSecurityGroupEgress

Description

Removes one or more egress rules from a security group for EC2-VPC. The values that you specify in the revoke request (for example, ports) must match the existing rule's values for the rule to be revoked. Each rule consists of the protocol and the CIDR range or source security group. For the TCP and UDP protocols, you must also specify the destination port or range of ports. For the ICMP protocol, you must also specify the ICMP type and code. Rule changes are propagated to instances within the security group as quickly as possible. However, a small delay might occur. Example 1 This example revokes the access that the specified security group has to the 205.192.0.016 and 205.159.0.016 address ranges on TCP port 80. https:/ec2.amazonaws.com?Action=RevokeSecurityGroupEgress &amp;GroupId=sg-1a2b3c4d &amp;IpPermissions.1.IpProtocol=tcp &amp;IpPermissions.1.FromPort=80 &amp;IpPermissions.1.ToPort=80 &amp;IpPermissions.1.IpRanges.1.CidrIp=205.192.0.0/16 &amp;IpPermissions.1.IpRanges.2.CidrIp=205.159.0.0/16 &amp;AUTHPARAMS Example 2 This example revokes the access that the specified security group has to the security group with the ID sg-9a8d7f5c on TCP port 1433. https:/ec2.amazonaws.com?Action=RevokeSecurityGroupEgress &amp;GroupId=sg-1a2b3c4d &amp;IpPermissions.1.IpProtocol=tcp &amp;IpPermissions.1.FromPort=1433 &amp;IpPermissions.1.ToPort=1433 &amp;IpPermissions.1.Groups.1.GroupId=sg-9a8d7f5c &amp;AUTHPARAMS.

Request

data RevokeSecurityGroupEgress

Request constructor

mkRevokeSecurityGroupEgress

Arguments

:: Text
rsgeGroupId

-> RevokeSecurityGroupEgress

Smart constructor for the minimum required parameters to construct a valid RevokeSecurityGroupEgress request.

Request lenses

rsgeGroupId :: Lens' RevokeSecurityGroupEgress Text

The ID of the security group.

rsgeSourceSecurityGroupName :: Lens' RevokeSecurityGroupEgress (Maybe Text)

EC2-Classic, default VPC

The name of the source security group. You can't specify a source security group and a CIDR IP address range.

rsgeSourceSecurityGroupOwnerId :: Lens' RevokeSecurityGroupEgress (Maybe Text)

The ID of the source security group. You can't specify a source security group and a CIDR IP address range.

rsgeIpProtocol :: Lens' RevokeSecurityGroupEgress (Maybe Text)

The IP protocol name (tcp, udp, icmp) or number (see Protocol Numbers). Use -1 to specify all.

rsgeFromPort :: Lens' RevokeSecurityGroupEgress (Maybe Integer)

The start of port range for the TCP and UDP protocols, or an ICMP type number. For the ICMP type number, use -1 to specify all ICMP types.

rsgeToPort :: Lens' RevokeSecurityGroupEgress (Maybe Integer)

The end of port range for the TCP and UDP protocols, or an ICMP code number. For the ICMP code number, use -1 to specify all ICMP codes for the ICMP type.

rsgeCidrIp :: Lens' RevokeSecurityGroupEgress (Maybe Text)

The CIDR IP address range. You can't specify this parameter when specifying a source security group.

rsgeIpPermissions :: Lens' RevokeSecurityGroupEgress [IpPermission]

Response

data RevokeSecurityGroupEgressResponse