Network.AWS.EC2.V2014_06_15.RevokeSecurityGroupIngress

Description

Removes one or more ingress rules from a security group. The values that you specify in the revoke request (for example, ports) must match the existing rule's values for the rule to be removed. Each rule consists of the protocol and the CIDR range or source security group. For the TCP and UDP protocols, you must also specify the destination port or range of ports. For the ICMP protocol, you must also specify the ICMP type and code. Rule changes are propagated to instances within the security group as quickly as possible. However, a small delay might occur. Example This example revokes TCP port 80 access from the 205.192.0.0/16 address range for the security group named websrv. If the security group is for a VPC, specify the ID of the security group instead of the name. https:/ec2.amazonaws.com?Action=RevokeSecurityGroupIngress &amp;GroupName=websrv &amp;IpProtocol=tcp &amp;FromPort=80 &amp;ToPort=80 &amp;CidrIp=205.192.0.0/16 &amp;AUTHPARAMS &lt;RevokeSecurityGroupIngressResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;/return&gt; &lt;/RevokeSecurityGroupIngressResponse&gt;.

Request

data RevokeSecurityGroupIngress

Request constructor

mkRevokeSecurityGroupIngress :: RevokeSecurityGroupIngress

Smart constructor for the minimum required parameters to construct a valid RevokeSecurityGroupIngress request.

Request lenses

rsgiGroupName :: Lens' RevokeSecurityGroupIngress (Maybe Text)

EC2-Classic, default VPC

The name of the security group.

rsgiGroupId :: Lens' RevokeSecurityGroupIngress (Maybe Text)

The ID of the security group.

rsgiSourceSecurityGroupName :: Lens' RevokeSecurityGroupIngress (Maybe Text)

EC2-Classic, default VPC

The name of the source security group. You can't specify a source security group and a CIDR IP address range.

rsgiSourceSecurityGroupOwnerId :: Lens' RevokeSecurityGroupIngress (Maybe Text)

The ID of the source security group. You can't specify a source security group and a CIDR IP address range.

rsgiIpProtocol :: Lens' RevokeSecurityGroupIngress (Maybe Text)

The IP protocol name (tcp, udp, icmp) or number (see Protocol Numbers). Use -1 to specify all.

rsgiFromPort :: Lens' RevokeSecurityGroupIngress (Maybe Integer)

The start of port range for the TCP and UDP protocols, or an ICMP type number. For the ICMP type number, use -1 to specify all ICMP types.

rsgiToPort :: Lens' RevokeSecurityGroupIngress (Maybe Integer)

The end of port range for the TCP and UDP protocols, or an ICMP code number. For the ICMP code number, use -1 to specify all ICMP codes for the ICMP type.

rsgiCidrIp :: Lens' RevokeSecurityGroupIngress (Maybe Text)

The CIDR IP address range. You can't specify this parameter when specifying a source security group.

rsgiIpPermissions :: Lens' RevokeSecurityGroupIngress [IpPermission]

Response

data RevokeSecurityGroupIngressResponse