Network.AWS.EC2.V2014_06_15.RunInstances

Description

Launches the specified number of instances using an AMI for which you have permissions. When you launch an instance, it enters the pending state. After the instance is ready for you, it enters the running state. To check the state of your instance, call DescribeInstances. If you don't specify a security group when launching an instance, Amazon EC2 uses the default security group. For more information, see Security Groups in the Amazon Elastic Compute Cloud User Guide. Linux instances have access to the public key of the key pair at boot. You can use this key to provide secure access to the instance. Amazon EC2 public images use this feature to provide secure access without passwords. For more information, see Key Pairs in the Amazon Elastic Compute Cloud User Guide. You can provide optional user data when launching an instance. For more information, see Instance Metadata in the Amazon Elastic Compute Cloud User Guide. If any of the AMIs have a product code attached for which the user has not subscribed, RunInstances fails. T2 instance types can only be launched into a VPC. If you do not have a default VPC, or if you do not specify a subnet ID in the request, RunInstances fails. For more information about troubleshooting, see What To Do If An Instance Immediately Terminates, and Troubleshooting Connecting to Your Instance in the Amazon Elastic Compute Cloud User Guide. This example launches three instances using the AMI with the ID ami-60a54009. https:/ec2.amazonaws.com?Action=RunInstances &amp;ImageId=ami-60a54009 &amp;MaxCount=3 &amp;MinCount=1 &amp;KeyName=my-key-pair &amp;Placement.AvailabilityZone=us-east-1d &amp;AUTHPARAMS This example launches an m1.small instance into a subnet. Because no network interface is specified, the default network interface is used. https:/ec2.amazonaws.com?Action=RunInstances &amp;ImageId=ami-31814f58 &amp;InstanceType=m1.small &amp;MaxCount=1 &amp;MinCount=1 &amp;KeyName=my-key-pair &amp;SubnetId=subnet-b2a249da &amp;AUTHPARAMS This example launches an m1.large instance into a subnet. The network interface specifies a primary private IP address of 10.0.2.106 and two secondary private IP addresses (10.0.2.107 and 10.0.2.108). https:/ec2.amazonaws.com?Action=RunInstances &amp;ImageId=ami-beb0caec &amp;InstanceType=m1.large &amp;MaxCount=1 &amp;MinCount=1 &amp;KeyName=my-key-pair &amp;NetworkInterface.0.DeviceIndex=0 &amp;NetworkInterface.0.PrivateIpAddresses.0.Primary=true &amp;NetworkInterface.0.PrivateIpAddresses.0.PrivateIpAddress=10.0.2.106 &amp;NetworkInterface.0.PrivateIpAddresses.1.Primary=false &amp;NetworkInterface.0.PrivateIpAddresses.1.PrivateIpAddress=10.0.2.107 &amp;NetworkInterface.0.PrivateIpAddresses.2.Primary=false &amp;NetworkInterface.0.PrivateIpAddresses.2.PrivateIpAddress=10.0.2.108 &amp;NetworkInterface.0.SubnetId=subnet-a61dafcf &amp;AUTHPARAMS This example launches a Dedicated Instance into the specified subnet. https:/ec2.amazonaws.com?Action=RunInstances &amp;ImageId=ami-2a1fec43 &amp;MaxCount=1 &amp;MinCount=1 &amp;KeyName=my-key-pair &amp;SubnetId=subnet-dea63cb7 &amp;Placement.Tenancy=dedicated &amp;AUTHPARAMS This request launches an instance into a nondefault subnet, and requests a public IP address for a new network interface with the device index of 0. https:/ec2.amazonaws.com?Action=RunInstances &amp;ImageId=ami-1a2b3c4d &amp;MaxCount=1 &amp;MinCount=1 &amp;NetworkInterface.0.DeviceIndex=0 &amp;NetworkInterface.0.AssociatePublicIpAddress=true &amp;NetworkInterface.0.SubnetId=subnet-1a2b3c4d &amp;AUTHPARAMS This request launches an m1.large instance with a block device mapping. There are two instance store volumes mapped to devsdc and devsdd, and a 100 GB Amazon EBS volume mapped to devsdf. https:/ec2.amazonaws.com?Action=RunInstances &amp;ImageId=ami-1a2b3c4d &amp;InstanceType=m1.large &amp;BlockDeviceMapping.1.DeviceName=%2Fdev%2Fsdc &amp;BlockDeviceMapping.1.VirtualName=ephemeral0 &amp;BlockDeviceMapping.2.DeviceName=%2Fdev%2Fsdd &amp;BlockDeviceMapping.2.VirtualName=ephemeral1 &amp;BlockDeviceMapping.3.DeviceName=%2Fdev%2Fsdf &amp;BlockDeviceMapping.3.Ebs.DeleteOnTermination=false &amp;BlockDeviceMapping.3.Ebs.VolumeSize=100 &amp;EbsOptimized=false &amp;MinCount=1 &amp;MaxCount=1 &amp;DisableApiTermination=false &amp;Monitoring.Enabled=false &amp;AUTHPARAMS.

Request

data RunInstances

Request constructor

mkRunInstances

Arguments

:: Text
ri3ImageId

-> Integer
ri3MinCount

-> Integer
ri3MaxCount

-> RunInstances

Smart constructor for the minimum required parameters to construct a valid RunInstances request.

Request lenses

ri3ImageId :: Lens' RunInstances Text

The ID of the AMI, which you can get by calling DescribeImages.

ri3MinCount :: Lens' RunInstances Integer

The minimum number of instances to launch. If you specify a minimum that is more instances than Amazon EC2 can launch in the target Availability Zone, Amazon EC2 launches no instances. Constraints: Between 1 and the maximum number you're allowed for the specified instance type. For more information about the default limits, and how to request an increase, see How many instances can I run in Amazon EC2 in the Amazon EC2 General FAQ.

ri3MaxCount :: Lens' RunInstances Integer

The maximum number of instances to launch. If you specify more instances than Amazon EC2 can launch in the target Availability Zone, Amazon EC2 launches the largest possible number of instances above MinCount. Constraints: Between 1 and the maximum number you're allowed for the specified instance type. For more information about the default limits, and how to request an increase, see How many instances can I run in Amazon EC2 in the Amazon EC2 General FAQ.

ri3KeyName :: Lens' RunInstances (Maybe Text)

The name of the key pair. You can create a key pair using CreateKeyPair or ImportKeyPair. If you launch an instance without specifying a key pair, you can't connect to the instance.

ri3SecurityGroups :: Lens' RunInstances [Text]

EC2-Classic, default VPC

One or more security group names. For a nondefault VPC, you must use security group IDs instead. Default: Amazon EC2 uses the default security group.

ri3SecurityGroupIds :: Lens' RunInstances [Text]

One or more security group IDs. You can create a security group using CreateSecurityGroup. Default: Amazon EC2 uses the default security group.

ri3UserData :: Lens' RunInstances (Maybe ByteString)

The user data for the instances. You can specify the user data as a string, or if the user data contents are in a file, you can use file://filename.

ri3InstanceType :: Lens' RunInstances (Maybe InstanceType)

The instance type. For more information, see Instance Types in the Amazon Elastic Compute Cloud User Guide. Default: m1.small.

ri3Placement :: Lens' RunInstances (Maybe Placement)

The placement for the instance.

ri3KernelId :: Lens' RunInstances (Maybe Text)

The ID of the kernel. We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB: A New Amazon Kernel Image in the Amazon Elastic Compute Cloud User Guide.

ri3RamdiskId :: Lens' RunInstances (Maybe Text)

The ID of the RAM disk.

ri3BlockDeviceMappings :: Lens' RunInstances [BlockDeviceMapping]

The block device mapping.

ri3Monitoring :: Lens' RunInstances (Maybe RunInstancesMonitoringEnabled)

The monitoring for the instance.

ri3SubnetId :: Lens' RunInstances (Maybe Text)

EC2-VPC

The ID of the subnet to launch the instance into.

ri3DisableApiTermination :: Lens' RunInstances (Maybe Bool)

If you set this parameter to true, you can't terminate the instance using the Amazon EC2 console, CLI, or API; otherwise, you can. If you set this parameter to true and then later want to be able to terminate the instance, you must first change the value of the disableApiTermination attribute to false using ModifyInstanceAttribute. Alternatively, if you set InstanceInitiatedShutdownBehavior to terminate, you can terminate the instance by running the shutdown command from the instance. Default: false.

ri3InstanceInitiatedShutdownBehavior :: Lens' RunInstances (Maybe ShutdownBehavior)

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown). Default: stop.

ri3PrivateIpAddress :: Lens' RunInstances (Maybe Text)

EC2-VPC

The primary IP address. You must specify a value from the IP address range of the subnet. Only one private IP address can be designated as primary. Therefore, you can't specify this parameter if PrivateIpAddresses.n.Primary is set to true and PrivateIpAddresses.n.PrivateIpAddress is set to an IP address. Default: We select an IP address from the IP address range of the subnet.

ri3ClientToken :: Lens' RunInstances (Maybe Text)

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency in the Amazon Elastic Compute Cloud User Guide. Constraints: Maximum 64 ASCII characters.

ri3AdditionalInfo :: Lens' RunInstances (Maybe Text)

Reserved.

ri3NetworkInterfaces :: Lens' RunInstances [InstanceNetworkInterfaceSpecification]

One or more network interfaces.

ri3IamInstanceProfile :: Lens' RunInstances (Maybe IamInstanceProfileSpecification)

The IAM instance profile.

ri3EbsOptimized :: Lens' RunInstances (Maybe Bool)

Indicates whether the instance is optimized for EBS I/O. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal Amazon EBS I/O performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS-optimized instance. Default: false.

Response

data RunInstancesResponse

One or more reservations.

Response constructor

mkRunInstancesResponse :: RunInstancesResponse

Smart constructor for the minimum required parameters to construct a valid RunInstancesResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

rirrReservationId :: Lens' RunInstancesResponse (Maybe Text)

The ID of the reservation.

rirrOwnerId :: Lens' RunInstancesResponse (Maybe Text)

The ID of the AWS account that owns the reservation.

rirrRequesterId :: Lens' RunInstancesResponse (Maybe Text)

The ID of the requester that launched the instances on your behalf (for example, AWS Management Console or Auto Scaling).

rirrGroups :: Lens' RunInstancesResponse [GroupIdentifier]

One or more security groups.

rirrInstances :: Lens' RunInstancesResponse [Instance]

One or more instances.