Network.AWS.EC2.V2014_06_15.TerminateInstances

Description

Shuts down one or more instances. This operation is idempotent; if you terminate an instance more than once, each call succeeds. Terminated instances remain visible after termination (for approximately one hour). By default, Amazon EC2 deletes all Amazon EBS volumes that were attached when the instance launched. Volumes attached after instance launch continue running. You can stop, start, and terminate EBS-backed instances. You can only terminate instance store-backed instances. What happens to an instance differs if you stop it or terminate it. For example, when you stop an instance, the root device and any other devices attached to the instance persist. When you terminate an instance, the root device and any other devices attached during the instance launch are automatically deleted. For more information about the differences between stopping and terminating instances, see Instance Lifecycle in the Amazon Elastic Compute Cloud User Guide. For more information about troubleshooting, see Troubleshooting Terminating Your Instance in the Amazon Elastic Compute Cloud User Guide. Example This example terminates the specified instance. https:/ec2.amazonaws.com?Action=TerminateInstances &amp;InstanceId.1=i-3ea74257 &amp;AUTHPARAMS &lt;TerminateInstancesResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;instancesSet&gt; &lt;item&gt; &lt;instanceId&gt;i-3ea74257&lt;/instanceId&gt; &lt;currentState&gt; &lt;code&gt;32&lt;code&gt; &lt;name&gt;shutting-down&lt;name&gt; &lt;currentState&gt; &lt;previousState&gt; &lt;code&gt;16&lt;code&gt; &lt;name&gt;running&lt;name&gt; &lt;previousState&gt; &lt;/item&gt; &lt;instancesSet&gt; &lt;TerminateInstancesResponse&gt;.

Request

data TerminateInstances

Request constructor

mkTerminateInstances

Arguments

:: [Text]
tiInstanceIds

-> TerminateInstances

Smart constructor for the minimum required parameters to construct a valid TerminateInstances request.

Request lenses

tiInstanceIds :: Lens' TerminateInstances [Text]

One or more instance IDs.

Response

data TerminateInstancesResponse

Response lenses

tirTerminatingInstances :: Lens' TerminateInstancesResponse [InstanceStateChange]

Information about one or more terminated instances.