Network.AWS.EC2.V2014_06_15.UnassignPrivateIpAddresses

Description

Unassigns one or more secondary private IP addresses from a network interface. Example The following example unassigns two secondary private IP addresses from the specified network interface. https:/ec2.amazonaws.com?Action=UnassignPrivateIpAddresses &amp;NetworkInterfaceId=eni-197d9972 &amp;PrivateIpAddress.0=10.0.2.60 &amp;PrivateIpAddress.1=10.0.2.65 &amp;AUTHPARAMS &lt;UnassignPrivateIpAddresses xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;UnassignPrivateIpAddresses&gt;.

Documentation

data UnassignPrivateIpAddresses

Constructors

UnassignPrivateIpAddresses

Fields

_upiarPrivateIpAddresses :: [Text]

The secondary private IP addresses to unassign from the network interface. You can specify this option multiple times to unassign more than one IP address.

_upiarNetworkInterfaceId :: Text

The ID of the network interface.

upiarPrivateIpAddresses :: Lens' UnassignPrivateIpAddresses [Text]

upiarNetworkInterfaceId :: Lens' UnassignPrivateIpAddresses Text

data UnassignPrivateIpAddressesResponse

Constructors

UnassignPrivateIpAddressesResponse