Network.AWS.EC2.AcceptVPCPeeringConnection

Description

Accept a VPC peering connection request. To accept a request, the VPC peering connection must be in the pending-acceptance state, and you must be the owner of the peer VPC. Use DescribeVpcPeeringConnections to view your outstanding VPC peering connection requests.

Creating a Request

acceptVPCPeeringConnection :: AcceptVPCPeeringConnection

Creates a value of AcceptVPCPeeringConnection with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

avpcVPCPeeringConnectionId - The ID of the VPC peering connection.

avpcDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data AcceptVPCPeeringConnection

Contains the parameters for AcceptVpcPeeringConnection.

See: acceptVPCPeeringConnection smart constructor.

Request Lenses

avpcVPCPeeringConnectionId :: Lens' AcceptVPCPeeringConnection (Maybe Text)

The ID of the VPC peering connection.

avpcDryRun :: Lens' AcceptVPCPeeringConnection (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

acceptVPCPeeringConnectionResponse

Arguments

:: Int
avpcrsResponseStatus

-> AcceptVPCPeeringConnectionResponse

Creates a value of AcceptVPCPeeringConnectionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

avpcrsVPCPeeringConnection - Information about the VPC peering connection.

avpcrsResponseStatus - -- | The response status code.

data AcceptVPCPeeringConnectionResponse

Contains the output of AcceptVpcPeeringConnection.

See: acceptVPCPeeringConnectionResponse smart constructor.

Response Lenses

avpcrsVPCPeeringConnection :: Lens' AcceptVPCPeeringConnectionResponse (Maybe VPCPeeringConnection)

Information about the VPC peering connection.

avpcrsResponseStatus :: Lens' AcceptVPCPeeringConnectionResponse Int

- | The response status code.