Network.AWS.EC2.AllocateAddress

Description

Acquires an Elastic IP address.

An Elastic IP address is for use either in the EC2-Classic platform or in a VPC. For more information, see Elastic IP Addresses in the Amazon ElasticCompute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-AllocateAddress.html

Request

data AllocateAddress

Request constructor

allocateAddress :: AllocateAddress

AllocateAddress constructor.

The fields accessible through corresponding lenses are:

aaDomain :: Maybe DomainType

aaDryRun :: Maybe Bool

Request lenses

aaDomain :: Lens' AllocateAddress (Maybe DomainType)

Set to vpc to allocate the address for use with instances in a VPC.

Default: The address is for use with instances in EC2-Classic.

aaDryRun :: Lens' AllocateAddress (Maybe Bool)

Response

data AllocateAddressResponse

Response constructor

allocateAddressResponse :: AllocateAddressResponse

AllocateAddressResponse constructor.

The fields accessible through corresponding lenses are:

aarAllocationId :: Maybe Text

aarDomain :: Maybe DomainType

aarPublicIp :: Maybe Text

Response lenses

aarAllocationId :: Lens' AllocateAddressResponse (Maybe Text)

EC2-VPC

The ID that AWS assigns to represent the allocation of the Elastic IP address for use with instances in a VPC.

aarDomain :: Lens' AllocateAddressResponse (Maybe DomainType)

Indicates whether this Elastic IP address is for use with instances in EC2-Classic (standard) or instances in a VPC (vpc).

aarPublicIp :: Lens' AllocateAddressResponse (Maybe Text)

The Elastic IP address.