Network.AWS.EC2.AssignPrivateIpAddresses

Description

Assigns one or more secondary private IP addresses to the specified network interface. You can specify one or more specific secondary IP addresses, or you can specify the number of secondary IP addresses to be automatically assigned within the subnet's CIDR block range. The number of secondary IP addresses that you can assign to an instance varies by instance type. For information about instance types, see Instance Types in the Amazon ElasticCompute Cloud User Guide for Linux. For more information about Elastic IP addresses, see Elastic IP Addresses in the Amazon Elastic Compute Cloud UserGuide for Linux.

AssignPrivateIpAddresses is available only in EC2-VPC.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-AssignPrivateIpAddresses.html

Request

data AssignPrivateIpAddresses

Request constructor

assignPrivateIpAddresses

Arguments

:: Text
apiaNetworkInterfaceId

-> AssignPrivateIpAddresses

AssignPrivateIpAddresses constructor.

The fields accessible through corresponding lenses are:

apiaAllowReassignment :: Maybe Bool

apiaNetworkInterfaceId :: Text

apiaPrivateIpAddresses :: [Text]

apiaSecondaryPrivateIpAddressCount :: Maybe Int

Request lenses

apiaAllowReassignment :: Lens' AssignPrivateIpAddresses (Maybe Bool)

Indicates whether to allow an IP address that is already assigned to another network interface or instance to be reassigned to the specified network interface.

apiaNetworkInterfaceId :: Lens' AssignPrivateIpAddresses Text

The ID of the network interface.

apiaPrivateIpAddresses :: Lens' AssignPrivateIpAddresses [Text]

One or more IP addresses to be assigned as a secondary private IP address to the network interface. You can't specify this parameter when also specifying a number of secondary IP addresses.

If you don't specify an IP address, Amazon EC2 automatically selects an IP address within the subnet range.

apiaSecondaryPrivateIpAddressCount :: Lens' AssignPrivateIpAddresses (Maybe Int)

The number of secondary IP addresses to assign to the network interface. You can't specify this parameter when also specifying private IP addresses.

Response

data AssignPrivateIpAddressesResponse

Response constructor

assignPrivateIpAddressesResponse :: AssignPrivateIpAddressesResponse

AssignPrivateIpAddressesResponse constructor.