Network.AWS.EC2.AssociateAddress

Description

Associates an Elastic IP address with an instance or a network interface.

An Elastic IP address is for use in either the EC2-Classic platform or in a VPC. For more information, see Elastic IP Addresses in the Amazon Elastic Compute Cloud User Guide .

EC2-Classic, VPC in an EC2-VPC-only account

If the Elastic IP address is already associated with a different instance, it is disassociated from that instance and associated with the specified instance. If you associate an Elastic IP address with an instance that has an existing Elastic IP address, the existing address is disassociated from the instance, but remains allocated to your account.

VPC in an EC2-Classic account

If you don't specify a private IP address, the Elastic IP address is associated with the primary IP address. If the Elastic IP address is already associated with a different instance or a network interface, you get an error unless you allow reassociation. You cannot associate an Elastic IP address with an instance or network interface that has an existing Elastic IP address.

Important: This is an idempotent operation. If you perform the operation more than once, Amazon EC2 doesn't return an error, and you may be charged for each time the Elastic IP address is remapped to the same instance. For more information, see the Elastic IP Addresses section of Amazon EC2 Pricing .

Creating a Request

associateAddress :: AssociateAddress

Creates a value of AssociateAddress with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aasInstanceId - The ID of the instance. This is required for EC2-Classic. For EC2-VPC, you can specify either the instance ID or the network interface ID, but not both. The operation fails if you specify an instance ID unless exactly one network interface is attached.

aasAllocationId - [EC2-VPC] The allocation ID. This is required for EC2-VPC.

aasNetworkInterfaceId - [EC2-VPC] The ID of the network interface. If the instance has more than one network interface, you must specify a network interface ID.

aasAllowReassociation - [EC2-VPC] For a VPC in an EC2-Classic account, specify true to allow an Elastic IP address that is already associated with an instance or network interface to be reassociated with the specified instance or network interface. Otherwise, the operation fails. In a VPC in an EC2-VPC-only account, reassociation is automatic, therefore you can specify false to ensure the operation fails if the Elastic IP address is already associated with another resource.

aasPrivateIPAddress - [EC2-VPC] The primary or secondary private IP address to associate with the Elastic IP address. If no private IP address is specified, the Elastic IP address is associated with the primary private IP address.

aasPublicIP - The Elastic IP address. This is required for EC2-Classic.

aasDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data AssociateAddress

Contains the parameters for AssociateAddress.

See: associateAddress smart constructor.

Request Lenses

aasInstanceId :: Lens' AssociateAddress (Maybe Text)

The ID of the instance. This is required for EC2-Classic. For EC2-VPC, you can specify either the instance ID or the network interface ID, but not both. The operation fails if you specify an instance ID unless exactly one network interface is attached.

aasAllocationId :: Lens' AssociateAddress (Maybe Text)

EC2-VPC

The allocation ID. This is required for EC2-VPC.

aasNetworkInterfaceId :: Lens' AssociateAddress (Maybe Text)

EC2-VPC

The ID of the network interface. If the instance has more than one network interface, you must specify a network interface ID.

aasAllowReassociation :: Lens' AssociateAddress (Maybe Bool)

EC2-VPC

For a VPC in an EC2-Classic account, specify true to allow an Elastic IP address that is already associated with an instance or network interface to be reassociated with the specified instance or network interface. Otherwise, the operation fails. In a VPC in an EC2-VPC-only account, reassociation is automatic, therefore you can specify false to ensure the operation fails if the Elastic IP address is already associated with another resource.

aasPrivateIPAddress :: Lens' AssociateAddress (Maybe Text)

EC2-VPC

The primary or secondary private IP address to associate with the Elastic IP address. If no private IP address is specified, the Elastic IP address is associated with the primary private IP address.

aasPublicIP :: Lens' AssociateAddress (Maybe Text)

The Elastic IP address. This is required for EC2-Classic.

aasDryRun :: Lens' AssociateAddress (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

associateAddressResponse

Arguments

:: Int
arsResponseStatus

-> AssociateAddressResponse

Creates a value of AssociateAddressResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

arsAssociationId - [EC2-VPC] The ID that represents the association of the Elastic IP address with an instance.

arsResponseStatus - -- | The response status code.

data AssociateAddressResponse

Contains the output of AssociateAddress.

See: associateAddressResponse smart constructor.

Response Lenses

arsAssociationId :: Lens' AssociateAddressResponse (Maybe Text)

EC2-VPC

The ID that represents the association of the Elastic IP address with an instance.

arsResponseStatus :: Lens' AssociateAddressResponse Int

- | The response status code.