Network.AWS.EC2.AssociateAddress

Description

Associates an Elastic IP address with an instance or a network interface.

An Elastic IP address is for use in either the EC2-Classic platform or in a VPC. For more information, see Elastic IP Addresses in the Amazon ElasticCompute Cloud User Guide for Linux.

EC2-Classic, VPC in an EC2-VPC-only account

If the Elastic IP address is already associated with a different instance, it is disassociated from that instance and associated with the specified instance.

VPC in an EC2-Classic account

If you don't specify a private IP address, the Elastic IP address is associated with the primary IP address. If the Elastic IP address is already associated with a different instance or a network interface, you get an error unless you allow reassociation.

This is an idempotent operation. If you perform the operation more than once, Amazon EC2 doesn't return an error.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-AssociateAddress.html

Request

data AssociateAddress

Request constructor

associateAddress :: AssociateAddress

AssociateAddress constructor.

The fields accessible through corresponding lenses are:

aa1AllocationId :: Maybe Text

aa1AllowReassociation :: Maybe Bool

aa1DryRun :: Maybe Bool

aa1InstanceId :: Maybe Text

aa1NetworkInterfaceId :: Maybe Text

aa1PrivateIpAddress :: Maybe Text

aa1PublicIp :: Maybe Text

Request lenses

aa1AllocationId :: Lens' AssociateAddress (Maybe Text)

EC2-VPC

The allocation ID. This is required for EC2-VPC.

aa1AllowReassociation :: Lens' AssociateAddress (Maybe Bool)

EC2-VPC

Allows an Elastic IP address that is already associated with an instance or network interface to be re-associated with the specified instance or network interface. Otherwise, the operation fails.

Default: false

aa1DryRun :: Lens' AssociateAddress (Maybe Bool)

aa1InstanceId :: Lens' AssociateAddress (Maybe Text)

The ID of the instance. This is required for EC2-Classic. For EC2-VPC, you can specify either the instance ID or the network interface ID, but not both. The operation fails if you specify an instance ID unless exactly one network interface is attached.

aa1NetworkInterfaceId :: Lens' AssociateAddress (Maybe Text)

EC2-VPC

The ID of the network interface. If the instance has more than one network interface, you must specify a network interface ID.

aa1PrivateIpAddress :: Lens' AssociateAddress (Maybe Text)

EC2-VPC

The primary or secondary private IP address to associate with the Elastic IP address. If no private IP address is specified, the Elastic IP address is associated with the primary private IP address.

aa1PublicIp :: Lens' AssociateAddress (Maybe Text)

The Elastic IP address. This is required for EC2-Classic.

Response

data AssociateAddressResponse

Response constructor

associateAddressResponse :: AssociateAddressResponse

AssociateAddressResponse constructor.

The fields accessible through corresponding lenses are:

aarAssociationId :: Maybe Text

Response lenses

aarAssociationId :: Lens' AssociateAddressResponse (Maybe Text)

EC2-VPC

The ID that represents the association of the Elastic IP address with an instance.