Network.AWS.EC2.AssociateDHCPOptions

Description

Associates a set of DHCP options (that you've previously created) with the specified VPC, or associates no DHCP options with the VPC.

After you associate the options with the VPC, any existing instances and all new instances that you launch in that VPC use the options. You don't need to restart or relaunch the instances. They automatically pick up the changes within a few hours, depending on how frequently the instance renews its DHCP lease. You can explicitly renew the lease using the operating system on the instance.

For more information, see DHCP Options Sets in the Amazon Virtual Private Cloud User Guide .

Creating a Request

associateDHCPOptions

Arguments

:: Text
adoDHCPOptionsId

-> Text
adoVPCId

-> AssociateDHCPOptions

Creates a value of AssociateDHCPOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

adoDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

adoDHCPOptionsId - The ID of the DHCP options set, or default to associate no DHCP options with the VPC.

adoVPCId - The ID of the VPC.

data AssociateDHCPOptions

Contains the parameters for AssociateDhcpOptions.

See: associateDHCPOptions smart constructor.

Request Lenses

adoDryRun :: Lens' AssociateDHCPOptions (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

adoDHCPOptionsId :: Lens' AssociateDHCPOptions Text

The ID of the DHCP options set, or default to associate no DHCP options with the VPC.

adoVPCId :: Lens' AssociateDHCPOptions Text

The ID of the VPC.

Destructuring the Response

associateDHCPOptionsResponse :: AssociateDHCPOptionsResponse

Creates a value of AssociateDHCPOptionsResponse with the minimum fields required to make a request.

data AssociateDHCPOptionsResponse

See: associateDHCPOptionsResponse smart constructor.