Network.AWS.EC2.AssociateRouteTable

Description

Associates a subnet with a route table. The subnet and route table must be in the same VPC. This association causes traffic originating from the subnet to be routed according to the routes in the route table. The action returns an association ID, which you need in order to disassociate the route table from the subnet later. A route table can be associated with multiple subnets.

For more information about route tables, see Route Tables in the Amazon Virtual Private Cloud User Guide .

Creating a Request

associateRouteTable

Arguments

:: Text
artRouteTableId

-> Text
artSubnetId

-> AssociateRouteTable

Creates a value of AssociateRouteTable with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

artDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

artRouteTableId - The ID of the route table.

artSubnetId - The ID of the subnet.

data AssociateRouteTable

Contains the parameters for AssociateRouteTable.

See: associateRouteTable smart constructor.

Request Lenses

artDryRun :: Lens' AssociateRouteTable (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

artRouteTableId :: Lens' AssociateRouteTable Text

The ID of the route table.

artSubnetId :: Lens' AssociateRouteTable Text

The ID of the subnet.

Destructuring the Response

associateRouteTableResponse

Arguments

:: Int
artrsResponseStatus

-> AssociateRouteTableResponse

Creates a value of AssociateRouteTableResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

artrsAssociationId - The route table association ID (needed to disassociate the route table).

artrsResponseStatus - -- | The response status code.

data AssociateRouteTableResponse

Contains the output of AssociateRouteTable.

See: associateRouteTableResponse smart constructor.

Response Lenses

artrsAssociationId :: Lens' AssociateRouteTableResponse (Maybe Text)

The route table association ID (needed to disassociate the route table).

artrsResponseStatus :: Lens' AssociateRouteTableResponse Int

- | The response status code.