Network.AWS.EC2.AssociateRouteTable

Description

Associates a subnet with a route table. The subnet and route table must be in the same VPC. This association causes traffic originating from the subnet to be routed according to the routes in the route table. The action returns an association ID, which you need in order to disassociate the route table from the subnet later. A route table can be associated with multiple subnets.

For more information about route tables, see Route Tables in the AmazonVirtual Private Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-AssociateRouteTable.html

Request

data AssociateRouteTable

Request constructor

associateRouteTable

Arguments

:: Text
artSubnetId

-> Text
artRouteTableId

-> AssociateRouteTable

AssociateRouteTable constructor.

The fields accessible through corresponding lenses are:

artDryRun :: Maybe Bool

artRouteTableId :: Text

artSubnetId :: Text

Request lenses

artDryRun :: Lens' AssociateRouteTable (Maybe Bool)

artRouteTableId :: Lens' AssociateRouteTable Text

The ID of the route table.

artSubnetId :: Lens' AssociateRouteTable Text

The ID of the subnet.

Response

data AssociateRouteTableResponse

Response constructor

associateRouteTableResponse :: AssociateRouteTableResponse

AssociateRouteTableResponse constructor.

The fields accessible through corresponding lenses are:

artrAssociationId :: Maybe Text

Response lenses

artrAssociationId :: Lens' AssociateRouteTableResponse (Maybe Text)

The route table association ID (needed to disassociate the route table).