Network.AWS.EC2.AssociateSubnetCidrBlock

Description

Associates a CIDR block with your subnet. You can only associate a single IPv6 CIDR block with your subnet. An IPv6 CIDR block must have a prefix length of /64.

Creating a Request

associateSubnetCidrBlock

Arguments

:: Text
ascbIPv6CidrBlock

-> Text
ascbSubnetId

-> AssociateSubnetCidrBlock

Creates a value of AssociateSubnetCidrBlock with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ascbIPv6CidrBlock - The IPv6 CIDR block for your subnet. The subnet must have a /64 prefix length.

ascbSubnetId - The ID of your subnet.

data AssociateSubnetCidrBlock

See: associateSubnetCidrBlock smart constructor.

Request Lenses

ascbIPv6CidrBlock :: Lens' AssociateSubnetCidrBlock Text

The IPv6 CIDR block for your subnet. The subnet must have a /64 prefix length.

ascbSubnetId :: Lens' AssociateSubnetCidrBlock Text

The ID of your subnet.

Destructuring the Response

associateSubnetCidrBlockResponse

Arguments

:: Int
ascbrsResponseStatus

-> AssociateSubnetCidrBlockResponse

Creates a value of AssociateSubnetCidrBlockResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ascbrsSubnetId - The ID of the subnet.

ascbrsIPv6CidrBlockAssociation - Information about the IPv6 CIDR block association.

ascbrsResponseStatus - -- | The response status code.

data AssociateSubnetCidrBlockResponse

See: associateSubnetCidrBlockResponse smart constructor.

Response Lenses

ascbrsSubnetId :: Lens' AssociateSubnetCidrBlockResponse (Maybe Text)

The ID of the subnet.

ascbrsIPv6CidrBlockAssociation :: Lens' AssociateSubnetCidrBlockResponse (Maybe SubnetIPv6CidrBlockAssociation)

Information about the IPv6 CIDR block association.

ascbrsResponseStatus :: Lens' AssociateSubnetCidrBlockResponse Int

- | The response status code.