Network.AWS.EC2.AttachInternetGateway

Description

Attaches an Internet gateway to a VPC, enabling connectivity between the Internet and the VPC. For more information about your VPC and Internet gateway, see the Amazon Virtual Private Cloud User Guide .

Creating a Request

attachInternetGateway

Arguments

:: Text
aigInternetGatewayId

-> Text
aigVPCId

-> AttachInternetGateway

Creates a value of AttachInternetGateway with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aigDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

aigInternetGatewayId - The ID of the Internet gateway.

aigVPCId - The ID of the VPC.

data AttachInternetGateway

Contains the parameters for AttachInternetGateway.

See: attachInternetGateway smart constructor.

Request Lenses

aigDryRun :: Lens' AttachInternetGateway (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

aigInternetGatewayId :: Lens' AttachInternetGateway Text

The ID of the Internet gateway.

aigVPCId :: Lens' AttachInternetGateway Text

The ID of the VPC.

Destructuring the Response

attachInternetGatewayResponse :: AttachInternetGatewayResponse

Creates a value of AttachInternetGatewayResponse with the minimum fields required to make a request.

data AttachInternetGatewayResponse

See: attachInternetGatewayResponse smart constructor.