Network.AWS.EC2.AttachVolume

Description

Attaches an EBS volume to a running or stopped instance and exposes it to the instance with the specified device name.

Encrypted EBS volumes may only be attached to instances that support Amazon EBS encryption. For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide .

For a list of supported device names, see Attaching an EBS Volume to an Instance . Any device names that aren't reserved for instance store volumes can be used for EBS volumes. For more information, see Amazon EC2 Instance Store in the Amazon Elastic Compute Cloud User Guide .

If a volume has an AWS Marketplace product code:

The volume can be attached only to a stopped instance.

AWS Marketplace product codes are copied from the volume to the instance.

You must be subscribed to the product.

The instance type and operating system of the instance must support the product. For example, you can't detach a volume from a Windows instance and attach it to a Linux instance.

For an overview of the AWS Marketplace, see Introducing AWS Marketplace .

For more information about EBS volumes, see Attaching Amazon EBS Volumes in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

attachVolume

Arguments

:: Text
avDevice

-> Text
avInstanceId

-> Text
avVolumeId

-> AttachVolume

Creates a value of AttachVolume with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

avDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

avDevice - The device name to expose to the instance (for example, devsdh or xvdh ).

avInstanceId - The ID of the instance.

avVolumeId - The ID of the EBS volume. The volume and instance must be within the same Availability Zone.

data AttachVolume

Contains the parameters for AttachVolume.

See: attachVolume smart constructor.

Request Lenses

avDryRun :: Lens' AttachVolume (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

avDevice :: Lens' AttachVolume Text

The device name to expose to the instance (for example, devsdh or xvdh ).

avInstanceId :: Lens' AttachVolume Text

The ID of the instance.

avVolumeId :: Lens' AttachVolume Text

The ID of the EBS volume. The volume and instance must be within the same Availability Zone.

Destructuring the Response

volumeAttachment :: VolumeAttachment

Creates a value of VolumeAttachment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

volInstanceId - The ID of the instance.

volDeleteOnTermination - Indicates whether the EBS volume is deleted on instance termination.

volState - The attachment state of the volume.

volDevice - The device name.

volVolumeId - The ID of the volume.

volAttachTime - The time stamp when the attachment initiated.

data VolumeAttachment

Describes volume attachment details.

See: volumeAttachment smart constructor.

Response Lenses

volInstanceId :: Lens' VolumeAttachment (Maybe Text)

The ID of the instance.

volDeleteOnTermination :: Lens' VolumeAttachment (Maybe Bool)

Indicates whether the EBS volume is deleted on instance termination.

volState :: Lens' VolumeAttachment (Maybe VolumeAttachmentState)

The attachment state of the volume.

volDevice :: Lens' VolumeAttachment (Maybe Text)

The device name.

volVolumeId :: Lens' VolumeAttachment (Maybe Text)

The ID of the volume.

volAttachTime :: Lens' VolumeAttachment (Maybe UTCTime)

The time stamp when the attachment initiated.