Network.AWS.EC2.AttachVolume

Description

Attaches an Amazon EBS volume to a running or stopped instance and exposes it to the instance with the specified device name.

Encrypted Amazon EBS volumes may only be attached to instances that support Amazon EBS encryption. For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide for Linux.

For a list of supported device names, see Attaching an Amazon EBS Volume toan Instance. Any device names that aren't reserved for instance store volumes can be used for Amazon EBS volumes. For more information, see Amazon EC2Instance Store in the Amazon Elastic Compute Cloud User Guide for Linux.

If a volume has an AWS Marketplace product code:

The volume can be attached only to a stopped instance. AWS Marketplace product codes are copied from the volume to the instance. You must be subscribed to the product. The instance type and operating system of the instance must support the product. For example, you can't detach a volume from a Windows instance and attach it to a Linux instance. For an overview of the AWS Marketplace, see Introducing AWS Marketplace.

For more information about Amazon EBS volumes, see Attaching Amazon EBSVolumes in the Amazon Elastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-AttachVolume.html

Request

data AttachVolume

Request constructor

attachVolume

Arguments

:: Text
avVolumeId

-> Text
avInstanceId

-> Text
avDevice

-> AttachVolume

AttachVolume constructor.

The fields accessible through corresponding lenses are:

avDevice :: Text

avDryRun :: Maybe Bool

avInstanceId :: Text

avVolumeId :: Text

Request lenses

avDevice :: Lens' AttachVolume Text

The device name to expose to the instance (for example, '/dev/sdh' or xvdh).

avDryRun :: Lens' AttachVolume (Maybe Bool)

avInstanceId :: Lens' AttachVolume Text

The ID of the instance.

avVolumeId :: Lens' AttachVolume Text

The ID of the Amazon EBS volume. The volume and instance must be within the same Availability Zone.

Response

data AttachVolumeResponse

Response constructor

attachVolumeResponse :: AttachVolumeResponse

AttachVolumeResponse constructor.

The fields accessible through corresponding lenses are:

avrAttachTime :: Maybe UTCTime

avrDeleteOnTermination :: Maybe Bool

avrDevice :: Maybe Text

avrInstanceId :: Maybe Text

avrState :: Maybe VolumeAttachmentState

avrVolumeId :: Maybe Text

Response lenses

avrAttachTime :: Lens' AttachVolumeResponse (Maybe UTCTime)

The time stamp when the attachment initiated.

avrDeleteOnTermination :: Lens' AttachVolumeResponse (Maybe Bool)

Indicates whether the Amazon EBS volume is deleted on instance termination.

avrDevice :: Lens' AttachVolumeResponse (Maybe Text)

The device name.

avrInstanceId :: Lens' AttachVolumeResponse (Maybe Text)

The ID of the instance.

avrState :: Lens' AttachVolumeResponse (Maybe VolumeAttachmentState)

The attachment state of the volume.

avrVolumeId :: Lens' AttachVolumeResponse (Maybe Text)

The ID of the volume.