Network.AWS.EC2.AuthorizeSecurityGroupEgress

Description

EC2-VPC only

Adds one or more egress rules to a security group for use with a VPC. Specifically, this action permits instances to send traffic to one or more destination IPv4 or IPv6 CIDR address ranges, or to one or more destination security groups for the same VPC. This action doesn't apply to security groups for use in EC2-Classic. For more information, see Security Groups for Your VPC in the Amazon Virtual Private Cloud User Guide . For more information about security group limits, see Amazon VPC Limits .

Each rule consists of the protocol (for example, TCP), plus either a CIDR range or a source group. For the TCP and UDP protocols, you must also specify the destination port or port range. For the ICMP protocol, you must also specify the ICMP type and code. You can use -1 for the type or code to mean all types or all codes.

Rule changes are propagated to affected instances as quickly as possible. However, a small delay might occur.

Creating a Request

authorizeSecurityGroupEgress

Arguments

:: Text
asgeGroupId

-> AuthorizeSecurityGroupEgress

Creates a value of AuthorizeSecurityGroupEgress with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

asgeFromPort - The start of port range for the TCP and UDP protocols, or an ICMP type number. We recommend that you specify the port range in a set of IP permissions instead.

asgeIPPermissions - A set of IP permissions. You can't specify a destination security group and a CIDR IP address range.

asgeIPProtocol - The IP protocol name or number. We recommend that you specify the protocol in a set of IP permissions instead.

asgeToPort - The end of port range for the TCP and UDP protocols, or an ICMP type number. We recommend that you specify the port range in a set of IP permissions instead.

asgeCidrIP - The CIDR IPv4 address range. We recommend that you specify the CIDR range in a set of IP permissions instead.

asgeSourceSecurityGroupOwnerId - The AWS account number for a destination security group. To authorize outbound access to a destination security group, we recommend that you use a set of IP permissions instead.

asgeSourceSecurityGroupName - The name of a destination security group. To authorize outbound access to a destination security group, we recommend that you use a set of IP permissions instead.

asgeDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

asgeGroupId - The ID of the security group.

data AuthorizeSecurityGroupEgress

Contains the parameters for AuthorizeSecurityGroupEgress.

See: authorizeSecurityGroupEgress smart constructor.

Request Lenses

asgeFromPort :: Lens' AuthorizeSecurityGroupEgress (Maybe Int)

The start of port range for the TCP and UDP protocols, or an ICMP type number. We recommend that you specify the port range in a set of IP permissions instead.

asgeIPPermissions :: Lens' AuthorizeSecurityGroupEgress [IPPermission]

A set of IP permissions. You can't specify a destination security group and a CIDR IP address range.

asgeIPProtocol :: Lens' AuthorizeSecurityGroupEgress (Maybe Text)

The IP protocol name or number. We recommend that you specify the protocol in a set of IP permissions instead.

asgeToPort :: Lens' AuthorizeSecurityGroupEgress (Maybe Int)

The end of port range for the TCP and UDP protocols, or an ICMP type number. We recommend that you specify the port range in a set of IP permissions instead.

asgeCidrIP :: Lens' AuthorizeSecurityGroupEgress (Maybe Text)

The CIDR IPv4 address range. We recommend that you specify the CIDR range in a set of IP permissions instead.

asgeSourceSecurityGroupOwnerId :: Lens' AuthorizeSecurityGroupEgress (Maybe Text)

The AWS account number for a destination security group. To authorize outbound access to a destination security group, we recommend that you use a set of IP permissions instead.

asgeSourceSecurityGroupName :: Lens' AuthorizeSecurityGroupEgress (Maybe Text)

The name of a destination security group. To authorize outbound access to a destination security group, we recommend that you use a set of IP permissions instead.

asgeDryRun :: Lens' AuthorizeSecurityGroupEgress (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

asgeGroupId :: Lens' AuthorizeSecurityGroupEgress Text

The ID of the security group.

Destructuring the Response

authorizeSecurityGroupEgressResponse :: AuthorizeSecurityGroupEgressResponse

Creates a value of AuthorizeSecurityGroupEgressResponse with the minimum fields required to make a request.

data AuthorizeSecurityGroupEgressResponse

See: authorizeSecurityGroupEgressResponse smart constructor.