Network.AWS.EC2.AuthorizeSecurityGroupIngress

Description

Adds one or more ingress rules to a security group.

Rule changes are propagated to instances within the security group as quickly as possible. However, a small delay might occur.

EC2-Classic

This action gives one or more IPv4 CIDR address ranges permission to access a security group in your account, or gives one or more security groups (called the source groups ) permission to access a security group for your account. A source group can be for your own AWS account, or another. You can have up to 100 rules per group.

EC2-VPC

This action gives one or more IPv4 or IPv6 CIDR address ranges permission to access a security group in your VPC, or gives one or more other security groups (called the source groups ) permission to access a security group for your VPC. The security groups must all be for the same VPC or a peer VPC in a VPC peering connection. For more information about VPC security group limits, see Amazon VPC Limits .

Creating a Request

authorizeSecurityGroupIngress :: AuthorizeSecurityGroupIngress

Creates a value of AuthorizeSecurityGroupIngress with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

asgiFromPort - The start of port range for the TCP and UDP protocols, or an ICMPICMPv6 type number. For the ICMPICMPv6 type number, use -1 to specify all types.

asgiIPPermissions - A set of IP permissions. Can be used to specify multiple rules in a single command.

asgiIPProtocol - The IP protocol name (tcp , udp , icmp ) or number (see Protocol Numbers ). (VPC only) Use -1 to specify all protocols. If you specify -1 , or a protocol number other than tcp , udp , icmp , or 58 (ICMPv6), traffic on all ports is allowed, regardless of any ports you specify. For tcp , udp , and icmp , you must specify a port range. For protocol 58 (ICMPv6), you can optionally specify a port range; if you don't, traffic for all types and codes is allowed.

asgiGroupId - The ID of the security group. Required for a nondefault VPC.

asgiToPort - The end of port range for the TCP and UDP protocols, or an ICMPICMPv6 code number. For the ICMPICMPv6 code number, use -1 to specify all codes.

asgiCidrIP - The CIDR IPv4 address range. You can't specify this parameter when specifying a source security group.

asgiSourceSecurityGroupOwnerId - [EC2-Classic] The AWS account number for the source security group, if the source security group is in a different account. You can't specify this parameter in combination with the following parameters: the CIDR IP address range, the IP protocol, the start of the port range, and the end of the port range. Creates rules that grant full ICMP, UDP, and TCP access. To create a rule with a specific IP protocol and port range, use a set of IP permissions instead.

asgiGroupName - [EC2-Classic, default VPC] The name of the security group.

asgiSourceSecurityGroupName - [EC2-Classic, default VPC] The name of the source security group. You can't specify this parameter in combination with the following parameters: the CIDR IP address range, the start of the port range, the IP protocol, and the end of the port range. Creates rules that grant full ICMP, UDP, and TCP access. To create a rule with a specific IP protocol and port range, use a set of IP permissions instead. For EC2-VPC, the source security group must be in the same VPC.

asgiDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data AuthorizeSecurityGroupIngress

Contains the parameters for AuthorizeSecurityGroupIngress.

See: authorizeSecurityGroupIngress smart constructor.

Request Lenses

asgiFromPort :: Lens' AuthorizeSecurityGroupIngress (Maybe Int)

The start of port range for the TCP and UDP protocols, or an ICMPICMPv6 type number. For the ICMPICMPv6 type number, use -1 to specify all types.

asgiIPPermissions :: Lens' AuthorizeSecurityGroupIngress [IPPermission]

A set of IP permissions. Can be used to specify multiple rules in a single command.

asgiIPProtocol :: Lens' AuthorizeSecurityGroupIngress (Maybe Text)

The IP protocol name (tcp , udp , icmp ) or number (see Protocol Numbers ). (VPC only) Use -1 to specify all protocols. If you specify -1 , or a protocol number other than tcp , udp , icmp , or 58 (ICMPv6), traffic on all ports is allowed, regardless of any ports you specify. For tcp , udp , and icmp , you must specify a port range. For protocol 58 (ICMPv6), you can optionally specify a port range; if you don't, traffic for all types and codes is allowed.

asgiGroupId :: Lens' AuthorizeSecurityGroupIngress (Maybe Text)

The ID of the security group. Required for a nondefault VPC.

asgiToPort :: Lens' AuthorizeSecurityGroupIngress (Maybe Int)

The end of port range for the TCP and UDP protocols, or an ICMPICMPv6 code number. For the ICMPICMPv6 code number, use -1 to specify all codes.

asgiCidrIP :: Lens' AuthorizeSecurityGroupIngress (Maybe Text)

The CIDR IPv4 address range. You can't specify this parameter when specifying a source security group.

asgiSourceSecurityGroupOwnerId :: Lens' AuthorizeSecurityGroupIngress (Maybe Text)

EC2-Classic

The AWS account number for the source security group, if the source security group is in a different account. You can't specify this parameter in combination with the following parameters: the CIDR IP address range, the IP protocol, the start of the port range, and the end of the port range. Creates rules that grant full ICMP, UDP, and TCP access. To create a rule with a specific IP protocol and port range, use a set of IP permissions instead.

asgiGroupName :: Lens' AuthorizeSecurityGroupIngress (Maybe Text)

EC2-Classic, default VPC

The name of the security group.

asgiSourceSecurityGroupName :: Lens' AuthorizeSecurityGroupIngress (Maybe Text)

EC2-Classic, default VPC

The name of the source security group. You can't specify this parameter in combination with the following parameters: the CIDR IP address range, the start of the port range, the IP protocol, and the end of the port range. Creates rules that grant full ICMP, UDP, and TCP access. To create a rule with a specific IP protocol and port range, use a set of IP permissions instead. For EC2-VPC, the source security group must be in the same VPC.

asgiDryRun :: Lens' AuthorizeSecurityGroupIngress (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

authorizeSecurityGroupIngressResponse :: AuthorizeSecurityGroupIngressResponse

Creates a value of AuthorizeSecurityGroupIngressResponse with the minimum fields required to make a request.

data AuthorizeSecurityGroupIngressResponse

See: authorizeSecurityGroupIngressResponse smart constructor.