Network.AWS.EC2.BundleInstance

Description

Bundles an Amazon instance store-backed Windows instance.

During bundling, only the root device volume (C:) is bundled. Data on other instance store volumes is not preserved.

For more information, see Creating an Instance Store-Backed Windows AMI .

Creating a Request

bundleInstance

Arguments

:: Text
biInstanceId

-> Storage
biStorage

-> BundleInstance

Creates a value of BundleInstance with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

biDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

biInstanceId - The ID of the instance to bundle. Type: String Default: None Required: Yes

biStorage - The bucket in which to store the AMI. You can specify a bucket that you already own or a new bucket that Amazon EC2 creates on your behalf. If you specify a bucket that belongs to someone else, Amazon EC2 returns an error.

data BundleInstance

Contains the parameters for BundleInstance.

See: bundleInstance smart constructor.

Request Lenses

biDryRun :: Lens' BundleInstance (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

biInstanceId :: Lens' BundleInstance Text

The ID of the instance to bundle. Type: String Default: None Required: Yes

biStorage :: Lens' BundleInstance Storage

The bucket in which to store the AMI. You can specify a bucket that you already own or a new bucket that Amazon EC2 creates on your behalf. If you specify a bucket that belongs to someone else, Amazon EC2 returns an error.

Destructuring the Response

bundleInstanceResponse

Arguments

:: Int
birsResponseStatus

-> BundleInstanceResponse

Creates a value of BundleInstanceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

birsBundleTask - Information about the bundle task.

birsResponseStatus - -- | The response status code.

data BundleInstanceResponse

Contains the output of BundleInstance.

See: bundleInstanceResponse smart constructor.

Response Lenses

birsBundleTask :: Lens' BundleInstanceResponse (Maybe BundleTask)

Information about the bundle task.

birsResponseStatus :: Lens' BundleInstanceResponse Int

- | The response status code.