Network.AWS.EC2.CancelConversionTask

Description

Cancels an active conversion task. The task can be the import of an instance or volume. The action removes all artifacts of the conversion, including a partially uploaded volume or instance. If the conversion is complete or is in the process of transferring the final disk image, the command fails and returns an exception.

For more information, see Importing a Virtual Machine Using the Amazon EC2 CLI .

Creating a Request

cancelConversionTask

Arguments

:: Text
cctConversionTaskId

-> CancelConversionTask

Creates a value of CancelConversionTask with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cctReasonMessage - The reason for canceling the conversion task.

cctDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cctConversionTaskId - The ID of the conversion task.

data CancelConversionTask

Contains the parameters for CancelConversionTask.

See: cancelConversionTask smart constructor.

Request Lenses

cctReasonMessage :: Lens' CancelConversionTask (Maybe Text)

The reason for canceling the conversion task.

cctDryRun :: Lens' CancelConversionTask (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cctConversionTaskId :: Lens' CancelConversionTask Text

The ID of the conversion task.

Destructuring the Response

cancelConversionTaskResponse :: CancelConversionTaskResponse

Creates a value of CancelConversionTaskResponse with the minimum fields required to make a request.

data CancelConversionTaskResponse

See: cancelConversionTaskResponse smart constructor.