Network.AWS.EC2.CancelExportTask

Description

Cancels an active export task. The request removes all artifacts of the export, including any partially-created Amazon S3 objects. If the export task is complete or is in the process of transferring the final disk image, the command fails and returns an error.

Creating a Request

cancelExportTask

Arguments

:: Text
cetExportTaskId

-> CancelExportTask

Creates a value of CancelExportTask with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cetExportTaskId - The ID of the export task. This is the ID returned by CreateInstanceExportTask .

data CancelExportTask

Contains the parameters for CancelExportTask.

See: cancelExportTask smart constructor.

Request Lenses

cetExportTaskId :: Lens' CancelExportTask Text

The ID of the export task. This is the ID returned by CreateInstanceExportTask .

Destructuring the Response

cancelExportTaskResponse :: CancelExportTaskResponse

Creates a value of CancelExportTaskResponse with the minimum fields required to make a request.

data CancelExportTaskResponse

See: cancelExportTaskResponse smart constructor.