Network.AWS.EC2.CancelReservedInstancesListing

Description

Cancels the specified Reserved Instance listing in the Reserved Instance Marketplace.

For more information, see Reserved Instance Marketplace in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

cancelReservedInstancesListing

Arguments

:: Text
crilReservedInstancesListingId

-> CancelReservedInstancesListing

Creates a value of CancelReservedInstancesListing with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crilReservedInstancesListingId - The ID of the Reserved Instance listing.

data CancelReservedInstancesListing

Contains the parameters for CancelReservedInstancesListing.

See: cancelReservedInstancesListing smart constructor.

Request Lenses

crilReservedInstancesListingId :: Lens' CancelReservedInstancesListing Text

The ID of the Reserved Instance listing.

Destructuring the Response

cancelReservedInstancesListingResponse

Arguments

:: Int
crilrsResponseStatus

-> CancelReservedInstancesListingResponse

Creates a value of CancelReservedInstancesListingResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crilrsReservedInstancesListings - The Reserved Instance listing.

crilrsResponseStatus - -- | The response status code.

data CancelReservedInstancesListingResponse

Contains the output of CancelReservedInstancesListing.

See: cancelReservedInstancesListingResponse smart constructor.

Response Lenses

crilrsReservedInstancesListings :: Lens' CancelReservedInstancesListingResponse [ReservedInstancesListing]

The Reserved Instance listing.

crilrsResponseStatus :: Lens' CancelReservedInstancesListingResponse Int

- | The response status code.