Network.AWS.EC2.CancelReservedInstancesListing

Description

Cancels the specified Reserved Instance listing in the Reserved Instance Marketplace.

For more information, see Reserved Instance Marketplace in the AmazonElastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CancelReservedInstancesListing.html

Request

data CancelReservedInstancesListing

Request constructor

cancelReservedInstancesListing

Arguments

:: Text
crilReservedInstancesListingId

-> CancelReservedInstancesListing

CancelReservedInstancesListing constructor.

The fields accessible through corresponding lenses are:

crilReservedInstancesListingId :: Text

Request lenses

crilReservedInstancesListingId :: Lens' CancelReservedInstancesListing Text

The ID of the Reserved Instance listing.

Response

data CancelReservedInstancesListingResponse

Response constructor

cancelReservedInstancesListingResponse :: CancelReservedInstancesListingResponse

CancelReservedInstancesListingResponse constructor.

The fields accessible through corresponding lenses are:

crilrReservedInstancesListings :: [ReservedInstancesListing]

Response lenses

crilrReservedInstancesListings :: Lens' CancelReservedInstancesListingResponse [ReservedInstancesListing]

The Reserved Instance listing.