Network.AWS.EC2.CopyImage

Description

Initiates the copy of an AMI from the specified source region to the current region. You specify the destination region by using its endpoint when making the request.

For more information about the prerequisites and limits when copying an AMI, see Copying an AMI in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

copyImage

Arguments

:: Text
ciName

-> Text
ciSourceImageId

-> Text
ciSourceRegion

-> CopyImage

Creates a value of CopyImage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ciClientToken - Unique, case-sensitive identifier you provide to ensure idempotency of the request. For more information, see How to Ensure Idempotency in the Amazon Elastic Compute Cloud User Guide .

ciEncrypted - Specifies whether the destination snapshots of the copied image should be encrypted. The default CMK for EBS is used unless a non-default AWS Key Management Service (AWS KMS) CMK is specified with KmsKeyId . For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide .

ciKMSKeyId - The full ARN of the AWS Key Management Service (AWS KMS) CMK to use when encrypting the snapshots of an image during a copy operation. This parameter is only required if you want to use a non-default CMK; if this parameter is not specified, the default CMK for EBS is used. The ARN contains the arn:aws:kms namespace, followed by the region of the CMK, the AWS account ID of the CMK owner, the key namespace, and then the CMK ID. For example, arn:aws:kms:us-east-1 :012345678910 :key/abcd1234-a123-456a-a12b-a123b4cd56ef . The specified CMK must exist in the region that the snapshot is being copied to. If a KmsKeyId is specified, the Encrypted flag must also be set.

ciDescription - A description for the new AMI in the destination region.

ciDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

ciName - The name of the new AMI in the destination region.

ciSourceImageId - The ID of the AMI to copy.

ciSourceRegion - The name of the region that contains the AMI to copy.

data CopyImage

Contains the parameters for CopyImage.

See: copyImage smart constructor.

Request Lenses

ciClientToken :: Lens' CopyImage (Maybe Text)

Unique, case-sensitive identifier you provide to ensure idempotency of the request. For more information, see How to Ensure Idempotency in the Amazon Elastic Compute Cloud User Guide .

ciEncrypted :: Lens' CopyImage (Maybe Bool)

Specifies whether the destination snapshots of the copied image should be encrypted. The default CMK for EBS is used unless a non-default AWS Key Management Service (AWS KMS) CMK is specified with KmsKeyId . For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide .

ciKMSKeyId :: Lens' CopyImage (Maybe Text)

The full ARN of the AWS Key Management Service (AWS KMS) CMK to use when encrypting the snapshots of an image during a copy operation. This parameter is only required if you want to use a non-default CMK; if this parameter is not specified, the default CMK for EBS is used. The ARN contains the arn:aws:kms namespace, followed by the region of the CMK, the AWS account ID of the CMK owner, the key namespace, and then the CMK ID. For example, arn:aws:kms:us-east-1 :012345678910 :key/abcd1234-a123-456a-a12b-a123b4cd56ef . The specified CMK must exist in the region that the snapshot is being copied to. If a KmsKeyId is specified, the Encrypted flag must also be set.

ciDescription :: Lens' CopyImage (Maybe Text)

A description for the new AMI in the destination region.

ciDryRun :: Lens' CopyImage (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

ciName :: Lens' CopyImage Text

The name of the new AMI in the destination region.

ciSourceImageId :: Lens' CopyImage Text

The ID of the AMI to copy.

ciSourceRegion :: Lens' CopyImage Text

The name of the region that contains the AMI to copy.

Destructuring the Response

copyImageResponse

Arguments

:: Int
coprsResponseStatus

-> CopyImageResponse

Creates a value of CopyImageResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

coprsImageId - The ID of the new AMI.

coprsResponseStatus - -- | The response status code.

data CopyImageResponse

Contains the output of CopyImage.

See: copyImageResponse smart constructor.

Response Lenses

coprsImageId :: Lens' CopyImageResponse (Maybe Text)

The ID of the new AMI.

coprsResponseStatus :: Lens' CopyImageResponse Int

- | The response status code.