Network.AWS.EC2.CopySnapshot

Description

Copies a point-in-time snapshot of an Amazon EBS volume and stores it in Amazon S3. You can copy the snapshot within the same region or from one region to another. You can use the snapshot to create Amazon EBS volumes or Amazon Machine Images (AMIs). The snapshot is copied to the regional endpoint that you send the HTTP request to. Copies of encrypted Amazon EBS snapshots remain encrypted. Copies of unencrypted snapshots remain unencrypted. For more information, see Copying an Amazon EBS Snapshot in the Amazon Elastic Compute Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CopySnapshot.html

Request

data CopySnapshot

Request constructor

copySnapshot

Arguments

:: Text
csSourceRegion

-> Text
csSourceSnapshotId

-> CopySnapshot

CopySnapshot constructor.

The fields accessible through corresponding lenses are:

csDescription :: Maybe Text

csDestinationRegion :: Maybe Text

csDryRun :: Maybe Bool

csPresignedUrl :: Maybe Text

csSourceRegion :: Text

csSourceSnapshotId :: Text

Request lenses

csDescription :: Lens' CopySnapshot (Maybe Text)

A description for the new Amazon EBS snapshot.

csDestinationRegion :: Lens' CopySnapshot (Maybe Text)

The destination region of the snapshot copy operation. This parameter is required in the PresignedUrl.

csDryRun :: Lens' CopySnapshot (Maybe Bool)

csPresignedUrl :: Lens' CopySnapshot (Maybe Text)

The pre-signed URL that facilitates copying an encrypted snapshot. This parameter is only required when copying an encrypted snapshot with the Amazon EC2 Query API; it is available as an optional parameter in all other cases. The PresignedUrl should use the snapshot source endpoint, the CopySnapshot action, and include the SourceRegion, SourceSnapshotId, and DestinationRegion parameters. The PresignedUrl must be signed using AWS Signature Version 4. Because Amazon EBS snapshots are stored in Amazon S3, the signing algorithm for this parameter uses the same logic that is described in Authenticating Requests by Using Query Parameters (AWS Signature Version 4) in the Amazon Simple Storage Service API Reference. An invalid or improperly signed PresignedUrl will cause the copy operation to fail asynchronously, and the snapshot will move to an error state.

csSourceRegion :: Lens' CopySnapshot Text

The ID of the region that contains the snapshot to be copied.

csSourceSnapshotId :: Lens' CopySnapshot Text

The ID of the Amazon EBS snapshot to copy.

Response

data CopySnapshotResponse

Response constructor

copySnapshotResponse :: CopySnapshotResponse

CopySnapshotResponse constructor.

The fields accessible through corresponding lenses are:

csrSnapshotId :: Maybe Text

Response lenses

csrSnapshotId :: Lens' CopySnapshotResponse (Maybe Text)

The ID of the new snapshot.