Network.AWS.EC2.CreateCustomerGateway

Description

Provides information to AWS about your VPN customer gateway device. The customer gateway is the appliance at your end of the VPN connection. (The device on the AWS side of the VPN connection is the virtual private gateway.) You must provide the Internet-routable IP address of the customer gateway's external interface. The IP address must be static and may be behind a device performing network address translation (NAT).

For devices that use Border Gateway Protocol (BGP), you can also provide the device's BGP Autonomous System Number (ASN). You can use an existing ASN assigned to your network. If you don't have an ASN already, you can use a private ASN (in the 64512 - 65534 range).

For more information about VPN customer gateways, see Adding a Hardware Virtual Private Gateway to Your VPC in the Amazon Virtual Private Cloud User Guide .

Important: You cannot create more than one customer gateway with the same VPN type, IP address, and BGP ASN parameter values. If you run an identical request more than one time, the first request creates the customer gateway, and subsequent requests return information about the existing customer gateway. The subsequent requests do not create new customer gateway resources.

Creating a Request

createCustomerGateway

Arguments

:: Int
ccgBGPASN

-> Text
ccgPublicIP

-> GatewayType
ccgType

-> CreateCustomerGateway

Creates a value of CreateCustomerGateway with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccgDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

ccgBGPASN - For devices that support BGP, the customer gateway's BGP ASN. Default: 65000

ccgPublicIP - The Internet-routable IP address for the customer gateway's outside interface. The address must be static.

ccgType - The type of VPN connection that this customer gateway supports (ipsec.1 ).

data CreateCustomerGateway

Contains the parameters for CreateCustomerGateway.

See: createCustomerGateway smart constructor.

Request Lenses

ccgDryRun :: Lens' CreateCustomerGateway (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

ccgBGPASN :: Lens' CreateCustomerGateway Int

For devices that support BGP, the customer gateway's BGP ASN. Default: 65000

ccgPublicIP :: Lens' CreateCustomerGateway Text

The Internet-routable IP address for the customer gateway's outside interface. The address must be static.

ccgType :: Lens' CreateCustomerGateway GatewayType

The type of VPN connection that this customer gateway supports (ipsec.1 ).

Destructuring the Response

createCustomerGatewayResponse

Arguments

:: Int
ccgrsResponseStatus

-> CreateCustomerGatewayResponse

Creates a value of CreateCustomerGatewayResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccgrsCustomerGateway - Information about the customer gateway.

ccgrsResponseStatus - -- | The response status code.

data CreateCustomerGatewayResponse

Contains the output of CreateCustomerGateway.

See: createCustomerGatewayResponse smart constructor.

Response Lenses

ccgrsCustomerGateway :: Lens' CreateCustomerGatewayResponse (Maybe CustomerGateway)

Information about the customer gateway.

ccgrsResponseStatus :: Lens' CreateCustomerGatewayResponse Int

- | The response status code.