Network.AWS.EC2.CreateCustomerGateway

Description

Provides information to AWS about your VPN customer gateway device. The customer gateway is the appliance at your end of the VPN connection. (The device on the AWS side of the VPN connection is the virtual private gateway.) You must provide the Internet-routable IP address of the customer gateway's external interface. The IP address must be static and can't be behind a device performing network address translation (NAT). For devices that use Border Gateway Protocol (BGP), you can also provide the device's BGP Autonomous System Number (ASN). You can use an existing ASN assigned to your network. If you don't have an ASN already, you can use a private ASN (in the 64512 - 65534 range). Amazon EC2 supports all 2-byte ASN numbers in the range of 1 - 65534, with the exception of 7224, which is reserved in the us-east-1 region, and 9059, which is reserved in the eu-west-1 region. For more information about VPN customer gateways, see Adding a Hardware Virtual Private Gateway to Your VPC in the Amazon Virtual Private Cloud User Guide. Example This example passes information to AWS about the customer gateway with the IP address 12.1.2.3 and BGP ASN 65534. https:/ec2.amazonaws.com?Action=CreateCustomerGateway &amp;Type=ipsec.1 &amp;IpAddress=12.1.2.3 &amp;BgpAsn=65534 &amp;AUTHPARAMS &lt;CreateCustomerGatewayResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;customerGateway&gt; &lt;customerGatewayId&gt;cgw-b4dc3961&lt;/customerGatewayId&gt; &lt;state&gt;pending&lt;state&gt; &lt;type&gt;ipsec.1&lt;type&gt; &lt;ipAddress&gt;12.1.2.3&lt;/ipAddress&gt; &lt;bgpAsn&gt;65534&lt;bgpAsn&gt; &lt;tagSet&gt; &lt;/customerGateway&gt; &lt;/CreateCustomerGatewayResponse&gt;.

Request

data CreateCustomerGateway

Request constructor

createCustomerGateway

Arguments

:: GatewayType
ccgType

-> Text
ccgPublicIp

-> Integer
ccgBgpAsn

-> CreateCustomerGateway

Smart constructor for the minimum required parameters to construct a valid CreateCustomerGateway request.

The fields accessible through corresponding lenses are:

Type :: GatewayType

PublicIp :: Text

BgpAsn :: Integer

Request lenses

ccgType :: Lens' CreateCustomerGateway GatewayType

The type of VPN connection that this customer gateway supports.

ccgPublicIp :: Lens' CreateCustomerGateway Text

The Internet-routable IP address for the customer gateway's outside interface. The address must be static.

ccgBgpAsn :: Lens' CreateCustomerGateway Integer

For devices that support BGP, the customer gateway's BGP ASN. Default: 65000.

Response

data CreateCustomerGatewayResponse

Response constructor

createCustomerGatewayResponse :: CreateCustomerGatewayResponse

Smart constructor for the minimum required parameters to construct a valid CreateCustomerGatewayResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

CustomerGateway :: Maybe CustomerGateway

Response lenses

ccgrCustomerGateway :: Lens' CreateCustomerGatewayResponse (Maybe CustomerGateway)

Information about the customer gateway.