Network.AWS.EC2.CreateDHCPOptions

Description

Creates a set of DHCP options for your VPC. After creating the set, you must associate it with the VPC, causing all existing and new instances that you launch in the VPC to use this set of DHCP options. The following are the individual DHCP options you can specify. For more information about the options, see RFC 2132 .

domain-name-servers - The IP addresses of up to four domain name servers, or AmazonProvidedDNS. The default DHCP option set specifies AmazonProvidedDNS. If specifying more than one domain name server, specify the IP addresses in a single parameter, separated by commas. If you want your instance to receive a custom DNS hostname as specified in domain-name , you must set domain-name-servers to a custom DNS server.

domain-name - If you're using AmazonProvidedDNS in us-east-1 , specify ec2.internal . If you're using AmazonProvidedDNS in another region, specify region.compute.internal (for example, ap-northeast-1.compute.internal ). Otherwise, specify a domain name (for example, MyCompany.com ). This value is used to complete unqualified DNS hostnames. Important : Some Linux operating systems accept multiple domain names separated by spaces. However, Windows and other Linux operating systems treat the value as a single domain, which results in unexpected behavior. If your DHCP options set is associated with a VPC that has instances with multiple operating systems, specify only one domain name.

ntp-servers - The IP addresses of up to four Network Time Protocol (NTP) servers.

netbios-name-servers - The IP addresses of up to four NetBIOS name servers.

netbios-node-type - The NetBIOS node type (1, 2, 4, or 8). We recommend that you specify 2 (broadcast and multicast are not currently supported). For more information about these node types, see RFC 2132 .

Your VPC automatically starts out with a set of DHCP options that includes only a DNS server that we provide (AmazonProvidedDNS). If you create a set of options, and if your VPC has an Internet gateway, make sure to set the domain-name-servers option either to AmazonProvidedDNS or to a domain name server of your choice. For more information about DHCP options, see DHCP Options Sets in the Amazon Virtual Private Cloud User Guide .

Creating a Request

createDHCPOptions :: CreateDHCPOptions

Creates a value of CreateDHCPOptions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdoDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cdoDHCPConfigurations - A DHCP configuration option.

data CreateDHCPOptions

Contains the parameters for CreateDhcpOptions.

See: createDHCPOptions smart constructor.

Request Lenses

cdoDryRun :: Lens' CreateDHCPOptions (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cdoDHCPConfigurations :: Lens' CreateDHCPOptions [NewDHCPConfiguration]

A DHCP configuration option.

Destructuring the Response

createDHCPOptionsResponse

Arguments

:: Int
cdorsResponseStatus

-> CreateDHCPOptionsResponse

Creates a value of CreateDHCPOptionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdorsDHCPOptions - A set of DHCP options.

cdorsResponseStatus - -- | The response status code.

data CreateDHCPOptionsResponse

Contains the output of CreateDhcpOptions.

See: createDHCPOptionsResponse smart constructor.

Response Lenses

cdorsDHCPOptions :: Lens' CreateDHCPOptionsResponse (Maybe DHCPOptions)

A set of DHCP options.

cdorsResponseStatus :: Lens' CreateDHCPOptionsResponse Int

- | The response status code.