Network.AWS.EC2.CreateEgressOnlyInternetGateway

Description

IPv6 only

Creates an egress-only Internet gateway for your VPC. An egress-only Internet gateway is used to enable outbound communication over IPv6 from instances in your VPC to the Internet, and prevents hosts outside of your VPC from initiating an IPv6 connection with your instance.

Creating a Request

createEgressOnlyInternetGateway

Arguments

:: Text
ceoigVPCId

-> CreateEgressOnlyInternetGateway

Creates a value of CreateEgressOnlyInternetGateway with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ceoigClientToken - Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency .

ceoigDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

ceoigVPCId - The ID of the VPC for which to create the egress-only Internet gateway.

data CreateEgressOnlyInternetGateway

See: createEgressOnlyInternetGateway smart constructor.

Request Lenses

ceoigClientToken :: Lens' CreateEgressOnlyInternetGateway (Maybe Text)

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency .

ceoigDryRun :: Lens' CreateEgressOnlyInternetGateway (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

ceoigVPCId :: Lens' CreateEgressOnlyInternetGateway Text

The ID of the VPC for which to create the egress-only Internet gateway.

Destructuring the Response

createEgressOnlyInternetGatewayResponse

Arguments

:: Int
ceoigrsResponseStatus

-> CreateEgressOnlyInternetGatewayResponse

Creates a value of CreateEgressOnlyInternetGatewayResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ceoigrsClientToken - Unique, case-sensitive identifier you provide to ensure the idempotency of the request.

ceoigrsEgressOnlyInternetGateway - Information about the egress-only Internet gateway.

ceoigrsResponseStatus - -- | The response status code.

data CreateEgressOnlyInternetGatewayResponse

See: createEgressOnlyInternetGatewayResponse smart constructor.

Response Lenses

ceoigrsClientToken :: Lens' CreateEgressOnlyInternetGatewayResponse (Maybe Text)

Unique, case-sensitive identifier you provide to ensure the idempotency of the request.

ceoigrsEgressOnlyInternetGateway :: Lens' CreateEgressOnlyInternetGatewayResponse (Maybe EgressOnlyInternetGateway)

Information about the egress-only Internet gateway.

ceoigrsResponseStatus :: Lens' CreateEgressOnlyInternetGatewayResponse Int

- | The response status code.