Network.AWS.EC2.CreateFlowLogs

Description

Creates one or more flow logs to capture IP traffic for a specific network interface, subnet, or VPC. Flow logs are delivered to a specified log group in Amazon CloudWatch Logs. If you specify a VPC or subnet in the request, a log stream is created in CloudWatch Logs for each network interface in the subnet or VPC. Log streams can include information about accepted and rejected traffic to a network interface. You can view the data in your log streams using Amazon CloudWatch Logs.

In your request, you must also specify an IAM role that has permission to publish logs to CloudWatch Logs.

Creating a Request

createFlowLogs

Arguments

:: Text
cflDeliverLogsPermissionARN

-> Text
cflLogGroupName

-> FlowLogsResourceType
cflResourceType

-> TrafficType
cflTrafficType

-> CreateFlowLogs

Creates a value of CreateFlowLogs with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cflClientToken - Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency .

cflDeliverLogsPermissionARN - The ARN for the IAM role that's used to post flow logs to a CloudWatch Logs log group.

cflLogGroupName - The name of the CloudWatch log group.

cflResourceIds - One or more subnet, network interface, or VPC IDs. Constraints: Maximum of 1000 resources

cflResourceType - The type of resource on which to create the flow log.

cflTrafficType - The type of traffic to log.

data CreateFlowLogs

Contains the parameters for CreateFlowLogs.

See: createFlowLogs smart constructor.

Request Lenses

cflClientToken :: Lens' CreateFlowLogs (Maybe Text)

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency .

cflDeliverLogsPermissionARN :: Lens' CreateFlowLogs Text

The ARN for the IAM role that's used to post flow logs to a CloudWatch Logs log group.

cflLogGroupName :: Lens' CreateFlowLogs Text

The name of the CloudWatch log group.

cflResourceIds :: Lens' CreateFlowLogs [Text]

One or more subnet, network interface, or VPC IDs. Constraints: Maximum of 1000 resources

cflResourceType :: Lens' CreateFlowLogs FlowLogsResourceType

The type of resource on which to create the flow log.

cflTrafficType :: Lens' CreateFlowLogs TrafficType

The type of traffic to log.

Destructuring the Response

createFlowLogsResponse

Arguments

:: Int
cflrsResponseStatus

-> CreateFlowLogsResponse

Creates a value of CreateFlowLogsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cflrsUnsuccessful - Information about the flow logs that could not be created successfully.

cflrsClientToken - Unique, case-sensitive identifier you provide to ensure the idempotency of the request.

cflrsFlowLogIds - The IDs of the flow logs.

cflrsResponseStatus - -- | The response status code.

data CreateFlowLogsResponse

Contains the output of CreateFlowLogs.

See: createFlowLogsResponse smart constructor.

Response Lenses

cflrsUnsuccessful :: Lens' CreateFlowLogsResponse [UnsuccessfulItem]

Information about the flow logs that could not be created successfully.

cflrsClientToken :: Lens' CreateFlowLogsResponse (Maybe Text)

Unique, case-sensitive identifier you provide to ensure the idempotency of the request.

cflrsFlowLogIds :: Lens' CreateFlowLogsResponse [Text]

The IDs of the flow logs.

cflrsResponseStatus :: Lens' CreateFlowLogsResponse Int

- | The response status code.