Network.AWS.EC2.CreateFpgaImage

Description

Creates an Amazon FPGA Image (AFI) from the specified design checkpoint (DCP).

The create operation is asynchronous. To verify that the AFI is ready for use, check the output logs.

An AFI contains the FPGA bitstream that is ready to download to an FPGA. You can securely deploy an AFI on one or more FPGA-accelerated instances. For more information, see the AWS FPGA Hardware Development Kit .

Creating a Request

createFpgaImage

Arguments

:: StorageLocation
cfiInputStorageLocation

-> CreateFpgaImage

Creates a value of CreateFpgaImage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cfiClientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency .

cfiLogsStorageLocation - The location in Amazon S3 for the output logs.

cfiName - A name for the AFI.

cfiDescription - A description for the AFI.

cfiDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cfiInputStorageLocation - The location of the encrypted design checkpoint in Amazon S3. The input must be a tarball.

data CreateFpgaImage

See: createFpgaImage smart constructor.

Request Lenses

cfiClientToken :: Lens' CreateFpgaImage (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see Ensuring Idempotency .

cfiLogsStorageLocation :: Lens' CreateFpgaImage (Maybe StorageLocation)

The location in Amazon S3 for the output logs.

cfiName :: Lens' CreateFpgaImage (Maybe Text)

A name for the AFI.

cfiDescription :: Lens' CreateFpgaImage (Maybe Text)

A description for the AFI.

cfiDryRun :: Lens' CreateFpgaImage (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cfiInputStorageLocation :: Lens' CreateFpgaImage StorageLocation

The location of the encrypted design checkpoint in Amazon S3. The input must be a tarball.

Destructuring the Response

createFpgaImageResponse

Arguments

:: Int
cfirsResponseStatus

-> CreateFpgaImageResponse

Creates a value of CreateFpgaImageResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cfirsFpgaImageId - The FPGA image identifier (AFI ID).

cfirsFpgaImageGlobalId - The global FPGA image identifier (AGFI ID).

cfirsResponseStatus - -- | The response status code.

data CreateFpgaImageResponse

See: createFpgaImageResponse smart constructor.

Response Lenses

cfirsFpgaImageId :: Lens' CreateFpgaImageResponse (Maybe Text)

The FPGA image identifier (AFI ID).

cfirsFpgaImageGlobalId :: Lens' CreateFpgaImageResponse (Maybe Text)

The global FPGA image identifier (AGFI ID).

cfirsResponseStatus :: Lens' CreateFpgaImageResponse Int

- | The response status code.