Network.AWS.EC2.CreateImage

Description

Creates an Amazon EBS-backed AMI from an Amazon EBS-backed instance that is either running or stopped.

If you customized your instance with instance store volumes or EBS volumes in addition to the root device volume, the new AMI contains block device mapping information for those volumes. When you launch an instance from this new AMI, the instance automatically launches with those additional volumes.

For more information, see Creating Amazon EBS-Backed Linux AMIs in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

createImage

Arguments

:: Text
ciiInstanceId

-> Text
ciiName

-> CreateImage

Creates a value of CreateImage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ciiNoReboot - By default, Amazon EC2 attempts to shut down and reboot the instance before creating the image. If the 'No Reboot' option is set, Amazon EC2 doesn't shut down the instance before creating the image. When this option is used, file system integrity on the created image can't be guaranteed.

ciiDescription - A description for the new image.

ciiBlockDeviceMappings - Information about one or more block device mappings.

ciiDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

ciiInstanceId - The ID of the instance.

ciiName - A name for the new image. Constraints: 3-128 alphanumeric characters, parentheses (()), square brackets ([]), spaces ( ), periods (.), slashes (/), dashes (-), single quotes ('), at-signs (@), or underscores(_)

data CreateImage

Contains the parameters for CreateImage.

See: createImage smart constructor.

Request Lenses

ciiNoReboot :: Lens' CreateImage (Maybe Bool)

By default, Amazon EC2 attempts to shut down and reboot the instance before creating the image. If the 'No Reboot' option is set, Amazon EC2 doesn't shut down the instance before creating the image. When this option is used, file system integrity on the created image can't be guaranteed.

ciiDescription :: Lens' CreateImage (Maybe Text)

A description for the new image.

ciiBlockDeviceMappings :: Lens' CreateImage [BlockDeviceMapping]

Information about one or more block device mappings.

ciiDryRun :: Lens' CreateImage (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

ciiInstanceId :: Lens' CreateImage Text

The ID of the instance.

ciiName :: Lens' CreateImage Text

A name for the new image. Constraints: 3-128 alphanumeric characters, parentheses (()), square brackets ([]), spaces ( ), periods (.), slashes (/), dashes (-), single quotes ('), at-signs (@), or underscores(_)

Destructuring the Response

createImageResponse

Arguments

:: Int
cirsResponseStatus

-> CreateImageResponse

Creates a value of CreateImageResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cirsImageId - The ID of the new AMI.

cirsResponseStatus - -- | The response status code.

data CreateImageResponse

Contains the output of CreateImage.

See: createImageResponse smart constructor.

Response Lenses

cirsImageId :: Lens' CreateImageResponse (Maybe Text)

The ID of the new AMI.

cirsResponseStatus :: Lens' CreateImageResponse Int

- | The response status code.