Network.AWS.EC2.CreateInstanceExportTask

Description

Exports a running or stopped instance to an S3 bucket.

For information about the supported operating systems, image formats, and known limitations for the types of instances you can export, see Exporting an Instance as a VM Using VM Import/Export in the VM Import/Export User Guide .

Creating a Request

createInstanceExportTask

Arguments

:: Text
cietInstanceId

-> CreateInstanceExportTask

Creates a value of CreateInstanceExportTask with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cietTargetEnvironment - The target virtualization environment.

cietExportToS3Task - The format and location for an instance export task.

cietDescription - A description for the conversion task or the resource being exported. The maximum length is 255 bytes.

cietInstanceId - The ID of the instance.

data CreateInstanceExportTask

Contains the parameters for CreateInstanceExportTask.

See: createInstanceExportTask smart constructor.

Request Lenses

cietTargetEnvironment :: Lens' CreateInstanceExportTask (Maybe ExportEnvironment)

The target virtualization environment.

cietExportToS3Task :: Lens' CreateInstanceExportTask (Maybe ExportToS3TaskSpecification)

The format and location for an instance export task.

cietDescription :: Lens' CreateInstanceExportTask (Maybe Text)

A description for the conversion task or the resource being exported. The maximum length is 255 bytes.

cietInstanceId :: Lens' CreateInstanceExportTask Text

The ID of the instance.

Destructuring the Response

createInstanceExportTaskResponse

Arguments

:: Int
cietrsResponseStatus

-> CreateInstanceExportTaskResponse

Creates a value of CreateInstanceExportTaskResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cietrsExportTask - Information about the instance export task.

cietrsResponseStatus - -- | The response status code.

data CreateInstanceExportTaskResponse

Contains the output for CreateInstanceExportTask.

See: createInstanceExportTaskResponse smart constructor.

Response Lenses

cietrsExportTask :: Lens' CreateInstanceExportTaskResponse (Maybe ExportTask)

Information about the instance export task.

cietrsResponseStatus :: Lens' CreateInstanceExportTaskResponse Int

- | The response status code.