Network.AWS.EC2.CreateInternetGateway

Description

Creates an Internet gateway for use with a VPC. After creating the Internet gateway, you attach it to a VPC using AttachInternetGateway .

For more information about your VPC and Internet gateway, see the Amazon Virtual Private Cloud User Guide .

Creating a Request

createInternetGateway :: CreateInternetGateway

Creates a value of CreateInternetGateway with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cigDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data CreateInternetGateway

Contains the parameters for CreateInternetGateway.

See: createInternetGateway smart constructor.

Request Lenses

cigDryRun :: Lens' CreateInternetGateway (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

createInternetGatewayResponse

Arguments

:: Int
cigrsResponseStatus

-> CreateInternetGatewayResponse

Creates a value of CreateInternetGatewayResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cigrsInternetGateway - Information about the Internet gateway.

cigrsResponseStatus - -- | The response status code.

data CreateInternetGatewayResponse

Contains the output of CreateInternetGateway.

See: createInternetGatewayResponse smart constructor.

Response Lenses

cigrsInternetGateway :: Lens' CreateInternetGatewayResponse (Maybe InternetGateway)

Information about the Internet gateway.

cigrsResponseStatus :: Lens' CreateInternetGatewayResponse Int

- | The response status code.