Network.AWS.EC2.CreateKeyPair

Description

Creates a 2048-bit RSA key pair with the specified name. Amazon EC2 stores the public key and displays the private key for you to save to a file. The private key is returned as an unencrypted PEM encoded PKCS#8 private key. If a key with the specified name already exists, Amazon EC2 returns an error.

You can have up to five thousand key pairs per region.

The key pair returned to you is available only in the region in which you create it. To create a key pair that is available in all regions, use ImportKeyPair .

For more information about key pairs, see Key Pairs in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

createKeyPair

Arguments

:: Text
ckpKeyName

-> CreateKeyPair

Creates a value of CreateKeyPair with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ckpDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

ckpKeyName - A unique name for the key pair. Constraints: Up to 255 ASCII characters

data CreateKeyPair

Contains the parameters for CreateKeyPair.

See: createKeyPair smart constructor.

Request Lenses

ckpDryRun :: Lens' CreateKeyPair (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

ckpKeyName :: Lens' CreateKeyPair Text

A unique name for the key pair. Constraints: Up to 255 ASCII characters

Destructuring the Response

createKeyPairResponse

Arguments

:: Int
ckprsResponseStatus

-> Text
ckprsKeyName

-> Text
ckprsKeyFingerprint

-> Text
ckprsKeyMaterial

-> CreateKeyPairResponse

Creates a value of CreateKeyPairResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ckprsResponseStatus - -- | The response status code.

ckprsKeyName - The name of the key pair.

ckprsKeyFingerprint - The SHA-1 digest of the DER encoded private key.

ckprsKeyMaterial - An unencrypted PEM encoded RSA private key.

data CreateKeyPairResponse

Describes a key pair.

See: createKeyPairResponse smart constructor.

Response Lenses

ckprsResponseStatus :: Lens' CreateKeyPairResponse Int

- | The response status code.

ckprsKeyName :: Lens' CreateKeyPairResponse Text

The name of the key pair.

ckprsKeyFingerprint :: Lens' CreateKeyPairResponse Text

The SHA-1 digest of the DER encoded private key.

ckprsKeyMaterial :: Lens' CreateKeyPairResponse Text

An unencrypted PEM encoded RSA private key.