Network.AWS.EC2.CreateNetworkACL

Description

Creates a network ACL in a VPC. Network ACLs provide an optional layer of security (in addition to security groups) for the instances in your VPC.

For more information about network ACLs, see Network ACLs in the Amazon Virtual Private Cloud User Guide .

Creating a Request

createNetworkACL

Arguments

:: Text
cnaVPCId

-> CreateNetworkACL

Creates a value of CreateNetworkACL with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cnaDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cnaVPCId - The ID of the VPC.

data CreateNetworkACL

Contains the parameters for CreateNetworkAcl.

See: createNetworkACL smart constructor.

Request Lenses

cnaDryRun :: Lens' CreateNetworkACL (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cnaVPCId :: Lens' CreateNetworkACL Text

The ID of the VPC.

Destructuring the Response

createNetworkACLResponse

Arguments

:: Int
cnarsResponseStatus

-> CreateNetworkACLResponse

Creates a value of CreateNetworkACLResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cnarsNetworkACL - Information about the network ACL.

cnarsResponseStatus - -- | The response status code.

data CreateNetworkACLResponse

Contains the output of CreateNetworkAcl.

See: createNetworkACLResponse smart constructor.

Response Lenses

cnarsNetworkACL :: Lens' CreateNetworkACLResponse (Maybe NetworkACL)

Information about the network ACL.

cnarsResponseStatus :: Lens' CreateNetworkACLResponse Int

- | The response status code.