Network.AWS.EC2.CreateNetworkInterface

Description

Creates a network interface in the specified subnet.

For more information about network interfaces, see Elastic Network Interfaces in the Amazon Virtual Private Cloud User Guide .

Creating a Request

createNetworkInterface

Arguments

:: Text
cniSubnetId

-> CreateNetworkInterface

Creates a value of CreateNetworkInterface with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cniGroups - The IDs of one or more security groups.

cniPrivateIPAddresses - One or more private IPv4 addresses.

cniIPv6AddressCount - The number of IPv6 addresses to assign to a network interface. Amazon EC2 automatically selects the IPv6 addresses from the subnet range. You can't use this option if specifying specific IPv6 addresses. If your subnet has the AssignIpv6AddressOnCreation attribute set to true , you can specify 0 to override this setting.

cniPrivateIPAddress - The primary private IPv4 address of the network interface. If you don't specify an IPv4 address, Amazon EC2 selects one for you from the subnet's IPv4 CIDR range. If you specify an IP address, you cannot indicate any IP addresses specified in privateIpAddresses as primary (only one IP address can be designated as primary).

cniSecondaryPrivateIPAddressCount - The number of secondary private IPv4 addresses to assign to a network interface. When you specify a number of secondary IPv4 addresses, Amazon EC2 selects these IP addresses within the subnet's IPv4 CIDR range. You can't specify this option and specify more than one private IP address using privateIpAddresses . The number of IP addresses you can assign to a network interface varies by instance type. For more information, see IP Addresses Per ENI Per Instance Type in the Amazon Virtual Private Cloud User Guide .

cniDescription - A description for the network interface.

cniDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cniIPv6Addresses - One or more specific IPv6 addresses from the IPv6 CIDR block range of your subnet. You can't use this option if you're specifying a number of IPv6 addresses.

cniSubnetId - The ID of the subnet to associate with the network interface.

data CreateNetworkInterface

Contains the parameters for CreateNetworkInterface.

See: createNetworkInterface smart constructor.

Request Lenses

cniGroups :: Lens' CreateNetworkInterface [Text]

The IDs of one or more security groups.

cniPrivateIPAddresses :: Lens' CreateNetworkInterface [PrivateIPAddressSpecification]

One or more private IPv4 addresses.

cniIPv6AddressCount :: Lens' CreateNetworkInterface (Maybe Int)

The number of IPv6 addresses to assign to a network interface. Amazon EC2 automatically selects the IPv6 addresses from the subnet range. You can't use this option if specifying specific IPv6 addresses. If your subnet has the AssignIpv6AddressOnCreation attribute set to true , you can specify 0 to override this setting.

cniPrivateIPAddress :: Lens' CreateNetworkInterface (Maybe Text)

The primary private IPv4 address of the network interface. If you don't specify an IPv4 address, Amazon EC2 selects one for you from the subnet's IPv4 CIDR range. If you specify an IP address, you cannot indicate any IP addresses specified in privateIpAddresses as primary (only one IP address can be designated as primary).

cniSecondaryPrivateIPAddressCount :: Lens' CreateNetworkInterface (Maybe Int)

The number of secondary private IPv4 addresses to assign to a network interface. When you specify a number of secondary IPv4 addresses, Amazon EC2 selects these IP addresses within the subnet's IPv4 CIDR range. You can't specify this option and specify more than one private IP address using privateIpAddresses . The number of IP addresses you can assign to a network interface varies by instance type. For more information, see IP Addresses Per ENI Per Instance Type in the Amazon Virtual Private Cloud User Guide .

cniDescription :: Lens' CreateNetworkInterface (Maybe Text)

A description for the network interface.

cniDryRun :: Lens' CreateNetworkInterface (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cniIPv6Addresses :: Lens' CreateNetworkInterface [InstanceIPv6Address]

One or more specific IPv6 addresses from the IPv6 CIDR block range of your subnet. You can't use this option if you're specifying a number of IPv6 addresses.

cniSubnetId :: Lens' CreateNetworkInterface Text

The ID of the subnet to associate with the network interface.

Destructuring the Response

createNetworkInterfaceResponse

Arguments

:: Int
cnirsResponseStatus

-> CreateNetworkInterfaceResponse

Creates a value of CreateNetworkInterfaceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cnirsNetworkInterface - Information about the network interface.

cnirsResponseStatus - -- | The response status code.

data CreateNetworkInterfaceResponse

Contains the output of CreateNetworkInterface.

See: createNetworkInterfaceResponse smart constructor.

Response Lenses

cnirsNetworkInterface :: Lens' CreateNetworkInterfaceResponse (Maybe NetworkInterface)

Information about the network interface.

cnirsResponseStatus :: Lens' CreateNetworkInterfaceResponse Int

- | The response status code.