Network.AWS.EC2.CreateNetworkInterfacePermission

Description

Grants an AWS authorized partner account permission to attach the specified network interface to an instance in their account.

You can grant permission to a single AWS account only, and only one account at a time.

Creating a Request

createNetworkInterfacePermission

Arguments

:: Text
cnipNetworkInterfaceId

-> InterfacePermissionType
cnipPermission

-> CreateNetworkInterfacePermission

Creates a value of CreateNetworkInterfacePermission with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cnipAWSAccountId - The AWS account ID.

cnipAWSService - The AWS service. Currently not supported.

cnipDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cnipNetworkInterfaceId - The ID of the network interface.

cnipPermission - The type of permission to grant.

data CreateNetworkInterfacePermission

Contains the parameters for CreateNetworkInterfacePermission.

See: createNetworkInterfacePermission smart constructor.

Request Lenses

cnipAWSAccountId :: Lens' CreateNetworkInterfacePermission (Maybe Text)

The AWS account ID.

cnipAWSService :: Lens' CreateNetworkInterfacePermission (Maybe Text)

The AWS service. Currently not supported.

cnipDryRun :: Lens' CreateNetworkInterfacePermission (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cnipNetworkInterfaceId :: Lens' CreateNetworkInterfacePermission Text

The ID of the network interface.

cnipPermission :: Lens' CreateNetworkInterfacePermission InterfacePermissionType

The type of permission to grant.

Destructuring the Response

createNetworkInterfacePermissionResponse

Arguments

:: Int
cniprsResponseStatus

-> CreateNetworkInterfacePermissionResponse

Creates a value of CreateNetworkInterfacePermissionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cniprsInterfacePermission - Information about the permission for the network interface.

cniprsResponseStatus - -- | The response status code.

data CreateNetworkInterfacePermissionResponse

Contains the output of CreateNetworkInterfacePermission.

See: createNetworkInterfacePermissionResponse smart constructor.

Response Lenses

cniprsInterfacePermission :: Lens' CreateNetworkInterfacePermissionResponse (Maybe NetworkInterfacePermission)

Information about the permission for the network interface.

cniprsResponseStatus :: Lens' CreateNetworkInterfacePermissionResponse Int

- | The response status code.