Network.AWS.EC2.CreatePlacementGroup

Description

Creates a placement group that you launch cluster instances into. You must give the group a name that's unique within the scope of your account.

For more information about placement groups and cluster instances, see Cluster Instances in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

createPlacementGroup

Arguments

:: Text
cpgGroupName

-> PlacementStrategy
cpgStrategy

-> CreatePlacementGroup

Creates a value of CreatePlacementGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cpgDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cpgGroupName - A name for the placement group. Constraints: Up to 255 ASCII characters

cpgStrategy - The placement strategy.

data CreatePlacementGroup

Contains the parameters for CreatePlacementGroup.

See: createPlacementGroup smart constructor.

Request Lenses

cpgDryRun :: Lens' CreatePlacementGroup (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cpgGroupName :: Lens' CreatePlacementGroup Text

A name for the placement group. Constraints: Up to 255 ASCII characters

cpgStrategy :: Lens' CreatePlacementGroup PlacementStrategy

The placement strategy.

Destructuring the Response

createPlacementGroupResponse :: CreatePlacementGroupResponse

Creates a value of CreatePlacementGroupResponse with the minimum fields required to make a request.

data CreatePlacementGroupResponse

See: createPlacementGroupResponse smart constructor.