Network.AWS.EC2.CreateReservedInstancesListing

Description

Creates a listing for Amazon EC2 Standard Reserved Instances to be sold in the Reserved Instance Marketplace. You can submit one Standard Reserved Instance listing at a time. To get a list of your Standard Reserved Instances, you can use the DescribeReservedInstances operation.

The Reserved Instance Marketplace matches sellers who want to resell Standard Reserved Instance capacity that they no longer need with buyers who want to purchase additional capacity. Reserved Instances bought and sold through the Reserved Instance Marketplace work like any other Reserved Instances.

To sell your Standard Reserved Instances, you must first register as a seller in the Reserved Instance Marketplace. After completing the registration process, you can create a Reserved Instance Marketplace listing of some or all of your Standard Reserved Instances, and specify the upfront price to receive for them. Your Standard Reserved Instance listings then become available for purchase. To view the details of your Standard Reserved Instance listing, you can use the DescribeReservedInstancesListings operation.

For more information, see Reserved Instance Marketplace in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

createReservedInstancesListing

Arguments

:: Text
crilClientToken

-> Int
crilInstanceCount

-> Text
crilReservedInstancesId

-> CreateReservedInstancesListing

Creates a value of CreateReservedInstancesListing with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crilClientToken - Unique, case-sensitive identifier you provide to ensure idempotency of your listings. This helps avoid duplicate listings. For more information, see Ensuring Idempotency .

crilInstanceCount - The number of instances that are a part of a Reserved Instance account to be listed in the Reserved Instance Marketplace. This number should be less than or equal to the instance count associated with the Reserved Instance ID specified in this call.

crilPriceSchedules - A list specifying the price of the Standard Reserved Instance for each month remaining in the Reserved Instance term.

crilReservedInstancesId - The ID of the active Standard Reserved Instance.

data CreateReservedInstancesListing

Contains the parameters for CreateReservedInstancesListing.

See: createReservedInstancesListing smart constructor.

Request Lenses

crilClientToken :: Lens' CreateReservedInstancesListing Text

Unique, case-sensitive identifier you provide to ensure idempotency of your listings. This helps avoid duplicate listings. For more information, see Ensuring Idempotency .

crilInstanceCount :: Lens' CreateReservedInstancesListing Int

The number of instances that are a part of a Reserved Instance account to be listed in the Reserved Instance Marketplace. This number should be less than or equal to the instance count associated with the Reserved Instance ID specified in this call.

crilPriceSchedules :: Lens' CreateReservedInstancesListing [PriceScheduleSpecification]

A list specifying the price of the Standard Reserved Instance for each month remaining in the Reserved Instance term.

crilReservedInstancesId :: Lens' CreateReservedInstancesListing Text

The ID of the active Standard Reserved Instance.

Destructuring the Response

createReservedInstancesListingResponse

Arguments

:: Int
crersResponseStatus

-> CreateReservedInstancesListingResponse

Creates a value of CreateReservedInstancesListingResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crersReservedInstancesListings - Information about the Standard Reserved Instance listing.

crersResponseStatus - -- | The response status code.

data CreateReservedInstancesListingResponse

Contains the output of CreateReservedInstancesListing.

See: createReservedInstancesListingResponse smart constructor.

Response Lenses

crersReservedInstancesListings :: Lens' CreateReservedInstancesListingResponse [ReservedInstancesListing]

Information about the Standard Reserved Instance listing.

crersResponseStatus :: Lens' CreateReservedInstancesListingResponse Int

- | The response status code.