Network.AWS.EC2.CreateReservedInstancesListing

Description

Creates a listing for Amazon EC2 Reserved Instances to be sold in the Reserved Instance Marketplace. You can submit one Reserved Instance listing at a time. To get a list of your Reserved Instances, you can use the DescribeReservedInstances operation.

The Reserved Instance Marketplace matches sellers who want to resell Reserved Instance capacity that they no longer need with buyers who want to purchase additional capacity. Reserved Instances bought and sold through the Reserved Instance Marketplace work like any other Reserved Instances.

To sell your Reserved Instances, you must first register as a Seller in the Reserved Instance Marketplace. After completing the registration process, you can create a Reserved Instance Marketplace listing of some or all of your Reserved Instances, and specify the upfront price to receive for them. Your Reserved Instance listings then become available for purchase. To view the details of your Reserved Instance listing, you can use the DescribeReservedInstancesListings operation.

For more information, see Reserved Instance Marketplace in the AmazonElastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateReservedInstancesListing.html

Request

data CreateReservedInstancesListing

Request constructor

createReservedInstancesListing

Arguments

:: Text
crilReservedInstancesId

-> Int
crilInstanceCount

-> Text
crilClientToken

-> CreateReservedInstancesListing

CreateReservedInstancesListing constructor.

The fields accessible through corresponding lenses are:

crilClientToken :: Text

crilInstanceCount :: Int

crilPriceSchedules :: [PriceScheduleSpecification]

crilReservedInstancesId :: Text

Request lenses

crilClientToken :: Lens' CreateReservedInstancesListing Text

Unique, case-sensitive identifier you provide to ensure idempotency of your listings. This helps avoid duplicate listings. For more information, see Ensuring Idempotency.

crilInstanceCount :: Lens' CreateReservedInstancesListing Int

The number of instances that are a part of a Reserved Instance account to be listed in the Reserved Instance Marketplace. This number should be less than or equal to the instance count associated with the Reserved Instance ID specified in this call.

crilPriceSchedules :: Lens' CreateReservedInstancesListing [PriceScheduleSpecification]

A list specifying the price of the Reserved Instance for each month remaining in the Reserved Instance term.

crilReservedInstancesId :: Lens' CreateReservedInstancesListing Text

The ID of the active Reserved Instance.

Response

data CreateReservedInstancesListingResponse

Response constructor

createReservedInstancesListingResponse :: CreateReservedInstancesListingResponse

CreateReservedInstancesListingResponse constructor.

The fields accessible through corresponding lenses are:

crilr1ReservedInstancesListings :: [ReservedInstancesListing]

Response lenses

crilr1ReservedInstancesListings :: Lens' CreateReservedInstancesListingResponse [ReservedInstancesListing]

Information about the Reserved Instances listing.