Network.AWS.EC2.CreateRoute

Description

Creates a route in a route table within a VPC.

You must specify one of the following targets: Internet gateway or virtual private gateway, NAT instance, NAT gateway, VPC peering connection, network interface, or egress-only Internet gateway.

When determining how to route traffic, we use the route with the most specific match. For example, traffic is destined for the IPv4 address 192.0.2.3 , and the route table includes the following two IPv4 routes:

192.0.2.0/24 (goes to some target A)

192.0.2.0/28 (goes to some target B)

Both routes apply to the traffic destined for 192.0.2.3 . However, the second route in the list covers a smaller number of IP addresses and is therefore more specific, so we use that route to determine where to target the traffic.

For more information about route tables, see Route Tables in the Amazon Virtual Private Cloud User Guide .

Creating a Request

createRoute

Arguments

:: Text
crRouteTableId

-> CreateRoute

Creates a value of CreateRoute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crVPCPeeringConnectionId - The ID of a VPC peering connection.

crInstanceId - The ID of a NAT instance in your VPC. The operation fails if you specify an instance ID unless exactly one network interface is attached.

crEgressOnlyInternetGatewayId - [IPv6 traffic only] The ID of an egress-only Internet gateway.

crDestinationIPv6CidrBlock - The IPv6 CIDR block used for the destination match. Routing decisions are based on the most specific match.

crNatGatewayId - [IPv4 traffic only] The ID of a NAT gateway.

crNetworkInterfaceId - The ID of a network interface.

crGatewayId - The ID of an Internet gateway or virtual private gateway attached to your VPC.

crDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

crDestinationCidrBlock - The IPv4 CIDR address block used for the destination match. Routing decisions are based on the most specific match.

crRouteTableId - The ID of the route table for the route.

data CreateRoute

Contains the parameters for CreateRoute.

See: createRoute smart constructor.

Request Lenses

crVPCPeeringConnectionId :: Lens' CreateRoute (Maybe Text)

The ID of a VPC peering connection.

crInstanceId :: Lens' CreateRoute (Maybe Text)

The ID of a NAT instance in your VPC. The operation fails if you specify an instance ID unless exactly one network interface is attached.

crEgressOnlyInternetGatewayId :: Lens' CreateRoute (Maybe Text)

IPv6 traffic only

The ID of an egress-only Internet gateway.

crDestinationIPv6CidrBlock :: Lens' CreateRoute (Maybe Text)

The IPv6 CIDR block used for the destination match. Routing decisions are based on the most specific match.

crNatGatewayId :: Lens' CreateRoute (Maybe Text)

IPv4 traffic only

The ID of a NAT gateway.

crNetworkInterfaceId :: Lens' CreateRoute (Maybe Text)

The ID of a network interface.

crGatewayId :: Lens' CreateRoute (Maybe Text)

The ID of an Internet gateway or virtual private gateway attached to your VPC.

crDryRun :: Lens' CreateRoute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

crDestinationCidrBlock :: Lens' CreateRoute (Maybe Text)

The IPv4 CIDR address block used for the destination match. Routing decisions are based on the most specific match.

crRouteTableId :: Lens' CreateRoute Text

The ID of the route table for the route.

Destructuring the Response

createRouteResponse

Arguments

:: Int
crrsResponseStatus

-> CreateRouteResponse

Creates a value of CreateRouteResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crrsReturn - Returns true if the request succeeds; otherwise, it returns an error.

crrsResponseStatus - -- | The response status code.

data CreateRouteResponse

Contains the output of CreateRoute.

See: createRouteResponse smart constructor.

Response Lenses

crrsReturn :: Lens' CreateRouteResponse (Maybe Bool)

Returns true if the request succeeds; otherwise, it returns an error.

crrsResponseStatus :: Lens' CreateRouteResponse Int

- | The response status code.