Network.AWS.EC2.CreateRoute

Description

Creates a route in a route table within a VPC.

You must specify one of the following targets: Internet gateway or virtual private gateway, NAT instance, VPC peering connection, or network interface.

When determining how to route traffic, we use the route with the most specific match. For example, let's say the traffic is destined for '192.0.2.3', and the route table includes the following two routes:

'192.0.2.0/24' (goes to some target A)

'192.0.2.0/28' (goes to some target B)

Both routes apply to the traffic destined for '192.0.2.3'. However, the second route in the list covers a smaller number of IP addresses and is therefore more specific, so we use that route to determine where to target the traffic.

For more information about route tables, see Route Tables in the AmazonVirtual Private Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateRoute.html

Request

data CreateRoute

Request constructor

createRoute

Arguments

:: Text
crRouteTableId

-> Text
crDestinationCidrBlock

-> CreateRoute

CreateRoute constructor.

The fields accessible through corresponding lenses are:

crDestinationCidrBlock :: Text

crDryRun :: Maybe Bool

crGatewayId :: Maybe Text

crInstanceId :: Maybe Text

crNetworkInterfaceId :: Maybe Text

crRouteTableId :: Text

crVpcPeeringConnectionId :: Maybe Text

Request lenses

crDestinationCidrBlock :: Lens' CreateRoute Text

The CIDR address block used for the destination match. Routing decisions are based on the most specific match.

crDryRun :: Lens' CreateRoute (Maybe Bool)

crGatewayId :: Lens' CreateRoute (Maybe Text)

The ID of an Internet gateway or virtual private gateway attached to your VPC.

crInstanceId :: Lens' CreateRoute (Maybe Text)

The ID of a NAT instance in your VPC. The operation fails if you specify an instance ID unless exactly one network interface is attached.

crNetworkInterfaceId :: Lens' CreateRoute (Maybe Text)

The ID of a network interface.

crRouteTableId :: Lens' CreateRoute Text

The ID of the route table for the route.

crVpcPeeringConnectionId :: Lens' CreateRoute (Maybe Text)

The ID of a VPC peering connection.

Response

data CreateRouteResponse

Response constructor

createRouteResponse :: CreateRouteResponse

CreateRouteResponse constructor.