Network.AWS.EC2.CreateRouteTable

Description

Creates a route table for the specified VPC. After you create a route table, you can add routes and associate the table with a subnet.

For more information about route tables, see Route Tables in the Amazon Virtual Private Cloud User Guide .

Creating a Request

createRouteTable

Arguments

:: Text
crtVPCId

-> CreateRouteTable

Creates a value of CreateRouteTable with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crtDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

crtVPCId - The ID of the VPC.

data CreateRouteTable

Contains the parameters for CreateRouteTable.

See: createRouteTable smart constructor.

Request Lenses

crtDryRun :: Lens' CreateRouteTable (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

crtVPCId :: Lens' CreateRouteTable Text

The ID of the VPC.

Destructuring the Response

createRouteTableResponse

Arguments

:: Int
crtrsResponseStatus

-> CreateRouteTableResponse

Creates a value of CreateRouteTableResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crtrsRouteTable - Information about the route table.

crtrsResponseStatus - -- | The response status code.

data CreateRouteTableResponse

Contains the output of CreateRouteTable.

See: createRouteTableResponse smart constructor.

Response Lenses

crtrsRouteTable :: Lens' CreateRouteTableResponse (Maybe RouteTable)

Information about the route table.

crtrsResponseStatus :: Lens' CreateRouteTableResponse Int

- | The response status code.