Network.AWS.EC2.CreateSnapshot

Description

Creates a snapshot of an Amazon EBS volume and stores it in Amazon S3. You can use snapshots for backups, to make copies of Amazon EBS volumes, and to save data before shutting down an instance.

When a snapshot is created, any AWS Marketplace product codes that are associated with the source volume are propagated to the snapshot.

You can take a snapshot of an attached volume that is in use. However, snapshots only capture data that has been written to your Amazon EBS volume at the time the snapshot command is issued; this may exclude any data that has been cached by any applications or the operating system. If you can pause any file systems on the volume long enough to take a snapshot, your snapshot should be complete. However, if you cannot pause all file writes to the volume, you should unmount the volume from within the instance, issue the snapshot command, and then remount the volume to ensure a consistent and complete snapshot. You may remount and use your volume while the snapshot status is pending.

To create a snapshot for Amazon EBS volumes that serve as root devices, you should stop the instance before taking the snapshot.

Snapshots that are taken from encrypted volumes are automatically encrypted. Volumes that are created from encrypted snapshots are also automatically encrypted. Your encrypted volumes and any associated snapshots always remain protected.

For more information, see Amazon Elastic Block Store and Amazon EBSEncryption in the Amazon Elastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateSnapshot.html

Request

data CreateSnapshot

Request constructor

createSnapshot

Arguments

:: Text
cs2VolumeId

-> CreateSnapshot

CreateSnapshot constructor.

The fields accessible through corresponding lenses are:

cs2Description :: Maybe Text

cs2DryRun :: Maybe Bool

cs2VolumeId :: Text

Request lenses

cs2Description :: Lens' CreateSnapshot (Maybe Text)

A description for the snapshot.

cs2DryRun :: Lens' CreateSnapshot (Maybe Bool)

cs2VolumeId :: Lens' CreateSnapshot Text

The ID of the Amazon EBS volume.

Response

data CreateSnapshotResponse

Response constructor

createSnapshotResponse

Arguments

:: Text
csr1SnapshotId

-> Text
csr1VolumeId

-> SnapshotState
csr1State

-> UTCTime
csr1StartTime

-> Text
csr1Progress

-> Text
csr1OwnerId

-> Text
csr1Description

-> Int
csr1VolumeSize

-> Bool
csr1Encrypted

-> CreateSnapshotResponse

CreateSnapshotResponse constructor.

The fields accessible through corresponding lenses are:

csr1Description :: Text

csr1Encrypted :: Bool

csr1KmsKeyId :: Maybe Text

csr1OwnerAlias :: Maybe Text

csr1OwnerId :: Text

csr1Progress :: Text

csr1SnapshotId :: Text

csr1StartTime :: UTCTime

csr1State :: SnapshotState

csr1Tags :: [Tag]

csr1VolumeId :: Text

csr1VolumeSize :: Int

Response lenses

csr1Description :: Lens' CreateSnapshotResponse Text

The description for the snapshot.

csr1Encrypted :: Lens' CreateSnapshotResponse Bool

Indicates whether the snapshot is encrypted.

csr1KmsKeyId :: Lens' CreateSnapshotResponse (Maybe Text)

The full ARN of the AWS Key Management Service (KMS) master key that was used to protect the volume encryption key for the parent volume.

csr1OwnerAlias :: Lens' CreateSnapshotResponse (Maybe Text)

The AWS account alias (for example, amazon, self) or AWS account ID that owns the snapshot.

csr1OwnerId :: Lens' CreateSnapshotResponse Text

The AWS account ID of the Amazon EBS snapshot owner.

csr1Progress :: Lens' CreateSnapshotResponse Text

The progress of the snapshot, as a percentage.

csr1SnapshotId :: Lens' CreateSnapshotResponse Text

The ID of the snapshot.

csr1StartTime :: Lens' CreateSnapshotResponse UTCTime

The time stamp when the snapshot was initiated.

csr1State :: Lens' CreateSnapshotResponse SnapshotState

The snapshot state.

csr1Tags :: Lens' CreateSnapshotResponse [Tag]

Any tags assigned to the snapshot.

csr1VolumeId :: Lens' CreateSnapshotResponse Text

The ID of the volume.

csr1VolumeSize :: Lens' CreateSnapshotResponse Int

The size of the volume, in GiB.