Network.AWS.EC2.CreateSpotDatafeedSubscription

Description

Creates a data feed for Spot Instances, enabling you to view Spot Instance usage logs. You can create one data feed per AWS account. For more information, see Spot Instance Data Feed in the Amazon Elastic Compute CloudUser Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateSpotDatafeedSubscription.html

Request

data CreateSpotDatafeedSubscription

Request constructor

createSpotDatafeedSubscription

Arguments

:: Text
csdsBucket

-> CreateSpotDatafeedSubscription

CreateSpotDatafeedSubscription constructor.

The fields accessible through corresponding lenses are:

csdsBucket :: Text

csdsDryRun :: Maybe Bool

csdsPrefix :: Maybe Text

Request lenses

csdsBucket :: Lens' CreateSpotDatafeedSubscription Text

The Amazon S3 bucket in which to store the Spot Instance data feed.

csdsDryRun :: Lens' CreateSpotDatafeedSubscription (Maybe Bool)

csdsPrefix :: Lens' CreateSpotDatafeedSubscription (Maybe Text)

A prefix for the data feed file names.

Response

data CreateSpotDatafeedSubscriptionResponse

Response constructor

createSpotDatafeedSubscriptionResponse :: CreateSpotDatafeedSubscriptionResponse

CreateSpotDatafeedSubscriptionResponse constructor.

The fields accessible through corresponding lenses are:

csdsrSpotDatafeedSubscription :: Maybe SpotDatafeedSubscription

Response lenses

csdsrSpotDatafeedSubscription :: Lens' CreateSpotDatafeedSubscriptionResponse (Maybe SpotDatafeedSubscription)

The Spot Instance data feed subscription.