Network.AWS.EC2.CreateSubnet

Description

Creates a subnet in an existing VPC.

When you create each subnet, you provide the VPC ID and the CIDR block you want for the subnet. After you create a subnet, you can't change its CIDR block. The subnet's IPv4 CIDR block can be the same as the VPC's IPv4 CIDR block (assuming you want only a single subnet in the VPC), or a subset of the VPC's IPv4 CIDR block. If you create more than one subnet in a VPC, the subnets' CIDR blocks must not overlap. The smallest IPv4 subnet (and VPC) you can create uses a 28 netmask (16 IPv4 addresses), and the largest uses a 16 netmask (65,536 IPv4 addresses).

If you've associated an IPv6 CIDR block with your VPC, you can create a subnet with an IPv6 CIDR block that uses a /64 prefix length.

Important: AWS reserves both the first four and the last IPv4 address in each subnet's CIDR block. They're not available for use.

If you add more than one subnet to a VPC, they're set up in a star topology with a logical router in the middle.

If you launch an instance in a VPC using an Amazon EBS-backed AMI, the IP address doesn't change if you stop and restart the instance (unlike a similar instance launched outside a VPC, which gets a new IP address when restarted). It's therefore possible to have a subnet with no running instances (they're all stopped), but no remaining IP addresses available.

For more information about subnets, see Your VPC and Subnets in the Amazon Virtual Private Cloud User Guide .

Creating a Request

createSubnet

Arguments

:: Text
cssCidrBlock

-> Text
cssVPCId

-> CreateSubnet

Creates a value of CreateSubnet with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cssIPv6CidrBlock - The IPv6 network range for the subnet, in CIDR notation. The subnet size must use a /64 prefix length.

cssAvailabilityZone - The Availability Zone for the subnet. Default: AWS selects one for you. If you create more than one subnet in your VPC, we may not necessarily select a different zone for each subnet.

cssDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cssCidrBlock - The IPv4 network range for the subnet, in CIDR notation. For example, 10.0.0.0/24 .

cssVPCId - The ID of the VPC.

data CreateSubnet

Contains the parameters for CreateSubnet.

See: createSubnet smart constructor.

Request Lenses

cssIPv6CidrBlock :: Lens' CreateSubnet (Maybe Text)

The IPv6 network range for the subnet, in CIDR notation. The subnet size must use a /64 prefix length.

cssAvailabilityZone :: Lens' CreateSubnet (Maybe Text)

The Availability Zone for the subnet. Default: AWS selects one for you. If you create more than one subnet in your VPC, we may not necessarily select a different zone for each subnet.

cssDryRun :: Lens' CreateSubnet (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cssCidrBlock :: Lens' CreateSubnet Text

The IPv4 network range for the subnet, in CIDR notation. For example, 10.0.0.0/24 .

cssVPCId :: Lens' CreateSubnet Text

The ID of the VPC.

Destructuring the Response

createSubnetResponse

Arguments

:: Int
crsResponseStatus

-> CreateSubnetResponse

Creates a value of CreateSubnetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crsSubnet - Information about the subnet.

crsResponseStatus - -- | The response status code.

data CreateSubnetResponse

Contains the output of CreateSubnet.

See: createSubnetResponse smart constructor.

Response Lenses

crsSubnet :: Lens' CreateSubnetResponse (Maybe Subnet)

Information about the subnet.

crsResponseStatus :: Lens' CreateSubnetResponse Int

- | The response status code.