Network.AWS.EC2.CreateTags

Description

Adds or overwrites one or more tags for the specified Amazon EC2 resource or resources. Each resource can have a maximum of 50 tags. Each tag consists of a key and optional value. Tag keys must be unique per resource.

For more information about tags, see Tagging Your Resources in the Amazon Elastic Compute Cloud User Guide . For more information about creating IAM policies that control users' access to resources based on tags, see Supported Resource-Level Permissions for Amazon EC2 API Actions in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

createTags :: CreateTags

Creates a value of CreateTags with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cResources - The IDs of one or more resources to tag. For example, ami-1a2b3c4d.

cTags - One or more tags. The value parameter is required, but if you don't want the tag to have a value, specify the parameter with no value, and we set the value to an empty string.

data CreateTags

Contains the parameters for CreateTags.

See: createTags smart constructor.

Request Lenses

cDryRun :: Lens' CreateTags (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cResources :: Lens' CreateTags [Text]

The IDs of one or more resources to tag. For example, ami-1a2b3c4d.

cTags :: Lens' CreateTags [Tag]

One or more tags. The value parameter is required, but if you don't want the tag to have a value, specify the parameter with no value, and we set the value to an empty string.

Destructuring the Response

createTagsResponse :: CreateTagsResponse

Creates a value of CreateTagsResponse with the minimum fields required to make a request.

data CreateTagsResponse

See: createTagsResponse smart constructor.