Network.AWS.EC2.CreateVPC

Description

Creates a VPC with the specified IPv4 CIDR block. The smallest VPC you can create uses a 28 netmask (16 IPv4 addresses), and the largest uses a 16 netmask (65,536 IPv4 addresses). To help you decide how big to make your VPC, see Your VPC and Subnets in the Amazon Virtual Private Cloud User Guide .

You can optionally request an Amazon-provided IPv6 CIDR block for the VPC. The IPv6 CIDR block uses a /56 prefix length, and is allocated from Amazon's pool of IPv6 addresses. You cannot choose the IPv6 range for your VPC.

By default, each instance you launch in the VPC has the default DHCP options, which includes only a default DNS server that we provide (AmazonProvidedDNS). For more information about DHCP options, see DHCP Options Sets in the Amazon Virtual Private Cloud User Guide .

You can specify the instance tenancy value for the VPC when you create it. You can't change this value for the VPC after you create it. For more information, see Dedicated Instances in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

createVPC

Arguments

:: Text
cvCidrBlock

-> CreateVPC

Creates a value of CreateVPC with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cvAmazonProvidedIPv6CidrBlock - Requests an Amazon-provided IPv6 CIDR block with a /56 prefix length for the VPC. You cannot specify the range of IP addresses, or the size of the CIDR block.

cvInstanceTenancy - The tenancy options for instances launched into the VPC. For default , instances are launched with shared tenancy by default. You can launch instances with any tenancy into a shared tenancy VPC. For dedicated , instances are launched as dedicated tenancy instances by default. You can only launch instances with a tenancy of dedicated or host into a dedicated tenancy VPC. Important: The host value cannot be used with this parameter. Use the default or dedicated values only. Default: default

cvDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cvCidrBlock - The IPv4 network range for the VPC, in CIDR notation. For example, 10.0.0.0/16 .

data CreateVPC

Contains the parameters for CreateVpc.

See: createVPC smart constructor.

Request Lenses

cvAmazonProvidedIPv6CidrBlock :: Lens' CreateVPC (Maybe Bool)

Requests an Amazon-provided IPv6 CIDR block with a /56 prefix length for the VPC. You cannot specify the range of IP addresses, or the size of the CIDR block.

cvInstanceTenancy :: Lens' CreateVPC (Maybe Tenancy)

The tenancy options for instances launched into the VPC. For default , instances are launched with shared tenancy by default. You can launch instances with any tenancy into a shared tenancy VPC. For dedicated , instances are launched as dedicated tenancy instances by default. You can only launch instances with a tenancy of dedicated or host into a dedicated tenancy VPC. Important: The host value cannot be used with this parameter. Use the default or dedicated values only. Default: default

cvDryRun :: Lens' CreateVPC (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cvCidrBlock :: Lens' CreateVPC Text

The IPv4 network range for the VPC, in CIDR notation. For example, 10.0.0.0/16 .

Destructuring the Response

createVPCResponse

Arguments

:: Int
cvrsResponseStatus

-> CreateVPCResponse

Creates a value of CreateVPCResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cvrsVPC - Information about the VPC.

cvrsResponseStatus - -- | The response status code.

data CreateVPCResponse

Contains the output of CreateVpc.

See: createVPCResponse smart constructor.

Response Lenses

cvrsVPC :: Lens' CreateVPCResponse (Maybe VPC)

Information about the VPC.

cvrsResponseStatus :: Lens' CreateVPCResponse Int

- | The response status code.