Network.AWS.EC2.CreateVPCEndpoint

Description

Creates a VPC endpoint for a specified AWS service. An endpoint enables you to create a private connection between your VPC and another AWS service in your account. You can specify an endpoint policy to attach to the endpoint that will control access to the service from your VPC. You can also specify the VPC route tables that use the endpoint.

Use DescribeVpcEndpointServices to get a list of supported AWS services.

Creating a Request

createVPCEndpoint

Arguments

:: Text
cveServiceName

-> Text
cveVPCId

-> CreateVPCEndpoint

Creates a value of CreateVPCEndpoint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cvePolicyDocument - A policy to attach to the endpoint that controls access to the service. The policy must be in valid JSON format. If this parameter is not specified, we attach a default policy that allows full access to the service.

cveClientToken - Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency .

cveDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cveRouteTableIds - One or more route table IDs.

cveServiceName - The AWS service name, in the form com.amazonaws.region .service . To get a list of available services, use the DescribeVpcEndpointServices request.

cveVPCId - The ID of the VPC in which the endpoint will be used.

data CreateVPCEndpoint

Contains the parameters for CreateVpcEndpoint.

See: createVPCEndpoint smart constructor.

Request Lenses

cvePolicyDocument :: Lens' CreateVPCEndpoint (Maybe Text)

A policy to attach to the endpoint that controls access to the service. The policy must be in valid JSON format. If this parameter is not specified, we attach a default policy that allows full access to the service.

cveClientToken :: Lens' CreateVPCEndpoint (Maybe Text)

Unique, case-sensitive identifier you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency .

cveDryRun :: Lens' CreateVPCEndpoint (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

cveRouteTableIds :: Lens' CreateVPCEndpoint [Text]

One or more route table IDs.

cveServiceName :: Lens' CreateVPCEndpoint Text

The AWS service name, in the form com.amazonaws.region .service . To get a list of available services, use the DescribeVpcEndpointServices request.

cveVPCId :: Lens' CreateVPCEndpoint Text

The ID of the VPC in which the endpoint will be used.

Destructuring the Response

createVPCEndpointResponse

Arguments

:: Int
cversResponseStatus

-> CreateVPCEndpointResponse

Creates a value of CreateVPCEndpointResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cversClientToken - Unique, case-sensitive identifier you provide to ensure the idempotency of the request.

cversVPCEndpoint - Information about the endpoint.

cversResponseStatus - -- | The response status code.

data CreateVPCEndpointResponse

Contains the output of CreateVpcEndpoint.

See: createVPCEndpointResponse smart constructor.

Response Lenses

cversClientToken :: Lens' CreateVPCEndpointResponse (Maybe Text)

Unique, case-sensitive identifier you provide to ensure the idempotency of the request.

cversVPCEndpoint :: Lens' CreateVPCEndpointResponse (Maybe VPCEndpoint)

Information about the endpoint.

cversResponseStatus :: Lens' CreateVPCEndpointResponse Int

- | The response status code.