Network.AWS.EC2.CreateVolume

Description

Creates an EBS volume that can be attached to an instance in the same Availability Zone. The volume is created in the regional endpoint that you send the HTTP request to. For more information see Regions and Endpoints .

You can create a new empty volume or restore a volume from an EBS snapshot. Any AWS Marketplace product codes from the snapshot are propagated to the volume.

You can create encrypted volumes with the Encrypted parameter. Encrypted volumes may only be attached to instances that support Amazon EBS encryption. Volumes that are created from encrypted snapshots are also automatically encrypted. For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide .

You can tag your volumes during creation. For more information, see Tagging Your Amazon EC2 Resources .

For more information, see Creating an Amazon EBS Volume in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

createVolume

Arguments

:: Text
creAvailabilityZone

-> CreateVolume

Creates a value of CreateVolume with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

creSize - The size of the volume, in GiBs. Constraints: 1-16384 for gp2 , 4-16384 for io1 , 500-16384 for st1 , 500-16384 for sc1 , and 1-1024 for standard . If you specify a snapshot, the volume size must be equal to or larger than the snapshot size. Default: If you're creating the volume from a snapshot and don't specify a volume size, the default is the snapshot size.

creIOPS - Only valid for Provisioned IOPS SSD volumes. The number of IO operations per second (IOPS) to provision for the volume, with a maximum ratio of 50 IOPSGiB. Constraint: Range is 100 to 20000 for Provisioned IOPS SSD volumes

creEncrypted - Specifies whether the volume should be encrypted. Encrypted Amazon EBS volumes may only be attached to instances that support Amazon EBS encryption. Volumes that are created from encrypted snapshots are automatically encrypted. There is no way to create an encrypted volume from an unencrypted snapshot or vice versa. If your AMI uses encrypted volumes, you can only launch it on supported instance types. For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide .

creTagSpecifications - The tags to apply to the volume during creation.

creKMSKeyId - The full ARN of the AWS Key Management Service (AWS KMS) customer master key (CMK) to use when creating the encrypted volume. This parameter is only required if you want to use a non-default CMK; if this parameter is not specified, the default CMK for EBS is used. The ARN contains the arn:aws:kms namespace, followed by the region of the CMK, the AWS account ID of the CMK owner, the key namespace, and then the CMK ID. For example, arn:aws:kms:us-east-1 :012345678910 :key/abcd1234-a123-456a-a12b-a123b4cd56ef . If a KmsKeyId is specified, the Encrypted flag must also be set.

creVolumeType - The volume type. This can be gp2 for General Purpose SSD, io1 for Provisioned IOPS SSD, st1 for Throughput Optimized HDD, sc1 for Cold HDD, or standard for Magnetic volumes. Default: standard

creDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

creSnapshotId - The snapshot from which to create the volume.

creAvailabilityZone - The Availability Zone in which to create the volume. Use DescribeAvailabilityZones to list the Availability Zones that are currently available to you.

data CreateVolume

Contains the parameters for CreateVolume.

See: createVolume smart constructor.

Request Lenses

creSize :: Lens' CreateVolume (Maybe Int)

The size of the volume, in GiBs. Constraints: 1-16384 for gp2 , 4-16384 for io1 , 500-16384 for st1 , 500-16384 for sc1 , and 1-1024 for standard . If you specify a snapshot, the volume size must be equal to or larger than the snapshot size. Default: If you're creating the volume from a snapshot and don't specify a volume size, the default is the snapshot size.

creIOPS :: Lens' CreateVolume (Maybe Int)

Only valid for Provisioned IOPS SSD volumes. The number of IO operations per second (IOPS) to provision for the volume, with a maximum ratio of 50 IOPSGiB. Constraint: Range is 100 to 20000 for Provisioned IOPS SSD volumes

creEncrypted :: Lens' CreateVolume (Maybe Bool)

Specifies whether the volume should be encrypted. Encrypted Amazon EBS volumes may only be attached to instances that support Amazon EBS encryption. Volumes that are created from encrypted snapshots are automatically encrypted. There is no way to create an encrypted volume from an unencrypted snapshot or vice versa. If your AMI uses encrypted volumes, you can only launch it on supported instance types. For more information, see Amazon EBS Encryption in the Amazon Elastic Compute Cloud User Guide .

creTagSpecifications :: Lens' CreateVolume [TagSpecification]

The tags to apply to the volume during creation.

creKMSKeyId :: Lens' CreateVolume (Maybe Text)

The full ARN of the AWS Key Management Service (AWS KMS) customer master key (CMK) to use when creating the encrypted volume. This parameter is only required if you want to use a non-default CMK; if this parameter is not specified, the default CMK for EBS is used. The ARN contains the arn:aws:kms namespace, followed by the region of the CMK, the AWS account ID of the CMK owner, the key namespace, and then the CMK ID. For example, arn:aws:kms:us-east-1 :012345678910 :key/abcd1234-a123-456a-a12b-a123b4cd56ef . If a KmsKeyId is specified, the Encrypted flag must also be set.

creVolumeType :: Lens' CreateVolume (Maybe VolumeType)

The volume type. This can be gp2 for General Purpose SSD, io1 for Provisioned IOPS SSD, st1 for Throughput Optimized HDD, sc1 for Cold HDD, or standard for Magnetic volumes. Default: standard

creDryRun :: Lens' CreateVolume (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

creSnapshotId :: Lens' CreateVolume (Maybe Text)

The snapshot from which to create the volume.

creAvailabilityZone :: Lens' CreateVolume Text

The Availability Zone in which to create the volume. Use DescribeAvailabilityZones to list the Availability Zones that are currently available to you.

Destructuring the Response

volume

Arguments

:: Text
vAvailabilityZone

-> UTCTime
vCreateTime

-> Bool
vEncrypted

-> Int
vSize

-> Text
vSnapshotId

-> VolumeState
vState

-> Text
vVolumeId

-> VolumeType
vVolumeType

-> Volume

Creates a value of Volume with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vAttachments - Information about the volume attachments.

vIOPS - The number of IO operations per second (IOPS) that the volume supports. For Provisioned IOPS SSD volumes, this represents the number of IOPS that are provisioned for the volume. For General Purpose SSD volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates IO credits for bursting. For more information on General Purpose SSD baseline performance, IO credits, and bursting, see <http:docs.aws.amazon.comAWSEC2latestUserGuideEBSVolumeTypes.html Amazon EBS Volume Types> in the Amazon Elastic Compute Cloud User Guide/ . Constraint: Range is 100-20000 IOPS for io1 volumes and 100-10000 IOPS for gp2 volumes. Condition: This parameter is required for requests to create io1 volumes; it is not used in requests to create gp2 , st1 , sc1 , or standard volumes.

vKMSKeyId - The full ARN of the AWS Key Management Service (AWS KMS) customer master key (CMK) that was used to protect the volume encryption key for the volume.

vTags - Any tags assigned to the volume.

vAvailabilityZone - The Availability Zone for the volume.

vCreateTime - The time stamp when volume creation was initiated.

vEncrypted - Indicates whether the volume will be encrypted.

vSize - The size of the volume, in GiBs.

vSnapshotId - The snapshot from which the volume was created, if applicable.

vState - The volume state.

vVolumeId - The ID of the volume.

vVolumeType - The volume type. This can be gp2 for General Purpose SSD, io1 for Provisioned IOPS SSD, st1 for Throughput Optimized HDD, sc1 for Cold HDD, or standard for Magnetic volumes.

data Volume

Describes a volume.

See: volume smart constructor.

Response Lenses

vAttachments :: Lens' Volume [VolumeAttachment]

Information about the volume attachments.

vIOPS :: Lens' Volume (Maybe Int)

The number of IO operations per second (IOPS) that the volume supports. For Provisioned IOPS SSD volumes, this represents the number of IOPS that are provisioned for the volume. For General Purpose SSD volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates IO credits for bursting. For more information on General Purpose SSD baseline performance, IO credits, and bursting, see <http:docs.aws.amazon.comAWSEC2latestUserGuideEBSVolumeTypes.html Amazon EBS Volume Types> in the Amazon Elastic Compute Cloud User Guide/ . Constraint: Range is 100-20000 IOPS for io1 volumes and 100-10000 IOPS for gp2 volumes. Condition: This parameter is required for requests to create io1 volumes; it is not used in requests to create gp2 , st1 , sc1 , or standard volumes.

vKMSKeyId :: Lens' Volume (Maybe Text)

The full ARN of the AWS Key Management Service (AWS KMS) customer master key (CMK) that was used to protect the volume encryption key for the volume.

vTags :: Lens' Volume [Tag]

Any tags assigned to the volume.

vAvailabilityZone :: Lens' Volume Text

The Availability Zone for the volume.

vCreateTime :: Lens' Volume UTCTime

The time stamp when volume creation was initiated.

vEncrypted :: Lens' Volume Bool

Indicates whether the volume will be encrypted.

vSize :: Lens' Volume Int

The size of the volume, in GiBs.

vSnapshotId :: Lens' Volume Text

The snapshot from which the volume was created, if applicable.

vState :: Lens' Volume VolumeState

The volume state.

vVolumeId :: Lens' Volume Text

The ID of the volume.

vVolumeType :: Lens' Volume VolumeType

The volume type. This can be gp2 for General Purpose SSD, io1 for Provisioned IOPS SSD, st1 for Throughput Optimized HDD, sc1 for Cold HDD, or standard for Magnetic volumes.