Network.AWS.EC2.CreateVolume

Description

Creates an Amazon EBS volume that can be attached to an instance in the same Availability Zone. The volume is created in the regional endpoint that you send the HTTP request to. For more information see Regions and Endpoints.

You can create a new empty volume or restore a volume from an Amazon EBS snapshot. Any AWS Marketplace product codes from the snapshot are propagated to the volume.

You can create encrypted volumes with the Encrypted parameter. Encrypted volumes may only be attached to instances that support Amazon EBS encryption. Volumes that are created from encrypted snapshots are also automatically encrypted. For more information, see Amazon EBS Encryption in the AmazonElastic Compute Cloud User Guide for Linux.

For more information, see Creating or Restoring an Amazon EBS Volume in the Amazon Elastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateVolume.html

Request

data CreateVolume

Request constructor

createVolume

Arguments

:: Text
cv1AvailabilityZone

-> CreateVolume

CreateVolume constructor.

The fields accessible through corresponding lenses are:

cv1AvailabilityZone :: Text

cv1DryRun :: Maybe Bool

cv1Encrypted :: Maybe Bool

cv1Iops :: Maybe Int

cv1KmsKeyId :: Maybe Text

cv1Size :: Maybe Int

cv1SnapshotId :: Maybe Text

cv1VolumeType :: Maybe VolumeType

Request lenses

cv1AvailabilityZone :: Lens' CreateVolume Text

The Availability Zone in which to create the volume. Use DescribeAvailabilityZones to list the Availability Zones that are currently available to you.

cv1DryRun :: Lens' CreateVolume (Maybe Bool)

cv1Encrypted :: Lens' CreateVolume (Maybe Bool)

Specifies whether the volume should be encrypted. Encrypted Amazon EBS volumes may only be attached to instances that support Amazon EBS encryption. Volumes that are created from encrypted snapshots are automatically encrypted. There is no way to create an encrypted volume from an unencrypted snapshot or vice versa. If your AMI uses encrypted volumes, you can only launch it on supported instance types. For more information, see Amazon EBSEncryption in the Amazon Elastic Compute Cloud User Guide for Linux.

cv1Iops :: Lens' CreateVolume (Maybe Int)

Only valid for Provisioned IOPS (SSD) volumes. The number of I/O operations per second (IOPS) to provision for the volume, with a maximum ratio of 30 IOPS/GiB.

Constraint: Range is 100 to 20000 for Provisioned IOPS (SSD) volumes

cv1KmsKeyId :: Lens' CreateVolume (Maybe Text)

The full ARN of the AWS Key Management Service (KMS) master key to use when creating the encrypted volume. This parameter is only required if you want to use a non-default master key; if this parameter is not specified, the default master key is used. The ARN contains the 'arn:aws:kms' namespace, followed by the region of the master key, the AWS account ID of the master key owner, the key namespace, and then the master key ID. For example, arn:aws:kms:us-east-1:012345678910:key/abcd1234-a123-456a-a12b-a123b4cd56ef.

cv1Size :: Lens' CreateVolume (Maybe Int)

The size of the volume, in GiBs.

Constraints: '1-1024' for standard volumes, '1-16384' for gp2 volumes, and '4-16384' for io1 volumes. If you specify a snapshot, the volume size must be equal to or larger than the snapshot size.

Default: If you're creating the volume from a snapshot and don't specify a volume size, the default is the snapshot size.

cv1SnapshotId :: Lens' CreateVolume (Maybe Text)

The snapshot from which to create the volume.

cv1VolumeType :: Lens' CreateVolume (Maybe VolumeType)

The volume type. This can be gp2 for General Purpose (SSD) volumes, io1 for Provisioned IOPS (SSD) volumes, or standard for Magnetic volumes.

Default: standard

Response

data CreateVolumeResponse

Response constructor

createVolumeResponse

Arguments

:: Text
cvrVolumeId

-> Int
cvrSize

-> Text
cvrSnapshotId

-> Text
cvrAvailabilityZone

-> VolumeState
cvrState

-> UTCTime
cvrCreateTime

-> VolumeType
cvrVolumeType

-> Bool
cvrEncrypted

-> CreateVolumeResponse

CreateVolumeResponse constructor.

The fields accessible through corresponding lenses are:

cvrAttachments :: [VolumeAttachment]

cvrAvailabilityZone :: Text

cvrCreateTime :: UTCTime

cvrEncrypted :: Bool

cvrIops :: Maybe Int

cvrKmsKeyId :: Maybe Text

cvrSize :: Int

cvrSnapshotId :: Text

cvrState :: VolumeState

cvrTags :: [Tag]

cvrVolumeId :: Text

cvrVolumeType :: VolumeType

Response lenses

cvrAttachments :: Lens' CreateVolumeResponse [VolumeAttachment]

cvrAvailabilityZone :: Lens' CreateVolumeResponse Text

The Availability Zone for the volume.

cvrCreateTime :: Lens' CreateVolumeResponse UTCTime

The time stamp when volume creation was initiated.

cvrEncrypted :: Lens' CreateVolumeResponse Bool

Indicates whether the volume will be encrypted.

cvrIops :: Lens' CreateVolumeResponse (Maybe Int)

The number of I/O operations per second (IOPS) that the volume supports. For Provisioned IOPS (SSD) volumes, this represents the number of IOPS that are provisioned for the volume. For General Purpose (SSD) volumes, this represents the baseline performance of the volume and the rate at which the volume accumulates I/O credits for bursting. For more information on General Purpose (SSD) baseline performance, IO credits, and bursting, see <http:docs.aws.amazon.comAWSEC2latestUserGuideEBSVolumeTypes.html Amazon EBSVolume Types> in the Amazon Elastic Compute Cloud User Guide for Linux/.

Constraint: Range is 100 to 20000 for Provisioned IOPS (SSD) volumes and 3 to 10000 for General Purpose (SSD) volumes.

Condition: This parameter is required for requests to create io1 volumes; it is not used in requests to create standard or gp2 volumes.

cvrKmsKeyId :: Lens' CreateVolumeResponse (Maybe Text)

The full ARN of the AWS Key Management Service (KMS) master key that was used to protect the volume encryption key for the volume.

cvrSize :: Lens' CreateVolumeResponse Int

The size of the volume, in GiBs.

cvrSnapshotId :: Lens' CreateVolumeResponse Text

The snapshot from which the volume was created, if applicable.

cvrState :: Lens' CreateVolumeResponse VolumeState

The volume state.

cvrTags :: Lens' CreateVolumeResponse [Tag]

Any tags assigned to the volume.

cvrVolumeId :: Lens' CreateVolumeResponse Text

The ID of the volume.

cvrVolumeType :: Lens' CreateVolumeResponse VolumeType

The volume type. This can be gp2 for General Purpose (SSD) volumes, io1 for Provisioned IOPS (SSD) volumes, or standard for Magnetic volumes.