Network.AWS.EC2.CreateVpcPeeringConnection

Description

Requests a VPC peering connection between two VPCs: a requester VPC that you own and a peer VPC with which to create the connection. The peer VPC can belong to another AWS account. The requester VPC and peer VPC cannot have overlapping CIDR blocks.

The owner of the peer VPC must accept the peering request to activate the peering connection. The VPC peering connection request expires after 7 days, after which it cannot be accepted or rejected.

A CreateVpcPeeringConnection request between VPCs with overlapping CIDR blocks results in the VPC peering connection having a status of failed.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateVpcPeeringConnection.html

Request

data CreateVpcPeeringConnection

Request constructor

createVpcPeeringConnection :: CreateVpcPeeringConnection

CreateVpcPeeringConnection constructor.

The fields accessible through corresponding lenses are:

cvpcDryRun :: Maybe Bool

cvpcPeerOwnerId :: Maybe Text

cvpcPeerVpcId :: Maybe Text

cvpcVpcId :: Maybe Text

Request lenses

cvpcDryRun :: Lens' CreateVpcPeeringConnection (Maybe Bool)

cvpcPeerOwnerId :: Lens' CreateVpcPeeringConnection (Maybe Text)

The AWS account ID of the owner of the peer VPC.

Default: Your AWS account ID

cvpcPeerVpcId :: Lens' CreateVpcPeeringConnection (Maybe Text)

The ID of the VPC with which you are creating the VPC peering connection.

cvpcVpcId :: Lens' CreateVpcPeeringConnection (Maybe Text)

The ID of the requester VPC.

Response

data CreateVpcPeeringConnectionResponse

Response constructor

createVpcPeeringConnectionResponse :: CreateVpcPeeringConnectionResponse

CreateVpcPeeringConnectionResponse constructor.

The fields accessible through corresponding lenses are:

cvpcrVpcPeeringConnection :: Maybe VpcPeeringConnection

Response lenses

cvpcrVpcPeeringConnection :: Lens' CreateVpcPeeringConnectionResponse (Maybe VpcPeeringConnection)

Information about the VPC peering connection.