Network.AWS.EC2.CreateVpnConnection

Description

Creates a VPN connection between an existing virtual private gateway and a VPN customer gateway. The only supported connection type is 'ipsec.1'.

The response includes information that you need to give to your network administrator to configure your customer gateway.

We strongly recommend that you use HTTPS when calling this operation because the response contains sensitive cryptographic information for configuring your customer gateway.

If you decide to shut down your VPN connection for any reason and later create a new VPN connection, you must reconfigure your customer gateway with the new information returned from this call.

For more information about VPN connections, see Adding a Hardware VirtualPrivate Gateway to Your VPC in the Amazon Virtual Private Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateVpnConnection.html

Request

data CreateVpnConnection

Request constructor

createVpnConnection

Arguments

:: Text
cvcType

-> Text
cvcCustomerGatewayId

-> Text
cvcVpnGatewayId

-> CreateVpnConnection

CreateVpnConnection constructor.

The fields accessible through corresponding lenses are:

cvcCustomerGatewayId :: Text

cvcDryRun :: Maybe Bool

cvcOptions :: Maybe VpnConnectionOptionsSpecification

cvcType :: Text

cvcVpnGatewayId :: Text

Request lenses

cvcCustomerGatewayId :: Lens' CreateVpnConnection Text

The ID of the customer gateway.

cvcDryRun :: Lens' CreateVpnConnection (Maybe Bool)

cvcOptions :: Lens' CreateVpnConnection (Maybe VpnConnectionOptionsSpecification)

Indicates whether the VPN connection requires static routes. If you are creating a VPN connection for a device that does not support BGP, you must specify true.

Default: false

cvcType :: Lens' CreateVpnConnection Text

The type of VPN connection ('ipsec.1').

cvcVpnGatewayId :: Lens' CreateVpnConnection Text

The ID of the virtual private gateway.

Response

data CreateVpnConnectionResponse

Response constructor

createVpnConnectionResponse :: CreateVpnConnectionResponse

CreateVpnConnectionResponse constructor.

The fields accessible through corresponding lenses are:

cvcrVpnConnection :: Maybe VpnConnection

Response lenses

cvcrVpnConnection :: Lens' CreateVpnConnectionResponse (Maybe VpnConnection)

Information about the VPN connection.