Network.AWS.EC2.CreateVpnConnectionRoute

Description

Creates a static route associated with a VPN connection between an existing virtual private gateway and a VPN customer gateway. The static route allows traffic to be routed from the virtual private gateway to the VPN customer gateway.

For more information about VPN connections, see Adding a Hardware VirtualPrivate Gateway to Your VPC in the Amazon Virtual Private Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateVpnConnectionRoute.html

Request

data CreateVpnConnectionRoute

Request constructor

createVpnConnectionRoute

Arguments

:: Text
cvcrVpnConnectionId

-> Text
cvcrDestinationCidrBlock

-> CreateVpnConnectionRoute

CreateVpnConnectionRoute constructor.

The fields accessible through corresponding lenses are:

cvcrDestinationCidrBlock :: Text

cvcrVpnConnectionId :: Text

Request lenses

cvcrDestinationCidrBlock :: Lens' CreateVpnConnectionRoute Text

The CIDR block associated with the local subnet of the customer network.

cvcrVpnConnectionId :: Lens' CreateVpnConnectionRoute Text

The ID of the VPN connection.

Response

data CreateVpnConnectionRouteResponse

Response constructor

createVpnConnectionRouteResponse :: CreateVpnConnectionRouteResponse

CreateVpnConnectionRouteResponse constructor.