Network.AWS.EC2.CreateVpnGateway

Description

Creates a virtual private gateway. A virtual private gateway is the endpoint on the VPC side of your VPN connection. You can create a virtual private gateway before creating the VPC itself.

For more information about virtual private gateways, see Adding a HardwareVirtual Private Gateway to Your VPC in the Amazon Virtual Private Cloud UserGuide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-CreateVpnGateway.html

Request

data CreateVpnGateway

Request constructor

createVpnGateway

Arguments

:: GatewayType
cvgType

-> CreateVpnGateway

CreateVpnGateway constructor.

The fields accessible through corresponding lenses are:

cvgAvailabilityZone :: Maybe Text

cvgDryRun :: Maybe Bool

cvgType :: GatewayType

Request lenses

cvgAvailabilityZone :: Lens' CreateVpnGateway (Maybe Text)

The Availability Zone for the virtual private gateway.

cvgDryRun :: Lens' CreateVpnGateway (Maybe Bool)

cvgType :: Lens' CreateVpnGateway GatewayType

The type of VPN connection this virtual private gateway supports.

Response

data CreateVpnGatewayResponse

Response constructor

createVpnGatewayResponse :: CreateVpnGatewayResponse

CreateVpnGatewayResponse constructor.

The fields accessible through corresponding lenses are:

cvgrVpnGateway :: Maybe VpnGateway

Response lenses

cvgrVpnGateway :: Lens' CreateVpnGatewayResponse (Maybe VpnGateway)

Information about the virtual private gateway.