Network.AWS.EC2.DeleteNatGateway

Description

Deletes the specified NAT gateway. Deleting a NAT gateway disassociates its Elastic IP address, but does not release the address from your account. Deleting a NAT gateway does not delete any NAT gateway routes in your route tables.

Creating a Request

deleteNatGateway

Arguments

:: Text
dngNatGatewayId

-> DeleteNatGateway

Creates a value of DeleteNatGateway with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dngNatGatewayId - The ID of the NAT gateway.

data DeleteNatGateway

Contains the parameters for DeleteNatGateway.

See: deleteNatGateway smart constructor.

Request Lenses

dngNatGatewayId :: Lens' DeleteNatGateway Text

The ID of the NAT gateway.

Destructuring the Response

deleteNatGatewayResponse

Arguments

:: Int
dngnrsResponseStatus

-> DeleteNatGatewayResponse

Creates a value of DeleteNatGatewayResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dngnrsNatGatewayId - The ID of the NAT gateway.

dngnrsResponseStatus - -- | The response status code.

data DeleteNatGatewayResponse

Contains the output of DeleteNatGateway.

See: deleteNatGatewayResponse smart constructor.

Response Lenses

dngnrsNatGatewayId :: Lens' DeleteNatGatewayResponse (Maybe Text)

The ID of the NAT gateway.

dngnrsResponseStatus :: Lens' DeleteNatGatewayResponse Int

- | The response status code.