Network.AWS.EC2.DeleteSecurityGroup

Description

Deletes a security group.

If you attempt to delete a security group that is associated with an instance, or is referenced by another security group, the operation fails with InvalidGroup.InUse in EC2-Classic or DependencyViolation in EC2-VPC.

Creating a Request

deleteSecurityGroup :: DeleteSecurityGroup

Creates a value of DeleteSecurityGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsgGroupId - The ID of the security group. Required for a nondefault VPC.

dsgGroupName - [EC2-Classic, default VPC] The name of the security group. You can specify either the security group name or the security group ID.

dsgDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data DeleteSecurityGroup

Contains the parameters for DeleteSecurityGroup.

See: deleteSecurityGroup smart constructor.

Request Lenses

dsgGroupId :: Lens' DeleteSecurityGroup (Maybe Text)

The ID of the security group. Required for a nondefault VPC.

dsgGroupName :: Lens' DeleteSecurityGroup (Maybe Text)

EC2-Classic, default VPC

The name of the security group. You can specify either the security group name or the security group ID.

dsgDryRun :: Lens' DeleteSecurityGroup (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

deleteSecurityGroupResponse :: DeleteSecurityGroupResponse

Creates a value of DeleteSecurityGroupResponse with the minimum fields required to make a request.

data DeleteSecurityGroupResponse

See: deleteSecurityGroupResponse smart constructor.