Network.AWS.EC2.DeleteSnapshot

Description

Deletes the specified snapshot.

When you make periodic snapshots of a volume, the snapshots are incremental, and only the blocks on the device that have changed since your last snapshot are saved in the new snapshot. When you delete a snapshot, only the data not needed for any other snapshot is removed. So regardless of which prior snapshots have been deleted, all active snapshots will have access to all the information needed to restore the volume.

You cannot delete a snapshot of the root device of an EBS volume used by a registered AMI. You must first de-register the AMI before you can delete the snapshot.

For more information, see Deleting an Amazon EBS Snapshot in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

deleteSnapshot

Arguments

:: Text
deleSnapshotId

-> DeleteSnapshot

Creates a value of DeleteSnapshot with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

deleDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

deleSnapshotId - The ID of the EBS snapshot.

data DeleteSnapshot

Contains the parameters for DeleteSnapshot.

See: deleteSnapshot smart constructor.

Request Lenses

deleDryRun :: Lens' DeleteSnapshot (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

deleSnapshotId :: Lens' DeleteSnapshot Text

The ID of the EBS snapshot.

Destructuring the Response

deleteSnapshotResponse :: DeleteSnapshotResponse

Creates a value of DeleteSnapshotResponse with the minimum fields required to make a request.

data DeleteSnapshotResponse

See: deleteSnapshotResponse smart constructor.