Network.AWS.EC2.DeleteTags

Description

Deletes the specified set of tags from the specified set of resources. This call is designed to follow a DescribeTags request.

For more information about tags, see Tagging Your Resources in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

deleteTags :: DeleteTags

Creates a value of DeleteTags with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtsDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dtsTags - One or more tags to delete. If you omit the value parameter, we delete the tag regardless of its value. If you specify this parameter with an empty string as the value, we delete the key only if its value is an empty string.

dtsResources - The ID of the resource. For example, ami-1a2b3c4d. You can specify more than one resource ID.

data DeleteTags

Contains the parameters for DeleteTags.

See: deleteTags smart constructor.

Request Lenses

dtsDryRun :: Lens' DeleteTags (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dtsTags :: Lens' DeleteTags [Tag]

One or more tags to delete. If you omit the value parameter, we delete the tag regardless of its value. If you specify this parameter with an empty string as the value, we delete the key only if its value is an empty string.

dtsResources :: Lens' DeleteTags [Text]

The ID of the resource. For example, ami-1a2b3c4d. You can specify more than one resource ID.

Destructuring the Response

deleteTagsResponse :: DeleteTagsResponse

Creates a value of DeleteTagsResponse with the minimum fields required to make a request.

data DeleteTagsResponse

See: deleteTagsResponse smart constructor.