Network.AWS.EC2.DeleteVolume

Description

Deletes the specified EBS volume. The volume must be in the available state (not attached to an instance).

For more information, see Deleting an Amazon EBS Volume in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

deleteVolume

Arguments

:: Text
dvvVolumeId

-> DeleteVolume

Creates a value of DeleteVolume with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvvDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dvvVolumeId - The ID of the volume.

data DeleteVolume

Contains the parameters for DeleteVolume.

See: deleteVolume smart constructor.

Request Lenses

dvvDryRun :: Lens' DeleteVolume (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dvvVolumeId :: Lens' DeleteVolume Text

The ID of the volume.

Destructuring the Response

deleteVolumeResponse :: DeleteVolumeResponse

Creates a value of DeleteVolumeResponse with the minimum fields required to make a request.

data DeleteVolumeResponse

See: deleteVolumeResponse smart constructor.