Network.AWS.EC2.DescribeAddresses

Description

Describes one or more of your Elastic IP addresses.

An Elastic IP address is for use in either the EC2-Classic platform or in a VPC. For more information, see Elastic IP Addresses in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

describeAddresses :: DescribeAddresses

Creates a value of DescribeAddresses with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

daFilters - One or more filters. Filter names and values are case-sensitive. * allocation-id - [EC2-VPC] The allocation ID for the address. * association-id - [EC2-VPC] The association ID for the address. * domain - Indicates whether the address is for use in EC2-Classic (standard ) or in a VPC (vpc ). * instance-id - The ID of the instance the address is associated with, if any. * network-interface-id - [EC2-VPC] The ID of the network interface that the address is associated with, if any. * network-interface-owner-id - The AWS account ID of the owner. * private-ip-address - [EC2-VPC] The private IP address associated with the Elastic IP address. * public-ip - The Elastic IP address.

daPublicIPs - [EC2-Classic] One or more Elastic IP addresses. Default: Describes all your Elastic IP addresses.

daAllocationIds - [EC2-VPC] One or more allocation IDs. Default: Describes all your Elastic IP addresses.

daDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data DescribeAddresses

Contains the parameters for DescribeAddresses.

See: describeAddresses smart constructor.

Request Lenses

daFilters :: Lens' DescribeAddresses [Filter]

One or more filters. Filter names and values are case-sensitive. * allocation-id - [EC2-VPC] The allocation ID for the address. * association-id - [EC2-VPC] The association ID for the address. * domain - Indicates whether the address is for use in EC2-Classic (standard ) or in a VPC (vpc ). * instance-id - The ID of the instance the address is associated with, if any. * network-interface-id - [EC2-VPC] The ID of the network interface that the address is associated with, if any. * network-interface-owner-id - The AWS account ID of the owner. * private-ip-address - [EC2-VPC] The private IP address associated with the Elastic IP address. * public-ip - The Elastic IP address.

daPublicIPs :: Lens' DescribeAddresses [Text]

EC2-Classic

One or more Elastic IP addresses. Default: Describes all your Elastic IP addresses.

daAllocationIds :: Lens' DescribeAddresses [Text]

EC2-VPC

One or more allocation IDs. Default: Describes all your Elastic IP addresses.

daDryRun :: Lens' DescribeAddresses (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

describeAddressesResponse

Arguments

:: Int
darsResponseStatus

-> DescribeAddressesResponse

Creates a value of DescribeAddressesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

darsAddresses - Information about one or more Elastic IP addresses.

darsResponseStatus - -- | The response status code.

data DescribeAddressesResponse

Contains the output of DescribeAddresses.

See: describeAddressesResponse smart constructor.

Response Lenses

darsAddresses :: Lens' DescribeAddressesResponse [Address]

Information about one or more Elastic IP addresses.

darsResponseStatus :: Lens' DescribeAddressesResponse Int

- | The response status code.