Network.AWS.EC2.DescribeAvailabilityZones

Description

Describes one or more of the Availability Zones that are available to you. The results include zones only for the region you're currently using. If there is an event impacting an Availability Zone, you can use this request to view the state and any provided message for that Availability Zone.

For more information, see Regions and Availability Zones in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

describeAvailabilityZones :: DescribeAvailabilityZones

Creates a value of DescribeAvailabilityZones with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dazZoneNames - The names of one or more Availability Zones.

dazFilters - One or more filters. * message - Information about the Availability Zone. * region-name - The name of the region for the Availability Zone (for example, us-east-1 ). * state - The state of the Availability Zone (available | information | impaired | unavailable ). * zone-name - The name of the Availability Zone (for example, us-east-1a ).

dazDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data DescribeAvailabilityZones

Contains the parameters for DescribeAvailabilityZones.

See: describeAvailabilityZones smart constructor.

Request Lenses

dazZoneNames :: Lens' DescribeAvailabilityZones [Text]

The names of one or more Availability Zones.

dazFilters :: Lens' DescribeAvailabilityZones [Filter]

One or more filters. * message - Information about the Availability Zone. * region-name - The name of the region for the Availability Zone (for example, us-east-1 ). * state - The state of the Availability Zone (available | information | impaired | unavailable ). * zone-name - The name of the Availability Zone (for example, us-east-1a ).

dazDryRun :: Lens' DescribeAvailabilityZones (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

describeAvailabilityZonesResponse

Arguments

:: Int
dazrsResponseStatus

-> DescribeAvailabilityZonesResponse

Creates a value of DescribeAvailabilityZonesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dazrsAvailabilityZones - Information about one or more Availability Zones.

dazrsResponseStatus - -- | The response status code.

data DescribeAvailabilityZonesResponse

Contains the output of DescribeAvailabiltyZones.

See: describeAvailabilityZonesResponse smart constructor.

Response Lenses

dazrsAvailabilityZones :: Lens' DescribeAvailabilityZonesResponse [AvailabilityZone]

Information about one or more Availability Zones.

dazrsResponseStatus :: Lens' DescribeAvailabilityZonesResponse Int

- | The response status code.