Network.AWS.EC2.DescribeAvailabilityZones

Description

Describes one or more of the Availability Zones that are available to you. The results include zones only for the region you're currently using. If there is an event impacting an Availability Zone, you can use this request to view the state and any provided message for that Availability Zone. For more information, see Regions and Availability Zones in the Amazon Elastic Compute Cloud User Guide.

Request

data DescribeAvailabilityZones

Request constructor

describeAvailabilityZones :: DescribeAvailabilityZones

DescribeAvailabilityZones constructor.

The fields accessible through corresponding lenses are:

dazDryRun :: Maybe Bool

dazFilters :: [Filter]

dazZoneNames :: [Text]

Request lenses

dazDryRun :: Lens' DescribeAvailabilityZones (Maybe Bool)

dazFilters :: Lens' DescribeAvailabilityZones [Filter]

One or more filters. message - Information about the Availability Zone. region-name - The name of the region for the Availability Zone (for example, us-east-1). state - The state of the Availability Zone (available | impaired | unavailable). zone-name - The name of the Availability Zone (for example, us-east-1a).

dazZoneNames :: Lens' DescribeAvailabilityZones [Text]

The names of one or more Availability Zones.

Response

data DescribeAvailabilityZonesResult

Response constructor

describeAvailabilityZonesResult :: DescribeAvailabilityZonesResult

DescribeAvailabilityZonesResult constructor.

The fields accessible through corresponding lenses are:

dazrAvailabilityZones :: [AvailabilityZone]

Response lenses

dazrAvailabilityZones :: Lens' DescribeAvailabilityZonesResult [AvailabilityZone]

Information about one or more Availability Zones.