Network.AWS.EC2.DescribeBundleTasks

Description

Describes one or more of your bundling tasks.

Creating a Request

describeBundleTasks :: DescribeBundleTasks

Creates a value of DescribeBundleTasks with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dbtBundleIds - One or more bundle task IDs. Default: Describes all your bundle tasks.

dbtFilters - One or more filters. * bundle-id - The ID of the bundle task. * error-code - If the task failed, the error code returned. * error-message - If the task failed, the error message returned. * instance-id - The ID of the instance. * progress - The level of task completion, as a percentage (for example, 20%). * s3-bucket - The Amazon S3 bucket to store the AMI. * s3-prefix - The beginning of the AMI name. * start-time - The time the task started (for example, 2013-09-15T17:15:20.000Z). * state - The state of the task (pending | waiting-for-shutdown | bundling | storing | cancelling | complete | failed ). * update-time - The time of the most recent update for the task.

dbtDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data DescribeBundleTasks

Contains the parameters for DescribeBundleTasks.

See: describeBundleTasks smart constructor.

Request Lenses

dbtBundleIds :: Lens' DescribeBundleTasks [Text]

One or more bundle task IDs. Default: Describes all your bundle tasks.

dbtFilters :: Lens' DescribeBundleTasks [Filter]

One or more filters. * bundle-id - The ID of the bundle task. * error-code - If the task failed, the error code returned. * error-message - If the task failed, the error message returned. * instance-id - The ID of the instance. * progress - The level of task completion, as a percentage (for example, 20%). * s3-bucket - The Amazon S3 bucket to store the AMI. * s3-prefix - The beginning of the AMI name. * start-time - The time the task started (for example, 2013-09-15T17:15:20.000Z). * state - The state of the task (pending | waiting-for-shutdown | bundling | storing | cancelling | complete | failed ). * update-time - The time of the most recent update for the task.

dbtDryRun :: Lens' DescribeBundleTasks (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

describeBundleTasksResponse

Arguments

:: Int
dbtrsResponseStatus

-> DescribeBundleTasksResponse

Creates a value of DescribeBundleTasksResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dbtrsBundleTasks - Information about one or more bundle tasks.

dbtrsResponseStatus - -- | The response status code.

data DescribeBundleTasksResponse

Contains the output of DescribeBundleTasks.

See: describeBundleTasksResponse smart constructor.

Response Lenses

dbtrsBundleTasks :: Lens' DescribeBundleTasksResponse [BundleTask]

Information about one or more bundle tasks.

dbtrsResponseStatus :: Lens' DescribeBundleTasksResponse Int

- | The response status code.