Network.AWS.EC2.DescribeClassicLinkInstances

Description

Describes one or more of your linked EC2-Classic instances. This request only returns information about EC2-Classic instances linked to a VPC through ClassicLink; you cannot use this request to return information about other instances.

Creating a Request

describeClassicLinkInstances :: DescribeClassicLinkInstances

Creates a value of DescribeClassicLinkInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcliFilters - One or more filters. * group-id - The ID of a VPC security group that's associated with the instance. * instance-id - The ID of the instance. * tag :key =value - The keyvalue combination of a tag assigned to the resource. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter. * vpc-id - The ID of the VPC that the instance is linked to.

dcliNextToken - The token to retrieve the next page of results.

dcliInstanceIds - One or more instance IDs. Must be instances linked to a VPC through ClassicLink.

dcliDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dcliMaxResults - The maximum number of results to return for the request in a single page. The remaining results of the initial request can be seen by sending another request with the returned NextToken value. This value can be between 5 and 1000; if MaxResults is given a value larger than 1000, only 1000 results are returned. You cannot specify this parameter and the instance IDs parameter in the same request. Constraint: If the value is greater than 1000, we return only 1000 items.

data DescribeClassicLinkInstances

Contains the parameters for DescribeClassicLinkInstances.

See: describeClassicLinkInstances smart constructor.

Request Lenses

dcliFilters :: Lens' DescribeClassicLinkInstances [Filter]

One or more filters. * group-id - The ID of a VPC security group that's associated with the instance. * instance-id - The ID of the instance. * tag :key =value - The keyvalue combination of a tag assigned to the resource. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter. * vpc-id - The ID of the VPC that the instance is linked to.

dcliNextToken :: Lens' DescribeClassicLinkInstances (Maybe Text)

The token to retrieve the next page of results.

dcliInstanceIds :: Lens' DescribeClassicLinkInstances [Text]

One or more instance IDs. Must be instances linked to a VPC through ClassicLink.

dcliDryRun :: Lens' DescribeClassicLinkInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dcliMaxResults :: Lens' DescribeClassicLinkInstances (Maybe Int)

The maximum number of results to return for the request in a single page. The remaining results of the initial request can be seen by sending another request with the returned NextToken value. This value can be between 5 and 1000; if MaxResults is given a value larger than 1000, only 1000 results are returned. You cannot specify this parameter and the instance IDs parameter in the same request. Constraint: If the value is greater than 1000, we return only 1000 items.

Destructuring the Response

describeClassicLinkInstancesResponse

Arguments

:: Int
dclirsResponseStatus

-> DescribeClassicLinkInstancesResponse

Creates a value of DescribeClassicLinkInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dclirsNextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

dclirsInstances - Information about one or more linked EC2-Classic instances.

dclirsResponseStatus - -- | The response status code.

data DescribeClassicLinkInstancesResponse

Contains the output of DescribeClassicLinkInstances.

See: describeClassicLinkInstancesResponse smart constructor.

Response Lenses

dclirsNextToken :: Lens' DescribeClassicLinkInstancesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

dclirsInstances :: Lens' DescribeClassicLinkInstancesResponse [ClassicLinkInstance]

Information about one or more linked EC2-Classic instances.

dclirsResponseStatus :: Lens' DescribeClassicLinkInstancesResponse Int

- | The response status code.