Network.AWS.EC2.DescribeConversionTasks

Description

Describes one or more of your conversion tasks. For more information, see the VM Import/Export User Guide .

For information about the import manifest referenced by this API action, see VM Import Manifest .

Creating a Request

describeConversionTasks :: DescribeConversionTasks

Creates a value of DescribeConversionTasks with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dctConversionTaskIds - One or more conversion task IDs.

dctDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data DescribeConversionTasks

Contains the parameters for DescribeConversionTasks.

See: describeConversionTasks smart constructor.

Request Lenses

dctConversionTaskIds :: Lens' DescribeConversionTasks [Text]

One or more conversion task IDs.

dctDryRun :: Lens' DescribeConversionTasks (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

describeConversionTasksResponse

Arguments

:: Int
dctrsResponseStatus

-> DescribeConversionTasksResponse

Creates a value of DescribeConversionTasksResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dctrsConversionTasks - Information about the conversion tasks.

dctrsResponseStatus - -- | The response status code.

data DescribeConversionTasksResponse

Contains the output for DescribeConversionTasks.

See: describeConversionTasksResponse smart constructor.

Response Lenses

dctrsConversionTasks :: Lens' DescribeConversionTasksResponse [ConversionTask]

Information about the conversion tasks.

dctrsResponseStatus :: Lens' DescribeConversionTasksResponse Int

- | The response status code.