Network.AWS.EC2.DescribeCustomerGateways

Description

Describes one or more of your VPN customer gateways.

For more information about VPN customer gateways, see Adding a Hardware Virtual Private Gateway to Your VPC in the Amazon Virtual Private Cloud User Guide .

Creating a Request

describeCustomerGateways :: DescribeCustomerGateways

Creates a value of DescribeCustomerGateways with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcgCustomerGatewayIds - One or more customer gateway IDs. Default: Describes all your customer gateways.

dcgFilters - One or more filters. * bgp-asn - The customer gateway's Border Gateway Protocol (BGP) Autonomous System Number (ASN). * customer-gateway-id - The ID of the customer gateway. * ip-address - The IP address of the customer gateway's Internet-routable external interface. * state - The state of the customer gateway (pending | available | deleting | deleted ). * type - The type of customer gateway. Currently, the only supported type is ipsec.1 . * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter.

dcgDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data DescribeCustomerGateways

Contains the parameters for DescribeCustomerGateways.

See: describeCustomerGateways smart constructor.

Request Lenses

dcgCustomerGatewayIds :: Lens' DescribeCustomerGateways [Text]

One or more customer gateway IDs. Default: Describes all your customer gateways.

dcgFilters :: Lens' DescribeCustomerGateways [Filter]

One or more filters. * bgp-asn - The customer gateway's Border Gateway Protocol (BGP) Autonomous System Number (ASN). * customer-gateway-id - The ID of the customer gateway. * ip-address - The IP address of the customer gateway's Internet-routable external interface. * state - The state of the customer gateway (pending | available | deleting | deleted ). * type - The type of customer gateway. Currently, the only supported type is ipsec.1 . * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter.

dcgDryRun :: Lens' DescribeCustomerGateways (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

describeCustomerGatewaysResponse

Arguments

:: Int
dcgrsResponseStatus

-> DescribeCustomerGatewaysResponse

Creates a value of DescribeCustomerGatewaysResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcgrsCustomerGateways - Information about one or more customer gateways.

dcgrsResponseStatus - -- | The response status code.

data DescribeCustomerGatewaysResponse

Contains the output of DescribeCustomerGateways.

See: describeCustomerGatewaysResponse smart constructor.

Response Lenses

dcgrsCustomerGateways :: Lens' DescribeCustomerGatewaysResponse [CustomerGateway]

Information about one or more customer gateways.

dcgrsResponseStatus :: Lens' DescribeCustomerGatewaysResponse Int

- | The response status code.