Network.AWS.EC2.DescribeExportTasks

Description

Describes one or more of your export tasks.

Creating a Request

describeExportTasks :: DescribeExportTasks

Creates a value of DescribeExportTasks with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

detExportTaskIds - One or more export task IDs.

data DescribeExportTasks

Contains the parameters for DescribeExportTasks.

See: describeExportTasks smart constructor.

Request Lenses

detExportTaskIds :: Lens' DescribeExportTasks [Text]

One or more export task IDs.

Destructuring the Response

describeExportTasksResponse

Arguments

:: Int
detrsResponseStatus

-> DescribeExportTasksResponse

Creates a value of DescribeExportTasksResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

detrsExportTasks - Information about the export tasks.

detrsResponseStatus - -- | The response status code.

data DescribeExportTasksResponse

Contains the output for DescribeExportTasks.

See: describeExportTasksResponse smart constructor.

Response Lenses

detrsExportTasks :: Lens' DescribeExportTasksResponse [ExportTask]

Information about the export tasks.

detrsResponseStatus :: Lens' DescribeExportTasksResponse Int

- | The response status code.