Network.AWS.EC2.DescribeFlowLogs

Description

Describes one or more flow logs. To view the information in your flow logs (the log streams for the network interfaces), you must use the CloudWatch Logs console or the CloudWatch Logs API.

Creating a Request

describeFlowLogs :: DescribeFlowLogs

Creates a value of DescribeFlowLogs with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dNextToken - The token to retrieve the next page of results.

dFlowLogIds - One or more flow log IDs.

dFilter - One or more filters. * deliver-log-status - The status of the logs delivery (SUCCESS | FAILED ). * flow-log-id - The ID of the flow log. * log-group-name - The name of the log group. * resource-id - The ID of the VPC, subnet, or network interface. * traffic-type - The type of traffic (ACCEPT | REJECT | ALL )

dMaxResults - The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned NextToken value. This value can be between 5 and 1000; if MaxResults is given a value larger than 1000, only 1000 results are returned. You cannot specify this parameter and the flow log IDs parameter in the same request.

data DescribeFlowLogs

Contains the parameters for DescribeFlowLogs.

See: describeFlowLogs smart constructor.

Request Lenses

dNextToken :: Lens' DescribeFlowLogs (Maybe Text)

The token to retrieve the next page of results.

dFlowLogIds :: Lens' DescribeFlowLogs [Text]

One or more flow log IDs.

dFilter :: Lens' DescribeFlowLogs [Filter]

One or more filters. * deliver-log-status - The status of the logs delivery (SUCCESS | FAILED ). * flow-log-id - The ID of the flow log. * log-group-name - The name of the log group. * resource-id - The ID of the VPC, subnet, or network interface. * traffic-type - The type of traffic (ACCEPT | REJECT | ALL )

dMaxResults :: Lens' DescribeFlowLogs (Maybe Int)

The maximum number of results to return for the request in a single page. The remaining results can be seen by sending another request with the returned NextToken value. This value can be between 5 and 1000; if MaxResults is given a value larger than 1000, only 1000 results are returned. You cannot specify this parameter and the flow log IDs parameter in the same request.

Destructuring the Response

describeFlowLogsResponse

Arguments

:: Int
dflsrsResponseStatus

-> DescribeFlowLogsResponse

Creates a value of DescribeFlowLogsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dflsrsNextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

dflsrsFlowLogs - Information about the flow logs.

dflsrsResponseStatus - -- | The response status code.

data DescribeFlowLogsResponse

Contains the output of DescribeFlowLogs.

See: describeFlowLogsResponse smart constructor.

Response Lenses

dflsrsNextToken :: Lens' DescribeFlowLogsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

dflsrsFlowLogs :: Lens' DescribeFlowLogsResponse [FlowLog]

Information about the flow logs.

dflsrsResponseStatus :: Lens' DescribeFlowLogsResponse Int

- | The response status code.