Network.AWS.EC2.DescribeFpgaImages

Description

Describes one or more available Amazon FPGA Images (AFIs). These include public AFIs, private AFIs that you own, and AFIs owned by other AWS accounts for which you have load permissions.

Creating a Request

describeFpgaImages :: DescribeFpgaImages

Creates a value of DescribeFpgaImages with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfiOwners - Filters the AFI by owner. Specify an AWS account ID, self (owner is the sender of the request), or an AWS owner alias (valid values are amazon | aws-marketplace ).

dfiFilters - One or more filters. * create-time - The creation time of the AFI. * fpga-image-id - The FPGA image identifier (AFI ID). * fpga-image-global-id - The global FPGA image identifier (AGFI ID). * name - The name of the AFI. * owner-id - The AWS account ID of the AFI owner. * product-code - The product code. * shell-version - The version of the AWS Shell that was used to create the bitstream. * state - The state of the AFI (pending | failed | available | unavailable ). * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter. * update-time - The time of the most recent update.

dfiNextToken - The token to retrieve the next page of results.

dfiDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dfiMaxResults - The maximum number of results to return in a single call.

dfiFpgaImageIds - One or more AFI IDs.

data DescribeFpgaImages

See: describeFpgaImages smart constructor.

Request Lenses

dfiOwners :: Lens' DescribeFpgaImages [Text]

Filters the AFI by owner. Specify an AWS account ID, self (owner is the sender of the request), or an AWS owner alias (valid values are amazon | aws-marketplace ).

dfiFilters :: Lens' DescribeFpgaImages [Filter]

One or more filters. * create-time - The creation time of the AFI. * fpga-image-id - The FPGA image identifier (AFI ID). * fpga-image-global-id - The global FPGA image identifier (AGFI ID). * name - The name of the AFI. * owner-id - The AWS account ID of the AFI owner. * product-code - The product code. * shell-version - The version of the AWS Shell that was used to create the bitstream. * state - The state of the AFI (pending | failed | available | unavailable ). * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter. * update-time - The time of the most recent update.

dfiNextToken :: Lens' DescribeFpgaImages (Maybe Text)

The token to retrieve the next page of results.

dfiDryRun :: Lens' DescribeFpgaImages (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dfiMaxResults :: Lens' DescribeFpgaImages (Maybe Natural)

The maximum number of results to return in a single call.

dfiFpgaImageIds :: Lens' DescribeFpgaImages [Text]

One or more AFI IDs.

Destructuring the Response

describeFpgaImagesResponse

Arguments

:: Int
dfirsResponseStatus

-> DescribeFpgaImagesResponse

Creates a value of DescribeFpgaImagesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfirsFpgaImages - Information about one or more FPGA images.

dfirsNextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

dfirsResponseStatus - -- | The response status code.

data DescribeFpgaImagesResponse

See: describeFpgaImagesResponse smart constructor.

Response Lenses

dfirsFpgaImages :: Lens' DescribeFpgaImagesResponse [FpgaImage]

Information about one or more FPGA images.

dfirsNextToken :: Lens' DescribeFpgaImagesResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

dfirsResponseStatus :: Lens' DescribeFpgaImagesResponse Int

- | The response status code.