Network.AWS.EC2.DescribeImages

Description

Describes one or more of the images (AMIs, AKIs, and ARIs) available to you. Images available to you include public images, private images that you own, and private images owned by other AWS accounts but for which you have explicit launch permissions.

Creating a Request

describeImages :: DescribeImages

Creates a value of DescribeImages with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

deseOwners - Filters the images by the owner. Specify an AWS account ID, self (owner is the sender of the request), or an AWS owner alias (valid values are amazon | aws-marketplace | microsoft ). Omitting this option returns all images for which you have launch permissions, regardless of ownership.

deseExecutableUsers - Scopes the images by users with explicit launch permissions. Specify an AWS account ID, self (the sender of the request), or all (public AMIs).

deseFilters - One or more filters. * architecture - The image architecture (i386 | x86_64 ). * block-device-mapping.delete-on-termination - A Boolean value that indicates whether the Amazon EBS volume is deleted on instance termination. * block-device-mapping.device-name - The device name for the EBS volume (for example, devsdh ). * block-device-mapping.snapshot-id - The ID of the snapshot used for the EBS volume. * block-device-mapping.volume-size - The volume size of the EBS volume, in GiB. * block-device-mapping.volume-type - The volume type of the EBS volume (gp2 | io1 | st1 | sc1 | standard ). * description - The description of the image (provided during image creation). * ena-support - A Boolean that indicates whether enhanced networking with ENA is enabled. * hypervisor - The hypervisor type (ovm | xen ). * image-id - The ID of the image. * image-type - The image type (machine | kernel | ramdisk ). * is-public - A Boolean that indicates whether the image is public. * kernel-id - The kernel ID. * manifest-location - The location of the image manifest. * name - The name of the AMI (provided during image creation). * owner-alias - String value from an Amazon-maintained list (amazon | aws-marketplace | microsoft ) of snapshot owners. Not to be confused with the user-configured AWS account alias, which is set from the IAM console. * owner-id - The AWS account ID of the image owner. * platform - The platform. To only list Windows-based AMIs, use windows . * product-code - The product code. * product-code.type - The type of the product code (devpay | marketplace ). * ramdisk-id - The RAM disk ID. * root-device-name - The name of the root device volume (for example, devsda1 ). * root-device-type - The type of the root device volume (ebs | instance-store ). * state - The state of the image (available | pending | failed ). * state-reason-code - The reason code for the state change. * state-reason-message - The message for the state change. * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter. * virtualization-type - The virtualization type (paravirtual | hvm ).

deseImageIds - One or more image IDs. Default: Describes all images available to you.

deseDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data DescribeImages

Contains the parameters for DescribeImages.

See: describeImages smart constructor.

Request Lenses

deseOwners :: Lens' DescribeImages [Text]

Filters the images by the owner. Specify an AWS account ID, self (owner is the sender of the request), or an AWS owner alias (valid values are amazon | aws-marketplace | microsoft ). Omitting this option returns all images for which you have launch permissions, regardless of ownership.

deseExecutableUsers :: Lens' DescribeImages [Text]

Scopes the images by users with explicit launch permissions. Specify an AWS account ID, self (the sender of the request), or all (public AMIs).

deseFilters :: Lens' DescribeImages [Filter]

One or more filters. * architecture - The image architecture (i386 | x86_64 ). * block-device-mapping.delete-on-termination - A Boolean value that indicates whether the Amazon EBS volume is deleted on instance termination. * block-device-mapping.device-name - The device name for the EBS volume (for example, devsdh ). * block-device-mapping.snapshot-id - The ID of the snapshot used for the EBS volume. * block-device-mapping.volume-size - The volume size of the EBS volume, in GiB. * block-device-mapping.volume-type - The volume type of the EBS volume (gp2 | io1 | st1 | sc1 | standard ). * description - The description of the image (provided during image creation). * ena-support - A Boolean that indicates whether enhanced networking with ENA is enabled. * hypervisor - The hypervisor type (ovm | xen ). * image-id - The ID of the image. * image-type - The image type (machine | kernel | ramdisk ). * is-public - A Boolean that indicates whether the image is public. * kernel-id - The kernel ID. * manifest-location - The location of the image manifest. * name - The name of the AMI (provided during image creation). * owner-alias - String value from an Amazon-maintained list (amazon | aws-marketplace | microsoft ) of snapshot owners. Not to be confused with the user-configured AWS account alias, which is set from the IAM console. * owner-id - The AWS account ID of the image owner. * platform - The platform. To only list Windows-based AMIs, use windows . * product-code - The product code. * product-code.type - The type of the product code (devpay | marketplace ). * ramdisk-id - The RAM disk ID. * root-device-name - The name of the root device volume (for example, devsda1 ). * root-device-type - The type of the root device volume (ebs | instance-store ). * state - The state of the image (available | pending | failed ). * state-reason-code - The reason code for the state change. * state-reason-message - The message for the state change. * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter. * virtualization-type - The virtualization type (paravirtual | hvm ).

deseImageIds :: Lens' DescribeImages [Text]

One or more image IDs. Default: Describes all images available to you.

deseDryRun :: Lens' DescribeImages (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

describeImagesResponse

Arguments

:: Int
diirsResponseStatus

-> DescribeImagesResponse

Creates a value of DescribeImagesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diirsImages - Information about one or more images.

diirsResponseStatus - -- | The response status code.

data DescribeImagesResponse

Contains the output of DescribeImages.

See: describeImagesResponse smart constructor.

Response Lenses

diirsImages :: Lens' DescribeImagesResponse [Image]

Information about one or more images.

diirsResponseStatus :: Lens' DescribeImagesResponse Int

- | The response status code.