Network.AWS.EC2.DescribeImportImageTasks

Description

Displays details about an import virtual machine or import snapshot tasks that are already created.

Creating a Request

describeImportImageTasks :: DescribeImportImageTasks

Creates a value of DescribeImportImageTasks with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diitFilters - Filter tasks using the task-state filter and one of the following values: active, completed, deleting, deleted.

diitImportTaskIds - A list of import image task IDs.

diitNextToken - A token that indicates the next page of results.

diitDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

diitMaxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value.

data DescribeImportImageTasks

Contains the parameters for DescribeImportImageTasks.

See: describeImportImageTasks smart constructor.

Request Lenses

diitFilters :: Lens' DescribeImportImageTasks [Filter]

Filter tasks using the task-state filter and one of the following values: active, completed, deleting, deleted.

diitImportTaskIds :: Lens' DescribeImportImageTasks [Text]

A list of import image task IDs.

diitNextToken :: Lens' DescribeImportImageTasks (Maybe Text)

A token that indicates the next page of results.

diitDryRun :: Lens' DescribeImportImageTasks (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

diitMaxResults :: Lens' DescribeImportImageTasks (Maybe Int)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value.

Destructuring the Response

describeImportImageTasksResponse

Arguments

:: Int
diitrsResponseStatus

-> DescribeImportImageTasksResponse

Creates a value of DescribeImportImageTasksResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diitrsNextToken - The token to use to get the next page of results. This value is null when there are no more results to return.

diitrsImportImageTasks - A list of zero or more import image tasks that are currently active or were completed or canceled in the previous 7 days.

diitrsResponseStatus - -- | The response status code.

data DescribeImportImageTasksResponse

Contains the output for DescribeImportImageTasks.

See: describeImportImageTasksResponse smart constructor.

Response Lenses

diitrsNextToken :: Lens' DescribeImportImageTasksResponse (Maybe Text)

The token to use to get the next page of results. This value is null when there are no more results to return.

diitrsImportImageTasks :: Lens' DescribeImportImageTasksResponse [ImportImageTask]

A list of zero or more import image tasks that are currently active or were completed or canceled in the previous 7 days.

diitrsResponseStatus :: Lens' DescribeImportImageTasksResponse Int

- | The response status code.