Network.AWS.EC2.DescribeInstanceAttribute

Description

Describes the specified attribute of the specified instance. You can specify only one attribute at a time. Valid attribute values are: instanceType | kernel | ramdisk | userData | disableApiTermination | instanceInitiatedShutdownBehavior | rootDeviceName | blockDeviceMapping | productCodes | sourceDestCheck | groupSet | ebsOptimized | sriovNetSupport

Creating a Request

describeInstanceAttribute

Arguments

:: InstanceAttributeName
diaAttribute

-> Text
diaInstanceId

-> DescribeInstanceAttribute

Creates a value of DescribeInstanceAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

diaDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

diaAttribute - The instance attribute. Note: The enaSupport attribute is not supported at this time.

diaInstanceId - The ID of the instance.

data DescribeInstanceAttribute

Contains the parameters for DescribeInstanceAttribute.

See: describeInstanceAttribute smart constructor.

Request Lenses

diaDryRun :: Lens' DescribeInstanceAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

diaAttribute :: Lens' DescribeInstanceAttribute InstanceAttributeName

The instance attribute. Note: The enaSupport attribute is not supported at this time.

diaInstanceId :: Lens' DescribeInstanceAttribute Text

The ID of the instance.

Destructuring the Response

describeInstanceAttributeResponse

Arguments

:: Int
desrsResponseStatus

-> DescribeInstanceAttributeResponse

Creates a value of DescribeInstanceAttributeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

desrsInstanceId - The ID of the instance.

desrsGroups - The security groups associated with the instance.

desrsEnaSupport - Indicates whether enhanced networking with ENA is enabled.

desrsSourceDestCheck - Indicates whether source/destination checking is enabled. A value of true means checking is enabled, and false means checking is disabled. This value must be false for a NAT instance to perform NAT.

desrsDisableAPITermination - If the value is true , you can't terminate the instance through the Amazon EC2 console, CLI, or API; otherwise, you can.

desrsRAMDiskId - The RAM disk ID.

desrsKernelId - The kernel ID.

desrsRootDeviceName - The name of the root device (for example, devsda1 or devxvda ).

desrsInstanceType - The instance type.

desrsSRIOVNetSupport - Indicates whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

desrsEBSOptimized - Indicates whether the instance is optimized for EBS I/O.

desrsUserData - The user data.

desrsInstanceInitiatedShutdownBehavior - Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

desrsProductCodes - A list of product codes.

desrsBlockDeviceMappings - The block device mapping of the instance.

desrsResponseStatus - -- | The response status code.

data DescribeInstanceAttributeResponse

Describes an instance attribute.

See: describeInstanceAttributeResponse smart constructor.

Response Lenses

desrsInstanceId :: Lens' DescribeInstanceAttributeResponse (Maybe Text)

The ID of the instance.

desrsGroups :: Lens' DescribeInstanceAttributeResponse [GroupIdentifier]

The security groups associated with the instance.

desrsEnaSupport :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeBooleanValue)

Indicates whether enhanced networking with ENA is enabled.

desrsSourceDestCheck :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeBooleanValue)

Indicates whether source/destination checking is enabled. A value of true means checking is enabled, and false means checking is disabled. This value must be false for a NAT instance to perform NAT.

desrsDisableAPITermination :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeBooleanValue)

If the value is true , you can't terminate the instance through the Amazon EC2 console, CLI, or API; otherwise, you can.

desrsRAMDiskId :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

The RAM disk ID.

desrsKernelId :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

The kernel ID.

desrsRootDeviceName :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

The name of the root device (for example, devsda1 or devxvda ).

desrsInstanceType :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

The instance type.

desrsSRIOVNetSupport :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

Indicates whether enhanced networking with the Intel 82599 Virtual Function interface is enabled.

desrsEBSOptimized :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeBooleanValue)

Indicates whether the instance is optimized for EBS I/O.

desrsUserData :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

The user data.

desrsInstanceInitiatedShutdownBehavior :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

desrsProductCodes :: Lens' DescribeInstanceAttributeResponse [ProductCode]

A list of product codes.

desrsBlockDeviceMappings :: Lens' DescribeInstanceAttributeResponse [InstanceBlockDeviceMapping]

The block device mapping of the instance.

desrsResponseStatus :: Lens' DescribeInstanceAttributeResponse Int

- | The response status code.