Network.AWS.EC2.DescribeInstanceAttribute

Description

Describes the specified attribute of the specified instance. You can specify only one attribute at a time. Valid attribute values are: instanceType | kernel | ramdisk | userData | disableApiTermination | instanceInitiatedShutdownBehavior | rootDeviceName | blockDeviceMapping | productCodes | sourceDestCheck | groupSet | ebsOptimized | sriovNetSupport

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeInstanceAttribute.html

Request

data DescribeInstanceAttribute

Request constructor

describeInstanceAttribute

Arguments

:: Text
diaInstanceId

-> InstanceAttributeName
diaAttribute

-> DescribeInstanceAttribute

DescribeInstanceAttribute constructor.

The fields accessible through corresponding lenses are:

diaAttribute :: InstanceAttributeName

diaDryRun :: Maybe Bool

diaInstanceId :: Text

Request lenses

diaAttribute :: Lens' DescribeInstanceAttribute InstanceAttributeName

The instance attribute.

diaDryRun :: Lens' DescribeInstanceAttribute (Maybe Bool)

diaInstanceId :: Lens' DescribeInstanceAttribute Text

The ID of the instance.

Response

data DescribeInstanceAttributeResponse

Response constructor

describeInstanceAttributeResponse :: DescribeInstanceAttributeResponse

DescribeInstanceAttributeResponse constructor.

The fields accessible through corresponding lenses are:

diar1BlockDeviceMappings :: [InstanceBlockDeviceMapping]

diar1DisableApiTermination :: Maybe AttributeBooleanValue

diar1EbsOptimized :: Maybe AttributeBooleanValue

diar1Groups :: [GroupIdentifier]

diar1InstanceId :: Maybe Text

diar1InstanceInitiatedShutdownBehavior :: Maybe AttributeValue

diar1InstanceType :: Maybe AttributeValue

diar1KernelId :: Maybe AttributeValue

diar1ProductCodes :: [ProductCode]

diar1RamdiskId :: Maybe AttributeValue

diar1RootDeviceName :: Maybe AttributeValue

diar1SourceDestCheck :: Maybe AttributeBooleanValue

diar1SriovNetSupport :: Maybe AttributeValue

diar1UserData :: Maybe AttributeValue

Response lenses

diar1BlockDeviceMappings :: Lens' DescribeInstanceAttributeResponse [InstanceBlockDeviceMapping]

The block device mapping of the instance.

diar1DisableApiTermination :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeBooleanValue)

If the value is true, you can't terminate the instance through the Amazon EC2 console, CLI, or API; otherwise, you can.

diar1EbsOptimized :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeBooleanValue)

Indicates whether the instance is optimized for EBS I/O.

diar1Groups :: Lens' DescribeInstanceAttributeResponse [GroupIdentifier]

The security groups associated with the instance.

diar1InstanceId :: Lens' DescribeInstanceAttributeResponse (Maybe Text)

The ID of the instance.

diar1InstanceInitiatedShutdownBehavior :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

diar1InstanceType :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

The instance type.

diar1KernelId :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

The kernel ID.

diar1ProductCodes :: Lens' DescribeInstanceAttributeResponse [ProductCode]

A list of product codes.

diar1RamdiskId :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

The RAM disk ID.

diar1RootDeviceName :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

The name of the root device (for example, '/dev/sda1' or '/dev/xvda').

diar1SourceDestCheck :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeBooleanValue)

Indicates whether source/destination checking is enabled. A value of true means checking is enabled, and false means checking is disabled. This value must be false for a NAT instance to perform NAT.

diar1SriovNetSupport :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

diar1UserData :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

The Base64-encoded MIME user data.