Network.AWS.EC2.DescribeInstanceStatus

Description

Describes the status of one or more instances. By default, only running instances are described, unless you specifically indicate to return the status of all instances.

Instance status includes the following components:

Status checks - Amazon EC2 performs status checks on running EC2 instances to identify hardware and software issues. For more information, see Status Checks for Your Instances and Troubleshooting Instances with Failed Status Checks in the Amazon Elastic Compute Cloud User Guide .

Scheduled events - Amazon EC2 can schedule events (such as reboot, stop, or terminate) for your instances related to hardware issues, software updates, or system maintenance. For more information, see Scheduled Events for Your Instances in the Amazon Elastic Compute Cloud User Guide .

Instance state - You can manage your instances from the moment you launch them through their termination. For more information, see Instance Lifecycle in the Amazon Elastic Compute Cloud User Guide .

This operation returns paginated results.

Creating a Request

describeInstanceStatus :: DescribeInstanceStatus

Creates a value of DescribeInstanceStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

disIncludeAllInstances - When true , includes the health status for all instances. When false , includes the health status for running instances only. Default: false

disFilters - One or more filters. * availability-zone - The Availability Zone of the instance. * event.code - The code for the scheduled event (instance-reboot | system-reboot | system-maintenance | instance-retirement | instance-stop ). * event.description - A description of the event. * event.not-after - The latest end time for the scheduled event (for example, 2014-09-15T17:15:20.000Z ). * event.not-before - The earliest start time for the scheduled event (for example, 2014-09-15T17:15:20.000Z ). * instance-state-code - The code for the instance state, as a 16-bit unsigned integer. The high byte is an opaque internal value and should be ignored. The low byte is set based on the state represented. The valid values are 0 (pending), 16 (running), 32 (shutting-down), 48 (terminated), 64 (stopping), and 80 (stopped). * instance-state-name - The state of the instance (pending | running | shutting-down | terminated | stopping | stopped ). * instance-status.reachability - Filters on instance status where the name is reachability (passed | failed | initializing | insufficient-data ). * instance-status.status - The status of the instance (ok | impaired | initializing | insufficient-data | not-applicable ). * system-status.reachability - Filters on system status where the name is reachability (passed | failed | initializing | insufficient-data ). * system-status.status - The system status of the instance (ok | impaired | initializing | insufficient-data | not-applicable ).

disNextToken - The token to retrieve the next page of results.

disInstanceIds - One or more instance IDs. Default: Describes all your instances. Constraints: Maximum 100 explicitly specified instance IDs.

disDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

disMaxResults - The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. This value can be between 5 and 1000. You cannot specify this parameter and the instance IDs parameter in the same call.

data DescribeInstanceStatus

Contains the parameters for DescribeInstanceStatus.

See: describeInstanceStatus smart constructor.

Request Lenses

disIncludeAllInstances :: Lens' DescribeInstanceStatus (Maybe Bool)

When true , includes the health status for all instances. When false , includes the health status for running instances only. Default: false

disFilters :: Lens' DescribeInstanceStatus [Filter]

One or more filters. * availability-zone - The Availability Zone of the instance. * event.code - The code for the scheduled event (instance-reboot | system-reboot | system-maintenance | instance-retirement | instance-stop ). * event.description - A description of the event. * event.not-after - The latest end time for the scheduled event (for example, 2014-09-15T17:15:20.000Z ). * event.not-before - The earliest start time for the scheduled event (for example, 2014-09-15T17:15:20.000Z ). * instance-state-code - The code for the instance state, as a 16-bit unsigned integer. The high byte is an opaque internal value and should be ignored. The low byte is set based on the state represented. The valid values are 0 (pending), 16 (running), 32 (shutting-down), 48 (terminated), 64 (stopping), and 80 (stopped). * instance-state-name - The state of the instance (pending | running | shutting-down | terminated | stopping | stopped ). * instance-status.reachability - Filters on instance status where the name is reachability (passed | failed | initializing | insufficient-data ). * instance-status.status - The status of the instance (ok | impaired | initializing | insufficient-data | not-applicable ). * system-status.reachability - Filters on system status where the name is reachability (passed | failed | initializing | insufficient-data ). * system-status.status - The system status of the instance (ok | impaired | initializing | insufficient-data | not-applicable ).

disNextToken :: Lens' DescribeInstanceStatus (Maybe Text)

The token to retrieve the next page of results.

disInstanceIds :: Lens' DescribeInstanceStatus [Text]

One or more instance IDs. Default: Describes all your instances. Constraints: Maximum 100 explicitly specified instance IDs.

disDryRun :: Lens' DescribeInstanceStatus (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

disMaxResults :: Lens' DescribeInstanceStatus (Maybe Int)

The maximum number of results to return in a single call. To retrieve the remaining results, make another call with the returned NextToken value. This value can be between 5 and 1000. You cannot specify this parameter and the instance IDs parameter in the same call.

Destructuring the Response

describeInstanceStatusResponse

Arguments

:: Int
disrsResponseStatus

-> DescribeInstanceStatusResponse

Creates a value of DescribeInstanceStatusResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

disrsInstanceStatuses - One or more instance status descriptions.

disrsNextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

disrsResponseStatus - -- | The response status code.

data DescribeInstanceStatusResponse

Contains the output of DescribeInstanceStatus.

See: describeInstanceStatusResponse smart constructor.

Response Lenses

disrsInstanceStatuses :: Lens' DescribeInstanceStatusResponse [InstanceStatus]

One or more instance status descriptions.

disrsNextToken :: Lens' DescribeInstanceStatusResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

disrsResponseStatus :: Lens' DescribeInstanceStatusResponse Int

- | The response status code.