Network.AWS.EC2.DescribeInternetGateways

Description

Describes one or more of your Internet gateways.

Creating a Request

describeInternetGateways :: DescribeInternetGateways

Creates a value of DescribeInternetGateways with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dFilters - One or more filters. * attachment.state - The current state of the attachment between the gateway and the VPC (available ). Present only if a VPC is attached. * attachment.vpc-id - The ID of an attached VPC. * internet-gateway-id - The ID of the Internet gateway. * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter.

dInternetGatewayIds - One or more Internet gateway IDs. Default: Describes all your Internet gateways.

dDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data DescribeInternetGateways

Contains the parameters for DescribeInternetGateways.

See: describeInternetGateways smart constructor.

Request Lenses

dFilters :: Lens' DescribeInternetGateways [Filter]

One or more filters. * attachment.state - The current state of the attachment between the gateway and the VPC (available ). Present only if a VPC is attached. * attachment.vpc-id - The ID of an attached VPC. * internet-gateway-id - The ID of the Internet gateway. * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter.

dInternetGatewayIds :: Lens' DescribeInternetGateways [Text]

One or more Internet gateway IDs. Default: Describes all your Internet gateways.

dDryRun :: Lens' DescribeInternetGateways (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

describeInternetGatewaysResponse

Arguments

:: Int
digrsResponseStatus

-> DescribeInternetGatewaysResponse

Creates a value of DescribeInternetGatewaysResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

digrsInternetGateways - Information about one or more Internet gateways.

digrsResponseStatus - -- | The response status code.

data DescribeInternetGatewaysResponse

Contains the output of DescribeInternetGateways.

See: describeInternetGatewaysResponse smart constructor.

Response Lenses

digrsInternetGateways :: Lens' DescribeInternetGatewaysResponse [InternetGateway]

Information about one or more Internet gateways.

digrsResponseStatus :: Lens' DescribeInternetGatewaysResponse Int

- | The response status code.