Network.AWS.EC2.DescribeInternetGateways

Description

Describes one or more of your Internet gateways.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeInternetGateways.html

Request

data DescribeInternetGateways

Request constructor

describeInternetGateways :: DescribeInternetGateways

DescribeInternetGateways constructor.

The fields accessible through corresponding lenses are:

dig1DryRun :: Maybe Bool

dig1Filters :: [Filter]

dig1InternetGatewayIds :: [Text]

Request lenses

dig1DryRun :: Lens' DescribeInternetGateways (Maybe Bool)

dig1Filters :: Lens' DescribeInternetGateways [Filter]

One or more filters.

'attachment.state' - The current state of the attachment between the gateway and the VPC (available). Present only if a VPC is attached.

'attachment.vpc-id' - The ID of an attached VPC.

'internet-gateway-id' - The ID of the Internet gateway.

tag:key=value - The key/value combination of a tag assigned to the resource.

'tag-key' - The key of a tag assigned to the resource. This filter is independent of the 'tag-value' filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag:key=value filter.

'tag-value' - The value of a tag assigned to the resource. This filter is independent of the 'tag-key' filter.

dig1InternetGatewayIds :: Lens' DescribeInternetGateways [Text]

One or more Internet gateway IDs.

Default: Describes all your Internet gateways.

Response

data DescribeInternetGatewaysResponse

Response constructor

describeInternetGatewaysResponse :: DescribeInternetGatewaysResponse

DescribeInternetGatewaysResponse constructor.

The fields accessible through corresponding lenses are:

digrInternetGateways :: [InternetGateway]

Response lenses

digrInternetGateways :: Lens' DescribeInternetGatewaysResponse [InternetGateway]

Information about one or more Internet gateways.