Network.AWS.EC2.DescribeKeyPairs

Description

Describes one or more of your key pairs.

For more information about key pairs, see Key Pairs in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

describeKeyPairs :: DescribeKeyPairs

Creates a value of DescribeKeyPairs with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dkpsFilters - One or more filters. * fingerprint - The fingerprint of the key pair. * key-name - The name of the key pair.

dkpsKeyNames - One or more key pair names. Default: Describes all your key pairs.

dkpsDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data DescribeKeyPairs

Contains the parameters for DescribeKeyPairs.

See: describeKeyPairs smart constructor.

Request Lenses

dkpsFilters :: Lens' DescribeKeyPairs [Filter]

One or more filters. * fingerprint - The fingerprint of the key pair. * key-name - The name of the key pair.

dkpsKeyNames :: Lens' DescribeKeyPairs [Text]

One or more key pair names. Default: Describes all your key pairs.

dkpsDryRun :: Lens' DescribeKeyPairs (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

describeKeyPairsResponse

Arguments

:: Int
dkprsResponseStatus

-> DescribeKeyPairsResponse

Creates a value of DescribeKeyPairsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dkprsKeyPairs - Information about one or more key pairs.

dkprsResponseStatus - -- | The response status code.

data DescribeKeyPairsResponse

Contains the output of DescribeKeyPairs.

See: describeKeyPairsResponse smart constructor.

Response Lenses

dkprsKeyPairs :: Lens' DescribeKeyPairsResponse [KeyPairInfo]

Information about one or more key pairs.

dkprsResponseStatus :: Lens' DescribeKeyPairsResponse Int

- | The response status code.