Network.AWS.EC2.DescribeNetworkACLs

Description

Describes one or more of your network ACLs.

For more information about network ACLs, see Network ACLs in the Amazon Virtual Private Cloud User Guide .

Creating a Request

describeNetworkACLs :: DescribeNetworkACLs

Creates a value of DescribeNetworkACLs with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dnaclFilters - One or more filters. * association.association-id - The ID of an association ID for the ACL. * association.network-acl-id - The ID of the network ACL involved in the association. * association.subnet-id - The ID of the subnet involved in the association. * default - Indicates whether the ACL is the default network ACL for the VPC. * entry.cidr - The IPv4 CIDR range specified in the entry. * entry.egress - Indicates whether the entry applies to egress traffic. * entry.icmp.code - The ICMP code specified in the entry, if any. * entry.icmp.type - The ICMP type specified in the entry, if any. * entry.ipv6-cidr - The IPv6 CIDR range specified in the entry. * entry.port-range.from - The start of the port range specified in the entry. * entry.port-range.to - The end of the port range specified in the entry. * entry.protocol - The protocol specified in the entry (tcp | udp | icmp or a protocol number). * entry.rule-action - Allows or denies the matching traffic (allow | deny ). * entry.rule-number - The number of an entry (in other words, rule) in the ACL's set of entries. * network-acl-id - The ID of the network ACL. * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter. * vpc-id - The ID of the VPC for the network ACL.

dnaclNetworkACLIds - One or more network ACL IDs. Default: Describes all your network ACLs.

dnaclDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data DescribeNetworkACLs

Contains the parameters for DescribeNetworkAcls.

See: describeNetworkACLs smart constructor.

Request Lenses

dnaclFilters :: Lens' DescribeNetworkACLs [Filter]

One or more filters. * association.association-id - The ID of an association ID for the ACL. * association.network-acl-id - The ID of the network ACL involved in the association. * association.subnet-id - The ID of the subnet involved in the association. * default - Indicates whether the ACL is the default network ACL for the VPC. * entry.cidr - The IPv4 CIDR range specified in the entry. * entry.egress - Indicates whether the entry applies to egress traffic. * entry.icmp.code - The ICMP code specified in the entry, if any. * entry.icmp.type - The ICMP type specified in the entry, if any. * entry.ipv6-cidr - The IPv6 CIDR range specified in the entry. * entry.port-range.from - The start of the port range specified in the entry. * entry.port-range.to - The end of the port range specified in the entry. * entry.protocol - The protocol specified in the entry (tcp | udp | icmp or a protocol number). * entry.rule-action - Allows or denies the matching traffic (allow | deny ). * entry.rule-number - The number of an entry (in other words, rule) in the ACL's set of entries. * network-acl-id - The ID of the network ACL. * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter. * vpc-id - The ID of the VPC for the network ACL.

dnaclNetworkACLIds :: Lens' DescribeNetworkACLs [Text]

One or more network ACL IDs. Default: Describes all your network ACLs.

dnaclDryRun :: Lens' DescribeNetworkACLs (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

describeNetworkACLsResponse

Arguments

:: Int
dnarsResponseStatus

-> DescribeNetworkACLsResponse

Creates a value of DescribeNetworkACLsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dnarsNetworkACLs - Information about one or more network ACLs.

dnarsResponseStatus - -- | The response status code.

data DescribeNetworkACLsResponse

Contains the output of DescribeNetworkAcls.

See: describeNetworkACLsResponse smart constructor.

Response Lenses

dnarsNetworkACLs :: Lens' DescribeNetworkACLsResponse [NetworkACL]

Information about one or more network ACLs.

dnarsResponseStatus :: Lens' DescribeNetworkACLsResponse Int

- | The response status code.