Network.AWS.EC2.DescribeNetworkInterfaceAttribute

Description

Describes a network interface attribute. You can specify only one attribute at a time.

Creating a Request

describeNetworkInterfaceAttribute

Arguments

:: Text
dniaNetworkInterfaceId

-> DescribeNetworkInterfaceAttribute

Creates a value of DescribeNetworkInterfaceAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dniaAttribute - The attribute of the network interface. This parameter is required.

dniaDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dniaNetworkInterfaceId - The ID of the network interface.

data DescribeNetworkInterfaceAttribute

Contains the parameters for DescribeNetworkInterfaceAttribute.

See: describeNetworkInterfaceAttribute smart constructor.

Request Lenses

dniaAttribute :: Lens' DescribeNetworkInterfaceAttribute (Maybe NetworkInterfaceAttribute)

The attribute of the network interface. This parameter is required.

dniaDryRun :: Lens' DescribeNetworkInterfaceAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dniaNetworkInterfaceId :: Lens' DescribeNetworkInterfaceAttribute Text

The ID of the network interface.

Destructuring the Response

describeNetworkInterfaceAttributeResponse

Arguments

:: Int
dniarsResponseStatus

-> DescribeNetworkInterfaceAttributeResponse

Creates a value of DescribeNetworkInterfaceAttributeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dniarsGroups - The security groups associated with the network interface.

dniarsSourceDestCheck - Indicates whether source/destination checking is enabled.

dniarsNetworkInterfaceId - The ID of the network interface.

dniarsAttachment - The attachment (if any) of the network interface.

dniarsDescription - The description of the network interface.

dniarsResponseStatus - -- | The response status code.

data DescribeNetworkInterfaceAttributeResponse

Contains the output of DescribeNetworkInterfaceAttribute.

See: describeNetworkInterfaceAttributeResponse smart constructor.

Response Lenses

dniarsGroups :: Lens' DescribeNetworkInterfaceAttributeResponse [GroupIdentifier]

The security groups associated with the network interface.

dniarsSourceDestCheck :: Lens' DescribeNetworkInterfaceAttributeResponse (Maybe AttributeBooleanValue)

Indicates whether source/destination checking is enabled.

dniarsNetworkInterfaceId :: Lens' DescribeNetworkInterfaceAttributeResponse (Maybe Text)

The ID of the network interface.

dniarsAttachment :: Lens' DescribeNetworkInterfaceAttributeResponse (Maybe NetworkInterfaceAttachment)

The attachment (if any) of the network interface.

dniarsDescription :: Lens' DescribeNetworkInterfaceAttributeResponse (Maybe AttributeValue)

The description of the network interface.

dniarsResponseStatus :: Lens' DescribeNetworkInterfaceAttributeResponse Int

- | The response status code.