Network.AWS.EC2.DescribePlacementGroups

Description

Describes one or more of your placement groups. For more information about placement groups and cluster instances, see Cluster Instances in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

describePlacementGroups :: DescribePlacementGroups

Creates a value of DescribePlacementGroups with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dpgsFilters - One or more filters. * group-name - The name of the placement group. * state - The state of the placement group (pending | available | deleting | deleted ). * strategy - The strategy of the placement group (cluster ).

dpgsGroupNames - One or more placement group names. Default: Describes all your placement groups, or only those otherwise specified.

dpgsDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data DescribePlacementGroups

Contains the parameters for DescribePlacementGroups.

See: describePlacementGroups smart constructor.

Request Lenses

dpgsFilters :: Lens' DescribePlacementGroups [Filter]

One or more filters. * group-name - The name of the placement group. * state - The state of the placement group (pending | available | deleting | deleted ). * strategy - The strategy of the placement group (cluster ).

dpgsGroupNames :: Lens' DescribePlacementGroups [Text]

One or more placement group names. Default: Describes all your placement groups, or only those otherwise specified.

dpgsDryRun :: Lens' DescribePlacementGroups (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

describePlacementGroupsResponse

Arguments

:: Int
dpgrsResponseStatus

-> DescribePlacementGroupsResponse

Creates a value of DescribePlacementGroupsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dpgrsPlacementGroups - One or more placement groups.

dpgrsResponseStatus - -- | The response status code.

data DescribePlacementGroupsResponse

Contains the output of DescribePlacementGroups.

See: describePlacementGroupsResponse smart constructor.

Response Lenses

dpgrsPlacementGroups :: Lens' DescribePlacementGroupsResponse [PlacementGroup]

One or more placement groups.

dpgrsResponseStatus :: Lens' DescribePlacementGroupsResponse Int

- | The response status code.