Network.AWS.EC2.DescribePlacementGroups

Description

Describes one or more of your placement groups. For more information about placement groups and cluster instances, see Cluster Instances in the AmazonElastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribePlacementGroups.html

Request

data DescribePlacementGroups

Request constructor

describePlacementGroups :: DescribePlacementGroups

DescribePlacementGroups constructor.

The fields accessible through corresponding lenses are:

dpg1DryRun :: Maybe Bool

dpg1Filters :: [Filter]

dpg1GroupNames :: [Text]

Request lenses

dpg1DryRun :: Lens' DescribePlacementGroups (Maybe Bool)

dpg1Filters :: Lens' DescribePlacementGroups [Filter]

One or more filters.

'group-name' - The name of the placement group.

state - The state of the placement group (pending | available | deleting | deleted).

strategy - The strategy of the placement group (cluster).

dpg1GroupNames :: Lens' DescribePlacementGroups [Text]

One or more placement group names.

Default: Describes all your placement groups, or only those otherwise specified.

Response

data DescribePlacementGroupsResponse

Response constructor

describePlacementGroupsResponse :: DescribePlacementGroupsResponse

DescribePlacementGroupsResponse constructor.

The fields accessible through corresponding lenses are:

dpgrPlacementGroups :: [PlacementGroup]

Response lenses

dpgrPlacementGroups :: Lens' DescribePlacementGroupsResponse [PlacementGroup]

One or more placement groups.