Network.AWS.EC2.DescribeRegions

Description

Describes one or more regions that are currently available to you.

For a list of the regions supported by Amazon EC2, see Regions and Endpoints .

Creating a Request

describeRegions :: DescribeRegions

Creates a value of DescribeRegions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsRegionNames - The names of one or more regions.

drsFilters - One or more filters. * endpoint - The endpoint of the region (for example, ec2.us-east-1.amazonaws.com ). * region-name - The name of the region (for example, us-east-1 ).

drsDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data DescribeRegions

Contains the parameters for DescribeRegions.

See: describeRegions smart constructor.

Request Lenses

drsRegionNames :: Lens' DescribeRegions [Text]

The names of one or more regions.

drsFilters :: Lens' DescribeRegions [Filter]

One or more filters. * endpoint - The endpoint of the region (for example, ec2.us-east-1.amazonaws.com ). * region-name - The name of the region (for example, us-east-1 ).

drsDryRun :: Lens' DescribeRegions (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

describeRegionsResponse

Arguments

:: Int
drrsResponseStatus

-> DescribeRegionsResponse

Creates a value of DescribeRegionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drrsRegions - Information about one or more regions.

drrsResponseStatus - -- | The response status code.

data DescribeRegionsResponse

Contains the output of DescribeRegions.

See: describeRegionsResponse smart constructor.

Response Lenses

drrsRegions :: Lens' DescribeRegionsResponse [RegionInfo]

Information about one or more regions.

drrsResponseStatus :: Lens' DescribeRegionsResponse Int

- | The response status code.