Network.AWS.EC2.DescribeReservedInstances

Description

Describes one or more of the Reserved Instances that you purchased.

For more information about Reserved Instances, see Reserved Instances in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

describeReservedInstances :: DescribeReservedInstances

Creates a value of DescribeReservedInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

driFilters - One or more filters. * availability-zone - The Availability Zone where the Reserved Instance can be used. * duration - The duration of the Reserved Instance (one year or three years), in seconds (31536000 | 94608000 ). * end - The time when the Reserved Instance expires (for example, 2015-08-07T11:54:42.000Z). * fixed-price - The purchase price of the Reserved Instance (for example, 9800.0). * instance-type - The instance type that is covered by the reservation. * scope - The scope of the Reserved Instance (Region or Availability Zone ). * product-description - The Reserved Instance product platform description. Instances that include (Amazon VPC) in the product platform description will only be displayed to EC2-Classic account holders and are for use with Amazon VPC (Linux/UNIX | Linux/UNIX (Amazon VPC) | SUSE Linux | SUSE Linux (Amazon VPC) | Red Hat Enterprise Linux | Red Hat Enterprise Linux (Amazon VPC) | Windows | Windows (Amazon VPC) | Windows with SQL Server Standard | Windows with SQL Server Standard (Amazon VPC) | Windows with SQL Server Web | Windows with SQL Server Web (Amazon VPC) | Windows with SQL Server Enterprise | Windows with SQL Server Enterprise (Amazon VPC) ). * reserved-instances-id - The ID of the Reserved Instance. * start - The time at which the Reserved Instance purchase request was placed (for example, 2014-08-07T11:54:42.000Z). * state - The state of the Reserved Instance (payment-pending | active | payment-failed | retired ). * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter. * usage-price - The usage price of the Reserved Instance, per hour (for example, 0.84).

driReservedInstancesIds - One or more Reserved Instance IDs. Default: Describes all your Reserved Instances, or only those otherwise specified.

driOfferingType - The Reserved Instance offering type. If you are using tools that predate the 2011-11-01 API version, you only have access to the Medium Utilization Reserved Instance offering type.

driOfferingClass - Describes whether the Reserved Instance is Standard or Convertible.

driDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data DescribeReservedInstances

Contains the parameters for DescribeReservedInstances.

See: describeReservedInstances smart constructor.

Request Lenses

driFilters :: Lens' DescribeReservedInstances [Filter]

One or more filters. * availability-zone - The Availability Zone where the Reserved Instance can be used. * duration - The duration of the Reserved Instance (one year or three years), in seconds (31536000 | 94608000 ). * end - The time when the Reserved Instance expires (for example, 2015-08-07T11:54:42.000Z). * fixed-price - The purchase price of the Reserved Instance (for example, 9800.0). * instance-type - The instance type that is covered by the reservation. * scope - The scope of the Reserved Instance (Region or Availability Zone ). * product-description - The Reserved Instance product platform description. Instances that include (Amazon VPC) in the product platform description will only be displayed to EC2-Classic account holders and are for use with Amazon VPC (Linux/UNIX | Linux/UNIX (Amazon VPC) | SUSE Linux | SUSE Linux (Amazon VPC) | Red Hat Enterprise Linux | Red Hat Enterprise Linux (Amazon VPC) | Windows | Windows (Amazon VPC) | Windows with SQL Server Standard | Windows with SQL Server Standard (Amazon VPC) | Windows with SQL Server Web | Windows with SQL Server Web (Amazon VPC) | Windows with SQL Server Enterprise | Windows with SQL Server Enterprise (Amazon VPC) ). * reserved-instances-id - The ID of the Reserved Instance. * start - The time at which the Reserved Instance purchase request was placed (for example, 2014-08-07T11:54:42.000Z). * state - The state of the Reserved Instance (payment-pending | active | payment-failed | retired ). * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter. * usage-price - The usage price of the Reserved Instance, per hour (for example, 0.84).

driReservedInstancesIds :: Lens' DescribeReservedInstances [Text]

One or more Reserved Instance IDs. Default: Describes all your Reserved Instances, or only those otherwise specified.

driOfferingType :: Lens' DescribeReservedInstances (Maybe OfferingTypeValues)

The Reserved Instance offering type. If you are using tools that predate the 2011-11-01 API version, you only have access to the Medium Utilization Reserved Instance offering type.

driOfferingClass :: Lens' DescribeReservedInstances (Maybe OfferingClassType)

Describes whether the Reserved Instance is Standard or Convertible.

driDryRun :: Lens' DescribeReservedInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

describeReservedInstancesResponse

Arguments

:: Int
drirsResponseStatus

-> DescribeReservedInstancesResponse

Creates a value of DescribeReservedInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drirsReservedInstances - A list of Reserved Instances.

drirsResponseStatus - -- | The response status code.

data DescribeReservedInstancesResponse

Contains the output for DescribeReservedInstances.

See: describeReservedInstancesResponse smart constructor.

Response Lenses

drirsReservedInstances :: Lens' DescribeReservedInstancesResponse [ReservedInstances]

A list of Reserved Instances.

drirsResponseStatus :: Lens' DescribeReservedInstancesResponse Int

- | The response status code.