Network.AWS.EC2.DescribeReservedInstancesListings

Description

Describes your account's Reserved Instance listings in the Reserved Instance Marketplace.

The Reserved Instance Marketplace matches sellers who want to resell Reserved Instance capacity that they no longer need with buyers who want to purchase additional capacity. Reserved Instances bought and sold through the Reserved Instance Marketplace work like any other Reserved Instances.

As a seller, you choose to list some or all of your Reserved Instances, and you specify the upfront price to receive for them. Your Reserved Instances are then listed in the Reserved Instance Marketplace and are available for purchase.

As a buyer, you specify the configuration of the Reserved Instance to purchase, and the Marketplace matches what you're searching for with what's available. The Marketplace first sells the lowest priced Reserved Instances to you, and continues to sell available Reserved Instance listings to you until your demand is met. You are charged based on the total price of all of the listings that you purchase.

For more information, see Reserved Instance Marketplace in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

describeReservedInstancesListings :: DescribeReservedInstancesListings

Creates a value of DescribeReservedInstancesListings with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drilFilters - One or more filters. * reserved-instances-id - The ID of the Reserved Instances. * reserved-instances-listing-id - The ID of the Reserved Instances listing. * status - The status of the Reserved Instance listing (pending | active | cancelled | closed ). * status-message - The reason for the status.

drilReservedInstancesId - One or more Reserved Instance IDs.

drilReservedInstancesListingId - One or more Reserved Instance listing IDs.

data DescribeReservedInstancesListings

Contains the parameters for DescribeReservedInstancesListings.

See: describeReservedInstancesListings smart constructor.

Request Lenses

drilFilters :: Lens' DescribeReservedInstancesListings [Filter]

One or more filters. * reserved-instances-id - The ID of the Reserved Instances. * reserved-instances-listing-id - The ID of the Reserved Instances listing. * status - The status of the Reserved Instance listing (pending | active | cancelled | closed ). * status-message - The reason for the status.

drilReservedInstancesId :: Lens' DescribeReservedInstancesListings (Maybe Text)

One or more Reserved Instance IDs.

drilReservedInstancesListingId :: Lens' DescribeReservedInstancesListings (Maybe Text)

One or more Reserved Instance listing IDs.

Destructuring the Response

describeReservedInstancesListingsResponse

Arguments

:: Int
drilrsResponseStatus

-> DescribeReservedInstancesListingsResponse

Creates a value of DescribeReservedInstancesListingsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drilrsReservedInstancesListings - Information about the Reserved Instance listing.

drilrsResponseStatus - -- | The response status code.

data DescribeReservedInstancesListingsResponse

Contains the output of DescribeReservedInstancesListings.

See: describeReservedInstancesListingsResponse smart constructor.

Response Lenses

drilrsReservedInstancesListings :: Lens' DescribeReservedInstancesListingsResponse [ReservedInstancesListing]

Information about the Reserved Instance listing.

drilrsResponseStatus :: Lens' DescribeReservedInstancesListingsResponse Int

- | The response status code.