Network.AWS.EC2.DescribeReservedInstancesModifications

Description

Describes the modifications made to your Reserved Instances. If no parameter is specified, information about all your Reserved Instances modification requests is returned. If a modification ID is specified, only information about the specific modification is returned.

For more information, see Modifying Reserved Instances in the Amazon Elastic Compute Cloud User Guide.

This operation returns paginated results.

Creating a Request

describeReservedInstancesModifications :: DescribeReservedInstancesModifications

Creates a value of DescribeReservedInstancesModifications with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drimFilters - One or more filters. * client-token - The idempotency token for the modification request. * create-date - The time when the modification request was created. * effective-date - The time when the modification becomes effective. * modification-result.reserved-instances-id - The ID for the Reserved Instances created as part of the modification request. This ID is only available when the status of the modification is fulfilled . * modification-result.target-configuration.availability-zone - The Availability Zone for the new Reserved Instances. * modification-result.target-configuration.instance-count - The number of new Reserved Instances. * modification-result.target-configuration.instance-type - The instance type of the new Reserved Instances. * modification-result.target-configuration.platform - The network platform of the new Reserved Instances (EC2-Classic | EC2-VPC ). * reserved-instances-id - The ID of the Reserved Instances modified. * reserved-instances-modification-id - The ID of the modification request. * status - The status of the Reserved Instances modification request (processing | fulfilled | failed ). * status-message - The reason for the status. * update-date - The time when the modification request was last updated.

drimReservedInstancesModificationIds - IDs for the submitted modification request.

drimNextToken - The token to retrieve the next page of results.

data DescribeReservedInstancesModifications

Contains the parameters for DescribeReservedInstancesModifications.

See: describeReservedInstancesModifications smart constructor.

Request Lenses

drimFilters :: Lens' DescribeReservedInstancesModifications [Filter]

One or more filters. * client-token - The idempotency token for the modification request. * create-date - The time when the modification request was created. * effective-date - The time when the modification becomes effective. * modification-result.reserved-instances-id - The ID for the Reserved Instances created as part of the modification request. This ID is only available when the status of the modification is fulfilled . * modification-result.target-configuration.availability-zone - The Availability Zone for the new Reserved Instances. * modification-result.target-configuration.instance-count - The number of new Reserved Instances. * modification-result.target-configuration.instance-type - The instance type of the new Reserved Instances. * modification-result.target-configuration.platform - The network platform of the new Reserved Instances (EC2-Classic | EC2-VPC ). * reserved-instances-id - The ID of the Reserved Instances modified. * reserved-instances-modification-id - The ID of the modification request. * status - The status of the Reserved Instances modification request (processing | fulfilled | failed ). * status-message - The reason for the status. * update-date - The time when the modification request was last updated.

drimReservedInstancesModificationIds :: Lens' DescribeReservedInstancesModifications [Text]

IDs for the submitted modification request.

drimNextToken :: Lens' DescribeReservedInstancesModifications (Maybe Text)

The token to retrieve the next page of results.

Destructuring the Response

describeReservedInstancesModificationsResponse

Arguments

:: Int
drimrsResponseStatus

-> DescribeReservedInstancesModificationsResponse

Creates a value of DescribeReservedInstancesModificationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drimrsNextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

drimrsReservedInstancesModifications - The Reserved Instance modification information.

drimrsResponseStatus - -- | The response status code.

data DescribeReservedInstancesModificationsResponse

Contains the output of DescribeReservedInstancesModifications.

See: describeReservedInstancesModificationsResponse smart constructor.

Response Lenses

drimrsNextToken :: Lens' DescribeReservedInstancesModificationsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

drimrsReservedInstancesModifications :: Lens' DescribeReservedInstancesModificationsResponse [ReservedInstancesModification]

The Reserved Instance modification information.

drimrsResponseStatus :: Lens' DescribeReservedInstancesModificationsResponse Int

- | The response status code.