Network.AWS.EC2.DescribeReservedInstancesModifications

Description

Describes the modifications made to your Reserved Instances. If no parameter is specified, information about all your Reserved Instances modification requests is returned. If a modification ID is specified, only information about the specific modification is returned.

For more information, see Modifying Reserved Instances in the Amazon Elastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeReservedInstancesModifications.html

Request

data DescribeReservedInstancesModifications

Request constructor

describeReservedInstancesModifications :: DescribeReservedInstancesModifications

DescribeReservedInstancesModifications constructor.

The fields accessible through corresponding lenses are:

drimFilters :: [Filter]

drimNextToken :: Maybe Text

drimReservedInstancesModificationIds :: [Text]

Request lenses

drimFilters :: Lens' DescribeReservedInstancesModifications [Filter]

One or more filters.

'client-token' - The idempotency token for the modification request.

'create-date' - The time when the modification request was created.

'effective-date' - The time when the modification becomes effective.

'modification-result.reserved-instances-id' - The ID for the Reserved Instances created as part of the modification request. This ID is only available when the status of the modification is fulfilled.

'modification-result.target-configuration.availability-zone' - The Availability Zone for the new Reserved Instances.

'modification-result.target-configuration.instance-count ' - The number of new Reserved Instances.

'modification-result.target-configuration.instance-type' - The instance type of the new Reserved Instances.

'modification-result.target-configuration.platform' - The network platform of the new Reserved Instances ('EC2-Classic' | 'EC2-VPC').

'reserved-instances-id' - The ID of the Reserved Instances modified.

'reserved-instances-modification-id' - The ID of the modification request.

status - The status of the Reserved Instances modification request (processing | fulfilled | failed).

'status-message' - The reason for the status.

'update-date' - The time when the modification request was last updated.

drimNextToken :: Lens' DescribeReservedInstancesModifications (Maybe Text)

The token to retrieve the next page of results.

drimReservedInstancesModificationIds :: Lens' DescribeReservedInstancesModifications [Text]

IDs for the submitted modification request.

Response

data DescribeReservedInstancesModificationsResponse

Response constructor

describeReservedInstancesModificationsResponse :: DescribeReservedInstancesModificationsResponse

DescribeReservedInstancesModificationsResponse constructor.

The fields accessible through corresponding lenses are:

drimrNextToken :: Maybe Text

drimrReservedInstancesModifications :: [ReservedInstancesModification]

Response lenses

drimrNextToken :: Lens' DescribeReservedInstancesModificationsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

drimrReservedInstancesModifications :: Lens' DescribeReservedInstancesModificationsResponse [ReservedInstancesModification]

The Reserved Instance modification information.