Network.AWS.EC2.DescribeReservedInstancesOfferings

Description

Describes Reserved Instance offerings that are available for purchase. With Reserved Instances, you purchase the right to launch instances for a period of time. During that time period, you do not receive insufficient capacity errors, and you pay a lower usage rate than the rate charged for On-Demand instances for the actual time used.

If you have listed your own Reserved Instances for sale in the Reserved Instance Marketplace, they will be excluded from these results. This is to ensure that you do not purchase your own Reserved Instances.

For more information, see Reserved Instance Marketplace in the Amazon Elastic Compute Cloud User Guide .

This operation returns paginated results.

Creating a Request

describeReservedInstancesOfferings :: DescribeReservedInstancesOfferings

Creates a value of DescribeReservedInstancesOfferings with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drioMaxDuration - The maximum duration (in seconds) to filter when searching for offerings. Default: 94608000 (3 years)

drioProductDescription - The Reserved Instance product platform description. Instances that include (Amazon VPC) in the description are for use with Amazon VPC.

drioFilters - One or more filters. * availability-zone - The Availability Zone where the Reserved Instance can be used. * duration - The duration of the Reserved Instance (for example, one year or three years), in seconds (31536000 | 94608000 ). * fixed-price - The purchase price of the Reserved Instance (for example, 9800.0). * instance-type - The instance type that is covered by the reservation. * marketplace - Set to true to show only Reserved Instance Marketplace offerings. When this filter is not used, which is the default behavior, all offerings from both AWS and the Reserved Instance Marketplace are listed. * product-description - The Reserved Instance product platform description. Instances that include (Amazon VPC) in the product platform description will only be displayed to EC2-Classic account holders and are for use with Amazon VPC. (Linux/UNIX | Linux/UNIX (Amazon VPC) | SUSE Linux | SUSE Linux (Amazon VPC) | Red Hat Enterprise Linux | Red Hat Enterprise Linux (Amazon VPC) | Windows | Windows (Amazon VPC) | Windows with SQL Server Standard | Windows with SQL Server Standard (Amazon VPC) | Windows with SQL Server Web | Windows with SQL Server Web (Amazon VPC) | Windows with SQL Server Enterprise | Windows with SQL Server Enterprise (Amazon VPC) ) * reserved-instances-offering-id - The Reserved Instances offering ID. * scope - The scope of the Reserved Instance (Availability Zone or Region ). * usage-price - The usage price of the Reserved Instance, per hour (for example, 0.84).

drioIncludeMarketplace - Include Reserved Instance Marketplace offerings in the response.

drioInstanceType - The instance type that the reservation will cover (for example, m1.small ). For more information, see Instance Types in the Amazon Elastic Compute Cloud User Guide .

drioNextToken - The token to retrieve the next page of results.

drioMinDuration - The minimum duration (in seconds) to filter when searching for offerings. Default: 2592000 (1 month)

drioAvailabilityZone - The Availability Zone in which the Reserved Instance can be used.

drioOfferingType - The Reserved Instance offering type. If you are using tools that predate the 2011-11-01 API version, you only have access to the Medium Utilization Reserved Instance offering type.

drioReservedInstancesOfferingIds - One or more Reserved Instances offering IDs.

drioInstanceTenancy - The tenancy of the instances covered by the reservation. A Reserved Instance with a tenancy of dedicated is applied to instances that run in a VPC on single-tenant hardware (i.e., Dedicated Instances). Important: The host value cannot be used with this parameter. Use the default or dedicated values only. Default: default

drioOfferingClass - The offering class of the Reserved Instance. Can be standard or convertible .

drioMaxInstanceCount - The maximum number of instances to filter when searching for offerings. Default: 20

drioDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

drioMaxResults - The maximum number of results to return for the request in a single page. The remaining results of the initial request can be seen by sending another request with the returned NextToken value. The maximum is 100. Default: 100

data DescribeReservedInstancesOfferings

Contains the parameters for DescribeReservedInstancesOfferings.

See: describeReservedInstancesOfferings smart constructor.

Request Lenses

drioMaxDuration :: Lens' DescribeReservedInstancesOfferings (Maybe Integer)

The maximum duration (in seconds) to filter when searching for offerings. Default: 94608000 (3 years)

drioProductDescription :: Lens' DescribeReservedInstancesOfferings (Maybe RIProductDescription)

The Reserved Instance product platform description. Instances that include (Amazon VPC) in the description are for use with Amazon VPC.

drioFilters :: Lens' DescribeReservedInstancesOfferings [Filter]

One or more filters. * availability-zone - The Availability Zone where the Reserved Instance can be used. * duration - The duration of the Reserved Instance (for example, one year or three years), in seconds (31536000 | 94608000 ). * fixed-price - The purchase price of the Reserved Instance (for example, 9800.0). * instance-type - The instance type that is covered by the reservation. * marketplace - Set to true to show only Reserved Instance Marketplace offerings. When this filter is not used, which is the default behavior, all offerings from both AWS and the Reserved Instance Marketplace are listed. * product-description - The Reserved Instance product platform description. Instances that include (Amazon VPC) in the product platform description will only be displayed to EC2-Classic account holders and are for use with Amazon VPC. (Linux/UNIX | Linux/UNIX (Amazon VPC) | SUSE Linux | SUSE Linux (Amazon VPC) | Red Hat Enterprise Linux | Red Hat Enterprise Linux (Amazon VPC) | Windows | Windows (Amazon VPC) | Windows with SQL Server Standard | Windows with SQL Server Standard (Amazon VPC) | Windows with SQL Server Web | Windows with SQL Server Web (Amazon VPC) | Windows with SQL Server Enterprise | Windows with SQL Server Enterprise (Amazon VPC) ) * reserved-instances-offering-id - The Reserved Instances offering ID. * scope - The scope of the Reserved Instance (Availability Zone or Region ). * usage-price - The usage price of the Reserved Instance, per hour (for example, 0.84).

drioIncludeMarketplace :: Lens' DescribeReservedInstancesOfferings (Maybe Bool)

Include Reserved Instance Marketplace offerings in the response.

drioInstanceType :: Lens' DescribeReservedInstancesOfferings (Maybe InstanceType)

The instance type that the reservation will cover (for example, m1.small ). For more information, see Instance Types in the Amazon Elastic Compute Cloud User Guide .

drioNextToken :: Lens' DescribeReservedInstancesOfferings (Maybe Text)

The token to retrieve the next page of results.

drioMinDuration :: Lens' DescribeReservedInstancesOfferings (Maybe Integer)

The minimum duration (in seconds) to filter when searching for offerings. Default: 2592000 (1 month)

drioAvailabilityZone :: Lens' DescribeReservedInstancesOfferings (Maybe Text)

The Availability Zone in which the Reserved Instance can be used.

drioOfferingType :: Lens' DescribeReservedInstancesOfferings (Maybe OfferingTypeValues)

The Reserved Instance offering type. If you are using tools that predate the 2011-11-01 API version, you only have access to the Medium Utilization Reserved Instance offering type.

drioReservedInstancesOfferingIds :: Lens' DescribeReservedInstancesOfferings [Text]

One or more Reserved Instances offering IDs.

drioInstanceTenancy :: Lens' DescribeReservedInstancesOfferings (Maybe Tenancy)

The tenancy of the instances covered by the reservation. A Reserved Instance with a tenancy of dedicated is applied to instances that run in a VPC on single-tenant hardware (i.e., Dedicated Instances). Important: The host value cannot be used with this parameter. Use the default or dedicated values only. Default: default

drioOfferingClass :: Lens' DescribeReservedInstancesOfferings (Maybe OfferingClassType)

The offering class of the Reserved Instance. Can be standard or convertible .

drioMaxInstanceCount :: Lens' DescribeReservedInstancesOfferings (Maybe Int)

The maximum number of instances to filter when searching for offerings. Default: 20

drioDryRun :: Lens' DescribeReservedInstancesOfferings (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

drioMaxResults :: Lens' DescribeReservedInstancesOfferings (Maybe Int)

The maximum number of results to return for the request in a single page. The remaining results of the initial request can be seen by sending another request with the returned NextToken value. The maximum is 100. Default: 100

Destructuring the Response

describeReservedInstancesOfferingsResponse

Arguments

:: Int
driorsResponseStatus

-> DescribeReservedInstancesOfferingsResponse

Creates a value of DescribeReservedInstancesOfferingsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

driorsNextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return.

driorsReservedInstancesOfferings - A list of Reserved Instances offerings.

driorsResponseStatus - -- | The response status code.

data DescribeReservedInstancesOfferingsResponse

Contains the output of DescribeReservedInstancesOfferings.

See: describeReservedInstancesOfferingsResponse smart constructor.

Response Lenses

driorsNextToken :: Lens' DescribeReservedInstancesOfferingsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

driorsReservedInstancesOfferings :: Lens' DescribeReservedInstancesOfferingsResponse [ReservedInstancesOffering]

A list of Reserved Instances offerings.

driorsResponseStatus :: Lens' DescribeReservedInstancesOfferingsResponse Int

- | The response status code.