Network.AWS.EC2.DescribeRouteTables

Description

Describes one or more of your route tables.

For more information about route tables, see Route Tables in the AmazonVirtual Private Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeRouteTables.html

Request

data DescribeRouteTables

Request constructor

describeRouteTables :: DescribeRouteTables

DescribeRouteTables constructor.

The fields accessible through corresponding lenses are:

drt2DryRun :: Maybe Bool

drt2Filters :: [Filter]

drt2RouteTableIds :: [Text]

Request lenses

drt2DryRun :: Lens' DescribeRouteTables (Maybe Bool)

drt2Filters :: Lens' DescribeRouteTables [Filter]

One or more filters.

'association.route-table-association-id' - The ID of an association ID for the route table.

'association.route-table-id' - The ID of the route table involved in the association.

'association.subnet-id' - The ID of the subnet involved in the association.

'association.main' - Indicates whether the route table is the main route table for the VPC.

'route-table-id' - The ID of the route table.

'route.destination-cidr-block' - The CIDR range specified in a route in the table.

'route.gateway-id' - The ID of a gateway specified in a route in the table.

'route.instance-id' - The ID of an instance specified in a route in the table.

'route.origin' - Describes how the route was created. CreateRouteTable indicates that the route was automatically created when the route table was created; CreateRoute indicates that the route was manually added to the route table; EnableVgwRoutePropagation indicates that the route was propagated by route propagation.

'route.state' - The state of a route in the route table (active | blackhole). The blackhole state indicates that the route's target isn't available (for example, the specified gateway isn't attached to the VPC, the specified NAT instance has been terminated, and so on).

'route.vpc-peering-connection-id' - The ID of a VPC peering connection specified in a route in the table.

tag:key=value - The key/value combination of a tag assigned to the resource.

'tag-key' - The key of a tag assigned to the resource. This filter is independent of the 'tag-value' filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag:key=value filter.

'tag-value' - The value of a tag assigned to the resource. This filter is independent of the 'tag-key' filter.

'vpc-id' - The ID of the VPC for the route table.

drt2RouteTableIds :: Lens' DescribeRouteTables [Text]

One or more route table IDs.

Default: Describes all your route tables.

Response

data DescribeRouteTablesResponse

Response constructor

describeRouteTablesResponse :: DescribeRouteTablesResponse

DescribeRouteTablesResponse constructor.

The fields accessible through corresponding lenses are:

drtrRouteTables :: [RouteTable]

Response lenses

drtrRouteTables :: Lens' DescribeRouteTablesResponse [RouteTable]

Information about one or more route tables.