Network.AWS.EC2.DescribeSecurityGroups

Description

Describes one or more of your security groups.

A security group is for use with instances either in the EC2-Classic platform or in a specific VPC. For more information, see Amazon EC2 SecurityGroups in the Amazon Elastic Compute Cloud User Guide for Linux and SecurityGroups for Your VPC in the Amazon Virtual Private Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeSecurityGroups.html

Request

data DescribeSecurityGroups

Request constructor

describeSecurityGroups :: DescribeSecurityGroups

DescribeSecurityGroups constructor.

The fields accessible through corresponding lenses are:

dsg1DryRun :: Maybe Bool

dsg1Filters :: [Filter]

dsg1GroupIds :: [Text]

dsg1GroupNames :: [Text]

Request lenses

dsg1DryRun :: Lens' DescribeSecurityGroups (Maybe Bool)

dsg1Filters :: Lens' DescribeSecurityGroups [Filter]

One or more filters.

description - The description of the security group.

'group-id' - The ID of the security group.

'group-name' - The name of the security group.

'ip-permission.cidr' - A CIDR range that has been granted permission.

'ip-permission.from-port' - The start of port range for the TCP and UDP protocols, or an ICMP type number.

'ip-permission.group-id' - The ID of a security group that has been granted permission.

'ip-permission.group-name' - The name of a security group that has been granted permission.

'ip-permission.protocol' - The IP protocol for the permission (tcp | udp | icmp or a protocol number).

'ip-permission.to-port' - The end of port range for the TCP and UDP protocols, or an ICMP code.

'ip-permission.user-id' - The ID of an AWS account that has been granted permission.

'owner-id' - The AWS account ID of the owner of the security group.

'tag-key' - The key of a tag assigned to the security group.

'tag-value' - The value of a tag assigned to the security group.

'vpc-id' - The ID of the VPC specified when the security group was created.

dsg1GroupIds :: Lens' DescribeSecurityGroups [Text]

One or more security group IDs. Required for security groups in a nondefault VPC.

Default: Describes all your security groups.

dsg1GroupNames :: Lens' DescribeSecurityGroups [Text]

EC2-Classic and default VPC only

One or more security group names. You can specify either the security group name or the security group ID. For security groups in a nondefault VPC, use the 'group-name' filter to describe security groups by name.

Default: Describes all your security groups.

Response

data DescribeSecurityGroupsResponse

Response constructor

describeSecurityGroupsResponse :: DescribeSecurityGroupsResponse

DescribeSecurityGroupsResponse constructor.

The fields accessible through corresponding lenses are:

dsgrSecurityGroups :: [SecurityGroup]

Response lenses

dsgrSecurityGroups :: Lens' DescribeSecurityGroupsResponse [SecurityGroup]

Information about one or more security groups.