Network.AWS.EC2.DescribeSnapshotAttribute

Description

Describes the specified attribute of the specified snapshot. You can specify only one attribute at a time.

For more information about EBS snapshots, see Amazon EBS Snapshots in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

describeSnapshotAttribute

Arguments

:: SnapshotAttributeName
dsaAttribute

-> Text
dsaSnapshotId

-> DescribeSnapshotAttribute

Creates a value of DescribeSnapshotAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsaDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dsaAttribute - The snapshot attribute you would like to view.

dsaSnapshotId - The ID of the EBS snapshot.

data DescribeSnapshotAttribute

Contains the parameters for DescribeSnapshotAttribute.

See: describeSnapshotAttribute smart constructor.

Request Lenses

dsaDryRun :: Lens' DescribeSnapshotAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dsaAttribute :: Lens' DescribeSnapshotAttribute SnapshotAttributeName

The snapshot attribute you would like to view.

dsaSnapshotId :: Lens' DescribeSnapshotAttribute Text

The ID of the EBS snapshot.

Destructuring the Response

describeSnapshotAttributeResponse

Arguments

:: Int
dsarsResponseStatus

-> DescribeSnapshotAttributeResponse

Creates a value of DescribeSnapshotAttributeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsarsCreateVolumePermissions - A list of permissions for creating volumes from the snapshot.

dsarsProductCodes - A list of product codes.

dsarsSnapshotId - The ID of the EBS snapshot.

dsarsResponseStatus - -- | The response status code.

data DescribeSnapshotAttributeResponse

Contains the output of DescribeSnapshotAttribute.

See: describeSnapshotAttributeResponse smart constructor.

Response Lenses

dsarsCreateVolumePermissions :: Lens' DescribeSnapshotAttributeResponse [CreateVolumePermission]

A list of permissions for creating volumes from the snapshot.

dsarsProductCodes :: Lens' DescribeSnapshotAttributeResponse [ProductCode]

A list of product codes.

dsarsSnapshotId :: Lens' DescribeSnapshotAttributeResponse (Maybe Text)

The ID of the EBS snapshot.

dsarsResponseStatus :: Lens' DescribeSnapshotAttributeResponse Int

- | The response status code.