Network.AWS.EC2.DescribeSpotPriceHistory

Description

Describes the Spot price history. For more information, see Spot Instance Pricing History in the Amazon Elastic Compute Cloud User Guide .

When you specify a start and end time, this operation returns the prices of the instance types within the time range that you specified and the time when the price changed. The price is valid within the time period that you specified; the response merely indicates the last time that the price changed.

This operation returns paginated results.

Creating a Request

describeSpotPriceHistory :: DescribeSpotPriceHistory

Creates a value of DescribeSpotPriceHistory with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsphInstanceTypes - Filters the results by the specified instance types. Note that T2 and HS1 instance types are not supported.

dsphStartTime - The date and time, up to the past 90 days, from which to start retrieving the price history data, in UTC format (for example, YYYY -MM -DD THH :MM :SS Z).

dsphFilters - One or more filters. * availability-zone - The Availability Zone for which prices should be returned. * instance-type - The type of instance (for example, m3.medium ). * product-description - The product description for the Spot price (Linux/UNIX | SUSE Linux | Windows | Linux/UNIX (Amazon VPC) | SUSE Linux (Amazon VPC) | Windows (Amazon VPC) ). * spot-price - The Spot price. The value must match exactly (or use wildcards; greater than or less than comparison is not supported). * timestamp - The timestamp of the Spot price history, in UTC format (for example, YYYY -MM -DD THH :MM :SS Z). You can use wildcards (* and ?). Greater than or less than comparison is not supported.

dsphNextToken - The token for the next set of results.

dsphAvailabilityZone - Filters the results by the specified Availability Zone.

dsphEndTime - The date and time, up to the current date, from which to stop retrieving the price history data, in UTC format (for example, YYYY -MM -DD THH :MM :SS Z).

dsphProductDescriptions - Filters the results by the specified basic product descriptions.

dsphDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dsphMaxResults - The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

data DescribeSpotPriceHistory

Contains the parameters for DescribeSpotPriceHistory.

See: describeSpotPriceHistory smart constructor.

Request Lenses

dsphInstanceTypes :: Lens' DescribeSpotPriceHistory [InstanceType]

Filters the results by the specified instance types. Note that T2 and HS1 instance types are not supported.

dsphStartTime :: Lens' DescribeSpotPriceHistory (Maybe UTCTime)

The date and time, up to the past 90 days, from which to start retrieving the price history data, in UTC format (for example, YYYY -MM -DD THH :MM :SS Z).

dsphFilters :: Lens' DescribeSpotPriceHistory [Filter]

One or more filters. * availability-zone - The Availability Zone for which prices should be returned. * instance-type - The type of instance (for example, m3.medium ). * product-description - The product description for the Spot price (Linux/UNIX | SUSE Linux | Windows | Linux/UNIX (Amazon VPC) | SUSE Linux (Amazon VPC) | Windows (Amazon VPC) ). * spot-price - The Spot price. The value must match exactly (or use wildcards; greater than or less than comparison is not supported). * timestamp - The timestamp of the Spot price history, in UTC format (for example, YYYY -MM -DD THH :MM :SS Z). You can use wildcards (* and ?). Greater than or less than comparison is not supported.

dsphNextToken :: Lens' DescribeSpotPriceHistory (Maybe Text)

The token for the next set of results.

dsphAvailabilityZone :: Lens' DescribeSpotPriceHistory (Maybe Text)

Filters the results by the specified Availability Zone.

dsphEndTime :: Lens' DescribeSpotPriceHistory (Maybe UTCTime)

The date and time, up to the current date, from which to stop retrieving the price history data, in UTC format (for example, YYYY -MM -DD THH :MM :SS Z).

dsphProductDescriptions :: Lens' DescribeSpotPriceHistory [Text]

Filters the results by the specified basic product descriptions.

dsphDryRun :: Lens' DescribeSpotPriceHistory (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dsphMaxResults :: Lens' DescribeSpotPriceHistory (Maybe Int)

The maximum number of results to return in a single call. Specify a value between 1 and 1000. The default value is 1000. To retrieve the remaining results, make another call with the returned NextToken value.

Destructuring the Response

describeSpotPriceHistoryResponse

Arguments

:: Int
dsphrsResponseStatus

-> DescribeSpotPriceHistoryResponse

Creates a value of DescribeSpotPriceHistoryResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsphrsNextToken - The token required to retrieve the next set of results. This value is null when there are no more results to return.

dsphrsSpotPriceHistory - The historical Spot prices.

dsphrsResponseStatus - -- | The response status code.

data DescribeSpotPriceHistoryResponse

Contains the output of DescribeSpotPriceHistory.

See: describeSpotPriceHistoryResponse smart constructor.

Response Lenses

dsphrsNextToken :: Lens' DescribeSpotPriceHistoryResponse (Maybe Text)

The token required to retrieve the next set of results. This value is null when there are no more results to return.

dsphrsSpotPriceHistory :: Lens' DescribeSpotPriceHistoryResponse [SpotPrice]

The historical Spot prices.

dsphrsResponseStatus :: Lens' DescribeSpotPriceHistoryResponse Int

- | The response status code.