Network.AWS.EC2.DescribeSpotPriceHistory

Description

Describes the Spot Price history. The prices returned are listed in chronological order, from the oldest to the most recent, for up to the past 90 days. For more information, see Spot Instance Pricing History in the Amazon Elastic Compute Cloud User Guide for Linux.

When you specify a start and end time, this operation returns the prices of the instance types within the time range that you specified and the time when the price changed. The price is valid within the time period that you specified; the response merely indicates the last time that the price changed.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeSpotPriceHistory.html

Request

data DescribeSpotPriceHistory

Request constructor

describeSpotPriceHistory :: DescribeSpotPriceHistory

DescribeSpotPriceHistory constructor.

The fields accessible through corresponding lenses are:

dsphAvailabilityZone :: Maybe Text

dsphDryRun :: Maybe Bool

dsphEndTime :: Maybe UTCTime

dsphFilters :: [Filter]

dsphInstanceTypes :: [InstanceType]

dsphMaxResults :: Maybe Int

dsphNextToken :: Maybe Text

dsphProductDescriptions :: [Text]

dsphStartTime :: Maybe UTCTime

Request lenses

dsphAvailabilityZone :: Lens' DescribeSpotPriceHistory (Maybe Text)

Filters the results by the specified Availability Zone.

dsphDryRun :: Lens' DescribeSpotPriceHistory (Maybe Bool)

dsphEndTime :: Lens' DescribeSpotPriceHistory (Maybe UTCTime)

The date and time, up to the current date, from which to stop retrieving the price history data.

dsphFilters :: Lens' DescribeSpotPriceHistory [Filter]

One or more filters.

'availability-zone' - The Availability Zone for which prices should be returned.

'instance-type' - The type of instance (for example, 'm1.small').

'product-description' - The product description for the Spot Price ('Linux/UNIX' | 'SUSE Linux' | Windows | 'Linux/UNIX (Amazon VPC)' | 'SUSE Linux (Amazon VPC)' | 'Windows (Amazon VPC)').

'spot-price' - The Spot Price. The value must match exactly (or use wildcards; greater than or less than comparison is not supported).

timestamp - The timestamp of the Spot Price history (for example, 2010-08-16T05:06:11.000Z). You can use wildcards (* and ?). Greater than or less than comparison is not supported.

dsphInstanceTypes :: Lens' DescribeSpotPriceHistory [InstanceType]

Filters the results by the specified instance types.

dsphMaxResults :: Lens' DescribeSpotPriceHistory (Maybe Int)

The maximum number of results to return for the request in a single page. The remaining results of the initial request can be seen by sending another request with the returned NextToken value. This value can be between 5 and 1000; if MaxResults is given a value larger than 1000, only 1000 results are returned.

dsphNextToken :: Lens' DescribeSpotPriceHistory (Maybe Text)

The token to retrieve the next page of results.

dsphProductDescriptions :: Lens' DescribeSpotPriceHistory [Text]

Filters the results by the specified basic product descriptions.

dsphStartTime :: Lens' DescribeSpotPriceHistory (Maybe UTCTime)

The date and time, up to the past 90 days, from which to start retrieving the price history data.

Response

data DescribeSpotPriceHistoryResponse

Response constructor

describeSpotPriceHistoryResponse :: DescribeSpotPriceHistoryResponse

DescribeSpotPriceHistoryResponse constructor.

The fields accessible through corresponding lenses are:

dsphrNextToken :: Maybe Text

dsphrSpotPriceHistory :: [SpotPrice]

Response lenses

dsphrNextToken :: Lens' DescribeSpotPriceHistoryResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return.

dsphrSpotPriceHistory :: Lens' DescribeSpotPriceHistoryResponse [SpotPrice]

The historical Spot Prices.