Network.AWS.EC2.DescribeStaleSecurityGroups

Description

EC2-VPC only

Describes the stale security group rules for security groups in a specified VPC. Rules are stale when they reference a deleted security group in a peer VPC, or a security group in a peer VPC for which the VPC peering connection has been deleted.

Creating a Request

describeStaleSecurityGroups

Arguments

:: Text
dssgVPCId

-> DescribeStaleSecurityGroups

Creates a value of DescribeStaleSecurityGroups with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dssgNextToken - The token for the next set of items to return. (You received this token from a prior call.)

dssgDryRun - Checks whether you have the required permissions for the operation, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

dssgMaxResults - The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results.

dssgVPCId - The ID of the VPC.

data DescribeStaleSecurityGroups

See: describeStaleSecurityGroups smart constructor.

Request Lenses

dssgNextToken :: Lens' DescribeStaleSecurityGroups (Maybe Text)

The token for the next set of items to return. (You received this token from a prior call.)

dssgDryRun :: Lens' DescribeStaleSecurityGroups (Maybe Bool)

Checks whether you have the required permissions for the operation, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

dssgMaxResults :: Lens' DescribeStaleSecurityGroups (Maybe Natural)

The maximum number of items to return for this request. The request returns a token that you can specify in a subsequent call to get the next set of results.

dssgVPCId :: Lens' DescribeStaleSecurityGroups Text

The ID of the VPC.

Destructuring the Response

describeStaleSecurityGroupsResponse

Arguments

:: Int
dssgrsResponseStatus

-> DescribeStaleSecurityGroupsResponse

Creates a value of DescribeStaleSecurityGroupsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dssgrsStaleSecurityGroupSet - Information about the stale security groups.

dssgrsNextToken - The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dssgrsResponseStatus - -- | The response status code.

data DescribeStaleSecurityGroupsResponse

See: describeStaleSecurityGroupsResponse smart constructor.

Response Lenses

dssgrsStaleSecurityGroupSet :: Lens' DescribeStaleSecurityGroupsResponse [StaleSecurityGroup]

Information about the stale security groups.

dssgrsNextToken :: Lens' DescribeStaleSecurityGroupsResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.

dssgrsResponseStatus :: Lens' DescribeStaleSecurityGroupsResponse Int

- | The response status code.