Network.AWS.EC2.DescribeTags

Description

Describes one or more of the tags for your EC2 resources.

For more information about tags, see Tagging Your Resources in the Amazon Elastic Compute Cloud User Guide .

This operation returns paginated results.

Creating a Request

describeTags :: DescribeTags

Creates a value of DescribeTags with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtFilters - One or more filters. * key - The tag key. * resource-id - The resource ID. * resource-type - The resource type (customer-gateway | dhcp-options | image | instance | internet-gateway | network-acl | network-interface | reserved-instances | route-table | security-group | snapshot | spot-instances-request | subnet | volume | vpc | vpn-connection | vpn-gateway ). * value - The tag value.

dtNextToken - The token to retrieve the next page of results.

dtDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dtMaxResults - The maximum number of results to return in a single call. This value can be between 5 and 1000. To retrieve the remaining results, make another call with the returned NextToken value.

data DescribeTags

Contains the parameters for DescribeTags.

See: describeTags smart constructor.

Request Lenses

dtFilters :: Lens' DescribeTags [Filter]

One or more filters. * key - The tag key. * resource-id - The resource ID. * resource-type - The resource type (customer-gateway | dhcp-options | image | instance | internet-gateway | network-acl | network-interface | reserved-instances | route-table | security-group | snapshot | spot-instances-request | subnet | volume | vpc | vpn-connection | vpn-gateway ). * value - The tag value.

dtNextToken :: Lens' DescribeTags (Maybe Text)

The token to retrieve the next page of results.

dtDryRun :: Lens' DescribeTags (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dtMaxResults :: Lens' DescribeTags (Maybe Int)

The maximum number of results to return in a single call. This value can be between 5 and 1000. To retrieve the remaining results, make another call with the returned NextToken value.

Destructuring the Response

describeTagsResponse

Arguments

:: Int
dtrsResponseStatus

-> DescribeTagsResponse

Creates a value of DescribeTagsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtrsNextToken - The token to use to retrieve the next page of results. This value is null when there are no more results to return..

dtrsTags - A list of tags.

dtrsResponseStatus - -- | The response status code.

data DescribeTagsResponse

Contains the output of DescribeTags.

See: describeTagsResponse smart constructor.

Response Lenses

dtrsNextToken :: Lens' DescribeTagsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return..

dtrsTags :: Lens' DescribeTagsResponse [TagDescription]

A list of tags.

dtrsResponseStatus :: Lens' DescribeTagsResponse Int

- | The response status code.