Network.AWS.EC2.DescribeTags

Description

Describes one or more of the tags for your EC2 resources. For more information about tags, see Tagging Your Resources in the Amazon Elastic Compute Cloud User Guide. Example This example describes all the tags in your account. https:/ec2.amazonaws.com?Action=DescribeTags &amp;AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE ami-1a2b3c4d image webserver ami-1a2b3c4d image stack Production i-5f4e3d2a instance webserver i-5f4e3d2a instance stack Production i-12345678 instance database_server i-12345678 instance stack Test Example This example describes only the tags for the AMI with ID ami-1a2b3c4d. https:/ec2.amazonaws.com?Action=DescribeTags &amp;Filter.1.Name=resource-id &amp;Filter.1.Value.1=ami-1a2b3c4d &amp;AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE ami-1a2b3c4d image webserver ami-1a2b3c4d image stack Production Example This example describes the tags for all your instances. https:/ec2.amazonaws.com?Action=DescribeTags &amp;Filter.1.Name=resource-type &amp;Filter.1.Value.1=instance &amp;AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE i-5f4e3d2a instance webserver i-5f4e3d2a instance stack Production i-12345678 instance database_server i-12345678 instance stack Test Example This example describes the tags for all your instances tagged with the key webserver. Note that you can use wildcards with filters, so you could specify the value as ?ebserver to find tags with the key webserver or Webserver. https:/ec2.amazonaws.com?Action=DescribeTags &amp;Filter.1.Name=key &amp;Filter.1.Value.1=webserver &amp;AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE i-5f4e3d2a instance webserver Example This example describes the tags for all your instances tagged with either stack=Test or stack=Production. https:/ec2.amazonaws.com?Action=DescribeTags &amp;Filter.1.Name=resource-type &amp;Filter.1.Value.1=instance &amp;Filter.2.Name=key &amp;Filter.2.Value.1=stack &amp;Filter.3.Name=value &amp;Filter.3.Value.1=Test &amp;Filter.3.Value.2=Production &amp;AUTHPARAMS 7a62c49f-347e-4fc4-9331-6e8eEXAMPLE i-5f4e3d2a instance stack Production i-12345678 instance stack Test Example This example describes the tags for all your instances tagged with Purpose=[empty string]. https:/ec2.amazonaws.com?Action=DescribeTags &amp;Filter.1.Name=resource-type &amp;Filter.1.Value.1=instance &amp;Filter.2.Name=key &amp;Filter.2.Value.1=Purpose &amp;Filter.3.Name=value &amp;Filter.3.Value.1= &amp;AUTHPARAMS.

Request

data DescribeTags

Request constructor

describeTags :: DescribeTags

Smart constructor for the minimum required parameters to construct a valid DescribeTags request.

The fields accessible through corresponding lenses are:

Filters :: [Filter]

MaxResults :: Maybe Integer

NextToken :: Maybe Text

Request lenses

dt1Filters :: Lens' DescribeTags [Filter]

One or more filters. key - The tag key. resource-id - The resource ID. resource-type - The resource type (customer-gateway | dhcp-options | image | instance | internet-gateway | network-acl | network-interface | reserved-instances | route-table | security-group | snapshot | spot-instances-request | subnet | volume | vpc | vpn-connection | vpn-gateway). value - The tag value.

dt1MaxResults :: Lens' DescribeTags (Maybe Integer)

The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results. If the value is greater than 1000, we return only 1000 items.

dt1NextToken :: Lens' DescribeTags (Maybe Text)

The token for the next set of items to return. (You received this token from a prior call.).

Response

data DescribeTagsResponse

Response constructor

describeTagsResponse :: DescribeTagsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeTagsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Tags :: [TagDescription]

NextToken :: Maybe Text

Response lenses

dtrTags :: Lens' DescribeTagsResponse [TagDescription]

A list of tags.

dtrNextToken :: Lens' DescribeTagsResponse (Maybe Text)

The token to use when requesting the next set of items. If there are no additional items to return, the string is empty.