Network.AWS.EC2.DescribeTags

Description

Describes one or more of the tags for your EC2 resources.

For more information about tags, see Tagging Your Resources in the AmazonElastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeTags.html

Request

data DescribeTags

Request constructor

describeTags :: DescribeTags

DescribeTags constructor.

The fields accessible through corresponding lenses are:

dtDryRun :: Maybe Bool

dtFilters :: [Filter]

dtMaxResults :: Maybe Int

dtNextToken :: Maybe Text

Request lenses

dtDryRun :: Lens' DescribeTags (Maybe Bool)

dtFilters :: Lens' DescribeTags [Filter]

One or more filters.

key - The tag key.

'resource-id' - The resource ID.

'resource-type' - The resource type ('customer-gateway' | 'dhcp-options' | image | 'instance' | 'internet-gateway' | 'network-acl' | 'network-interface' | 'reserved-instances' | 'route-table' | 'security-group' | snapshot | 'spot-instances-request' | subnet | volume | vpc | 'vpn-connection' | 'vpn-gateway').

value - The tag value.

dtMaxResults :: Lens' DescribeTags (Maybe Int)

The maximum number of results to return for the request in a single page. The remaining results of the initial request can be seen by sending another request with the returned NextToken value. This value can be between 5 and 1000; if MaxResults is given a value larger than 1000, only 1000 results are returned.

dtNextToken :: Lens' DescribeTags (Maybe Text)

The token to retrieve the next page of results.

Response

data DescribeTagsResponse

Response constructor

describeTagsResponse :: DescribeTagsResponse

DescribeTagsResponse constructor.

The fields accessible through corresponding lenses are:

dtrNextToken :: Maybe Text

dtrTags :: [TagDescription]

Response lenses

dtrNextToken :: Lens' DescribeTagsResponse (Maybe Text)

The token to use to retrieve the next page of results. This value is null when there are no more results to return..

dtrTags :: Lens' DescribeTagsResponse [TagDescription]

A list of tags.