Network.AWS.EC2.DescribeVolumeAttribute

Description

Describes the specified attribute of the specified volume. You can specify only one attribute at a time.

For more information about EBS volumes, see Amazon EBS Volumes in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

describeVolumeAttribute

Arguments

:: Text
dvaVolumeId

-> DescribeVolumeAttribute

Creates a value of DescribeVolumeAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvaAttribute - The attribute of the volume. This parameter is required.

dvaDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dvaVolumeId - The ID of the volume.

data DescribeVolumeAttribute

Contains the parameters for DescribeVolumeAttribute.

See: describeVolumeAttribute smart constructor.

Request Lenses

dvaAttribute :: Lens' DescribeVolumeAttribute (Maybe VolumeAttributeName)

The attribute of the volume. This parameter is required.

dvaDryRun :: Lens' DescribeVolumeAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dvaVolumeId :: Lens' DescribeVolumeAttribute Text

The ID of the volume.

Destructuring the Response

describeVolumeAttributeResponse

Arguments

:: Int
dvarsResponseStatus

-> DescribeVolumeAttributeResponse

Creates a value of DescribeVolumeAttributeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvarsVolumeId - The ID of the volume.

dvarsProductCodes - A list of product codes.

dvarsAutoEnableIO - The state of autoEnableIO attribute.

dvarsResponseStatus - -- | The response status code.

data DescribeVolumeAttributeResponse

Contains the output of DescribeVolumeAttribute.

See: describeVolumeAttributeResponse smart constructor.

Response Lenses

dvarsVolumeId :: Lens' DescribeVolumeAttributeResponse (Maybe Text)

The ID of the volume.

dvarsProductCodes :: Lens' DescribeVolumeAttributeResponse [ProductCode]

A list of product codes.

dvarsAutoEnableIO :: Lens' DescribeVolumeAttributeResponse (Maybe AttributeBooleanValue)

The state of autoEnableIO attribute.

dvarsResponseStatus :: Lens' DescribeVolumeAttributeResponse Int

- | The response status code.