Network.AWS.EC2.DescribeVolumes

Description

Describes the specified EBS volumes.

If you are describing a long list of volumes, you can paginate the output to make the list more manageable. The MaxResults parameter sets the maximum number of results returned in a single page. If the list of results exceeds your MaxResults value, then that number of results is returned along with a NextToken value that can be passed to a subsequent DescribeVolumes request to retrieve the remaining results.

For more information about EBS volumes, see Amazon EBS Volumes in the Amazon Elastic Compute Cloud User Guide .

This operation returns paginated results.

Creating a Request

describeVolumes :: DescribeVolumes

Creates a value of DescribeVolumes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

desFilters - One or more filters. * attachment.attach-time - The time stamp when the attachment initiated. * attachment.delete-on-termination - Whether the volume is deleted on instance termination. * attachment.device - The device name that is exposed to the instance (for example, devsda1 ). * attachment.instance-id - The ID of the instance the volume is attached to. * attachment.status - The attachment state (attaching | attached | detaching | detached ). * availability-zone - The Availability Zone in which the volume was created. * create-time - The time stamp when the volume was created. * encrypted - The encryption status of the volume. * size - The size of the volume, in GiB. * snapshot-id - The snapshot from which the volume was created. * status - The status of the volume (creating | available | in-use | deleting | deleted | error ). * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter. * volume-id - The volume ID. * volume-type - The Amazon EBS volume type. This can be gp2 for General Purpose SSD, io1 for Provisioned IOPS SSD, st1 for Throughput Optimized HDD, sc1 for Cold HDD, or standard for Magnetic volumes.

desVolumeIds - One or more volume IDs.

desNextToken - The NextToken value returned from a previous paginated DescribeVolumes request where MaxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the NextToken value. This value is null when there are no more results to return.

desDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

desMaxResults - The maximum number of volume results returned by DescribeVolumes in paginated output. When this parameter is used, DescribeVolumes only returns MaxResults results in a single page along with a NextToken response element. The remaining results of the initial request can be seen by sending another DescribeVolumes request with the returned NextToken value. This value can be between 5 and 500; if MaxResults is given a value larger than 500, only 500 results are returned. If this parameter is not used, then DescribeVolumes returns all results. You cannot specify this parameter and the volume IDs parameter in the same request.

data DescribeVolumes

Contains the parameters for DescribeVolumes.

See: describeVolumes smart constructor.

Request Lenses

desFilters :: Lens' DescribeVolumes [Filter]

One or more filters. * attachment.attach-time - The time stamp when the attachment initiated. * attachment.delete-on-termination - Whether the volume is deleted on instance termination. * attachment.device - The device name that is exposed to the instance (for example, devsda1 ). * attachment.instance-id - The ID of the instance the volume is attached to. * attachment.status - The attachment state (attaching | attached | detaching | detached ). * availability-zone - The Availability Zone in which the volume was created. * create-time - The time stamp when the volume was created. * encrypted - The encryption status of the volume. * size - The size of the volume, in GiB. * snapshot-id - The snapshot from which the volume was created. * status - The status of the volume (creating | available | in-use | deleting | deleted | error ). * tag :key =value - The keyvalue combination of a tag assigned to the resource. Specify the key of the tag in the filter name and the value of the tag in the filter value. For example, for the tag Purpose=X, specify tag:Purpose for the filter name and X for the filter value. * tag-key - The key of a tag assigned to the resource. This filter is independent of the tag-value filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag :key =value/ filter. * tag-value - The value of a tag assigned to the resource. This filter is independent of the tag-key filter. * volume-id - The volume ID. * volume-type - The Amazon EBS volume type. This can be gp2 for General Purpose SSD, io1 for Provisioned IOPS SSD, st1 for Throughput Optimized HDD, sc1 for Cold HDD, or standard for Magnetic volumes.

desVolumeIds :: Lens' DescribeVolumes [Text]

One or more volume IDs.

desNextToken :: Lens' DescribeVolumes (Maybe Text)

The NextToken value returned from a previous paginated DescribeVolumes request where MaxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the NextToken value. This value is null when there are no more results to return.

desDryRun :: Lens' DescribeVolumes (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

desMaxResults :: Lens' DescribeVolumes (Maybe Int)

The maximum number of volume results returned by DescribeVolumes in paginated output. When this parameter is used, DescribeVolumes only returns MaxResults results in a single page along with a NextToken response element. The remaining results of the initial request can be seen by sending another DescribeVolumes request with the returned NextToken value. This value can be between 5 and 500; if MaxResults is given a value larger than 500, only 500 results are returned. If this parameter is not used, then DescribeVolumes returns all results. You cannot specify this parameter and the volume IDs parameter in the same request.

Destructuring the Response

describeVolumesResponse

Arguments

:: Int
dvvrsResponseStatus

-> DescribeVolumesResponse

Creates a value of DescribeVolumesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvvrsNextToken - The NextToken value to include in a future DescribeVolumes request. When the results of a DescribeVolumes request exceed MaxResults , this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

dvvrsVolumes - Information about the volumes.

dvvrsResponseStatus - -- | The response status code.

data DescribeVolumesResponse

Contains the output of DescribeVolumes.

See: describeVolumesResponse smart constructor.

Response Lenses

dvvrsNextToken :: Lens' DescribeVolumesResponse (Maybe Text)

The NextToken value to include in a future DescribeVolumes request. When the results of a DescribeVolumes request exceed MaxResults , this value can be used to retrieve the next page of results. This value is null when there are no more results to return.

dvvrsVolumes :: Lens' DescribeVolumesResponse [Volume]

Information about the volumes.

dvvrsResponseStatus :: Lens' DescribeVolumesResponse Int

- | The response status code.