Network.AWS.EC2.DescribeVolumesModifications

Description

Reports the current modification status of EBS volumes.

Current-generation EBS volumes support modification of attributes including type, size, and (for io1 volumes) IOPS provisioning while either attached to or detached from an instance. Following an action from the API or the console to modify a volume, the status of the modification may be modifying , optimizing , completed , or failed . If a volume has never been modified, then certain elements of the returned VolumeModification objects are null.

You can also use CloudWatch Events to check the status of a modification to an EBS volume. For information about CloudWatch Events, see the Amazon CloudWatch Events User Guide . For more information, see Monitoring Volume Modifications" .

Creating a Request

describeVolumesModifications :: DescribeVolumesModifications

Creates a value of DescribeVolumesModifications with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvmFilters - One or more filters. Supported filters: volume-id , modification-state , target-size , target-iops , target-volume-type , original-size , original-iops , original-volume-type , start-time .

dvmVolumeIds - One or more volume IDs for which in-progress modifications will be described.

dvmNextToken - The nextToken value returned by a previous paginated request.

dvmDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dvmMaxResults - The maximum number of results (up to a limit of 500) to be returned in a paginated request.

data DescribeVolumesModifications

See: describeVolumesModifications smart constructor.

Request Lenses

dvmFilters :: Lens' DescribeVolumesModifications [Filter]

One or more filters. Supported filters: volume-id , modification-state , target-size , target-iops , target-volume-type , original-size , original-iops , original-volume-type , start-time .

dvmVolumeIds :: Lens' DescribeVolumesModifications [Text]

One or more volume IDs for which in-progress modifications will be described.

dvmNextToken :: Lens' DescribeVolumesModifications (Maybe Text)

The nextToken value returned by a previous paginated request.

dvmDryRun :: Lens' DescribeVolumesModifications (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dvmMaxResults :: Lens' DescribeVolumesModifications (Maybe Int)

The maximum number of results (up to a limit of 500) to be returned in a paginated request.

Destructuring the Response

describeVolumesModificationsResponse

Arguments

:: Int
dvmrsResponseStatus

-> DescribeVolumesModificationsResponse

Creates a value of DescribeVolumesModificationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvmrsVolumesModifications - A list of returned VolumeModification objects.

dvmrsNextToken - Token for pagination, null if there are no more results

dvmrsResponseStatus - -- | The response status code.

data DescribeVolumesModificationsResponse

See: describeVolumesModificationsResponse smart constructor.

Response Lenses

dvmrsVolumesModifications :: Lens' DescribeVolumesModificationsResponse [VolumeModification]

A list of returned VolumeModification objects.

dvmrsNextToken :: Lens' DescribeVolumesModificationsResponse (Maybe Text)

Token for pagination, null if there are no more results

dvmrsResponseStatus :: Lens' DescribeVolumesModificationsResponse Int

- | The response status code.