Network.AWS.EC2.DescribeVpcPeeringConnections

Description

Describes one or more of your VPC peering connections.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeVpcPeeringConnections.html

Request

data DescribeVpcPeeringConnections

Request constructor

describeVpcPeeringConnections :: DescribeVpcPeeringConnections

DescribeVpcPeeringConnections constructor.

The fields accessible through corresponding lenses are:

dvpc1DryRun :: Maybe Bool

dvpc1Filters :: [Filter]

dvpc1VpcPeeringConnectionIds :: [Text]

Request lenses

dvpc1DryRun :: Lens' DescribeVpcPeeringConnections (Maybe Bool)

dvpc1Filters :: Lens' DescribeVpcPeeringConnections [Filter]

One or more filters.

'accepter-vpc-info.cidr-block' - The CIDR block of the peer VPC.

'accepter-vpc-info.owner-id' - The AWS account ID of the owner of the peer VPC.

'accepter-vpc-info.vpc-id' - The ID of the peer VPC.

'expiration-time' - The expiration date and time for the VPC peering connection.

'requester-vpc-info.cidr-block' - The CIDR block of the requester's VPC.

'requester-vpc-info.owner-id' - The AWS account ID of the owner of the requester VPC.

'requester-vpc-info.vpc-id' - The ID of the requester VPC.

'status-code' - The status of the VPC peering connection ('pending-acceptance' | failed | expired | provisioning | active | deleted | rejected).

'status-message' - A message that provides more information about the status of the VPC peering connection, if applicable.

tag:key=value - The key/value combination of a tag assigned to the resource.

'tag-key' - The key of a tag assigned to the resource. This filter is independent of the 'tag-value' filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag:key=value filter.

'tag-value' - The value of a tag assigned to the resource. This filter is independent of the 'tag-key' filter.

'vpc-peering-connection-id' - The ID of the VPC peering connection.

dvpc1VpcPeeringConnectionIds :: Lens' DescribeVpcPeeringConnections [Text]

One or more VPC peering connection IDs.

Default: Describes all your VPC peering connections.

Response

data DescribeVpcPeeringConnectionsResponse

Response constructor

describeVpcPeeringConnectionsResponse :: DescribeVpcPeeringConnectionsResponse

DescribeVpcPeeringConnectionsResponse constructor.

The fields accessible through corresponding lenses are:

dvpcrVpcPeeringConnections :: [VpcPeeringConnection]

Response lenses

dvpcrVpcPeeringConnections :: Lens' DescribeVpcPeeringConnectionsResponse [VpcPeeringConnection]

Information about the VPC peering connections.