Network.AWS.EC2.DescribeVpnGateways

Description

Describes one or more of your virtual private gateways.

For more information about virtual private gateways, see Adding an IPsecHardware VPN to Your VPC in the Amazon Virtual Private Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeVpnGateways.html

Request

data DescribeVpnGateways

Request constructor

describeVpnGateways :: DescribeVpnGateways

DescribeVpnGateways constructor.

The fields accessible through corresponding lenses are:

dvg2DryRun :: Maybe Bool

dvg2Filters :: [Filter]

dvg2VpnGatewayIds :: [Text]

Request lenses

dvg2DryRun :: Lens' DescribeVpnGateways (Maybe Bool)

dvg2Filters :: Lens' DescribeVpnGateways [Filter]

One or more filters.

'attachment.state' - The current state of the attachment between the gateway and the VPC (attaching | attached | detaching | detached).

'attachment.vpc-id' - The ID of an attached VPC.

'availability-zone' - The Availability Zone for the virtual private gateway.

state - The state of the virtual private gateway (pending | available | deleting | deleted).

tag:key=value - The key/value combination of a tag assigned to the resource.

'tag-key' - The key of a tag assigned to the resource. This filter is independent of the 'tag-value' filter. For example, if you use both the filter "tag-key=Purpose" and the filter "tag-value=X", you get any resources assigned both the tag key Purpose (regardless of what the tag's value is), and the tag value X (regardless of what the tag's key is). If you want to list only resources where Purpose is X, see the tag:key=value filter.

'tag-value' - The value of a tag assigned to the resource. This filter is independent of the 'tag-key' filter.

'type' - The type of virtual private gateway. Currently the only supported type is 'ipsec.1'.

'vpn-gateway-id' - The ID of the virtual private gateway.

dvg2VpnGatewayIds :: Lens' DescribeVpnGateways [Text]

One or more virtual private gateway IDs.

Default: Describes all your virtual private gateways.

Response

data DescribeVpnGatewaysResponse

Response constructor

describeVpnGatewaysResponse :: DescribeVpnGatewaysResponse

DescribeVpnGatewaysResponse constructor.

The fields accessible through corresponding lenses are:

dvgrVpnGateways :: [VpnGateway]

Response lenses

dvgrVpnGateways :: Lens' DescribeVpnGatewaysResponse [VpnGateway]

Information about one or more virtual private gateways.