Network.AWS.EC2.DetachInternetGateway

Description

Detaches an Internet gateway from a VPC, disabling connectivity between the Internet and the VPC. The VPC must not contain any running instances with Elastic IP addresses or public IPv4 addresses.

Creating a Request

detachInternetGateway

Arguments

:: Text
digInternetGatewayId

-> Text
digVPCId

-> DetachInternetGateway

Creates a value of DetachInternetGateway with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

digDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

digInternetGatewayId - The ID of the Internet gateway.

digVPCId - The ID of the VPC.

data DetachInternetGateway

Contains the parameters for DetachInternetGateway.

See: detachInternetGateway smart constructor.

Request Lenses

digDryRun :: Lens' DetachInternetGateway (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

digInternetGatewayId :: Lens' DetachInternetGateway Text

The ID of the Internet gateway.

digVPCId :: Lens' DetachInternetGateway Text

The ID of the VPC.

Destructuring the Response

detachInternetGatewayResponse :: DetachInternetGatewayResponse

Creates a value of DetachInternetGatewayResponse with the minimum fields required to make a request.

data DetachInternetGatewayResponse

See: detachInternetGatewayResponse smart constructor.