Network.AWS.EC2.DetachNetworkInterface

Description

Detaches a network interface from an instance.

Creating a Request

detachNetworkInterface

Arguments

:: Text
dniAttachmentId

-> DetachNetworkInterface

Creates a value of DetachNetworkInterface with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dniForce - Specifies whether to force a detachment.

dniDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dniAttachmentId - The ID of the attachment.

data DetachNetworkInterface

Contains the parameters for DetachNetworkInterface.

See: detachNetworkInterface smart constructor.

Request Lenses

dniForce :: Lens' DetachNetworkInterface (Maybe Bool)

Specifies whether to force a detachment.

dniDryRun :: Lens' DetachNetworkInterface (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dniAttachmentId :: Lens' DetachNetworkInterface Text

The ID of the attachment.

Destructuring the Response

detachNetworkInterfaceResponse :: DetachNetworkInterfaceResponse

Creates a value of DetachNetworkInterfaceResponse with the minimum fields required to make a request.

data DetachNetworkInterfaceResponse

See: detachNetworkInterfaceResponse smart constructor.