Network.AWS.EC2.DetachVolume

Description

Detaches an EBS volume from an instance. Make sure to unmount any file systems on the device within your operating system before detaching the volume. Failure to do so can result in the volume becoming stuck in the busy state while detaching. If this happens, detachment can be delayed indefinitely until you unmount the volume, force detachment, reboot the instance, or all three. If an EBS volume is the root device of an instance, it can't be detached while the instance is running. To detach the root volume, stop the instance first.

When a volume with an AWS Marketplace product code is detached from an instance, the product code is no longer associated with the instance.

For more information, see Detaching an Amazon EBS Volume in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

detachVolume

Arguments

:: Text
dvVolumeId

-> DetachVolume

Creates a value of DetachVolume with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvInstanceId - The ID of the instance.

dvForce - Forces detachment if the previous detachment attempt did not occur cleanly (for example, logging into an instance, unmounting the volume, and detaching normally). This option can lead to data loss or a corrupted file system. Use this option only as a last resort to detach a volume from a failed instance. The instance won't have an opportunity to flush file system caches or file system metadata. If you use this option, you must perform file system check and repair procedures.

dvDevice - The device name.

dvDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dvVolumeId - The ID of the volume.

data DetachVolume

Contains the parameters for DetachVolume.

See: detachVolume smart constructor.

Request Lenses

dvInstanceId :: Lens' DetachVolume (Maybe Text)

The ID of the instance.

dvForce :: Lens' DetachVolume (Maybe Bool)

Forces detachment if the previous detachment attempt did not occur cleanly (for example, logging into an instance, unmounting the volume, and detaching normally). This option can lead to data loss or a corrupted file system. Use this option only as a last resort to detach a volume from a failed instance. The instance won't have an opportunity to flush file system caches or file system metadata. If you use this option, you must perform file system check and repair procedures.

dvDevice :: Lens' DetachVolume (Maybe Text)

The device name.

dvDryRun :: Lens' DetachVolume (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

dvVolumeId :: Lens' DetachVolume Text

The ID of the volume.

Destructuring the Response

volumeAttachment :: VolumeAttachment

Creates a value of VolumeAttachment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

volInstanceId - The ID of the instance.

volDeleteOnTermination - Indicates whether the EBS volume is deleted on instance termination.

volState - The attachment state of the volume.

volDevice - The device name.

volVolumeId - The ID of the volume.

volAttachTime - The time stamp when the attachment initiated.

data VolumeAttachment

Describes volume attachment details.

See: volumeAttachment smart constructor.

Response Lenses

volInstanceId :: Lens' VolumeAttachment (Maybe Text)

The ID of the instance.

volDeleteOnTermination :: Lens' VolumeAttachment (Maybe Bool)

Indicates whether the EBS volume is deleted on instance termination.

volState :: Lens' VolumeAttachment (Maybe VolumeAttachmentState)

The attachment state of the volume.

volDevice :: Lens' VolumeAttachment (Maybe Text)

The device name.

volVolumeId :: Lens' VolumeAttachment (Maybe Text)

The ID of the volume.

volAttachTime :: Lens' VolumeAttachment (Maybe UTCTime)

The time stamp when the attachment initiated.