Network.AWS.EC2.DetachVpnGateway

Description

Detaches a virtual private gateway from a VPC. You do this if you're planning to turn off the VPC and not use it anymore. You can confirm a virtual private gateway has been completely detached from a VPC by describing the virtual private gateway (any attachments to the virtual private gateway are also described).

You must wait for the attachment's state to switch to detached before you can delete the VPC or attach a different VPC to the virtual private gateway.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DetachVpnGateway.html

Request

data DetachVpnGateway

Request constructor

detachVpnGateway

Arguments

:: Text
dvg1VpnGatewayId

-> Text
dvg1VpcId

-> DetachVpnGateway

DetachVpnGateway constructor.

The fields accessible through corresponding lenses are:

dvg1DryRun :: Maybe Bool

dvg1VpcId :: Text

dvg1VpnGatewayId :: Text

Request lenses

dvg1DryRun :: Lens' DetachVpnGateway (Maybe Bool)

dvg1VpcId :: Lens' DetachVpnGateway Text

The ID of the VPC.

dvg1VpnGatewayId :: Lens' DetachVpnGateway Text

The ID of the virtual private gateway.

Response

data DetachVpnGatewayResponse

Response constructor

detachVpnGatewayResponse :: DetachVpnGatewayResponse

DetachVpnGatewayResponse constructor.