Network.AWS.EC2.DisassociateAddress

Description

Disassociates an Elastic IP address from the instance or network interface it's associated with.

An Elastic IP address is for use in either the EC2-Classic platform or in a VPC. For more information, see Elastic IP Addresses in the Amazon Elastic Compute Cloud User Guide .

This is an idempotent operation. If you perform the operation more than once, Amazon EC2 doesn't return an error.

Creating a Request

disassociateAddress :: DisassociateAddress

Creates a value of DisassociateAddress with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dasAssociationId - [EC2-VPC] The association ID. Required for EC2-VPC.

dasPublicIP - [EC2-Classic] The Elastic IP address. Required for EC2-Classic.

dasDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data DisassociateAddress

Contains the parameters for DisassociateAddress.

See: disassociateAddress smart constructor.

Request Lenses

dasAssociationId :: Lens' DisassociateAddress (Maybe Text)

EC2-VPC

The association ID. Required for EC2-VPC.

dasPublicIP :: Lens' DisassociateAddress (Maybe Text)

EC2-Classic

The Elastic IP address. Required for EC2-Classic.

dasDryRun :: Lens' DisassociateAddress (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

disassociateAddressResponse :: DisassociateAddressResponse

Creates a value of DisassociateAddressResponse with the minimum fields required to make a request.

data DisassociateAddressResponse

See: disassociateAddressResponse smart constructor.