Network.AWS.EC2.DisassociateAddress

Description

Disassociates an Elastic IP address from the instance or network interface it's associated with.

An Elastic IP address is for use in either the EC2-Classic platform or in a VPC. For more information, see Elastic IP Addresses in the Amazon ElasticCompute Cloud User Guide for Linux.

This is an idempotent operation. If you perform the operation more than once, Amazon EC2 doesn't return an error.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DisassociateAddress.html

Request

data DisassociateAddress

Request constructor

disassociateAddress :: DisassociateAddress

DisassociateAddress constructor.

The fields accessible through corresponding lenses are:

da1AssociationId :: Maybe Text

da1DryRun :: Maybe Bool

da1PublicIp :: Maybe Text

Request lenses

da1AssociationId :: Lens' DisassociateAddress (Maybe Text)

EC2-VPC

The association ID. Required for EC2-VPC.

da1DryRun :: Lens' DisassociateAddress (Maybe Bool)

da1PublicIp :: Lens' DisassociateAddress (Maybe Text)

EC2-Classic

The Elastic IP address. Required for EC2-Classic.

Response

data DisassociateAddressResponse

Response constructor

disassociateAddressResponse :: DisassociateAddressResponse

DisassociateAddressResponse constructor.