Network.AWS.EC2.DisassociateRouteTable

Description

Disassociates a subnet from a route table.

After you perform this action, the subnet no longer uses the routes in the route table. Instead, it uses the routes in the VPC's main route table. For more information about route tables, see Route Tables in the Amazon Virtual Private Cloud User Guide .

Creating a Request

disassociateRouteTable

Arguments

:: Text
drtAssociationId

-> DisassociateRouteTable

Creates a value of DisassociateRouteTable with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drtDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

drtAssociationId - The association ID representing the current association between the route table and subnet.

data DisassociateRouteTable

Contains the parameters for DisassociateRouteTable.

See: disassociateRouteTable smart constructor.

Request Lenses

drtDryRun :: Lens' DisassociateRouteTable (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

drtAssociationId :: Lens' DisassociateRouteTable Text

The association ID representing the current association between the route table and subnet.

Destructuring the Response

disassociateRouteTableResponse :: DisassociateRouteTableResponse

Creates a value of DisassociateRouteTableResponse with the minimum fields required to make a request.

data DisassociateRouteTableResponse

See: disassociateRouteTableResponse smart constructor.