Network.AWS.EC2.GetConsoleOutput

Description

Gets the console output for the specified instance.

Instances do not have a physical monitor through which you can view their console output. They also lack physical controls that allow you to power up, reboot, or shut them down. To allow these actions, we provide them through the Amazon EC2 API and command line interface.

Instance console output is buffered and posted shortly after instance boot, reboot, and termination. Amazon EC2 preserves the most recent 64 KB output which is available for at least one hour after the most recent post.

For Linux instances, the instance console output displays the exact console output that would normally be displayed on a physical monitor attached to a computer. This output is buffered because the instance produces it and then posts it to a store where the instance's owner can retrieve it.

For Windows instances, the instance console output includes output from the EC2Config service.

Creating a Request

getConsoleOutput

Arguments

:: Text
gcoInstanceId

-> GetConsoleOutput

Creates a value of GetConsoleOutput with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcoDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

gcoInstanceId - The ID of the instance.

data GetConsoleOutput

Contains the parameters for GetConsoleOutput.

See: getConsoleOutput smart constructor.

Request Lenses

gcoDryRun :: Lens' GetConsoleOutput (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

gcoInstanceId :: Lens' GetConsoleOutput Text

The ID of the instance.

Destructuring the Response

getConsoleOutputResponse

Arguments

:: Int
gcorsResponseStatus

-> GetConsoleOutputResponse

Creates a value of GetConsoleOutputResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcorsInstanceId - The ID of the instance.

gcorsOutput - The console output, Base64-encoded. If using a command line tool, the tool decodes the output for you.

gcorsTimestamp - The time the output was last updated.

gcorsResponseStatus - -- | The response status code.

data GetConsoleOutputResponse

Contains the output of GetConsoleOutput.

See: getConsoleOutputResponse smart constructor.

Response Lenses

gcorsInstanceId :: Lens' GetConsoleOutputResponse (Maybe Text)

The ID of the instance.

gcorsOutput :: Lens' GetConsoleOutputResponse (Maybe Text)

The console output, Base64-encoded. If using a command line tool, the tool decodes the output for you.

gcorsTimestamp :: Lens' GetConsoleOutputResponse (Maybe UTCTime)

The time the output was last updated.

gcorsResponseStatus :: Lens' GetConsoleOutputResponse Int

- | The response status code.