Network.AWS.EC2.GetConsoleOutput

Description

Gets the console output for the specified instance.

Instances do not have a physical monitor through which you can view their console output. They also lack physical controls that allow you to power up, reboot, or shut them down. To allow these actions, we provide them through the Amazon EC2 API and command line interface.

Instance console output is buffered and posted shortly after instance boot, reboot, and termination. Amazon EC2 preserves the most recent 64 KB output which is available for at least one hour after the most recent post.

For Linux instances, the instance console output displays the exact console output that would normally be displayed on a physical monitor attached to a computer. This output is buffered because the instance produces it and then posts it to a store where the instance's owner can retrieve it.

For Windows instances, the instance console output includes output from the EC2Config service.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-GetConsoleOutput.html

Request

data GetConsoleOutput

Request constructor

getConsoleOutput

Arguments

:: Text
gcoInstanceId

-> GetConsoleOutput

GetConsoleOutput constructor.

The fields accessible through corresponding lenses are:

gcoDryRun :: Maybe Bool

gcoInstanceId :: Text

Request lenses

gcoDryRun :: Lens' GetConsoleOutput (Maybe Bool)

gcoInstanceId :: Lens' GetConsoleOutput Text

The ID of the instance.

Response

data GetConsoleOutputResponse

Response constructor

getConsoleOutputResponse :: GetConsoleOutputResponse

GetConsoleOutputResponse constructor.

The fields accessible through corresponding lenses are:

gcorInstanceId :: Maybe Text

gcorOutput :: Maybe Text

gcorTimestamp :: Maybe UTCTime

Response lenses

gcorInstanceId :: Lens' GetConsoleOutputResponse (Maybe Text)

The ID of the instance.

gcorOutput :: Lens' GetConsoleOutputResponse (Maybe Text)

The console output, Base64 encoded.

gcorTimestamp :: Lens' GetConsoleOutputResponse (Maybe UTCTime)

The time the output was last updated.