Network.AWS.EC2.GetHostReservationPurchasePreview

Description

Preview a reservation purchase with configurations that match those of your Dedicated Host. You must have active Dedicated Hosts in your account before you purchase a reservation.

This is a preview of the PurchaseHostReservation action and does not result in the offering being purchased.

Creating a Request

getHostReservationPurchasePreview

Arguments

:: Text
ghrppOfferingId

-> GetHostReservationPurchasePreview

Creates a value of GetHostReservationPurchasePreview with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ghrppHostIdSet - The IDs of the Dedicated Hosts that the reservation will be associated with.

ghrppOfferingId - The offering ID of the reservation.

data GetHostReservationPurchasePreview

See: getHostReservationPurchasePreview smart constructor.

Request Lenses

ghrppHostIdSet :: Lens' GetHostReservationPurchasePreview [Text]

The IDs of the Dedicated Hosts that the reservation will be associated with.

ghrppOfferingId :: Lens' GetHostReservationPurchasePreview Text

The offering ID of the reservation.

Destructuring the Response

getHostReservationPurchasePreviewResponse

Arguments

:: Int
ghrpprsResponseStatus

-> GetHostReservationPurchasePreviewResponse

Creates a value of GetHostReservationPurchasePreviewResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ghrpprsCurrencyCode - The currency in which the totalUpfrontPrice and totalHourlyPrice amounts are specified. At this time, the only supported currency is USD .

ghrpprsTotalHourlyPrice - The potential total hourly price of the reservation per hour.

ghrpprsTotalUpfrontPrice - The potential total upfront price. This is billed immediately.

ghrpprsPurchase - The purchase information of the Dedicated Host Reservation and the Dedicated Hosts associated with it.

ghrpprsResponseStatus - -- | The response status code.

data GetHostReservationPurchasePreviewResponse

See: getHostReservationPurchasePreviewResponse smart constructor.

Response Lenses

ghrpprsCurrencyCode :: Lens' GetHostReservationPurchasePreviewResponse (Maybe CurrencyCodeValues)

The currency in which the totalUpfrontPrice and totalHourlyPrice amounts are specified. At this time, the only supported currency is USD .

ghrpprsTotalHourlyPrice :: Lens' GetHostReservationPurchasePreviewResponse (Maybe Text)

The potential total hourly price of the reservation per hour.

ghrpprsTotalUpfrontPrice :: Lens' GetHostReservationPurchasePreviewResponse (Maybe Text)

The potential total upfront price. This is billed immediately.

ghrpprsPurchase :: Lens' GetHostReservationPurchasePreviewResponse [Purchase]

The purchase information of the Dedicated Host Reservation and the Dedicated Hosts associated with it.

ghrpprsResponseStatus :: Lens' GetHostReservationPurchasePreviewResponse Int

- | The response status code.