Network.AWS.EC2.GetPasswordData

Description

Retrieves the encrypted administrator password for an instance running Windows.

The Windows password is generated at boot if the EC2Config service plugin, Ec2SetPassword , is enabled. This usually only happens the first time an AMI is launched, and then Ec2SetPassword is automatically disabled. The password is not generated for rebundled AMIs unless Ec2SetPassword is enabled before bundling.

The password is encrypted using the key pair that you specified when you launched the instance. You must provide the corresponding key pair file.

Password generation and encryption takes a few moments. We recommend that you wait up to 15 minutes after launching an instance before trying to retrieve the generated password.

Creating a Request

getPasswordData

Arguments

:: Text
gpdInstanceId

-> GetPasswordData

Creates a value of GetPasswordData with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gpdDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

gpdInstanceId - The ID of the Windows instance.

data GetPasswordData

Contains the parameters for GetPasswordData.

See: getPasswordData smart constructor.

Request Lenses

gpdDryRun :: Lens' GetPasswordData (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

gpdInstanceId :: Lens' GetPasswordData Text

The ID of the Windows instance.

Destructuring the Response

getPasswordDataResponse

Arguments

:: Int
gpdrsResponseStatus

-> Text
gpdrsInstanceId

-> Text
gpdrsPasswordData

-> UTCTime
gpdrsTimestamp

-> GetPasswordDataResponse

Creates a value of GetPasswordDataResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gpdrsResponseStatus - -- | The response status code.

gpdrsInstanceId - The ID of the Windows instance.

gpdrsPasswordData - The password of the instance.

gpdrsTimestamp - The time the data was last updated.

data GetPasswordDataResponse

Contains the output of GetPasswordData.

See: getPasswordDataResponse smart constructor.

Response Lenses

gpdrsResponseStatus :: Lens' GetPasswordDataResponse Int

- | The response status code.

gpdrsInstanceId :: Lens' GetPasswordDataResponse Text

The ID of the Windows instance.

gpdrsPasswordData :: Lens' GetPasswordDataResponse Text

The password of the instance.

gpdrsTimestamp :: Lens' GetPasswordDataResponse UTCTime

The time the data was last updated.