Network.AWS.EC2.GetPasswordData

Description

Retrieves the encrypted administrator password for an instance running Windows.

The Windows password is generated at boot if the EC2Config service plugin, Ec2SetPassword, is enabled. This usually only happens the first time an AMI is launched, and then Ec2SetPassword is automatically disabled. The password is not generated for rebundled AMIs unless Ec2SetPassword is enabled before bundling.

The password is encrypted using the key pair that you specified when you launched the instance. You must provide the corresponding key pair file.

Password generation and encryption takes a few moments. We recommend that you wait up to 15 minutes after launching an instance before trying to retrieve the generated password.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-GetPasswordData.html

Request

data GetPasswordData

Request constructor

getPasswordData

Arguments

:: Text
gpdInstanceId

-> GetPasswordData

GetPasswordData constructor.

The fields accessible through corresponding lenses are:

gpdDryRun :: Maybe Bool

gpdInstanceId :: Text

Request lenses

gpdDryRun :: Lens' GetPasswordData (Maybe Bool)

gpdInstanceId :: Lens' GetPasswordData Text

The ID of the Windows instance.

Response

data GetPasswordDataResponse

Response constructor

getPasswordDataResponse

Arguments

:: Text
gpdrInstanceId

-> UTCTime
gpdrTimestamp

-> Text
gpdrPasswordData

-> GetPasswordDataResponse

GetPasswordDataResponse constructor.

The fields accessible through corresponding lenses are:

gpdrInstanceId :: Text

gpdrPasswordData :: Text

gpdrTimestamp :: UTCTime

Response lenses

gpdrInstanceId :: Lens' GetPasswordDataResponse Text

The ID of the Windows instance.

gpdrPasswordData :: Lens' GetPasswordDataResponse Text

The password of the instance.

gpdrTimestamp :: Lens' GetPasswordDataResponse UTCTime

The time the data was last updated.