Network.AWS.EC2.ImportKeyPair

Description

Imports the public key from an RSA key pair that you created with a third-party tool. Compare this with CreateKeyPair, in which AWS creates the key pair and gives the keys to you (AWS keeps a copy of the public key). With ImportKeyPair, you create the key pair and give AWS just the public key. The private key is never transferred between you and AWS. For more information about key pairs, see <http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/ec2-key-pairs.html Key Pairs> in the Amazon Elastic Compute Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-ImportKeyPair.html

Request

data ImportKeyPair

Request constructor

importKeyPair

Arguments

:: Text
ikpKeyName

-> Base64
ikpPublicKeyMaterial

-> ImportKeyPair

ImportKeyPair constructor.

The fields accessible through corresponding lenses are:

ikpDryRun :: Maybe Bool

ikpKeyName :: Text

ikpPublicKeyMaterial :: Base64

Request lenses

ikpDryRun :: Lens' ImportKeyPair (Maybe Bool)

ikpKeyName :: Lens' ImportKeyPair Text

A unique name for the key pair.

ikpPublicKeyMaterial :: Lens' ImportKeyPair Base64

The public key. You must base64 encode the public key material before sending it to AWS.

Response

data ImportKeyPairResponse

Response constructor

importKeyPairResponse :: ImportKeyPairResponse

ImportKeyPairResponse constructor.

The fields accessible through corresponding lenses are:

ikprKeyFingerprint :: Maybe Text

ikprKeyName :: Maybe Text

Response lenses

ikprKeyFingerprint :: Lens' ImportKeyPairResponse (Maybe Text)

The MD5 public key fingerprint as specified in section 4 of RFC 4716.

ikprKeyName :: Lens' ImportKeyPairResponse (Maybe Text)

The key pair name you provided.