Network.AWS.EC2.ModifyIdentityIdFormat

Description

Modifies the ID format of a resource for a specified IAM user, IAM role, or the root user for an account; or all IAM users, IAM roles, and the root user for an account. You can specify that resources should receive longer IDs (17-character IDs) when they are created.

The following resource types support longer IDs: instance | reservation | snapshot | volume . For more information, see Resource IDs in the Amazon Elastic Compute Cloud User Guide .

This setting applies to the principal specified in the request; it does not apply to the principal that makes the request.

Resources created with longer IDs are visible to all IAM roles and users, regardless of these settings and provided that they have permission to use the relevant Describe command for the resource type.

Creating a Request

modifyIdentityIdFormat

Arguments

:: Text
miifPrincipalARN

-> Text
miifResource

-> Bool
miifUseLongIds

-> ModifyIdentityIdFormat

Creates a value of ModifyIdentityIdFormat with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

miifPrincipalARN - The ARN of the principal, which can be an IAM user, IAM role, or the root user. Specify all to modify the ID format for all IAM users, IAM roles, and the root user of the account.

miifResource - The type of resource: instance | reservation | snapshot | volume

miifUseLongIds - Indicates whether the resource should use longer IDs (17-character IDs)

data ModifyIdentityIdFormat

Contains the parameters of ModifyIdentityIdFormat.

See: modifyIdentityIdFormat smart constructor.

Request Lenses

miifPrincipalARN :: Lens' ModifyIdentityIdFormat Text

The ARN of the principal, which can be an IAM user, IAM role, or the root user. Specify all to modify the ID format for all IAM users, IAM roles, and the root user of the account.

miifResource :: Lens' ModifyIdentityIdFormat Text

The type of resource: instance | reservation | snapshot | volume

miifUseLongIds :: Lens' ModifyIdentityIdFormat Bool

Indicates whether the resource should use longer IDs (17-character IDs)

Destructuring the Response

modifyIdentityIdFormatResponse :: ModifyIdentityIdFormatResponse

Creates a value of ModifyIdentityIdFormatResponse with the minimum fields required to make a request.

data ModifyIdentityIdFormatResponse

See: modifyIdentityIdFormatResponse smart constructor.