Network.AWS.EC2.ModifyImageAttribute

Description

Modifies the specified attribute of the specified AMI. You can specify only one attribute at a time.

Creating a Request

modifyImageAttribute

Arguments

:: Text
miaImageId

-> ModifyImageAttribute

Creates a value of ModifyImageAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

miaAttribute - The name of the attribute to modify.

miaUserIds - One or more AWS account IDs. This is only valid when modifying the launchPermission attribute.

miaUserGroups - One or more user groups. This is only valid when modifying the launchPermission attribute.

miaValue - The value of the attribute being modified. This is only valid when modifying the description attribute.

miaLaunchPermission - A launch permission modification.

miaOperationType - The operation type.

miaProductCodes - One or more product codes. After you add a product code to an AMI, it can't be removed. This is only valid when modifying the productCodes attribute.

miaDescription - A description for the AMI.

miaDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

miaImageId - The ID of the AMI.

data ModifyImageAttribute

Contains the parameters for ModifyImageAttribute.

See: modifyImageAttribute smart constructor.

Request Lenses

miaAttribute :: Lens' ModifyImageAttribute (Maybe Text)

The name of the attribute to modify.

miaUserIds :: Lens' ModifyImageAttribute [Text]

One or more AWS account IDs. This is only valid when modifying the launchPermission attribute.

miaUserGroups :: Lens' ModifyImageAttribute [Text]

One or more user groups. This is only valid when modifying the launchPermission attribute.

miaValue :: Lens' ModifyImageAttribute (Maybe Text)

The value of the attribute being modified. This is only valid when modifying the description attribute.

miaLaunchPermission :: Lens' ModifyImageAttribute (Maybe LaunchPermissionModifications)

A launch permission modification.

miaOperationType :: Lens' ModifyImageAttribute (Maybe OperationType)

The operation type.

miaProductCodes :: Lens' ModifyImageAttribute [Text]

One or more product codes. After you add a product code to an AMI, it can't be removed. This is only valid when modifying the productCodes attribute.

miaDescription :: Lens' ModifyImageAttribute (Maybe AttributeValue)

A description for the AMI.

miaDryRun :: Lens' ModifyImageAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

miaImageId :: Lens' ModifyImageAttribute Text

The ID of the AMI.

Destructuring the Response

modifyImageAttributeResponse :: ModifyImageAttributeResponse

Creates a value of ModifyImageAttributeResponse with the minimum fields required to make a request.

data ModifyImageAttributeResponse

See: modifyImageAttributeResponse smart constructor.