Network.AWS.EC2.ModifyInstanceAttribute

Description

Modifies the specified attribute of the specified instance. You can specify only one attribute at a time.

To modify some attributes, the instance must be stopped. For more information, see Modifying Attributes of a Stopped Instance in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

modifyInstanceAttribute

Arguments

:: Text
mInstanceId

-> ModifyInstanceAttribute

Creates a value of ModifyInstanceAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mGroups - [EC2-VPC] Changes the security groups of the instance. You must specify at least one security group, even if it's just the default security group for the VPC. You must specify the security group ID, not the security group name.

mAttribute - The name of the attribute.

mEnaSupport - Set to true to enable enhanced networking with ENA for the instance. This option is supported only for HVM instances. Specifying this option with a PV instance can make it unreachable.

mSourceDestCheck - Specifies whether source/destination checking is enabled. A value of true means that checking is enabled, and false means checking is disabled. This value must be false for a NAT instance to perform NAT.

mDisableAPITermination - If the value is true , you can't terminate the instance using the Amazon EC2 console, CLI, or API; otherwise, you can. You cannot use this parameter for Spot Instances.

mKernel - Changes the instance's kernel to the specified value. We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB .

mRAMDisk - Changes the instance's RAM disk to the specified value. We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB .

mValue - A new value for the attribute. Use only with the kernel , ramdisk , userData , disableApiTermination , or instanceInitiatedShutdownBehavior attribute.

mInstanceType - Changes the instance type to the specified value. For more information, see Instance Types . If the instance type is not valid, the error returned is InvalidInstanceAttributeValue .

mSRIOVNetSupport - Set to simple to enable enhanced networking with the Intel 82599 Virtual Function interface for the instance. There is no way to disable enhanced networking with the Intel 82599 Virtual Function interface at this time. This option is supported only for HVM instances. Specifying this option with a PV instance can make it unreachable.

mEBSOptimized - Specifies whether the instance is optimized for EBS IO. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal EBS IO performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS Optimized instance.

mUserData - Changes the instance's user data to the specified value. If you are using an AWS SDK or command line tool, Base64-encoding is performed for you, and you can load the text from a file. Otherwise, you must provide Base64-encoded text.

mInstanceInitiatedShutdownBehavior - Specifies whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

mBlockDeviceMappings - Modifies the DeleteOnTermination attribute for volumes that are currently attached. The volume must be owned by the caller. If no value is specified for DeleteOnTermination , the default is true and the volume is deleted when the instance is terminated. To add instance store volumes to an Amazon EBS-backed instance, you must add them when you launch the instance. For more information, see Updating the Block Device Mapping when Launching an Instance in the Amazon Elastic Compute Cloud User Guide .

mDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

mInstanceId - The ID of the instance.

data ModifyInstanceAttribute

Contains the parameters for ModifyInstanceAttribute.

See: modifyInstanceAttribute smart constructor.

Request Lenses

mGroups :: Lens' ModifyInstanceAttribute [Text]

EC2-VPC

Changes the security groups of the instance. You must specify at least one security group, even if it's just the default security group for the VPC. You must specify the security group ID, not the security group name.

mAttribute :: Lens' ModifyInstanceAttribute (Maybe InstanceAttributeName)

The name of the attribute.

mEnaSupport :: Lens' ModifyInstanceAttribute (Maybe AttributeBooleanValue)

Set to true to enable enhanced networking with ENA for the instance. This option is supported only for HVM instances. Specifying this option with a PV instance can make it unreachable.

mSourceDestCheck :: Lens' ModifyInstanceAttribute (Maybe AttributeBooleanValue)

Specifies whether source/destination checking is enabled. A value of true means that checking is enabled, and false means checking is disabled. This value must be false for a NAT instance to perform NAT.

mDisableAPITermination :: Lens' ModifyInstanceAttribute (Maybe AttributeBooleanValue)

If the value is true , you can't terminate the instance using the Amazon EC2 console, CLI, or API; otherwise, you can. You cannot use this parameter for Spot Instances.

mKernel :: Lens' ModifyInstanceAttribute (Maybe AttributeValue)

Changes the instance's kernel to the specified value. We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB .

mRAMDisk :: Lens' ModifyInstanceAttribute (Maybe AttributeValue)

Changes the instance's RAM disk to the specified value. We recommend that you use PV-GRUB instead of kernels and RAM disks. For more information, see PV-GRUB .

mValue :: Lens' ModifyInstanceAttribute (Maybe Text)

A new value for the attribute. Use only with the kernel , ramdisk , userData , disableApiTermination , or instanceInitiatedShutdownBehavior attribute.

mInstanceType :: Lens' ModifyInstanceAttribute (Maybe AttributeValue)

Changes the instance type to the specified value. For more information, see Instance Types . If the instance type is not valid, the error returned is InvalidInstanceAttributeValue .

mSRIOVNetSupport :: Lens' ModifyInstanceAttribute (Maybe AttributeValue)

Set to simple to enable enhanced networking with the Intel 82599 Virtual Function interface for the instance. There is no way to disable enhanced networking with the Intel 82599 Virtual Function interface at this time. This option is supported only for HVM instances. Specifying this option with a PV instance can make it unreachable.

mEBSOptimized :: Lens' ModifyInstanceAttribute (Maybe AttributeBooleanValue)

Specifies whether the instance is optimized for EBS IO. This optimization provides dedicated throughput to Amazon EBS and an optimized configuration stack to provide optimal EBS IO performance. This optimization isn't available with all instance types. Additional usage charges apply when using an EBS Optimized instance.

mUserData :: Lens' ModifyInstanceAttribute (Maybe BlobAttributeValue)

Changes the instance's user data to the specified value. If you are using an AWS SDK or command line tool, Base64-encoding is performed for you, and you can load the text from a file. Otherwise, you must provide Base64-encoded text.

mInstanceInitiatedShutdownBehavior :: Lens' ModifyInstanceAttribute (Maybe AttributeValue)

Specifies whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

mBlockDeviceMappings :: Lens' ModifyInstanceAttribute [InstanceBlockDeviceMappingSpecification]

Modifies the DeleteOnTermination attribute for volumes that are currently attached. The volume must be owned by the caller. If no value is specified for DeleteOnTermination , the default is true and the volume is deleted when the instance is terminated. To add instance store volumes to an Amazon EBS-backed instance, you must add them when you launch the instance. For more information, see Updating the Block Device Mapping when Launching an Instance in the Amazon Elastic Compute Cloud User Guide .

mDryRun :: Lens' ModifyInstanceAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

mInstanceId :: Lens' ModifyInstanceAttribute Text

The ID of the instance.

Destructuring the Response

modifyInstanceAttributeResponse :: ModifyInstanceAttributeResponse

Creates a value of ModifyInstanceAttributeResponse with the minimum fields required to make a request.

data ModifyInstanceAttributeResponse

See: modifyInstanceAttributeResponse smart constructor.