Network.AWS.EC2.ModifyNetworkInterfaceAttribute

Description

Modifies the specified network interface attribute. You can specify only one attribute at a time.

Creating a Request

modifyNetworkInterfaceAttribute

Arguments

:: Text
mniaNetworkInterfaceId

-> ModifyNetworkInterfaceAttribute

Creates a value of ModifyNetworkInterfaceAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mniaGroups - Changes the security groups for the network interface. The new set of groups you specify replaces the current set. You must specify at least one group, even if it's just the default security group in the VPC. You must specify the ID of the security group, not the name.

mniaSourceDestCheck - Indicates whether sourcedestination checking is enabled. A value of true means checking is enabled, and false means checking is disabled. This value must be false for a NAT instance to perform NAT. For more information, see <http:docs.aws.amazon.comAmazonVPClatestUserGuideVPC_NAT_Instance.html NAT Instances> in the Amazon Virtual Private Cloud User Guide/ .

mniaAttachment - Information about the interface attachment. If modifying the 'delete on termination' attribute, you must specify the ID of the interface attachment.

mniaDescription - A description for the network interface.

mniaDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

mniaNetworkInterfaceId - The ID of the network interface.

data ModifyNetworkInterfaceAttribute

Contains the parameters for ModifyNetworkInterfaceAttribute.

See: modifyNetworkInterfaceAttribute smart constructor.

Request Lenses

mniaGroups :: Lens' ModifyNetworkInterfaceAttribute [Text]

Changes the security groups for the network interface. The new set of groups you specify replaces the current set. You must specify at least one group, even if it's just the default security group in the VPC. You must specify the ID of the security group, not the name.

mniaSourceDestCheck :: Lens' ModifyNetworkInterfaceAttribute (Maybe AttributeBooleanValue)

Indicates whether sourcedestination checking is enabled. A value of true means checking is enabled, and false means checking is disabled. This value must be false for a NAT instance to perform NAT. For more information, see <http:docs.aws.amazon.comAmazonVPClatestUserGuideVPC_NAT_Instance.html NAT Instances> in the Amazon Virtual Private Cloud User Guide/ .

mniaAttachment :: Lens' ModifyNetworkInterfaceAttribute (Maybe NetworkInterfaceAttachmentChanges)

Information about the interface attachment. If modifying the 'delete on termination' attribute, you must specify the ID of the interface attachment.

mniaDescription :: Lens' ModifyNetworkInterfaceAttribute (Maybe AttributeValue)

A description for the network interface.

mniaDryRun :: Lens' ModifyNetworkInterfaceAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

mniaNetworkInterfaceId :: Lens' ModifyNetworkInterfaceAttribute Text

The ID of the network interface.

Destructuring the Response

modifyNetworkInterfaceAttributeResponse :: ModifyNetworkInterfaceAttributeResponse

Creates a value of ModifyNetworkInterfaceAttributeResponse with the minimum fields required to make a request.

data ModifyNetworkInterfaceAttributeResponse

See: modifyNetworkInterfaceAttributeResponse smart constructor.