Network.AWS.EC2.ModifyNetworkInterfaceAttribute

Description

Modifies the specified network interface attribute. You can specify only one attribute at a time.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-ModifyNetworkInterfaceAttribute.html

Request

data ModifyNetworkInterfaceAttribute

Request constructor

modifyNetworkInterfaceAttribute

Arguments

:: Text
mniaNetworkInterfaceId

-> ModifyNetworkInterfaceAttribute

ModifyNetworkInterfaceAttribute constructor.

The fields accessible through corresponding lenses are:

mniaAttachment :: Maybe NetworkInterfaceAttachmentChanges

mniaDescription :: Maybe AttributeValue

mniaDryRun :: Maybe Bool

mniaGroups :: [Text]

mniaNetworkInterfaceId :: Text

mniaSourceDestCheck :: Maybe AttributeBooleanValue

Request lenses

mniaAttachment :: Lens' ModifyNetworkInterfaceAttribute (Maybe NetworkInterfaceAttachmentChanges)

Information about the interface attachment. If modifying the 'delete on termination' attribute, you must specify the ID of the interface attachment.

mniaDescription :: Lens' ModifyNetworkInterfaceAttribute (Maybe AttributeValue)

A description for the network interface.

mniaDryRun :: Lens' ModifyNetworkInterfaceAttribute (Maybe Bool)

mniaGroups :: Lens' ModifyNetworkInterfaceAttribute [Text]

Changes the security groups for the network interface. The new set of groups you specify replaces the current set. You must specify at least one group, even if it's just the default security group in the VPC. You must specify the ID of the security group, not the name.

mniaNetworkInterfaceId :: Lens' ModifyNetworkInterfaceAttribute Text

The ID of the network interface.

mniaSourceDestCheck :: Lens' ModifyNetworkInterfaceAttribute (Maybe AttributeBooleanValue)

Indicates whether source/destination checking is enabled. A value of true means checking is enabled, and false means checking is disabled. This value must be false for a NAT instance to perform NAT. For more information, see NAT Instances in the Amazon Virtual Private Cloud User Guide.

Response

data ModifyNetworkInterfaceAttributeResponse

Response constructor

modifyNetworkInterfaceAttributeResponse :: ModifyNetworkInterfaceAttributeResponse

ModifyNetworkInterfaceAttributeResponse constructor.