Network.AWS.EC2.ModifyReservedInstances

Description

Modifies the Availability Zone, instance count, instance type, or network platform (EC2-Classic or EC2-VPC) of your Standard Reserved Instances. The Reserved Instances to be modified must be identical, except for Availability Zone, network platform, and instance type.

For more information, see Modifying Reserved Instances in the Amazon Elastic Compute Cloud User Guide.

Creating a Request

modifyReservedInstances :: ModifyReservedInstances

Creates a value of ModifyReservedInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mriClientToken - A unique, case-sensitive token you provide to ensure idempotency of your modification request. For more information, see Ensuring Idempotency .

mriReservedInstancesIds - The IDs of the Reserved Instances to modify.

mriTargetConfigurations - The configuration settings for the Reserved Instances to modify.

data ModifyReservedInstances

Contains the parameters for ModifyReservedInstances.

See: modifyReservedInstances smart constructor.

Request Lenses

mriClientToken :: Lens' ModifyReservedInstances (Maybe Text)

A unique, case-sensitive token you provide to ensure idempotency of your modification request. For more information, see Ensuring Idempotency .

mriReservedInstancesIds :: Lens' ModifyReservedInstances [Text]

The IDs of the Reserved Instances to modify.

mriTargetConfigurations :: Lens' ModifyReservedInstances [ReservedInstancesConfiguration]

The configuration settings for the Reserved Instances to modify.

Destructuring the Response

modifyReservedInstancesResponse

Arguments

:: Int
mrirsResponseStatus

-> ModifyReservedInstancesResponse

Creates a value of ModifyReservedInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mrirsReservedInstancesModificationId - The ID for the modification.

mrirsResponseStatus - -- | The response status code.

data ModifyReservedInstancesResponse

Contains the output of ModifyReservedInstances.

See: modifyReservedInstancesResponse smart constructor.

Response Lenses

mrirsReservedInstancesModificationId :: Lens' ModifyReservedInstancesResponse (Maybe Text)

The ID for the modification.

mrirsResponseStatus :: Lens' ModifyReservedInstancesResponse Int

- | The response status code.