Network.AWS.EC2.ModifySnapshotAttribute

Description

Adds or removes permission settings for the specified snapshot. You may add or remove specified AWS account IDs from a snapshot's list of create volume permissions, but you cannot do both in a single API call. If you need to both add and remove account IDs for a snapshot, you must use multiple API calls.

For more information on modifying snapshot permissions, see Sharing Snapshots in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

modifySnapshotAttribute

Arguments

:: Text
msaSnapshotId

-> ModifySnapshotAttribute

Creates a value of ModifySnapshotAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

msaAttribute - The snapshot attribute to modify.

msaCreateVolumePermission - A JSON representation of the snapshot attribute modification.

msaUserIds - The account ID to modify for the snapshot.

msaGroupNames - The group to modify for the snapshot.

msaOperationType - The type of operation to perform to the attribute.

msaDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

msaSnapshotId - The ID of the snapshot.

data ModifySnapshotAttribute

Contains the parameters for ModifySnapshotAttribute.

See: modifySnapshotAttribute smart constructor.

Request Lenses

msaAttribute :: Lens' ModifySnapshotAttribute (Maybe SnapshotAttributeName)

The snapshot attribute to modify.

msaCreateVolumePermission :: Lens' ModifySnapshotAttribute (Maybe CreateVolumePermissionModifications)

A JSON representation of the snapshot attribute modification.

msaUserIds :: Lens' ModifySnapshotAttribute [Text]

The account ID to modify for the snapshot.

msaGroupNames :: Lens' ModifySnapshotAttribute [Text]

The group to modify for the snapshot.

msaOperationType :: Lens' ModifySnapshotAttribute (Maybe OperationType)

The type of operation to perform to the attribute.

msaDryRun :: Lens' ModifySnapshotAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

msaSnapshotId :: Lens' ModifySnapshotAttribute Text

The ID of the snapshot.

Destructuring the Response

modifySnapshotAttributeResponse :: ModifySnapshotAttributeResponse

Creates a value of ModifySnapshotAttributeResponse with the minimum fields required to make a request.

data ModifySnapshotAttributeResponse

See: modifySnapshotAttributeResponse smart constructor.