Network.AWS.EC2.ModifySubnetAttribute

Description

Modifies a subnet attribute. You can only modify one attribute at a time.

Creating a Request

modifySubnetAttribute

Arguments

:: Text
msaSubnetId

-> ModifySubnetAttribute

Creates a value of ModifySubnetAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

msaAssignIPv6AddressOnCreation - Specify true to indicate that network interfaces created in the specified subnet should be assigned an IPv6 address. This includes a network interface that's created when launching an instance into the subnet (the instance therefore receives an IPv6 address). If you enable the IPv6 addressing feature for your subnet, your network interface or instance only receives an IPv6 address if it's created using version 2016-11-15 or later of the Amazon EC2 API.

msaMapPublicIPOnLaunch - Specify true to indicate that network interfaces created in the specified subnet should be assigned a public IPv4 address. This includes a network interface that's created when launching an instance into the subnet (the instance therefore receives a public IPv4 address).

msaSubnetId - The ID of the subnet.

data ModifySubnetAttribute

Contains the parameters for ModifySubnetAttribute.

See: modifySubnetAttribute smart constructor.

Request Lenses

msaAssignIPv6AddressOnCreation :: Lens' ModifySubnetAttribute (Maybe AttributeBooleanValue)

Specify true to indicate that network interfaces created in the specified subnet should be assigned an IPv6 address. This includes a network interface that's created when launching an instance into the subnet (the instance therefore receives an IPv6 address). If you enable the IPv6 addressing feature for your subnet, your network interface or instance only receives an IPv6 address if it's created using version 2016-11-15 or later of the Amazon EC2 API.

msaMapPublicIPOnLaunch :: Lens' ModifySubnetAttribute (Maybe AttributeBooleanValue)

Specify true to indicate that network interfaces created in the specified subnet should be assigned a public IPv4 address. This includes a network interface that's created when launching an instance into the subnet (the instance therefore receives a public IPv4 address).

msaSubnetId :: Lens' ModifySubnetAttribute Text

The ID of the subnet.

Destructuring the Response

modifySubnetAttributeResponse :: ModifySubnetAttributeResponse

Creates a value of ModifySubnetAttributeResponse with the minimum fields required to make a request.

data ModifySubnetAttributeResponse

See: modifySubnetAttributeResponse smart constructor.