Network.AWS.EC2.ModifyVPCEndpoint

Description

Modifies attributes of a specified VPC endpoint. You can modify the policy associated with the endpoint, and you can add and remove route tables associated with the endpoint.

Creating a Request

modifyVPCEndpoint

Arguments

:: Text
mveVPCEndpointId

-> ModifyVPCEndpoint

Creates a value of ModifyVPCEndpoint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mvePolicyDocument - A policy document to attach to the endpoint. The policy must be in valid JSON format.

mveRemoveRouteTableIds - One or more route table IDs to disassociate from the endpoint.

mveResetPolicy - Specify true to reset the policy document to the default policy. The default policy allows access to the service.

mveAddRouteTableIds - One or more route tables IDs to associate with the endpoint.

mveDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

mveVPCEndpointId - The ID of the endpoint.

data ModifyVPCEndpoint

Contains the parameters for ModifyVpcEndpoint.

See: modifyVPCEndpoint smart constructor.

Request Lenses

mvePolicyDocument :: Lens' ModifyVPCEndpoint (Maybe Text)

A policy document to attach to the endpoint. The policy must be in valid JSON format.

mveRemoveRouteTableIds :: Lens' ModifyVPCEndpoint [Text]

One or more route table IDs to disassociate from the endpoint.

mveResetPolicy :: Lens' ModifyVPCEndpoint (Maybe Bool)

Specify true to reset the policy document to the default policy. The default policy allows access to the service.

mveAddRouteTableIds :: Lens' ModifyVPCEndpoint [Text]

One or more route tables IDs to associate with the endpoint.

mveDryRun :: Lens' ModifyVPCEndpoint (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

mveVPCEndpointId :: Lens' ModifyVPCEndpoint Text

The ID of the endpoint.

Destructuring the Response

modifyVPCEndpointResponse

Arguments

:: Int
mversResponseStatus

-> ModifyVPCEndpointResponse

Creates a value of ModifyVPCEndpointResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mversReturn - Returns true if the request succeeds; otherwise, it returns an error.

mversResponseStatus - -- | The response status code.

data ModifyVPCEndpointResponse

Contains the output of ModifyVpcEndpoint.

See: modifyVPCEndpointResponse smart constructor.

Response Lenses

mversReturn :: Lens' ModifyVPCEndpointResponse (Maybe Bool)

Returns true if the request succeeds; otherwise, it returns an error.

mversResponseStatus :: Lens' ModifyVPCEndpointResponse Int

- | The response status code.