Network.AWS.EC2.ModifyVolume

Description

You can modify several parameters of an existing EBS volume, including volume size, volume type, and IOPS capacity. If your EBS volume is attached to a current-generation EC2 instance type, you may be able to apply these changes without stopping the instance or detaching the volume from it. For more information about modifying an EBS volume running Linux, see Modifying the Size, IOPS, or Type of an EBS Volume on Linux . For more information about modifying an EBS volume running Windows, see Modifying the Size, IOPS, or Type of an EBS Volume on Windows .

When you complete a resize operation on your volume, you need to extend the volume's file-system size to take advantage of the new storage capacity. For information about extending a Linux file system, see Extending a Linux File System . For information about extending a Windows file system, see Extending a Windows File System .

You can use CloudWatch Events to check the status of a modification to an EBS volume. For information about CloudWatch Events, see the Amazon CloudWatch Events User Guide . You can also track the status of a modification using the DescribeVolumesModifications API. For information about tracking status changes using either method, see Monitoring Volume Modifications .

Creating a Request

modifyVolume

Arguments

:: Text
mvVolumeId

-> ModifyVolume

Creates a value of ModifyVolume with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mvSize - Target size in GiB of the volume to be modified. Target volume size must be greater than or equal to than the existing size of the volume. For information about available EBS volume sizes, see http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/EBSVolumeTypes.html . Default: If no size is specified, the existing size is retained.

mvIOPS - Target IOPS rate of the volume to be modified. Only valid for Provisioned IOPS SSD (io1 ) volumes. For more information about io1 IOPS configuration, see http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/EBSVolumeTypes.html#EBSVolumeTypes_piops . Default: If no IOPS value is specified, the existing value is retained.

mvVolumeType - Target EBS volume type of the volume to be modified The API does not support modifications for volume type standard . You also cannot change the type of a volume to standard . Default: If no type is specified, the existing type is retained.

mvDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

mvVolumeId - Undocumented member.

data ModifyVolume

See: modifyVolume smart constructor.

Request Lenses

mvSize :: Lens' ModifyVolume (Maybe Int)

Target size in GiB of the volume to be modified. Target volume size must be greater than or equal to than the existing size of the volume. For information about available EBS volume sizes, see http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/EBSVolumeTypes.html . Default: If no size is specified, the existing size is retained.

mvIOPS :: Lens' ModifyVolume (Maybe Int)

Target IOPS rate of the volume to be modified. Only valid for Provisioned IOPS SSD (io1 ) volumes. For more information about io1 IOPS configuration, see http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/EBSVolumeTypes.html#EBSVolumeTypes_piops . Default: If no IOPS value is specified, the existing value is retained.

mvVolumeType :: Lens' ModifyVolume (Maybe VolumeType)

Target EBS volume type of the volume to be modified The API does not support modifications for volume type standard . You also cannot change the type of a volume to standard . Default: If no type is specified, the existing type is retained.

mvDryRun :: Lens' ModifyVolume (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

mvVolumeId :: Lens' ModifyVolume Text

Undocumented member.

Destructuring the Response

modifyVolumeResponse

Arguments

:: Int
mvrsResponseStatus

-> ModifyVolumeResponse

Creates a value of ModifyVolumeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mvrsVolumeModification - A VolumeModification object.

mvrsResponseStatus - -- | The response status code.

data ModifyVolumeResponse

See: modifyVolumeResponse smart constructor.

Response Lenses

mvrsVolumeModification :: Lens' ModifyVolumeResponse (Maybe VolumeModification)

A VolumeModification object.

mvrsResponseStatus :: Lens' ModifyVolumeResponse Int

- | The response status code.