Network.AWS.EC2.ModifyVolumeAttribute

Description

Modifies a volume attribute.

By default, all IO operations for the volume are suspended when the data on the volume is determined to be potentially inconsistent, to prevent undetectable, latent data corruption. The IO access to the volume can be resumed by first enabling I/O access and then checking the data consistency on your volume.

You can change the default behavior to resume I/O operations. We recommend that you change this only for boot volumes or for volumes that are stateless or disposable.

Creating a Request

modifyVolumeAttribute

Arguments

:: Text
mvaVolumeId

-> ModifyVolumeAttribute

Creates a value of ModifyVolumeAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

mvaAutoEnableIO - Indicates whether the volume should be auto-enabled for I/O operations.

mvaDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

mvaVolumeId - The ID of the volume.

data ModifyVolumeAttribute

Contains the parameters for ModifyVolumeAttribute.

See: modifyVolumeAttribute smart constructor.

Request Lenses

mvaAutoEnableIO :: Lens' ModifyVolumeAttribute (Maybe AttributeBooleanValue)

Indicates whether the volume should be auto-enabled for I/O operations.

mvaDryRun :: Lens' ModifyVolumeAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

mvaVolumeId :: Lens' ModifyVolumeAttribute Text

The ID of the volume.

Destructuring the Response

modifyVolumeAttributeResponse :: ModifyVolumeAttributeResponse

Creates a value of ModifyVolumeAttributeResponse with the minimum fields required to make a request.

data ModifyVolumeAttributeResponse

See: modifyVolumeAttributeResponse smart constructor.