Network.AWS.EC2.PurchaseReservedInstancesOffering

Description

Purchases a Reserved Instance for use with your account. With Reserved Instances, you pay a lower hourly rate compared to On-Demand instance pricing.

Use DescribeReservedInstancesOfferings to get a list of Reserved Instance offerings that match your specifications. After you've purchased a Reserved Instance, you can check for your new Reserved Instance with DescribeReservedInstances .

For more information, see Reserved Instances and Reserved Instance Marketplace in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

purchaseReservedInstancesOffering

Arguments

:: Int
prioInstanceCount

-> Text
prioReservedInstancesOfferingId

-> PurchaseReservedInstancesOffering

Creates a value of PurchaseReservedInstancesOffering with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

prioLimitPrice - Specified for Reserved Instance Marketplace offerings to limit the total order and ensure that the Reserved Instances are not purchased at unexpected prices.

prioDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

prioInstanceCount - The number of Reserved Instances to purchase.

prioReservedInstancesOfferingId - The ID of the Reserved Instance offering to purchase.

data PurchaseReservedInstancesOffering

Contains the parameters for PurchaseReservedInstancesOffering.

See: purchaseReservedInstancesOffering smart constructor.

Request Lenses

prioLimitPrice :: Lens' PurchaseReservedInstancesOffering (Maybe ReservedInstanceLimitPrice)

Specified for Reserved Instance Marketplace offerings to limit the total order and ensure that the Reserved Instances are not purchased at unexpected prices.

prioDryRun :: Lens' PurchaseReservedInstancesOffering (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

prioInstanceCount :: Lens' PurchaseReservedInstancesOffering Int

The number of Reserved Instances to purchase.

prioReservedInstancesOfferingId :: Lens' PurchaseReservedInstancesOffering Text

The ID of the Reserved Instance offering to purchase.

Destructuring the Response

purchaseReservedInstancesOfferingResponse

Arguments

:: Int
priorsResponseStatus

-> PurchaseReservedInstancesOfferingResponse

Creates a value of PurchaseReservedInstancesOfferingResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

priorsReservedInstancesId - The IDs of the purchased Reserved Instances.

priorsResponseStatus - -- | The response status code.

data PurchaseReservedInstancesOfferingResponse

Contains the output of PurchaseReservedInstancesOffering.

See: purchaseReservedInstancesOfferingResponse smart constructor.

Response Lenses

priorsReservedInstancesId :: Lens' PurchaseReservedInstancesOfferingResponse (Maybe Text)

The IDs of the purchased Reserved Instances.

priorsResponseStatus :: Lens' PurchaseReservedInstancesOfferingResponse Int

- | The response status code.