Network.AWS.EC2.PurchaseScheduledInstances

Description

Purchases one or more Scheduled Instances with the specified schedule.

Scheduled Instances enable you to purchase Amazon EC2 compute capacity by the hour for a one-year term. Before you can purchase a Scheduled Instance, you must call DescribeScheduledInstanceAvailability to check for available schedules and obtain a purchase token. After you purchase a Scheduled Instance, you must call RunScheduledInstances during each scheduled time period.

After you purchase a Scheduled Instance, you can't cancel, modify, or resell your purchase.

Creating a Request

purchaseScheduledInstances

Arguments

:: NonEmpty PurchaseRequest
psiPurchaseRequests

-> PurchaseScheduledInstances

Creates a value of PurchaseScheduledInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

psiClientToken - Unique, case-sensitive identifier that ensures the idempotency of the request. For more information, see Ensuring Idempotency .

psiDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

psiPurchaseRequests - One or more purchase requests.

data PurchaseScheduledInstances

Contains the parameters for PurchaseScheduledInstances.

See: purchaseScheduledInstances smart constructor.

Request Lenses

psiClientToken :: Lens' PurchaseScheduledInstances (Maybe Text)

Unique, case-sensitive identifier that ensures the idempotency of the request. For more information, see Ensuring Idempotency .

psiDryRun :: Lens' PurchaseScheduledInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

psiPurchaseRequests :: Lens' PurchaseScheduledInstances (NonEmpty PurchaseRequest)

One or more purchase requests.

Destructuring the Response

purchaseScheduledInstancesResponse

Arguments

:: Int
psirsResponseStatus

-> PurchaseScheduledInstancesResponse

Creates a value of PurchaseScheduledInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

psirsScheduledInstanceSet - Information about the Scheduled Instances.

psirsResponseStatus - -- | The response status code.

data PurchaseScheduledInstancesResponse

Contains the output of PurchaseScheduledInstances.

See: purchaseScheduledInstancesResponse smart constructor.

Response Lenses

psirsScheduledInstanceSet :: Lens' PurchaseScheduledInstancesResponse [ScheduledInstance]

Information about the Scheduled Instances.

psirsResponseStatus :: Lens' PurchaseScheduledInstancesResponse Int

- | The response status code.