Network.AWS.EC2.RebootInstances

Description

Requests a reboot of one or more instances. This operation is asynchronous; it only queues a request to reboot the specified instances. The operation succeeds if the instances are valid and belong to you. Requests to reboot terminated instances are ignored.

If an instance does not cleanly shut down within four minutes, Amazon EC2 performs a hard reboot.

For more information about troubleshooting, see Getting Console Output and Rebooting Instances in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

rebootInstances :: RebootInstances

Creates a value of RebootInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rebDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rebInstanceIds - One or more instance IDs.

data RebootInstances

Contains the parameters for RebootInstances.

See: rebootInstances smart constructor.

Request Lenses

rebDryRun :: Lens' RebootInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rebInstanceIds :: Lens' RebootInstances [Text]

One or more instance IDs.

Destructuring the Response

rebootInstancesResponse :: RebootInstancesResponse

Creates a value of RebootInstancesResponse with the minimum fields required to make a request.

data RebootInstancesResponse

See: rebootInstancesResponse smart constructor.