Network.AWS.EC2.RegisterImage

Description

Registers an AMI. When you're creating an AMI, this is the final step you must complete before you can launch an instance from the AMI. For more information about creating AMIs, see Creating Your Own AMIs in the Amazon Elastic Compute Cloud User Guide .

You can also use RegisterImage to create an Amazon EBS-backed Linux AMI from a snapshot of a root device volume. You specify the snapshot using the block device mapping. For more information, see Launching a Linux Instance from a Backup in the Amazon Elastic Compute Cloud User Guide .

You can't register an image where a secondary (non-root) snapshot has AWS Marketplace product codes.

Some Linux distributions, such as Red Hat Enterprise Linux (RHEL) and SUSE Linux Enterprise Server (SLES), use the EC2 billing product code associated with an AMI to verify the subscription status for package updates. Creating an AMI from an EBS snapshot does not maintain this billing code, and subsequent instances launched from such an AMI will not be able to connect to package update infrastructure. To create an AMI that must retain billing codes, see CreateImage .

If needed, you can deregister an AMI at any time. Any modifications you make to an AMI backed by an instance store volume invalidates its registration. If you make changes to an image, deregister the previous image and register the new image.

Creating a Request

registerImage

Arguments

:: Text
riName

-> RegisterImage

Creates a value of RegisterImage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

riVirtualizationType - The type of virtualization. Default: paravirtual

riImageLocation - The full path to your AMI manifest in Amazon S3 storage.

riEnaSupport - Set to true to enable enhanced networking with ENA for the AMI and any instances that you launch from the AMI. This option is supported only for HVM AMIs. Specifying this option with a PV AMI can make instances launched from the AMI unreachable.

riBillingProducts - The billing product codes. Your account must be authorized to specify billing product codes. Otherwise, you can use the AWS Marketplace to bill for the use of an AMI.

riRAMDiskId - The ID of the RAM disk.

riKernelId - The ID of the kernel.

riRootDeviceName - The name of the root device (for example, devsda1 , or devxvda ).

riSRIOVNetSupport - Set to simple to enable enhanced networking with the Intel 82599 Virtual Function interface for the AMI and any instances that you launch from the AMI. There is no way to disable sriovNetSupport at this time. This option is supported only for HVM AMIs. Specifying this option with a PV AMI can make instances launched from the AMI unreachable.

riArchitecture - The architecture of the AMI. Default: For Amazon EBS-backed AMIs, i386 . For instance store-backed AMIs, the architecture specified in the manifest file.

riDescription - A description for your AMI.

riBlockDeviceMappings - One or more block device mapping entries.

riDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

riName - A name for your AMI. Constraints: 3-128 alphanumeric characters, parentheses (()), square brackets ([]), spaces ( ), periods (.), slashes (/), dashes (-), single quotes ('), at-signs (@), or underscores(_)

data RegisterImage

Contains the parameters for RegisterImage.

See: registerImage smart constructor.

Request Lenses

riVirtualizationType :: Lens' RegisterImage (Maybe Text)

The type of virtualization. Default: paravirtual

riImageLocation :: Lens' RegisterImage (Maybe Text)

The full path to your AMI manifest in Amazon S3 storage.

riEnaSupport :: Lens' RegisterImage (Maybe Bool)

Set to true to enable enhanced networking with ENA for the AMI and any instances that you launch from the AMI. This option is supported only for HVM AMIs. Specifying this option with a PV AMI can make instances launched from the AMI unreachable.

riBillingProducts :: Lens' RegisterImage [Text]

The billing product codes. Your account must be authorized to specify billing product codes. Otherwise, you can use the AWS Marketplace to bill for the use of an AMI.

riRAMDiskId :: Lens' RegisterImage (Maybe Text)

The ID of the RAM disk.

riKernelId :: Lens' RegisterImage (Maybe Text)

The ID of the kernel.

riRootDeviceName :: Lens' RegisterImage (Maybe Text)

The name of the root device (for example, devsda1 , or devxvda ).

riSRIOVNetSupport :: Lens' RegisterImage (Maybe Text)

Set to simple to enable enhanced networking with the Intel 82599 Virtual Function interface for the AMI and any instances that you launch from the AMI. There is no way to disable sriovNetSupport at this time. This option is supported only for HVM AMIs. Specifying this option with a PV AMI can make instances launched from the AMI unreachable.

riArchitecture :: Lens' RegisterImage (Maybe ArchitectureValues)

The architecture of the AMI. Default: For Amazon EBS-backed AMIs, i386 . For instance store-backed AMIs, the architecture specified in the manifest file.

riDescription :: Lens' RegisterImage (Maybe Text)

A description for your AMI.

riBlockDeviceMappings :: Lens' RegisterImage [BlockDeviceMapping]

One or more block device mapping entries.

riDryRun :: Lens' RegisterImage (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

riName :: Lens' RegisterImage Text

A name for your AMI. Constraints: 3-128 alphanumeric characters, parentheses (()), square brackets ([]), spaces ( ), periods (.), slashes (/), dashes (-), single quotes ('), at-signs (@), or underscores(_)

Destructuring the Response

registerImageResponse

Arguments

:: Int
rirsResponseStatus

-> RegisterImageResponse

Creates a value of RegisterImageResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rirsImageId - The ID of the newly registered AMI.

rirsResponseStatus - -- | The response status code.

data RegisterImageResponse

Contains the output of RegisterImage.

See: registerImageResponse smart constructor.

Response Lenses

rirsImageId :: Lens' RegisterImageResponse (Maybe Text)

The ID of the newly registered AMI.

rirsResponseStatus :: Lens' RegisterImageResponse Int

- | The response status code.