Network.AWS.EC2.RegisterImage

Description

Registers an AMI. When you're creating an AMI, this is the final step you must complete before you can launch an instance from the AMI. For more information about creating AMIs, see Creating Your Own AMIs in the Amazon Elastic Compute Cloud User Guide. For Amazon EBS-backed instances, CreateImage creates and registers the AMI in a single request, so you don't have to register the AMI yourself. You can also use RegisterImage to create an Amazon EBS-backed AMI from a snapshot of a root device volume. For more information, see Launching an Instance from a Snapshot in the Amazon Elastic Compute Cloud User Guide. If needed, you can deregister an AMI at any time. Any modifications you make to an AMI backed by an instance store volume invalidates its registration. If you make changes to an image, deregister the previous image and register the new image. You can't register an image where a secondary (non-root) snapshot has AWS Marketplace product codes. Example 1 This example registers the AMI specified in the my-new-image.manifest.xml manifest file, located in the bucket called myawsbucket. https:/ec2.amazonaws.com?Action=RegisterImage &amp;ImageLocation=myawsbucket/my-new-image.manifest.xml &amp;AUTHPARAMS &lt;RegisterImageResponse xmlns="http://ec2.amazonaws.com/doc/2014-02-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;imageId&gt;ami-1a2b3c4d&lt;imageId&gt; &lt;RegisterImageResponse&gt; Example 2 This example registers an Amazon EBS snapshot to create an AMI backed by Amazon EBS. https:/ec2.amazonaws.com?Action=RegisterImage &amp;RootDeviceName=devsda1 &amp;BlockDeviceMapping.1.DeviceName=devsda1 &amp;BlockDeviceMapping.1.Ebs.SnapshotId=snap-1a2b3c4d &amp;Name=MyImage &amp;AUTHPARAMS &lt;RegisterImageResponse xmlns="http://ec2.amazonaws.com/doc/2014-02-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;imageId&gt;ami-1a2b3c4d&lt;imageId&gt; &lt;RegisterImageResponse&gt; Example 3 This example registers the AMI with an Amazon EBS snapshot as the root device, a separate snapshot as a secondary device, and an empty 100 GiB Amazon EBS volume as a storage device. https:/ec2.amazonaws.com?Action=RegisterImage &amp;RootDeviceName=devsda1 &amp;BlockDeviceMapping.1.DeviceName=devsda1 &amp;BlockDeviceMapping.1.Ebs.SnapshotId=snap-1a2b3c4d &amp;BlockDeviceMapping.2.DeviceName=devsdb &amp;BlockDeviceMapping.2.Ebs.SnapshotId=snap-2a2b3c4d &amp;BlockDeviceMapping.3.DeviceName=devsdc &amp;BlockDeviceMapping.3.Ebs.VolumeSize=100 &amp;Name=MyImage &amp;AUTHPARAMS &lt;RegisterImageResponse xmlns="http://ec2.amazonaws.com/doc/2014-02-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;imageId&gt;ami-1a2b3c4d&lt;imageId&gt; &lt;RegisterImageResponse&gt;.

Request

data RegisterImage

Request constructor

mkRegisterImage

Arguments

:: Text
ri2Name

-> RegisterImage

Smart constructor for the minimum required parameters to construct a valid RegisterImage request.

The fields accessible through corresponding lenses are:

ImageLocation :: Maybe Text

Name :: Text

Description :: Maybe Text

Architecture :: Maybe ArchitectureValues

KernelId :: Maybe Text

RamdiskId :: Maybe Text

RootDeviceName :: Maybe Text

BlockDeviceMappings :: [BlockDeviceMapping]

VirtualizationType :: Maybe VirtualizationType

SriovNetSupport :: Maybe Text

Request lenses

ri2ImageLocation :: Lens' RegisterImage (Maybe Text)

The full path to your AMI manifest in Amazon S3 storage.

ri2Name :: Lens' RegisterImage Text

A name for your AMI. Constraints: 3-128 alphanumeric characters, parenthesis (()), commas (,), slashes (/), dashes (-), or underscores (_).

ri2Description :: Lens' RegisterImage (Maybe Text)

A description for your AMI.

ri2Architecture :: Lens' RegisterImage (Maybe ArchitectureValues)

The architecture of the AMI. Default: For Amazon EBS-backed AMIs, i386. For instance store-backed AMIs, the architecture specified in the manifest file.

ri2KernelId :: Lens' RegisterImage (Maybe Text)

The ID of the kernel.

ri2RamdiskId :: Lens' RegisterImage (Maybe Text)

The ID of the RAM disk.

ri2RootDeviceName :: Lens' RegisterImage (Maybe Text)

The name of the root device (for example, devsda1, or xvda).

ri2BlockDeviceMappings :: Lens' RegisterImage [BlockDeviceMapping]

One or more block device mapping entries.

ri2VirtualizationType :: Lens' RegisterImage (Maybe VirtualizationType)

The type of virtualization.

ri2SriovNetSupport :: Lens' RegisterImage (Maybe Text)

Set to simple to enable enhanced networking for the AMI and any instances that you launch from the AMI. There is no way to disable enhanced networking at this time. This option is supported only for HVM AMIs. Specifying this option with a PV AMI can make instances launched from the AMI unreachable.

Response

data RegisterImageResponse

Response constructor

mkRegisterImageResponse :: RegisterImageResponse

Smart constructor for the minimum required parameters to construct a valid RegisterImageResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ImageId :: Maybe Text

Response lenses

rirImageId :: Lens' RegisterImageResponse (Maybe Text)

The ID of the newly registered AMI.