Network.AWS.EC2.RegisterImage

Description

Registers an AMI. When you're creating an AMI, this is the final step you must complete before you can launch an instance from the AMI. For more information about creating AMIs, see Creating Your Own AMIs in the AmazonElastic Compute Cloud User Guide for Linux.

For Amazon EBS-backed instances, CreateImage creates and registers the AMI in a single request, so you don't have to register the AMI yourself.

You can also use RegisterImage to create an Amazon EBS-backed AMI from a snapshot of a root device volume. For more information, see Launching anInstance from a Snapshot in the Amazon Elastic Compute Cloud User Guide forLinux.

If needed, you can deregister an AMI at any time. Any modifications you make to an AMI backed by an instance store volume invalidates its registration. If you make changes to an image, deregister the previous image and register the new image.

You can't register an image where a secondary (non-root) snapshot has AWS Marketplace product codes.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-RegisterImage.html

Request

data RegisterImage

Request constructor

registerImage

Arguments

:: Text
ri1Name

-> RegisterImage

RegisterImage constructor.

The fields accessible through corresponding lenses are:

ri1Architecture :: Maybe ArchitectureValues

ri1BlockDeviceMappings :: [BlockDeviceMapping]

ri1Description :: Maybe Text

ri1DryRun :: Maybe Bool

ri1ImageLocation :: Maybe Text

ri1KernelId :: Maybe Text

ri1Name :: Text

ri1RamdiskId :: Maybe Text

ri1RootDeviceName :: Maybe Text

ri1SriovNetSupport :: Maybe Text

ri1VirtualizationType :: Maybe Text

Request lenses

ri1Architecture :: Lens' RegisterImage (Maybe ArchitectureValues)

The architecture of the AMI.

Default: For Amazon EBS-backed AMIs, i386. For instance store-backed AMIs, the architecture specified in the manifest file.

ri1BlockDeviceMappings :: Lens' RegisterImage [BlockDeviceMapping]

One or more block device mapping entries.

ri1Description :: Lens' RegisterImage (Maybe Text)

A description for your AMI.

ri1DryRun :: Lens' RegisterImage (Maybe Bool)

ri1ImageLocation :: Lens' RegisterImage (Maybe Text)

The full path to your AMI manifest in Amazon S3 storage.

ri1KernelId :: Lens' RegisterImage (Maybe Text)

The ID of the kernel.

ri1Name :: Lens' RegisterImage Text

A name for your AMI.

Constraints: 3-128 alphanumeric characters, parentheses (()), square brackets ([]), spaces ( ), periods (.), slashes (/), dashes (-), single quotes ('), at-signs (@), or underscores(_)

ri1RamdiskId :: Lens' RegisterImage (Maybe Text)

The ID of the RAM disk.

ri1RootDeviceName :: Lens' RegisterImage (Maybe Text)

The name of the root device (for example, '/dev/sda1', or '/dev/xvda').

ri1SriovNetSupport :: Lens' RegisterImage (Maybe Text)

Set to simple to enable enhanced networking for the AMI and any instances that you launch from the AMI.

There is no way to disable enhanced networking at this time.

This option is supported only for HVM AMIs. Specifying this option with a PV AMI can make instances launched from the AMI unreachable.

ri1VirtualizationType :: Lens' RegisterImage (Maybe Text)

The type of virtualization.

Default: paravirtual

Response

data RegisterImageResponse

Response constructor

registerImageResponse :: RegisterImageResponse

RegisterImageResponse constructor.

The fields accessible through corresponding lenses are:

rirImageId :: Maybe Text

Response lenses

rirImageId :: Lens' RegisterImageResponse (Maybe Text)

The ID of the newly registered AMI.