Network.AWS.EC2.ReleaseAddress

Description

Releases the specified Elastic IP address.

EC2-Classic, default VPC

Releasing an Elastic IP address automatically disassociates it from any instance that it's associated with. To disassociate an Elastic IP address without releasing it, use DisassociateAddress .

Nondefault VPC

You must use DisassociateAddress to disassociate the Elastic IP address before you can release it. Otherwise, Amazon EC2 returns an error (InvalidIPAddress.InUse ).

After releasing an Elastic IP address, it is released to the IP address pool. Be sure to update your DNS records and any servers or devices that communicate with the address. If you attempt to release an Elastic IP address that you already released, you'll get an AuthFailure error if the address is already allocated to another AWS account.

EC2-VPC

After you release an Elastic IP address for use in a VPC, you might be able to recover it. For more information, see AllocateAddress .

Creating a Request

releaseAddress :: ReleaseAddress

Creates a value of ReleaseAddress with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

raAllocationId - [EC2-VPC] The allocation ID. Required for EC2-VPC.

raPublicIP - [EC2-Classic] The Elastic IP address. Required for EC2-Classic.

raDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

data ReleaseAddress

Contains the parameters for ReleaseAddress.

See: releaseAddress smart constructor.

Request Lenses

raAllocationId :: Lens' ReleaseAddress (Maybe Text)

EC2-VPC

The allocation ID. Required for EC2-VPC.

raPublicIP :: Lens' ReleaseAddress (Maybe Text)

EC2-Classic

The Elastic IP address. Required for EC2-Classic.

raDryRun :: Lens' ReleaseAddress (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

Destructuring the Response

releaseAddressResponse :: ReleaseAddressResponse

Creates a value of ReleaseAddressResponse with the minimum fields required to make a request.

data ReleaseAddressResponse

See: releaseAddressResponse smart constructor.