Network.AWS.EC2.ReleaseAddress

Description

Releases the specified Elastic IP address.

After releasing an Elastic IP address, it is released to the IP address pool and might be unavailable to you. Be sure to update your DNS records and any servers or devices that communicate with the address. If you attempt to release an Elastic IP address that you already released, you'll get an AuthFailure error if the address is already allocated to another AWS account.

EC2-Classic, default VPC

Releasing an Elastic IP address automatically disassociates it from any instance that it's associated with. To disassociate an Elastic IP address without releasing it, use DisassociateAddress.

Nondefault VPC

You must use DisassociateAddress to disassociate the Elastic IP address before you try to release it. Otherwise, Amazon EC2 returns an error (InUse).

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-ReleaseAddress.html

Request

data ReleaseAddress

Request constructor

releaseAddress :: ReleaseAddress

ReleaseAddress constructor.

The fields accessible through corresponding lenses are:

raAllocationId :: Maybe Text

raDryRun :: Maybe Bool

raPublicIp :: Maybe Text

Request lenses

raAllocationId :: Lens' ReleaseAddress (Maybe Text)

EC2-VPC

The allocation ID. Required for EC2-VPC.

raDryRun :: Lens' ReleaseAddress (Maybe Bool)

raPublicIp :: Lens' ReleaseAddress (Maybe Text)

EC2-Classic

The Elastic IP address. Required for EC2-Classic.

Response

data ReleaseAddressResponse

Response constructor

releaseAddressResponse :: ReleaseAddressResponse

ReleaseAddressResponse constructor.