Network.AWS.EC2.ReplaceNetworkACLEntry

Description

Replaces an entry (rule) in a network ACL. For more information about network ACLs, see Network ACLs in the Amazon Virtual Private Cloud User Guide .

Creating a Request

replaceNetworkACLEntry

Arguments

:: Bool
rnaeEgress

-> Text
rnaeNetworkACLId

-> Text
rnaeProtocol

-> RuleAction
rnaeRuleAction

-> Int
rnaeRuleNumber

-> ReplaceNetworkACLEntry

Creates a value of ReplaceNetworkACLEntry with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rnaeIPv6CidrBlock - The IPv6 network range to allow or deny, in CIDR notation (for example 2001:bd8:1234:1a00::/64 ).

rnaeICMPTypeCode - ICMP protocol: The ICMP or ICMPv6 type and code. Required if specifying the ICMP (1) protocol, or protocol 58 (ICMPv6) with an IPv6 CIDR block.

rnaePortRange - TCP or UDP protocols: The range of ports the rule applies to. Required if specifying TCP (6) or UDP (17) for the protocol.

rnaeCidrBlock - The IPv4 network range to allow or deny, in CIDR notation (for example 172.16.0.0/24 ).

rnaeDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rnaeEgress - Indicates whether to replace the egress rule. Default: If no value is specified, we replace the ingress rule.

rnaeNetworkACLId - The ID of the ACL.

rnaeProtocol - The IP protocol. You can specify all or -1 to mean all protocols. If you specify all , -1 , or a protocol number other than tcp , udp , or icmp , traffic on all ports is allowed, regardless of any ports or ICMP types or codes you specify. If you specify protocol 58 (ICMPv6) and specify an IPv4 CIDR block, traffic for all ICMP types and codes allowed, regardless of any that you specify. If you specify protocol 58 (ICMPv6) and specify an IPv6 CIDR block, you must specify an ICMP type and code.

rnaeRuleAction - Indicates whether to allow or deny the traffic that matches the rule.

rnaeRuleNumber - The rule number of the entry to replace.

data ReplaceNetworkACLEntry

Contains the parameters for ReplaceNetworkAclEntry.

See: replaceNetworkACLEntry smart constructor.

Request Lenses

rnaeIPv6CidrBlock :: Lens' ReplaceNetworkACLEntry (Maybe Text)

The IPv6 network range to allow or deny, in CIDR notation (for example 2001:bd8:1234:1a00::/64 ).

rnaeICMPTypeCode :: Lens' ReplaceNetworkACLEntry (Maybe ICMPTypeCode)

ICMP protocol: The ICMP or ICMPv6 type and code. Required if specifying the ICMP (1) protocol, or protocol 58 (ICMPv6) with an IPv6 CIDR block.

rnaePortRange :: Lens' ReplaceNetworkACLEntry (Maybe PortRange)

TCP or UDP protocols: The range of ports the rule applies to. Required if specifying TCP (6) or UDP (17) for the protocol.

rnaeCidrBlock :: Lens' ReplaceNetworkACLEntry (Maybe Text)

The IPv4 network range to allow or deny, in CIDR notation (for example 172.16.0.0/24 ).

rnaeDryRun :: Lens' ReplaceNetworkACLEntry (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rnaeEgress :: Lens' ReplaceNetworkACLEntry Bool

Indicates whether to replace the egress rule. Default: If no value is specified, we replace the ingress rule.

rnaeNetworkACLId :: Lens' ReplaceNetworkACLEntry Text

The ID of the ACL.

rnaeProtocol :: Lens' ReplaceNetworkACLEntry Text

The IP protocol. You can specify all or -1 to mean all protocols. If you specify all , -1 , or a protocol number other than tcp , udp , or icmp , traffic on all ports is allowed, regardless of any ports or ICMP types or codes you specify. If you specify protocol 58 (ICMPv6) and specify an IPv4 CIDR block, traffic for all ICMP types and codes allowed, regardless of any that you specify. If you specify protocol 58 (ICMPv6) and specify an IPv6 CIDR block, you must specify an ICMP type and code.

rnaeRuleAction :: Lens' ReplaceNetworkACLEntry RuleAction

Indicates whether to allow or deny the traffic that matches the rule.

rnaeRuleNumber :: Lens' ReplaceNetworkACLEntry Int

The rule number of the entry to replace.

Destructuring the Response

replaceNetworkACLEntryResponse :: ReplaceNetworkACLEntryResponse

Creates a value of ReplaceNetworkACLEntryResponse with the minimum fields required to make a request.

data ReplaceNetworkACLEntryResponse

See: replaceNetworkACLEntryResponse smart constructor.