Network.AWS.EC2.ReplaceNetworkAclEntry

Description

Replaces an entry (rule) in a network ACL. For more information about network ACLs, see Network ACLs in the Amazon Virtual Private Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-ReplaceNetworkAclEntry.html

Request

data ReplaceNetworkAclEntry

Request constructor

replaceNetworkAclEntry

Arguments

:: Text
rnaeNetworkAclId

-> Int
rnaeRuleNumber

-> Text
rnaeProtocol

-> RuleAction
rnaeRuleAction

-> Bool
rnaeEgress

-> Text
rnaeCidrBlock

-> ReplaceNetworkAclEntry

ReplaceNetworkAclEntry constructor.

The fields accessible through corresponding lenses are:

rnaeCidrBlock :: Text

rnaeDryRun :: Maybe Bool

rnaeEgress :: Bool

rnaeIcmpTypeCode :: Maybe IcmpTypeCode

rnaeNetworkAclId :: Text

rnaePortRange :: Maybe PortRange

rnaeProtocol :: Text

rnaeRuleAction :: RuleAction

rnaeRuleNumber :: Int

Request lenses

rnaeCidrBlock :: Lens' ReplaceNetworkAclEntry Text

The network range to allow or deny, in CIDR notation.

rnaeDryRun :: Lens' ReplaceNetworkAclEntry (Maybe Bool)

rnaeEgress :: Lens' ReplaceNetworkAclEntry Bool

Indicates whether to replace the egress rule.

Default: If no value is specified, we replace the ingress rule.

rnaeIcmpTypeCode :: Lens' ReplaceNetworkAclEntry (Maybe IcmpTypeCode)

ICMP protocol: The ICMP type and code. Required if specifying 1 (ICMP) for the protocol.

rnaeNetworkAclId :: Lens' ReplaceNetworkAclEntry Text

The ID of the ACL.

rnaePortRange :: Lens' ReplaceNetworkAclEntry (Maybe PortRange)

TCP or UDP protocols: The range of ports the rule applies to. Required if specifying 6 (TCP) or 17 (UDP) for the protocol.

rnaeProtocol :: Lens' ReplaceNetworkAclEntry Text

The IP protocol. You can specify all or '-1' to mean all protocols.

rnaeRuleAction :: Lens' ReplaceNetworkAclEntry RuleAction

Indicates whether to allow or deny the traffic that matches the rule.

rnaeRuleNumber :: Lens' ReplaceNetworkAclEntry Int

The rule number of the entry to replace.

Response

data ReplaceNetworkAclEntryResponse

Response constructor

replaceNetworkAclEntryResponse :: ReplaceNetworkAclEntryResponse

ReplaceNetworkAclEntryResponse constructor.