Network.AWS.EC2.ReplaceRoute

Description

Replaces an existing route within a route table in a VPC. You must provide only one of the following: Internet gateway or virtual private gateway, NAT instance, NAT gateway, VPC peering connection, network interface, or egress-only Internet gateway.

For more information about route tables, see Route Tables in the Amazon Virtual Private Cloud User Guide .

Creating a Request

replaceRoute

Arguments

:: Text
rrRouteTableId

-> ReplaceRoute

Creates a value of ReplaceRoute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rrVPCPeeringConnectionId - The ID of a VPC peering connection.

rrInstanceId - The ID of a NAT instance in your VPC.

rrEgressOnlyInternetGatewayId - [IPv6 traffic only] The ID of an egress-only Internet gateway.

rrDestinationIPv6CidrBlock - The IPv6 CIDR address block used for the destination match. The value you provide must match the CIDR of an existing route in the table.

rrNatGatewayId - [IPv4 traffic only] The ID of a NAT gateway.

rrNetworkInterfaceId - The ID of a network interface.

rrGatewayId - The ID of an Internet gateway or virtual private gateway.

rrDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rrDestinationCidrBlock - The IPv4 CIDR address block used for the destination match. The value you provide must match the CIDR of an existing route in the table.

rrRouteTableId - The ID of the route table.

data ReplaceRoute

Contains the parameters for ReplaceRoute.

See: replaceRoute smart constructor.

Request Lenses

rrVPCPeeringConnectionId :: Lens' ReplaceRoute (Maybe Text)

The ID of a VPC peering connection.

rrInstanceId :: Lens' ReplaceRoute (Maybe Text)

The ID of a NAT instance in your VPC.

rrEgressOnlyInternetGatewayId :: Lens' ReplaceRoute (Maybe Text)

IPv6 traffic only

The ID of an egress-only Internet gateway.

rrDestinationIPv6CidrBlock :: Lens' ReplaceRoute (Maybe Text)

The IPv6 CIDR address block used for the destination match. The value you provide must match the CIDR of an existing route in the table.

rrNatGatewayId :: Lens' ReplaceRoute (Maybe Text)

IPv4 traffic only

The ID of a NAT gateway.

rrNetworkInterfaceId :: Lens' ReplaceRoute (Maybe Text)

The ID of a network interface.

rrGatewayId :: Lens' ReplaceRoute (Maybe Text)

The ID of an Internet gateway or virtual private gateway.

rrDryRun :: Lens' ReplaceRoute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rrDestinationCidrBlock :: Lens' ReplaceRoute (Maybe Text)

The IPv4 CIDR address block used for the destination match. The value you provide must match the CIDR of an existing route in the table.

rrRouteTableId :: Lens' ReplaceRoute Text

The ID of the route table.

Destructuring the Response

replaceRouteResponse :: ReplaceRouteResponse

Creates a value of ReplaceRouteResponse with the minimum fields required to make a request.

data ReplaceRouteResponse

See: replaceRouteResponse smart constructor.