Network.AWS.EC2.ReplaceRouteTableAssociation

Description

Changes the route table associated with a given subnet in a VPC. After the operation completes, the subnet uses the routes in the new route table it's associated with. For more information about route tables, see Route Tables in the Amazon Virtual Private Cloud User Guide .

You can also use ReplaceRouteTableAssociation to change which table is the main route table in the VPC. You just specify the main route table's association ID and the route table to be the new main route table.

Creating a Request

replaceRouteTableAssociation

Arguments

:: Text
rrtaAssociationId

-> Text
rrtaRouteTableId

-> ReplaceRouteTableAssociation

Creates a value of ReplaceRouteTableAssociation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rrtaDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rrtaAssociationId - The association ID.

rrtaRouteTableId - The ID of the new route table to associate with the subnet.

data ReplaceRouteTableAssociation

Contains the parameters for ReplaceRouteTableAssociation.

See: replaceRouteTableAssociation smart constructor.

Request Lenses

rrtaDryRun :: Lens' ReplaceRouteTableAssociation (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rrtaAssociationId :: Lens' ReplaceRouteTableAssociation Text

The association ID.

rrtaRouteTableId :: Lens' ReplaceRouteTableAssociation Text

The ID of the new route table to associate with the subnet.

Destructuring the Response

replaceRouteTableAssociationResponse

Arguments

:: Int
rrtarsResponseStatus

-> ReplaceRouteTableAssociationResponse

Creates a value of ReplaceRouteTableAssociationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rrtarsNewAssociationId - The ID of the new association.

rrtarsResponseStatus - -- | The response status code.

data ReplaceRouteTableAssociationResponse

Contains the output of ReplaceRouteTableAssociation.

See: replaceRouteTableAssociationResponse smart constructor.

Response Lenses

rrtarsNewAssociationId :: Lens' ReplaceRouteTableAssociationResponse (Maybe Text)

The ID of the new association.

rrtarsResponseStatus :: Lens' ReplaceRouteTableAssociationResponse Int

- | The response status code.