Network.AWS.EC2.ReportInstanceStatus

Description

Submits feedback about the status of an instance. The instance must be in the running state. If your experience with the instance differs from the instance status returned by DescribeInstanceStatus , use ReportInstanceStatus to report your experience with the instance. Amazon EC2 collects this information to improve the accuracy of status checks.

Use of this action does not change the value returned by DescribeInstanceStatus .

Creating a Request

reportInstanceStatus

Arguments

:: ReportStatusType
risStatus

-> ReportInstanceStatus

Creates a value of ReportInstanceStatus with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

risStartTime - The time at which the reported instance health state began.

risEndTime - The time at which the reported instance health state ended.

risDescription - Descriptive text about the health state of your instance.

risDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

risInstances - One or more instances.

risReasonCodes - One or more reason codes that describes the health state of your instance. * instance-stuck-in-state : My instance is stuck in a state. * unresponsive : My instance is unresponsive. * not-accepting-credentials : My instance is not accepting my credentials. * password-not-available : A password is not available for my instance. * performance-network : My instance is experiencing performance problems which I believe are network related. * performance-instance-store : My instance is experiencing performance problems which I believe are related to the instance stores. * performance-ebs-volume : My instance is experiencing performance problems which I believe are related to an EBS volume. * performance-other : My instance is experiencing performance problems. * other : [explain using the description parameter]

risStatus - The status of all instances listed.

data ReportInstanceStatus

Contains the parameters for ReportInstanceStatus.

See: reportInstanceStatus smart constructor.

Request Lenses

risStartTime :: Lens' ReportInstanceStatus (Maybe UTCTime)

The time at which the reported instance health state began.

risEndTime :: Lens' ReportInstanceStatus (Maybe UTCTime)

The time at which the reported instance health state ended.

risDescription :: Lens' ReportInstanceStatus (Maybe Text)

Descriptive text about the health state of your instance.

risDryRun :: Lens' ReportInstanceStatus (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

risInstances :: Lens' ReportInstanceStatus [Text]

One or more instances.

risReasonCodes :: Lens' ReportInstanceStatus [ReportInstanceReasonCodes]

One or more reason codes that describes the health state of your instance. * instance-stuck-in-state : My instance is stuck in a state. * unresponsive : My instance is unresponsive. * not-accepting-credentials : My instance is not accepting my credentials. * password-not-available : A password is not available for my instance. * performance-network : My instance is experiencing performance problems which I believe are network related. * performance-instance-store : My instance is experiencing performance problems which I believe are related to the instance stores. * performance-ebs-volume : My instance is experiencing performance problems which I believe are related to an EBS volume. * performance-other : My instance is experiencing performance problems. * other : [explain using the description parameter]

risStatus :: Lens' ReportInstanceStatus ReportStatusType

The status of all instances listed.

Destructuring the Response

reportInstanceStatusResponse :: ReportInstanceStatusResponse

Creates a value of ReportInstanceStatusResponse with the minimum fields required to make a request.

data ReportInstanceStatusResponse

See: reportInstanceStatusResponse smart constructor.