Network.AWS.EC2.ReportInstanceStatus

Description

Submits feedback about the status of an instance. The instance must be in the running state. If your experience with the instance differs from the instance status returned by DescribeInstanceStatus, use ReportInstanceStatus to report your experience with the instance. Amazon EC2 collects this information to improve the accuracy of status checks.

Use of this action does not change the value returned by DescribeInstanceStatus.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-ReportInstanceStatus.html

Request

data ReportInstanceStatus

Request constructor

reportInstanceStatus

Arguments

:: ReportStatusType
risStatus

-> ReportInstanceStatus

ReportInstanceStatus constructor.

The fields accessible through corresponding lenses are:

risDescription :: Maybe Text

risDryRun :: Maybe Bool

risEndTime :: Maybe UTCTime

risInstances :: [Text]

risReasonCodes :: [ReportInstanceReasonCodes]

risStartTime :: Maybe UTCTime

risStatus :: ReportStatusType

Request lenses

risDescription :: Lens' ReportInstanceStatus (Maybe Text)

Descriptive text about the health state of your instance.

risDryRun :: Lens' ReportInstanceStatus (Maybe Bool)

risEndTime :: Lens' ReportInstanceStatus (Maybe UTCTime)

The time at which the reported instance health state ended.

risInstances :: Lens' ReportInstanceStatus [Text]

One or more instances.

risReasonCodes :: Lens' ReportInstanceStatus [ReportInstanceReasonCodes]

One or more reason codes that describes the health state of your instance.

'instance-stuck-in-state': My instance is stuck in a state.

unresponsive: My instance is unresponsive.

'not-accepting-credentials': My instance is not accepting my credentials.

'password-not-available': A password is not available for my instance.

'performance-network': My instance is experiencing performance problems which I believe are network related.

'performance-instance-store': My instance is experiencing performance problems which I believe are related to the instance stores.

'performance-ebs-volume': My instance is experiencing performance problems which I believe are related to an EBS volume.

'performance-other': My instance is experiencing performance problems.

other: [explain using the description parameter]

risStartTime :: Lens' ReportInstanceStatus (Maybe UTCTime)

The time at which the reported instance health state began.

risStatus :: Lens' ReportInstanceStatus ReportStatusType

The status of all instances listed.

Response

data ReportInstanceStatusResponse

Response constructor

reportInstanceStatusResponse :: ReportInstanceStatusResponse

ReportInstanceStatusResponse constructor.