Network.AWS.EC2.RequestSpotInstances

Description

Creates a Spot instance request. Spot instances are instances that Amazon EC2 launches when the bid price that you specify exceeds the current Spot price. Amazon EC2 periodically sets the Spot price based on available Spot Instance capacity and current Spot instance requests. For more information, see Spot Instance Requests in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

requestSpotInstances

Arguments

:: Text
rsisSpotPrice

-> RequestSpotInstances

Creates a value of RequestSpotInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsisBlockDurationMinutes - The required duration for the Spot instances (also known as Spot blocks), in minutes. This value must be a multiple of 60 (60, 120, 180, 240, 300, or 360). The duration period starts as soon as your Spot instance receives its instance ID. At the end of the duration period, Amazon EC2 marks the Spot instance for termination and provides a Spot instance termination notice, which gives the instance a two-minute warning before it terminates. Note that you can't specify an Availability Zone group or a launch group if you specify a duration.

rsisClientToken - Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency in the Amazon Elastic Compute Cloud User Guide .

rsisInstanceCount - The maximum number of Spot instances to launch. Default: 1

rsisLaunchSpecification - The launch specification.

rsisAvailabilityZoneGroup - The user-specified name for a logical grouping of bids. When you specify an Availability Zone group in a Spot Instance request, all Spot instances in the request are launched in the same Availability Zone. Instance proximity is maintained with this parameter, but the choice of Availability Zone is not. The group applies only to bids for Spot Instances of the same instance type. Any additional Spot instance requests that are specified with the same Availability Zone group name are launched in that same Availability Zone, as long as at least one instance from the group is still active. If there is no active instance running in the Availability Zone group that you specify for a new Spot instance request (all instances are terminated, the bid is expired, or the bid falls below current market), then Amazon EC2 launches the instance in any Availability Zone where the constraint can be met. Consequently, the subsequent set of Spot instances could be placed in a different zone from the original request, even if you specified the same Availability Zone group. Default: Instances are launched in any available Availability Zone.

rsisValidUntil - The end date of the request. If this is a one-time request, the request remains active until all instances launch, the request is canceled, or this date is reached. If the request is persistent, it remains active until it is canceled or this date and time is reached. Default: The request is effective indefinitely.

rsisLaunchGroup - The instance launch group. Launch groups are Spot instances that launch together and terminate together. Default: Instances are launched and terminated individually

rsisType - The Spot instance request type. Default: one-time

rsisValidFrom - The start date of the request. If this is a one-time request, the request becomes active at this date and time and remains active until all instances launch, the request expires, or the request is canceled. If the request is persistent, the request becomes active at this date and time and remains active until it expires or is canceled. Default: The request is effective indefinitely.

rsisDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rsisSpotPrice - The maximum hourly price (bid) for any Spot instance launched to fulfill the request.

data RequestSpotInstances

Contains the parameters for RequestSpotInstances.

See: requestSpotInstances smart constructor.

Request Lenses

rsisBlockDurationMinutes :: Lens' RequestSpotInstances (Maybe Int)

The required duration for the Spot instances (also known as Spot blocks), in minutes. This value must be a multiple of 60 (60, 120, 180, 240, 300, or 360). The duration period starts as soon as your Spot instance receives its instance ID. At the end of the duration period, Amazon EC2 marks the Spot instance for termination and provides a Spot instance termination notice, which gives the instance a two-minute warning before it terminates. Note that you can't specify an Availability Zone group or a launch group if you specify a duration.

rsisClientToken :: Lens' RequestSpotInstances (Maybe Text)

Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to Ensure Idempotency in the Amazon Elastic Compute Cloud User Guide .

rsisInstanceCount :: Lens' RequestSpotInstances (Maybe Int)

The maximum number of Spot instances to launch. Default: 1

rsisLaunchSpecification :: Lens' RequestSpotInstances (Maybe RequestSpotLaunchSpecification)

The launch specification.

rsisAvailabilityZoneGroup :: Lens' RequestSpotInstances (Maybe Text)

The user-specified name for a logical grouping of bids. When you specify an Availability Zone group in a Spot Instance request, all Spot instances in the request are launched in the same Availability Zone. Instance proximity is maintained with this parameter, but the choice of Availability Zone is not. The group applies only to bids for Spot Instances of the same instance type. Any additional Spot instance requests that are specified with the same Availability Zone group name are launched in that same Availability Zone, as long as at least one instance from the group is still active. If there is no active instance running in the Availability Zone group that you specify for a new Spot instance request (all instances are terminated, the bid is expired, or the bid falls below current market), then Amazon EC2 launches the instance in any Availability Zone where the constraint can be met. Consequently, the subsequent set of Spot instances could be placed in a different zone from the original request, even if you specified the same Availability Zone group. Default: Instances are launched in any available Availability Zone.

rsisValidUntil :: Lens' RequestSpotInstances (Maybe UTCTime)

The end date of the request. If this is a one-time request, the request remains active until all instances launch, the request is canceled, or this date is reached. If the request is persistent, it remains active until it is canceled or this date and time is reached. Default: The request is effective indefinitely.

rsisLaunchGroup :: Lens' RequestSpotInstances (Maybe Text)

The instance launch group. Launch groups are Spot instances that launch together and terminate together. Default: Instances are launched and terminated individually

rsisType :: Lens' RequestSpotInstances (Maybe SpotInstanceType)

The Spot instance request type. Default: one-time

rsisValidFrom :: Lens' RequestSpotInstances (Maybe UTCTime)

The start date of the request. If this is a one-time request, the request becomes active at this date and time and remains active until all instances launch, the request expires, or the request is canceled. If the request is persistent, the request becomes active at this date and time and remains active until it expires or is canceled. Default: The request is effective indefinitely.

rsisDryRun :: Lens' RequestSpotInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rsisSpotPrice :: Lens' RequestSpotInstances Text

The maximum hourly price (bid) for any Spot instance launched to fulfill the request.

Destructuring the Response

requestSpotInstancesResponse

Arguments

:: Int
rsirsResponseStatus

-> RequestSpotInstancesResponse

Creates a value of RequestSpotInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsirsSpotInstanceRequests - One or more Spot instance requests.

rsirsResponseStatus - -- | The response status code.

data RequestSpotInstancesResponse

Contains the output of RequestSpotInstances.

See: requestSpotInstancesResponse smart constructor.

Response Lenses

rsirsSpotInstanceRequests :: Lens' RequestSpotInstancesResponse [SpotInstanceRequest]

One or more Spot instance requests.

rsirsResponseStatus :: Lens' RequestSpotInstancesResponse Int

- | The response status code.