Network.AWS.EC2.ResetInstanceAttribute

Description

Resets an attribute of an instance to its default value. To reset the kernel or ramdisk , the instance must be in a stopped state. To reset the sourceDestCheck , the instance can be either running or stopped.

The sourceDestCheck attribute controls whether sourcedestination checking is enabled. The default value is true , which means checking is enabled. This value must be false for a NAT instance to perform NAT. For more information, see <http:docs.aws.amazon.comAmazonVPClatestUserGuideVPC_NAT_Instance.html NAT Instances> in the Amazon Virtual Private Cloud User Guide/ .

Creating a Request

resetInstanceAttribute

Arguments

:: InstanceAttributeName
riaAttribute

-> Text
riaInstanceId

-> ResetInstanceAttribute

Creates a value of ResetInstanceAttribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

riaDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

riaAttribute - The attribute to reset. Important: You can only reset the following attributes: kernel | ramdisk | sourceDestCheck . To change an instance attribute, use ModifyInstanceAttribute .

riaInstanceId - The ID of the instance.

data ResetInstanceAttribute

Contains the parameters for ResetInstanceAttribute.

See: resetInstanceAttribute smart constructor.

Request Lenses

riaDryRun :: Lens' ResetInstanceAttribute (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

riaAttribute :: Lens' ResetInstanceAttribute InstanceAttributeName

The attribute to reset. Important: You can only reset the following attributes: kernel | ramdisk | sourceDestCheck . To change an instance attribute, use ModifyInstanceAttribute .

riaInstanceId :: Lens' ResetInstanceAttribute Text

The ID of the instance.

Destructuring the Response

resetInstanceAttributeResponse :: ResetInstanceAttributeResponse

Creates a value of ResetInstanceAttributeResponse with the minimum fields required to make a request.

data ResetInstanceAttributeResponse

See: resetInstanceAttributeResponse smart constructor.