Network.AWS.EC2.ResetInstanceAttribute

Description

Resets an attribute of an instance to its default value. To reset the kernel or ramdisk, the instance must be in a stopped state. To reset the SourceDestCheck, the instance can be either running or stopped.

The SourceDestCheck attribute controls whether source/destination checking is enabled. The default value is true, which means checking is enabled. This value must be false for a NAT instance to perform NAT. For more information, see NAT Instances in the Amazon Virtual Private Cloud User Guide.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-ResetInstanceAttribute.html

Request

data ResetInstanceAttribute

Request constructor

resetInstanceAttribute

Arguments

:: Text
riaInstanceId

-> InstanceAttributeName
riaAttribute

-> ResetInstanceAttribute

ResetInstanceAttribute constructor.

The fields accessible through corresponding lenses are:

riaAttribute :: InstanceAttributeName

riaDryRun :: Maybe Bool

riaInstanceId :: Text

Request lenses

riaAttribute :: Lens' ResetInstanceAttribute InstanceAttributeName

The attribute to reset.

riaDryRun :: Lens' ResetInstanceAttribute (Maybe Bool)

riaInstanceId :: Lens' ResetInstanceAttribute Text

The ID of the instance.

Response

data ResetInstanceAttributeResponse

Response constructor

resetInstanceAttributeResponse :: ResetInstanceAttributeResponse

ResetInstanceAttributeResponse constructor.