Network.AWS.EC2.RevokeSecurityGroupEgress

Description

EC2-VPC only

Removes one or more egress rules from a security group for EC2-VPC. This action doesn't apply to security groups for use in EC2-Classic. The values that you specify in the revoke request (for example, ports) must match the existing rule's values for the rule to be revoked.

Each rule consists of the protocol and the IPv4 or IPv6 CIDR range or source security group. For the TCP and UDP protocols, you must also specify the destination port or range of ports. For the ICMP protocol, you must also specify the ICMP type and code.

Rule changes are propagated to instances within the security group as quickly as possible. However, a small delay might occur.

Creating a Request

revokeSecurityGroupEgress

Arguments

:: Text
rsgeGroupId

-> RevokeSecurityGroupEgress

Creates a value of RevokeSecurityGroupEgress with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsgeFromPort - The start of port range for the TCP and UDP protocols, or an ICMP type number. We recommend that you specify the port range in a set of IP permissions instead.

rsgeIPPermissions - A set of IP permissions. You can't specify a destination security group and a CIDR IP address range.

rsgeIPProtocol - The IP protocol name or number. We recommend that you specify the protocol in a set of IP permissions instead.

rsgeToPort - The end of port range for the TCP and UDP protocols, or an ICMP type number. We recommend that you specify the port range in a set of IP permissions instead.

rsgeCidrIP - The CIDR IP address range. We recommend that you specify the CIDR range in a set of IP permissions instead.

rsgeSourceSecurityGroupOwnerId - The AWS account number for a destination security group. To revoke outbound access to a destination security group, we recommend that you use a set of IP permissions instead.

rsgeSourceSecurityGroupName - The name of a destination security group. To revoke outbound access to a destination security group, we recommend that you use a set of IP permissions instead.

rsgeDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rsgeGroupId - The ID of the security group.

data RevokeSecurityGroupEgress

Contains the parameters for RevokeSecurityGroupEgress.

See: revokeSecurityGroupEgress smart constructor.

Request Lenses

rsgeFromPort :: Lens' RevokeSecurityGroupEgress (Maybe Int)

The start of port range for the TCP and UDP protocols, or an ICMP type number. We recommend that you specify the port range in a set of IP permissions instead.

rsgeIPPermissions :: Lens' RevokeSecurityGroupEgress [IPPermission]

A set of IP permissions. You can't specify a destination security group and a CIDR IP address range.

rsgeIPProtocol :: Lens' RevokeSecurityGroupEgress (Maybe Text)

The IP protocol name or number. We recommend that you specify the protocol in a set of IP permissions instead.

rsgeToPort :: Lens' RevokeSecurityGroupEgress (Maybe Int)

The end of port range for the TCP and UDP protocols, or an ICMP type number. We recommend that you specify the port range in a set of IP permissions instead.

rsgeCidrIP :: Lens' RevokeSecurityGroupEgress (Maybe Text)

The CIDR IP address range. We recommend that you specify the CIDR range in a set of IP permissions instead.

rsgeSourceSecurityGroupOwnerId :: Lens' RevokeSecurityGroupEgress (Maybe Text)

The AWS account number for a destination security group. To revoke outbound access to a destination security group, we recommend that you use a set of IP permissions instead.

rsgeSourceSecurityGroupName :: Lens' RevokeSecurityGroupEgress (Maybe Text)

The name of a destination security group. To revoke outbound access to a destination security group, we recommend that you use a set of IP permissions instead.

rsgeDryRun :: Lens' RevokeSecurityGroupEgress (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rsgeGroupId :: Lens' RevokeSecurityGroupEgress Text

The ID of the security group.

Destructuring the Response

revokeSecurityGroupEgressResponse :: RevokeSecurityGroupEgressResponse

Creates a value of RevokeSecurityGroupEgressResponse with the minimum fields required to make a request.

data RevokeSecurityGroupEgressResponse

See: revokeSecurityGroupEgressResponse smart constructor.