Network.AWS.EC2.RunScheduledInstances

Description

Launches the specified Scheduled Instances.

Before you can launch a Scheduled Instance, you must purchase it and obtain an identifier using PurchaseScheduledInstances .

You must launch a Scheduled Instance during its scheduled time period. You can't stop or reboot a Scheduled Instance, but you can terminate it as needed. If you terminate a Scheduled Instance before the current scheduled time period ends, you can launch it again after a few minutes. For more information, see Scheduled Instances in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

runScheduledInstances

Arguments

:: ScheduledInstancesLaunchSpecification
rsiLaunchSpecification

-> Text
rsiScheduledInstanceId

-> RunScheduledInstances

Creates a value of RunScheduledInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rsiClientToken - Unique, case-sensitive identifier that ensures the idempotency of the request. For more information, see Ensuring Idempotency .

rsiInstanceCount - The number of instances. Default: 1

rsiDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rsiLaunchSpecification - The launch specification. You must match the instance type, Availability Zone, network, and platform of the schedule that you purchased.

rsiScheduledInstanceId - The Scheduled Instance ID.

data RunScheduledInstances

Contains the parameters for RunScheduledInstances.

See: runScheduledInstances smart constructor.

Request Lenses

rsiClientToken :: Lens' RunScheduledInstances (Maybe Text)

Unique, case-sensitive identifier that ensures the idempotency of the request. For more information, see Ensuring Idempotency .

rsiInstanceCount :: Lens' RunScheduledInstances (Maybe Int)

The number of instances. Default: 1

rsiDryRun :: Lens' RunScheduledInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

rsiLaunchSpecification :: Lens' RunScheduledInstances ScheduledInstancesLaunchSpecification

The launch specification. You must match the instance type, Availability Zone, network, and platform of the schedule that you purchased.

rsiScheduledInstanceId :: Lens' RunScheduledInstances Text

The Scheduled Instance ID.

Destructuring the Response

runScheduledInstancesResponse

Arguments

:: Int
rrsResponseStatus

-> RunScheduledInstancesResponse

Creates a value of RunScheduledInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rrsInstanceIdSet - The IDs of the newly launched instances.

rrsResponseStatus - -- | The response status code.

data RunScheduledInstancesResponse

Contains the output of RunScheduledInstances.

See: runScheduledInstancesResponse smart constructor.

Response Lenses

rrsInstanceIdSet :: Lens' RunScheduledInstancesResponse [Text]

The IDs of the newly launched instances.

rrsResponseStatus :: Lens' RunScheduledInstancesResponse Int

- | The response status code.