Network.AWS.EC2.StartInstances

Description

Starts an Amazon EBS-backed AMI that you've previously stopped.

Instances that use Amazon EBS volumes as their root devices can be quickly stopped and started. When an instance is stopped, the compute resources are released and you are not billed for hourly instance usage. However, your root partition Amazon EBS volume remains, continues to persist your data, and you are charged for Amazon EBS volume usage. You can restart your instance at any time. Each time you transition an instance from stopped to started, Amazon EC2 charges a full instance hour, even if transitions happen multiple times within a single hour.

Before stopping an instance, make sure it is in a state from which it can be restarted. Stopping an instance does not preserve data stored in RAM.

Performing this operation on an instance that uses an instance store as its root device returns an error.

For more information, see Stopping Instances in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

startInstances :: StartInstances

Creates a value of StartInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sAdditionalInfo - Reserved.

sDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

sInstanceIds - One or more instance IDs.

data StartInstances

Contains the parameters for StartInstances.

See: startInstances smart constructor.

Request Lenses

sAdditionalInfo :: Lens' StartInstances (Maybe Text)

Reserved.

sDryRun :: Lens' StartInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

sInstanceIds :: Lens' StartInstances [Text]

One or more instance IDs.

Destructuring the Response

startInstancesResponse

Arguments

:: Int
srsResponseStatus

-> StartInstancesResponse

Creates a value of StartInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srsStartingInstances - Information about one or more started instances.

srsResponseStatus - -- | The response status code.

data StartInstancesResponse

Contains the output of StartInstances.

See: startInstancesResponse smart constructor.

Response Lenses

srsStartingInstances :: Lens' StartInstancesResponse [InstanceStateChange]

Information about one or more started instances.

srsResponseStatus :: Lens' StartInstancesResponse Int

- | The response status code.