Network.AWS.EC2.StopInstances

Description

Stops an Amazon EBS-backed instance.

We don't charge hourly usage for a stopped instance, or data transfer fees; however, your root partition Amazon EBS volume remains, continues to persist your data, and you are charged for Amazon EBS volume usage. Each time you transition an instance from stopped to started, Amazon EC2 charges a full instance hour, even if transitions happen multiple times within a single hour.

You can't start or stop Spot instances, and you can't stop instance store-backed instances.

When you stop an instance, we shut it down. You can restart your instance at any time. Before stopping an instance, make sure it is in a state from which it can be restarted. Stopping an instance does not preserve data stored in RAM.

Stopping an instance is different to rebooting or terminating it. For example, when you stop an instance, the root device and any other devices attached to the instance persist. When you terminate an instance, the root device and any other devices attached during the instance launch are automatically deleted. For more information about the differences between rebooting, stopping, and terminating instances, see Instance Lifecycle in the Amazon Elastic Compute Cloud User Guide .

When you stop an instance, we attempt to shut it down forcibly after a short while. If your instance appears stuck in the stopping state after a period of time, there may be an issue with the underlying host computer. For more information, see Troubleshooting Stopping Your Instance in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

stopInstances :: StopInstances

Creates a value of StopInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

siForce - Forces the instances to stop. The instances do not have an opportunity to flush file system caches or file system metadata. If you use this option, you must perform file system check and repair procedures. This option is not recommended for Windows instances. Default: false

siDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

siInstanceIds - One or more instance IDs.

data StopInstances

Contains the parameters for StopInstances.

See: stopInstances smart constructor.

Request Lenses

siForce :: Lens' StopInstances (Maybe Bool)

Forces the instances to stop. The instances do not have an opportunity to flush file system caches or file system metadata. If you use this option, you must perform file system check and repair procedures. This option is not recommended for Windows instances. Default: false

siDryRun :: Lens' StopInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

siInstanceIds :: Lens' StopInstances [Text]

One or more instance IDs.

Destructuring the Response

stopInstancesResponse

Arguments

:: Int
sirsResponseStatus

-> StopInstancesResponse

Creates a value of StopInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sirsStoppingInstances - Information about one or more stopped instances.

sirsResponseStatus - -- | The response status code.

data StopInstancesResponse

Contains the output of StopInstances.

See: stopInstancesResponse smart constructor.

Response Lenses

sirsStoppingInstances :: Lens' StopInstancesResponse [InstanceStateChange]

Information about one or more stopped instances.

sirsResponseStatus :: Lens' StopInstancesResponse Int

- | The response status code.