Network.AWS.EC2.StopInstances

Description

Stops an Amazon EBS-backed instance. Each time you transition an instance from stopped to started, Amazon EC2 charges a full instance hour, even if transitions happen multiple times within a single hour.

You can't start or stop Spot Instances.

Instances that use Amazon EBS volumes as their root devices can be quickly stopped and started. When an instance is stopped, the compute resources are released and you are not billed for hourly instance usage. However, your root partition Amazon EBS volume remains, continues to persist your data, and you are charged for Amazon EBS volume usage. You can restart your instance at any time.

Before stopping an instance, make sure it is in a state from which it can be restarted. Stopping an instance does not preserve data stored in RAM.

Performing this operation on an instance that uses an instance store as its root device returns an error.

You can stop, start, and terminate EBS-backed instances. You can only terminate instance store-backed instances. What happens to an instance differs if you stop it or terminate it. For example, when you stop an instance, the root device and any other devices attached to the instance persist. When you terminate an instance, the root device and any other devices attached during the instance launch are automatically deleted. For more information about the differences between stopping and terminating instances, see Instance Lifecycle in the Amazon Elastic Compute Cloud UserGuide for Linux.

For more information about troubleshooting, see Troubleshooting StoppingYour Instance in the Amazon Elastic Compute Cloud User Guide for Linux.

http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-StopInstances.html

Request

data StopInstances

Request constructor

stopInstances :: StopInstances

StopInstances constructor.

The fields accessible through corresponding lenses are:

siDryRun :: Maybe Bool

siForce :: Maybe Bool

siInstanceIds :: [Text]

Request lenses

siDryRun :: Lens' StopInstances (Maybe Bool)

siForce :: Lens' StopInstances (Maybe Bool)

Forces the instances to stop. The instances do not have an opportunity to flush file system caches or file system metadata. If you use this option, you must perform file system check and repair procedures. This option is not recommended for Windows instances.

Default: false

siInstanceIds :: Lens' StopInstances [Text]

One or more instance IDs.

Response

data StopInstancesResponse

Response constructor

stopInstancesResponse :: StopInstancesResponse

StopInstancesResponse constructor.

The fields accessible through corresponding lenses are:

sirStoppingInstances :: [InstanceStateChange]

Response lenses

sirStoppingInstances :: Lens' StopInstancesResponse [InstanceStateChange]

Information about one or more stopped instances.