Network.AWS.EC2.TerminateInstances

Description

Shuts down one or more instances. This operation is idempotent; if you terminate an instance more than once, each call succeeds.

If you specify multiple instances and the request fails (for example, because of a single incorrect instance ID), none of the instances are terminated.

Terminated instances remain visible after termination (for approximately one hour).

By default, Amazon EC2 deletes all EBS volumes that were attached when the instance launched. Volumes attached after instance launch continue running.

You can stop, start, and terminate EBS-backed instances. You can only terminate instance store-backed instances. What happens to an instance differs if you stop it or terminate it. For example, when you stop an instance, the root device and any other devices attached to the instance persist. When you terminate an instance, any attached EBS volumes with the DeleteOnTermination block device mapping parameter set to true are automatically deleted. For more information about the differences between stopping and terminating instances, see Instance Lifecycle in the Amazon Elastic Compute Cloud User Guide .

For more information about troubleshooting, see Troubleshooting Terminating Your Instance in the Amazon Elastic Compute Cloud User Guide .

Creating a Request

terminateInstances :: TerminateInstances

Creates a value of TerminateInstances with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tiDryRun - Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

tiInstanceIds - One or more instance IDs. Constraints: Up to 1000 instance IDs. We recommend breaking up this request into smaller batches.

data TerminateInstances

Contains the parameters for TerminateInstances.

See: terminateInstances smart constructor.

Request Lenses

tiDryRun :: Lens' TerminateInstances (Maybe Bool)

Checks whether you have the required permissions for the action, without actually making the request, and provides an error response. If you have the required permissions, the error response is DryRunOperation . Otherwise, it is UnauthorizedOperation .

tiInstanceIds :: Lens' TerminateInstances [Text]

One or more instance IDs. Constraints: Up to 1000 instance IDs. We recommend breaking up this request into smaller batches.

Destructuring the Response

terminateInstancesResponse

Arguments

:: Int
tirsResponseStatus

-> TerminateInstancesResponse

Creates a value of TerminateInstancesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tirsTerminatingInstances - Information about one or more terminated instances.

tirsResponseStatus - -- | The response status code.

data TerminateInstancesResponse

Contains the output of TerminateInstances.

See: terminateInstancesResponse smart constructor.

Response Lenses

tirsTerminatingInstances :: Lens' TerminateInstancesResponse [InstanceStateChange]

Information about one or more terminated instances.

tirsResponseStatus :: Lens' TerminateInstancesResponse Int

- | The response status code.