Network.AWS.EC2.V2014_06_15.AttachNetworkInterface

Description

Attaches a network interface to an instance. Example This example attaches the specified network interface to the specified instance. https:/ec2.amazonaws.com?Action=AttachNetworkInterface &amp;DeviceIndex=1 &amp;InstanceId=i-9cc316fe &amp;NetworkInterfaceId=eni-ffda3197 &amp;AUTHPARAMS &lt;AttachNetworkInterfaceResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;ace8cd1e-e685-4e44-90fb-92014d907212&lt;/requestId&gt; &lt;attachmentId&gt;eni-attach-d94b09b0&lt;/attachmentId&gt; &lt;/AttachNetworkInterfaceResponse&gt;.

Request

data AttachNetworkInterface

Request constructor

mkAttachNetworkInterface

Arguments

:: Text
aniNetworkInterfaceId

-> Text
aniInstanceId

-> Integer
aniDeviceIndex

-> AttachNetworkInterface

Smart constructor for the minimum required parameters to construct a valid AttachNetworkInterface request.

Request lenses

aniNetworkInterfaceId :: Lens' AttachNetworkInterface Text

The ID of the network interface.

aniInstanceId :: Lens' AttachNetworkInterface Text

The ID of the instance.

aniDeviceIndex :: Lens' AttachNetworkInterface Integer

The index of the device for the network interface attachment.

Response

data AttachNetworkInterfaceResponse

Response constructor

mkAttachNetworkInterfaceResponse :: AttachNetworkInterfaceResponse

Smart constructor for the minimum required parameters to construct a valid AttachNetworkInterfaceResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

anirAttachmentId :: Lens' AttachNetworkInterfaceResponse (Maybe Text)

The ID of the network interface attachment.