Network.AWS.EC2.V2014_06_15.AttachVpnGateway

Description

Attaches a virtual private gateway to a VPC. For more information, see Adding a Hardware Virtual Private Gateway to Your VPC in the Amazon Virtual Private Cloud User Guide. Example This example attaches the virtual private gateway with the ID vgw-8db04f81 to the VPC with the ID vpc-1a2b3c4d. https:/ec2.amazonaws.com?Action=AttachVpnGateway &amp;VpnGatewayId=vgw-8db04f81 &amp;VpcId=vpc-1a2b3c4d &amp;AUTHPARAMS &lt;AttachVpnGatewayResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;7a62c49f-347e-4fc4-9331-6e8eEXAMPLE&lt;/requestId&gt; &lt;attachment&gt; &lt;vpcId&gt;vpc-1a2b3c4d&lt;/vpcId&gt; &lt;state&gt;attaching&lt;state&gt; &lt;attachment&gt; &lt;/AttachVpnGatewayResponse&gt;.

Request

data AttachVpnGateway

Request constructor

mkAttachVpnGateway

Arguments

:: Text
avgVpnGatewayId

-> Text
avgVpcId

-> AttachVpnGateway

Smart constructor for the minimum required parameters to construct a valid AttachVpnGateway request.

Request lenses

avgVpnGatewayId :: Lens' AttachVpnGateway Text

The ID of the virtual private gateway.

avgVpcId :: Lens' AttachVpnGateway Text

The ID of the VPC.

Response

data AttachVpnGatewayResponse

Response lenses

avgrVpcAttachment :: Lens' AttachVpnGatewayResponse (Maybe VpcAttachment)

Information about the attachment.