Network.AWS.EC2.V2014_06_15.ConfirmProductInstance

Description

Determines whether a product code is associated with an instance. This action can only be used by the owner of the product code. It is useful when a product code owner needs to verify whether another user's instance is eligible for support. Example This example determines whether the specified product code is associated with the specified instance. https:/ec2.amazonaws.com?Action=ConfirmProductInstance &amp;ProductCode=774F4FF8 &amp;InstanceId=i-10a64379 &amp;AUTHPARAMS &lt;ConfirmProductInstanceResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;/return&gt; &lt;ownerId&gt;111122223333&lt;/ownerId&gt; &lt;/ConfirmProductInstanceResponse&gt;.

Request

data ConfirmProductInstance

Request constructor

mkConfirmProductInstance

Arguments

:: Text
cpiProductCode

-> Text
cpiInstanceId

-> ConfirmProductInstance

Smart constructor for the minimum required parameters to construct a valid ConfirmProductInstance request.

Request lenses

cpiProductCode :: Lens' ConfirmProductInstance Text

The product code. This must be an Amazon DevPay product code that you own.

cpiInstanceId :: Lens' ConfirmProductInstance Text

The ID of the instance.

Response

data ConfirmProductInstanceResponse

Response lenses

cpirsOwnerId :: Lens' ConfirmProductInstanceResponse (Maybe Text)

The AWS account ID of the instance owner. This is only present if the product code is attached to the instance.