Network.AWS.EC2.V2014_06_15.CreateNetworkAcl

Description

Creates a network ACL in a VPC. Network ACLs provide an optional layer of security (in addition to security groups) for the instances in your VPC. For more information about network ACLs, see Network ACLs in the Amazon Virtual Private Cloud User Guide. Example This example creates a network ACL in the specified VPC. The response includes a default entry for egress, and another for ingress, each with a very high rule number. These are the last entries we process to decide whether traffic is allowed in or out of an associated subnet. If the traffic doesn't match any rules with a lower rule number, then these default entries ultimately deny the traffic. https:/ec2.amazonaws.com?Action=CreateNetworkAcl &amp;VpcId=vpc-11ad4878 &amp;AUTHPARAMS 59dbff89-35bd-4eac-99ed-be587EXAMPLE acl-5fb85d36 vpc-11ad4878 false 32767 all deny true 0.0.0.0/0 32767 all deny false 0.0.0.0/0.

Request

data CreateNetworkAcl

Request constructor

mkCreateNetworkAcl

Arguments

:: Text
cnaVpcId

-> CreateNetworkAcl

Smart constructor for the minimum required parameters to construct a valid CreateNetworkAcl request.

Request lenses

cnaVpcId :: Lens' CreateNetworkAcl Text

The ID of the VPC.

Response

data CreateNetworkAclResponse

Response constructor

mkCreateNetworkAclResponse :: CreateNetworkAclResponse

Smart constructor for the minimum required parameters to construct a valid CreateNetworkAclResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

cnarNetworkAcl :: Lens' CreateNetworkAclResponse (Maybe NetworkAcl)

Information about the network ACL.