Network.AWS.EC2.V2014_06_15.CreatePlacementGroup

Description

Creates a placement group that you launch cluster instances into. You must give the group a name that's unique within the scope of your account. For more information about placement groups and cluster instances, see Cluster Instances in the Amazon Elastic Compute Cloud User Guide. Example This example creates a placement group named XYZ-cluster. https:/ec2.amazonaws.com?Action=CreatePlacementGroup &amp;GroupName=XYZ-cluster &amp;Strategy=cluster &amp;AUTHPARAMS &lt;CreatePlacementGroupResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;d4904fd9-82c2-4ea5-adfe-a9cc3EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;CreatePlacementGroupResponse&gt;.

Request

data CreatePlacementGroup

Request constructor

mkCreatePlacementGroup

Arguments

:: Text
cpgGroupName

-> PlacementStrategy
cpgStrategy

-> CreatePlacementGroup

Smart constructor for the minimum required parameters to construct a valid CreatePlacementGroup request.

The fields accessible through corresponding lenses are:

GroupName :: Text

Strategy :: PlacementStrategy

Request lenses

cpgGroupName :: Lens' CreatePlacementGroup Text

A name for the placement group. Constraints: Up to 255 ASCII characters.

cpgStrategy :: Lens' CreatePlacementGroup PlacementStrategy

The placement strategy.

Response

data CreatePlacementGroupResponse

Response constructor

mkCreatePlacementGroupResponse :: CreatePlacementGroupResponse

Smart constructor for the minimum required parameters to construct a valid CreatePlacementGroupResponse response.

This constructor is provided for convenience and testing purposes.