Network.AWS.EC2.V2014_06_15.CreateSpotDatafeedSubscription

Description

Creates a datafeed for Spot Instances, enabling you to view Spot Instance usage logs. You can create one data feed per AWS account. For more information, see Spot Instances in the Amazon Elastic Compute Cloud User Guide. Example This example creates a Spot Instance datafeed for the account. https:/ec2.amazonaws.com?Action=CreateSpotDatafeedSubscription &amp;Bucket=my-s3-bucket &amp;AUTHPARAMS &lt;CreateSpotDatafeedSubscriptionResponse xmlns="http://ec2.amazonaws.com/doc/2013-06-15/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;spotDatafeedSubscription&gt; &lt;ownerId&gt;123456789012&lt;/ownerId&gt; &lt;bucket&gt;my-s3-bucket&lt;/bucket&gt; &lt;prefix&gt;spotdata_&lt;prefix&gt; &lt;state&gt;Active&lt;state&gt; &lt;/spotDatafeedSubscription&gt; &lt;/CreateSpotDatafeedSubscriptionResponse&gt;.

Documentation

createSpotDatafeedSubscription

Arguments

:: Text
_csdsrBucket

-> CreateSpotDatafeedSubscription

Minimum specification for a CreateSpotDatafeedSubscription request.

data CreateSpotDatafeedSubscription

Constructors

CreateSpotDatafeedSubscription

Fields

_csdsrBucket :: Text

The Amazon S3 bucket in which to store the Spot Instance datafeed. Constraints: Must be a valid bucket associated with your AWS account.

_csdsrDryRun :: Maybe Bool

_csdsrPrefix :: Maybe Text

A prefix for the datafeed file names.

csdsrPrefix :: Lens' CreateSpotDatafeedSubscription (Maybe Text)

csdsrDryRun :: Lens' CreateSpotDatafeedSubscription (Maybe Bool)

csdsrBucket :: Lens' CreateSpotDatafeedSubscription Text

data CreateSpotDatafeedSubscriptionResponse

Constructors

CreateSpotDatafeedSubscriptionResponse

Fields

_csdssSpotDatafeedSubscription :: Maybe SpotDatafeedSubscription

The Spot Instance datafeed subscription.

csdssSpotDatafeedSubscription :: Iso' CreateSpotDatafeedSubscriptionResponse (Maybe SpotDatafeedSubscription)