Network.AWS.EC2.V2014_06_15.DescribeInstanceAttribute

Description

Describes the specified attribute of the specified instance. You can specify only one attribute at a time. Example 1 This example lists the instance type of the specified instance. https:/ec2.amazonaws.com?Action=DescribeInstanceAttribute &amp;InstanceId=i-10a64379 &amp;Attribute=instanceType &amp;AUTHPARAMS &lt;DescribeInstanceAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;instanceId&gt;i-10a64379&lt;/instanceId&gt; &lt;instanceType&gt; &lt;value&gt;t1.micro&lt;value&gt; &lt;instanceType&gt; &lt;/DescribeInstanceAttributeResponse&gt; Example 2 This example lists the current value of the InstanceInitiatedShutdownBehavior attribute for the specified instance. https:/ec2.amazonaws.com?Action=DescribeInstanceAttribute &amp;InstanceId=i-10a64379 &amp;Attribute=instanceInitiatedShutdownBehavior &amp;AUTHPARAMS &lt;DescribeInstanceAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;instanceId&gt;i-10a64379&lt;/instanceId&gt; &lt;instanceInitiatedShutdownBehavior&gt; &lt;value&gt;stop&lt;/value&gt; &lt;/instanceInitiatedShutdownBehavior&gt; &lt;/DescribeInstanceAttributeResponse&gt; Example 3 This example lists the current value of the DisableApiTermination attribute for the specified instance. https:/ec2.amazonaws.com?Action=DescribeInstanceAttribute &amp;InstanceId=i-10a64379 &amp;Attribute=disableApiTermination &amp;AUTHPARAMS &lt;DescribeInstanceAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;instanceId&gt;i-10a64379&lt;/instanceId&gt; &lt;disableApiTermination&gt; &lt;value&gt;false&lt;/value&gt; &lt;disableApiTermination&gt; &lt;DescribeInstanceAttributeResponse&gt;.

Documentation

describeInstanceAttribute

Arguments

:: InstanceAttributeName
_diarAttribute

-> Text
_diarInstanceId

-> DescribeInstanceAttribute

Minimum specification for a DescribeInstanceAttribute request.

data DescribeInstanceAttribute

Constructors

DescribeInstanceAttribute

Fields

_diarAttribute :: InstanceAttributeName

The instance attribute.

_diarInstanceId :: Text

The ID of the instance.

_diarDryRun :: Maybe Bool

diarInstanceId :: Lens' DescribeInstanceAttribute Text

diarDryRun :: Lens' DescribeInstanceAttribute (Maybe Bool)

diarAttribute :: Lens' DescribeInstanceAttribute InstanceAttributeName

data DescribeInstanceAttributeResponse

Constructors

DescribeInstanceAttributeResponse

Fields

_iaSourceDestCheck :: Maybe AttributeBooleanValue

Indicates whether source/destination checking is enabled. A value of true means checking is enabled, and false means checking is disabled. This value must be false for a NAT instance to perform NAT.

_iaDisableApiTermination :: Maybe AttributeBooleanValue

If the value is true, you can't terminate the instance through the Amazon EC2 console, CLI, or API; otherwise, you can.

_iaEbsOptimized :: Maybe AttributeBooleanValue

Indicates whether the instance is optimized for EBS I/O.

_iaRamdiskId :: Maybe AttributeValue

The RAM disk ID.

_iaKernelId :: Maybe AttributeValue

The kernel ID.

_iaRootDeviceName :: Maybe AttributeValue

The name of the root device (for example, devsda1).

_iaInstanceType :: Maybe AttributeValue

The instance type.

_iaSriovNetSupport :: Maybe AttributeValue

_iaUserData :: Maybe AttributeValue

The Base64-encoded MIME user data.

_iaInstanceInitiatedShutdownBehavior :: Maybe AttributeValue

Indicates whether an instance stops or terminates when you initiate shutdown from the instance (using the operating system command for system shutdown).

_iaBlockDeviceMappings :: [InstanceBlockDeviceMapping]

The block device mapping of the instance.

_iaProductCodes :: [ProductCode]

A list of product codes.

_iaInstanceId :: Maybe Text

The ID of the instance.

iaUserData :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

iaSriovNetSupport :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

iaSourceDestCheck :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeBooleanValue)

iaRootDeviceName :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

iaRamdiskId :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

iaProductCodes :: Lens' DescribeInstanceAttributeResponse [ProductCode]

iaKernelId :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

iaInstanceType :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

iaInstanceInitiatedShutdownBehavior :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeValue)

iaInstanceId :: Lens' DescribeInstanceAttributeResponse (Maybe Text)

iaEbsOptimized :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeBooleanValue)

iaDisableApiTermination :: Lens' DescribeInstanceAttributeResponse (Maybe AttributeBooleanValue)

iaBlockDeviceMappings :: Lens' DescribeInstanceAttributeResponse [InstanceBlockDeviceMapping]