Network.AWS.EC2.V2014_06_15.DetachInternetGateway

Description

Detaches an Internet gateway from a VPC, disabling connectivity between the Internet and the VPC. The VPC must not contain any running instances with Elastic IP addresses. Example The example detaches the specified Internet gateway from the specified VPC. https:/ec2.amazonaws.com?Action=DetachInternetGateway &amp;InternetGatewayId=igw-eaad4883 &amp;VpcId=vpc-11ad4878 &amp;AUTHPARAMS &lt;DetachInternetGatewayResponse xmlns="http://ec2.amazonaws.com/doc/2014-06-15/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;DetachInternetGatewayResponse&gt;.

Request

data DetachInternetGateway

Request constructor

mkDetachInternetGateway

Arguments

:: Text
dig2InternetGatewayId

-> Text
dig2VpcId

-> DetachInternetGateway

Smart constructor for the minimum required parameters to construct a valid DetachInternetGateway request.

Request lenses

dig2InternetGatewayId :: Lens' DetachInternetGateway Text

The ID of the Internet gateway.

dig2VpcId :: Lens' DetachInternetGateway Text

The ID of the VPC.

Response

data DetachInternetGatewayResponse

Response constructor

mkDetachInternetGatewayResponse :: DetachInternetGatewayResponse

Smart constructor for the minimum required parameters to construct a valid DetachInternetGatewayResponse response.

This constructor is provided for convenience and testing purposes.