Network.AWS.EC2.V2014_06_15.ImportVolume

Description

Creates an import volume task using metadata from the specified disk image. After importing the image, you then upload it using the ec2-import-volume command in the Amazon EC2 command-line interface (CLI) tools. For more information, see Using the Command Line Tools to Import Your Virtual Machine to Amazon EC2 in the Amazon Elastic Compute Cloud User Guide. Example This example creates an import volume task that migrates a Windows Server 2008 SP2 (32-bit) volume into the AWS us-east-1 region. https:/ec2.amazonaws.com?Action=ImportVolume &amp;AvailabilityZone=us-east-1c &amp;Image.Format=VMDK &amp;Image.Bytes=128696320 &amp;Image.ImportManifestUrl=https:/s3.amazonaws.commyawsbucket​a3a5e1b6-590d-43cc-97c1-15c7325d3f41​Win_2008_Server_Data_Center_SP2_32-bit.​vmdkmanifest.xml?AWSAccessKeyId=​AKIAIOSFODNN7EXAMPLE&amp;​Expires=1294855591&amp;​Signature=5snej01TlTtL0uR7KExtEXAMPLE%3D &amp;VolumeSize=8 &amp;AUTHPARAMS&gt; import-i-fh95npoc 2010-12-22T12:01Z 0 us-east-1c VDMK 128696320 https:/s3.amazonaws.commyawsbucket​a3a5e1b6-590d-43cc-97c1-15c7325d3f41​Win_2008_Server_Data_Center_SP2_32-bit.​vmdkmanifest.xml?AWSAccessKeyId=​AKIAIOSFODNN7EXAMPLE&amp;​Expires=1294855591&amp;​Signature=5snej01TlTtL0uR7KExtEXAMPLE%3D ccb1b0536a4a70e86016b85229b5c6b10b14a4eb 8 vol-34d8a2ff active.

Request

data ImportVolume

Request constructor

mkImportVolume

Arguments

:: Text
ivAvailabilityZone

-> DiskImageDetail
ivImage

-> VolumeDetail
ivVolume

-> ImportVolume

Smart constructor for the minimum required parameters to construct a valid ImportVolume request.

Request lenses

ivAvailabilityZone :: Lens' ImportVolume Text

The Availability Zone for the resulting Amazon EBS volume.

ivImage :: Lens' ImportVolume DiskImageDetail

ivDescription :: Lens' ImportVolume (Maybe Text)

An optional description for the volume being imported.

ivVolume :: Lens' ImportVolume VolumeDetail

Response

data ImportVolumeResponse

Response lenses

ivrConversionTask :: Lens' ImportVolumeResponse (Maybe ConversionTask)