Network.AWS.EC2.V2014_06_15.ModifyReservedInstances

Description

Modifies the Availability Zone, instance count, instance type, or network platform (EC2-Classic or EC2-VPC) of your Reserved Instances. The Reserved Instances to be modified must be identical, except for Availability Zone, network platform, and instance type. Example https:/ec2.amazonaws.com?Action=ModifyReservedInstances &amp;ClientToken=myClientToken &amp;ReservedInstancesConfigurationSetItemType.0.AvailabilityZone=us-east-1a &amp;ReservedInstancesConfigurationSetItemType.0.InstanceCount=1 &amp;ReservedInstancesConfigurationSetItemType.0.Platform=EC2-VPC &amp;ReservedInstancesConfigurationSetItemType.0.InstanceType=m1.small &amp;ReservedInstancesId.0=d16f7a91-4d0f-4f19-9d7f-a74d26b1ccfa &amp;AUTHPARAMS bef729b6-0731-4489-8881-2258746ae163 rimod-3aae219d-3d63-47a9-a7e9-e764example.

Request

data ModifyReservedInstances

Request constructor

mkModifyReservedInstances

Arguments

:: [Text]
mriReservedInstancesIds

-> [ReservedInstancesConfiguration]
mriTargetConfigurations

-> ModifyReservedInstances

Smart constructor for the minimum required parameters to construct a valid ModifyReservedInstances request.

Request lenses

mriClientToken :: Lens' ModifyReservedInstances (Maybe Text)

A unique, case-sensitive token you provide to ensure idempotency of your modification request.

mriReservedInstancesIds :: Lens' ModifyReservedInstances [Text]

The IDs of the Reserved Instances to modify.

mriTargetConfigurations :: Lens' ModifyReservedInstances [ReservedInstancesConfiguration]

The configuration settings for the Reserved Instances to modify.

Response

data ModifyReservedInstancesResponse

Response lenses

mrirReservedInstancesModificationId :: Lens' ModifyReservedInstancesResponse (Maybe Text)

The ID for the modification.