Network.AWS.EC2.V2014_06_15.ResetInstanceAttribute

Description

Resets an attribute of an instance to its default value. To reset the kernel or RAM disk, the instance must be in a stopped state. To reset the SourceDestCheck, the instance can be either running or stopped. The SourceDestCheck attribute controls whether source/destination checking is enabled. The default value is true, which means checking is enabled. This value must be false for a NAT instance to perform NAT. For more information, see NAT Instances in the Amazon Virtual Private Cloud User Guide. Example This example resets the sourceDestCheck attribute. https:/ec2.amazonaws.com?Action=ResetInstanceAttribute &amp;InstanceId=i-1a2b3c4d &amp;Attribute=sourceDestCheck &amp;AUTHPARAMS &lt;ResetInstanceAttributeResponse xmlns="http://ec2.amazonaws.com/doc/2013-10-01/"&gt; &lt;requestId&gt;59dbff89-35bd-4eac-99ed-be587EXAMPLE&lt;/requestId&gt; &lt;return&gt;true&lt;return&gt; &lt;ResetInstanceAttributeResponse&gt;.

Documentation

resetInstanceAttribute

Arguments

:: InstanceAttributeName
_riarAttribute

-> Text
_riarInstanceId

-> ResetInstanceAttribute

Minimum specification for a ResetInstanceAttribute request.

data ResetInstanceAttribute

Constructors

ResetInstanceAttribute

Fields

_riarAttribute :: InstanceAttributeName

The attribute to reset.

_riarInstanceId :: Text

The ID of the instance.

_riarDryRun :: Maybe Bool

riarInstanceId :: Lens' ResetInstanceAttribute Text

riarDryRun :: Lens' ResetInstanceAttribute (Maybe Bool)

riarAttribute :: Lens' ResetInstanceAttribute InstanceAttributeName

data ResetInstanceAttributeResponse

Constructors

ResetInstanceAttributeResponse